FICTIONAL IV tobramycin model for test purpose

A FICTIONAL two-compartment population pharmacokinetic model of intravenous tobramycin.

mod_tobramycin_2cpt_fictional

Format

A list of six objects:

ppk_model

A RxODE model implementing the structural population pharmacokinetics model with the individual model (i.e. the model of inter-individual variability) and the covariates

error_model

A function of the residual error model

theta

A named vector of the population estimates of the fixed effects parameters (called THETAs, following NONMEM terminology)

omega

A named square variance-covariance matrix of the population parameters inter-individual variability

covariates

A character vector of the covariates of the model

sigma

The estimates of the parameters of the residual error model

Source

https://www.page-meeting.org/pdf_assets/1954-2017_05_01_poster_Tobramycin.pdf