Estimate the optimal dose for a selected target area under the time-concentration curve (AUC)

Estimates the optimal dose for a selected target area under the time-concentration curve (AUC) given a population pharmacokinetic model, a set of individual parameters, and a target AUC.

poso_dose_auc(
  object,
  time_auc,
  target_auc,
  adapt = FALSE,
  estim_method = "map",
  p = NULL,
  greater_than = TRUE,
  starting_time = 0,
  interdose_interval = NULL,
  add_dose = NULL,
  duration = NULL,
  starting_dose = 100,
  indiv_param = NULL
)

Arguments

object
A posologyr list, created by the posologyr function.

time_auc
Numeric. The target AUC is computed from 0 to time_auc.

target_auc
Numeric. The target AUC.

adapt
A boolean. If estim_method="map", should the estimation be performed with the adaptive MAP method (as opposed to the standard MAP)? A column named AMS is required in the patient record to define the segments for the adaptive MAP approach.

estim_method
A character string. An estimation method to be used for the individual parameters. The default method "map" is the Maximum A Posteriori estimation, the method "prior" simulates from the prior population model, and "sir" uses the Sequential Importance Resampling algorithm to estimate the a posteriori distribution of the individual parameters. This argument is ignored if indiv_param is provided.

p
Numeric. The proportion of the distribution of AUC to consider for the optimization.

greater_than
A boolean. If TRUE: targets a dose leading to a proportion p of the AUCs be greater than target_auc. Respectively, lower if FALSE.

starting_time
Numeric. First point in time of the AUC, for multiple dose regimen. The default is zero.

interdose_interval
Numeric. Time for the interdose interval for multiple dose regimen. Must be provided when add_dose is used.

add_dose
Numeric. Additional doses administered at inter-dose interval after the first dose. Optional.

duration
Numeric. Duration of infusion, for zero-order administrations.

starting_dose
Numeric. Starting dose for the optimization algorithm.

indiv_param
Optional. A set of individual parameters : THETA, estimates of ETA, and covariates.

Value

A list containing the following components:

dose

Numeric. An optimal dose for the selected target AUC.

type_of_estimate

Character string. The type of estimate of the individual parameters. Either a point estimate, or a distribution.

auc_estimate

A vector of numeric estimates of the AUC. Either a single value (for a point estimate of ETA), or a distribution.

indiv_param

A data.frame. The set of individual parameters used for the determination of the optimal dose : THETA, estimates of ETA, and covariates

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the optimal dose to reach an AUC(0-12h) of 45 h.mg/l
poso_dose_auc(patient01_tobra,time_auc=12,target_auc=45)

#> $dose
#> [1] 244.7216
#> 
#> $type_of_estimate
#> [1] "point estimate"
#> 
#> $auc_estimate
#> [1] 45
#> 
#> $indiv_param
#>   THETA_ke CLCREAT WT THETA_V THETA_k12 THETA_k21      ETA_ke      ETA_V
#> 1     0.21      80 65    19.8     0.041      0.12 -0.09255706 0.06709432
#>