Estimate the optimal dose for a selected target concentration at a selected point in time

Estimates the optimal dose for a selected target concentration at a selected point in time given a population pharmacokinetic model, a set of individual parameters, a selected point in time, and a target concentration.

poso_dose_ctime(
  object,
  time_c,
  target_conc,
  param_map = NULL,
  adapt = FALSE,
  starting_dose = 100,
  interdose_interval = NULL,
  add_dose = NULL,
  duration = NULL
)

Arguments

object
A posologyr list, created by the posologyr function.

time_c
Numeric. Point in time for which the dose is to be optimized.

target_conc
Numeric. Target concentration.

param_map
A vector of individual parameters. May be omitted, in which case the poso_estim_map function will be called.

adapt
A boolean. If param_map is omitted, should the estimation be performed with the adaptive MAP method (as opposed to the standard MAP)? A column AMS is required in the patient record to define the segments for the adaptive MAP approach.

starting_dose
Numeric. Starting dose for the optimization algorithm.

interdose_interval
Numeric. Time for the interdose interval for multiple dose regimen. Must be provided when add_dose is used.

add_dose
Numeric. Additional doses administered at inter-dose interval after the first dose. Optional.

duration
Numeric. Duration of infusion, for zero-order administrations.

Value

A numeric optimal dose to reach the target concentration at the selected point in time.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the optimal dose to reach a concentration of 80 mg/l
# one hour after starting the 30-minutes infusion
poso_dose_ctime(patient01_tobra,time_c=1,duration=0.5,target_conc=80)

#> [1] 1648.628