Estimate the Maximum A Posteriori individual parameters

Estimates the Maximum A Posteriori (MAP) individual parameters, also known as Empirical Bayes Estimates (EBE).

poso_estim_map(
  object = NULL,
  return_model = TRUE,
  return_fim = FALSE,
  return_rse = FALSE
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

return_fim
A boolean. Returns the Fisher Information Matrix (FIM) if set to TRUE.

return_rse
A boolean. Returns the relative standard errors (RSE) of the MAP estimates of ETA if set to TRUE.

Value

A named list consisting of one or more of the following elements depending on the input parameters of the function: $eta a named vector of the MAP estimates of the individual values of ETA, $model an RxODE model using the estimated ETAs, $fim the Fisher information matrix, $rse a named vector of RSEs of the MAP estimates of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the Maximum A Posteriori individual parameters
poso_estim_map(patient01_tobra)

#> $eta
#>     ETA_ke      ETA_V    ETA_k12    ETA_k21 
#> -0.6828811 -0.0663349  0.0000000  0.0000000 
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#>   THETA_ke    CLCREAT         WT    THETA_V  THETA_k12  THETA_k21     ETA_ke 
#>  0.2100000 80.0000000 65.0000000 19.8000000  0.0410000  0.1200000 -0.6828811 
#>      ETA_V 
#> -0.0663349 
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2 x 13
#>    time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr periph
#>   <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1     1 -1.39  2.97  -3.19  -2.12 0.126  18.2 0.041  0.12 24.3  442.    13.8
#> 2    14 -1.39  2.97  -3.19  -2.12 0.126  18.2 0.041  0.12  4.52  82.3   46.4
#> # … with 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>