Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object

A posologyr list, created by the posologyr function.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01,n_iter=50,n_chains=2)
#> $eta
#>          ETA_Cl       ETA_Vc        ETA_Ka
#> 1    0.03926763 -0.188756893  0.1896393994
#> 2    0.53744549 -0.169158315  0.2560000513
#> 3    0.71998273 -0.135407799  0.3204400005
#> 4    0.72278720 -0.408261092  0.1524069434
#> 5    0.38991455 -0.790830165  0.0458275680
#> 6    0.65417995 -0.246787939  0.4931803615
#> 7    0.84636250 -0.207523402  0.8727107966
#> 8    0.66365477 -0.243488880  0.9092672406
#> 9    0.40235860 -0.003161621  1.0040249609
#> 10   0.28326539 -0.729230820  0.1403776702
#> 11   0.56140322 -0.219766667  0.3927323960
#> 12   0.64144749 -0.342470687  0.0735283237
#> 13   0.57084326 -0.345171124  0.0375327391
#> 14   0.35145656 -0.747923722  0.1865718213
#> 15   0.43966691 -0.398720455  0.0986135864
#> 16   0.65518178 -0.389980227  0.8378350420
#> 17   0.46204558 -0.194034540  0.3659159912
#> 18   0.46204558 -0.160276470  0.5141421120
#> 19   0.60168564 -0.137561337  0.5029282585
#> 20   0.74530041 -0.205767316  0.1495770828
#> 21  -0.03268826 -0.592623178 -0.2447507829
#> 22   0.68580701 -0.492638879  0.4070733838
#> 23   0.78527418 -0.134301437  0.4880541268
#> 24   0.53075012 -0.733346910 -0.3689969194
#> 25   0.51366734 -0.677647158  0.0533651652
#> 26   0.69390138 -0.714998894 -0.5542648710
#> 27   0.66637861 -0.389439766  0.1954435258
#> 28   0.46047361 -0.272838621  0.0425944761
#> 29   0.68052634 -0.164642204  0.1930095709
#> 30   0.28625883 -0.449265730 -0.2066403274
#> 31   0.61230200 -0.762095829  0.0064243526
#> 32   0.45438875 -0.220880547  0.7111558512
#> 33   0.55104895 -0.332519822 -0.1621549391
#> 34   0.21374670 -0.361151970  0.7160555080
#> 35   0.60166687 -0.373723564  0.1729947591
#> 36   0.65620697 -0.549374230 -0.1911316607
#> 37   0.69017742 -0.250745083  0.0160378865
#> 38   0.46605734 -0.557967704  0.7261591702
#> 39   0.41988507 -0.890937089 -0.4653491629
#> 40   0.28366786 -1.163967050 -0.5132717613
#> 41  -0.02799889 -0.152387084  0.1254344824
#> 42   0.44985925 -0.914372069 -0.3904514155
#> 43   0.52867942 -0.334254087  0.0408791096
#> 44   0.37973158 -0.857753482  0.0408791096
#> 45   0.58542819 -0.671495310 -0.3241788073
#> 46   0.55439888 -0.675067381  0.2914979091
#> 47   0.46747598 -0.216742009  0.1831775848
#> 48   0.73826524  0.215515729  1.5660961476
#> 49   0.52273963 -0.421116022  0.4419044318
#> 50   0.80679857 -0.258267164  0.7234769752
#> 51   0.00000000  0.000000000  0.0000000000
#> 52   0.69498072 -0.461255606 -0.7068519339
#> 53   0.54091095 -0.874502718 -0.0861773070
#> 54   0.32729853 -0.092238888  0.3161653870
#> 55   0.87295701 -0.077363952  0.7280567643
#> 56   0.59825688 -0.015170733  0.8137985671
#> 57   0.47732443 -0.291157303  0.3488387977
#> 58   0.56820843 -0.557942009 -0.1667473898
#> 59   0.44939896 -0.333771255  0.6700227696
#> 60   0.56772951 -0.584908483 -0.0492405139
#> 61   0.46207964 -0.442919807  0.5403885186
#> 62   0.48069457 -0.859506567 -0.3339390149
#> 63   0.60816987 -0.272546492  0.7301667451
#> 64   0.82438640 -0.185170596  0.1461508754
#> 65   0.52026398 -0.505085205 -0.0364893919
#> 66   0.54295579 -0.344190423 -0.1664341832
#> 67   0.51754755 -0.515820099  0.1061896138
#> 68   0.35752384 -0.399486680 -0.0173187181
#> 69   0.68660796 -0.443976263 -0.1700871967
#> 70   0.68971536 -0.234233928  0.4977315935
#> 71   0.69841756 -0.238632971  0.0927516126
#> 72   0.40896650 -0.863221137  0.1690039533
#> 73   0.63048814 -0.643166667  0.0558624583
#> 74   0.33679333 -0.166124662  0.4170367468
#> 75   0.38440918 -0.093759541  0.3383663860
#> 76   0.22824456 -1.156644911 -0.0433045763
#> 77   0.56368738 -0.704270025 -0.4633508334
#> 78   0.48022725 -0.743788355 -0.1244119525
#> 79   0.47166803 -0.266875721  0.2304452867
#> 80  -0.24890496 -0.739155267 -0.0006679183
#> 81   0.45229051 -0.634511397  0.0645328905
#> 82   0.47050079 -0.421235978  0.0131119326
#> 83   0.23864895 -0.727426148 -0.3183925773
#> 84   0.55162003 -0.063515078  0.5512533561
#> 85   0.51361280 -0.455362058  0.0824205264
#> 86   0.26249041 -0.311410386  0.0420984255
#> 87   0.88173066 -0.017193498  0.5384668296
#> 88   0.61222149 -0.325941297 -0.1019496719
#> 89   0.54520013 -0.360665959  0.2061863141
#> 90   0.40076565 -0.525804295 -0.2493611747
#> 91  -0.14974917 -0.433221231  0.1920209380
#> 92   0.46836950 -0.498759935 -0.0077979127
#> 93   0.35355259 -0.880203256 -0.3475043627
#> 94   0.31282928 -0.951155764  0.1124920152
#> 95   0.38559362 -0.925947228 -0.5329128492
#> 96   0.42994491 -0.313548689  0.0786226682
#> 97   0.47875651 -0.201088714  0.2745326634
#> 98   0.64955938 -0.255152536  0.2775474261
#> 99   0.72971610 -0.111874101  0.1434326421
#> 100  0.62248895 -0.247504202  0.0393017883
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl   ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>    <dbl>  <dbl>
#>  1      1        4       70        1 0.0393 -0.189   0.190 
#>  2      2        4       70        1 0.537  -0.169   0.256 
#>  3      3        4       70        1 0.720  -0.135   0.320 
#>  4      4        4       70        1 0.723  -0.408   0.152 
#>  5      5        4       70        1 0.390  -0.791   0.0458
#>  6      6        4       70        1 0.654  -0.247   0.493 
#>  7      7        4       70        1 0.846  -0.208   0.873 
#>  8      8        4       70        1 0.664  -0.243   0.909 
#>  9      9        4       70        1 0.402  -0.00316 1.00  
#> 10     10        4       70        1 0.283  -0.729   0.140 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  4.16  58.0  1.21 0.0718 19.7     8.20e+2 1143.
#> 2      1    14     4    70     1  4.16  58.0  1.21 0.0718 13.7     1.23e-4  793.
#> 3      2     1     4    70     1  6.85  59.1  1.29 0.116  19.7     7.72e+2 1166.
#> 4      2    14     4    70     1  6.85  59.1  1.29 0.116   7.56    3.93e-5  447.
#> 5      3     1     4    70     1  8.22  61.1  1.38 0.134  19.6     7.26e+2 1200.
#> 6      3    14     4    70     1  8.22  61.1  1.38 0.134   5.71    1.21e-5  349.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>