Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e3,n_resample=1e2)
#> $eta
#>        ETA_Cl      ETA_Vc        ETA_Ka
#> 1   0.2679994 -0.13112123  5.967401e-01
#> 2   0.4611585 -0.41091511  4.165087e-02
#> 3   0.7111648 -0.25208914  2.856139e-01
#> 4   0.4682708 -0.73882245 -3.124134e-01
#> 5   0.3937168 -0.64124249  9.705775e-02
#> 6   0.3937168 -0.64124249  9.705775e-02
#> 7   0.2725197 -0.05999858  6.499541e-01
#> 8   0.3937168 -0.64124249  9.705775e-02
#> 9   0.8156592 -0.18909842  9.992836e-02
#> 10  0.5995893 -0.60269376 -5.743076e-01
#> 11  0.6691375 -0.59079497  4.271345e-01
#> 12  0.5766533 -0.65742620 -1.647240e-01
#> 13  0.8594692  0.06125631  4.171537e-01
#> 14  0.3518958 -0.82409806 -2.902659e-01
#> 15  0.5936046 -0.65734376 -2.499049e-01
#> 16  0.5496519 -0.45696681  3.700461e-03
#> 17  0.4682708 -0.73882245 -3.124134e-01
#> 18  0.4032793 -0.42624606  4.532158e-02
#> 19  0.9167175 -0.05896651  1.417114e-01
#> 20  0.4519942 -0.80640277 -1.886841e-01
#> 21  0.4032793 -0.42624606  4.532158e-02
#> 22  0.4652994 -0.26727955  2.626289e-01
#> 23  0.5452217 -0.17237583  4.468084e-01
#> 24  0.4453857 -0.37948184 -5.103432e-02
#> 25  0.5210469 -0.22149257  4.702114e-01
#> 26  0.7446348 -0.23615738 -9.303725e-02
#> 27  0.4519942 -0.80640277 -1.886841e-01
#> 28  0.3518958 -0.82409806 -2.902659e-01
#> 29  0.7489848 -0.29036293  1.232090e-01
#> 30  0.6773654 -0.23087861  8.419649e-01
#> 31  0.4196083 -0.49975075  2.290292e-01
#> 32  0.2679994 -0.13112123  5.967401e-01
#> 33  0.1080232 -0.71959270 -1.345680e-01
#> 34  0.2437701 -0.65438519  2.772438e-01
#> 35  0.5344965 -0.70674608 -5.314845e-01
#> 36  0.5648737 -0.53074719 -7.768911e-02
#> 37  0.5398181 -0.90618020 -4.369519e-01
#> 38  0.5766533 -0.65742620 -1.647240e-01
#> 39  0.5146523 -0.47545167 -3.661290e-03
#> 40  0.5146523 -0.47545167 -3.661290e-03
#> 41  0.5146523 -0.47545167 -3.661290e-03
#> 42  0.6716634 -0.37950288  1.149491e-01
#> 43  0.4502671 -0.79074224  4.627639e-01
#> 44  0.4601762 -0.64703527  3.208253e-01
#> 45  0.5344965 -0.70674608 -5.314845e-01
#> 46  0.5671956 -0.37418593  2.945781e-01
#> 47  0.7111648 -0.25208914  2.856139e-01
#> 48  0.6165432 -0.54224186  2.248402e-01
#> 49  0.5146523 -0.47545167 -3.661290e-03
#> 50  0.5344965 -0.70674608 -5.314845e-01
#> 51  0.6773654 -0.23087861  8.419649e-01
#> 52  0.5766533 -0.65742620 -1.647240e-01
#> 53  0.5648737 -0.53074719 -7.768911e-02
#> 54  0.3937168 -0.64124249  9.705775e-02
#> 55  0.4519942 -0.80640277 -1.886841e-01
#> 56  0.5496519 -0.45696681  3.700461e-03
#> 57  0.6174819 -0.06857429  6.233681e-01
#> 58  0.7489848 -0.29036293  1.232090e-01
#> 59  0.5766533 -0.65742620 -1.647240e-01
#> 60  0.3439026 -0.35373879  6.817992e-01
#> 61  0.5648737 -0.53074719 -7.768911e-02
#> 62  0.5496519 -0.45696681  3.700461e-03
#> 63  0.5671956 -0.37418593  2.945781e-01
#> 64  0.7111648 -0.25208914  2.856139e-01
#> 65  0.5676750 -0.35110421  1.421802e-01
#> 66  0.5425279 -0.21511041  4.128249e-01
#> 67  0.9652298  0.07906872  5.650511e-01
#> 68  0.5496519 -0.45696681  3.700461e-03
#> 69  0.5648737 -0.53074719 -7.768911e-02
#> 70  0.8594692  0.06125631  4.171537e-01
#> 71  0.2844637 -0.52664027 -7.479763e-02
#> 72  0.4682708 -0.73882245 -3.124134e-01
#> 73  0.5452217 -0.17237583  4.468084e-01
#> 74  0.4910282 -0.04503763  5.934957e-01
#> 75  0.3672503 -0.18246860 -1.035631e-06
#> 76  0.2136315 -0.90686326 -3.987184e-01
#> 77  0.4519942 -0.80640277 -1.886841e-01
#> 78  0.3518958 -0.82409806 -2.902659e-01
#> 79  0.5766533 -0.65742620 -1.647240e-01
#> 80  0.4733822 -0.39712517  1.575903e-01
#> 81  0.6716634 -0.37950288  1.149491e-01
#> 82  0.3241315 -0.77792334 -1.097274e-01
#> 83  0.3197599 -0.10474425  2.612698e-01
#> 84  0.5146523 -0.47545167 -3.661290e-03
#> 85  0.5676750 -0.35110421  1.421802e-01
#> 86  0.4601762 -0.64703527  3.208253e-01
#> 87  0.5676750 -0.35110421  1.421802e-01
#> 88  0.5146523 -0.47545167 -3.661290e-03
#> 89  0.7111648 -0.25208914  2.856139e-01
#> 90  0.5425279 -0.21511041  4.128249e-01
#> 91  0.6716634 -0.37950288  1.149491e-01
#> 92  0.5936046 -0.65734376 -2.499049e-01
#> 93  0.5344965 -0.70674608 -5.314845e-01
#> 94  0.3518958 -0.82409806 -2.902659e-01
#> 95  0.4453857 -0.37948184 -5.103432e-02
#> 96  0.4601762 -0.64703527  3.208253e-01
#> 97  0.4682708 -0.73882245 -3.124134e-01
#> 98  0.7111648 -0.25208914  2.856139e-01
#> 99  0.7489848 -0.29036293  1.232090e-01
#> 100 0.7800076 -0.33988351  6.324268e-01
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.268 -0.131   0.597 
#>  2      2        4       70        1  0.461 -0.411   0.0417
#>  3      3        4       70        1  0.711 -0.252   0.286 
#>  4      4        4       70        1  0.468 -0.739  -0.312 
#>  5      5        4       70        1  0.394 -0.641   0.0971
#>  6      6        4       70        1  0.394 -0.641   0.0971
#>  7      7        4       70        1  0.273 -0.0600  0.650 
#>  8      8        4       70        1  0.394 -0.641   0.0971
#>  9      9        4       70        1  0.816 -0.189   0.0999
#> 10     10        4       70        1  0.600 -0.603  -0.574 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  5.23  61.4  1.82 0.0852 23.0  5.30e+2 1413.
#> 2      1    14     4    70     1  5.23  61.4  1.82 0.0852 10.6  2.95e-8  651.
#> 3      2     1     4    70     1  6.34  46.4  1.04 0.137  21.8  9.25e+2 1013.
#> 4      2    14     4    70     1  6.34  46.4  1.04 0.137   7.57 1.20e-3  352.
#> 5      3     1     4    70     1  8.15  54.4  1.33 0.150  21.5  7.51e+2 1169.
#> 6      3    14     4    70     1  8.15  54.4  1.33 0.150   5.29 2.31e-5  288.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>