Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object = NULL, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a dataframe of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> [[1]]
#>           ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1    0.037433542 -0.061208382       0       0
#> 2    0.121807131  0.187765368       0       0
#> 3    0.197609023 -0.058318849       0       0
#> 4    0.380623459 -0.024912476       0       0
#> 5    0.124912770 -0.128968120       0       0
#> 6    0.162351938  0.047276715       0       0
#> 7    0.346921242 -0.027515451       0       0
#> 8    0.299858140  0.034996395       0       0
#> 9    0.202729597 -0.017101161       0       0
#> 10  -0.110615877 -0.122249706       0       0
#> 11  -0.051212457 -0.019314078       0       0
#> 12  -0.423298574 -0.033986303       0       0
#> 13   0.066291229 -0.011284800       0       0
#> 14  -0.183899965 -0.126117075       0       0
#> 15   0.096891602  0.084419971       0       0
#> 16  -0.318883506 -0.070038388       0       0
#> 17  -0.244616050 -0.266723495       0       0
#> 18   0.249608370  0.065088146       0       0
#> 19  -0.074209710 -0.022361194       0       0
#> 20  -0.198842419 -0.040687061       0       0
#> 21   0.187720959  0.149339967       0       0
#> 22   0.069858455  0.103015715       0       0
#> 23   0.539135592  0.097435141       0       0
#> 24  -0.170840487  0.022107251       0       0
#> 25   0.285229204  0.011103174       0       0
#> 26  -0.306305078 -0.009333569       0       0
#> 27  -0.047824811 -0.044299626       0       0
#> 28   0.310516499  0.033266726       0       0
#> 29  -0.096069282  0.103061217       0       0
#> 30   0.193859796 -0.012722076       0       0
#> 31   0.101926356  0.111265670       0       0
#> 32  -0.175711473  0.007965615       0       0
#> 33   0.229401942 -0.012820619       0       0
#> 34   0.047327493 -0.130398994       0       0
#> 35  -0.052212990  0.155355375       0       0
#> 36   0.030154932 -0.085607707       0       0
#> 37  -0.295714812  0.112662353       0       0
#> 38  -0.094811245 -0.035291109       0       0
#> 39   0.001461592  0.065651611       0       0
#> 40   0.045074311  0.133943120       0       0
#> 41  -0.489578740  0.091235409       0       0
#> 42   0.391581606  0.204584554       0       0
#> 43   0.460044805  0.048173234       0       0
#> 44  -0.075852211 -0.061576686       0       0
#> 45  -0.269403216 -0.130035377       0       0
#> 46  -0.190399962 -0.021043917       0       0
#> 47   0.293748340  0.056715454       0       0
#> 48  -0.197172350 -0.091546148       0       0
#> 49  -0.059161768  0.074657562       0       0
#> 50   0.479028999 -0.094573326       0       0
#> 51  -0.016970478 -0.045936785       0       0
#> 52   0.011103256 -0.147960249       0       0
#> 53   0.190016207 -0.140892319       0       0
#> 54  -0.103544653 -0.074828794       0       0
#> 55   0.571281468  0.134079999       0       0
#> 56  -0.544548067 -0.076107813       0       0
#> 57  -0.169868563 -0.004247273       0       0
#> 58  -0.186434520 -0.040242035       0       0
#> 59   0.173039498 -0.082316662       0       0
#> 60   0.199290015 -0.010824657       0       0
#> 61   0.130551374  0.007754021       0       0
#> 62   0.258084362  0.198574866       0       0
#> 63   0.101336370 -0.247994101       0       0
#> 64   0.189218652 -0.058742876       0       0
#> 65   0.351578103  0.036435099       0       0
#> 66  -0.068631711 -0.067824717       0       0
#> 67   0.013666738  0.019338544       0       0
#> 68   0.525113113 -0.196119578       0       0
#> 69  -0.126594114  0.096433493       0       0
#> 70  -0.059445226 -0.130139793       0       0
#> 71  -0.109231994 -0.094699628       0       0
#> 72   0.488399723 -0.011413582       0       0
#> 73   0.070005838  0.171417131       0       0
#> 74   0.168724026 -0.089491113       0       0
#> 75  -0.038489538 -0.117422708       0       0
#> 76   0.002219412 -0.080415772       0       0
#> 77  -0.093805862  0.049518725       0       0
#> 78   0.244481230 -0.002472742       0       0
#> 79   0.101373890  0.027464757       0       0
#> 80  -0.012198038 -0.063065408       0       0
#> 81   0.180934756  0.160968756       0       0
#> 82  -0.504771801 -0.043123922       0       0
#> 83   0.433595317 -0.007714621       0       0
#> 84  -0.100788553 -0.155363481       0       0
#> 85  -0.345994665  0.028885477       0       0
#> 86  -0.224119360  0.077625496       0       0
#> 87   0.014103653 -0.135546222       0       0
#> 88  -0.149945429  0.192589918       0       0
#> 89   0.774157112  0.093947186       0       0
#> 90   0.042831695 -0.002773085       0       0
#> 91  -0.337617079 -0.048864039       0       0
#> 92   0.365030033 -0.140001951       0       0
#> 93  -0.054145847  0.052117258       0       0
#> 94   0.116170797 -0.100670758       0       0
#> 95   0.199193199  0.105672026       0       0
#> 96   0.328778427  0.053775602       0       0
#> 97   0.101928204  0.038616646       0       0
#> 98  -0.013335559  0.017072070       0       0
#> 99  -0.377772687 -0.184910237       0       0
#> 100  0.078350397  0.023993292       0       0
#> 
#> [[2]]
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12  0.0374 -0.0612
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.122   0.188 
#>  3      3     0.21      80    65    19.8     0.041      0.12  0.198  -0.0583
#>  4      4     0.21      80    65    19.8     0.041      0.12  0.381  -0.0249
#>  5      5     0.21      80    65    19.8     0.041      0.12  0.125  -0.129 
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.162   0.0473
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.347  -0.0275
#>  8      8     0.21      80    65    19.8     0.041      0.12  0.300   0.0350
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.203  -0.0171
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.111  -0.122 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21     Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.259  18.3 0.041  0.12 21.9   400. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.259  18.3 0.041  0.12  1.06   19.4
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.281  23.5 0.041  0.12 16.8   394. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.281  23.5 0.041  0.12  0.664  15.6
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.303  18.4 0.041  0.12 21.1   387. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.303  18.4 0.041  0.12  0.692  12.7
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>