Creates a posologyr list from a prior model and an individual event record

Creates a list for a posologyr prior model, an individual event record, and an rxSolve solve object, created from the prior ppk model and the individual event record.

posologyr(prior_model = NULL, dat = NULL, nocb = FALSE)

Arguments

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

posologyr will check the validity of the compiled rxode2 model. If prior_model$ppk_model$isValid() returns FALSE, posologyr will call rxode2 to recompile the model before solving it using prior_model and dat.

Value

A list of eight objects: the posologyr prior population pharmacokinetics model given as prior_model parameter (6 objects), the individual event record (tdm_data) given as dat parameter, and the solved model (solved_ppk_model).

ppk_model

A rxode2 model implementing the structural population pharmacokinetics model with the individual model (i.e. the model of inter-individual variability) and the covariates

error_model

A function of the residual error model

theta

A named vector of the population estimates of the fixed effects parameters (called THETAs, following NONMEM terminology)

omega

A named square variance-covariance matrix of the population parameters inter-individual variability

covariates

A character vector of the covariates of the model

sigma

The estimates of the parameters of the residual error model

tdm_data

A dataframe. The individual subject dataset given as dat parameter

solved_ppk_model

An rxSolve solve object, created with prior_ppk_model and using dat as the event record.