Estimate the Maximum A Posteriori individual parameters

Estimates the Maximum A Posteriori (MAP) individual parameters, also known as Empirical Bayes Estimates (EBE).

poso_estim_map(
  solved_model = solved_ppk_model,
  prior_model = prior_ppk_model,
  dat = dat_posologyr,
  return_model = TRUE
)

Arguments

solved_model
An rxSolve solve object, created with the prior RxODE structural population pharmacokinetics model and the prior estimates of the population parameters from the prior_model, using dat as the event record

prior_model
A posologyr prior population pharmacokinetics model, a list of six objects (see 'Details' for the description of the object)

dat
Dataframe. An individual subject dataset following the structure of NONMEM/RxODE event records

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a named vector of the MAP estimates of the individual values of ETA. If return_model is set to TRUE, a list of the MAP estimates of the individual values of ETA, and a RxODE model using the estimated ETAs.

Details

The default values of the arguments solved_model, prior_model and dat correspond to the objects created by the convenience function load_ppk_model

The posologyr prior population pharmacokinetics model is a list of six objects:

$ppk_model

A RxODE model implementing the structural population pharmacokinetics model with the individual model (i.e. the model of inter-individual variability) and the covariates

$error_model

A function of the residual error model

$psi

A named vector of the population estimates of the fixed effects parameters (called THETAs, following NONMEM terminology)

$omega

A named square variance-covariance matrix of the population parameters inter-individual variability

$covariates

A character vector of the covariates of the model

$xi

The estimates of the parameters of the residual error model

Examples

# df_michel: event table for Michel, following a 30 minutes intravenous
# infusion of tobramycin
df_michel <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Michel's event record
load_ppk_model(prior_model=mod_tobramycin_2cpt_fictional,dat=df_michel)
# estimate the Maximum A Posteriori individual parameters
poso_estim_map()