Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.65120181 -0.273267108  0.459275888
#> 2    0.66069219 -0.248866426  0.549421024
#> 3    0.69069460 -0.420602608  0.006668362
#> 4    0.53797883 -0.493106818  0.038990056
#> 5    0.55169299 -0.250765563  0.316301781
#> 6    0.70575754 -0.252964319  0.572506294
#> 7    0.32344429 -0.544656959 -0.081130544
#> 8    0.03138412 -0.055615527  0.468496235
#> 9   -0.03315853  0.209067647  0.514254617
#> 10   0.12843529  0.370413565  0.674240986
#> 11   0.42002669 -0.619608925  0.395228164
#> 12   0.16253919  0.018197804  0.495726374
#> 13   0.38678980 -0.551684394  0.007253592
#> 14   0.61131335 -0.269376710  0.217027304
#> 15   0.40489996 -0.540615146  0.158578177
#> 16   0.54324536 -0.443019060 -0.015914959
#> 17   0.56378024 -0.311859641 -0.154926427
#> 18   0.63589541 -0.387388084  0.228702411
#> 19   0.46073058 -0.542591188 -0.016818494
#> 20   0.24838775 -0.428480524 -0.005353467
#> 21   0.71587823 -0.069383779  0.898800956
#> 22   0.43744306 -0.730910395  0.541692595
#> 23   0.43714922 -0.683285409 -0.078260573
#> 24   0.53860782 -0.240475794  0.240883853
#> 25   0.63264396 -0.320150745  0.800959845
#> 26   0.72383214  0.086119478  0.822361452
#> 27   0.53928380 -0.214704361  1.146664869
#> 28   0.55221924 -0.009330131  0.734623580
#> 29   0.17970483 -0.110808854  0.426664131
#> 30   0.39681386 -0.850789284  0.280751077
#> 31   0.20550842 -0.935249971 -0.573503235
#> 32   0.25518128 -1.080846104 -0.185530931
#> 33   0.17171316 -0.655561158 -0.245025608
#> 34   0.42356535 -0.507945446 -0.061310174
#> 35   0.32000795 -0.629583357 -0.189755423
#> 36   0.52738074 -0.484755694 -0.310228221
#> 37   0.78653970 -0.275902792  0.649738204
#> 38   0.71345540 -0.328768020  0.316595496
#> 39   0.37507061 -0.760217532 -0.114786303
#> 40   0.80046022 -0.094198022  1.087438737
#> 41   0.72252686  0.004334631  0.637750515
#> 42   0.66391939 -0.248869388  0.539049602
#> 43   0.77467423 -0.266721637  0.379630873
#> 44   0.66268111 -0.234145727  0.086664726
#> 45   0.46435207 -0.335418141 -0.125720496
#> 46   0.60241819 -0.299612662  0.291075685
#> 47   0.43387337 -0.036792142  0.493449116
#> 48   0.51476863  0.008450996  0.772894228
#> 49   0.25820950 -0.303984492  0.250242535
#> 50   0.42774139 -0.569393766 -0.022473142
#> 51   0.00000000  0.000000000  0.000000000
#> 52   0.61467582 -0.312438165  0.504161669
#> 53   0.54754283 -0.370198992  0.988490408
#> 54   0.51883442 -0.715452322 -0.193260004
#> 55   0.85120325 -0.383567389 -0.786379975
#> 56   0.73044107 -0.173852286  0.668239117
#> 57   0.49469549  0.030426913  0.899914280
#> 58   0.36811892 -0.502982509 -0.045203102
#> 59   0.25919685 -0.288431756  0.491912157
#> 60   0.39188739 -0.378938007  0.063179108
#> 61   0.16721684 -0.796357840  0.230394493
#> 62   0.50785551 -0.125674836  0.489363725
#> 63   0.54384689 -0.902721264 -0.687435742
#> 64   0.62592061 -0.505330702 -0.049882332
#> 65   0.67568980 -0.485005468 -0.012298971
#> 66   0.54535939 -0.417329297 -0.105475684
#> 67   0.66682230 -0.496478956  0.053455655
#> 68   0.88533493 -0.139082808  0.346635496
#> 69   0.39976760 -0.556410824  0.141116440
#> 70   0.52123491 -0.402476584  0.402136570
#> 71   0.42800476 -0.549876504  0.184259402
#> 72   0.73532060 -0.474195037  0.034792772
#> 73   0.66383728 -0.326956059  0.561318316
#> 74   0.49328804 -0.301589796  0.073751634
#> 75   0.70120408 -0.209029088  0.395837757
#> 76   0.61275123 -0.342608595  0.537993672
#> 77   0.79374274 -0.148121495  0.427778796
#> 78   0.64682657 -0.294440074  0.287842344
#> 79   0.75987962 -0.257442905  0.912061248
#> 80   0.39717354 -0.890151425 -0.374138111
#> 81   0.43350414 -0.322013763  0.374606859
#> 82   0.52244274 -0.241791619  0.658445545
#> 83   0.77092902 -0.332870687  0.167583633
#> 84   0.55551474 -0.303778941  0.513188296
#> 85   0.33049253 -0.551942039  0.454581736
#> 86   0.55492269 -0.491297992 -0.180564860
#> 87   0.81005344 -0.510306577  0.978760962
#> 88   0.69927540 -0.533671866  0.578156468
#> 89   0.54017310 -0.727394597 -0.014506255
#> 90   0.38468999 -0.770101973 -0.156933740
#> 91   0.61072261 -0.513695434  0.122170833
#> 92   0.43441212 -0.891381142 -0.714400341
#> 93   0.31784375 -0.679759484 -0.189567008
#> 94   0.48252058 -0.636848367  0.568949465
#> 95   0.15747125 -0.618328910 -0.303730109
#> 96   0.71214760 -0.495600504  0.174553045
#> 97   0.49698752 -0.627757936 -0.012386134
#> 98   0.65421094 -0.502036871  0.083115024
#> 99   0.56630291 -0.050317386  0.736078571
#> 100  0.44393783 -0.116359522  0.270660612
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1      1        4       70        1  0.651  -0.273   0.459  
#>  2      2        4       70        1  0.661  -0.249   0.549  
#>  3      3        4       70        1  0.691  -0.421   0.00667
#>  4      4        4       70        1  0.538  -0.493   0.0390 
#>  5      5        4       70        1  0.552  -0.251   0.316  
#>  6      6        4       70        1  0.706  -0.253   0.573  
#>  7      7        4       70        1  0.323  -0.545  -0.0811 
#>  8      8        4       70        1  0.0314 -0.0556  0.468  
#>  9      9        4       70        1 -0.0332  0.209   0.514  
#> 10     10        4       70        1  0.128   0.370   0.674  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  7.67  53.3  1.58 0.144 24.2  6.26e+2 1286.
#> 2      1    14     4    70     1  7.67  53.3  1.58 0.144  5.70 7.24e-7  304.
#> 3      2     1     4    70     1  7.74  54.6  1.73 0.142 24.7  5.63e+2 1346.
#> 4      2    14     4    70     1  7.74  54.6  1.73 0.142  5.67 9.41e-8  310.
#> 5      3     1     4    70     1  7.98  46.0  1.01 0.174 21.2  9.50e+2  974.
#> 6      3    14     4    70     1  7.98  46.0  1.01 0.174  4.83 1.97e-3  222.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>