Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object

A posologyr list, created by the posologyr function.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01,n_iter=50,n_chains=2)
#> $eta
#>           ETA_Cl        ETA_Vc       ETA_Ka
#> 1    0.416019337 -0.3902578777  0.624772262
#> 2    0.432320375 -0.2966178639  0.644796722
#> 3   -0.046898522 -0.0404657023  0.476360264
#> 4    0.622623303 -0.7361009269 -0.369941943
#> 5    0.665343907 -0.3189699287  0.369448960
#> 6    0.656281345 -0.3140743848  0.351950715
#> 7    0.638054231 -0.3166902651  0.461544067
#> 8    0.539187933 -0.1582770675  0.513126834
#> 9    0.915136980  0.2367669311  0.872158040
#> 10   0.677932543 -0.3311873106  0.080620685
#> 11   0.546986396 -0.4947699626 -0.181239211
#> 12   0.746970720 -0.3258886262  0.692913831
#> 13   0.367771064 -0.3777398153  0.134602645
#> 14  -0.268374765 -0.1861140825  0.391295705
#> 15   0.662666073 -0.2775079800  0.177794007
#> 16   0.637794452 -0.2624623093 -0.048136931
#> 17   0.637794452 -0.3749306867  0.008425417
#> 18   0.721409640 -0.3221324134  0.301892164
#> 19   0.590619934 -0.5085287068  0.393381959
#> 20   0.724608383 -0.2014835292  0.266874718
#> 21   0.754515983  0.1029482658  1.028943476
#> 22   0.482441538  0.2008397232  0.726197772
#> 23   0.621494677 -0.3912070156  0.390252913
#> 24   0.550802970 -0.5445453151 -0.312267157
#> 25   0.645386974 -0.3732372207  0.482720987
#> 26   0.703588788 -0.2561724173  0.299562065
#> 27   0.430063553 -0.3465609175  0.357287878
#> 28   0.390684188 -0.6645068398  0.116162049
#> 29   0.364107472 -0.0945282496  0.620044586
#> 30   0.454023845 -0.0499938882  0.682508526
#> 31   0.516310028 -0.5574635801  0.059051391
#> 32   0.502010155 -0.5507109018  0.068763121
#> 33   0.515258657 -0.6138908670  0.187709452
#> 34   0.670563622 -0.4511662677  0.277636217
#> 35   0.382653793 -0.6017413866  0.227219731
#> 36   0.665744426 -0.4760386394  0.358555437
#> 37   0.582403990 -0.6086837996 -0.316431712
#> 38   0.739272806 -0.0628610662  0.566942146
#> 39   0.682512219 -0.6283052474 -0.225567967
#> 40   0.402949681 -0.3489537667  0.432620228
#> 41   0.306048212 -0.3431528919 -0.129940823
#> 42   0.174436838 -0.9314419210 -0.599373981
#> 43   0.230293508 -0.5978804941 -0.185177894
#> 44   0.837816216 -0.2692977262  0.598171477
#> 45   0.810105758 -0.2200132553  0.407144313
#> 46   0.662200357 -0.2200132553  0.407144313
#> 47   0.317356421 -0.4507724323  0.984820721
#> 48   0.417311926 -0.3467554450  0.287728023
#> 49   0.517213097 -0.6910140473 -0.157841079
#> 50   0.411061838 -0.8529742307 -0.316346560
#> 51   0.000000000  0.0000000000  0.000000000
#> 52   0.518135178 -0.3535230785 -0.035181695
#> 53   0.664071704 -0.7047329747 -0.229918007
#> 54   0.470116789 -0.4906795157 -0.109112807
#> 55   0.435849616 -0.5855867936  0.427491497
#> 56   0.514216922 -0.0277436352  0.668338556
#> 57   0.241195037  0.0466160943  0.676560501
#> 58   0.284042264 -0.8251864524  0.390893813
#> 59   0.536465999 -0.5223693732  0.027510546
#> 60   0.553476300 -0.4879044587 -0.381604438
#> 61   0.540452520 -0.3955475380  0.658720590
#> 62   0.601521507 -0.4374729071  0.129909557
#> 63   0.502664607 -0.1844258085  0.536550864
#> 64   0.430790409 -0.6088709959 -0.259923507
#> 65   0.538720476 -0.7846091582  0.443690516
#> 66  -0.290960082 -0.5545462390  0.169010651
#> 67   0.353456349 -0.9092447088 -0.138978178
#> 68   0.494804919 -0.6798125935  0.106657491
#> 69   0.278505368 -0.2975321574  0.579436497
#> 70   0.475091298 -0.5241493482  0.078749716
#> 71   0.546186171 -0.4744800803  0.109129515
#> 72   0.565366425 -0.0134268936  0.641595130
#> 73   0.146241875 -0.6265786013  0.121274701
#> 74   0.550856639 -0.6299571919 -0.257587888
#> 75   0.696291290  0.0234798532  1.282101572
#> 76   0.612339782 -0.0813289637  0.623532155
#> 77  -0.308468929 -0.4355959861  0.001465851
#> 78   0.523284886 -0.3820813303 -0.137315104
#> 79   0.595505164 -0.5104705180 -0.135773094
#> 80   0.592287904 -0.6970521595 -0.454971365
#> 81   0.269985444  0.0004669408  0.946560454
#> 82   0.432240664 -0.5216702617 -0.133428625
#> 83   0.422710634 -0.4204325672  0.075089686
#> 84   0.130027756 -1.3695058602 -0.482367737
#> 85   0.594533220 -0.1889614129  0.698500684
#> 86   0.157654939 -0.2345571777  0.217816341
#> 87   0.431518469 -0.8545690830 -0.332640215
#> 88   0.298895362 -0.6250880231  0.098705760
#> 89   0.511737345 -0.7304173510  0.147765432
#> 90   0.170955930 -0.9633149485 -0.693502600
#> 91   0.407821136 -0.4644838495  0.258541756
#> 92   0.496804742 -0.3770899055  0.014278484
#> 93   0.688022183 -0.4812440185  0.087399352
#> 94   0.728726058 -0.4380779116  0.083188629
#> 95   0.005517799 -0.1801129877  0.399683992
#> 96   0.519811319 -0.0631444185  0.944602854
#> 97   0.249623476 -0.0958928475  0.841843788
#> 98   0.447775756 -0.7985109153  0.095041931
#> 99   0.447975022 -0.4648224938  0.074005272
#> 100  0.393545618 -0.4455849416  0.154190052
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.416  -0.390   0.625 
#>  2      2        4       70        1  0.432  -0.297   0.645 
#>  3      3        4       70        1 -0.0469 -0.0405  0.476 
#>  4      4        4       70        1  0.623  -0.736  -0.370 
#>  5      5        4       70        1  0.665  -0.319   0.369 
#>  6      6        4       70        1  0.656  -0.314   0.352 
#>  7      7        4       70        1  0.638  -0.317   0.462 
#>  8      8        4       70        1  0.539  -0.158   0.513 
#>  9      9        4       70        1  0.915   0.237   0.872 
#> 10     10        4       70        1  0.678  -0.331   0.0806
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  6.06  47.4  1.87 0.128  29.6  5.11e+2 1402.
#> 2      1    14     4    70     1  6.06  47.4  1.87 0.128   7.80 1.44e-8  370.
#> 3      2     1     4    70     1  6.16  52.0  1.91 0.118  27.3  4.97e+2 1421.
#> 4      2    14     4    70     1  6.16  52.0  1.91 0.118   8.04 8.49e-9  418.
#> 5      3     1     4    70     1  3.82  67.2  1.61 0.0568 20.1  6.14e+2 1350.
#> 6      3    14     4    70     1  3.82  67.2  1.61 0.0568 14.1  4.98e-7  950.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>