Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object

A posologyr list, created by the posologyr function.

return_model

A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01,n_iter=100,n_chains=1)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.60406708 -0.433013436  0.026529886
#> 2    0.57877233 -0.379211686  0.208372321
#> 3    0.55928293 -0.421542048  0.856239972
#> 4    0.52608549 -0.353082621  0.362278079
#> 5    0.70306022 -0.086110808  0.657288211
#> 6    0.70325033 -0.145806225  0.499464762
#> 7    0.63099672 -0.183851386  0.288790513
#> 8    0.47352081 -0.092917300  0.435600692
#> 9    0.65919795  0.019777900  0.862831771
#> 10   0.50675453 -0.286542781  0.765727564
#> 11   0.88884849  0.105241630  0.904947622
#> 12   0.60365532 -0.453258751  0.729427652
#> 13   0.56167302  0.178971351  0.633103201
#> 14   0.65604920 -0.395410667  0.364296007
#> 15   0.57298835  0.076896036  0.593077067
#> 16   0.71565253 -0.402847899  0.375975334
#> 17   0.43336729 -0.251281617  0.388826992
#> 18   0.70447388 -0.580798412 -0.266853316
#> 19   0.46891978 -0.281593736  0.570173059
#> 20   0.65868734 -0.312681242  0.122802501
#> 21   0.28947410 -0.202183139  0.457518033
#> 22   0.67475926 -0.265651881  0.378770526
#> 23   0.38510354 -0.397015061 -0.027442803
#> 24   0.36473898 -0.771318439 -0.077944090
#> 25   0.35196595 -0.967387446 -0.136036390
#> 26   0.37297867 -0.924508770  0.250121858
#> 27   0.44140869 -0.555942763  0.327974569
#> 28   0.49833237 -0.449091313  0.197394518
#> 29   0.82868164 -0.163805423  0.412522486
#> 30   0.73115811 -0.122066487  0.744081489
#> 31   0.69453953 -0.503229246 -0.023994864
#> 32   0.60737392 -0.577852790 -0.105992906
#> 33   0.26135897 -0.874720479  0.267915914
#> 34   0.63962352 -0.482508001  0.391420612
#> 35   0.52153360 -0.634445397 -0.055538807
#> 36   0.63902708 -0.637418702 -0.166526766
#> 37   0.46521269 -0.543609849  0.089049818
#> 38   0.44159998 -0.360582714 -0.099034938
#> 39   0.44110240 -0.557704226 -0.037039400
#> 40   0.41555064 -0.263625574  0.657111303
#> 41   0.57118486 -0.164413299  0.389152634
#> 42   0.62765642 -0.565306296  0.387297266
#> 43   0.40032715 -0.306104159  0.396666973
#> 44   0.61975696 -0.390594861  0.122645010
#> 45   0.51892472 -0.815885350 -0.324992145
#> 46   0.45555991 -0.839285870  0.069479800
#> 47   0.61851885 -0.555830615 -0.442407427
#> 48   0.51522526 -0.577841549 -0.138743884
#> 49   0.60718627 -0.315322496  0.444566322
#> 50   0.41880980 -0.471648847 -0.110262766
#> 51   0.40295473 -0.947360686 -0.338549697
#> 52   0.45415164 -0.633962663 -0.185671202
#> 53   0.31928391 -0.428936026  0.176799204
#> 54   0.49720037 -0.237095228  0.622886178
#> 55   0.83334183 -0.241852424  0.209004550
#> 56   0.46313777 -0.563950947  0.497716401
#> 57   0.38051700 -0.549570156  0.305157666
#> 58   0.61182907 -0.581050337  0.201163516
#> 59   0.50561362 -0.712185727  0.185142870
#> 60   0.50561362 -0.712185727  0.185142870
#> 61   0.38524302 -0.299541374  0.311488228
#> 62   0.61146974 -0.485359431 -0.257594699
#> 63   0.60041252 -0.558512101  0.256308597
#> 64   0.43645096 -0.337804296  0.311727451
#> 65   0.22146459  0.040102491  0.378487632
#> 66   0.22509292 -0.189643712  0.527072665
#> 67   0.39777259 -0.934838062 -0.411297826
#> 68   0.41453973 -0.724980234 -0.224222485
#> 69   0.90363597  0.017529748  0.639910784
#> 70   0.65519640 -0.367900551  0.330085066
#> 71   0.55394854 -0.427187694  0.253540241
#> 72   0.61379089 -0.450819801  0.936431646
#> 73   0.78834502 -0.122931773  0.474590413
#> 74   0.36276009 -0.177455042  0.159510316
#> 75   0.65365290 -0.517731246  0.009539724
#> 76   0.62841200 -0.200695619  0.541372179
#> 77   0.79620785 -0.073132525  0.520925671
#> 78   0.64724214 -0.001129221  0.405426509
#> 79   0.28312113 -0.533599175 -0.384423232
#> 80   0.54240102 -0.405658822  0.648357003
#> 81   0.46879795 -0.382860761  0.247655153
#> 82   0.16306785 -0.295024994  0.213674210
#> 83   0.41010306 -0.577600853 -0.187650498
#> 84   0.58253473 -0.262152811  0.699681930
#> 85   0.72223995  0.060961748  1.106174740
#> 86  -0.10479347 -0.995666674 -0.235343276
#> 87   0.38340992 -0.763363332 -0.256435202
#> 88   0.49093138 -0.598778777 -0.093409520
#> 89   0.41528793 -0.992627419 -0.383807702
#> 90   0.48243944 -0.727559567 -0.253023230
#> 91   0.38035411 -0.717302747 -0.108011932
#> 92   0.61575057 -0.486931339  0.783756947
#> 93   0.47482386 -0.572067571  0.650494951
#> 94  -0.02270018 -0.653661830  0.098762009
#> 95   0.49936961 -0.095715893  0.431685386
#> 96   0.61500312 -0.601703707  0.215095538
#> 97   0.48443271 -0.648204313 -0.042676297
#> 98   0.69189072 -0.475965193  0.208978860
#> 99   0.59708373 -0.307300778  0.202473372
#> 100  0.74843492 -0.279824225  0.455863998
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1  0.604 -0.433  0.0265
#>  2      2        4       70        1  0.579 -0.379  0.208 
#>  3      3        4       70        1  0.559 -0.422  0.856 
#>  4      4        4       70        1  0.526 -0.353  0.362 
#>  5      5        4       70        1  0.703 -0.0861 0.657 
#>  6      6        4       70        1  0.703 -0.146  0.499 
#>  7      7        4       70        1  0.631 -0.184  0.289 
#>  8      8        4       70        1  0.474 -0.0929 0.436 
#>  9      9        4       70        1  0.659  0.0198 0.863 
#> 10     10        4       70        1  0.507 -0.287  0.766 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  7.32  45.4  1.03 0.161 21.9  9.36e+2  992.
#> 2      1    14     4    70     1  7.32  45.4  1.03 0.161  5.70 1.49e-3  259.
#> 3      2     1     4    70     1  7.14  47.9  1.23 0.149 23.3  8.07e+2 1118.
#> 4      2    14     4    70     1  7.14  47.9  1.23 0.149  6.13 8.97e-5  294.
#> 5      3     1     4    70     1  7.00  45.9  2.35 0.152 33.1  3.62e+2 1521.
#> 6      3    14     4    70     1  7.00  45.9  2.35 0.152  5.73 1.16e-9  263.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>