Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object

A posologyr list, created by the posologyr function.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01,n_iter=50,n_chains=2)
#> $eta
#>         ETA_Cl        ETA_Vc       ETA_Ka
#> 1    0.3238863 -0.5585124254  0.209157180
#> 2    0.6017776 -0.2683308105  0.623888571
#> 3    0.5611388 -0.7892369213  0.315536932
#> 4    0.3504784 -0.5569167085  0.315536932
#> 5    0.6083799 -0.5916193707  0.184732474
#> 6    0.8711761 -0.4628650986  0.282646620
#> 7    0.2907463 -0.4318604038  0.237837693
#> 8    0.1595889 -0.2269152778  0.497627708
#> 9    0.7133925 -0.3412328305  0.251301514
#> 10   0.1870007  0.0925710020  0.961906621
#> 11   0.5439533  0.1194961301  0.718975081
#> 12   0.9059145 -0.4212330276  1.012582903
#> 13   0.6854089 -0.3308852741  0.553763363
#> 14   0.5056763 -0.4947213951  0.286557147
#> 15   0.4898571 -0.6497817132  0.231320255
#> 16   0.6155042 -0.2792322594  0.155788725
#> 17   0.3935371 -0.4342835756  0.182620447
#> 18   0.6842239 -0.0887674287  0.650337031
#> 19   0.7389174  0.0122226752  0.640299180
#> 20   0.2756066  0.0004049675  0.710214490
#> 21   0.3130429 -0.2994741522  0.563676956
#> 22   0.1392189 -0.0041115355  0.860035645
#> 23   0.2898325 -0.7053913746 -0.447954794
#> 24   0.1786605 -0.4460793380  0.028630842
#> 25   0.2118972 -0.6767753119 -0.013984120
#> 26   0.4395534 -0.5959675025  0.748863274
#> 27   0.6270103 -0.1038027466  0.712193834
#> 28   0.3077762 -0.6774449696 -0.146149240
#> 29   0.5895113 -0.3777764349  0.180361617
#> 30   0.5844998 -0.3483122524  0.177369630
#> 31   0.4359790 -0.0063451751  0.471143703
#> 32   0.6243621 -0.4371553120  0.592078977
#> 33   0.4992273 -0.6709365201  0.223530864
#> 34   0.6840308 -0.1743361482  1.007113276
#> 35   0.5365004 -0.1602087940  0.388601484
#> 36   0.7186238 -0.3751035993  0.677415382
#> 37   0.3973808 -0.4347011609  0.839438005
#> 38   0.5485218 -0.3186967277  0.630186750
#> 39   0.4479711 -0.7924255336 -0.244623814
#> 40   0.6145049 -0.6777925673 -0.104850534
#> 41   0.1319254 -1.0488117242 -0.537889422
#> 42   0.4535826 -0.4262845391 -0.372451030
#> 43   0.4140482 -0.4608446476  0.718529969
#> 44   0.5085390 -0.6340906962 -0.157165960
#> 45   0.4326348 -0.8493890284 -0.181601866
#> 46   0.5353348 -0.4827527929 -0.113430937
#> 47   0.3645399 -0.7723512360 -0.173015431
#> 48   0.2432677 -0.0495813801  0.626233763
#> 49   0.6551392 -0.4579111019  0.142203685
#> 50   0.6460410 -0.5376533503  0.475771320
#> 51   0.0000000  0.0000000000  0.000000000
#> 52   0.2116665 -0.3454099113 -0.384671469
#> 53   0.6785068 -0.0502909108  0.762152652
#> 54   0.5479687 -0.6157738520  0.663919184
#> 55   0.3043990 -0.4625885690 -0.068128632
#> 56   0.5165459 -0.2251918071  0.016281467
#> 57   0.5559584 -0.1407629259  0.598384156
#> 58   0.3181573 -1.0340981323  0.014939212
#> 59   0.4416977 -0.7764961405 -0.689452969
#> 60   0.8820281 -0.1198204576  1.388197497
#> 61   0.5365055 -0.3359773366 -0.099262437
#> 62   0.4141312 -0.6934280451  0.383799426
#> 63   0.4654175 -0.4648016228  0.125857433
#> 64   0.3303768 -0.5415317564 -0.052116575
#> 65   0.3864776 -0.4201019359 -0.004631594
#> 66   0.8508690 -0.2095702851  0.590071103
#> 67   0.7951547 -0.3009603517  0.009676640
#> 68   0.4467241 -0.5377492315  0.144945100
#> 69   0.2759834 -0.6737724292  0.062978150
#> 70   0.2616939 -0.1080786893  0.577118605
#> 71   0.5791030 -0.3264128336  0.514876245
#> 72   0.3086654 -0.0133096235  0.662795187
#> 73   0.4224723 -0.4799271676 -0.204535840
#> 74   0.4391923 -0.8533047620 -0.111972090
#> 75   0.6724919 -0.3434447595  0.389469207
#> 76   0.4920381 -0.2507474070  0.632043077
#> 77   0.6450751 -0.1177837412  0.149754733
#> 78   0.6477189 -0.1269213152  0.601568600
#> 79   0.5630401 -0.2388083013  0.290737118
#> 80   0.7080336 -0.3854677371  0.304632304
#> 81   0.4437239 -0.0781802348  0.282863178
#> 82   0.2964664 -0.1778830679  0.919741189
#> 83   0.5021206 -0.0845423913  0.780023144
#> 84   0.6761041 -0.1427978275 -0.038134668
#> 85   0.7678501 -0.0499908065  0.305437396
#> 86  -0.1127638 -0.6652859474  0.055592146
#> 87   0.7506229 -0.1793914653  0.659460999
#> 88   0.4257469 -0.7901385678 -0.219103462
#> 89   0.6543571 -0.5264065110  0.116460391
#> 90   0.3693862 -1.2575896755 -0.352844667
#> 91   0.3239895 -0.7449559341  0.404806787
#> 92   0.1685155 -0.5195756806 -0.525269349
#> 93   0.7251414 -0.0541899142  0.579559448
#> 94   0.7452169  0.0623578078  0.571063542
#> 95   0.7202601 -0.5271995433  0.200187655
#> 96   0.6804118 -0.7040034327  0.378526683
#> 97   0.2381274 -0.8215143326 -0.615115182
#> 98   0.5914397 -0.3003144359  0.936368733
#> 99   0.4675894 -0.5873164217  0.094959231
#> 100  0.4349647 -0.7021722092  0.347033093
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1  0.324 -0.559   0.209
#>  2      2        4       70        1  0.602 -0.268   0.624
#>  3      3        4       70        1  0.561 -0.789   0.316
#>  4      4        4       70        1  0.350 -0.557   0.316
#>  5      5        4       70        1  0.608 -0.592   0.185
#>  6      6        4       70        1  0.871 -0.463   0.283
#>  7      7        4       70        1  0.291 -0.432   0.238
#>  8      8        4       70        1  0.160 -0.227   0.498
#>  9      9        4       70        1  0.713 -0.341   0.251
#> 10     10        4       70        1  0.187  0.0926  0.962
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  5.53  40.0  1.23 0.138 28.1  8.06e+2 1123.
#> 2      1    14     4    70     1  5.53  40.0  1.23 0.138  8.42 8.85e-5  337.
#> 3      2     1     4    70     1  7.30  53.5  1.87 0.136 26.1  5.11e+2 1396.
#> 4      2    14     4    70     1  7.30  53.5  1.87 0.136  6.18 1.47e-8  331.
#> 5      3     1     4    70     1  7.01  31.8  1.37 0.221 36.2  7.29e+2 1152.
#> 6      3    14     4    70     1  7.01  31.8  1.37 0.221  3.62 1.33e-5  115.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>