Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> $eta
#>            ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1    0.0028240733 -9.132942e-02       0       0
#> 2   -0.2489454084  1.230239e-02       0       0
#> 3   -0.2489454084  1.230239e-02       0       0
#> 4   -0.0925856057  6.614605e-03       0       0
#> 5   -0.1781639924  2.117573e-01       0       0
#> 6   -0.0207046995  1.506828e-01       0       0
#> 7    0.0007591407 -2.184966e-02       0       0
#> 8   -0.0624978169 -5.034992e-02       0       0
#> 9   -0.0624978169 -1.955098e-02       0       0
#> 10  -0.1366116769  9.230302e-02       0       0
#> 11  -0.1366116769  9.230302e-02       0       0
#> 12  -0.1096378489  1.137113e-01       0       0
#> 13   0.0101027869  5.899011e-02       0       0
#> 14   0.0850371558 -1.277500e-01       0       0
#> 15  -0.1941980218  1.130402e-01       0       0
#> 16  -0.1039537318  1.555941e-01       0       0
#> 17  -0.1480552761  1.029628e-01       0       0
#> 18  -0.2332220459  1.471527e-01       0       0
#> 19  -0.1310229736  2.470534e-01       0       0
#> 20  -0.2113737314  2.935424e-01       0       0
#> 21  -0.0934560373  2.207580e-01       0       0
#> 22  -0.1283760974  1.798664e-01       0       0
#> 23  -0.0014231989  1.476839e-02       0       0
#> 24  -0.1345349240  1.176902e-01       0       0
#> 25  -0.0532615384 -1.452925e-01       0       0
#> 26  -0.0575695009 -7.813169e-02       0       0
#> 27  -0.1196716059 -1.253926e-01       0       0
#> 28  -0.1665738825  1.039369e-01       0       0
#> 29  -0.1817818729  1.152903e-01       0       0
#> 30  -0.1293794423  6.740359e-02       0       0
#> 31  -0.0868761792 -1.554233e-01       0       0
#> 32   0.0604633981 -1.590911e-01       0       0
#> 33  -0.1206688372  1.647560e-01       0       0
#> 34  -0.1372700370  5.458725e-02       0       0
#> 35  -0.1276055074  3.502155e-01       0       0
#> 36  -0.3754475353 -8.420146e-02       0       0
#> 37  -0.1186629710 -5.032287e-03       0       0
#> 38  -0.1480655520  5.316428e-02       0       0
#> 39  -0.1480655520  1.777902e-01       0       0
#> 40  -0.0809981555 -5.004507e-02       0       0
#> 41  -0.1316673653  9.258328e-02       0       0
#> 42  -0.0738432965  9.064428e-02       0       0
#> 43   0.0783273718 -1.399295e-01       0       0
#> 44  -0.0748657770  3.117769e-02       0       0
#> 45   0.0113657724  3.117769e-02       0       0
#> 46  -0.1869188765  2.142252e-01       0       0
#> 47  -0.0640940738 -1.010721e-01       0       0
#> 48   0.0023152252  1.136332e-01       0       0
#> 49  -0.0585427715 -4.079201e-04       0       0
#> 50  -0.1218156923 -1.171134e-01       0       0
#> 51  -0.1859857655 -8.512338e-02       0       0
#> 52  -0.0573785123 -2.828685e-02       0       0
#> 53  -0.2122265684  1.080527e-01       0       0
#> 54  -0.0739667105  4.646610e-02       0       0
#> 55  -0.1554983998 -2.836382e-02       0       0
#> 56  -0.1737533622  2.073277e-01       0       0
#> 57  -0.2203384007  1.822113e-01       0       0
#> 58  -0.1289475926  1.822113e-01       0       0
#> 59  -0.0794589510 -2.838256e-02       0       0
#> 60  -0.3040694326  9.264992e-02       0       0
#> 61  -0.3040694326  9.264992e-02       0       0
#> 62  -0.1436155439  1.842640e-03       0       0
#> 63  -0.0426903872 -8.388697e-02       0       0
#> 64  -0.0426903872 -3.276283e-02       0       0
#> 65  -0.1624727286  1.487077e-01       0       0
#> 66  -0.2550526617  1.355878e-01       0       0
#> 67  -0.0578005628  1.011494e-01       0       0
#> 68  -0.0601564910  1.575240e-01       0       0
#> 69  -0.1271127714  1.066575e-01       0       0
#> 70  -0.2038700963  2.166668e-01       0       0
#> 71  -0.1094008156  4.139219e-02       0       0
#> 72  -0.1336273735 -5.938690e-03       0       0
#> 73  -0.0351587896  6.510674e-02       0       0
#> 74  -0.0446517307 -2.886281e-02       0       0
#> 75  -0.1058622872 -8.111862e-02       0       0
#> 76  -0.2624211209  1.021064e-01       0       0
#> 77   0.0697524698 -8.558628e-02       0       0
#> 78   0.0095825018 -4.092919e-03       0       0
#> 79  -0.1037866241  1.698252e-05       0       0
#> 80  -0.1004550084 -6.670059e-02       0       0
#> 81  -0.1313160117  1.470053e-01       0       0
#> 82  -0.1766371289  6.673063e-02       0       0
#> 83   0.0610220634 -5.767447e-02       0       0
#> 84  -0.0189298344  1.666612e-01       0       0
#> 85  -0.0636679082  8.112038e-02       0       0
#> 86  -0.0390504700 -6.583650e-02       0       0
#> 87  -0.1588867639  1.675774e-01       0       0
#> 88  -0.1204752274  1.961381e-01       0       0
#> 89  -0.1633470059 -1.021132e-01       0       0
#> 90  -0.1345625189 -5.597937e-02       0       0
#> 91  -0.2646595249  2.916910e-01       0       0
#> 92  -0.0822247728  1.525062e-01       0       0
#> 93  -0.1677425440  9.005230e-02       0       0
#> 94  -0.0433808844 -5.557628e-02       0       0
#> 95  -0.0579058658  1.402773e-01       0       0
#> 96  -0.2040259732  1.890903e-01       0       0
#> 97  -0.3939027827  4.052407e-01       0       0
#> 98  -0.2163840729  1.840259e-01       0       0
#> 99  -0.0041973764  1.244836e-02       0       0
#> 100 -0.1206505988  3.441737e-01       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21    ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>     <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12  0.00282  -0.0913 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.249     0.0123 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.249     0.0123 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0926    0.00661
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.178     0.212  
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0207    0.151  
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.000759 -0.0218 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0625   -0.0503 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0625   -0.0196 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.137     0.0923 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.250  17.8 0.041  0.12 90.7  1612. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.250  17.8 0.041  0.12  4.77   84.7
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.194  19.7 0.041  0.12 85.3  1680. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.194  19.7 0.041  0.12  7.70  152. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.194  19.7 0.041  0.12 85.3  1680. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.194  19.7 0.041  0.12  7.70  152. 
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>