Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

Usage

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- function() {
  ini({
    THETA_Cl <- 4.0
    THETA_Vc <- 70.0
    THETA_Ka <- 1.0
    ETA_Cl ~ 0.2
    ETA_Vc ~ 0.2
    ETA_Ka ~ 0.2
    prop.sd <- sqrt(0.05)
  })
  model({
    TVCl <- THETA_Cl
    TVVc <- THETA_Vc
    TVKa <- THETA_Ka

    Cl <- TVCl*exp(ETA_Cl)
    Vc <- TVVc*exp(ETA_Vc)
    Ka <- TVKa*exp(ETA_Ka)

    K20 <- Cl/Vc
    Cc <- centr/Vc

    d/dt(depot) = -Ka*depot
    d/dt(centr) = Ka*depot - K20*centr
    Cc ~ prop(prop.sd)
  })
}
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#>  
#>  
#>  
#>  
#> $eta
#>         ETA_Cl      ETA_Vc        ETA_Ka
#> 1    0.1844208 -0.43805309  0.5589599433
#> 2    0.1702946 -0.65343916  0.1772172345
#> 3    0.1547268 -0.71269139  0.0521570306
#> 4    0.3819019 -0.59969530  0.1515927910
#> 5    0.6566451 -0.38319787 -0.0714505765
#> 6    0.3781985 -0.69843562 -0.1871125603
#> 7    0.4369133 -0.73026322 -0.0592865960
#> 8   -0.1574308 -0.04286925  0.2664079840
#> 9    0.2520540 -0.76552810  0.7851955315
#> 10   0.3463545 -0.87047463 -0.1839729890
#> 11   0.6784458 -0.35395065 -0.0047872087
#> 12   0.6072037 -0.69603594 -0.1395369306
#> 13   0.7438610 -0.43034035 -0.0700554309
#> 14   0.7438610 -0.17441492  0.1601942250
#> 15   0.5072936 -0.57262383 -0.0736425268
#> 16   0.6238319 -0.51164915  0.0493567341
#> 17   0.5018634  0.01850689  0.7438441110
#> 18   0.8231393 -0.23037458  0.4495101423
#> 19   0.7339266 -0.23037458  0.4495101423
#> 20   0.6192160 -0.41009690 -0.1031305035
#> 21   0.3083109 -0.97581677 -0.4396286302
#> 22   0.4254984 -0.99078584 -0.3567426205
#> 23   0.5828894 -0.67236291 -0.2972581086
#> 24   0.5828226 -0.65971302 -0.0800784307
#> 25   0.6305524 -0.34233136  0.0213730534
#> 26   0.5052878 -0.48230648  0.3752040101
#> 27   0.4204808 -0.62579618  0.1195915534
#> 28   0.6403227 -0.17561995  0.5538499595
#> 29   0.5901522 -0.32024753  0.3430964679
#> 30   0.4009766 -0.19878880  0.3441395366
#> 31   0.5584779 -0.37425079  0.2437752596
#> 32   0.7375154 -0.57724261 -0.2563781613
#> 33   0.2769516 -0.91104803 -0.4564882387
#> 34   0.3578926 -1.09047058 -0.5538847410
#> 35   0.5198541 -0.54839228  0.2429224252
#> 36   0.6113854 -0.34950276  0.5131425888
#> 37   0.5604968 -0.09046750  0.5611541593
#> 38   0.1814259 -0.24608783  0.4342569698
#> 39   0.4102265 -0.08196872  0.8200357479
#> 40   0.6392396 -0.29629692  0.5648894331
#> 41   0.4753098 -0.74649139 -0.2040274604
#> 42   0.6456799 -0.56160960 -0.4158774867
#> 43   0.4983860 -0.58955742 -0.3342173927
#> 44   0.6255700 -0.46159965 -0.2173076636
#> 45   0.4563637 -0.59827848 -0.0304088043
#> 46   0.2812374 -0.52037136  0.1561507754
#> 47   0.4992919 -0.66218053  0.3170746691
#> 48   0.1671170 -0.81543614  0.0007025171
#> 49   0.2701360 -0.70903040 -0.4097055185
#> 50   0.4467154 -0.30572110  0.8988434537
#> 51   0.0000000  0.00000000  0.0000000000
#> 52   0.4432430 -0.19959941 -0.0182360984
#> 53   0.6507309 -0.23441443  0.5517944803
#> 54   0.7883359 -0.26204071  0.0441133809
#> 55   0.7841662 -0.42529071  0.3095634469
#> 56   0.3094279 -0.28857171  0.1862077489
#> 57   0.7510996 -0.35096414  0.2429222616
#> 58   0.4643780 -0.14385902  0.2549082856
#> 59   0.8734574 -0.03900315  0.5617039297
#> 60   0.3957196 -0.70907875 -0.1982888720
#> 61   0.5726123 -0.41609870  0.5798524130
#> 62   0.6503596 -0.24594603  0.2186818875
#> 63   0.5975696 -0.19069546  0.2061462199
#> 64   0.5891815 -0.52517129  0.2635254393
#> 65   0.4646913 -0.73255335  0.4256222090
#> 66   0.6844460 -0.05074335  1.1366032540
#> 67   0.4581476 -0.22463899  0.4254299350
#> 68   0.7046095  0.19784770  0.7410062393
#> 69   0.5571614 -0.03723236  0.0843635788
#> 70   0.7551196  0.01634314  0.4715239820
#> 71   0.8003644  0.02486649  0.7843992012
#> 72   0.6050803 -0.47232483  0.1225824277
#> 73   0.5764811 -0.52297163  0.2357184626
#> 74   0.4445525 -0.50004446  0.1803458273
#> 75   0.5956160 -0.25531119  0.3832833207
#> 76   0.6455107 -0.57474626  0.0369711416
#> 77   0.7193022 -0.35355033  0.2163061638
#> 78   0.9101741  0.14641048  0.6897415101
#> 79   0.4845060 -0.43303924  0.4692957211
#> 80   0.5835763 -0.26022039  0.0899865641
#> 81   0.4486918 -0.40186572  0.1217597636
#> 82   0.5834377 -0.71840370 -0.2779697962
#> 83   0.4502321 -0.14998718  0.0415319870
#> 84   0.3982273 -0.42345381  0.2799825343
#> 85   0.2774437 -0.63031671  0.2306799747
#> 86   0.5425262 -0.40931725 -0.0575336196
#> 87   0.7647374 -0.11205551  0.5384090961
#> 88   0.4876078 -0.78084003 -0.1802243063
#> 89   0.5263299 -0.39573556 -0.0032796324
#> 90  -0.4628368 -0.22318771  0.3111785403
#> 91   0.3519512 -0.64220949  0.2113238574
#> 92   0.3852560 -0.56464553 -0.0445950666
#> 93   0.8507789  0.06801904  0.5116342655
#> 94   0.4247760 -0.68044889  0.3097495127
#> 95   0.9184119 -0.06631377  0.8903765697
#> 96   0.7878075 -0.20266299  1.3348373375
#> 97   0.6881082 -0.21275893  1.0036443668
#> 98   0.6547887 -0.41678319  0.0918646560
#> 99   0.4965296 -0.64374584 -0.3236602162
#> 100  0.4101506 -0.60223557  0.5909147000
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 8
#>    sim.id THETA_Cl THETA_Vc THETA_Ka prop.sd ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1   0.224  0.184 -0.438   0.559 
#>  2      2        4       70        1   0.224  0.170 -0.653   0.177 
#>  3      3        4       70        1   0.224  0.155 -0.713   0.0522
#>  4      4        4       70        1   0.224  0.382 -0.600   0.152 
#>  5      5        4       70        1   0.224  0.657 -0.383  -0.0715
#>  6      6        4       70        1   0.224  0.378 -0.698  -0.187 
#>  7      7        4       70        1   0.224  0.437 -0.730  -0.0593
#>  8      8        4       70        1   0.224 -0.157 -0.0429  0.266 
#>  9      9        4       70        1   0.224  0.252 -0.766   0.785 
#> 10     10        4       70        1   0.224  0.346 -0.870  -0.184 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 14
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20  rxCc    Cc   depot
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>
#> 1      1     1     4    70     1  4.81  45.2  1.75 0.106  30.4  30.4 5.56e+2
#> 2      1    14     4    70     1  4.81  45.2  1.75 0.106  10.9  10.9 7.51e-8
#> 3      2     1     4    70     1  4.74  36.4  1.19 0.130  30.4  30.4 8.29e+2
#> 4      2    14     4    70     1  4.74  36.4  1.19 0.130  10.3  10.3 1.51e-4
#> 5      3     1     4    70     1  4.67  34.3  1.05 0.136  29.7  29.7 9.18e+2
#> 6      3    14     4    70     1  4.67  34.3  1.05 0.136  10.3  10.3 1.03e-3
#> # ℹ 194 more rows
#> # ℹ 2 more variables: centr <dbl>, AUC <dbl>
#>