Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object

A posologyr list, created by the posologyr function.

return_model

A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01,n_iter=50,n_chains=2)
#> $eta
#>          ETA_Cl      ETA_Vc       ETA_Ka
#> 1    0.37471900 -0.14875290  0.436805005
#> 2    0.67510280  0.14650223  0.698530331
#> 3    0.91193351  0.01160183  0.575539320
#> 4    0.49221634 -0.35249339  0.588392527
#> 5    0.58743536 -0.42993827  0.309147366
#> 6    0.47628389 -0.31504362  0.281201413
#> 7    0.33717301 -0.51960018  0.039319134
#> 8    0.71539823 -0.05141592  0.673564858
#> 9    0.48774094 -0.68050571  0.408583224
#> 10   0.62517132 -0.32165262  0.351470633
#> 11   0.34471349 -0.31650119  0.765549483
#> 12   0.71671800 -0.04360327  0.581410734
#> 13   0.37629737 -0.19456192  0.131284738
#> 14   0.40340067 -0.03720956  0.840004378
#> 15   0.41694682 -0.21240821  0.361942726
#> 16   0.59826217 -0.45546047  0.045956188
#> 17   0.63555062 -0.35949182  0.079740637
#> 18   0.68743702 -0.46143790  0.248936325
#> 19   0.64950265 -0.22498780  0.462450088
#> 20   0.76712441 -0.07488553  0.478805927
#> 21   0.68485820 -0.37216352  0.117018290
#> 22   0.61152116 -0.36323724 -0.072790076
#> 23   0.59206159 -0.48250655  0.135347521
#> 24   0.61386998 -0.62328488  0.200404251
#> 25   0.72534973 -0.32537382  0.235989131
#> 26   0.63343337 -0.60745250  0.235989131
#> 27   0.49642635  0.03559871  0.370318166
#> 28   0.64188053 -0.15406271  0.266221336
#> 29   0.71258118 -0.21421865  0.903773900
#> 30   0.39196836 -0.13261376  0.549351069
#> 31   0.70882639 -0.36090791 -0.017587301
#> 32   0.50354297 -0.36090791 -0.085596988
#> 33   0.54467838 -0.35195412  0.130096842
#> 34   0.36055931 -0.66964962  0.038646849
#> 35   0.24677630 -0.62089956  0.255135870
#> 36   0.58352887 -0.56981007  0.041058780
#> 37   0.56612975 -0.87877936 -0.498024942
#> 38   0.53163085 -0.50025837 -0.113864097
#> 39   0.45308944 -0.49287027  0.090162311
#> 40   0.76020783 -0.16372184  0.270979664
#> 41   0.87501644  0.20139869  0.933329758
#> 42   0.70646857  0.04383275  0.740642152
#> 43   0.43553750 -0.08383218  0.266928467
#> 44   0.55734556 -0.23663495  0.233625957
#> 45   0.57094731 -0.33443314  0.681239741
#> 46   0.66576712 -0.17585919  0.427174045
#> 47   0.38599508 -0.27607443  0.633485222
#> 48   0.44490247 -0.30400987  1.127464190
#> 49   0.48492641 -0.01614474  1.308331858
#> 50   0.48602197 -0.10356493  1.168619626
#> 51   0.00000000  0.00000000  0.000000000
#> 52   0.49909660 -0.70992895 -0.351661430
#> 53   0.27512167 -0.76327578 -0.231790075
#> 54   0.57253873 -0.34494230  0.194216600
#> 55   0.21643466 -0.84449311  0.622770534
#> 56   0.23598972 -0.31502599  0.428991376
#> 57   0.26083324 -1.40081183 -0.781487426
#> 58   0.29119114 -0.75971295 -0.351190932
#> 59   0.57157074 -0.34563549  0.165134393
#> 60   0.42247138 -0.76987677 -0.358044853
#> 61   0.22150258 -1.51263174 -0.727661676
#> 62   0.33044213 -1.20095971  0.885917541
#> 63   0.20322001 -0.58593510  0.133059340
#> 64   0.32535962 -0.32535114  0.197624161
#> 65   0.76697966  0.04404603  0.853248676
#> 66   0.57428715 -0.53668907  0.622896278
#> 67   0.35251999 -0.67897293 -0.106268201
#> 68   0.77128474 -0.39357104  0.356568946
#> 69   0.49695649 -0.55844669  0.537349779
#> 70   0.37635355 -0.83566002 -0.226351535
#> 71  -0.19923572  0.29902312  0.657580239
#> 72   0.55383125 -0.72302552 -0.175613982
#> 73   0.49356910 -0.63514750  0.093656808
#> 74   0.51358279 -0.10490507  0.527064260
#> 75   0.53126903 -0.72264720 -0.110887341
#> 76   0.36093648 -0.62484987  0.592960985
#> 77   0.65207867 -0.56774676  0.216836691
#> 78   0.39456816 -0.48811801 -0.012403511
#> 79   0.54442192 -0.53126718  0.199867623
#> 80   0.58359434 -0.29894726  0.123741062
#> 81   0.44440087 -0.69747473 -0.140745870
#> 82   0.37224085 -1.00967706 -0.376320322
#> 83  -0.03499881 -0.41234399 -0.016745398
#> 84   0.49210261 -0.61433797 -0.159877888
#> 85   0.36175514 -0.73389346  0.604837267
#> 86   0.54317113 -0.80215748 -0.389976106
#> 87   0.15409428 -0.51258795  0.290328993
#> 88   0.74567117 -0.22346027  0.177935332
#> 89   0.60046702 -0.58516766  0.112057384
#> 90   0.74864971 -0.19994574  0.396530689
#> 91   0.25382207 -0.07980309  0.223101653
#> 92   0.59799244 -0.53299831  0.005411019
#> 93   0.42113377 -0.66819026 -0.383290231
#> 94   0.56098020 -0.42742048  0.625012364
#> 95   0.56909134 -0.66762760  0.548720788
#> 96   0.50262612 -0.14047916  0.222845596
#> 97   0.48108015 -0.46409367  0.439845278
#> 98   0.65801570 -0.45973322  0.430144021
#> 99   0.52414038 -0.22704576  0.041316183
#> 100  0.72610536  0.05562278  0.815060598
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1  0.375 -0.149  0.437 
#>  2      2        4       70        1  0.675  0.147  0.699 
#>  3      3        4       70        1  0.912  0.0116 0.576 
#>  4      4        4       70        1  0.492 -0.352  0.588 
#>  5      5        4       70        1  0.587 -0.430  0.309 
#>  6      6        4       70        1  0.476 -0.315  0.281 
#>  7      7        4       70        1  0.337 -0.520  0.0393
#>  8      8        4       70        1  0.715 -0.0514 0.674 
#>  9      9        4       70        1  0.488 -0.681  0.409 
#> 10     10        4       70        1  0.625 -0.322  0.351 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  5.82  60.3  1.55 0.0965 21.5  6.42e+2 1299.
#> 2      1    14     4    70     1  5.82  60.3  1.55 0.0965  9.39 1.17e-6  566.
#> 3      2     1     4    70     1  7.86  81.0  2.01 0.0969 18.1  4.62e+2 1469.
#> 4      2    14     4    70     1  7.86  81.0  2.01 0.0969  6.84 1.95e-9  554.
#> 5      3     1     4    70     1  9.96  70.8  1.78 0.141  19.3  5.45e+2 1363.
#> 6      3    14     4    70     1  9.96  70.8  1.78 0.141   4.44 5.03e-8  314.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>