Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.78365662 -0.186118930  0.516686138
#> 2    0.75811108 -0.041364876  0.805579171
#> 3    0.69563422 -0.188096139  0.347594020
#> 4    0.52454892 -0.125009452  0.358098539
#> 5    0.43013695 -0.583340189  0.385141438
#> 6    0.39911422 -0.631693594 -0.058220651
#> 7    0.65593098 -0.362137522  0.576327730
#> 8    0.43982605 -0.264987921  0.650979616
#> 9    0.50223182 -0.003663647  1.214374643
#> 10   0.54321077 -0.053454923  0.453921069
#> 11   0.36228886 -0.347663713  0.339231288
#> 12   0.71276062 -0.081433880  0.330007065
#> 13   0.84663089 -0.329592922  0.679286369
#> 14   0.56208161 -0.611973436 -0.055532465
#> 15   0.39976817 -0.786165844 -0.610759736
#> 16   0.54054223  0.036828234  0.721525519
#> 17   0.80042264 -0.143944283  0.521809481
#> 18   0.71831526 -0.284204223  0.380236085
#> 19   0.28489415 -0.428781678  0.272274403
#> 20   0.23147842 -0.530851416  0.131600495
#> 21   0.46497054 -0.674667837  0.040905095
#> 22   0.42081085 -0.570211931 -0.169483812
#> 23   0.07454651 -0.829627186 -0.498726044
#> 24   0.42454053 -0.745699918  0.176755138
#> 25   0.54738132 -0.376413809  0.482671761
#> 26   0.71740608 -0.243993313 -0.032485226
#> 27   0.71835591 -0.289010826  0.001122633
#> 28   0.62924943 -0.025070433  0.833694422
#> 29   0.67326526 -0.264997192  0.556012534
#> 30   0.85015578 -0.094534785  0.919082937
#> 31   0.59463842 -0.396135192 -0.181850798
#> 32   0.63223054 -0.379255981  0.530370329
#> 33   0.28609410 -0.521198286  0.143251122
#> 34   0.14077570 -0.302585518  0.147500971
#> 35   0.53904335 -0.195522384  0.349637002
#> 36   0.84499541 -0.051492823  0.722499412
#> 37   0.71659998 -0.110553186  1.016070227
#> 38   0.67611024 -0.235377902  0.183506045
#> 39   0.66745084 -0.137535706  0.267879264
#> 40   0.78821954 -0.001436581  0.098999366
#> 41   0.63845452 -0.213859807  0.598767553
#> 42   0.26803660 -0.452500120  0.061171704
#> 43   0.55372007 -0.543018188  0.032888008
#> 44   0.55372007 -0.684357313 -0.145682732
#> 45   0.57128133 -0.374015796  0.397056888
#> 46   0.24034944 -0.800361175 -0.112777227
#> 47   0.25351547 -1.022823579 -0.032474717
#> 48   0.43877545 -0.518247968 -0.313514369
#> 49   0.41526239  0.113026348  0.976513714
#> 50   0.72416126 -0.435067213 -0.229173386
#> 51   0.00000000  0.000000000  0.000000000
#> 52   0.38682555 -0.581436584  0.692977446
#> 53   0.75374143 -0.246669684  0.222273217
#> 54   0.48561792 -0.491039798  0.065104968
#> 55   0.61561907 -0.505586703  0.247848780
#> 56   0.62105304 -0.219417159  1.056123987
#> 57   0.39464399 -0.875741248 -0.069444515
#> 58   0.59521177 -0.455015533  0.014270930
#> 59   0.59490800 -0.489988746  0.052238934
#> 60   0.35442982 -0.014347182  0.525799711
#> 61   0.69769240 -0.357413051  0.428962285
#> 62   0.64329007 -0.622428195  0.529449153
#> 63   0.25135639 -1.015968957  0.096584698
#> 64   0.22310827 -0.568261157 -0.311379752
#> 65   0.56534669 -0.390476371  0.261591575
#> 66   0.42248686 -0.236268012  0.609057056
#> 67   0.71693695  0.057310706  0.963218934
#> 68  -0.03983019 -0.864204714 -0.286742004
#> 69   0.40605738 -0.496555452  0.308390617
#> 70   0.97411504  0.060590335  0.937151230
#> 71   0.75059867 -0.223647762  0.303906272
#> 72   0.44412941 -0.660676723  0.033986931
#> 73   0.54201327 -0.414829248 -0.144789975
#> 74   0.24527137 -0.505220965  0.549269586
#> 75   0.39085863 -0.376632389  0.772924633
#> 76   0.47428917 -0.370670913 -0.059741912
#> 77   0.64136745 -0.223143950  0.097530278
#> 78   0.48951935 -0.809765275 -0.421642962
#> 79   0.50193963 -0.355520421  0.145459774
#> 80   0.39908212 -0.243646455 -0.213024059
#> 81   0.32025859 -0.670126173 -0.134155760
#> 82   0.58815468 -0.166398660  0.290742220
#> 83   0.70122675 -0.270063195  0.363966755
#> 84   0.55778111 -0.525971379  0.091416198
#> 85   0.73398579 -0.640853200  0.033450618
#> 86   0.55397138 -0.210517383  1.102932383
#> 87   0.38297010 -0.304011654  0.170230193
#> 88   0.83285298 -0.191268756  0.706018839
#> 89   0.55383932 -0.433057110  0.157362844
#> 90   0.35097754 -0.365609240  0.753834791
#> 91   0.48618249 -0.432583940  0.289646416
#> 92   0.79428087 -0.215122900  0.388115322
#> 93   0.42597052 -0.927107140 -0.324166931
#> 94   0.59199166 -0.294525292  0.107078967
#> 95   0.72039526 -0.114893837  0.480393266
#> 96   0.19579723 -0.929093706 -0.458867046
#> 97   0.58339373 -0.576135830 -0.105848207
#> 98   0.62286842 -0.164355919  0.079027869
#> 99   0.55917314 -0.181140732  0.372152680
#> 100  0.58942037 -0.727616474  0.203640917
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl   ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>    <dbl>   <dbl>
#>  1      1        4       70        1  0.784 -0.186    0.517 
#>  2      2        4       70        1  0.758 -0.0414   0.806 
#>  3      3        4       70        1  0.696 -0.188    0.348 
#>  4      4        4       70        1  0.525 -0.125    0.358 
#>  5      5        4       70        1  0.430 -0.583    0.385 
#>  6      6        4       70        1  0.399 -0.632   -0.0582
#>  7      7        4       70        1  0.656 -0.362    0.576 
#>  8      8        4       70        1  0.440 -0.265    0.651 
#>  9      9        4       70        1  0.502 -0.00366  1.21  
#> 10     10        4       70        1  0.543 -0.0535   0.454 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc     depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>     <dbl> <dbl>
#> 1      1     1     4    70     1  8.76  58.1  1.68 0.151 22.7   5.86e+ 2 1320.
#> 2      1    14     4    70     1  8.76  58.1  1.68 0.151  4.76  2.01e- 7  277.
#> 3      2     1     4    70     1  8.54  67.2  2.24 0.127 22.5   3.93e+ 2 1511.
#> 4      2    14     4    70     1  8.54  67.2  2.24 0.127  5.50 -3.82e-10  369.
#> 5      3     1     4    70     1  8.02  58.0  1.42 0.138 21.0   7.06e+ 2 1216.
#> 6      3    14     4    70     1  8.02  58.0  1.42 0.138  5.71  7.18e- 6  331.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>