Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>           ETA_Cl      ETA_Vc       ETA_Ka
#> 1    0.644403296 -0.44063648  0.373475433
#> 2    0.662398028 -0.43445101  0.040259679
#> 3    0.648919712 -0.57792586 -0.011018296
#> 4    0.594538738 -0.56906367  0.226230407
#> 5    0.533629869 -0.56906367 -0.084346015
#> 6    0.389083916 -0.75730560 -0.152346128
#> 7    0.434933747 -0.45674204 -0.048582878
#> 8    0.469198135 -0.41460754  0.293865970
#> 9    0.766353108 -0.32917521 -0.102100434
#> 10   0.624760181 -0.75211861 -0.278799818
#> 11   0.402898625 -0.75211861 -0.452129877
#> 12   0.358829408 -0.82374294  0.268653274
#> 13   0.376747084 -0.84098150 -0.170769231
#> 14   0.574596653 -0.58855827 -0.170769231
#> 15   0.642168525 -0.10469632  0.369684592
#> 16   0.333933510 -0.26916804  0.155481498
#> 17   0.675248060 -0.53530196  0.646996254
#> 18   0.691649654 -0.36706351  0.269392579
#> 19   0.317768904 -0.36968500 -0.253973471
#> 20   0.557692277 -0.81251128  0.063928229
#> 21   0.395240605 -0.92669431 -0.593400988
#> 22   0.463475117 -0.60888312 -0.086399519
#> 23   0.456250149 -0.63181228 -0.008460352
#> 24   0.577808514 -0.50073497  0.509811872
#> 25   0.419482878 -0.38751034  0.194042524
#> 26   0.694934779 -0.37136034  0.083101050
#> 27   0.297506881 -0.43332838  0.220050925
#> 28   0.074411626 -0.06499652  0.627308087
#> 29   0.687094052 -0.36404598  0.554461003
#> 30   0.621632380 -0.41755698  1.098191419
#> 31   0.878740209  0.30683520  0.940706611
#> 32   0.643979340 -0.28454040  0.693691561
#> 33   0.808568593 -0.23932030 -0.070109513
#> 34   1.029521020 -0.44742083  0.516868882
#> 35   0.641019702 -0.26917146  0.298627649
#> 36   0.580503314 -0.12585224  0.174406247
#> 37   0.550758040 -0.31795770  0.673943380
#> 38   0.199311579 -0.22615355  0.248419751
#> 39   0.718396696 -0.35051081  0.514103068
#> 40   0.718396696 -0.11122251  0.604666613
#> 41   0.666006443 -0.17442191  0.554817644
#> 42   0.250841288 -0.27703722 -0.041297573
#> 43   0.511454848 -0.43042823  0.283856449
#> 44   0.511454848 -0.43042823  0.117965648
#> 45   0.406637503 -0.88260801 -0.297722533
#> 46   0.481625791 -0.99291844 -0.504463098
#> 47   0.343419891 -0.78908690 -0.486476835
#> 48  -0.055344694 -0.11136147  0.919148466
#> 49   0.512249019 -0.10133309  0.806821308
#> 50   0.016559159 -0.54303685 -0.164148074
#> 51   0.000000000  0.00000000  0.000000000
#> 52   0.586456446 -0.50022100  0.217835075
#> 53   0.421088683 -0.35160929  0.224185896
#> 54   0.516894150 -0.14729129  0.490218039
#> 55  -0.007281453 -0.23261736  0.330673044
#> 56   0.428056004 -0.05661282  0.177141546
#> 57   0.140295691 -1.01943398 -0.040704593
#> 58   0.528432060 -0.34905514  0.235113885
#> 59   0.369135203 -0.33348361  0.107499562
#> 60   0.641196674 -0.30093361  0.683749346
#> 61   0.554825957 -0.49603813  0.176318074
#> 62   0.336965260 -0.26408877  0.158810860
#> 63   0.157332263 -0.35630028 -0.124685522
#> 64   0.164187133 -0.48769363 -0.008611331
#> 65   0.496007500 -0.42060302 -0.091257423
#> 66   0.787628226 -0.40617545  0.170304079
#> 67   0.648489245 -0.74478311 -0.241691834
#> 68   0.663969763 -0.01057230  0.667678657
#> 69   0.305650920 -0.03131123  0.829223834
#> 70   0.712563872 -0.47176739  0.742439009
#> 71   0.614993969 -0.40962734  0.071172414
#> 72   0.390223102 -0.48544405 -0.046449094
#> 73   0.452095769 -0.53125472  0.094291034
#> 74   0.489314354 -0.40997066  0.320927380
#> 75   0.563830880 -0.56174512 -0.274679008
#> 76   0.625119785 -0.10564391  0.612476741
#> 77   0.388705704 -0.39451289  0.372715350
#> 78   0.546832259 -0.64510517 -0.145415492
#> 79   0.563875028 -0.70893457 -0.451584949
#> 80   0.680002670 -0.41764447  0.781183983
#> 81   0.535723554 -0.45787131  0.125489802
#> 82   0.503044444 -0.34133459 -0.086515016
#> 83   0.597385085  0.14817678  0.686893569
#> 84   0.846419345 -0.22283546  0.594544296
#> 85   0.387795561 -0.42201072 -0.241920441
#> 86   0.768642137  0.16066538  0.829214979
#> 87   0.411073159 -0.14263383  0.208045462
#> 88   0.728055879 -0.36779910  0.189796018
#> 89   0.353629538 -0.87513086 -0.030465443
#> 90  -0.016826320 -0.35981051  0.236534275
#> 91   0.423927625 -0.69045011 -0.176552648
#> 92   0.765783878  0.06454790  0.724637151
#> 93   0.468127548 -0.51151795 -0.013656261
#> 94   0.592519547 -0.65147044  0.302331730
#> 95   0.291764886 -0.66381676  0.242247219
#> 96   0.510943903 -0.54124020  0.006863447
#> 97   0.423100508 -0.82056667 -0.139749478
#> 98   0.455905210 -0.46738856 -0.065281084
#> 99   0.272332767 -0.73163121 -0.249687367
#> 100  0.352242369 -0.56400632  0.264659641
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1  0.644 -0.441  0.373 
#>  2      2        4       70        1  0.662 -0.434  0.0403
#>  3      3        4       70        1  0.649 -0.578 -0.0110
#>  4      4        4       70        1  0.595 -0.569  0.226 
#>  5      5        4       70        1  0.534 -0.569 -0.0843
#>  6      6        4       70        1  0.389 -0.757 -0.152 
#>  7      7        4       70        1  0.435 -0.457 -0.0486
#>  8      8        4       70        1  0.469 -0.415  0.294 
#>  9      9        4       70        1  0.766 -0.329 -0.102 
#> 10     10        4       70        1  0.625 -0.752 -0.279 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  7.62  45.1 1.45  0.169 27.0      6.88e+2 1216.
#> 2      1    14     4    70     1  7.62  45.1 1.45  0.169  4.91     4.32e-6  221.
#> 3      2     1     4    70     1  7.76  45.3 1.04  0.171 22.0      9.26e+2  997.
#> 4      2    14     4    70     1  7.76  45.3 1.04  0.171  5.02     1.23e-3  228.
#> 5      3     1     4    70     1  7.65  39.3 0.989 0.195 24.3      9.62e+2  955.
#> 6      3    14     4    70     1  7.65  39.3 0.989 0.195  4.35     2.51e-3  171.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>