Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object

A posologyr list, created by the posologyr function.

return_model

A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01,n_iter=100)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.54456882 -0.236744562  0.457154644
#> 2    0.33802280 -0.401540413  0.137784321
#> 3    0.49971679 -0.262334629  0.293253295
#> 4    0.31145267 -0.322551956  0.254203231
#> 5    0.36354372 -1.101829913 -0.302527411
#> 6    0.86288987 -0.190847212  0.285774395
#> 7    0.81267562 -0.085691236  0.814555395
#> 8    0.29020795 -0.471975364  0.099967543
#> 9    0.52171560 -0.444367078 -0.051587461
#> 10   0.21634699 -1.006927871 -0.388364220
#> 11   0.29258357 -0.914036262  0.100619344
#> 12   0.65080306 -0.461918648  0.100619344
#> 13   0.27769101 -0.208047707  0.375552764
#> 14   0.39848800  0.049550248  0.979227098
#> 15   0.52694710 -0.365307705  0.774438959
#> 16   0.51312859 -0.171314983  1.113944215
#> 17   0.32520744 -0.524979036  0.010094882
#> 18   0.42873195 -0.796848433 -0.309881494
#> 19   0.52311503 -0.635900842  0.226404366
#> 20   0.81704852 -0.205578962  0.280779166
#> 21   0.70813041 -0.323348253  0.076326702
#> 22   0.43107635 -0.180494055  0.239477590
#> 23   0.53176527 -0.745004396 -0.187607774
#> 24   0.27055086 -0.744833195  0.051009083
#> 25   0.47480003 -0.742389782 -0.430141403
#> 26   0.49238638 -0.796200751 -0.421371714
#> 27   0.31703075 -0.473988274  0.647297363
#> 28   0.20391992 -0.714193356 -0.172833502
#> 29   0.29472892 -0.762877004 -0.351167719
#> 30   0.02367438 -0.691650424 -0.460412887
#> 31   0.39652116 -0.952722739 -0.523226062
#> 32   0.45979783 -0.925131042 -0.276841819
#> 33   0.12126738 -0.920732691 -0.569103573
#> 34   0.24824033 -0.743929730 -0.100159765
#> 35   0.38655690 -0.563147229  0.167762369
#> 36   0.41408442 -0.018464713  0.463914345
#> 37   0.50387284  0.016030637  0.635254419
#> 38   0.64753935 -0.474192563  0.052599533
#> 39   0.67026524 -0.437520077  0.191439630
#> 40   0.57294710 -0.271869380  0.191439630
#> 41   0.65543445 -0.369130689  0.162610734
#> 42   0.50615748 -0.208285819  0.162610734
#> 43   0.57151872 -0.627404776 -0.012100672
#> 44   0.84109768 -0.668549423 -0.242389385
#> 45   0.47740249 -0.681254114 -0.151378353
#> 46   0.33430558 -0.356802270  0.537189741
#> 47   0.60533622 -0.038810510  0.688866410
#> 48   0.61938588 -0.553127324  0.545327650
#> 49   0.73152049 -0.060848226  0.356804118
#> 50   0.54195993 -0.112587476  0.253992989
#> 51   0.67267113 -0.219684920  0.551150625
#> 52   0.79456813 -0.212129966  0.699069845
#> 53   0.54662052 -0.090783094  0.635974699
#> 54   0.14136285  0.008859487  0.722307755
#> 55   0.55325623 -0.625942940 -0.282383890
#> 56   0.66886826 -0.396873873  0.311606307
#> 57   0.69529018 -0.058984728  0.372196268
#> 58   0.57847298 -0.828111099 -0.482021439
#> 59   0.44461620 -0.695279966 -0.323840106
#> 60   0.56210044 -0.319257797  0.243878797
#> 61   0.51136551 -0.402656607  0.245238013
#> 62   0.67954926 -0.338382796 -0.036154152
#> 63   0.71609704 -0.249556585  0.492009723
#> 64   0.40817311 -0.349027378 -0.001126585
#> 65   0.52666077  0.122993501  0.502259387
#> 66   0.49814324 -0.436185069 -0.199881645
#> 67  -0.28689052 -0.737592908 -0.073176725
#> 68   0.62787331 -0.323390199  0.181683758
#> 69   0.81751069 -0.267573064  0.696591676
#> 70   0.31381696 -0.584812216  0.437805926
#> 71   0.31502958 -0.951906823  0.466021433
#> 72   0.33441671 -0.804171862 -0.510768246
#> 73   0.48117197 -0.569616376 -0.033082140
#> 74   0.20958046 -0.045111067  0.199771081
#> 75   0.48897380 -0.885811704 -0.219878711
#> 76   0.57106634 -0.527238934  0.231452925
#> 77   0.70477132 -0.327633656  0.231452925
#> 78   0.46588637 -0.694602707  0.237542094
#> 79   0.46003874 -0.759977938 -0.164555271
#> 80   0.60058095 -0.575364632  0.322298800
#> 81   0.40940008 -0.580097997  0.185682702
#> 82   0.42807584 -0.881019026 -0.424732022
#> 83   0.28633066 -0.722113266 -0.228152625
#> 84   0.63709793 -0.353391262  0.150048354
#> 85   0.35208532 -0.257446105  0.174777936
#> 86   0.77360942  0.003460442  0.447733314
#> 87   0.37507490 -0.207555830  0.049008180
#> 88   0.21466309 -1.046893612 -0.002584371
#> 89  -0.17174778 -0.663055539  0.712557985
#> 90   0.04533325 -0.113654747  0.560180691
#> 91   0.25972980 -0.594621795  0.470179381
#> 92   0.50699026 -0.708085902  0.324075270
#> 93   0.24229011 -0.192289209  0.537557063
#> 94   0.31697522 -0.014540881  0.402688967
#> 95  -0.60707104 -0.208230165  0.756718629
#> 96   0.49097809 -1.190818002  0.023155139
#> 97   0.49097809 -0.538616487  0.862304110
#> 98   0.54997209 -0.255747953  0.637399065
#> 99   0.67605895 -0.308144894  0.050945957
#> 100  0.56362387 -0.301438960  0.093027651
#> 101  0.79456813 -0.212129966  0.699069845
#> 102  0.14136285  0.008859487  0.722307755
#> 103  0.66886826 -0.396873873  0.311606307
#> 104  0.57847298 -0.828111099 -0.482021439
#> 105  0.56210044 -0.319257797  0.243878797
#> 106  0.67954926 -0.338382796 -0.036154152
#> 107  0.40817311 -0.349027378 -0.001126585
#> 108  0.49814324 -0.436185069 -0.199881645
#> 109  0.62787331 -0.323390199  0.181683758
#> 110  0.31381696 -0.584812216  0.437805926
#> 111  0.33441671 -0.804171862 -0.510768246
#> 112  0.20958046 -0.045111067  0.199771081
#> 113  0.57106634 -0.527238934  0.231452925
#> 114  0.46588637 -0.694602707  0.237542094
#> 115  0.60058095 -0.575364632  0.322298800
#> 116  0.42807584 -0.881019026 -0.424732022
#> 117  0.63709793 -0.353391262  0.150048354
#> 118  0.77360942  0.003460442  0.447733314
#> 119  0.21466309 -1.046893612 -0.002584371
#> 120  0.04533325 -0.113654747  0.560180691
#> 121  0.50699026 -0.708085902  0.324075270
#> 122  0.31697522 -0.014540881  0.402688967
#> 123  0.49097809 -1.190818002  0.023155139
#> 124  0.54997209 -0.255747953  0.637399065
#> 125  0.56362387 -0.301438960  0.093027651
#> 126  0.00000000  0.000000000  0.000000000
#> 127  0.71973321 -0.446005733  0.417433999
#> 128  0.44894609 -0.830053480 -0.391715331
#> 129  0.33327581 -0.478948370  0.013440124
#> 130  0.52932071 -0.195969576  0.448632828
#> 131  0.65042804 -0.387933695  0.234263871
#> 132  0.76418122  0.036129809  1.173055618
#> 133  0.47196528 -0.630803174  0.101305853
#> 134  0.47792197 -0.307607570  0.407150992
#> 135  0.20398202 -0.567261880 -0.182972989
#> 136  0.54248210 -0.905967988  0.336850618
#> 137  0.75865568 -0.085768726  0.369384791
#> 138  0.59929993  0.009878784  0.880519836
#> 139  0.57927357 -0.280998873  0.197206694
#> 140  0.57959576 -1.002001914 -0.305460888
#> 141  0.32435120 -0.957039970 -0.357815247
#> 142  0.42855751 -0.390066833  1.053802111
#> 143  0.68427085 -0.308081572  0.453280513
#> 144  0.55985844 -0.198827188  0.077611219
#> 145  0.40280698 -0.884867830 -0.571439798
#> 146  0.76525277 -0.148786957  0.419927769
#> 147  0.47787775 -0.490190001 -0.131123366
#> 148  0.28365961 -0.615162050  0.243850395
#> 149  0.26061192 -0.014105729  0.275042212
#> 150  0.64631105 -0.227667576  0.170696478
#> 151  0.35348863 -0.551517091 -0.328378846
#> 152 -0.02321881 -0.343291666  0.546652183
#> 153  0.50700160 -0.430169241  0.057300389
#> 154  0.26230442 -0.705803557  0.012554363
#> 155  0.42513013 -0.121320274  0.483656892
#> 156  0.60354964 -0.383775777  0.031529575
#> 157  0.72240020 -0.118662030  0.556411241
#> 158  0.67608946  0.076628484  0.588092508
#> 159  0.57481685 -0.359179801  0.298376116
#> 160  0.32716614 -0.146141698  0.325006243
#> 161  0.64897046 -0.571928553 -0.391279262
#> 162  0.39727425 -0.328998987 -0.054516441
#> 163  0.58186115 -0.597472457  0.373086957
#> 164  0.40202423 -0.613776399  0.032689064
#> 165  0.39654203 -0.948863531 -0.058702498
#> 166  0.54416725 -0.744531791 -0.042599613
#> 167  0.36535755 -0.126009997  0.594409422
#> 168  0.76670310 -0.275970677  0.434561938
#> 169  0.75145386 -0.333842002  0.248984410
#> 170  0.75038469 -0.123782550  0.884078330
#> 171  0.69852560 -0.094934414  0.870930998
#> 172  0.63247266 -0.618360736  0.375418289
#> 173  0.44364344 -0.426219550  0.499177851
#> 174  0.73127849 -0.075425079  0.434010527
#> 175  0.71017760 -0.413251912 -0.035843417
#> 176  0.15131973 -0.483974299  0.628148194
#> 177  0.38767105 -0.036128649  0.415266212
#> 178  0.37159860 -0.735905125 -0.253381970
#> 179  0.48857822 -0.640826479  0.096365269
#> 180  0.54705814 -0.159204317  0.717854198
#> 181  0.43582956 -0.810652201 -0.184110145
#> 182  0.14935089 -1.012780977 -0.726131352
#> 183  0.39172808 -1.100300871 -0.286540089
#> 184  0.53919563 -0.391171612 -0.098356341
#> 185  0.52728728 -0.318826308  0.411184408
#> 186 -0.16422338 -0.212907700  0.620332644
#> 187  0.66170908 -0.200729874  0.177405550
#> 188  0.67066154 -0.392527359  0.008162344
#> 189  0.25535049 -0.443622890 -0.078663934
#> 190  0.67072719 -0.605057073  0.012074582
#> 191  0.18406579 -0.486382635  0.048840452
#> 192  0.46121198 -0.758475052 -0.173477525
#> 193  0.55950969 -0.492943084  0.481603995
#> 194  0.01758102 -1.002251679 -0.302855548
#> 195  0.29248209 -0.710734689 -0.339932169
#> 196  0.55709050 -0.543367570  0.118126249
#> 197  0.27558751 -0.254013957  1.078827548
#> 198  0.56533961 -0.198750785  0.398870305
#> 199  0.27160227 -1.002120978  0.240895967
#> 200  0.48551515 -0.492480140  0.326374870
#> 201  0.85345454 -0.074871437  0.145024771
#> 202  0.44894609 -0.830053480 -0.391715331
#> 203  0.19001086 -0.555328007 -0.457850531
#> 204  0.65042804 -0.387933695  0.234263871
#> 205  0.63936292 -0.042828789  0.792785997
#> 206  0.47792197 -0.307607570  0.407150992
#> 207  0.10711794 -0.965919550 -0.182972989
#> 208  0.75865568 -0.085768726  0.369384791
#> 209  0.60409055  0.066719138  0.989541439
#> 210  0.57959576 -1.002001914 -0.305460888
#> 211  0.73601536 -0.372811692  0.670178300
#> 212  0.68427085 -0.308081572  0.453280513
#> 213  0.56975684 -0.118092126  0.441230989
#> 214  0.76525277 -0.148786957  0.419927769
#> 215  0.25847215 -0.346697290  0.196816955
#> 216  0.26061192 -0.014105729  0.275042212
#> 217  0.65517146 -0.497861730  0.361156376
#> 218 -0.02321881 -0.343291666  0.546652183
#> 219  0.51327841 -0.538878398  0.409253471
#> 220  0.42513013 -0.121320274  0.483656892
#> 221  0.75279647 -0.255961286  0.361098019
#> 222  0.67608946  0.076628484  0.588092508
#> 223  0.47533355 -0.098310015  0.515733976
#> 224  0.64897046 -0.571928553 -0.391279262
#> 225  0.14862146 -0.339755728  0.131313670
#> 226  0.40202423 -0.613776399  0.032689064
#> 227  0.58058587 -0.719399507  0.130649173
#> 228  0.36535755 -0.126009997  0.594409422
#> 229  0.56325066 -0.506972602  0.196053670
#> 230  0.75038469 -0.123782550  0.884078330
#> 231  0.72370480 -0.381240794 -0.095853160
#> 232  0.44364344 -0.426219550  0.499177851
#> 233  0.30539567 -0.200428559  0.472711816
#> 234  0.15131973 -0.483974299  0.628148194
#> 235  0.44938537 -0.653403476 -0.223423766
#> 236  0.48857822 -0.640826479  0.096365269
#> 237  0.73997102 -0.104210709  0.078926116
#> 238  0.14935089 -1.012780977 -0.726131352
#> 239  0.29517164 -0.637390400  0.374179180
#> 240  0.52728728 -0.318826308  0.411184408
#> 241  0.46102204 -0.561981256 -0.184897617
#> 242  0.67066154 -0.392527359  0.008162344
#> 243  0.56094630 -0.561227872 -0.068573248
#> 244  0.18406579 -0.486382635  0.048840452
#> 245  0.69856313 -0.492943084  0.481603995
#> 246  0.01758102 -1.002251679 -0.302855548
#> 247  0.46953921 -0.718093458 -0.339932169
#> 248  0.27558751 -0.254013957  1.078827548
#> 249  0.56533961 -0.505947750  0.215224679
#> 250  0.48551515 -0.492480140  0.326374870
#> 251  0.00000000  0.000000000  0.000000000
#> 252  0.66785024 -0.541063965 -0.013031616
#> 253  0.32573782 -0.198922977  0.406022218
#> 254  0.31038657 -1.009956699 -0.094186151
#> 255  0.39640658 -0.632119027 -0.403518082
#> 256  0.08203623 -0.347566877  0.315680747
#> 257  0.75295946 -0.306865668  0.049959540
#> 258  0.38141101 -0.334266628  0.401609039
#> 259  0.57325013 -0.337082006 -0.033323428
#> 260  0.63971627  0.099550283  0.682872132
#> 261  0.51965916 -0.653578896  0.129458209
#> 262  0.60146225 -0.491562183 -0.361706948
#> 263  0.37871893 -0.607725279  0.571340653
#> 264  0.63260023 -0.703710582 -0.141578924
#> 265  0.75722708 -0.117387618  0.290611083
#> 266  0.69763789 -0.031952138  0.740975681
#> 267  0.55499433 -0.637869959 -0.212443162
#> 268  0.36869533 -0.688416758  0.076680669
#> 269  0.32413515 -0.508582003  0.095467657
#> 270  0.44267156 -0.249578146  0.293117461
#> 271  0.58031980 -0.674323020 -0.095436592
#> 272  0.45609170 -0.316448405  0.607438117
#> 273  0.48512297  0.154382133  0.753977304
#> 274  0.79578808 -0.490360213  0.147466468
#> 275  0.21239267 -0.684835864 -0.314015237
#> 276  0.57352570 -0.682916991  0.210804787
#> 277  0.56059178 -0.597713232  0.813384771
#> 278  0.62885113 -0.167853706  0.438118530
#> 279  0.41787315 -0.161398955  0.389888814
#> 280  0.54200127 -0.122173471  0.066871345
#> 281  0.42661611 -0.529977020 -0.094590928
#> 282  0.67436621 -0.379616076 -0.022379756
#> 283  0.64714261 -0.400997723 -0.042609696
#> 284  0.61033702 -0.221291875  0.300996436
#> 285  0.46342875 -0.674840597  0.195110725
#> 286  0.61755575 -0.460119181 -0.101496816
#> 287  0.44346670 -0.621708083 -0.401501522
#> 288  0.66562425 -0.247406412  0.125637348
#> 289  0.57224843 -0.448275700  0.032163102
#> 290  0.44194804 -0.839442781 -0.160673814
#> 291  0.71401841 -0.203851452  0.134470356
#> 292  0.74031708  0.002806418  0.541459530
#> 293  0.72966533  0.123213052  0.498290888
#> 294  0.56472417 -0.638768530  0.206870883
#> 295  0.79207446 -0.322945323  0.300388458
#> 296  0.66541091 -0.170497605  0.455905257
#> 297  0.74370300 -0.417832563  0.130786263
#> 298  0.38945841 -0.391371985  0.034204097
#> 299  0.48076405 -0.249450775  0.286175073
#> 300  0.46566274 -0.223040843  0.142280685
#> 301 -0.02321881 -0.343291666  0.546652183
#> 302  0.26230442 -0.705803557  0.012554363
#> 303  0.60354964 -0.383775777  0.031529575
#> 304  0.67608946  0.076628484  0.588092508
#> 305  0.32716614 -0.146141698  0.325006243
#> 306  0.39727425 -0.328998987 -0.054516441
#> 307  0.40202423 -0.613776399  0.032689064
#> 308  0.54416725 -0.744531791 -0.042599613
#> 309  0.76670310 -0.275970677  0.434561938
#> 310  0.75038469 -0.123782550  0.884078330
#> 311  0.63247266 -0.618360736  0.375418289
#> 312  0.73127849 -0.075425079  0.434010527
#> 313  0.15131973 -0.483974299  0.628148194
#> 314  0.37159860 -0.735905125 -0.253381970
#> 315  0.54705814 -0.159204317  0.717854198
#> 316  0.14935089 -1.012780977 -0.726131352
#> 317  0.53919563 -0.391171612 -0.098356341
#> 318 -0.16422338 -0.212907700  0.620332644
#> 319  0.67066154 -0.392527359  0.008162344
#> 320  0.67072719 -0.605057073  0.012074582
#> 321  0.46121198 -0.758475052 -0.173477525
#> 322  0.01758102 -1.002251679 -0.302855548
#> 323  0.55709050 -0.543367570  0.118126249
#> 324  0.56533961 -0.198750785  0.398870305
#> 325  0.48551515 -0.492480140  0.326374870
#> 326  0.15532625 -0.231935466  0.008341107
#> 327  0.36174623  0.099266081  0.665597551
#> 328  0.31038657 -1.009956699 -0.094186151
#> 329  0.43498491 -0.102466953  0.723228271
#> 330  0.28367552 -0.611083745  0.106373612
#> 331  0.38141101 -0.334266628  0.401609039
#> 332  0.27014874 -0.557904856 -0.169520606
#> 333  0.60497471 -0.488720665 -0.293090825
#> 334  0.60146225 -0.491562183 -0.361706948
#> 335  0.09089794 -0.658411935 -0.064952637
#> 336  0.69853688 -0.473432366 -0.131771917
#> 337  0.69763789 -0.031952138  0.740975681
#> 338  0.45511627 -0.776497885 -0.212520026
#> 339  0.42978401 -0.236330768  0.606573711
#> 340  0.44267156 -0.249578146  0.293117461
#> 341  0.55176180 -0.476876872 -0.031423581
#> 342  0.85028272  0.012006798  0.396762960
#> 343  0.79578808 -0.490360213  0.147466468
#> 344 -0.11306736 -0.942334206  0.037702486
#> 345  0.20601100 -0.793558522  0.440218725
#> 346  0.62885113 -0.167853706  0.438118530
#> 347  0.09668311  0.003454291  0.698645110
#> 348  0.50399823 -0.551904521  0.049443453
#> 349  0.67436621 -0.379616076 -0.022379756
#> 350  0.59358002 -0.494952813  0.056930499
#> 351  0.66311155 -0.343567793  0.313387267
#> 352  0.61755575 -0.460119181 -0.101496816
#> 353  0.58649283 -0.338084972 -0.001450145
#> 354  0.23038089 -0.882145289 -0.602178046
#> 355  0.44194804 -0.839442781 -0.160673814
#> 356  0.60445465 -0.169201739  0.692782068
#> 357  0.77143735  0.114138509  0.687243248
#> 358  0.56472417 -0.638768530  0.206870883
#> 359  0.77399339 -0.312307155  0.267886314
#> 360  0.58043216 -0.430542711  0.601859487
#> 361  0.38945841 -0.391371985  0.034204097
#> 362  0.51706464 -0.674469247 -0.031371269
#> 363  0.71099812 -0.218708896  0.095396511
#> 364  0.52645033 -0.473584335 -0.144496556
#> 365  0.45099313 -0.249759058  0.241260040
#> 366  0.69546290 -0.162214788  0.656817663
#> 367  0.70355242  0.114623019  0.646468829
#> 368  0.40936848 -0.092108128  0.732256043
#> 369  0.57341779 -0.031701995  0.407947628
#> 370  0.33007984 -0.625933377 -0.111002171
#> 371  0.34081097 -0.654411158 -0.205682608
#> 372  0.48448063 -0.433959477  0.177192852
#> 373  0.36093024 -0.857493135 -0.399830704
#> 374  0.34684012 -0.434629897 -0.026670630
#> 375  0.46950992 -0.619321899 -0.134628904
#> 376  0.67276730 -0.147119398  0.314221020
#> 377  0.53798054 -0.244203619  0.166652137
#> 378  0.44179113 -0.855558338 -0.469819619
#> 379  0.54669059 -0.135297089  0.936665109
#> 380  0.30097395 -0.219791694  0.144141108
#> 381  0.61315291 -0.590006223  0.188581006
#> 382  0.51837575 -0.398265647  0.040527846
#> 383  0.16808151 -0.691752231 -0.413220410
#> 384  0.75766358 -0.147996818  0.151820888
#> 385  0.37830860 -0.304624175  0.580242812
#> 386  0.64275181 -0.605750855  0.792861147
#> 387  0.32737537 -0.380495920 -0.191807824
#> 388  0.44408541 -0.725014623 -0.360153786
#> 389  0.42329314 -0.438900224  0.266297893
#> 390  0.50406550 -0.714867875 -0.306053922
#> 391  0.52010960 -0.042538893  0.476870185
#> 392  0.79830859 -0.176889352  0.256241774
#> 393  0.39599292 -0.307469179  0.078580526
#> 394  0.75947379 -0.317391692  0.342369169
#> 395  0.53590307 -0.426801784  0.188999331
#> 396  0.49886314 -0.074410578  0.366313209
#> 397  0.85175791  0.066603038  0.585827745
#> 398  0.54493819 -0.690024497 -0.074526471
#> 399  0.33855311 -0.579301015  0.336883205
#> 400  0.37969117 -0.810228167 -0.218240050
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 400 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.545 -0.237   0.457 
#>  2      2        4       70        1  0.338 -0.402   0.138 
#>  3      3        4       70        1  0.500 -0.262   0.293 
#>  4      4        4       70        1  0.311 -0.323   0.254 
#>  5      5        4       70        1  0.364 -1.10   -0.303 
#>  6      6        4       70        1  0.863 -0.191   0.286 
#>  7      7        4       70        1  0.813 -0.0857  0.815 
#>  8      8        4       70        1  0.290 -0.472   0.100 
#>  9      9        4       70        1  0.522 -0.444  -0.0516
#> 10     10        4       70        1  0.216 -1.01   -0.388 
#> # … with 390 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 800 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  6.90  55.2  1.58 0.125 23.5      6.28e+2 1296.
#> 2      1    14     4    70     1  6.90  55.2  1.58 0.125  7.07     7.58e-7  390.
#> 3      2     1     4    70     1  5.61  46.9  1.15 0.120 23.1      8.57e+2 1084.
#> 4      2    14     4    70     1  5.61  46.9  1.15 0.120  9.19     2.84e-4  431.
#> 5      3     1     4    70     1  6.59  53.8  1.34 0.122 22.1      7.45e+2 1188.
#> 6      3    14     4    70     1  6.59  53.8  1.34 0.122  7.59     2.01e-5  409.
#> # … with 794 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>