Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.40724811 -0.563030161  0.024021160
#> 2    0.81114568 -0.110337316  0.056504832
#> 3    0.12305559 -0.798025341  0.214300318
#> 4    0.51633018 -0.151828205  0.307573339
#> 5    0.77447898 -0.108730610  0.551635983
#> 6    0.49079173 -0.354675200  0.003646606
#> 7    0.71255279 -0.459976731  0.130391402
#> 8    0.64376520 -0.465368428  0.357712480
#> 9    0.67891164 -0.369511650  0.464462086
#> 10   0.50999711  0.006100256  0.501523061
#> 11   0.60211068 -0.506164720  0.161508835
#> 12   0.53467781 -0.290158899  0.653308924
#> 13   0.30551284 -0.646386356  0.310843478
#> 14   0.48252547 -0.612463109 -0.359678265
#> 15   0.60616342 -0.188623774  0.650670197
#> 16   0.41325485 -0.269308778  0.215313102
#> 17   0.80534116 -0.509250930 -0.134419998
#> 18   0.29525589 -0.870778188 -0.266469345
#> 19   0.30843774 -0.774518817 -0.092544206
#> 20   0.44341975 -0.456903155  0.044147116
#> 21   0.45556449 -0.368768240 -0.181832064
#> 22   0.49986244 -0.372678275  0.281556420
#> 23   0.68558958 -0.466350022 -0.007472740
#> 24   0.62021839 -0.663196418 -0.338663195
#> 25   0.54346411 -0.736113314 -0.154516098
#> 26   0.46810716 -0.562021559  0.050963946
#> 27   0.42819893 -0.648301575 -0.257027130
#> 28   0.57031956 -0.374778034 -0.040288335
#> 29   0.61054638 -0.077630968  0.534770044
#> 30   0.51449289 -0.259913093  0.750363813
#> 31   0.74930802 -0.164666311  0.167144258
#> 32   0.76727874 -0.391483870  0.294896119
#> 33   0.69212535 -0.079457901  0.313828707
#> 34   0.62488375 -0.354105032  0.421647317
#> 35   0.56914053  0.007012069  0.791853732
#> 36   0.64130561 -0.377613884 -0.012506102
#> 37   0.36059214 -0.129325952  0.356552754
#> 38   0.55722775 -0.276262116  0.044992769
#> 39   0.50017650 -0.649341860 -0.463561650
#> 40   0.61704468 -0.663278730 -0.271667859
#> 41   0.35985516 -0.492073132 -0.196825367
#> 42   0.30906490 -0.602324893 -0.342556745
#> 43   0.55790578 -0.457062133  0.420471747
#> 44   0.35764039 -0.428515713  0.059715457
#> 45   0.12293162 -0.608056645  0.438637670
#> 46   0.65159622 -0.230761434 -0.102654171
#> 47   0.57516879 -0.230715910  0.911452424
#> 48   0.49877334 -0.079345798  0.424548880
#> 49   0.86132480 -0.239350275  0.718566397
#> 50   0.74385923  0.072715489  0.836524138
#> 51   0.00000000  0.000000000  0.000000000
#> 52   0.45746971 -0.897543636  0.568924579
#> 53   0.25706727 -0.488742654 -0.126433915
#> 54   0.77709488 -0.038371275  0.416589300
#> 55   0.50230392 -0.282611828  0.386640710
#> 56   0.54861797 -0.620359156  0.109686401
#> 57   0.75013086 -0.307488976  0.038291109
#> 58   0.51554245 -0.002879128  0.780816140
#> 59   0.42249259 -0.033981756  0.446226653
#> 60   0.69115757 -0.034362509  0.741642418
#> 61   0.33264648 -0.173253981  0.496776974
#> 62   0.40911373 -0.495052869 -0.362851153
#> 63   0.51458281 -0.526967721 -0.094508739
#> 64   0.54648326 -0.488712420 -0.187693787
#> 65   0.62519583 -0.175074775  0.807044469
#> 66   0.47492816 -0.331933105  0.387466961
#> 67   0.65235268 -0.413473484  0.510944182
#> 68   0.54857774 -0.324282272  0.650774667
#> 69   0.70025067 -0.137400166  0.489193284
#> 70   0.64082333 -0.281925642  0.091037845
#> 71   0.01065260 -0.404277767  0.259191020
#> 72   0.52900391 -0.221640608  0.104667645
#> 73   0.72398979 -0.688791705 -0.071975868
#> 74   0.38182074 -0.813560664 -0.214194461
#> 75   0.18317080 -0.644669717  0.054994854
#> 76   0.81953866 -0.129080906  0.705106121
#> 77   0.41834539 -0.719796557  0.082003237
#> 78   0.53704078 -0.414576656 -0.100080012
#> 79   0.57609252 -0.580794185 -0.234245184
#> 80   0.48264993 -0.840443977 -0.122120935
#> 81   0.68066536 -0.280088969  0.274558701
#> 82   0.20082927 -0.795288416  0.127442150
#> 83   0.66103074 -0.222148406  0.239650188
#> 84   0.70714278 -0.219377644  0.789180126
#> 85   0.32535128 -0.831682608 -0.182062192
#> 86   0.57269176 -0.382498665  0.570136217
#> 87   0.38256449 -0.670831758  0.423190048
#> 88   0.53889312 -0.514421231 -0.208715285
#> 89   0.72907564 -0.364021492  0.542533185
#> 90   0.71364531 -0.178311121  0.671033998
#> 91   0.55489160 -0.292349780  0.442871675
#> 92   0.13813960 -0.877512133  0.110538811
#> 93   0.73298678 -0.022147763  0.348447964
#> 94  -0.44378584 -0.478718355  0.989763748
#> 95   0.09672523  0.029995441  0.586702566
#> 96   0.34230907 -0.337439274  0.359174943
#> 97   0.74623185 -0.364881750  0.057345358
#> 98   0.53219919 -0.732652539 -0.400099299
#> 99   0.35553164 -0.634254439 -0.461993056
#> 100  0.71376140 -0.568168626 -0.118299509
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl   ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>    <dbl>   <dbl>
#>  1      1        4       70        1  0.407 -0.563   0.0240 
#>  2      2        4       70        1  0.811 -0.110   0.0565 
#>  3      3        4       70        1  0.123 -0.798   0.214  
#>  4      4        4       70        1  0.516 -0.152   0.308  
#>  5      5        4       70        1  0.774 -0.109   0.552  
#>  6      6        4       70        1  0.491 -0.355   0.00365
#>  7      7        4       70        1  0.713 -0.460   0.130  
#>  8      8        4       70        1  0.644 -0.465   0.358  
#>  9      9        4       70        1  0.679 -0.370   0.464  
#> 10     10        4       70        1  0.510  0.00610 0.502  
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  6.01  39.9  1.02 0.151 25.0  938.         995.
#> 2      1    14     4    70     1  6.01  39.9  1.02 0.151  7.40   0.00155    295.
#> 3      2     1     4    70     1  9.00  62.7  1.06 0.144 16.3  915.        1020.
#> 4      2    14     4    70     1  9.00  62.7  1.06 0.144  5.13   0.000971   321.
#> 5      3     1     4    70     1  4.52  31.5  1.24 0.144 35.7  802.        1124.
#> 6      3    14     4    70     1  4.52  31.5  1.24 0.144  9.97   0.0000811  314.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>