Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> $eta
#>            ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.1321786514  0.136649342       0       0
#> 2   -0.0766561657 -0.067676990       0       0
#> 3   -0.0783912008  0.050137486       0       0
#> 4   -0.1739583067  0.147562598       0       0
#> 5   -0.2386348253  0.162899548       0       0
#> 6   -0.0638498943  0.129531907       0       0
#> 7   -0.0792847947  0.059858992       0       0
#> 8   -0.1613289849  0.221907189       0       0
#> 9   -0.0498492664  0.006913345       0       0
#> 10  -0.1058638243  0.060336266       0       0
#> 11  -0.0653855508 -0.013031366       0       0
#> 12  -0.0226469763  0.028581412       0       0
#> 13  -0.1713516040  0.330830812       0       0
#> 14  -0.1019327259  0.092063892       0       0
#> 15  -0.1249022437 -0.059838588       0       0
#> 16  -0.1236883457 -0.108441613       0       0
#> 17  -0.1950191130  0.138927947       0       0
#> 18  -0.1134636926  0.180800436       0       0
#> 19  -0.3482282710  0.217439394       0       0
#> 20  -0.0784967391  0.100125645       0       0
#> 21  -0.0940507097  0.003298516       0       0
#> 22  -0.0441450225 -0.223201709       0       0
#> 23  -0.0651870020 -0.092898755       0       0
#> 24  -0.0651870020  0.024094078       0       0
#> 25   0.0278765757 -0.254446636       0       0
#> 26  -0.2698477909 -0.111094483       0       0
#> 27  -0.1316052599  0.042630581       0       0
#> 28  -0.2018341622  0.036448432       0       0
#> 29  -0.1158456621  0.114480517       0       0
#> 30  -0.1189730930  0.165267268       0       0
#> 31  -0.0885568721  0.123722896       0       0
#> 32  -0.2659446476  0.208237589       0       0
#> 33  -0.0901330670  0.188370400       0       0
#> 34  -0.0915369607 -0.023767885       0       0
#> 35  -0.1956246802  0.066370001       0       0
#> 36  -0.0734836518  0.066370001       0       0
#> 37   0.0203713752 -0.205791249       0       0
#> 38  -0.1921745579  0.183146513       0       0
#> 39  -0.2029980356 -0.060118066       0       0
#> 40  -0.0435365863  0.032490835       0       0
#> 41  -0.1573001730 -0.053312709       0       0
#> 42  -0.0310667412 -0.094201509       0       0
#> 43  -0.1637349096  0.223258850       0       0
#> 44  -0.1191602918  0.148539192       0       0
#> 45  -0.1188820716  0.148539192       0       0
#> 46  -0.1051653881  0.143456244       0       0
#> 47  -0.1051653881  0.102367715       0       0
#> 48  -0.0189470518  0.003119051       0       0
#> 49  -0.1506548436  0.249709994       0       0
#> 50  -0.0435408484  0.008217985       0       0
#> 51  -0.1570260432  0.019997537       0       0
#> 52  -0.0557748996 -0.036638426       0       0
#> 53  -0.0254366665 -0.102895640       0       0
#> 54  -0.2230335121  0.134265323       0       0
#> 55  -0.1997595205  0.020873275       0       0
#> 56  -0.1631581268  0.135899340       0       0
#> 57  -0.0917446162  0.048794510       0       0
#> 58  -0.0731610120  0.070093268       0       0
#> 59  -0.0964136333  0.126845099       0       0
#> 60  -0.0923344172  0.156158251       0       0
#> 61  -0.0731662557 -0.013406232       0       0
#> 62  -0.1184032709 -0.025900751       0       0
#> 63  -0.1663696130  0.193455286       0       0
#> 64  -0.1349960779  0.171368166       0       0
#> 65  -0.1947354197  0.088151965       0       0
#> 66  -0.0506294465 -0.215472885       0       0
#> 67   0.0566001528  0.036487971       0       0
#> 68  -0.0389876324  0.006933513       0       0
#> 69  -0.1071884067  0.054637288       0       0
#> 70  -0.3334923325  0.174196232       0       0
#> 71  -0.1974054749  0.149493688       0       0
#> 72  -0.1974054749  0.149493688       0       0
#> 73  -0.0657492698  0.137339013       0       0
#> 74  -0.2179300579  0.196280114       0       0
#> 75  -0.1867702946  0.140736601       0       0
#> 76  -0.1328586617 -0.011752652       0       0
#> 77  -0.0996342150  0.106283479       0       0
#> 78  -0.1792367452  0.258435579       0       0
#> 79  -0.3555337157 -0.077561459       0       0
#> 80  -0.2573127440  0.074225168       0       0
#> 81  -0.2127249560 -0.123898251       0       0
#> 82  -0.1664956898  0.093149696       0       0
#> 83  -0.0651493720  0.146185145       0       0
#> 84  -0.1623145614  0.117993225       0       0
#> 85  -0.1552199472  0.117993225       0       0
#> 86  -0.0965219768  0.107897042       0       0
#> 87  -0.2867982133  0.294401585       0       0
#> 88  -0.1870872816  0.051641128       0       0
#> 89  -0.1066112905 -0.027498560       0       0
#> 90  -0.1168997367 -0.027498560       0       0
#> 91  -0.1732872107  0.023155220       0       0
#> 92  -0.0434289841  0.106079198       0       0
#> 93  -0.1709741250  0.134107371       0       0
#> 94  -0.1481157592  0.100794064       0       0
#> 95  -0.0990623974  0.071243579       0       0
#> 96   0.0545272808 -0.029816015       0       0
#> 97   0.0005907839 -0.090886517       0       0
#> 98   0.0673163572  0.055715936       0       0
#> 99  -0.1432774640  0.131031907       0       0
#> 100 -0.1397820932  0.043031985       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.132   0.137  
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.0767 -0.0677 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.0784  0.0501 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.174   0.148  
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.239   0.163  
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0638  0.130  
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0793  0.0599 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.161   0.222  
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0498  0.00691
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.106   0.0603 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.218  22.3 0.041  0.12 73.9  1650.
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.218  22.3 0.041  0.12  5.25  117.
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.231  18.2 0.041  0.12 89.9  1635.
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.231  18.2 0.041  0.12  5.66  103.
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.230  20.5 0.041  0.12 79.9  1635.
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.230  20.5 0.041  0.12  5.05  103.
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>