Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object

A posologyr list, created by the posologyr function.

return_model

A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01,n_iter=50,n_chains=2)
#> $eta
#>           ETA_Cl       ETA_Vc        ETA_Ka
#> 1    0.721331085 -0.115437266  0.1990524519
#> 2    0.648567024 -0.371608358  0.1672634095
#> 3    0.488750963 -0.542870410 -0.3130186578
#> 4    0.595393064 -0.545640704  0.0499065495
#> 5    0.298694053 -1.137253926 -0.6674955046
#> 6    0.738670208 -0.022505470  0.5698362982
#> 7    0.375926444 -0.214949292  0.1803104323
#> 8    0.697179267 -0.177819741  0.4645566583
#> 9    0.672311004 -0.479178417  0.0491444684
#> 10   0.404795132 -0.696672802  0.0523540379
#> 11   0.583878047  0.029257397  1.2371970580
#> 12   0.649682784 -0.271891750  0.2421446389
#> 13   0.355878316 -0.688876475 -0.1616677731
#> 14   0.572892351 -0.378797641  0.5133812124
#> 15   0.472085376 -0.337400534  0.2310086235
#> 16   0.667121491 -0.530567081 -0.1064582135
#> 17   0.703678833 -0.185781472  0.4175047084
#> 18   0.413905700 -0.590603577 -0.0800574460
#> 19   0.705914976 -0.066543242  0.1534606502
#> 20   0.614282815 -0.233263796  0.2956362262
#> 21   0.520175413 -0.494502997  0.1336127906
#> 22   0.555287017 -0.056992112  0.4584460725
#> 23   0.338455540 -0.931944369  0.4132288735
#> 24   0.204914959 -0.626303917 -0.2400682876
#> 25   0.343511954 -0.611490352  0.9255531176
#> 26   0.525473352 -0.330728511  0.4667312020
#> 27   0.400222630 -0.547862140  0.1833433117
#> 28   0.744573956 -0.543745445 -0.6934890598
#> 29   0.425606233 -0.849598539 -0.4469413116
#> 30   0.706873683  0.188430850  0.6266989415
#> 31   0.813685419 -0.319102809 -0.0546961790
#> 32   0.314495552 -1.023701550 -0.3853925017
#> 33   0.452507680 -0.373737629 -0.1559376280
#> 34   0.591661616 -0.326841277  1.0479066508
#> 35   0.544931546 -0.827680322  0.5584299781
#> 36   0.167008695 -0.365554543  0.4046703010
#> 37   0.217425487 -0.022025537  0.5954474453
#> 38  -0.418436730 -0.430663802 -0.0771634924
#> 39  -0.200620055 -0.172938304  0.1967968332
#> 40   0.081410796 -0.551338929  0.0345282254
#> 41   0.676926138 -0.506767678 -0.1427925426
#> 42   0.453562738 -0.655677415 -0.3012065742
#> 43   0.539227176 -0.653357845 -0.0000247292
#> 44   0.461129198 -0.691223857 -0.1580856614
#> 45   0.543995299 -0.361353337 -0.1815432390
#> 46   0.658625322 -0.394584410  0.0289192953
#> 47   0.362125517 -0.258363760  0.1596536857
#> 48   0.751265147 -0.359969009  0.2111917218
#> 49   0.555706423  0.184776719  1.1697002814
#> 50   0.689822708 -0.694418746  0.2672257004
#> 51   0.000000000  0.000000000  0.0000000000
#> 52   0.282210332 -0.011552000  0.8146013973
#> 53   0.689648895 -0.727229582 -0.0079869486
#> 54   0.536808523 -0.522108495  0.3379460637
#> 55   0.683747928 -0.492416012  0.3171341583
#> 56   0.462376719 -0.481951289  0.0348586032
#> 57   0.748860724 -0.342765515 -0.1430524161
#> 58   0.422240935 -0.428187844  0.4127676739
#> 59   0.648177411 -0.484999024  0.0587154389
#> 60   0.203655786 -0.725349415  0.1274532294
#> 61   0.634771430 -0.575517812  0.0727592187
#> 62   0.467329957 -0.564214296  0.2864737415
#> 63   0.376767285  0.170746371  0.5445691743
#> 64   0.719526205 -0.213360602  0.5221861194
#> 65   0.725452546 -0.205583755  0.0819310650
#> 66   0.827772520 -0.202455834  0.2134778396
#> 67   0.307445843 -0.392406700  0.4370617637
#> 68   0.348569379 -0.089920457  0.5895866088
#> 69   0.277805296 -0.391803057  0.2443182134
#> 70   0.200839227 -0.663859358 -0.0557775029
#> 71   0.143351261 -0.969634463 -0.3862894167
#> 72   0.800990005 -0.313316456  0.4020057868
#> 73   0.782186543 -0.211152586  0.4846553027
#> 74  -0.006935032 -0.351988788  0.2205150825
#> 75   0.465007520 -0.356214739  0.1362886267
#> 76   0.103234322 -0.360866065  0.0966142656
#> 77   0.158650301 -0.702076511 -0.3915699664
#> 78   0.764216476 -0.288677942  0.1973818597
#> 79   0.351694677 -0.730143376 -0.1208636058
#> 80   0.701317316 -0.006239072  0.5862644181
#> 81   0.611904216 -0.289402559  0.5123936844
#> 82   0.649998342 -0.365427715  0.0272778380
#> 83   0.420298824 -0.986520965 -0.3353715407
#> 84   0.112252718 -0.058054975  0.3298147856
#> 85   0.390775657 -0.494827765  0.5060244710
#> 86   0.665811764 -0.348720738  0.1826807697
#> 87   0.550407963 -0.673571295 -0.0310263285
#> 88   0.299350469 -0.726402450  0.0274809445
#> 89   0.750678714 -0.167446928  0.3756614455
#> 90   0.684836531 -0.367090092  0.1814580207
#> 91   0.716330724 -0.217837727  1.2622548526
#> 92   0.473511863 -0.790083420 -0.3441425868
#> 93   0.816331017 -0.111818044  0.4904620880
#> 94   0.782929088 -0.345228661  0.5144971440
#> 95   0.867240329  0.070980506  0.8379076884
#> 96   0.625927263 -0.683970750  0.1408159743
#> 97   0.440109331  0.048106904  1.2618230315
#> 98   0.340682750 -0.344108320  0.3337058298
#> 99   0.685673844 -0.161253462  0.1407372887
#> 100  0.583992931 -0.451672573  0.5902987603
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.721 -0.115   0.199 
#>  2      2        4       70        1  0.649 -0.372   0.167 
#>  3      3        4       70        1  0.489 -0.543  -0.313 
#>  4      4        4       70        1  0.595 -0.546   0.0499
#>  5      5        4       70        1  0.299 -1.14   -0.667 
#>  6      6        4       70        1  0.739 -0.0225  0.570 
#>  7      7        4       70        1  0.376 -0.215   0.180 
#>  8      8        4       70        1  0.697 -0.178   0.465 
#>  9      9        4       70        1  0.672 -0.479   0.0491
#> 10     10        4       70        1  0.405 -0.697   0.0524
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  8.23  62.4 1.22  0.132 18.0   813.       1120.
#> 2      1    14     4    70     1  8.23  62.4 1.22  0.132  5.86    0.000105  366.
#> 3      2     1     4    70     1  7.65  48.3 1.18  0.158 22.5   836.       1086.
#> 4      2    14     4    70     1  7.65  48.3 1.18  0.158  5.41    0.000177  261.
#> 5      3     1     4    70     1  6.52  40.7 0.731 0.160 19.3  1162.        784.
#> 6      3    14     4    70     1  6.52  40.7 0.731 0.160  6.95    0.0865    283.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>