Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.730595594 -0.718302023  0.444837320
#> 2    0.745776882 -0.331930012  0.414656149
#> 3    0.856606369  0.082307011  0.090094048
#> 4   -0.079980505 -0.590162530 -0.001739851
#> 5    0.669049865  0.034096793  1.042534295
#> 6    0.669049865 -0.362718181  0.213511174
#> 7    0.256493979 -0.663827118  0.057663578
#> 8   -0.079969722 -0.973732056 -0.440229013
#> 9    0.671533199 -0.699360305 -0.571595690
#> 10   0.789260811 -0.699360305 -0.571595690
#> 11   0.671074483 -0.390399862  0.349553100
#> 12   0.725284202 -0.349481968  0.411444057
#> 13   0.317651039 -0.136073974  0.105940319
#> 14   0.732136017 -0.589280286  0.357950774
#> 15   0.690828342 -0.526961909  0.039123930
#> 16   0.722722504 -0.270222511  0.337313246
#> 17   0.356181369 -0.562909182 -0.488648238
#> 18   0.473677818 -0.335527921  0.274967506
#> 19   0.525542076 -0.394765300  0.097952646
#> 20   0.788713970  0.010519841  0.522989812
#> 21   0.375544870 -0.653146516 -0.071004677
#> 22   0.463154812 -0.701497958 -0.235348627
#> 23   0.489998566 -0.919814723  0.339711303
#> 24   0.417705488 -0.478160455  0.217683462
#> 25   0.506953099 -0.669672571  0.142214030
#> 26   0.373777544 -0.358983977  0.142214030
#> 27   0.360887426 -0.329741296 -0.141013263
#> 28   0.484650202 -0.625586276  0.081541145
#> 29   0.520990780 -0.430394853  0.388520739
#> 30   0.754127462 -0.331561493  0.513405645
#> 31   0.416165995  0.104489700  0.320447077
#> 32   0.550026133 -0.427586933  0.377118992
#> 33   0.367997051 -0.826842822 -0.045802030
#> 34   0.169287244 -0.750549473 -0.084951715
#> 35   0.627837597 -0.157764491  0.439709606
#> 36   0.617553821 -0.563450614  0.838201292
#> 37   0.614459013 -0.335265784  0.335775737
#> 38   0.614459013 -0.515428029 -0.421991789
#> 39   0.463191943 -0.591978484  0.259520179
#> 40   0.390360349 -0.586877060  0.027197322
#> 41   0.344079170 -0.420742569  0.100979274
#> 42   0.459651489 -0.445893864 -0.153948330
#> 43   0.641951427 -0.440474211 -0.238399213
#> 44   0.272295075 -0.423945531  0.235046953
#> 45   0.811189746 -0.141026581  0.195365491
#> 46   0.738051420 -0.162560308  0.195365491
#> 47   0.615618279 -0.815129853 -0.007723661
#> 48   0.336356291 -0.579090948 -0.472522731
#> 49   0.311311902 -0.300205914 -0.006782243
#> 50   0.642472381 -0.300205914 -0.006782243
#> 51   0.000000000  0.000000000  0.000000000
#> 52   0.569162169 -0.584883683 -0.131603932
#> 53   0.881900433 -0.008700325  0.935075724
#> 54   0.720694737 -0.122193232  0.843496738
#> 55   0.719683632 -0.304495848 -0.045471488
#> 56   0.987232885 -0.200327641 -0.255389496
#> 57   0.468380051 -0.513855433  0.391708190
#> 58   0.433072281 -0.432021291  0.413030815
#> 59   0.642484420 -0.907529287 -0.458212991
#> 60   0.603387485 -0.161238857  0.324321935
#> 61   0.677210436 -0.044038564  0.458973035
#> 62   0.822902074 -0.046347371  0.484966733
#> 63   0.394163329 -0.915680444 -0.445875679
#> 64   0.522761395 -0.235042268  0.418485818
#> 65   0.889670712 -0.019577152  1.301530335
#> 66   0.732179099 -0.319674002  0.506809967
#> 67   0.622472705 -0.255111070  1.116989589
#> 68   0.628527738 -0.585453920  0.136518206
#> 69   0.239731846 -0.720955167 -0.461945797
#> 70   0.414062644 -0.755474610 -0.328137733
#> 71   0.102014974 -0.546496015  0.054279220
#> 72   0.765845825 -0.519855924  0.035627171
#> 73   0.674835160  0.149381944  0.633172616
#> 74   0.539381782 -0.112166987  0.242179440
#> 75   0.669772445 -0.242900754  0.349802889
#> 76   0.381645278  0.233005028  0.595014972
#> 77   0.702089774 -0.149720433  1.024215772
#> 78   0.474300696 -0.649773710  0.400690716
#> 79   0.297992078 -0.662002681 -0.176129289
#> 80   0.640858644 -0.295306515  0.432950414
#> 81   0.828623627 -0.270134317  0.583977901
#> 82   0.192248816 -0.913853561 -0.313438705
#> 83   0.624491441 -0.265272567  0.307176777
#> 84   0.704626081 -0.467864746  0.708728449
#> 85   0.784259439  0.068002391  1.056689360
#> 86   0.504886756 -0.419112014  0.149839371
#> 87   0.353608311 -0.532949271 -0.483181142
#> 88   0.457492426 -0.561278583  0.401066918
#> 89  -0.108792241 -0.399011823  0.083429018
#> 90   0.381005850 -1.257842606 -1.058975031
#> 91   0.380671365 -0.303467226  0.214926720
#> 92   0.237662709 -0.445359794  0.563910986
#> 93   0.454983824 -0.833925286  0.155866468
#> 94   0.525759864 -0.554253780 -0.276753584
#> 95   0.525218047 -0.407158739  0.148590320
#> 96   0.389938991 -1.357824357 -1.236362675
#> 97   0.485282920 -0.567861376  0.310769649
#> 98   0.004509157 -0.186143095  0.831885108
#> 99   0.281712732 -0.196184418  1.205492256
#> 100  0.452742016 -0.657757040  0.294297568
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1      1        4       70        1  0.731  -0.718   0.445  
#>  2      2        4       70        1  0.746  -0.332   0.415  
#>  3      3        4       70        1  0.857   0.0823  0.0901 
#>  4      4        4       70        1 -0.0800 -0.590  -0.00174
#>  5      5        4       70        1  0.669   0.0341  1.04   
#>  6      6        4       70        1  0.669  -0.363   0.214  
#>  7      7        4       70        1  0.256  -0.664   0.0577 
#>  8      8        4       70        1 -0.0800 -0.974  -0.440  
#>  9      9        4       70        1  0.672  -0.699  -0.572  
#> 10     10        4       70        1  0.789  -0.699  -0.572  
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc      depot  centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>      <dbl>  <dbl>
#> 1      1     1     4    70     1  8.31  34.1  1.56 0.243 35.8     6.36e+2 1221. 
#> 2      1    14     4    70     1  8.31  34.1  1.56 0.243  2.45    9.88e-7   83.5
#> 3      2     1     4    70     1  8.43  50.2  1.51 0.168 24.8     6.58e+2 1244. 
#> 4      2    14     4    70     1  8.43  50.2  1.51 0.168  4.45    1.87e-6  224. 
#> 5      3     1     4    70     1  9.42  76.0  1.09 0.124 13.8     8.91e+2 1050. 
#> 6      3    14     4    70     1  9.42  76.0  1.09 0.124  5.40    5.91e-4  410. 
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>