Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>         ETA_Cl       ETA_Vc        ETA_Ka
#> 1   0.48747162 -0.866147326 -0.5148822378
#> 2   0.54299573 -0.392200711  0.1391080418
#> 3   0.52666204 -0.031115077  0.5301588455
#> 4   0.25413716 -0.508285071  0.5359432291
#> 5   0.47029293 -0.314136393  0.3519474931
#> 6   0.74280555  0.008876238  0.7146346048
#> 7   0.90671221  0.004857614  0.9363851290
#> 8   0.51935206 -0.812877693  0.0596885038
#> 9   0.65179481 -0.779522872 -0.6768563530
#> 10  0.19443078 -0.103913969  0.4029671691
#> 11  0.58872816 -0.522299060  0.1431930866
#> 12  0.08675310 -0.155935649  0.2507841428
#> 13  0.68006681 -0.102275844  0.1369114375
#> 14  0.41945979 -0.603832195  0.2345546632
#> 15  0.72435499 -0.008598737  0.5527271840
#> 16  0.64420959 -0.371946865  0.0467461412
#> 17  0.45506754 -0.878092770 -0.3981455306
#> 18  0.44252543 -0.599406424 -0.2092167612
#> 19  0.67973214 -0.513833557  0.0188154775
#> 20  0.50075599 -0.513833557  0.3637833787
#> 21  0.44226290 -0.630449137 -0.2611577160
#> 22  0.29687114 -0.759905181 -0.1432580189
#> 23  0.50036830 -0.517770907  0.0032531511
#> 24  0.56719813 -0.533076261 -0.2107537889
#> 25  0.77039959 -0.250765882  0.4386840495
#> 26  0.56574478 -0.353593587 -0.0301892165
#> 27  0.39278276 -0.195071139  0.0668401651
#> 28  0.76900112 -0.300705290  0.2884971342
#> 29  0.58774359 -0.573895907  0.3824831509
#> 30  0.66604274 -0.092670539  0.4371855103
#> 31  0.51064767 -0.221216207  0.3114095812
#> 32  0.49269139 -0.761764083 -0.4402387948
#> 33  0.28372526 -1.096732333 -0.1203977101
#> 34  0.46038798 -0.925483987 -0.6258142745
#> 35  0.64898487 -0.394431784 -0.0651116098
#> 36  0.73602126 -0.694016886 -0.4330030842
#> 37  0.42837923 -0.644271535  0.0672122369
#> 38  0.24800895 -0.410060692  0.6372687191
#> 39  0.41773403 -0.603288742  0.2591052874
#> 40  0.52808063 -0.486906487 -0.1182104720
#> 41  0.15045427 -0.398853905  0.3282555133
#> 42  0.58320496 -0.264613869  0.0383569413
#> 43  0.27591557 -0.707263212  0.0001470677
#> 44  0.58819749 -0.435688669  0.0984299222
#> 45  0.74164068 -0.032620485  0.3519780228
#> 46  0.40017600 -0.557771991  0.3202409809
#> 47  0.66087670 -0.362074814  0.5310081714
#> 48  0.48350338 -0.148117276  0.3721329479
#> 49  0.39416869 -0.514463939  0.0606267071
#> 50  0.46425854 -0.813689717 -0.1020371466
#> 51  0.00000000  0.000000000  0.0000000000
#> 52  0.31446918 -1.259763600 -0.9432173057
#> 53  0.42184709 -1.053082775 -0.4186749663
#> 54  0.52188460 -1.014146572 -0.7050755852
#> 55  0.60083241 -0.660276762 -0.5600860845
#> 56  0.52385793 -0.511491011 -0.0722396498
#> 57  0.89553457 -0.008483781  0.8801940239
#> 58  0.56856914  0.147043300  0.4276336519
#> 59  0.52971371 -0.281352784  0.4210025016
#> 60  0.68057769 -0.423773019  0.2603109932
#> 61  0.54225408 -0.461221542  0.1227508085
#> 62  0.57283954 -0.553607408 -0.0043469607
#> 63  0.56573637  0.192757623  0.6017297894
#> 64  0.57367612 -0.212248619  0.6733902186
#> 65  0.57589025 -0.357760456  0.0959394042
#> 66  0.63562483 -0.213558424  0.8283504241
#> 67  0.47691229 -0.535873797 -0.0115474806
#> 68  0.56485738 -0.493990150 -0.1710502625
#> 69  0.40178164 -0.500716229  0.3608945801
#> 70  0.38028621  0.026593668  0.7422818941
#> 71  0.34564679 -0.077237758  0.3852234780
#> 72  0.76682459 -0.193848945  0.8298390783
#> 73  0.64798802 -0.149156053  0.4856996263
#> 74  0.61149169 -0.250009836  0.6338052078
#> 75  0.62056656 -0.287624576  0.9575780192
#> 76  0.31198549 -0.554091299  0.2316936862
#> 77  0.23404949 -1.139763936  0.0587131850
#> 78  0.37220221 -0.376986360  0.1820067707
#> 79  0.68491636 -0.424281587  0.4328655552
#> 80  0.53923876 -0.212260626  0.2625299622
#> 81  0.48699376 -0.722712722  0.4595245804
#> 82  0.42552361 -0.478336821  0.1059992134
#> 83  0.42168759 -0.763804365 -0.4067374160
#> 84  0.21747472 -1.054465590 -0.6516872250
#> 85  0.57330752 -0.395040303  0.2080234828
#> 86  0.24832514 -0.237224197  0.3747729888
#> 87  0.76696258 -0.218673284  0.4813282918
#> 88  0.01955981 -0.175063723  0.3441998945
#> 89  0.31248894 -0.849808557 -0.1298725523
#> 90  0.54518137 -0.655685262  0.0288587505
#> 91  0.69577452 -0.113785622  0.6636087296
#> 92  0.50963179 -0.723381355 -0.4276183657
#> 93  0.61021506 -0.601103312  0.4374209588
#> 94  0.59594112 -0.582941344 -0.0821742896
#> 95  0.70716705 -0.042619387  0.3249657310
#> 96  0.85127679 -0.108758219  1.2741669148
#> 97  0.47726424 -0.330848332  0.5394421500
#> 98  0.44774364 -0.134397718  0.1968189751
#> 99  0.54171323 -0.472395723 -0.2192865279
#> 100 0.55480883 -0.415287961 -0.0665943202
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl   ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>    <dbl>   <dbl>
#>  1      1        4       70        1  0.487 -0.866   -0.515 
#>  2      2        4       70        1  0.543 -0.392    0.139 
#>  3      3        4       70        1  0.527 -0.0311   0.530 
#>  4      4        4       70        1  0.254 -0.508    0.536 
#>  5      5        4       70        1  0.470 -0.314    0.352 
#>  6      6        4       70        1  0.743  0.00888  0.715 
#>  7      7        4       70        1  0.907  0.00486  0.936 
#>  8      8        4       70        1  0.519 -0.813    0.0597
#>  9      9        4       70        1  0.652 -0.780   -0.677 
#> 10     10        4       70        1  0.194 -0.104    0.403 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  6.51  29.4 0.598 0.221  22.3  1.28e+3  655.
#> 2      1    14     4    70     1  6.51  29.4 0.598 0.221   5.12 5.42e-1  151.
#> 3      2     1     4    70     1  6.88  47.3 1.15  0.146  22.7  8.56e+2 1073.
#> 4      2    14     4    70     1  6.88  47.3 1.15  0.146   6.54 2.78e-4  309.
#> 5      3     1     4    70     1  6.77  67.9 1.70  0.0998 20.0  5.76e+2 1360.
#> 6      3    14     4    70     1  6.77  67.9 1.70  0.0998  7.94 1.47e-7  539.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>