Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)
#> $eta
#>            ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.1409800411  0.037558269       0       0
#> 2   -0.2283745494  0.023711446       0       0
#> 3   -0.2308836395  0.171930568       0       0
#> 4   -0.0358581283  0.024701470       0       0
#> 5   -0.1056562516  0.013800503       0       0
#> 6   -0.1396772109 -0.012849857       0       0
#> 7   -0.2808109221  0.139146295       0       0
#> 8   -0.2357913906  0.200573365       0       0
#> 9   -0.1318348168  0.010601267       0       0
#> 10  -0.1585776226  0.144816817       0       0
#> 11  -0.1809144300  0.142592468       0       0
#> 12  -0.1298200801  0.128657106       0       0
#> 13  -0.0866979937  0.179833788       0       0
#> 14  -0.0793986987 -0.008652249       0       0
#> 15  -0.1700192309  0.110124964       0       0
#> 16  -0.0401165781 -0.053818064       0       0
#> 17  -0.1485179736  0.129957638       0       0
#> 18  -0.0836838625  0.059029370       0       0
#> 19  -0.0519806316  0.069616216       0       0
#> 20  -0.1194580136 -0.021205833       0       0
#> 21  -0.1539048028 -0.102944439       0       0
#> 22  -0.0772646703 -0.102944439       0       0
#> 23  -0.0772646703  0.087329637       0       0
#> 24  -0.0428591089  0.044282296       0       0
#> 25  -0.3130404345 -0.078491280       0       0
#> 26  -0.2907533632 -0.078491280       0       0
#> 27  -0.0785567940 -0.058497539       0       0
#> 28  -0.1910553579  0.190665935       0       0
#> 29  -0.0703327499  0.190665935       0       0
#> 30  -0.2296501110  0.069501919       0       0
#> 31  -0.0390881161 -0.108466147       0       0
#> 32   0.0832968134 -0.143901623       0       0
#> 33  -0.0155783628  0.035157717       0       0
#> 34  -0.2345591149  0.133525223       0       0
#> 35  -0.1488350113  0.186215782       0       0
#> 36  -0.0124057869  0.200919671       0       0
#> 37  -0.1772688766  0.083081017       0       0
#> 38  -0.0564042909  0.165786335       0       0
#> 39  -0.0459774772  0.016720866       0       0
#> 40  -0.0362917100 -0.027571260       0       0
#> 41  -0.1656751339  0.037742787       0       0
#> 42  -0.1519668069  0.144675527       0       0
#> 43  -0.0148669337  0.048549887       0       0
#> 44  -0.1391144161 -0.223538043       0       0
#> 45  -0.1300285119  0.006215513       0       0
#> 46  -0.0393515826  0.099890592       0       0
#> 47  -0.0393515826 -0.073878174       0       0
#> 48  -0.1863646899  0.140094549       0       0
#> 49  -0.1167367173  0.217399397       0       0
#> 50  -0.0744851722  0.008221857       0       0
#> 51  -0.0668270726  0.096094352       0       0
#> 52  -0.1352040739 -0.127117931       0       0
#> 53  -0.0580509676 -0.040438046       0       0
#> 54  -0.0587570262  0.059478434       0       0
#> 55  -0.1081632128  0.084554486       0       0
#> 56  -0.1930841289  0.163114113       0       0
#> 57  -0.1766019620  0.143323034       0       0
#> 58  -0.1780676657  0.119689979       0       0
#> 59  -0.2439365954  0.030444184       0       0
#> 60  -0.1056146491  0.080241203       0       0
#> 61  -0.0946137929  0.020569717       0       0
#> 62  -0.0776294419  0.088782184       0       0
#> 63   0.0193265898  0.240392852       0       0
#> 64  -0.0110144125  0.136381381       0       0
#> 65  -0.1049616805  0.083894119       0       0
#> 66  -0.0849123961 -0.001123942       0       0
#> 67  -0.1778625929 -0.026566468       0       0
#> 68  -0.0827212072  0.017251709       0       0
#> 69   0.0003650363  0.023933238       0       0
#> 70  -0.1639291297 -0.197028052       0       0
#> 71  -0.1932599406  0.142056721       0       0
#> 72  -0.0952636300  0.104833460       0       0
#> 73  -0.0952636300  0.200345180       0       0
#> 74  -0.0934792655  0.035678799       0       0
#> 75  -0.1025983846  0.042487826       0       0
#> 76  -0.2822993727 -0.035591576       0       0
#> 77  -0.1171231321 -0.120859031       0       0
#> 78  -0.0414841739  0.131477611       0       0
#> 79  -0.1854663451 -0.018996583       0       0
#> 80  -0.1098618802  0.197397653       0       0
#> 81  -0.1835329969 -0.156901459       0       0
#> 82  -0.0926510904  0.037175366       0       0
#> 83  -0.1818479543  0.139661375       0       0
#> 84  -0.0423990227  0.089055445       0       0
#> 85  -0.0136656436 -0.109719344       0       0
#> 86  -0.0325342507 -0.035481571       0       0
#> 87  -0.1165910439  0.096749011       0       0
#> 88  -0.0002389761 -0.125755068       0       0
#> 89   0.0775239030  0.015997678       0       0
#> 90  -0.2212582677 -0.002507005       0       0
#> 91  -0.1018817445 -0.002507005       0       0
#> 92  -0.1754673456  0.224442826       0       0
#> 93  -0.0403606943  0.112038100       0       0
#> 94  -0.1564217795  0.091845288       0       0
#> 95  -0.1032956943  0.147697638       0       0
#> 96  -0.0418837365 -0.085257815       0       0
#> 97  -0.1484972199 -0.117065721       0       0
#> 98   0.0167205251  0.115827190       0       0
#> 99  -0.1312935632  0.198582722       0       0
#> 100 -0.0960945185  0.032682782       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.141   0.0376
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.228   0.0237
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.231   0.172 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0359  0.0247
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.106   0.0138
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.140  -0.0128
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.281   0.139 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.236   0.201 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.132   0.0106
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.159   0.145 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.216  20.2 0.041  0.12 81.8  1653.
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.216  20.2 0.041  0.12  5.92  120.
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.198  19.9 0.041  0.12 84.0  1675.
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.198  19.9 0.041  0.12  7.28  145.
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.198  23.1 0.041  0.12 72.5  1676.
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.198  23.1 0.041  0.12  6.31  146.
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>