Estimate the posterior distribution of individual parameters

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (burn-in included) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a dataframe of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> [[1]]
#>        ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1  -0.9741824 -0.029869659       0       0
#> 2  -0.7660286  0.046042672       0       0
#> 3  -0.9772889 -0.076432065       0       0
#> 4  -0.6147328 -0.076432065       0       0
#> 5  -0.9323364  0.034228221       0       0
#> 6  -0.7244219  0.081861525       0       0
#> 7  -0.6863382 -0.185831716       0       0
#> 8  -0.5036758 -0.191031686       0       0
#> 9  -0.7665573 -0.021185418       0       0
#> 10 -0.5878894 -0.079335070       0       0
#> 11 -0.8235777 -0.150278489       0       0
#> 12 -0.5690378  0.101032836       0       0
#> 13 -0.7820964 -0.199110352       0       0
#> 14 -0.7743651 -0.195684668       0       0
#> 15 -0.7544893 -0.125096241       0       0
#> 16 -0.8592159  0.007920796       0       0
#> 17 -0.8051639 -0.010641688       0       0
#> 18 -0.7175788 -0.021172050       0       0
#> 19 -0.6479250 -0.172188974       0       0
#> 20 -0.7045912 -0.215060547       0       0
#> 21 -0.6965113 -0.100623723       0       0
#> 22 -0.8544401 -0.078674616       0       0
#> 23 -0.7340197 -0.054716318       0       0
#> 24 -0.6885733 -0.046396853       0       0
#> 25 -0.7455571  0.035853409       0       0
#> 26 -0.6660415  0.035853409       0       0
#> 27 -0.9077300  0.089722290       0       0
#> 28 -0.6119565  0.042967247       0       0
#> 29 -0.6798548  0.054331802       0       0
#> 30 -0.7382587  0.001116925       0       0
#> 31 -0.8925374 -0.045611491       0       0
#> 32 -0.6998561 -0.099805253       0       0
#> 33 -0.6103461 -0.099805253       0       0
#> 34 -0.6040937 -0.093500355       0       0
#> 35 -0.5634308 -0.093500355       0       0
#> 36 -0.7850921 -0.029370195       0       0
#> 37 -0.9018960 -0.045955413       0       0
#> 38 -0.7341137  0.030229742       0       0
#> 39 -0.9245045  0.039674754       0       0
#> 40 -0.7469656  0.029432049       0       0
#> 41 -0.8341472 -0.074599932       0       0
#> 42 -0.6399317  0.066093101       0       0
#> 43 -0.7550950  0.061291082       0       0
#> 44 -0.7334481 -0.110638542       0       0
#> 45 -0.9835479  0.069186377       0       0
#> 46 -0.6434427  0.017669476       0       0
#> 47 -0.5851910 -0.230835391       0       0
#> 48 -0.6928989 -0.285173832       0       0
#> 49 -0.6932060 -0.026563503       0       0
#> 50 -0.7978958  0.016208426       0       0
#> 51 -0.6794383  0.016208426       0       0
#> 
#> [[2]]
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 51 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21 ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>  <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.974 -0.0299
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.766  0.0460
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.977 -0.0764
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.615 -0.0764
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.932  0.0342
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.724  0.0819
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.686 -0.186 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.504 -0.191 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.767 -0.0212
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.588 -0.0793
#> # … with 41 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 102 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21     ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.0940  18.9 0.041  0.12 24.0  453. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.0940  18.9 0.041  0.12  6.37 120. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.116   20.4 0.041  0.12 21.8  445. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.116   20.4 0.041  0.12  4.55  92.7
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.0937  18.0 0.041  0.12 25.1  453. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.0937  18.0 0.041  0.12  6.69 121. 
#> # … with 96 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>