Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> $eta
#>           ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1   -0.133286732  0.1764954106       0       0
#> 2   -0.053451054  0.0285385428       0       0
#> 3   -0.171923011  0.0718074174       0       0
#> 4   -0.196212734  0.1436619659       0       0
#> 5   -0.192653300 -0.0482667455       0       0
#> 6   -0.033691254  0.0423827318       0       0
#> 7   -0.032833997  0.2107458578       0       0
#> 8   -0.013667442  0.2489123712       0       0
#> 9   -0.050305335  0.1518013314       0       0
#> 10  -0.199952522  0.0519538009       0       0
#> 11  -0.129050697  0.0502281354       0       0
#> 12  -0.039839440  0.0502281354       0       0
#> 13  -0.205755990  0.0512868287       0       0
#> 14  -0.151002776  0.0099093179       0       0
#> 15  -0.168575175  0.0908389809       0       0
#> 16  -0.168575175  0.0908389809       0       0
#> 17  -0.166020389  0.1026108393       0       0
#> 18  -0.103835817  0.0658913257       0       0
#> 19  -0.139812637  0.0949386242       0       0
#> 20  -0.113322609 -0.0447104342       0       0
#> 21  -0.077224437  0.0006587715       0       0
#> 22  -0.188598831  0.0760204797       0       0
#> 23  -0.258973438  0.0408293098       0       0
#> 24   0.028073391  0.0407717259       0       0
#> 25  -0.017291551 -0.0409516030       0       0
#> 26  -0.112114604  0.0077871230       0       0
#> 27  -0.135145809  0.2263966360       0       0
#> 28  -0.060136968  0.0918778847       0       0
#> 29  -0.152338927  0.0151904257       0       0
#> 30  -0.229389870  0.0926136640       0       0
#> 31  -0.236143513  0.1025631977       0       0
#> 32  -0.218222446  0.1948121023       0       0
#> 33  -0.144211209  0.1306194897       0       0
#> 34  -0.306242441  0.0464062740       0       0
#> 35  -0.073323122  0.0479038042       0       0
#> 36  -0.073323122  0.0479038042       0       0
#> 37  -0.135691113  0.1106539883       0       0
#> 38  -0.274377999  0.2461807446       0       0
#> 39  -0.166177868  0.2713676826       0       0
#> 40  -0.152823933  0.1439235370       0       0
#> 41  -0.169204389 -0.0745389118       0       0
#> 42  -0.094800562  0.1010374611       0       0
#> 43  -0.094276236 -0.0481199319       0       0
#> 44  -0.315166760  0.2827343629       0       0
#> 45  -0.006250597 -0.0440187569       0       0
#> 46  -0.104444114  0.0582752677       0       0
#> 47  -0.104444114  0.0582752677       0       0
#> 48  -0.062984551 -0.0170284300       0       0
#> 49   0.025179027 -0.1453738303       0       0
#> 50  -0.073372860  0.1282848695       0       0
#> 51  -0.071546593  0.1464507661       0       0
#> 52  -0.108855463  0.0024290379       0       0
#> 53  -0.093212977  0.1165065330       0       0
#> 54  -0.022259005 -0.1047194890       0       0
#> 55  -0.044239849 -0.0396974703       0       0
#> 56  -0.185026036 -0.0794869456       0       0
#> 57  -0.182554480  0.1929855404       0       0
#> 58  -0.129999574  0.0358252249       0       0
#> 59  -0.129999574  0.0358252249       0       0
#> 60   0.059053001 -0.0897855351       0       0
#> 61  -0.102150083  0.1372334830       0       0
#> 62  -0.455023607  0.0653782024       0       0
#> 63  -0.203831461  0.2288699840       0       0
#> 64  -0.054135691  0.1673201905       0       0
#> 65  -0.216324600  0.0795013963       0       0
#> 66  -0.216324600  0.0795013963       0       0
#> 67  -0.032920936 -0.0323670471       0       0
#> 68  -0.227188417  0.0480335045       0       0
#> 69  -0.035630757 -0.0086226559       0       0
#> 70  -0.144983407  0.1583285667       0       0
#> 71  -0.157531012  0.2376915155       0       0
#> 72  -0.332721099  0.2073072464       0       0
#> 73  -0.145619175  0.2542773129       0       0
#> 74  -0.442353655  0.0629541301       0       0
#> 75  -0.089555822  0.2230677277       0       0
#> 76  -0.209968835  0.1292452425       0       0
#> 77  -0.107677082  0.0330585838       0       0
#> 78  -0.032488352  0.0300978913       0       0
#> 79  -0.137670401 -0.1487371687       0       0
#> 80  -0.034923844  0.0696949841       0       0
#> 81  -0.122398806  0.0998019186       0       0
#> 82   0.003134745  0.0125773250       0       0
#> 83  -0.155631870  0.2379140556       0       0
#> 84  -0.226648264  0.2223635644       0       0
#> 85  -0.155110710  0.1595457975       0       0
#> 86  -0.109532067  0.0759047661       0       0
#> 87  -0.073626271 -0.0657847104       0       0
#> 88  -0.073626271 -0.0657847104       0       0
#> 89  -0.102059334  0.1622787169       0       0
#> 90  -0.085345195  0.0435712946       0       0
#> 91  -0.125431487  0.0240981279       0       0
#> 92  -0.069597127  0.0240981279       0       0
#> 93  -0.062777338 -0.0524457988       0       0
#> 94  -0.141833689  0.0748578040       0       0
#> 95  -0.293223207  0.1974011480       0       0
#> 96  -0.174865809  0.1227730500       0       0
#> 97  -0.118586214  0.1279384433       0       0
#> 98  -0.094747261  0.1943115150       0       0
#> 99  -0.192223721  0.2092883214       0       0
#> 100 -0.259028307  0.1350657231       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.133   0.176 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.0535  0.0285
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.172   0.0718
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.196   0.144 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.193  -0.0483
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0337  0.0424
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0328  0.211 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0137  0.249 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0503  0.152 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.200   0.0520
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.218  23.2 0.041  0.12 71.1  1651. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.218  23.2 0.041  0.12  5.06  117. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.236  20.0 0.041  0.12 81.3  1628. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.236  20.0 0.041  0.12  4.86   97.3
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.210  20.9 0.041  0.12 79.4  1661. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.210  20.9 0.041  0.12  6.13  128. 
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>