Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains
Number of Markov chains

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)
#> $eta
#>            ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1   -0.1339842376 -0.0240080043       0       0
#> 2   -0.1339842376 -0.0240080043       0       0
#> 3   -0.0585281738  0.2368249248       0       0
#> 4   -0.1178905257  0.1765369364       0       0
#> 5   -0.0716090296  0.1035020995       0       0
#> 6   -0.1245689943  0.0691088742       0       0
#> 7   -0.0616196220 -0.0520088921       0       0
#> 8   -0.0616196220  0.0607565750       0       0
#> 9   -0.0662610815 -0.0319748076       0       0
#> 10  -0.1423072975  0.1131977775       0       0
#> 11  -0.1916138598  0.0179801727       0       0
#> 12  -0.0782578352  0.0845259746       0       0
#> 13  -0.0864397128 -0.0029797384       0       0
#> 14  -0.0864397128 -0.0029797384       0       0
#> 15  -0.1051436024  0.1778431791       0       0
#> 16  -0.0813339675  0.0936308776       0       0
#> 17  -0.1025954837  0.1227632373       0       0
#> 18  -0.2815552637  0.1823477925       0       0
#> 19  -0.0217575274  0.0211829503       0       0
#> 20  -0.2990798075  0.2344762741       0       0
#> 21  -0.0312882347 -0.1381478941       0       0
#> 22   0.1352234936  0.0537511638       0       0
#> 23  -0.1478449074  0.1025183278       0       0
#> 24  -0.0644902654 -0.0358812581       0       0
#> 25  -0.0463843604 -0.0438695147       0       0
#> 26  -0.0233289095  0.0136787944       0       0
#> 27  -0.1451871662  0.0390937850       0       0
#> 28  -0.1677174460  0.1881379546       0       0
#> 29  -0.1724192684  0.1585847489       0       0
#> 30  -0.0070421642 -0.0466577724       0       0
#> 31  -0.0609499926  0.0821683555       0       0
#> 32  -0.0713476505  0.1590719784       0       0
#> 33  -0.2235070023  0.1164991354       0       0
#> 34   0.0164255528  0.1648498572       0       0
#> 35  -0.1270626371  0.0540073237       0       0
#> 36  -0.1139748974  0.1959614787       0       0
#> 37  -0.0895403162  0.0677617298       0       0
#> 38  -0.1254654154  0.0454991234       0       0
#> 39  -0.1097148787 -0.0592738435       0       0
#> 40  -0.0876521886  0.0610037043       0       0
#> 41  -0.0043306637  0.2005297001       0       0
#> 42  -0.0043306637  0.0229046230       0       0
#> 43  -0.0653391757 -0.1698234278       0       0
#> 44  -0.2557745938 -0.0162138197       0       0
#> 45  -0.0460107148 -0.0359337985       0       0
#> 46  -0.1819264447  0.2031958731       0       0
#> 47  -0.2059728543  0.1054288613       0       0
#> 48  -0.0227899627  0.1620601113       0       0
#> 49  -0.0785262460  0.0531532761       0       0
#> 50  -0.0167783537  0.1420163046       0       0
#> 51   0.0590437604  0.0194050978       0       0
#> 52  -0.1580536691  0.0854044750       0       0
#> 53  -0.2161353247  0.1489761286       0       0
#> 54  -0.1780099983  0.2972714439       0       0
#> 55  -0.0863740822  0.0965100850       0       0
#> 56  -0.1520884574  0.1549122544       0       0
#> 57  -0.1071387770  0.2749835914       0       0
#> 58  -0.1053012075  0.1536633864       0       0
#> 59  -0.0802596404  0.1214093164       0       0
#> 60  -0.1019848194  0.0285118366       0       0
#> 61  -0.1343807226 -0.0086497276       0       0
#> 62  -0.1886518900  0.1172492164       0       0
#> 63  -0.1352540702  0.1165339063       0       0
#> 64  -0.0851314632 -0.0597476739       0       0
#> 65  -0.1103918659  0.1105720086       0       0
#> 66  -0.1298437717  0.2608663323       0       0
#> 67  -0.0845717385  0.0169503677       0       0
#> 68   0.0170929312 -0.1422359058       0       0
#> 69  -0.1795804409  0.0947107281       0       0
#> 70  -0.2003023087  0.1499253565       0       0
#> 71  -0.1243456485 -0.0184811493       0       0
#> 72  -0.0732181917  0.1190307083       0       0
#> 73  -0.1443994546  0.1933390316       0       0
#> 74  -0.0475885888 -0.1096225960       0       0
#> 75  -0.0987229629 -0.1274852876       0       0
#> 76  -0.0730914509 -0.1043259773       0       0
#> 77  -0.2014611085  0.0151632418       0       0
#> 78  -0.2014611085  0.0151632418       0       0
#> 79  -0.1629704903  0.1355410856       0       0
#> 80  -0.0101842219  0.2117489982       0       0
#> 81  -0.1585582656  0.0212950371       0       0
#> 82  -0.1213906210 -0.0392772974       0       0
#> 83  -0.0462115487 -0.0668262704       0       0
#> 84  -0.1499819156  0.1682671036       0       0
#> 85  -0.1769032376  0.1092324625       0       0
#> 86  -0.1231793455  0.2208021026       0       0
#> 87  -0.2145427346  0.1870740705       0       0
#> 88   0.0127733296 -0.0446390277       0       0
#> 89  -0.1743027534 -0.0256025850       0       0
#> 90  -0.1937577541  0.1002157563       0       0
#> 91  -0.0328619830  0.0100501088       0       0
#> 92  -0.0586588924 -0.0206003414       0       0
#> 93  -0.0451073788 -0.0203855966       0       0
#> 94  -0.0952715040  0.0716418618       0       0
#> 95  -0.1613730136  0.0696647185       0       0
#> 96  -0.0873755330 -0.1036405957       0       0
#> 97  -0.1987178155  0.0645958557       0       0
#> 98  -0.0628939707 -0.0293101247       0       0
#> 99  -0.2763179613  0.0895455259       0       0
#> 100 -0.0966211337  0.1173227644       0       0
#> 101 -0.1580536691  0.0854044750       0       0
#> 102 -0.1780099983  0.2972714439       0       0
#> 103 -0.1520884574  0.1549122544       0       0
#> 104 -0.1053012075  0.1536633864       0       0
#> 105 -0.1019848194  0.0285118366       0       0
#> 106 -0.1886518900  0.1172492164       0       0
#> 107 -0.0851314632 -0.0597476739       0       0
#> 108 -0.1298437717  0.2608663323       0       0
#> 109  0.0170929312 -0.1422359058       0       0
#> 110 -0.2003023087  0.1499253565       0       0
#> 111 -0.0732181917  0.1190307083       0       0
#> 112 -0.0475885888 -0.1096225960       0       0
#> 113 -0.0730914509 -0.1043259773       0       0
#> 114 -0.2014611085  0.0151632418       0       0
#> 115 -0.0101842219  0.2117489982       0       0
#> 116 -0.1213906210 -0.0392772974       0       0
#> 117 -0.1499819156  0.1682671036       0       0
#> 118 -0.1231793455  0.2208021026       0       0
#> 119  0.0127733296 -0.0446390277       0       0
#> 120 -0.1937577541  0.1002157563       0       0
#> 121 -0.0586588924 -0.0206003414       0       0
#> 122 -0.0952715040  0.0716418618       0       0
#> 123 -0.0873755330 -0.1036405957       0       0
#> 124 -0.0628939707 -0.0293101247       0       0
#> 125 -0.0966211337  0.1173227644       0       0
#> 126  0.0000000000  0.0000000000       0       0
#> 127 -0.0296424592  0.0521056523       0       0
#> 128 -0.1342186293 -0.0258455850       0       0
#> 129 -0.1088745619  0.1287116157       0       0
#> 130 -0.0953692663  0.0064467345       0       0
#> 131 -0.1030734978  0.0014860337       0       0
#> 132 -0.0119222870 -0.0492118850       0       0
#> 133 -0.2679291473  0.1054656211       0       0
#> 134 -0.0664757932  0.1473918184       0       0
#> 135 -0.0291650757  0.0384192838       0       0
#> 136 -0.2959733059  0.0369026904       0       0
#> 137 -0.0889209885  0.0080674721       0       0
#> 138 -0.1003629358 -0.0118667064       0       0
#> 139 -0.0129047769 -0.0413836915       0       0
#> 140 -0.2075256190  0.0273080746       0       0
#> 141 -0.0428056413  0.0925208295       0       0
#> 142  0.0149231215 -0.0855381576       0       0
#> 143 -0.0502422583  0.1298952956       0       0
#> 144  0.0480846997  0.0173536381       0       0
#> 145 -0.1800968628  0.1504910359       0       0
#> 146 -0.0911650394 -0.0283975066       0       0
#> 147 -0.1482534565  0.1982192899       0       0
#> 148 -0.1063051217  0.2189986579       0       0
#> 149 -0.2313854107  0.1438520503       0       0
#> 150 -0.1394676234  0.1824376332       0       0
#> 151 -0.1112660765  0.1084484258       0       0
#> 152 -0.1390971951  0.2039193293       0       0
#> 153 -0.0447032562  0.1962938279       0       0
#> 154 -0.1709059809  0.1016839243       0       0
#> 155 -0.1692501300  0.1504821653       0       0
#> 156 -0.1249564431  0.1209860772       0       0
#> 157 -0.0844683982  0.1582890989       0       0
#> 158 -0.1715885198  0.2957256445       0       0
#> 159 -0.0656301352 -0.1544050612       0       0
#> 160 -0.2633145260 -0.0355648621       0       0
#> 161 -0.0471129181  0.0425799596       0       0
#> 162 -0.0140137129  0.1248664232       0       0
#> 163 -0.3080492330 -0.0939264768       0       0
#> 164 -0.1382591358  0.2486139490       0       0
#> 165 -0.1261504947  0.1666221159       0       0
#> 166 -0.0564732871  0.0133935871       0       0
#> 167 -0.1303727674  0.0876975224       0       0
#> 168 -0.2220281148  0.2055487178       0       0
#> 169 -0.1699399120  0.0984500638       0       0
#> 170 -0.0489949923  0.1040227632       0       0
#> 171 -0.0778221530  0.1587258183       0       0
#> 172 -0.0739877990  0.0829425240       0       0
#> 173 -0.3513938313  0.1328251574       0       0
#> 174  0.0379416821  0.0727381394       0       0
#> 175 -0.1705966528  0.1369492451       0       0
#> 176 -0.1156804276  0.0852589574       0       0
#> 177  0.0864987169 -0.0970121734       0       0
#> 178 -0.1736972762  0.2012617579       0       0
#> 179 -0.1909990422  0.1658665118       0       0
#> 180 -0.1128288713  0.2217197084       0       0
#> 181 -0.0691751248  0.0786930784       0       0
#> 182 -0.0378581924  0.0388996978       0       0
#> 183 -0.1382232666  0.2044950511       0       0
#> 184 -0.2031666059  0.2024592385       0       0
#> 185 -0.1571802144  0.1040325669       0       0
#> 186 -0.2170589413 -0.1311912920       0       0
#> 187 -0.0853559097 -0.0114611011       0       0
#> 188 -0.1907114860 -0.0586702364       0       0
#> 189 -0.0337326484  0.0830594006       0       0
#> 190 -0.0005160908  0.1832537026       0       0
#> 191 -0.1964577385  0.0918949353       0       0
#> 192 -0.0574230908  0.0342584954       0       0
#> 193 -0.1043610073  0.2420415603       0       0
#> 194 -0.0685976682  0.0243994597       0       0
#> 195 -0.0593047944  0.0911189762       0       0
#> 196 -0.0738205543  0.0509652375       0       0
#> 197 -0.1885382471  0.0906015641       0       0
#> 198 -0.1658126987  0.1224463251       0       0
#> 199 -0.2187864100  0.1825328378       0       0
#> 200 -0.1192820106  0.0955464603       0       0
#> 201 -0.0432211143 -0.1741533313       0       0
#> 202 -0.1342186293 -0.0258455850       0       0
#> 203 -0.2333400279  0.0223363115       0       0
#> 204 -0.1030734978  0.0014860337       0       0
#> 205  0.0083754436 -0.0504434025       0       0
#> 206 -0.0664757932  0.1473918184       0       0
#> 207 -0.1310247982  0.0492506075       0       0
#> 208 -0.0889209885  0.0080674721       0       0
#> 209  0.0179933581 -0.0013907640       0       0
#> 210 -0.2075256190  0.0273080746       0       0
#> 211 -0.0901631696  0.0703248872       0       0
#> 212 -0.0502422583  0.1298952956       0       0
#> 213 -0.1374840341  0.1226970987       0       0
#> 214 -0.0911650394 -0.0283975066       0       0
#> 215 -0.0409759903  0.0932890623       0       0
#> 216 -0.2313854107  0.1438520503       0       0
#> 217 -0.1780220831  0.1389497349       0       0
#> 218 -0.1390971951  0.2039193293       0       0
#> 219 -0.0859838196  0.1037937445       0       0
#> 220 -0.1692501300  0.1504821653       0       0
#> 221 -0.1685398838  0.2716031593       0       0
#> 222 -0.1715885198  0.2957256445       0       0
#> 223 -0.0972812361  0.0700864391       0       0
#> 224 -0.0471129181  0.0425799596       0       0
#> 225 -0.0055301910  0.1248664232       0       0
#> 226 -0.1382591358  0.2486139490       0       0
#> 227 -0.1660939214  0.1347783864       0       0
#> 228 -0.1303727674  0.0876975224       0       0
#> 229 -0.1417750404  0.2055487178       0       0
#> 230 -0.0489949923  0.1040227632       0       0
#> 231 -0.0486283811  0.0677277714       0       0
#> 232 -0.3513938313  0.1328251574       0       0
#> 233 -0.0061474947  0.0878291287       0       0
#> 234 -0.1156804276  0.0852589574       0       0
#> 235 -0.1374876704  0.1913997326       0       0
#> 236 -0.1909990422  0.1658665118       0       0
#> 237 -0.0568238736  0.1349721455       0       0
#> 238 -0.0378581924  0.0388996978       0       0
#> 239 -0.1648656580  0.2044950511       0       0
#> 240 -0.1571802144  0.1040325669       0       0
#> 241 -0.0708029822 -0.0056359173       0       0
#> 242 -0.1907114860 -0.0586702364       0       0
#> 243 -0.0030151737  0.0512450965       0       0
#> 244 -0.1964577385  0.0918949353       0       0
#> 245 -0.1613290621 -0.0464968365       0       0
#> 246 -0.0685976682  0.0243994597       0       0
#> 247 -0.1773801003  0.1522616736       0       0
#> 248 -0.1885382471  0.0906015641       0       0
#> 249 -0.1671185714  0.1144418851       0       0
#> 250 -0.1192820106  0.0955464603       0       0
#> 251  0.0000000000  0.0000000000       0       0
#> 252  0.0154691587  0.0277387536       0       0
#> 253 -0.0812530547  0.0649308872       0       0
#> 254 -0.1055439828  0.0468133621       0       0
#> 255 -0.0809369592 -0.0386576320       0       0
#> 256  0.0073726417 -0.1168985519       0       0
#> 257 -0.2363956673  0.1393350532       0       0
#> 258 -0.0577965280  0.0039990824       0       0
#> 259 -0.3959608143  0.1087004143       0       0
#> 260 -0.1621174138  0.1027818602       0       0
#> 261 -0.1932521168  0.1665438679       0       0
#> 262 -0.1043022711  0.1770697683       0       0
#> 263 -0.0919261387 -0.0258548022       0       0
#> 264 -0.2357348028  0.1363820673       0       0
#> 265 -0.1145893533  0.0620935502       0       0
#> 266 -0.4032603768  0.0473198206       0       0
#> 267 -0.0817183716  0.0423737790       0       0
#> 268 -0.2532836617 -0.0123251750       0       0
#> 269 -0.0991357342  0.0568143608       0       0
#> 270 -0.1395884765 -0.0711519879       0       0
#> 271 -0.0918348222  0.0519602638       0       0
#> 272 -0.1527643093  0.1589616917       0       0
#> 273 -0.0074618176  0.1863680870       0       0
#> 274 -0.1154236824  0.0031835631       0       0
#> 275 -0.1788975791  0.1687101862       0       0
#> 276 -0.1671174428 -0.0434625733       0       0
#> 277 -0.0968150775  0.0716098842       0       0
#> 278 -0.0843042946  0.0377786790       0       0
#> 279 -0.3090148191 -0.1793631557       0       0
#> 280 -0.1307852194  0.0852151104       0       0
#> 281  0.0477765115 -0.0969090778       0       0
#> 282 -0.0521089616  0.2271655405       0       0
#> 283 -0.3143745750  0.1797056981       0       0
#> 284  0.0068076974  0.0519503662       0       0
#> 285  0.1191796414 -0.0662989312       0       0
#> 286 -0.1016142694  0.0658053734       0       0
#> 287  0.0332370046 -0.0630132064       0       0
#> 288  0.0110548558 -0.0415690636       0       0
#> 289 -0.1337752154  0.0507821244       0       0
#> 290 -0.0313224715 -0.1479045855       0       0
#> 291 -0.2472250257  0.1729007654       0       0
#> 292 -0.1050180009  0.1872384185       0       0
#> 293 -0.1890254108 -0.0095187949       0       0
#> 294 -0.0858140840  0.0670972939       0       0
#> 295 -0.2410449562 -0.1463867587       0       0
#> 296 -0.2774717737  0.0033952273       0       0
#> 297 -0.0251516178  0.0806280498       0       0
#> 298 -0.0844244913  0.1431100961       0       0
#> 299 -0.0659902477 -0.0001592934       0       0
#> 300 -0.1941741500  0.1951496374       0       0
#> 301 -0.1390971951  0.2039193293       0       0
#> 302 -0.1709059809  0.1016839243       0       0
#> 303 -0.1249564431  0.1209860772       0       0
#> 304 -0.1715885198  0.2957256445       0       0
#> 305 -0.2633145260 -0.0355648621       0       0
#> 306 -0.0140137129  0.1248664232       0       0
#> 307 -0.1382591358  0.2486139490       0       0
#> 308 -0.0564732871  0.0133935871       0       0
#> 309 -0.2220281148  0.2055487178       0       0
#> 310 -0.0489949923  0.1040227632       0       0
#> 311 -0.0739877990  0.0829425240       0       0
#> 312  0.0379416821  0.0727381394       0       0
#> 313 -0.1156804276  0.0852589574       0       0
#> 314 -0.1736972762  0.2012617579       0       0
#> 315 -0.1128288713  0.2217197084       0       0
#> 316 -0.0378581924  0.0388996978       0       0
#> 317 -0.2031666059  0.2024592385       0       0
#> 318 -0.2170589413 -0.1311912920       0       0
#> 319 -0.1907114860 -0.0586702364       0       0
#> 320 -0.0005160908  0.1832537026       0       0
#> 321 -0.0574230908  0.0342584954       0       0
#> 322 -0.0685976682  0.0243994597       0       0
#> 323 -0.0738205543  0.0509652375       0       0
#> 324 -0.1658126987  0.1224463251       0       0
#> 325 -0.1192820106  0.0955464603       0       0
#> 326 -0.0255631215  0.0656884443       0       0
#> 327 -0.0475500739  0.0320198183       0       0
#> 328 -0.1055439828  0.0468133621       0       0
#> 329 -0.2847874912 -0.0475892512       0       0
#> 330 -0.1768556081  0.1773275335       0       0
#> 331 -0.0577965280  0.0039990824       0       0
#> 332 -0.1011293733  0.2099344940       0       0
#> 333 -0.0140025300 -0.0368351905       0       0
#> 334 -0.1043022711  0.1770697683       0       0
#> 335 -0.1203055575  0.0640972919       0       0
#> 336 -0.1172094280  0.1101288825       0       0
#> 337 -0.4032603768  0.0473198206       0       0
#> 338 -0.1270949336  0.0405607128       0       0
#> 339 -0.0399218310 -0.1357024401       0       0
#> 340 -0.1395884765 -0.0711519879       0       0
#> 341 -0.0504915694  0.0858336510       0       0
#> 342 -0.2049822694  0.2179075033       0       0
#> 343 -0.1154236824  0.0031835631       0       0
#> 344 -0.1087693786  0.1195305137       0       0
#> 345 -0.0310606693 -0.0050615552       0       0
#> 346 -0.0843042946  0.0377786790       0       0
#> 347 -0.1785208193  0.0796425321       0       0
#> 348 -0.0927019754  0.0720783157       0       0
#> 349 -0.0521089616  0.2271655405       0       0
#> 350 -0.0763759420  0.0478467407       0       0
#> 351 -0.1840101149 -0.0296689357       0       0
#> 352 -0.1016142694  0.0658053734       0       0
#> 353  0.1099523912 -0.0980152535       0       0
#> 354 -0.1574161374  0.3194511225       0       0
#> 355 -0.0313224715 -0.1479045855       0       0
#> 356 -0.0366911548  0.0360191581       0       0
#> 357  0.0288975784 -0.1490050697       0       0
#> 358 -0.0858140840  0.0670972939       0       0
#> 359 -0.0284680180  0.0975978472       0       0
#> 360 -0.1632375230  0.0842311562       0       0
#> 361 -0.0844244913  0.1431100961       0       0
#> 362 -0.0801999930 -0.0779680643       0       0
#> 363 -0.1463256338  0.0481223602       0       0
#> 364  0.0786061071 -0.0995790288       0       0
#> 365 -0.1760498122  0.2113679038       0       0
#> 366 -0.0752142003  0.1408952764       0       0
#> 367 -0.0478046076  0.1915626918       0       0
#> 368 -0.1868579781  0.2339749008       0       0
#> 369 -0.0780740565  0.0995527963       0       0
#> 370  0.0713241844 -0.1822148429       0       0
#> 371 -0.0621800559  0.1614892100       0       0
#> 372 -0.0779839929  0.0093092455       0       0
#> 373 -0.0170864051 -0.0091218439       0       0
#> 374 -0.0648306287  0.3110951494       0       0
#> 375 -0.0828272679  0.1709771690       0       0
#> 376 -0.1282608118  0.2401701471       0       0
#> 377 -0.1114297449  0.2200945479       0       0
#> 378 -0.1498893506  0.1505438959       0       0
#> 379 -0.1709977074  0.1219712990       0       0
#> 380 -0.0489058784  0.0775182416       0       0
#> 381  0.0070880433  0.1314675547       0       0
#> 382 -0.1946235917  0.0867387486       0       0
#> 383 -0.2097290458  0.1789626597       0       0
#> 384  0.0240656567 -0.0234584742       0       0
#> 385 -0.0927070320 -0.0120963419       0       0
#> 386 -0.0428307323 -0.1025342429       0       0
#> 387 -0.2011378976 -0.0370917899       0       0
#> 388 -0.2029157058  0.1266477443       0       0
#> 389 -0.2258002808 -0.0105604332       0       0
#> 390 -0.0909630652  0.2286227296       0       0
#> 391 -0.1871406052  0.0404112281       0       0
#> 392 -0.1169533010  0.1523508726       0       0
#> 393 -0.0967884191  0.0924424243       0       0
#> 394 -0.1547325705  0.0281438730       0       0
#> 395 -0.0481173514  0.1177147292       0       0
#> 396 -0.0794334754  0.0572356192       0       0
#> 397 -0.2128809235  0.0135138155       0       0
#> 398 -0.1551166631  0.0686672008       0       0
#> 399 -0.1897370028  0.0547116765       0       0
#> 400 -0.1810620664  0.1481843199       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 400 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.134  -0.0240
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.134  -0.0240
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.0585  0.237 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.118   0.177 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.0716  0.104 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.125   0.0691
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0616 -0.0520
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0616  0.0608
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0663 -0.0320
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.142   0.113 
#> # … with 390 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 800 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.218  19.0 0.041  0.12 86.9  1651. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.218  19.0 0.041  0.12  6.19  118. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.218  19.0 0.041  0.12 86.9  1651. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.218  19.0 0.041  0.12  6.19  118. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.235  24.7 0.041  0.12 66.1  1630. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.235  24.7 0.041  0.12  3.99   98.5
#> # … with 794 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>