Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>           ETA_Cl        ETA_Vc       ETA_Ka
#> 1    0.646021282 -0.4724015023  0.473063075
#> 2    0.378487565 -0.6967355621  0.393183154
#> 3    0.475003297 -0.5052637311  0.078008674
#> 4    0.634131105 -0.0864150588  0.376553950
#> 5    0.797764248 -0.2364824264  0.572000140
#> 6    0.396556745 -0.6195718407  0.024599628
#> 7    0.579987568 -0.4929216103  0.084219806
#> 8    0.608420444 -0.4662602129  0.808175282
#> 9    0.007836430 -0.0510104582  0.279352859
#> 10   0.804925237 -0.4167468602  0.484541994
#> 11  -0.005682223  0.3539826250  0.810135361
#> 12   0.579626245 -0.4020341400  0.891088281
#> 13   0.537333747 -0.2609807576  0.146013004
#> 14   0.730754914 -0.3331394345  0.228059899
#> 15   0.575929584 -0.7825765801 -0.378554607
#> 16  -0.065402061 -0.3189258445 -0.404156524
#> 17   0.354457096 -1.1332973412 -0.540783899
#> 18   0.153901067 -1.2179942544 -0.464008710
#> 19   0.646803698 -0.6084269750 -0.352891511
#> 20   0.166156551 -0.3991253040  0.027592629
#> 21   0.700261947 -0.0918039509  0.665031487
#> 22   0.729849306 -0.1684943960  0.437840144
#> 23   0.737707158 -0.1849317861  0.687968200
#> 24   0.348811869 -0.0155364177  0.394521906
#> 25   0.582304443 -0.5979027840  0.738205251
#> 26   0.736982387 -0.2563820931  0.056355627
#> 27   0.085262635 -0.6214524868  0.110589592
#> 28   0.474588128 -0.5411578645  0.070581848
#> 29   0.447599394 -0.5559358172  0.290528254
#> 30   0.517237632 -0.4836616389  0.784676723
#> 31   0.360833308 -0.5708436754  0.744348291
#> 32   0.214714120 -0.2020638562  0.722475447
#> 33   0.667308992 -0.3688887230  0.349613872
#> 34   0.682795245 -0.2596972463  0.209526144
#> 35   0.339340337 -0.4262521738  0.240692938
#> 36   0.414266780 -1.3774922423 -0.912409121
#> 37   0.802768079 -0.0949812169  0.400847295
#> 38   0.769826460 -0.1888449682  0.573288991
#> 39   0.643392351 -0.0700033961  0.726572265
#> 40   0.876844278 -0.1447780322  0.852961790
#> 41   0.275640640 -0.2730771135  0.542602318
#> 42   0.461257587 -0.6288929777 -0.136077863
#> 43   0.492902054 -0.6206455235 -0.154104409
#> 44   0.290239448 -0.9145544611 -0.280529575
#> 45   0.158101008 -0.3564639053  0.109355824
#> 46   0.695101561 -0.2280786538  0.302917168
#> 47   0.573290786 -0.6247377522 -0.431134461
#> 48   0.551360533 -0.4753500891  0.073888346
#> 49   0.558087947 -0.2417935044  0.242492021
#> 50   0.000791646 -0.5583404238  0.181877158
#> 51   0.000000000  0.0000000000  0.000000000
#> 52   0.277918971 -0.2706687496  0.257456083
#> 53   0.621224167 -0.1925623381  0.591810076
#> 54   0.580867435 -0.4381389208  0.792439247
#> 55   0.653064381 -0.3382776473  0.021558065
#> 56   0.805530769  0.0654004101  1.005799593
#> 57   0.643025297 -0.3156295552  0.335793771
#> 58   0.438097256 -0.3604453967  0.052894407
#> 59   0.430024275 -0.9970191500 -0.132616493
#> 60   0.422350222 -0.6809052142  0.318325951
#> 61   0.344770746 -1.1706841683 -0.294251535
#> 62   0.336477826 -0.8090337556 -0.285254090
#> 63   0.303790274 -0.6311926116  0.139234724
#> 64   0.444555829 -0.3790527338  0.777742012
#> 65   0.313294462 -0.5300702473  0.079667834
#> 66   0.632792635 -0.0003166909  0.908040726
#> 67   0.303414693 -0.7444105067 -0.516781512
#> 68   0.233394267 -0.6288221643 -0.200971914
#> 69   0.349356138 -0.7347891023  0.053512487
#> 70   0.770694101 -0.4303131061  0.213296896
#> 71   0.286903915 -0.7088464065 -0.481942914
#> 72   0.451673573 -0.4252449946 -0.328929941
#> 73   0.727283033 -0.2677692868  0.173534562
#> 74   0.531442729 -0.3771398693 -0.002644498
#> 75   0.592025346  0.2278992130  0.775068009
#> 76   0.572278658 -0.4079719723 -0.001835657
#> 77   0.615584179 -0.2477442022  0.188222448
#> 78   0.485352672 -0.1308976545  0.002464575
#> 79   0.503608362 -0.7295507414 -0.613991785
#> 80   0.344032436 -0.7736610506 -0.022794295
#> 81   0.663780753 -0.2419387235 -0.030525928
#> 82   0.667902482 -0.1363015450  0.186136440
#> 83   0.253901537 -0.6834447593  0.108977389
#> 84   0.505991485 -0.5203064153  0.031846045
#> 85   0.703445351 -0.3973243534 -0.130499002
#> 86   0.273890284 -0.5891078199 -0.100458282
#> 87  -0.072741851  0.1802530549  1.186648765
#> 88   0.689881847 -0.0780486637  0.843484628
#> 89   0.482370867 -0.3044304029  0.225308590
#> 90   0.340273811 -0.9093371694 -0.268262597
#> 91   0.631008693 -0.5013894932  0.300204324
#> 92   0.647966260 -0.3341376762  0.394508712
#> 93   0.398009432 -0.7332264514 -0.189871991
#> 94   0.576504350 -0.0541545964  0.367447852
#> 95   0.794058091 -0.3166032867  0.340036803
#> 96   0.375297599 -0.5462481264 -0.002508713
#> 97   0.381875489 -0.1697845499  0.025610384
#> 98   0.262389820 -0.1961849717 -0.091324823
#> 99   0.543383893 -0.3562377100  0.434667262
#> 100  0.698206036 -0.2195123382  0.489633993
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1 0.646   -0.472  0.473 
#>  2      2        4       70        1 0.378   -0.697  0.393 
#>  3      3        4       70        1 0.475   -0.505  0.0780
#>  4      4        4       70        1 0.634   -0.0864 0.377 
#>  5      5        4       70        1 0.798   -0.236  0.572 
#>  6      6        4       70        1 0.397   -0.620  0.0246
#>  7      7        4       70        1 0.580   -0.493  0.0842
#>  8      8        4       70        1 0.608   -0.466  0.808 
#>  9      9        4       70        1 0.00784 -0.0510 0.279 
#> 10     10        4       70        1 0.805   -0.417  0.485 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  7.63  43.6  1.60 0.175 29.3      6.16e+2 1278.
#> 2      1    14     4    70     1  7.63  43.6  1.60 0.175  4.65     5.35e-7  203.
#> 3      2     1     4    70     1  5.84  34.9  1.48 0.167 35.3      6.73e+2 1230.
#> 4      2    14     4    70     1  5.84  34.9  1.48 0.167  6.47     2.90e-6  226.
#> 5      3     1     4    70     1  6.43  42.2  1.08 0.152 24.4      9.00e+2 1030.
#> 6      3    14     4    70     1  6.43  42.2  1.08 0.152  6.79     7.08e-4  287.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>