Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, , a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> $eta
#>         ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.7242713 -0.107637143       0       0
#> 2   -0.8278383  0.027342718       0       0
#> 3   -0.7477279 -0.145526228       0       0
#> 4   -0.8673390 -0.100988596       0       0
#> 5   -0.5360899 -0.041391626       0       0
#> 6   -0.7115416 -0.093147864       0       0
#> 7   -0.6089565  0.003010461       0       0
#> 8   -0.6461666  0.003010461       0       0
#> 9   -0.7794456 -0.155428303       0       0
#> 10  -0.5305989 -0.236164013       0       0
#> 11  -0.7666045 -0.106852469       0       0
#> 12  -0.7504021 -0.118973712       0       0
#> 13  -0.6392790 -0.112120762       0       0
#> 14  -0.6392790 -0.160045292       0       0
#> 15  -0.6239354 -0.022013115       0       0
#> 16  -0.6490068 -0.138497021       0       0
#> 17  -0.5479493 -0.094788303       0       0
#> 18  -0.5060595 -0.289458184       0       0
#> 19  -0.5781066 -0.144083522       0       0
#> 20  -0.5416755 -0.147758666       0       0
#> 21  -0.7853289 -0.129526762       0       0
#> 22  -0.6011904 -0.017831843       0       0
#> 23  -0.7011763 -0.130980831       0       0
#> 24  -0.8513197 -0.103271579       0       0
#> 25  -0.6985567 -0.129746386       0       0
#> 26  -0.7333925  0.018055580       0       0
#> 27  -0.7874712 -0.015601429       0       0
#> 28  -0.5730317 -0.038079022       0       0
#> 29  -0.6085487 -0.038079022       0       0
#> 30  -0.6085487 -0.038079022       0       0
#> 31  -0.5401041 -0.056602280       0       0
#> 32  -0.7941336 -0.098529169       0       0
#> 33  -0.6075635 -0.074927430       0       0
#> 34  -0.6916945 -0.074927430       0       0
#> 35  -0.6006121 -0.124736540       0       0
#> 36  -0.8181305 -0.092353358       0       0
#> 37  -0.7357908 -0.048047580       0       0
#> 38  -0.7131440 -0.048047580       0       0
#> 39  -0.5587179  0.017494463       0       0
#> 40  -0.6445806 -0.007243311       0       0
#> 41  -0.6759268 -0.070326987       0       0
#> 42  -0.7131980 -0.125300759       0       0
#> 43  -0.5265338 -0.179995942       0       0
#> 44  -0.7946921  0.057031372       0       0
#> 45  -0.7336823 -0.112187192       0       0
#> 46  -0.6824082 -0.113872529       0       0
#> 47  -0.7549182 -0.026701430       0       0
#> 48  -0.6487215  0.064519536       0       0
#> 49  -0.6305436 -0.069061343       0       0
#> 50  -0.5806598 -0.065098690       0       0
#> 51  -0.8068941  0.018801259       0       0
#> 52  -0.5930030  0.020284310       0       0
#> 53  -0.6866284 -0.153926898       0       0
#> 54  -0.8054235 -0.065939697       0       0
#> 55  -0.5932835 -0.065939697       0       0
#> 56  -0.6027725  0.035381586       0       0
#> 57  -0.7302901 -0.115949401       0       0
#> 58  -0.8569987  0.002511304       0       0
#> 59  -0.7285732  0.015036903       0       0
#> 60  -0.7901761  0.029906421       0       0
#> 61  -0.6035795 -0.101776835       0       0
#> 62  -0.8208260 -0.128572446       0       0
#> 63  -0.6210468 -0.128572446       0       0
#> 64  -0.6370756 -0.160902388       0       0
#> 65  -0.7882230  0.009737403       0       0
#> 66  -0.6453699  0.062509292       0       0
#> 67  -0.6807815 -0.119104382       0       0
#> 68  -0.5464620 -0.237656019       0       0
#> 69  -0.5372703 -0.215961445       0       0
#> 70  -0.5372703 -0.337810564       0       0
#> 71  -0.6580490 -0.144931928       0       0
#> 72  -0.7357714 -0.016529306       0       0
#> 73  -0.6654675 -0.087792301       0       0
#> 74  -0.6908835 -0.087792301       0       0
#> 75  -0.5758215 -0.087792301       0       0
#> 76  -0.6162495 -0.035675413       0       0
#> 77  -0.6915594 -0.204930764       0       0
#> 78  -0.6915594 -0.018808807       0       0
#> 79  -0.6585006  0.019949111       0       0
#> 80  -0.7906132  0.044468357       0       0
#> 81  -0.6380662 -0.025123061       0       0
#> 82  -0.6681001 -0.037265701       0       0
#> 83  -0.6681001 -0.037265701       0       0
#> 84  -0.6112908 -0.037265701       0       0
#> 85  -0.6112908 -0.037265701       0       0
#> 86  -0.6500896 -0.021289366       0       0
#> 87  -0.8097439 -0.069891823       0       0
#> 88  -0.6486687 -0.106619534       0       0
#> 89  -0.6958658 -0.142765181       0       0
#> 90  -0.5771566 -0.142765181       0       0
#> 91  -0.7526130  0.008605212       0       0
#> 92  -0.7656118  0.025488930       0       0
#> 93  -0.6283229 -0.025790827       0       0
#> 94  -0.5798611 -0.052624532       0       0
#> 95  -0.8096747 -0.130851583       0       0
#> 96  -0.8166673  0.126937396       0       0
#> 97  -0.7291972 -0.042997033       0       0
#> 98  -0.4579325 -0.149896177       0       0
#> 99  -0.4568377 -0.223748430       0       0
#> 100 -0.6511197 -0.211784073       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21 ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>  <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.724 -0.108  
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.828  0.0273 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.748 -0.146  
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.867 -0.101  
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.536 -0.0414 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.712 -0.0931 
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.609  0.00301
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.646  0.00301
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.779 -0.155  
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.531 -0.236  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.121  17.5 0.041  0.12 25.4  444. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.121  17.5 0.041  0.12  5.00  87.4
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.109  20.0 0.041  0.12 22.4  448. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.109  20.0 0.041  0.12  5.03 101. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.118  16.8 0.041  0.12 26.4  445. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.118  16.8 0.041  0.12  5.37  90.4
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>