Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.462039849 -0.310156609  0.365557204
#> 2    0.566222669 -0.211596202  0.721770219
#> 3    0.271193605 -0.066183273  0.397411308
#> 4    0.218067226 -0.978433123 -0.130568590
#> 5    0.698635991 -0.466866312 -0.092581868
#> 6    0.770130685 -0.602319587  0.293671763
#> 7    0.529512928 -0.557567038 -0.156822370
#> 8    0.682665083 -0.413377712  0.267487374
#> 9    0.766482068 -0.361097016 -0.072685037
#> 10   0.481795995 -0.277508294 -0.087563829
#> 11   0.398170628 -0.634869025 -0.074846367
#> 12   0.769477638 -0.503950596 -0.179345945
#> 13  -0.002831251 -0.379985754  0.135416058
#> 14   0.568026181 -0.144729793  0.222277424
#> 15   0.717399239 -0.174673627  0.229789053
#> 16   0.745633282 -0.352406167  0.229789053
#> 17   0.486143699 -0.469886462 -0.118160523
#> 18   0.486143699 -0.395568746 -0.012031306
#> 19   0.663196976 -0.425232366 -0.012031306
#> 20   0.448188202 -0.328562039  0.139339133
#> 21   0.663319552 -0.057395503  0.484218441
#> 22   0.257981633 -1.456825748 -0.017266727
#> 23   0.758442007  0.005174785  0.269716753
#> 24   0.672997276 -0.044756826  0.567777658
#> 25   0.709657320 -0.151278115  0.369785738
#> 26   0.393722582 -0.104424335  0.543095714
#> 27   0.687046191 -0.309553543  0.158837120
#> 28   0.333588582 -0.704703478  0.065056264
#> 29   0.777380573 -0.285602774  0.737952760
#> 30   0.488803345 -0.122452716  0.762464532
#> 31   0.507577963 -0.290045323  0.067587156
#> 32   0.556946192 -0.391059994  0.198289708
#> 33   0.595329405 -0.020866384  0.632988706
#> 34   0.757081119 -0.203101901  0.391361604
#> 35   0.771881843 -0.201126935  0.168972034
#> 36   0.518000858 -0.751717352 -0.018816111
#> 37   0.508232083 -0.878659297 -0.452884895
#> 38   0.388566163 -0.539810617  0.236352000
#> 39   0.496730242 -0.506918334 -0.156565343
#> 40   0.238499821 -0.984461682  0.021667438
#> 41   0.405667578 -0.771134617 -0.308838722
#> 42   0.468103075 -0.578104151  0.135716305
#> 43   0.500348971 -0.256558141 -0.063438209
#> 44   0.507846506 -0.527752663  0.543061325
#> 45   0.309692923 -0.935847448  0.541115590
#> 46   0.248755062 -0.394010041  0.474026059
#> 47   0.061339901 -0.786322341  0.153460887
#> 48   0.484786239 -0.812661879 -0.115372607
#> 49   0.274323986 -0.552326565 -0.210621827
#> 50   0.397043839 -0.497612787  0.612019836
#> 51   0.000000000  0.000000000  0.000000000
#> 52   0.450708308 -0.559030710  0.227646405
#> 53   0.455193290 -0.355979262  0.390414205
#> 54   0.514985791 -0.801214004 -0.461886752
#> 55   0.869183132 -0.237929567  0.696031815
#> 56   0.404772645 -0.106527131  0.539769867
#> 57   0.790741858  0.018604421  0.741385555
#> 58   0.857566351  0.053911854  0.392730399
#> 59   0.527238692 -0.562928254  0.401175767
#> 60   0.829257146 -0.230504712  0.618959991
#> 61   0.716999992 -0.295122242  0.810662658
#> 62   0.243415189  0.192394632  0.843186051
#> 63   0.753132456 -0.214507377  0.327346234
#> 64   0.563719520 -0.265455087  0.558921756
#> 65   0.366984792 -0.659975695  0.081170539
#> 66   0.237392703 -0.502462417  0.296247762
#> 67   0.597457081 -0.421138675  0.140530753
#> 68   0.408913132 -0.752264776 -0.149765608
#> 69   0.615143450 -0.301884398  0.213840366
#> 70   0.413377187 -0.749511880 -0.146990710
#> 71   0.822432624 -0.318379796 -0.003545124
#> 72   0.480197782 -0.379509612  0.201008754
#> 73   0.715578103 -0.348013426  0.615844633
#> 74   0.596626187 -0.180968002  0.476115044
#> 75   0.646690903 -0.279006842  0.608039721
#> 76   0.574451432 -0.071257454  0.393090054
#> 77   0.636696293 -0.221790501  0.360798212
#> 78   0.527223230 -0.756289776  0.007486144
#> 79   0.536094427 -0.407850711  0.528503610
#> 80   0.385790907 -0.700792287 -0.048581253
#> 81   0.566080782 -0.834564585 -0.517097658
#> 82   0.705334012 -0.126579533  0.708944308
#> 83   0.407893230  0.022629156  0.698392285
#> 84   0.578859928 -0.360093151  0.631554259
#> 85   0.543682595 -0.133994131  0.564413498
#> 86   0.553918002 -0.264942753  0.692284864
#> 87   0.593799053 -0.184818534  0.173544596
#> 88  -0.072467830  0.073137228  0.609499042
#> 89   0.696943467 -0.233933072  0.143950035
#> 90   0.591038942 -0.278851614  0.008580279
#> 91   0.359704973 -0.225112393  0.779201310
#> 92   0.780256621 -0.395464923  0.223281739
#> 93   0.354742081 -0.379799026  0.515941739
#> 94   0.351957332 -0.907839209 -0.492628787
#> 95   0.598503039 -0.373803648  0.481316373
#> 96   0.654138895 -0.223731559  0.232157083
#> 97   0.558374891 -0.066003459  0.397512033
#> 98   0.593651204 -0.699729027 -0.249165465
#> 99   0.706781448 -0.307514051  0.635145690
#> 100  0.806558309  0.002982834  0.413223389
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.462 -0.310   0.366 
#>  2      2        4       70        1  0.566 -0.212   0.722 
#>  3      3        4       70        1  0.271 -0.0662  0.397 
#>  4      4        4       70        1  0.218 -0.978  -0.131 
#>  5      5        4       70        1  0.699 -0.467  -0.0926
#>  6      6        4       70        1  0.770 -0.602   0.294 
#>  7      7        4       70        1  0.530 -0.558  -0.157 
#>  8      8        4       70        1  0.683 -0.413   0.267 
#>  9      9        4       70        1  0.766 -0.361  -0.0727
#> 10     10        4       70        1  0.482 -0.278  -0.0876
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc    depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>    <dbl> <dbl>
#> 1      1     1     4    70     1  6.35  51.3  1.44 0.124  24.1  6.93e+ 2 1236.
#> 2      1    14     4    70     1  6.35  51.3  1.44 0.124   7.78 5.05e- 6  399.
#> 3      2     1     4    70     1  7.05  56.7  2.06 0.124  25.8  4.46e+ 2 1464.
#> 4      2    14     4    70     1  7.05  56.7  2.06 0.124   6.80 9.84e-10  385.
#> 5      3     1     4    70     1  5.25  65.5  1.49 0.0801 19.6  6.70e+ 2 1282.
#> 6      3    14     4    70     1  5.25  65.5  1.49 0.0801 10.7  2.66e- 6  703.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>