Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)
#> $eta
#>            ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1   -1.950189e-01  0.3156729564       0       0
#> 2   -1.460268e-01  0.2147044370       0       0
#> 3   -5.893795e-02  0.0722355624       0       0
#> 4   -1.878721e-01 -0.0373552432       0       0
#> 5   -1.721257e-01 -0.0568925645       0       0
#> 6   -1.980520e-01  0.1387226289       0       0
#> 7   -9.442073e-02  0.1081748558       0       0
#> 8   -1.135914e-01  0.1674259992       0       0
#> 9   -1.520109e-01  0.0039841217       0       0
#> 10  -1.526992e-01  0.2414070381       0       0
#> 11  -1.877580e-01  0.2067834474       0       0
#> 12  -1.662047e-01  0.1763698910       0       0
#> 13   5.393249e-02 -0.0651686201       0       0
#> 14   5.393249e-02 -0.1870401972       0       0
#> 15  -2.491424e-01  0.2260884289       0       0
#> 16  -6.976054e-02  0.1328914628       0       0
#> 17  -1.900870e-02 -0.2555750206       0       0
#> 18  -1.207843e-02 -0.1368064189       0       0
#> 19  -3.097968e-01 -0.1169050961       0       0
#> 20  -1.125314e-01 -0.0005526712       0       0
#> 21  -2.133114e-01  0.0455102886       0       0
#> 22  -1.722911e-01  0.1738413204       0       0
#> 23  -1.768691e-01  0.0805462796       0       0
#> 24  -1.177285e-01  0.0527625484       0       0
#> 25  -6.507507e-02  0.0816438095       0       0
#> 26  -6.700171e-02 -0.0096500690       0       0
#> 27  -1.278392e-01  0.0457837833       0       0
#> 28  -9.350060e-02  0.1374257077       0       0
#> 29  -6.308643e-02  0.1550305893       0       0
#> 30  -2.239380e-01  0.0125021133       0       0
#> 31  -2.955545e-01  0.1780389859       0       0
#> 32  -2.867533e-01  0.2051951137       0       0
#> 33  -1.552622e-01  0.2895211985       0       0
#> 34  -1.573529e-01 -0.0048128306       0       0
#> 35  -2.006019e-01  0.2482761234       0       0
#> 36  -3.384191e-02  0.0015587098       0       0
#> 37  -1.251850e-01  0.0072601517       0       0
#> 38  -1.251850e-01  0.0072601517       0       0
#> 39  -3.318011e-01  0.2213954824       0       0
#> 40  -4.082441e-01  0.0561072999       0       0
#> 41  -7.125126e-02 -0.1800576648       0       0
#> 42   1.027128e-01 -0.1718575792       0       0
#> 43  -1.778100e-01  0.0220037897       0       0
#> 44  -2.039719e-02 -0.0788126332       0       0
#> 45  -3.180246e-02 -0.0193845664       0       0
#> 46  -1.038989e-01 -0.0441307554       0       0
#> 47  -6.737169e-02  0.0656427020       0       0
#> 48  -8.130035e-02  0.1169443886       0       0
#> 49  -9.803831e-02 -0.0736743037       0       0
#> 50  -1.934003e-01 -0.0308701871       0       0
#> 51   5.645442e-02  0.1915410038       0       0
#> 52  -4.699767e-02  0.1005811472       0       0
#> 53  -6.598708e-02 -0.0274368471       0       0
#> 54  -6.578057e-02 -0.0108202599       0       0
#> 55  -1.819959e-02  0.1190767888       0       0
#> 56  -1.701600e-01  0.1318283147       0       0
#> 57  -1.820270e-01  0.0454828847       0       0
#> 58  -1.093857e-01  0.2373215955       0       0
#> 59  -2.305989e-01  0.0365670396       0       0
#> 60  -1.747081e-01  0.1485030366       0       0
#> 61  -1.306338e-01  0.0492112141       0       0
#> 62  -4.553789e-02 -0.1273388434       0       0
#> 63  -1.099989e-01  0.0023628410       0       0
#> 64  -5.816382e-03  0.0023628410       0       0
#> 65  -5.902833e-02  0.0507614342       0       0
#> 66  -5.857426e-02  0.1985749911       0       0
#> 67  -1.381417e-01  0.1116290569       0       0
#> 68  -1.326100e-01  0.1867349032       0       0
#> 69  -1.477226e-01  0.1577386179       0       0
#> 70  -8.325310e-02  0.1147222003       0       0
#> 71  -1.167717e-01  0.0382524624       0       0
#> 72  -1.324119e-01  0.0626711405       0       0
#> 73  -1.644969e-01  0.1308885386       0       0
#> 74  -1.403334e-01  0.0616292946       0       0
#> 75  -9.341283e-02  0.1600374354       0       0
#> 76  -2.405377e-01  0.1154528871       0       0
#> 77   4.022034e-03 -0.0688878995       0       0
#> 78  -1.136872e-01  0.1458833858       0       0
#> 79  -2.286511e-01  0.0078765123       0       0
#> 80  -1.518435e-01  0.0670537617       0       0
#> 81  -1.250268e-01  0.0670537617       0       0
#> 82  -1.250268e-01  0.0670537617       0       0
#> 83  -1.357869e-01  0.1161299044       0       0
#> 84  -1.230608e-01  0.1672578618       0       0
#> 85  -2.152295e-01 -0.0043746855       0       0
#> 86  -6.910311e-02  0.1301465831       0       0
#> 87  -1.221936e-01  0.2126529397       0       0
#> 88  -1.453723e-01  0.1066820408       0       0
#> 89  -6.008309e-02 -0.1103935320       0       0
#> 90   1.265005e-02  0.0157623987       0       0
#> 91   3.716144e-02  0.0767031091       0       0
#> 92  -1.875723e-01  0.0390821139       0       0
#> 93   2.641561e-05  0.2192652327       0       0
#> 94  -1.372463e-01  0.2287624858       0       0
#> 95  -1.073238e-01  0.0853917961       0       0
#> 96  -1.390589e-01  0.0521680283       0       0
#> 97  -3.178890e-02  0.0187395717       0       0
#> 98   3.825520e-02 -0.0119471098       0       0
#> 99   1.494050e-02  0.0142295591       0       0
#> 100 -1.803369e-01  0.1695834770       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.195   0.316  
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.146   0.215  
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.0589  0.0722 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.188  -0.0374 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.172  -0.0569 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.198   0.139  
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0944  0.108  
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.114   0.167  
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.152   0.00398
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.153   0.241  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.205  26.7 0.041  0.12 62.4  1667. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.205  26.7 0.041  0.12  5.06  135. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.215  24.1 0.041  0.12 68.6  1654. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.215  24.1 0.041  0.12  5.01  121. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.235  20.9 0.041  0.12 77.9  1630. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.235  20.9 0.041  0.12  4.71   98.6
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>