Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>          ETA_Cl        ETA_Vc       ETA_Ka
#> 1    0.35412029 -1.2906852322 -0.448977280
#> 2    0.31880547 -0.2018615639  0.244514184
#> 3    0.51363185 -0.3298573078  0.170558309
#> 4    0.69866615 -0.2543627155  0.337819767
#> 5    0.36243779 -0.1859957999  0.188713544
#> 6    0.45111849 -0.3966956963  0.140822531
#> 7    0.53988358 -0.7337493494  0.043605789
#> 8    0.71083883 -0.0128645728  0.455431870
#> 9    0.82022114 -0.2657718315  0.254460371
#> 10   0.82022114 -0.6101332055 -0.490651790
#> 11   0.68252982 -0.4684292968 -0.123011374
#> 12   0.41554758 -0.2943865465  0.095912374
#> 13   0.54374974 -0.6726220944 -0.163708589
#> 14   0.67915435 -0.1599975063  0.281847217
#> 15   0.67188819 -0.2156304343  0.054956831
#> 16   0.63635332 -0.5285875463 -0.298996196
#> 17   0.45009199 -0.7131615225 -0.124057502
#> 18   0.34880148 -0.5854912525 -0.037014293
#> 19   0.58882382 -0.6992395737  0.584562095
#> 20   0.57580440 -0.4162729718  0.436539115
#> 21   0.47547828 -0.2720032092  0.475976514
#> 22   0.01277662 -0.1754439810  0.113662730
#> 23   0.28247224 -0.2912164143  0.326788167
#> 24   0.46760853 -0.5781230473  0.359356281
#> 25   0.46760853 -0.5781230473  0.206352809
#> 26   0.39502473 -0.5505958181  0.206352809
#> 27   0.41187541 -0.5241273090  0.153886389
#> 28   0.54690319 -0.1693323360  0.447495618
#> 29   0.53790164 -0.4312735171 -0.167302600
#> 30   0.41763103 -0.6099771506 -0.167302600
#> 31   0.83406634 -0.0604378458  0.717511877
#> 32   0.36297537 -0.3890430612 -0.007643202
#> 33   0.76413546 -0.1600770433  0.812383757
#> 34   0.50974562 -0.5666682601  0.293112755
#> 35   0.66869152 -0.2835244617  0.394024926
#> 36   0.83873896  0.0286219169  0.390190272
#> 37   0.86097978 -0.0146203534  0.350715242
#> 38   0.61754832 -0.0814862042  0.740060445
#> 39   0.49023129 -0.0814862042  0.802193582
#> 40   0.62984970 -0.1048941606  0.780885904
#> 41   0.54173775 -0.5407463605 -0.320294244
#> 42   0.64603272 -0.2836758511  0.267360390
#> 43   0.35614754 -0.5093510700  0.327008934
#> 44   0.19912045 -0.5552647164  0.751807264
#> 45   0.40221759 -0.2096851418  0.251687386
#> 46   0.69736574 -0.3090726296  0.340678112
#> 47   0.47779892 -0.1832153892  0.248065913
#> 48   0.45688291 -0.5449638221  0.094291419
#> 49   0.53412364 -0.2289844718  0.387416145
#> 50   0.48352343 -0.5639541375 -0.240545569
#> 51   0.00000000  0.0000000000  0.000000000
#> 52   0.21529803  0.0087699749  0.284688575
#> 53   0.47378849 -0.0863820930  0.723504149
#> 54   0.40099341 -0.3802525561  0.276208411
#> 55   0.77753643 -0.4486310692 -0.070054080
#> 56   0.53701351 -0.3061453343  0.320486410
#> 57   0.58227260 -0.0874936661  0.281681893
#> 58   0.46123610 -0.3331219549  0.571807489
#> 59   0.65073469 -0.0504892556  0.253784885
#> 60   0.36760694 -0.0210348202  0.252147300
#> 61   0.35052441 -0.1382730487  0.157804399
#> 62   0.66471525  0.0040464685  0.480441919
#> 63   0.36552344 -0.3010552958  0.450657863
#> 64   0.42964260 -0.9769458688 -0.592421541
#> 65   0.65230362 -0.0966153007  0.285487167
#> 66   0.13858396 -0.6089862694  0.593135324
#> 67   0.63335405 -0.3471258579  1.052441903
#> 68   0.61698943 -0.2455412039  0.515656365
#> 69   0.77683229  0.1078930638  1.084353724
#> 70   0.58175542 -0.1595756529  0.636226296
#> 71   0.56914380  0.1420376829  0.522645545
#> 72   0.53890168 -0.0465795869  0.797251546
#> 73   0.58563947 -0.3105646874  0.454221728
#> 74   0.53004211 -0.4630677064  0.158360323
#> 75   0.81884509 -0.3616943254  0.276759417
#> 76   0.59164393 -0.6003328632  0.273085181
#> 77   0.54744441 -0.6334244121 -0.576298349
#> 78   0.29372045 -0.9027152590 -0.026411488
#> 79   0.40386146 -0.8981866397 -0.206904893
#> 80   0.44377815 -0.6138155578  0.393884008
#> 81   0.60889706 -0.4408503788  0.010651017
#> 82  -0.11171000 -0.6623318207 -0.146964419
#> 83   0.68087705 -0.2647230565  0.116734627
#> 84   0.43970014 -0.3892180604  0.733057457
#> 85   0.86811356  0.1011129417  0.724599091
#> 86   0.72683394 -0.0006635716  0.506215141
#> 87   0.71529045  0.1188175637  0.980621453
#> 88   0.47066397 -0.2079713720  0.676060652
#> 89   0.37774382 -0.3782803324  0.306365108
#> 90   0.63415655 -0.5422667909  0.122355552
#> 91   0.64704212 -0.3971194066  0.147541272
#> 92   0.54522344 -0.8479479058  0.118638594
#> 93   0.53076360 -0.5048640246  0.129187028
#> 94   0.62866043 -0.4649398764 -0.128413055
#> 95   0.41008587 -0.5729194800  0.289427669
#> 96   0.34084475 -0.6250779579  0.139876051
#> 97   0.49778599 -0.5435637943 -0.213136481
#> 98   0.63671672 -0.2426665180  0.059119653
#> 99   0.17276353 -0.7771478161  0.187266313
#> 100  0.39482642 -1.1602833270 -1.020179515
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.354 -1.29   -0.449 
#>  2      2        4       70        1  0.319 -0.202   0.245 
#>  3      3        4       70        1  0.514 -0.330   0.171 
#>  4      4        4       70        1  0.699 -0.254   0.338 
#>  5      5        4       70        1  0.362 -0.186   0.189 
#>  6      6        4       70        1  0.451 -0.397   0.141 
#>  7      7        4       70        1  0.540 -0.734   0.0436
#>  8      8        4       70        1  0.711 -0.0129  0.455 
#>  9      9        4       70        1  0.820 -0.266   0.254 
#> 10     10        4       70        1  0.820 -0.610  -0.491 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc     depot  centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>     <dbl>  <dbl>
#> 1      1     1     4    70     1  5.70  19.3 0.638 0.296  34.8    1.24e+3  669. 
#> 2      1    14     4    70     1  5.70  19.3 0.638 0.296   3.28   3.10e-1   63.2
#> 3      2     1     4    70     1  5.50  57.2 1.28  0.0962 20.4    7.81e+2 1168. 
#> 4      2    14     4    70     1  5.50  57.2 1.28  0.0962 10.1    4.82e-5  576. 
#> 5      3     1     4    70     1  6.69  50.3 1.19  0.133  21.9    8.34e+2 1100. 
#> 6      3    14     4    70     1  6.69  50.3 1.19  0.133   7.21   1.68e-4  363. 
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>