Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  dat = NULL,
  prior_model = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  n_chains = 4,
  nocb = FALSE,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max = 3)
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

burn_in

Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter

Total number of iterations (following the burn-in iterations) for each Markov chain of the Metropolis-Hastings algorithm.

n_chains

Number of Markov chains

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

control

A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_mcmc(dat=df_patient01,prior_model=mod_run001,
n_iter=50,n_chains=2)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.49736166 -0.262121438  1.012791880
#> 2    0.54867885 -0.313105643  0.700858819
#> 3    0.35618499 -0.428159537  0.312859730
#> 4    0.68348235 -0.725390484 -0.355357424
#> 5    0.53854472 -0.725390484  0.052456140
#> 6    0.76378421 -0.207208125  0.118263957
#> 7    0.34178653 -0.606758120 -0.009326937
#> 8    0.71300224  0.073971130  0.965766216
#> 9    0.56473442 -0.442291495  0.187630445
#> 10   0.75335923 -0.007895713  0.612019539
#> 11   0.61808333  0.098674909  0.844866760
#> 12   0.85168297  0.081762586  0.328986718
#> 13   0.54928142 -0.099485832  0.739588632
#> 14   0.53339934 -0.245302394  0.462482036
#> 15  -0.03385513 -0.499919754 -0.067405380
#> 16   0.16336486 -0.073058695  0.458429533
#> 17   0.68364663 -0.161355561  0.913277515
#> 18   0.52989970 -0.092357651  0.486759238
#> 19   0.52204473  0.120074745  0.771758754
#> 20   0.50494570 -0.412120495 -0.022826847
#> 21   0.60283779 -0.174731915  0.734969825
#> 22   0.21846006 -0.948046174 -0.443361940
#> 23   0.36434913 -0.820263597  0.246778461
#> 24   0.25858386 -0.345394639  0.371309690
#> 25   0.55594589 -0.427938369  0.278366240
#> 26   0.83529453  0.049160104  0.550076922
#> 27   0.60677631 -0.088650526  0.418853980
#> 28   0.74920862 -0.038164327  0.679137405
#> 29   0.59444010 -0.415678065  0.162725448
#> 30   0.47286379 -0.563601975  0.162725448
#> 31   0.25385690 -0.239083530  0.335122495
#> 32   0.67400948 -0.269812832  0.339485713
#> 33   0.57709621 -0.132746082  0.693889638
#> 34   0.57709621 -0.440217754  0.693889638
#> 35   0.54314647 -0.497463421  0.067216964
#> 36   0.42395755 -0.580508660 -0.201718524
#> 37   0.49189322  0.216248595  0.928388159
#> 38   0.72275366  0.100873423  0.735143212
#> 39   0.76643086 -0.007212241  0.939468417
#> 40   0.72911004  0.019521980  0.480476616
#> 41   0.62720098 -0.371735990  0.111745310
#> 42   0.64644424 -0.175250875  0.886925363
#> 43   0.65646514 -0.278625493  0.412340584
#> 44   0.63514102 -0.146320876  0.647829405
#> 45   0.50500610 -0.431836723  0.510020600
#> 46   0.16194699 -1.003972434 -0.239243585
#> 47   0.47637019 -0.814945833 -0.542386231
#> 48   0.04990222 -1.237399504 -0.792847300
#> 49   0.16252571 -0.110445392  0.907665005
#> 50   0.76267378  0.055284486  0.517292001
#> 51   0.00000000  0.000000000  0.000000000
#> 52  -0.32510845 -0.247505533  0.203168958
#> 53   0.41565543 -0.971672357 -0.373031810
#> 54   0.54520690 -0.430499756  0.399877250
#> 55   0.34777650 -0.102875040  0.934708353
#> 56   0.62483321 -0.472529432  0.041681308
#> 57   0.67252155 -0.190584947  0.216975990
#> 58   0.60080782 -0.467557073  0.055296919
#> 59   0.36321747 -0.587474319  0.380827704
#> 60   0.64741941 -0.481480329  0.779594815
#> 61   0.63160323 -0.673916617  0.163354881
#> 62   0.50393097 -0.477923184  0.111312936
#> 63   0.58119086 -0.345136198  0.207976659
#> 64   0.32537071 -0.645524365  0.773049372
#> 65   0.34209301 -0.190589916  0.503819926
#> 66   0.45089573 -0.448928390  0.320597640
#> 67   0.39165903 -0.187480975  0.655541929
#> 68   0.29384770 -0.737768703  0.101634813
#> 69   0.45991758 -0.604094043  0.019797880
#> 70   0.35161437 -0.821770632 -0.302467002
#> 71   0.65951226 -0.606347919 -0.110533359
#> 72   0.64332975 -0.581616255  0.008811059
#> 73   0.59524686 -0.722079523 -0.473902807
#> 74   0.46116937 -0.637414125  0.259448880
#> 75   0.35255296 -0.736687702 -0.286711440
#> 76   0.58420542 -0.638167580 -0.054221560
#> 77   0.03675916 -0.949568514  0.110610862
#> 78   0.72860194 -0.164145530  0.451100331
#> 79   0.72423340 -0.231529186  0.280245634
#> 80   0.41967496 -0.357759456  0.049850726
#> 81   0.70890501 -0.390941524  0.359375875
#> 82   0.47072632 -0.398318466  0.721909252
#> 83   0.58288271 -0.475525140  0.159237297
#> 84   0.58094183 -0.451730153  0.458914153
#> 85   0.73445793 -0.422510113  0.282273638
#> 86   0.47088580 -0.257518422  0.413667028
#> 87   0.41767648 -0.425489607  0.063308776
#> 88   0.54899227 -0.720928960 -0.344959777
#> 89   0.54576298 -0.189540066  0.015207382
#> 90   0.69256150 -0.234299049  0.526472741
#> 91   0.31618207 -0.424795896  0.429293273
#> 92   0.49028868 -0.279826340  0.044334433
#> 93   0.51648170 -0.362280053  0.243448317
#> 94   0.36094833 -0.529721548  0.446227081
#> 95   0.22528867 -0.188420404  0.380832433
#> 96   0.50372692 -0.548128256  0.621383745
#> 97   0.57413960 -0.332209792  0.409978711
#> 98   0.58110199 -0.527798199 -0.108794061
#> 99   0.74944672 -0.342941324  0.825504385
#> 100  0.67503782 -0.278951230  0.080529833
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl   ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>    <dbl>    <dbl>
#>  1      1        4       70        1  0.497 -0.262    1.01   
#>  2      2        4       70        1  0.549 -0.313    0.701  
#>  3      3        4       70        1  0.356 -0.428    0.313  
#>  4      4        4       70        1  0.683 -0.725   -0.355  
#>  5      5        4       70        1  0.539 -0.725    0.0525 
#>  6      6        4       70        1  0.764 -0.207    0.118  
#>  7      7        4       70        1  0.342 -0.607   -0.00933
#>  8      8        4       70        1  0.713  0.0740   0.966  
#>  9      9        4       70        1  0.565 -0.442    0.188  
#> 10     10        4       70        1  0.753 -0.00790  0.612  
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc    depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>    <dbl> <dbl>
#> 1      1     1     4    70     1  6.58  53.9  2.75 0.122 30.1  2.74e+ 2 1623.
#> 2      1    14     4    70     1  6.58  53.9  2.75 0.122  7.25 9.82e-10  390.
#> 3      2     1     4    70     1  6.92  51.2  2.02 0.135 28.2  4.60e+ 2 1444.
#> 4      2    14     4    70     1  6.92  51.2  2.02 0.135  6.52 1.83e- 9  334.
#> 5      3     1     4    70     1  5.71  45.6  1.37 0.125 26.3  7.31e+ 2 1200.
#> 6      3    14     4    70     1  5.71  45.6  1.37 0.125  8.63 1.39e- 5  394.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>