Estimate the posterior distribution of individual parameters

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, , a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> $eta
#>         ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.6498041 -0.019123331       0       0
#> 2   -0.8726928 -0.114834983       0       0
#> 3   -0.7000905  0.038378752       0       0
#> 4   -0.7748301  0.038378752       0       0
#> 5   -1.0279211 -0.042231020       0       0
#> 6   -0.8238482  0.029212434       0       0
#> 7   -0.6958233 -0.198202756       0       0
#> 8   -0.7621952 -0.024493862       0       0
#> 9   -0.7060572 -0.024493862       0       0
#> 10  -0.6188322 -0.154716884       0       0
#> 11  -0.6824858 -0.061913107       0       0
#> 12  -0.6802727 -0.026607424       0       0
#> 13  -0.6408240  0.009261701       0       0
#> 14  -0.7036764 -0.037291901       0       0
#> 15  -0.7641644  0.041348321       0       0
#> 16  -0.7906135  0.035170473       0       0
#> 17  -0.6917822  0.018103086       0       0
#> 18  -0.8280606  0.077822431       0       0
#> 19  -0.5645744  0.005909890       0       0
#> 20  -0.7316147 -0.098506071       0       0
#> 21  -0.8313347  0.048431376       0       0
#> 22  -0.7073688 -0.269221987       0       0
#> 23  -0.5556130 -0.132429341       0       0
#> 24  -0.5556130 -0.141193076       0       0
#> 25  -0.6574548 -0.131354858       0       0
#> 26  -0.6708323 -0.102555684       0       0
#> 27  -0.7177551 -0.029394710       0       0
#> 28  -0.7625431 -0.029648339       0       0
#> 29  -0.6678881 -0.160590792       0       0
#> 30  -0.7581188 -0.117883314       0       0
#> 31  -0.9838439  0.138578082       0       0
#> 32  -0.8031070 -0.079461892       0       0
#> 33  -0.5401895 -0.202305386       0       0
#> 34  -0.6200695 -0.363334725       0       0
#> 35  -0.5255017 -0.311962484       0       0
#> 36  -0.7897493 -0.096633042       0       0
#> 37  -0.8128752 -0.018582278       0       0
#> 38  -0.6668746 -0.190288098       0       0
#> 39  -0.9607821  0.074353413       0       0
#> 40  -0.9887673 -0.034891651       0       0
#> 41  -0.6093634 -0.212660092       0       0
#> 42  -0.6069450 -0.144049906       0       0
#> 43  -0.7415761 -0.088399036       0       0
#> 44  -0.7240133  0.001674777       0       0
#> 45  -0.8703200  0.095523172       0       0
#> 46  -0.7354410 -0.041145420       0       0
#> 47  -0.6474873  0.026072172       0       0
#> 48  -0.6474873  0.026072172       0       0
#> 49  -0.6132784 -0.131790189       0       0
#> 50  -0.6135150 -0.206185990       0       0
#> 51  -0.4854143 -0.212897927       0       0
#> 52  -0.6847829 -0.161358506       0       0
#> 53  -0.6158017 -0.151834818       0       0
#> 54  -0.5557262 -0.072676033       0       0
#> 55  -0.7142444 -0.072676033       0       0
#> 56  -0.7577482 -0.051281001       0       0
#> 57  -0.7826980 -0.066958718       0       0
#> 58  -0.7903927 -0.240913060       0       0
#> 59  -0.5963035 -0.048216729       0       0
#> 60  -0.6915951 -0.143045779       0       0
#> 61  -0.7179101 -0.012505207       0       0
#> 62  -0.7409172 -0.068582191       0       0
#> 63  -0.7656620 -0.072957088       0       0
#> 64  -1.0167370 -0.011503630       0       0
#> 65  -0.8154897 -0.211725787       0       0
#> 66  -0.9003160  0.122219048       0       0
#> 67  -0.6347426  0.122219048       0       0
#> 68  -0.6347426  0.122219048       0       0
#> 69  -0.6349425  0.122219048       0       0
#> 70  -0.6901412  0.013066412       0       0
#> 71  -0.6901412  0.013066412       0       0
#> 72  -0.6727888 -0.193193367       0       0
#> 73  -0.4619593 -0.287590272       0       0
#> 74  -0.5832460  0.002189720       0       0
#> 75  -0.7533788 -0.037000680       0       0
#> 76  -0.5885625 -0.130133626       0       0
#> 77  -0.6540931 -0.010675616       0       0
#> 78  -0.6625451 -0.113100654       0       0
#> 79  -0.8676963 -0.025827860       0       0
#> 80  -0.7963855 -0.035622335       0       0
#> 81  -1.0366785  0.172960899       0       0
#> 82  -0.6039496 -0.149338284       0       0
#> 83  -0.7106454 -0.149338284       0       0
#> 84  -0.6795982 -0.149338284       0       0
#> 85  -0.7222484 -0.071912207       0       0
#> 86  -0.6578094  0.038568374       0       0
#> 87  -0.8148733 -0.021171646       0       0
#> 88  -0.6668316 -0.021171646       0       0
#> 89  -0.6668316 -0.056844933       0       0
#> 90  -0.5951934 -0.121709758       0       0
#> 91  -0.6714689 -0.083635615       0       0
#> 92  -0.9497881 -0.074701057       0       0
#> 93  -0.6186365 -0.204667776       0       0
#> 94  -0.7513817 -0.121052715       0       0
#> 95  -0.9118165 -0.115281708       0       0
#> 96  -1.0141270 -0.102471062       0       0
#> 97  -0.7957962 -0.218148602       0       0
#> 98  -0.7948310 -0.090045497       0       0
#> 99  -0.7876236 -0.044114627       0       0
#> 100 -0.6522871  0.070888475       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21 ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>  <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.650 -0.0191
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.873 -0.115 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.700  0.0384
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.775  0.0384
#>  5      5     0.21      80    65    19.8     0.041      0.12 -1.03  -0.0422
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.824  0.0292
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.696 -0.198 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.762 -0.0245
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.706 -0.0245
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.619 -0.155 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.130  19.1 0.041  0.12 23.1  441. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.130  19.1 0.041  0.12  4.10  78.4
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.104  17.4 0.041  0.12 25.9  449. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.104  17.4 0.041  0.12  6.14 107. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.124  20.2 0.041  0.12 21.9  443. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.124  20.2 0.041  0.12  4.17  84.4
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>