Estimate the posterior distribution of individual parameters by MCMC

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> $eta
#>          ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.08297106  0.074875694       0       0
#> 2   -0.12778479  0.149721532       0       0
#> 3   -0.12778479  0.149721532       0       0
#> 4   -0.19543360  0.216125443       0       0
#> 5   -0.30157767  0.133464337       0       0
#> 6   -0.39091823  0.146170241       0       0
#> 7   -0.20430068  0.193504570       0       0
#> 8   -0.04758191 -0.014025348       0       0
#> 9   -0.22130863  0.009828456       0       0
#> 10  -0.09599184  0.130590601       0       0
#> 11  -0.14529119  0.125363008       0       0
#> 12  -0.09055217  0.116385796       0       0
#> 13  -0.09055217  0.001120463       0       0
#> 14  -0.16302623  0.094779480       0       0
#> 15  -0.14108022  0.055356046       0       0
#> 16  -0.02014768  0.034136484       0       0
#> 17  -0.14171973  0.250155409       0       0
#> 18  -0.15973120  0.143092143       0       0
#> 19  -0.14060217  0.020600713       0       0
#> 20  -0.05160062  0.125410295       0       0
#> 21  -0.18398055  0.232805941       0       0
#> 22  -0.20164627  0.153775869       0       0
#> 23  -0.23120250  0.153775869       0       0
#> 24  -0.35553380  0.347442537       0       0
#> 25  -0.54999106  0.445197409       0       0
#> 26  -0.43871938 -0.077803630       0       0
#> 27  -0.05743687  0.047415764       0       0
#> 28  -0.12673840  0.132297886       0       0
#> 29  -0.25667973  0.132297886       0       0
#> 30  -0.05324201  0.011463814       0       0
#> 31  -0.01599183  0.033834332       0       0
#> 32  -0.06501035 -0.001528406       0       0
#> 33   0.08691710 -0.098324365       0       0
#> 34  -0.10185367  0.051077313       0       0
#> 35  -0.07050116  0.171224936       0       0
#> 36  -0.07368928  0.137239720       0       0
#> 37   0.02169420  0.014095350       0       0
#> 38  -0.03818037  0.042911723       0       0
#> 39  -0.14192799  0.059686770       0       0
#> 40  -0.10964943  0.330036936       0       0
#> 41  -0.20910289  0.305106108       0       0
#> 42  -0.20910289  0.107231411       0       0
#> 43  -0.15012174  0.049542693       0       0
#> 44  -0.07140896 -0.226095273       0       0
#> 45  -0.09040741  0.003376791       0       0
#> 46  -0.11120984  0.214691631       0       0
#> 47  -0.11139750 -0.055424087       0       0
#> 48  -0.11139750  0.139650578       0       0
#> 49   0.03302239  0.008615949       0       0
#> 50  -0.04030810  0.047087143       0       0
#> 51  -0.20615896  0.115448656       0       0
#> 52  -0.20615896  0.115448656       0       0
#> 53  -0.44993882  0.439926328       0       0
#> 54  -0.40111995  0.385338863       0       0
#> 55  -0.30098787 -0.025453263       0       0
#> 56  -0.08796956  0.012589161       0       0
#> 57  -0.06195373  0.076226708       0       0
#> 58  -0.01095918 -0.054684524       0       0
#> 59  -0.32794141  0.044707160       0       0
#> 60  -0.19930206 -0.009803977       0       0
#> 61  -0.06538041  0.124135363       0       0
#> 62  -0.21416555  0.133020462       0       0
#> 63  -0.14346710  0.215390455       0       0
#> 64  -0.11822373  0.163230776       0       0
#> 65  -0.18415335  0.147367064       0       0
#> 66  -0.14942872  0.074647320       0       0
#> 67   0.02456805  0.002287553       0       0
#> 68  -0.19151512  0.006171535       0       0
#> 69  -0.19151512  0.006171535       0       0
#> 70  -0.06653245  0.078297700       0       0
#> 71  -0.03702906  0.042454637       0       0
#> 72  -0.04598042 -0.068573247       0       0
#> 73  -0.19065168  0.165715699       0       0
#> 74  -0.02495695  0.081074523       0       0
#> 75  -0.23455340  0.154871104       0       0
#> 76  -0.28335871  0.242695819       0       0
#> 77  -0.14051822  0.021221142       0       0
#> 78  -0.08424683  0.075144372       0       0
#> 79  -0.07636411  0.062138840       0       0
#> 80  -0.13993651 -0.082070679       0       0
#> 81  -0.19201666 -0.047312797       0       0
#> 82   0.02855052 -0.039635755       0       0
#> 83  -0.12544709  0.040827228       0       0
#> 84  -0.12544709  0.040827228       0       0
#> 85  -0.07538614  0.108263217       0       0
#> 86  -0.16204471  0.211802966       0       0
#> 87  -0.11134431  0.162537187       0       0
#> 88  -0.15295862  0.218852495       0       0
#> 89  -0.28907294 -0.066707843       0       0
#> 90  -0.15367868 -0.005065933       0       0
#> 91  -0.20341598  0.184732000       0       0
#> 92  -0.04236193  0.054850967       0       0
#> 93  -0.09506188  0.030191625       0       0
#> 94  -0.02786753 -0.059016652       0       0
#> 95  -0.21761429  0.013335126       0       0
#> 96   0.05059013 -0.023678803       0       0
#> 97  -0.13408154 -0.019245503       0       0
#> 98  -0.01317292 -0.084574362       0       0
#> 99  -0.09767827  0.180995437       0       0
#> 100 -0.14714020  0.134667847       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.0830  0.0749 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.128   0.150  
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.128   0.150  
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.195   0.216  
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.302   0.133  
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.391   0.146  
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.204   0.194  
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0476 -0.0140 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.221   0.00983
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.0960  0.131  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.229  21.0 0.041  0.12 78.0  1637.
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.229  21.0 0.041  0.12  4.98  104.
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.219  22.6 0.041  0.12 72.9  1649.
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.219  22.6 0.041  0.12  5.13  116.
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.219  22.6 0.041  0.12 72.9  1649.
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.219  22.6 0.041  0.12  5.13  116.
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>