Estimate the posterior distribution of individual parameters

Estimates the posterior distribution of individual parameters by Markov Chain Monte Carlo (using a Metropolis-Hastings algorithm)

poso_estim_mcmc(
  object = NULL,
  return_model = TRUE,
  burn_in = 50,
  n_iter = 1000,
  control = list(n_kernel = c(2, 2, 2), stepsize_rw = 0.4, proba_mcmc = 0.3, nb_max =
    3)
)

Arguments

object
A posologyr list, created by the posologyr function.

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

burn_in
Number of burn-in iterations for the Metropolis-Hastings algorithm.

n_iter
Total number of iterations (following the burn-in iterations) for the Metropolis-Hastings algorithm.

control
A list of parameters controlling the Metropolis-Hastings algorithm.

Value

If return_model is set to FALSE, , a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Markov Chain Monte Carlo. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

References

Comets E, Lavenu A, Lavielle M. Parameter estimation in nonlinear mixed effect models using saemix, an R implementation of the SAEM algorithm. Journal of Statistical Software 80, 3 (2017), 1-41.

Author

Emmanuelle Comets, Audrey Lavenu, Marc Lavielle, Cyril Leven

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_mcmc(patient01_tobra,n_iter=100)

#> $eta
#>         ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1   -0.6381666  9.065365e-02       0       0
#> 2   -0.6731583 -1.146333e-01       0       0
#> 3   -0.6731583 -1.419003e-01       0       0
#> 4   -0.6862326 -9.461728e-02       0       0
#> 5   -0.6114712 -1.645414e-01       0       0
#> 6   -0.6433670 -4.613243e-02       0       0
#> 7   -0.6174869 -2.449728e-01       0       0
#> 8   -0.6651733 -4.163777e-02       0       0
#> 9   -0.8186136 -3.006557e-02       0       0
#> 10  -0.8615869  7.263232e-03       0       0
#> 11  -0.6470773  7.263232e-03       0       0
#> 12  -0.7211588 -1.094773e-01       0       0
#> 13  -0.8937909  1.712095e-01       0       0
#> 14  -0.8937909  4.213045e-05       0       0
#> 15  -0.9266940 -5.998789e-02       0       0
#> 16  -0.7213404 -9.814386e-02       0       0
#> 17  -0.7848108  9.614708e-02       0       0
#> 18  -0.7532213  1.327953e-01       0       0
#> 19  -0.6495835 -1.160261e-01       0       0
#> 20  -0.5376823  8.410343e-02       0       0
#> 21  -0.5955230 -3.643603e-02       0       0
#> 22  -0.7733515  4.855292e-02       0       0
#> 23  -0.7684615  9.445805e-03       0       0
#> 24  -0.7091451 -3.029286e-02       0       0
#> 25  -0.9465389 -1.122217e-01       0       0
#> 26  -1.0130203  1.315535e-01       0       0
#> 27  -0.6906981 -2.005114e-01       0       0
#> 28  -0.6439371 -7.798115e-02       0       0
#> 29  -0.7206642 -1.689968e-01       0       0
#> 30  -0.6662685 -1.575577e-01       0       0
#> 31  -0.7228031 -1.701583e-01       0       0
#> 32  -0.5578124 -1.029549e-01       0       0
#> 33  -0.6899597 -7.855518e-02       0       0
#> 34  -0.6104904 -7.855518e-02       0       0
#> 35  -0.6232232 -3.189767e-03       0       0
#> 36  -0.9053593 -2.574484e-02       0       0
#> 37  -0.6073892 -3.302036e-01       0       0
#> 38  -0.5642459 -3.302036e-01       0       0
#> 39  -0.5995749 -1.354402e-01       0       0
#> 40  -0.5823417 -1.354402e-01       0       0
#> 41  -0.7005495 -5.116759e-02       0       0
#> 42  -0.6713170 -1.212088e-01       0       0
#> 43  -0.6910966 -1.537153e-01       0       0
#> 44  -0.6350737 -1.236954e-01       0       0
#> 45  -0.6200674 -1.568070e-01       0       0
#> 46  -0.5160199 -1.300688e-01       0       0
#> 47  -0.7402212 -9.863474e-02       0       0
#> 48  -0.5390520 -1.706246e-01       0       0
#> 49  -0.8278856 -6.206851e-02       0       0
#> 50  -0.7194493  3.419470e-02       0       0
#> 51  -0.6800186 -2.456173e-01       0       0
#> 52  -0.7478221 -1.397691e-01       0       0
#> 53  -0.5942199 -8.668983e-02       0       0
#> 54  -0.8418560 -1.233689e-02       0       0
#> 55  -0.7884497  6.296014e-03       0       0
#> 56  -0.6905696  2.063062e-02       0       0
#> 57  -0.7575073  1.413624e-01       0       0
#> 58  -0.6734610  7.047396e-02       0       0
#> 59  -0.8027106  1.005481e-02       0       0
#> 60  -0.8538511  3.096529e-02       0       0
#> 61  -0.7409749 -2.135867e-01       0       0
#> 62  -0.7802968 -2.444471e-01       0       0
#> 63  -0.7611151 -9.960848e-02       0       0
#> 64  -0.8129495 -9.960848e-02       0       0
#> 65  -0.8550350 -4.224679e-02       0       0
#> 66  -0.5888025  3.176179e-02       0       0
#> 67  -0.6691780 -3.119307e-02       0       0
#> 68  -0.7231984 -3.341893e-02       0       0
#> 69  -0.6006068 -3.028801e-02       0       0
#> 70  -0.6006068 -6.766917e-02       0       0
#> 71  -0.6197647 -1.614533e-01       0       0
#> 72  -0.4554496 -2.312705e-01       0       0
#> 73  -0.6163415  1.614054e-01       0       0
#> 74  -0.6648039 -2.365468e-01       0       0
#> 75  -0.8276255 -3.622960e-01       0       0
#> 76  -1.1033645 -4.093172e-03       0       0
#> 77  -0.7385356 -5.677748e-02       0       0
#> 78  -0.6389406 -6.734581e-02       0       0
#> 79  -0.5087915 -1.056159e-01       0       0
#> 80  -0.6792508 -1.056159e-01       0       0
#> 81  -0.6803951 -8.063604e-02       0       0
#> 82  -0.6803951 -1.615757e-01       0       0
#> 83  -0.6803951 -1.207801e-01       0       0
#> 84  -0.7476382 -8.099634e-02       0       0
#> 85  -0.6370372 -7.572499e-02       0       0
#> 86  -0.5175773 -2.435055e-01       0       0
#> 87  -0.8930536 -1.316951e-01       0       0
#> 88  -0.8932813 -8.675212e-02       0       0
#> 89  -0.7254553 -1.712862e-01       0       0
#> 90  -0.6614055 -2.370985e-02       0       0
#> 91  -0.6596331 -8.213673e-02       0       0
#> 92  -0.6646366 -8.549519e-02       0       0
#> 93  -0.6036015 -4.344545e-02       0       0
#> 94  -0.5866273 -1.278401e-01       0       0
#> 95  -0.8044575  1.154252e-01       0       0
#> 96  -0.7336767 -1.382086e-01       0       0
#> 97  -0.7003092 -2.242560e-02       0       0
#> 98  -0.7003092 -1.337348e-01       0       0
#> 99  -0.8206183  3.976953e-02       0       0
#> 100 -0.8828548  6.985901e-02       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21 ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>  <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.638  0.0907 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.673 -0.115  
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.673 -0.142  
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.686 -0.0946 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.611 -0.165  
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.643 -0.0461 
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.617 -0.245  
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.665 -0.0416 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.819 -0.0301 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.862  0.00726
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.132  21.3 0.041  0.12 20.6  440. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.132  21.3 0.041  0.12  3.61  77.0
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.127  17.4 0.041  0.12 25.4  442. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.127  17.4 0.041  0.12  4.68  81.2
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.127  16.9 0.041  0.12 26.1  442. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.127  16.9 0.041  0.12  4.81  81.2
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>