Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e4,n_resample=1e3)
#> $eta
#>            ETA_Cl        ETA_Vc        ETA_Ka
#> 1     0.488882429 -0.5193379922  0.4697085801
#> 2     0.680957131 -0.3978890187  0.2411645751
#> 3     0.738994509 -0.3456147424  0.2305001090
#> 4     0.302204538 -0.5275800520  0.5682345954
#> 5     0.579914838 -0.5647093714 -0.4385272876
#> 6     0.605055322 -0.5322610153 -0.3150155113
#> 7     0.449543625 -0.3228331516  0.4156916002
#> 8     0.591657423 -0.4847878407 -0.0476600846
#> 9     0.589824620  0.0464889159  0.5077102662
#> 10    0.505836187  0.1067886369  0.3805368846
#> 11    0.597500360 -0.5132245184  0.1778545384
#> 12    0.247662693 -0.7495314232  0.7625694462
#> 13    0.717663662 -0.1293603263  0.4726001780
#> 14    0.759083818  0.1016986566  0.7734819130
#> 15    0.466699894 -0.6966513181  0.1089128478
#> 16    0.333802104 -0.2067233435  0.2066754602
#> 17    0.465280564 -0.8706595991 -0.1487998034
#> 18    0.288557568 -0.7975681029 -0.0204031299
#> 19    0.349285088  0.0787114845  0.6099091051
#> 20    0.673228723 -0.5474619466 -0.1724927482
#> 21    0.433004971 -0.8951925425 -0.1675247852
#> 22    0.342391378  0.0690850001  0.7194203829
#> 23    0.761229087 -0.0935241378  0.2202250086
#> 24    0.817114376 -0.0170463170  1.1503714554
#> 25    0.066450161 -0.3365425895  0.1489922426
#> 26    0.769782586 -0.1686125625  0.6506453126
#> 27    0.929393572  0.0498340878  0.6505997200
#> 28    0.596799985 -0.3875585591  0.2931267203
#> 29    0.639525260 -0.4080283523  0.3554821128
#> 30   -0.163788987 -0.4232718582  0.2824373181
#> 31    0.522603666 -0.4364880541  0.0846083449
#> 32    0.240807694 -0.5286082474  0.0437859239
#> 33    0.332173671 -0.2535307008  0.6809187932
#> 34    0.522304299 -0.5863683365  0.1231520968
#> 35    0.343988237 -0.9046336017 -0.2727816058
#> 36    0.718837058 -0.0786315190  0.7932792965
#> 37   -0.110588845 -0.2207403382  1.1134429167
#> 38    0.344678571 -0.5611603413 -0.1118324237
#> 39    0.742933727 -0.4000697522  0.0191115481
#> 40    0.334299112 -0.3112158025  0.3427953502
#> 41    0.036734153 -0.3914604092  0.5671411047
#> 42    0.510612301 -0.3902438063  0.3496449006
#> 43    0.817114376 -0.0170463170  1.1503714554
#> 44    0.528971410 -0.5174896290  0.2876276409
#> 45    0.093936704  0.2054153700  1.6016675542
#> 46    0.517729812 -0.7246092841 -0.2274196621
#> 47    0.577857272 -0.4809237716  0.5418112635
#> 48    0.603922600 -0.1262038342  0.7283619389
#> 49    0.759083818  0.1016986566  0.7734819130
#> 50    0.633580780 -0.1864317184  0.2498479446
#> 51    0.284857140 -0.8295482444 -0.5586929057
#> 52    0.529520011 -0.4107775991 -0.3113407731
#> 53    0.702938702 -0.3353722047  0.1331238216
#> 54    0.419439372 -0.3715063372  0.2704695862
#> 55    0.470857068 -0.7012752901 -0.0533944483
#> 56    0.442899340 -0.2764638475  0.3015319417
#> 57    0.644880526 -0.6166015840 -0.1176687865
#> 58    0.702648989  0.0801524848  0.6705259113
#> 59    0.500657439 -0.5747495210 -0.1204946640
#> 60    0.149309717 -0.8804546325 -0.3804730303
#> 61    0.533803409 -0.6204993838 -0.0267670499
#> 62    0.697170215 -0.3207257006  0.2505779678
#> 63    0.148195879 -0.7686583193 -0.4140991924
#> 64    0.738994509 -0.3456147424  0.2305001090
#> 65    0.408135672 -0.5242783082 -0.4598463664
#> 66    0.522603666 -0.4364880541  0.0846083449
#> 67    0.611148515 -0.4980743716 -0.1933281665
#> 68    0.532783979 -0.7035839040  0.4237405259
#> 69    0.558125200 -0.5932552569 -0.1266574887
#> 70    0.587098113 -0.5366719140  0.3834410678
#> 71    0.389535392 -0.2281267365  0.9080321640
#> 72    0.325005447 -0.2148081007  0.5270248761
#> 73    0.753195634 -0.3088516806  0.2159745071
#> 74    0.593459310 -0.1399539375  0.1666339577
#> 75    0.702938702 -0.3353722047  0.1331238216
#> 76    0.785419819 -0.3161356791 -0.0243700099
#> 77    0.627236699 -0.5685253270  0.3974905464
#> 78    0.462249269 -0.0284248628  0.3974446408
#> 79    0.804483325 -0.2503472868  0.1463935698
#> 80    0.653998535 -0.3959659836  0.4507163848
#> 81    0.446231244 -0.5845835171  0.2900176672
#> 82    0.898678701 -0.0610141755  0.8619640118
#> 83    0.870197982 -0.0616065924  0.6103987386
#> 84    0.770009274  0.0394608417  0.6567141262
#> 85    0.447318165 -0.7575562894 -0.3961136656
#> 86    0.392152164 -0.7059789514 -0.1980735628
#> 87    0.536525205 -0.5428362715  0.0786830139
#> 88    0.560545465 -0.7264377439 -0.1610015840
#> 89    0.460984611 -0.7121593338  0.1812313102
#> 90    0.669437530 -0.5646352736 -0.1491879479
#> 91    0.381029996 -0.2775796444  0.9396813329
#> 92   -0.002422997 -0.3730502508 -0.0124539351
#> 93    0.673228723 -0.5474619466 -0.1724927482
#> 94    0.608775887  0.2154992384  0.7241718332
#> 95    0.264110165 -0.7276533046 -0.3813094748
#> 96    0.502653327 -0.2256884178  0.4189055560
#> 97    0.354231697 -0.3586118805  0.3236512453
#> 98    0.562736524 -0.3754320839  0.1735595175
#> 99    0.260693905 -0.3086383028  0.6412809765
#> 100   0.481049275 -0.6559345549 -0.3337964889
#> 101   0.171417922 -0.8975780351 -0.1921506445
#> 102   0.610246316 -0.1860627135  0.4337096531
#> 103   0.233612689 -0.2594227402  0.3935431877
#> 104   0.517729812 -0.7246092841 -0.2274196621
#> 105   0.481049275 -0.6559345549 -0.3337964889
#> 106   0.467774196 -0.3884306138  0.4189471096
#> 107   0.933663474 -0.0713366781  0.3319437929
#> 108   0.466480646 -0.5580765499  0.1352134014
#> 109   0.448426371  0.0348655548  0.6780068873
#> 110   0.697620813 -0.1081130362  0.7089193134
#> 111   0.769782586 -0.1686125625  0.6506453126
#> 112   0.574447770 -0.4359195219 -0.1549239484
#> 113   0.527590926 -0.3786855055  0.0842204892
#> 114   0.536525205 -0.5428362715  0.0786830139
#> 115   0.808513994 -0.1933859208  0.2626094252
#> 116   0.808513994 -0.1933859208  0.2626094252
#> 117   0.669437530 -0.5646352736 -0.1491879479
#> 118   0.085522053 -0.2676986922  0.1819445693
#> 119   0.621275263 -0.1504336522  0.5449235100
#> 120   0.389614578  0.0545158385  0.6258871771
#> 121   0.434697333 -0.6688966580 -0.0145753373
#> 122   0.187838216 -0.5367150766 -0.1145077690
#> 123   0.680957131 -0.3978890187  0.2411645751
#> 124   0.332601918 -0.5180550896  0.4274471664
#> 125   0.894095454 -0.0319822433  0.5434940593
#> 126   0.585650672 -0.5201446497 -0.2474993570
#> 127   0.517729812 -0.7246092841 -0.2274196621
#> 128   0.624595403 -0.1038784281  0.5045013528
#> 129   0.375813909 -0.4448287203 -0.2514912185
#> 130   0.240995164 -0.3536188152  0.4261236801
#> 131   0.707976084 -0.1547232289  0.9765115024
#> 132   0.591799827 -0.5214798767 -0.0200362404
#> 133   0.285969772 -0.9389469777 -0.6994155174
#> 134   0.395239832 -0.3128308637  0.0323138994
#> 135   0.768206703 -0.1490874141  0.7074006829
#> 136   0.768206703 -0.1490874141  0.7074006829
#> 137   0.439997537 -0.5244168946  0.1084983526
#> 138   0.618493603  0.0565361911  0.6671997635
#> 139   0.458422412 -0.8531389306 -0.0367465993
#> 140   0.430711649 -1.0222303007 -0.3379985011
#> 141   0.674062889 -0.3269624188  0.3076873928
#> 142   0.328600063 -1.1815729977 -0.7530961858
#> 143   0.593459310 -0.1399539375  0.1666339577
#> 144   0.394533982 -0.5724687622 -0.1421417708
#> 145   0.249120125 -0.6243714565 -0.2503183831
#> 146   0.424528339 -0.3819410103 -0.1156833698
#> 147   0.694080726 -0.4885010854  0.0295663880
#> 148   1.041663388  0.1324037334  0.6965064444
#> 149   0.687147345 -0.4511434243 -0.2804474807
#> 150   0.653998535 -0.3959659836  0.4507163848
#> 151   0.458082582 -0.3045358157  0.2935757906
#> 152   0.397391943 -0.7129855438  0.0132751773
#> 153   0.491230322 -0.2029236489  0.4834774071
#> 154   0.301988041 -0.8116186751 -0.3280006498
#> 155   0.870197982 -0.0616065924  0.6103987386
#> 156   0.523318304 -0.2444993751 -0.0041720767
#> 157   0.693496018 -0.0022729054  0.4342803054
#> 158   0.190389371 -0.4487198739  0.4136979344
#> 159   0.348437627 -0.7643806147  0.7537150329
#> 160   0.148195879 -0.7686583193 -0.4140991924
#> 161   0.583074813 -0.4261627843  0.2454969749
#> 162   0.562893493 -0.2266771303  0.2395295164
#> 163   0.380495653 -0.3604301134  1.1273208456
#> 164   0.196137033 -0.8667457869 -0.3483336238
#> 165   0.312920532 -0.7835286318  0.2547111660
#> 166   0.463256770 -0.2437855381  0.0967967867
#> 167   0.477380758 -0.5764138300 -0.1984008386
#> 168   0.633580780 -0.1864317184  0.2498479446
#> 169   0.597500360 -0.5132245184  0.1778545384
#> 170   0.564117418 -0.4800938994  0.5678678935
#> 171   0.603922600 -0.1262038342  0.7283619389
#> 172   0.457672728 -0.4224624292  0.5714255913
#> 173   0.399213789 -1.2008687423 -0.5623802136
#> 174   0.730247532 -0.0722159148  0.4691592867
#> 175   0.815574995  0.2382713157  0.9584291442
#> 176   0.391186251 -0.6977422613 -0.0765389788
#> 177   0.663337897 -0.1044994685  0.3748580912
#> 178   0.496758665 -0.2754850490  0.1508734219
#> 179   0.492459107 -0.7812901676 -0.3082812756
#> 180   0.439801829 -0.3101206378  0.4056870977
#> 181   0.698393748 -0.2646586256  0.4288310627
#> 182   0.808513994 -0.1933859208  0.2626094252
#> 183   0.815574995  0.2382713157  0.9584291442
#> 184   0.467577527 -1.0332432322 -0.4897257934
#> 185   0.602818315 -0.4270626748  0.4848650341
#> 186   0.435720947 -0.1018609561  0.5755311622
#> 187   0.389535392 -0.2281267365  0.9080321640
#> 188   0.558125200 -0.5932552569 -0.1266574887
#> 189   0.686823053 -0.3367375570 -0.0003031393
#> 190   0.596799985 -0.3875585591  0.2931267203
#> 191   0.336026131 -0.2972518269  0.3600423353
#> 192   0.653998535 -0.3959659836  0.4507163848
#> 193   0.769782586 -0.1686125625  0.6506453126
#> 194   0.465879291 -0.1356635227  0.4694093687
#> 195   0.224564460 -0.1844717426  0.3019321814
#> 196   0.562893493 -0.2266771303  0.2395295164
#> 197   0.675311541 -0.0777936693  0.6716356933
#> 198   0.894703004 -0.0300343280  0.3765102950
#> 199   0.288339646 -0.7742779253  0.0477430965
#> 200   0.833167253 -0.1898677425  0.2502804779
#> 201   0.237850296 -0.3970333238 -0.2315123709
#> 202   0.768206703 -0.1490874141  0.7074006829
#> 203   0.605234096 -0.2617282588  1.6421873962
#> 204   0.657195360 -0.4731286483  0.0495742550
#> 205   0.372132611 -0.5927318739  0.2454892372
#> 206   0.815693725 -0.3278289705  0.0488447426
#> 207   0.662702054 -0.3143555196  0.1635569079
#> 208   0.768298343  0.0964523655  0.6873290244
#> 209   0.728248541 -0.2913468093  0.8580562593
#> 210   0.768206703 -0.1490874141  0.7074006829
#> 211   0.918337593  0.0048262394  0.4565741613
#> 212   0.652818412  0.1260472232  0.7752869101
#> 213   0.224796181 -0.0052082793  0.6230605472
#> 214   0.596213414 -0.0150436912  0.5492284038
#> 215   0.330532220 -0.9992477517 -0.3728109851
#> 216   0.611148515 -0.4980743716 -0.1933281665
#> 217   0.386236288 -0.1732939204  0.6741637326
#> 218   0.694080726 -0.4885010854  0.0295663880
#> 219   0.644880526 -0.6166015840 -0.1176687865
#> 220   0.729421382 -0.5059200176  0.2014080073
#> 221   0.323027389 -0.8717531865 -0.1412601138
#> 222   0.472047325 -0.3962825897 -0.0453881979
#> 223   0.397391943 -0.7129855438  0.0132751773
#> 224   0.807897808 -0.0412132001  0.7779760793
#> 225   0.669437530 -0.5646352736 -0.1491879479
#> 226   0.108277584 -0.4906887395 -0.1290319741
#> 227   0.558125200 -0.5932552569 -0.1266574887
#> 228   0.122377088 -0.5964236790 -0.0321425038
#> 229   0.429301348 -0.2389244962  0.3440752741
#> 230   0.774552866 -0.0073385684  0.3348274345
#> 231   0.558125200 -0.5932552569 -0.1266574887
#> 232   0.785807440 -0.2348240754  0.2184634849
#> 233   0.588573123 -0.5722319774 -0.3093570027
#> 234   0.306964764 -0.7840284346 -0.2674382861
#> 235   0.421985847 -0.0115256480  0.6095580777
#> 236   0.657833726 -0.5436154694  1.0753539448
#> 237   0.368325324 -0.6207678688  0.6354332680
#> 238   0.562736524 -0.3754320839  0.1735595175
#> 239   0.371297983 -0.2393095559 -0.1762307018
#> 240   0.876206746 -0.0939347728  0.2328133671
#> 241   0.505607231 -0.4985081768  0.4959870777
#> 242   0.734123417 -0.2267625232 -0.0691651203
#> 243   0.589080118 -0.2685322342  0.5058685150
#> 244   0.581292121 -0.1933829741  0.3587734341
#> 245   0.516392048 -0.1501869327  0.5603172645
#> 246   0.574447770 -0.4359195219 -0.1549239484
#> 247   0.639525260 -0.4080283523  0.3554821128
#> 248   0.939525131 -0.0172080395  0.5105003030
#> 249   0.645894488 -0.1268588095  0.7516447565
#> 250   0.470857068 -0.7012752901 -0.0533944483
#> 251   0.556717645 -0.0251348861  0.5587452385
#> 252   0.536340709 -0.5396689831 -0.1031563429
#> 253   0.270151604 -0.4499442075  0.0617140453
#> 254   0.246183374 -0.4990790881  0.2673362926
#> 255   0.532783979 -0.7035839040  0.4237405259
#> 256   0.395573782 -0.5718703160  0.4420127687
#> 257   0.481049275 -0.6559345549 -0.3337964889
#> 258   0.301988041 -0.8116186751 -0.3280006498
#> 259   0.465280564 -0.8706595991 -0.1487998034
#> 260   0.408673934  0.0419123925  0.7019354495
#> 261   0.352100755 -0.1984750363  1.0737858020
#> 262   0.807897808 -0.0412132001  0.7779760793
#> 263   0.528498362 -0.4419003697  0.0952939115
#> 264   0.610246316 -0.1860627135  0.4337096531
#> 265   0.644880526 -0.6166015840 -0.1176687865
#> 266   0.730247532 -0.0722159148  0.4691592867
#> 267   0.674848866 -0.1995803634  0.2130140496
#> 268   0.392660562 -0.5127695111 -0.1550251785
#> 269   0.491197064 -0.0847027696  0.6029650930
#> 270   0.804483325 -0.2503472868  0.1463935698
#> 271   0.323027389 -0.8717531865 -0.1412601138
#> 272   0.527590926 -0.3786855055  0.0842204892
#> 273   0.381813533 -0.6617286529 -0.0859965480
#> 274   0.589080118 -0.2685322342  0.5058685150
#> 275   0.536340709 -0.5396689831 -0.1031563429
#> 276   0.725903823 -0.1800053211  0.2857344701
#> 277   0.702648989  0.0801524848  0.6705259113
#> 278   0.583074813 -0.4261627843  0.2454969749
#> 279   0.505607231 -0.4985081768  0.4959870777
#> 280   0.525382404 -0.8405352664 -0.2096030585
#> 281   0.167392354 -0.4601450850 -0.0471884460
#> 282   0.698393748 -0.2646586256  0.4288310627
#> 283   0.410480046 -0.7539354910 -0.3976715506
#> 284   0.477380758 -0.5764138300 -0.1984008386
#> 285   0.674062889 -0.3269624188  0.3076873928
#> 286   0.520385335 -0.6450328646 -0.2310408704
#> 287   0.491518914 -0.3144624345 -0.0419174778
#> 288   0.622554753 -0.4631156454 -0.0203813983
#> 289   0.697620813 -0.1081130362  0.7089193134
#> 290   0.560027587 -0.7393498046  0.0150354612
#> 291   0.323027389 -0.8717531865 -0.1412601138
#> 292   0.657195360 -0.4731286483  0.0495742550
#> 293   0.104107772 -0.8267930698  0.2208496268
#> 294   0.432887938 -0.6735779583 -0.2924426378
#> 295   0.468876217 -0.7623249658 -0.5006200387
#> 296   0.413899814 -0.9316270055 -0.2657991690
#> 297   0.687766986 -0.1744295809  0.1795048277
#> 298   0.588757333 -0.1690655741  0.4176794292
#> 299   0.532783979 -0.7035839040  0.4237405259
#> 300   0.833167253 -0.1898677425  0.2502804779
#> 301   0.347702645 -0.5324314316 -0.1203097888
#> 302   0.453430627  0.0294929443  0.5822793441
#> 303   0.319685363 -0.0669331895  0.3420955764
#> 304   0.697170215 -0.3207257006  0.2505779678
#> 305   0.076828766 -0.2879429414  0.7969695896
#> 306   0.591799827 -0.5214798767 -0.0200362404
#> 307   0.338158886 -0.7291699835 -0.1507744529
#> 308   0.392152164 -0.7059789514 -0.1980735628
#> 309   0.277683065 -0.2907208549  0.0281656679
#> 310   0.817114376 -0.0170463170  1.1503714554
#> 311   0.618093752 -0.0999739095  0.7011603195
#> 312   0.639525260 -0.4080283523  0.3554821128
#> 313   0.522304299 -0.5863683365  0.1231520968
#> 314   0.288508453 -0.9425838614  0.4964016766
#> 315   0.665371416 -0.1999262764  0.3430625803
#> 316   0.402837826 -0.3194836650 -0.1002610449
#> 317   0.419439372 -0.3715063372  0.2704695862
#> 318   0.339695740  0.2595097197  1.1744083125
#> 319   0.645164420 -0.6030845721 -0.3922604042
#> 320   0.274281253 -0.8605212265  0.0321174604
#> 321   0.500657439 -0.5747495210 -0.1204946640
#> 322   0.462701696 -0.4281632969 -0.0147505513
#> 323   0.488131128 -0.6414815386 -0.5540853349
#> 324   0.272367621 -0.7398832923 -0.1146232394
#> 325   0.401643305 -0.8449718885  0.1228071811
#> 326   0.588573123 -0.5722319774 -0.3093570027
#> 327   0.328712758 -0.5555038644 -0.0446441700
#> 328   0.458082582 -0.3045358157  0.2935757906
#> 329   0.684208818 -0.1188061151  0.3432462985
#> 330   0.674062889 -0.3269624188  0.3076873928
#> 331   0.267986819 -0.5003097674  0.1321731877
#> 332   0.929393572  0.0498340878  0.6505997200
#> 333   0.469150502 -0.1664647073  0.5398124647
#> 334   0.616990258 -0.4983052106  0.2313510359
#> 335   0.517729812 -0.7246092841 -0.2274196621
#> 336   0.602818315 -0.4270626748  0.4848650341
#> 337   0.610246316 -0.1860627135  0.4337096531
#> 338   0.535981993 -0.2486159456  0.6553858636
#> 339   0.385867843 -0.6747857895  0.3751614879
#> 340   0.642170860  0.0965705575  0.4854810120
#> 341   0.548082399 -0.1853274301  0.0859794429
#> 342   0.663337897 -0.1044994685  0.3748580912
#> 343   0.407212450 -1.0912330861 -0.7624121405
#> 344   0.528442569 -0.6526586024  0.6057006421
#> 345   0.904778967  0.0296756959  0.7565975846
#> 346   0.410480046 -0.7539354910 -0.3976715506
#> 347   0.465154629 -0.7307856272 -0.1620789431
#> 348   0.449543625 -0.3228331516  0.4156916002
#> 349   0.210559914 -0.5094163850 -0.0761820811
#> 350   0.605734330 -0.2233366348  0.7011156505
#> 351   0.284002293 -0.8833362685 -0.1975182652
#> 352   0.622554753 -0.4631156454 -0.0203813983
#> 353   0.538076485 -0.4017223189  0.5092397823
#> 354   0.510612301 -0.3902438063  0.3496449006
#> 355   0.338571214 -0.6269812518  0.4258033072
#> 356   0.392152164 -0.7059789514 -0.1980735628
#> 357   0.531515759  0.2258983145  1.1159692407
#> 358   0.492459107 -0.7812901676 -0.3082812756
#> 359   0.680957131 -0.3978890187  0.2411645751
#> 360   0.272367621 -0.7398832923 -0.1146232394
#> 361   0.466480646 -0.5580765499  0.1352134014
#> 362   0.639759303 -0.2424789254  0.6590879831
#> 363   0.256564035 -0.8976261037 -0.1356577280
#> 364   0.167635725 -0.3630207013  0.0808469057
#> 365   0.241178272 -0.5524596103 -0.2549310537
#> 366   0.591799827 -0.5214798767 -0.0200362404
#> 367  -0.670733099 -0.2519980590  0.7597303056
#> 368   0.587098113 -0.5366719140  0.3834410678
#> 369   0.610246316 -0.1860627135  0.4337096531
#> 370   0.581501050 -0.3293810457 -0.0472220054
#> 371   0.312342450 -0.9142750510 -0.3858332044
#> 372   0.432887938 -0.6735779583 -0.2924426378
#> 373   0.370387992 -0.4418381120 -0.3437525975
#> 374   0.164408691 -0.5713865823 -0.0282953633
#> 375   0.402463061 -0.9010098581  0.4687659938
#> 376   0.684208818 -0.1188061151  0.3432462985
#> 377   0.532783979 -0.7035839040  0.4237405259
#> 378   0.566822925  0.2620748333  0.7426129395
#> 379   0.621275263 -0.1504336522  0.5449235100
#> 380   0.434162269 -0.1380952683  0.2698538793
#> 381   0.691651403 -0.4028217874  0.2339681067
#> 382   0.660500498  0.1199449497  1.0905322441
#> 383  -0.073336670 -0.1023191636 -0.1678809732
#> 384   0.752331820 -0.1084064761  0.4378842849
#> 385   0.289348004 -0.7501651485 -0.3192711152
#> 386   0.413899814 -0.9316270055 -0.2657991690
#> 387   0.315262395 -0.6464464354 -0.1430879655
#> 388   0.946745705 -0.0635777344  0.2111267830
#> 389   0.536340709 -0.5396689831 -0.1031563429
#> 390   0.591657423 -0.4847878407 -0.0476600846
#> 391   0.095285419 -0.5107157942  0.6557267635
#> 392   0.657195360 -0.4731286483  0.0495742550
#> 393   0.665371416 -0.1999262764  0.3430625803
#> 394   0.407399548 -0.3267445054  0.1982631094
#> 395   0.685903377 -0.4500405912 -0.3315394269
#> 396   0.555670645 -0.4245912238  0.3808923244
#> 397  -0.247992530 -1.0222162463  0.0577241627
#> 398   0.538076485 -0.4017223189  0.5092397823
#> 399   0.623973197  0.0524165500  0.6011384278
#> 400   0.694080726 -0.4885010854  0.0295663880
#> 401   0.257259443 -0.7542340669 -0.1890850451
#> 402   0.579006445 -0.2799553132  0.2410091812
#> 403   0.654701133 -0.0433314184  0.3506233672
#> 404   0.402786455 -0.6608291423 -0.3508473313
#> 405   0.579006445 -0.2799553132  0.2410091812
#> 406   0.736998118 -0.2018309371  0.4542345700
#> 407   0.452467474 -0.1784820010  0.4596974630
#> 408   0.035057585  0.1354902801  0.8358830059
#> 409   0.470857068 -0.7012752901 -0.0533944483
#> 410   0.621275263 -0.1504336522  0.5449235100
#> 411   0.589080118 -0.2685322342  0.5058685150
#> 412   0.684208818 -0.1188061151  0.3432462985
#> 413   0.663810696 -0.5855056230  0.0519867840
#> 414   0.653998535 -0.3959659836  0.4507163848
#> 415   0.314996538 -0.1744875602  0.4885058256
#> 416   0.401643305 -0.8449718885  0.1228071811
#> 417   0.349287267 -0.8461984521  0.3441944278
#> 418   0.588573123 -0.5722319774 -0.3093570027
#> 419   0.457784887 -0.5173326337  0.4564668523
#> 420   0.585650672 -0.5201446497 -0.2474993570
#> 421   0.603979971 -0.3473924941  0.3619800010
#> 422   0.933604954  0.1194894254  0.8940527786
#> 423   0.399213789 -1.2008687423 -0.5623802136
#> 424   0.177356952 -0.2415943695  0.4826777447
#> 425   0.544485818 -0.7940140376 -0.1421224748
#> 426   0.675311541 -0.0777936693  0.6716356933
#> 427   0.536340709 -0.5396689831 -0.1031563429
#> 428   0.407212450 -1.0912330861 -0.7624121405
#> 429   0.578834969 -0.6785538316  0.1530313621
#> 430   0.252045602 -0.1515315228  0.5890552544
#> 431   0.527590926 -0.3786855055  0.0842204892
#> 432   0.898678701 -0.0610141755  0.8619640118
#> 433   0.591657423 -0.4847878407 -0.0476600846
#> 434   0.652424159 -0.3146580890  0.0980775675
#> 435   0.433466574 -0.8250102022 -0.3472046272
#> 436   0.373223944 -0.7716511902 -0.0665658750
#> 437   0.678219293 -0.2237884238  0.3211206868
#> 438   0.465280564 -0.8706595991 -0.1487998034
#> 439   0.536340709 -0.5396689831 -0.1031563429
#> 440   0.401643305 -0.8449718885  0.1228071811
#> 441   0.273906118 -0.3885641604  0.2082639643
#> 442   0.194679300 -0.3759818207  0.0834623471
#> 443   0.759083818  0.1016986566  0.7734819130
#> 444   0.397391943 -0.7129855438  0.0132751773
#> 445   0.605734330 -0.2233366348  0.7011156505
#> 446   0.654304344 -0.3935799893 -0.1632488727
#> 447   0.562736524 -0.3754320839  0.1735595175
#> 448   0.785807440 -0.2348240754  0.2184634849
#> 449   0.416396108 -0.5511128913 -0.1727096988
#> 450   0.453883098 -0.6201890579  0.2865912648
#> 451   0.808513994 -0.1933859208  0.2626094252
#> 452   0.685903377 -0.4500405912 -0.3315394269
#> 453   0.631904840 -0.3981817205  0.1307246569
#> 454   0.043172889 -0.6408377806 -0.3432323917
#> 455   0.560545465 -0.7264377439 -0.1610015840
#> 456   0.170224059 -0.6571227311 -0.1335722354
#> 457   0.466480646 -0.5580765499  0.1352134014
#> 458   0.272244239 -0.3232586995 -0.1025886003
#> 459   0.500657439 -0.5747495210 -0.1204946640
#> 460   0.284246286 -0.9143444520 -0.2475787869
#> 461   0.463256770 -0.2437855381  0.0967967867
#> 462   0.560545465 -0.7264377439 -0.1610015840
#> 463   0.611148515 -0.4980743716 -0.1933281665
#> 464   0.454307592 -0.8972618251  0.1278634208
#> 465   0.589080118 -0.2685322342  0.5058685150
#> 466   0.687766986 -0.1744295809  0.1795048277
#> 467   0.482123357 -0.4410708328 -0.1082549750
#> 468   0.250265403 -0.0641320189  0.5660519159
#> 469   0.639525260 -0.4080283523  0.3554821128
#> 470   0.395857968 -0.0191022639  0.4792134611
#> 471   0.289225580 -0.7339787680  0.0204835628
#> 472   0.415357039 -0.4989053520 -0.0545371188
#> 473   0.697170215 -0.3207257006  0.2505779678
#> 474   0.494941099 -0.0360353620  0.5034048667
#> 475   0.038005960 -0.7769977631  0.0397904449
#> 476   0.678219293 -0.2237884238  0.3211206868
#> 477   0.408673934  0.0419123925  0.7019354495
#> 478   0.734123417 -0.2267625232 -0.0691651203
#> 479   0.514067852  0.1067177773  0.4714774555
#> 480   0.284246286 -0.9143444520 -0.2475787869
#> 481   0.898678701 -0.0610141755  0.8619640118
#> 482   0.489287397 -0.3939663196  0.2726586813
#> 483   0.697170215 -0.3207257006  0.2505779678
#> 484   0.374288089 -0.6214921678 -0.0346210045
#> 485   0.033086279 -0.3741044785  0.1628108274
#> 486   0.742933727 -0.4000697522  0.0191115481
#> 487   0.657195360 -0.4731286483  0.0495742550
#> 488   0.687147345 -0.4511434243 -0.2804474807
#> 489   0.520489739 -0.2472146679  0.6667194644
#> 490   0.769782586 -0.1686125625  0.6506453126
#> 491   0.686204268 -0.3094016773  0.0464832512
#> 492   0.730247532 -0.0722159148  0.4691592867
#> 493   0.675311541 -0.0777936693  0.6716356933
#> 494   0.815693725 -0.3278289705  0.0488447426
#> 495   0.669437530 -0.5646352736 -0.1491879479
#> 496   0.522655275 -0.3260064198  0.6905237658
#> 497   0.588510267 -0.7169186403  0.0075476769
#> 498   0.217707606 -1.0923886999 -0.0862845527
#> 499   0.904803499  0.1328547294  1.1375854587
#> 500   0.416943214 -0.4307702534  0.1287895435
#> 501   0.343988237 -0.9046336017 -0.2727816058
#> 502   0.465280564 -0.8706595991 -0.1487998034
#> 503   0.702938702 -0.3353722047  0.1331238216
#> 504   0.558125200 -0.5932552569 -0.1266574887
#> 505   0.616990258 -0.4983052106  0.2313510359
#> 506   0.482123357 -0.4410708328 -0.1082549750
#> 507   0.481049275 -0.6559345549 -0.3337964889
#> 508   0.768298343  0.0964523655  0.6873290244
#> 509   0.296566597 -0.6756576771  0.1025375120
#> 510   0.168076913 -0.7074389692 -0.2533546140
#> 511   0.621275263 -0.1504336522  0.5449235100
#> 512   0.645164420 -0.6030845721 -0.3922604042
#> 513   0.381813533 -0.6617286529 -0.0859965480
#> 514   0.809858463 -0.3660113873 -0.1276043982
#> 515   0.410480046 -0.7539354910 -0.3976715506
#> 516   0.175073730 -0.2043978649  0.5063663702
#> 517   0.808513994 -0.1933859208  0.2626094252
#> 518   0.738994509 -0.3456147424  0.2305001090
#> 519   0.694080726 -0.4885010854  0.0295663880
#> 520   0.171417922 -0.8975780351 -0.1921506445
#> 521   0.027245070 -0.2969614736  0.2053150242
#> 522   0.397391943 -0.7129855438  0.0132751773
#> 523   0.449543625 -0.3228331516  0.4156916002
#> 524   0.515323402 -0.3263329025  0.4791680315
#> 525   0.871887568 -0.3450877243 -0.1158360705
#> 526   0.236092975 -0.6676559228  0.4374049081
#> 527   0.517029234 -0.3685395351  0.1059089952
#> 528   0.520385335 -0.6450328646 -0.2310408704
#> 529   0.282980585 -1.0196573689 -0.4798614281
#> 530   0.536525205 -0.5428362715  0.0786830139
#> 531   0.657195360 -0.4731286483  0.0495742550
#> 532   0.346556695 -0.4449005735  0.7535324919
#> 533   0.623973197  0.0524165500  0.6011384278
#> 534   0.488428364 -0.5695050155 -0.4502866480
#> 535   0.572300042 -0.1696826331  0.6170550928
#> 536   0.734123417 -0.2267625232 -0.0691651203
#> 537   0.466699894 -0.6966513181  0.1089128478
#> 538   0.348947888 -0.4700803449  0.0659565168
#> 539   0.465154629 -0.7307856272 -0.1620789431
#> 540   0.660500498  0.1199449497  1.0905322441
#> 541   0.536525205 -0.5428362715  0.0786830139
#> 542   0.678219293 -0.2237884238  0.3211206868
#> 543   0.521209825  0.0146763749  0.7320883630
#> 544  -0.027973582 -0.1380653803  0.4876286162
#> 545   0.763996556  0.0079128300  0.3777336245
#> 546   0.147603032 -0.5886716652  0.2698118169
#> 547   0.506731218 -0.1057660348  0.1066274141
#> 548   0.631904840 -0.3981817205  0.1307246569
#> 549   0.693496018 -0.0022729054  0.4342803054
#> 550   0.306964764 -0.7840284346 -0.2674382861
#> 551   0.665371416 -0.1999262764  0.3430625803
#> 552   0.478925528 -0.5567453820  0.1756855069
#> 553   0.629474503 -0.2106696909  0.3533249641
#> 554   0.684181135 -0.1948213194  0.5353497196
#> 555   0.544485818 -0.7940140376 -0.1421224748
#> 556   0.416396108 -0.5511128913 -0.1727096988
#> 557   0.429422361 -0.1768137734  0.2841671951
#> 558   0.175354875 -0.1618084653  0.2998456418
#> 559   0.512456472 -0.2474317176  0.2893920438
#> 560   0.059022661 -0.7044381255  0.0149381661
#> 561   0.603979971 -0.3473924941  0.3619800010
#> 562   0.512456472 -0.2474317176  0.2893920438
#> 563   0.330532220 -0.9992477517 -0.3728109851
#> 564   0.564117418 -0.4800938994  0.5678678935
#> 565   0.609627642 -0.2430697440  0.2933453356
#> 566   0.481588092 -0.0713767347  0.7979786076
#> 567   0.894095454 -0.0319822433  0.5434940593
#> 568   0.174533000 -0.2001064031  0.0481308289
#> 569   0.785807440 -0.2348240754  0.2184634849
#> 570   0.234193165 -0.3488187259  0.0569303957
#> 571   0.267139840 -0.6835784954  0.0500371798
#> 572   0.202115229 -0.0416151234  0.5139412571
#> 573   0.590119334 -0.3397513282  0.4611017137
#> 574   0.795150471 -0.3730309717  0.1328327254
#> 575   0.252500888 -0.9238692993 -0.0663791813
#> 576   0.697170215 -0.3207257006  0.2505779678
#> 577   0.298255081 -0.5110643830  0.2251713729
#> 578  -0.002422997 -0.3730502508 -0.0124539351
#> 579   0.725903823 -0.1800053211  0.2857344701
#> 580   0.556717645 -0.0251348861  0.5587452385
#> 581   0.719112671 -0.0675928401  0.3555808645
#> 582   0.170499736 -0.6237989593 -0.3771765035
#> 583   0.306964764 -0.7840284346 -0.2674382861
#> 584   0.273592516 -0.7767521286 -0.0204619562
#> 585   0.562766156 -0.1706411329  0.2236790431
#> 586   0.395573782 -0.5718703160  0.4420127687
#> 587   0.347856422 -0.2484288569  0.1360828005
#> 588   0.298255081 -0.5110643830  0.2251713729
#> 589   0.500657439 -0.5747495210 -0.1204946640
#> 590   0.027682346 -0.0453573528  0.4964859004
#> 591   0.598304069 -0.2895384670  0.2036346123
#> 592   0.598304069 -0.2895384670  0.2036346123
#> 593   0.741139550 -0.5074998759  0.0464920276
#> 594   0.407212450 -1.0912330861 -0.7624121405
#> 595   0.624598593 -0.6598360291  0.0982058587
#> 596   0.562893493 -0.2266771303  0.2395295164
#> 597   0.268291608 -0.2905291363  0.7870757472
#> 598   0.940000784  0.0166907794  0.5198577634
#> 599   0.577857272 -0.4809237716  0.5418112635
#> 600   0.474012544 -0.2199027205  0.4862187145
#> 601   0.686823053 -0.3367375570 -0.0003031393
#> 602   0.560027587 -0.7393498046  0.0150354612
#> 603   0.803703663 -0.1679292206  0.3443489539
#> 604   0.536525205 -0.5428362715  0.0786830139
#> 605   0.528498362 -0.4419003697  0.0952939115
#> 606   0.500657439 -0.5747495210 -0.1204946640
#> 607   0.844404683 -0.2231441912  0.2455988990
#> 608   0.698696637 -0.4808382743  0.1725247175
#> 609   0.593551427 -0.5067191040 -0.1571875980
#> 610   0.645164420 -0.6030845721 -0.3922604042
#> 611   0.447318165 -0.7575562894 -0.3961136656
#> 612   0.087529265 -0.0969955507  0.3188764135
#> 613   0.651336428  0.0004473797  0.8042997415
#> 614   0.448331278 -0.7699793348  0.0959832806
#> 615   0.452054549 -0.4023036795 -0.1903799127
#> 616   0.629474503 -0.2106696909  0.3533249641
#> 617   0.627111795 -0.7062635255  0.5695974784
#> 618   0.585650672 -0.5201446497 -0.2474993570
#> 619   0.605055322 -0.5322610153 -0.3150155113
#> 620   0.536525205 -0.5428362715  0.0786830139
#> 621   0.446231244 -0.5845835171  0.2900176672
#> 622   0.510612301 -0.3902438063  0.3496449006
#> 623   0.419034155 -0.5461234509  0.2001759020
#> 624   0.730247532 -0.0722159148  0.4691592867
#> 625   0.568947270 -0.0037942421  0.1944902615
#> 626   0.536340709 -0.5396689831 -0.1031563429
#> 627   0.684181135 -0.1948213194  0.5353497196
#> 628   0.687766986 -0.1744295809  0.1795048277
#> 629   0.111909424 -0.3794173617  0.1743290424
#> 630   0.470857068 -0.7012752901 -0.0533944483
#> 631   0.597500360 -0.5132245184  0.1778545384
#> 632   0.500657439 -0.5747495210 -0.1204946640
#> 633   0.562736524 -0.3754320839  0.1735595175
#> 634   0.321129642 -0.1589595436  0.3357257963
#> 635   0.572300042 -0.1696826331  0.6170550928
#> 636   0.205413835 -0.4200271383  0.1835644491
#> 637   0.591799827 -0.5214798767 -0.0200362404
#> 638   0.492459107 -0.7812901676 -0.3082812756
#> 639   0.552037204 -0.1138421777  0.2499620272
#> 640   0.609627642 -0.2430697440  0.2933453356
#> 641   0.421012186 -0.4418044818  0.5253837157
#> 642   0.572300042 -0.1696826331  0.6170550928
#> 643   0.325489212 -0.1349528418  0.4620437098
#> 644   0.806153018  0.2563447510  0.7583404309
#> 645   0.402786455 -0.6608291423 -0.3508473313
#> 646   0.575259811 -0.4392507702  0.4314924072
#> 647   0.556717645 -0.0251348861  0.5587452385
#> 648   0.662702054 -0.3143555196  0.1635569079
#> 649   0.489287397 -0.3939663196  0.2726586813
#> 650   0.402786455 -0.6608291423 -0.3508473313
#> 651   0.720856664 -0.1513000982 -0.0167185990
#> 652   0.374288089 -0.6214921678 -0.0346210045
#> 653   0.760262426 -0.3371682014 -0.0219582202
#> 654   0.645894488 -0.1268588095  0.7516447565
#> 655   0.171522789 -0.1903695470  0.3063764580
#> 656   0.272594411 -0.9332411047 -0.2349142149
#> 657   0.536525205 -0.5428362715  0.0786830139
#> 658   0.624598593 -0.6598360291  0.0982058587
#> 659   0.240456965 -0.4720554658  0.2472364109
#> 660   0.654701133 -0.0433314184  0.3506233672
#> 661   0.456915272 -0.4835954518  0.5371498164
#> 662   0.555800904 -0.3108431257  0.3983919466
#> 663   0.172903430 -0.7801752258 -0.0251037661
#> 664   0.650580791 -0.1375828466 -0.0375821722
#> 665   0.731608138  0.0329892857  0.2872646388
#> 666   0.282980585 -1.0196573689 -0.4798614281
#> 667   0.464903082 -0.3317178903  0.1999157390
#> 668   0.452467474 -0.1784820010  0.4596974630
#> 669   0.125048061 -0.3889889134  0.9938137263
#> 670   0.375276407 -0.1571292215  0.1890324689
#> 671   0.833167253 -0.1898677425  0.2502804779
#> 672   0.295725849 -0.3208077151  0.6034699976
#> 673   0.605734330 -0.2233366348  0.7011156505
#> 674   0.252500888 -0.9238692993 -0.0663791813
#> 675   0.467269986 -0.0830655622  0.7464704477
#> 676   0.342610009 -0.5090802046  0.1352897750
#> 677   0.528498362 -0.4419003697  0.0952939115
#> 678   0.753195634 -0.3088516806  0.2159745071
#> 679   0.528498362 -0.4419003697  0.0952939115
#> 680   0.870197982 -0.0616065924  0.6103987386
#> 681   0.769847112 -0.2751131112  0.3641180521
#> 682   0.416943214 -0.4307702534  0.1287895435
#> 683   0.759110391 -0.4428705464 -0.2819422482
#> 684   0.401643305 -0.8449718885  0.1228071811
#> 685   0.536525205 -0.5428362715  0.0786830139
#> 686   0.710098254  0.0882260982  0.6885070519
#> 687   0.402216686  0.0861020719  0.2831909166
#> 688   0.515323402 -0.3263329025  0.4791680315
#> 689   0.597500360 -0.5132245184  0.1778545384
#> 690   0.131041553 -0.5454202565 -0.0221906085
#> 691   0.492459107 -0.7812901676 -0.3082812756
#> 692   0.502653327 -0.2256884178  0.4189055560
#> 693   0.288557568 -0.7975681029 -0.0204031299
#> 694   0.833167253 -0.1898677425  0.2502804779
#> 695   0.146860003 -0.1579344308  0.6409481980
#> 696   0.429301348 -0.2389244962  0.3440752741
#> 697   0.492459107 -0.7812901676 -0.3082812756
#> 698   0.652818412  0.1260472232  0.7752869101
#> 699   0.603979971 -0.3473924941  0.3619800010
#> 700   0.491230322 -0.2029236489  0.4834774071
#> 701   0.918337593  0.0048262394  0.4565741613
#> 702   0.330532220 -0.9992477517 -0.3728109851
#> 703   0.277042249 -0.5171216763  0.4945059108
#> 704   0.596799985 -0.3875585591  0.2931267203
#> 705   0.366595760  0.0309824615  0.4612996437
#> 706   0.527590926 -0.3786855055  0.0842204892
#> 707   0.591657423 -0.4847878407 -0.0476600846
#> 708   0.817114376 -0.0170463170  1.1503714554
#> 709   0.770009274  0.0394608417  0.6567141262
#> 710   0.770061967 -0.4345266763 -0.2988973579
#> 711   0.149309717 -0.8804546325 -0.3804730303
#> 712   0.662702054 -0.3143555196  0.1635569079
#> 713   0.251251122 -0.2204147899  0.1458340053
#> 714   0.520489739 -0.2472146679  0.6667194644
#> 715   0.662702054 -0.3143555196  0.1635569079
#> 716   0.687147345 -0.4511434243 -0.2804474807
#> 717   0.267139840 -0.6835784954  0.0500371798
#> 718   0.343988237 -0.9046336017 -0.2727816058
#> 719   0.413899814 -0.9316270055 -0.2657991690
#> 720   0.249013001 -0.5668000980  0.0662146320
#> 721   0.520385335 -0.6450328646 -0.2310408704
#> 722   0.474012544 -0.2199027205  0.4862187145
#> 723   0.284246286 -0.9143444520 -0.2475787869
#> 724   0.572300042 -0.1696826331  0.6170550928
#> 725   0.583074813 -0.4261627843  0.2454969749
#> 726   0.277042249 -0.5171216763  0.4945059108
#> 727   0.752858352 -0.3379186784  0.3363154173
#> 728   0.522603666 -0.4364880541  0.0846083449
#> 729   0.585650672 -0.5201446497 -0.2474993570
#> 730   0.389535392 -0.2281267365  0.9080321640
#> 731   0.774552866 -0.0073385684  0.3348274345
#> 732   0.456665400 -0.4148207717 -0.0547547775
#> 733   0.496786862 -0.6414704437  0.6565489321
#> 734   0.752858352 -0.3379186784  0.3363154173
#> 735   0.081946339 -0.6966007147  0.0884504870
#> 736   0.769782586 -0.1686125625  0.6506453126
#> 737   0.804483325 -0.2503472868  0.1463935698
#> 738  -0.115234676 -0.4903953671  0.5838896132
#> 739   0.639525260 -0.4080283523  0.3554821128
#> 740   0.933604954  0.1194894254  0.8940527786
#> 741   0.590119334 -0.3397513282  0.4611017137
#> 742   0.718837058 -0.0786315190  0.7932792965
#> 743   0.718837058 -0.0786315190  0.7932792965
#> 744   0.253451488 -0.5435863832 -0.4053643691
#> 745   0.770009274  0.0394608417  0.6567141262
#> 746   0.759083818  0.1016986566  0.7734819130
#> 747   0.438094416 -0.5295589485  0.2709644714
#> 748   0.347702645 -0.5324314316 -0.1203097888
#> 749   0.559212271 -0.3353427437  0.1128377420
#> 750   0.237871104 -0.6028036009  0.4747265952
#> 751   0.447318165 -0.7575562894 -0.3961136656
#> 752   0.151793727 -1.1302943011 -0.3311175380
#> 753   0.892215540 -0.0878459660 -0.0285618820
#> 754   0.769782586 -0.1686125625  0.6506453126
#> 755   0.645164420 -0.6030845721 -0.3922604042
#> 756   0.500657439 -0.5747495210 -0.1204946640
#> 757   0.605055322 -0.5322610153 -0.3150155113
#> 758   0.544485818 -0.7940140376 -0.1421224748
#> 759   0.656085948 -0.3277176693 -0.0968727406
#> 760   0.555670645 -0.4245912238  0.3808923244
#> 761   0.301313635 -0.4656013816 -0.0188831513
#> 762   0.348947888 -0.4700803449  0.0659565168
#> 763   0.807897808 -0.0412132001  0.7779760793
#> 764   0.407212450 -1.0912330861 -0.7624121405
#> 765   0.433935749 -0.3568793547 -0.2829107011
#> 766   0.814031579 -0.3373251752  0.3076491659
#> 767   0.605734330 -0.2233366348  0.7011156505
#> 768   0.330532220 -0.9992477517 -0.3728109851
#> 769   0.628534000 -0.6357460902  0.2621836738
#> 770   0.505607231 -0.4985081768  0.4959870777
#> 771   0.562736524 -0.3754320839  0.1735595175
#> 772   0.674062889 -0.3269624188  0.3076873928
#> 773   0.674848866 -0.1995803634  0.2130140496
#> 774   0.779787759 -0.1164619613  0.1437038880
#> 775   0.427784806 -0.7726107181  0.3243965594
#> 776   0.338158886 -0.7291699835 -0.1507744529
#> 777   0.605734330 -0.2233366348  0.7011156505
#> 778   0.464903082 -0.3317178903  0.1999157390
#> 779   0.537309489 -0.2144362185  0.7130372819
#> 780   0.149309717 -0.8804546325 -0.3804730303
#> 781   0.558125200 -0.5932552569 -0.1266574887
#> 782   0.752858352 -0.3379186784  0.3363154173
#> 783   0.330532220 -0.9992477517 -0.3728109851
#> 784   0.665371416 -0.1999262764  0.3430625803
#> 785   0.808513994 -0.1933859208  0.2626094252
#> 786   0.590119334 -0.3397513282  0.4611017137
#> 787   0.342610009 -0.5090802046  0.1352897750
#> 788   0.241178272 -0.5524596103 -0.2549310537
#> 789   0.694321760 -0.0863740773  0.4165528615
#> 790   0.854474253  0.0915562378  0.3862174539
#> 791   0.179922990 -0.5700697392 -0.2612464337
#> 792   0.406783420  0.1644128313  0.9123704238
#> 793   0.669437530 -0.5646352736 -0.1491879479
#> 794   0.517729812 -0.7246092841 -0.2274196621
#> 795   0.609478948 -0.1243873861  0.7515703360
#> 796   0.558125200 -0.5932552569 -0.1266574887
#> 797   0.449543625 -0.3228331516  0.4156916002
#> 798   0.528277709 -0.2061848132  0.7754128103
#> 799   0.666471264 -0.4254874872  0.5029643444
#> 800   0.510612301 -0.3902438063  0.3496449006
#> 801   0.482123357 -0.4410708328 -0.1082549750
#> 802  -0.702718257 -1.0606433261 -0.5244804347
#> 803   0.455082038 -0.5002082173  0.7281425202
#> 804   0.472702920 -0.0644746461  0.2854822774
#> 805   0.583563548 -0.0218995081  0.1703976961
#> 806   0.770009274  0.0394608417  0.6567141262
#> 807   0.407212450 -1.0912330861 -0.7624121405
#> 808   0.804483325 -0.2503472868  0.1463935698
#> 809   0.413899814 -0.9316270055 -0.2657991690
#> 810   0.673228723 -0.5474619466 -0.1724927482
#> 811   0.585807567 -0.8143902490 -0.2409367288
#> 812   0.248986427 -1.1325082776 -0.3071458207
#> 813   0.410846421 -0.5988103166  0.0127316215
#> 814   0.759083818  0.1016986566  0.7734819130
#> 815   0.328600063 -1.1815729977 -0.7530961858
#> 816   0.769847112 -0.2751131112  0.3641180521
#> 817   0.514431599 -0.2992562024  0.1540125294
#> 818   0.332601918 -0.5180550896  0.4274471664
#> 819   0.344866101 -0.6806141811  0.3258019225
#> 820   0.097836965 -0.5657126921 -0.2784893043
#> 821   0.803703663 -0.1679292206  0.3443489539
#> 822   0.568947270 -0.0037942421  0.1944902615
#> 823   0.434697333 -0.6688966580 -0.0145753373
#> 824   0.736998118 -0.2018309371  0.4542345700
#> 825   0.605734330 -0.2233366348  0.7011156505
#> 826   0.478925528 -0.5567453820  0.1756855069
#> 827   0.597472574 -0.0322423695  0.3678340283
#> 828   0.605055322 -0.5322610153 -0.3150155113
#> 829   0.556717645 -0.0251348861  0.5587452385
#> 830   0.318375496 -0.6975985146  0.7748980752
#> 831   0.904778967  0.0296756959  0.7565975846
#> 832   0.583074813 -0.4261627843  0.2454969749
#> 833   0.301988041 -0.8116186751 -0.3280006498
#> 834   0.416943214 -0.4307702534  0.1287895435
#> 835   0.697170215 -0.3207257006  0.2505779678
#> 836   0.559212271 -0.3353427437  0.1128377420
#> 837   0.343548202 -0.5554889871  0.3187968977
#> 838   0.537309489 -0.2144362185  0.7130372819
#> 839   0.517729812 -0.7246092841 -0.2274196621
#> 840   0.811010098 -0.3366069419  0.1345965309
#> 841   0.693496018 -0.0022729054  0.4342803054
#> 842   0.492459107 -0.7812901676 -0.3082812756
#> 843   0.522655275 -0.3260064198  0.6905237658
#> 844   0.904778967  0.0296756959  0.7565975846
#> 845   0.594507470 -0.4153846406 -0.2326481100
#> 846   1.022047053  0.2812304348  0.4992084876
#> 847   0.574331059 -0.2241538070  0.1106412392
#> 848   0.610988105 -0.0315779674  0.3479972481
#> 849   0.319392183 -0.1512543275  0.4574933637
#> 850   0.628534000 -0.6357460902  0.2621836738
#> 851   0.273592516 -0.7767521286 -0.0204619562
#> 852   0.780857140 -0.1596013852  0.5037488254
#> 853   0.395132025  0.2861099371  0.7182717288
#> 854   0.687766986 -0.1744295809  0.1795048277
#> 855   0.570331929 -0.3194761321  0.2639507761
#> 856   0.477380758 -0.5764138300 -0.1984008386
#> 857   0.605234096 -0.2617282588  1.6421873962
#> 858   0.576781395 -0.4575097468 -0.2177860721
#> 859   0.288339646 -0.7742779253  0.0477430965
#> 860   0.334092000 -0.3779038264 -0.5259367163
#> 861   0.686204268 -0.3094016773  0.0464832512
#> 862   0.616990258 -0.4983052106  0.2313510359
#> 863   0.691651403 -0.4028217874  0.2339681067
#> 864   0.414482921 -0.5166498863  0.4733016120
#> 865   0.785807440 -0.2348240754  0.2184634849
#> 866   0.593551427 -0.5067191040 -0.1571875980
#> 867   0.410480046 -0.7539354910 -0.3976715506
#> 868   0.478925528 -0.5567453820  0.1756855069
#> 869   0.488131128 -0.6414815386 -0.5540853349
#> 870   0.691651403 -0.4028217874  0.2339681067
#> 871   0.728248541 -0.2913468093  0.8580562593
#> 872   0.466480646 -0.5580765499  0.1352134014
#> 873   0.698696637 -0.4808382743  0.1725247175
#> 874   0.211198960 -0.8622169679 -0.4641729673
#> 875   0.434697333 -0.6688966580 -0.0145753373
#> 876   0.807897808 -0.0412132001  0.7779760793
#> 877   0.719112671 -0.0675928401  0.3555808645
#> 878   0.547315019 -0.1314325234  0.9333921931
#> 879   0.533803409 -0.6204993838 -0.0267670499
#> 880   0.338158886 -0.7291699835 -0.1507744529
#> 881   0.597472574 -0.0322423695  0.3678340283
#> 882   0.464903082 -0.3317178903  0.1999157390
#> 883   0.419034155 -0.5461234509  0.2001759020
#> 884   0.528971410 -0.5174896290  0.2876276409
#> 885   0.457672728 -0.4224624292  0.5714255913
#> 886   0.532341784 -0.8267097747 -0.4872490353
#> 887   0.624598593 -0.6598360291  0.0982058587
#> 888   0.397391943 -0.7129855438  0.0132751773
#> 889   0.697170215 -0.3207257006  0.2505779678
#> 890   0.807897808 -0.0412132001  0.7779760793
#> 891   0.512456472 -0.2474317176  0.2893920438
#> 892   0.939525131 -0.0172080395  0.5105003030
#> 893   0.391186251 -0.6977422613 -0.0765389788
#> 894   0.240456965 -0.4720554658  0.2472364109
#> 895   0.591799827 -0.5214798767 -0.0200362404
#> 896   0.359422466 -0.3452159329  0.0345580656
#> 897   0.547307507 -0.2648833238  0.1971811122
#> 898   0.420172866 -0.2268676232  0.3083622396
#> 899   0.244062537 -0.1741964681  0.4389565209
#> 900   0.433466574 -0.8250102022 -0.3472046272
#> 901   0.517729812 -0.7246092841 -0.2274196621
#> 902   0.469150502 -0.1664647073  0.5398124647
#> 903   0.641874809  0.0628317146  0.7006256645
#> 904   0.698393748 -0.2646586256  0.4288310627
#> 905   0.433466574 -0.8250102022 -0.3472046272
#> 906   0.414482921 -0.5166498863  0.4733016120
#> 907   0.500657439 -0.5747495210 -0.1204946640
#> 908   0.560545465 -0.7264377439 -0.1610015840
#> 909   0.180156049 -0.4381688603 -0.1921772649
#> 910   0.328837937 -0.5620882160  0.0425337552
#> 911   0.438094416 -0.5295589485  0.2709644714
#> 912   0.359558941 -0.6572539005 -0.0086467380
#> 913   0.528498362 -0.4419003697  0.0952939115
#> 914   0.446231244 -0.5845835171  0.2900176672
#> 915   0.607620249 -0.2187979537  0.0515247533
#> 916   0.656085948 -0.3277176693 -0.0968727406
#> 917   0.325005447 -0.2148081007  0.5270248761
#> 918   0.451720146 -0.7673846123 -0.6620129948
#> 919   0.481733590 -0.3066480260 -0.1710173394
#> 920   0.741139550 -0.5074998759  0.0464920276
#> 921   0.787519247 -0.3710958566  0.6817245798
#> 922   0.492459107 -0.7812901676 -0.3082812756
#> 923   0.692394136 -0.5233258967 -0.1450490264
#> 924   0.308259988 -1.0527318048 -0.0833072487
#> 925   0.760262426 -0.3371682014 -0.0219582202
#> 926  -0.089940154 -0.5930545528 -0.3382996605
#> 927   0.644880526 -0.6166015840 -0.1176687865
#> 928   0.528498362 -0.4419003697  0.0952939115
#> 929   0.457672728 -0.4224624292  0.5714255913
#> 930   0.529343249 -0.5969728595  0.5182698410
#> 931   0.806438852 -0.3138086692  0.0096489959
#> 932   0.651336428  0.0004473797  0.8042997415
#> 933   0.481588092 -0.0713767347  0.7979786076
#> 934   0.626322657  0.3313825617  0.5777648848
#> 935   0.631904840 -0.3981817205  0.1307246569
#> 936   0.238040631 -0.5196477235  0.2891656508
#> 937   0.532341784 -0.8267097747 -0.4872490353
#> 938   0.512456472 -0.2474317176  0.2893920438
#> 939   0.433004971 -0.8951925425 -0.1675247852
#> 940   0.517029234 -0.3685395351  0.1059089952
#> 941   0.410846421 -0.5988103166  0.0127316215
#> 942   0.691651403 -0.4028217874  0.2339681067
#> 943  -0.069618910 -0.5034214985 -0.1373018797
#> 944   0.652818412  0.1260472232  0.7752869101
#> 945   0.759083818  0.1016986566  0.7734819130
#> 946   0.228180489 -0.1610871569  0.1647488520
#> 947   0.410480046 -0.7539354910 -0.3976715506
#> 948   0.694080726 -0.4885010854  0.0295663880
#> 949   0.606977145 -0.3346482519  0.1101031134
#> 950   0.161414567  0.1603809466  1.1176780659
#> 951   0.517029234 -0.3685395351  0.1059089952
#> 952   0.451909843 -0.3329682324  0.5239021938
#> 953   0.559212271 -0.3353427437  0.1128377420
#> 954   0.741139550 -0.5074998759  0.0464920276
#> 955   0.678219293 -0.2237884238  0.3211206868
#> 956   0.493075177 -0.4950746181  0.6921841363
#> 957   0.730247532 -0.0722159148  0.4691592867
#> 958   0.330532220 -0.9992477517 -0.3728109851
#> 959   0.768206703 -0.1490874141  0.7074006829
#> 960   0.611148515 -0.4980743716 -0.1933281665
#> 961   0.780857140 -0.1596013852  0.5037488254
#> 962   0.387484836 -0.4631240952 -0.3865704478
#> 963   0.454767307 -0.2404364367  0.8410652047
#> 964   0.589080118 -0.2685322342  0.5058685150
#> 965   0.991594942  0.0738285475  0.4827134830
#> 966   0.894095454 -0.0319822433  0.5434940593
#> 967   0.407162840 -0.2984657771  0.3915937030
#> 968   0.504631271 -0.2081012276  0.5340849911
#> 969   0.796058569 -0.1196986716  0.6321305640
#> 970   0.488882429 -0.5193379922  0.4697085801
#> 971   0.429301348 -0.2389244962  0.3440752741
#> 972   0.435000271 -0.5836262312 -0.2692582388
#> 973   0.736998118 -0.2018309371  0.4542345700
#> 974   0.301313635 -0.4656013816 -0.0188831513
#> 975   0.412119882 -0.3144410948  0.0857610668
#> 976   0.528971410 -0.5174896290  0.2876276409
#> 977   0.628534000 -0.6357460902  0.2621836738
#> 978   0.492459107 -0.7812901676 -0.3082812756
#> 979   0.929393572  0.0498340878  0.6505997200
#> 980   0.702938702 -0.3353722047  0.1331238216
#> 981   0.598304069 -0.2895384670  0.2036346123
#> 982   0.323027389 -0.8717531865 -0.1412601138
#> 983   0.502653327 -0.2256884178  0.4189055560
#> 984   0.212132841 -0.8114041507 -0.3289232355
#> 985   0.702938702 -0.3353722047  0.1331238216
#> 986   0.579914838 -0.5647093714 -0.4385272876
#> 987   0.493492802 -0.4308856335  0.0853502508
#> 988   0.665371416 -0.1999262764  0.3430625803
#> 989   0.940000784  0.0166907794  0.5198577634
#> 990   0.470857068 -0.7012752901 -0.0533944483
#> 991   0.374288089 -0.6214921678 -0.0346210045
#> 992   0.768206703 -0.1490874141  0.7074006829
#> 993   0.438094416 -0.5295589485  0.2709644714
#> 994   0.205413835 -0.4200271383  0.1835644491
#> 995   1.000449256  0.0327683183  1.2732107628
#> 996   0.415123286 -0.6238842539  0.8919144054
#> 997   0.533803409 -0.6204993838 -0.0267670499
#> 998   0.434791167 -0.0653124458  0.4149432432
#> 999   0.255763102 -0.9562549754  0.1390847080
#> 1000  0.322784983 -0.2299078576 -0.0785279129
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.489 -0.519   0.470 
#>  2      2        4       70        1  0.681 -0.398   0.241 
#>  3      3        4       70        1  0.739 -0.346   0.231 
#>  4      4        4       70        1  0.302 -0.528   0.568 
#>  5      5        4       70        1  0.580 -0.565  -0.439 
#>  6      6        4       70        1  0.605 -0.532  -0.315 
#>  7      7        4       70        1  0.450 -0.323   0.416 
#>  8      8        4       70        1  0.592 -0.485  -0.0477
#>  9      9        4       70        1  0.590  0.0465  0.508 
#> 10     10        4       70        1  0.506  0.107   0.381 
#> # ℹ 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  6.52  41.6  1.60 0.157 30.9      6.19e+2 1286.
#> 2      1    14     4    70     1  6.52  41.6  1.60 0.157  6.18     5.76e-7  257.
#> 3      2     1     4    70     1  7.90  47.0  1.27 0.168 24.0      7.83e+2 1130.
#> 4      2    14     4    70     1  7.90  47.0  1.27 0.168  4.86     5.11e-5  229.
#> 5      3     1     4    70     1  8.38  49.5  1.26 0.169 22.7      7.91e+2 1122.
#> 6      3    14     4    70     1  8.38  49.5  1.26 0.169  4.56     6.15e-5  226.
#> # ℹ 1,994 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>