Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e4,n_resample=1e3)
#> $eta
#>             ETA_Cl        ETA_Vc       ETA_Ka
#> 1     0.4612301580 -0.8266936173 -0.182978516
#> 2     0.1959846528 -0.6997620654  0.037540719
#> 3     0.2660681313 -0.9977712466 -0.706348134
#> 4     0.5675679045 -0.3567872073  0.162467454
#> 5     0.4710697846 -0.4021786007 -0.342486217
#> 6     0.3538709453 -0.3188540057 -0.102352066
#> 7     0.7706232318 -0.4977973123 -0.023642689
#> 8     0.7079470723 -0.3610009017  0.054809659
#> 9     0.2914733422 -0.6114695636  0.335015957
#> 10    0.6387659438 -0.2033085268  0.679854181
#> 11    0.4612301580 -0.8266936173 -0.182978516
#> 12    0.4999591026 -0.3962704644  0.089793454
#> 13    0.6521806063 -0.5018948742  0.251868877
#> 14    0.3672107140 -0.4440604481  0.121514292
#> 15    0.0488564185 -0.7636693617  0.427754919
#> 16    0.2380292240 -1.1313646689 -0.267291115
#> 17    0.6315012280 -0.2936321734  0.051795860
#> 18    0.3829950246 -0.6163755048 -0.136090471
#> 19    0.6315012280 -0.2936321734  0.051795860
#> 20    0.1387404806 -0.0341937587  0.335084325
#> 21    0.0224550110 -0.0379199018  0.397123094
#> 22    0.6034435686 -0.2664556975 -0.063182885
#> 23    0.3904617691 -0.6666031984  0.362021487
#> 24    0.3396425059 -0.8992973305  0.079639646
#> 25    0.4963165193 -0.9024584988 -0.432998206
#> 26    0.5606104482 -0.5431091365 -0.047739529
#> 27    0.7315915381 -0.3415518429  0.270006275
#> 28    0.3251882231 -0.9207862919 -0.257631383
#> 29    0.4387311824 -0.3761303477  0.086555293
#> 30    0.4612301580 -0.8266936173 -0.182978516
#> 31    0.5861602717 -0.3091343286  0.005318896
#> 32    0.4121024649 -0.5023118697  0.129264960
#> 33    0.5404524973 -0.2887551269  0.524244602
#> 34    0.3452685533 -0.8682725442  0.023360403
#> 35    0.6077947216  0.0976074613  0.605894608
#> 36    0.2437039479 -0.8735791708 -0.333067194
#> 37    0.4942300146 -0.5472533066  0.173991435
#> 38    0.5554633160 -0.4042291502  0.238831270
#> 39    0.5130901912 -0.6956931443 -0.387849868
#> 40    0.2951641499 -0.5809686276  0.853894256
#> 41    0.4487280285 -0.6416461856 -0.548265580
#> 42    0.4612301580 -0.8266936173 -0.182978516
#> 43    0.3592394535 -0.5291305225  0.841490579
#> 44    0.4496193483 -0.8687630562 -0.663601640
#> 45    0.5640185118 -0.2810665525  0.318625912
#> 46    0.3819707368 -0.4714347125  0.065445159
#> 47    0.7447310321 -0.3027233112  0.150157652
#> 48    0.8379872574 -0.0636708756  0.724445615
#> 49    0.8379872574 -0.0636708756  0.724445615
#> 50    0.1781241826 -0.7981848670 -0.221891850
#> 51    0.5606104482 -0.5431091365 -0.047739529
#> 52    0.4964923238 -0.1140863890  0.196617941
#> 53    0.6881770303 -0.3671060831  0.650232891
#> 54    0.4918395074 -0.7274498988 -0.515014191
#> 55    0.5521112637 -0.7143361247 -0.246402103
#> 56    0.7133112968 -0.3624525615  0.156763361
#> 57    0.4397322136 -0.1694135529  0.810600503
#> 58    0.6684053489 -0.2343678166  0.445877920
#> 59    0.3021588644 -0.5850086317 -0.084909256
#> 60    0.5977612699 -0.1934951342  0.871409391
#> 61    0.5158058677 -0.2089929930  0.353776671
#> 62    0.2229564502 -0.6318692436 -0.236788182
#> 63    0.5115836024 -0.4606996165  0.138310253
#> 64    0.5453186151 -0.3793113858 -0.201669514
#> 65    0.5587754769 -0.4648886906 -0.167808408
#> 66    0.3464268584 -0.2018025609  0.192930448
#> 67    0.6684053489 -0.2343678166  0.445877920
#> 68    0.2741892159 -0.6330273384 -0.144712243
#> 69    0.7541737546  0.0187041829  0.887661749
#> 70    0.4512243571 -0.5642080520  0.282374240
#> 71    0.7952227215  0.0514016010  0.443435577
#> 72    0.5620086118 -0.5161998124 -0.268943157
#> 73    0.3632348346 -0.6057520394 -0.297950635
#> 74    0.4511050887 -0.3470161697 -0.127273192
#> 75    0.3323222125 -0.6751968339 -0.203651281
#> 76    0.5128731412 -0.5714692443  0.593586894
#> 77    0.4907960469 -0.3395639097  0.547603664
#> 78    0.7651701689 -0.1287963737  0.699075614
#> 79    0.5970873749 -0.3835577815  0.307353771
#> 80    0.5222381232 -0.5496771215 -0.174818644
#> 81    0.2721856898 -0.4400797994  0.314442520
#> 82    0.5655694810 -0.1025663062  0.474757193
#> 83    0.2303086825 -0.9505179789 -0.455017524
#> 84    0.5675679045 -0.3567872073  0.162467454
#> 85    0.7315915381 -0.3415518429  0.270006275
#> 86    0.7356217386  0.0088383892  0.461115656
#> 87    0.6177955869  0.1310475622  0.907012410
#> 88    0.5496123451 -0.3832373540 -0.036795978
#> 89    0.5288688780 -0.6965293158 -0.553470408
#> 90    0.3634196553 -0.7090044770 -0.076622925
#> 91    0.4726622084 -0.3960308383  0.860530506
#> 92    0.3247114635 -0.7293943299  0.564508342
#> 93    0.2299388870 -0.6025852330 -0.145624430
#> 94    0.5500423383 -0.6936777151 -0.480943339
#> 95    0.5203758470 -0.6901594509 -0.403859292
#> 96    0.5107362214 -0.7175435587 -0.327359272
#> 97    0.6574742157 -0.2130316864  0.883428515
#> 98    0.7399748982 -0.4677290038 -0.171550495
#> 99    0.3396425059 -0.8992973305  0.079639646
#> 100   0.5313440212 -0.3719077756  0.382800342
#> 101   0.5603085629 -0.3604006928  0.078987476
#> 102   0.7452143671 -0.0657351560  0.536759692
#> 103   0.3995414054 -0.4300629199 -0.261875291
#> 104   0.5202420068 -0.5526890882 -0.060786945
#> 105   0.3522133107  0.2296984190  0.893968900
#> 106   0.6812904347 -0.3787693450  0.291462506
#> 107   0.5620086118 -0.5161998124 -0.268943157
#> 108   0.6812904347 -0.3787693450  0.291462506
#> 109   0.5909206006 -0.4237803856  0.441608608
#> 110   0.8833045341 -0.1821553627  0.371552772
#> 111   0.4265475284 -0.1110251873 -0.029666707
#> 112   0.6813008547 -0.4139869542  0.279512205
#> 113   0.2068455211 -0.6690426375 -0.117525455
#> 114   0.2778644926 -1.0023382906 -0.183814326
#> 115   0.7133112968 -0.3624525615  0.156763361
#> 116   0.4895781649 -0.9959492524 -0.526412091
#> 117   0.4512243571 -0.5642080520  0.282374240
#> 118   0.3021588644 -0.5850086317 -0.084909256
#> 119   0.5360182941 -0.0929116561  0.772572629
#> 120   0.7315915381 -0.3415518429  0.270006275
#> 121   0.8118390992 -0.1778821450  0.481104821
#> 122   0.4874021470 -0.6631147391  0.141783620
#> 123   0.5620086118 -0.5161998124 -0.268943157
#> 124   0.5309061499 -0.2505007187  0.341126802
#> 125   0.7721265604 -0.4075470274  0.054000678
#> 126   0.1164874297 -1.1611624476  0.398454494
#> 127   0.0874580596 -0.8262299949 -0.167859728
#> 128   0.3251882231 -0.9207862919 -0.257631383
#> 129   0.6095869478 -0.2006885115  0.295312809
#> 130   0.3223965898 -1.0439073172 -0.085900882
#> 131   0.7925304284 -0.4840640091 -0.509036854
#> 132   0.5587754769 -0.4648886906 -0.167808408
#> 133   0.7892499713 -0.1126804606  0.708594072
#> 134   0.5365409940 -0.5350451143  0.277928300
#> 135   0.8688149331  0.2245192842  0.544588562
#> 136   0.6902224608 -0.1659294804  0.678752772
#> 137   0.3337570768 -0.9784971099 -0.312577459
#> 138   0.6148690903 -0.5823673824  0.040912244
#> 139   0.1760854823 -0.8067109541  0.084575271
#> 140   0.7444548814  0.1621567902  1.236579224
#> 141   0.2859039670 -0.6264780875 -0.218667495
#> 142   0.5059671584 -0.8780493780 -0.497646111
#> 143   0.4028401444  0.0761703921  0.948124431
#> 144   0.8379872574 -0.0636708756  0.724445615
#> 145   0.5186643291 -0.8245968671 -0.393220231
#> 146   0.4069755579 -0.7050958352  0.160030040
#> 147   0.4124736720 -0.7795627631 -0.362014511
#> 148   0.3256735825 -0.5714883764  0.002526752
#> 149   0.6455381556 -0.6264384398  0.138794122
#> 150   0.6512473853 -0.4976792479 -0.229140339
#> 151   0.7466039792  0.0837357091  0.408284486
#> 152   0.7578727507 -0.0940537694  0.393884972
#> 153   0.6096544795 -0.5283424110 -0.253708134
#> 154   0.5130901912 -0.6956931443 -0.387849868
#> 155   0.7133112968 -0.3624525615  0.156763361
#> 156   0.5772597592 -0.5310496685 -0.011217417
#> 157   0.6059892497 -0.4136253684  0.261799739
#> 158   0.3828680835 -0.8881955781  0.157635487
#> 159   0.5606104482 -0.5431091365 -0.047739529
#> 160   0.7809258920 -0.0433188826  0.904293046
#> 161   0.6106988042 -0.3769071167  0.308107305
#> 162   0.0434113842 -0.8085925840  0.291430860
#> 163   0.7213475598 -0.0168634669  0.259713854
#> 164   0.5785663253 -0.6357008414 -0.187580094
#> 165   0.4402964466 -0.7332791382  0.130280377
#> 166   0.3904617691 -0.6666031984  0.362021487
#> 167   0.5309061499 -0.2505007187  0.341126802
#> 168   0.5202420068 -0.5526890882 -0.060786945
#> 169   0.4164300405 -0.6122437278  0.103330667
#> 170   0.4007428174 -0.4393026909 -0.209757029
#> 171   0.7809258920 -0.0433188826  0.904293046
#> 172   0.6025975164 -0.4332131966 -0.184087305
#> 173   0.7269321116 -0.1839742007  0.165963248
#> 174   0.5087101205 -0.5323985366  0.626365704
#> 175   0.6809904355 -0.4785510663  0.014105108
#> 176   0.8951350330 -0.2279528723  0.136845365
#> 177   0.2023379532 -1.0899083594 -0.472802394
#> 178   0.7532238093 -0.2968851873  0.731283481
#> 179   0.6148690903 -0.5823673824  0.040912244
#> 180   0.4639915456 -0.3629787490  0.183917234
#> 181   0.7603256213 -0.0404097236  0.639798945
#> 182   0.5087101205 -0.5323985366  0.626365704
#> 183   0.1981413887 -0.9443231226 -0.172959105
#> 184   0.7079470723 -0.3610009017  0.054809659
#> 185   0.6816311544 -0.4163212925 -0.039952101
#> 186   0.5469899170 -0.5060702167  0.009217203
#> 187   0.0842805200 -0.5106383003  0.078387945
#> 188   0.6512473853 -0.4976792479 -0.229140339
#> 189   0.7498927555  0.0204683159  0.860562828
#> 190   0.4779236979 -0.6180979657  0.079214321
#> 191   0.3819707368 -0.4714347125  0.065445159
#> 192   0.6156863956 -0.3796418179  0.159697216
#> 193   0.5675679045 -0.3567872073  0.162467454
#> 194   0.5934359465 -0.5461291342  0.026508610
#> 195   0.5115836024 -0.4606996165  0.138310253
#> 196   0.5882837385 -0.6418476051 -0.018261584
#> 197   0.4386385297 -0.4885824165 -0.334170408
#> 198   0.4999591026 -0.3962704644  0.089793454
#> 199   0.4444328321 -0.6957581523 -0.123427136
#> 200   0.5446996968 -0.3425771304 -0.051992062
#> 201   0.4719819958 -0.6187990705 -0.428769462
#> 202   0.6130700631 -1.2382233084 -1.168777248
#> 203   0.4515881776 -0.6474046162 -0.098629093
#> 204   0.7354159932 -0.1775566189  0.506677755
#> 205   0.6381345370 -0.0483201903  0.617653108
#> 206   0.3271295871 -0.5868085592  0.273707106
#> 207   0.6904297723 -0.3070054620  0.180239495
#> 208   0.5797912015 -0.3931698753  0.404208285
#> 209   0.4964923238 -0.1140863890  0.196617941
#> 210   0.5899349781 -0.3600133725  0.498371966
#> 211   0.6870921626 -0.3140007764  0.517553168
#> 212  -0.0156916677 -0.7071026798  0.217635545
#> 213   0.6902224608 -0.1659294804  0.678752772
#> 214   0.4100912237 -0.1128031588  0.406341873
#> 215   0.7122902082 -0.3065758680  0.213855214
#> 216   0.3394912834 -0.7387385184  0.050673834
#> 217   0.6537424531 -0.3155047364  0.070232978
#> 218   0.6531250380 -0.2789371029  0.140430161
#> 219   0.7883989122 -0.0084062074  0.229403835
#> 220   0.3206137874 -0.8751105230 -0.681833334
#> 221   0.5249367851 -0.3992391628  0.182632745
#> 222   0.6936561990 -0.2798397199  0.283327349
#> 223   0.4357295817 -0.3263478693  0.529916515
#> 224   0.7358519587 -0.5046057936 -0.324576437
#> 225   0.5837333514 -0.3474906378  0.397398414
#> 226   0.4956072301 -0.3538656407  0.123767159
#> 227   0.6809904355 -0.4785510663  0.014105108
#> 228   0.6441846061 -0.4211754596  0.295867945
#> 229   0.5500423383 -0.6936777151 -0.480943339
#> 230   0.5069126625 -0.3631824793  0.447047972
#> 231   0.6006863698 -0.3167464495  0.128717422
#> 232   0.8379872574 -0.0636708756  0.724445615
#> 233   0.6387659438 -0.2033085268  0.679854181
#> 234   0.6969570093 -0.4149651672  0.165800524
#> 235   0.5350188143 -0.5943631190  0.142519340
#> 236   0.6571376499 -0.6216386730  0.080795598
#> 237   0.5045753649 -0.5232586258 -0.027909337
#> 238   0.1591385259 -0.5430370969 -0.081496085
#> 239   0.5979016092 -0.1165199855  0.380247601
#> 240   0.7399748982 -0.4677290038 -0.171550495
#> 241   0.8331948033 -0.2560867114  0.202384057
#> 242   0.5469899170 -0.5060702167  0.009217203
#> 243   0.6904297723 -0.3070054620  0.180239495
#> 244   0.6521806063 -0.5018948742  0.251868877
#> 245   0.3729634294 -0.2452538411  0.505813250
#> 246   0.7552566728 -0.3201711083  0.101159247
#> 247   0.7354159932 -0.1775566189  0.506677755
#> 248   0.4710697846 -0.4021786007 -0.342486217
#> 249   0.7133112968 -0.3624525615  0.156763361
#> 250   0.7603256213 -0.0404097236  0.639798945
#> 251   0.7651701689 -0.1287963737  0.699075614
#> 252   0.7714275127 -0.1012506151  0.694102324
#> 253   0.6902224608 -0.1659294804  0.678752772
#> 254   0.6969570093 -0.4149651672  0.165800524
#> 255   0.5632663960 -0.3669338532  0.051737548
#> 256   0.7344514722 -0.2672070366  0.185346709
#> 257   0.6574742157 -0.2130316864  0.883428515
#> 258   0.7101601638 -0.2982947346  0.710823914
#> 259   0.7498927555  0.0204683159  0.860562828
#> 260   0.1781241826 -0.7981848670 -0.221891850
#> 261   0.7953399242 -0.2301559491  0.271117077
#> 262   0.7464102362 -0.2447132481  0.620854295
#> 263   0.4462779142 -0.7519690667 -0.399616516
#> 264   0.4496193483 -0.8687630562 -0.663601640
#> 265   0.4402964466 -0.7332791382  0.130280377
#> 266   0.3404121028 -0.8744866236  0.475497208
#> 267   0.4895781649 -0.9959492524 -0.526412091
#> 268   0.5107362214 -0.7175435587 -0.327359272
#> 269   0.7067691217 -0.3506600000  0.016518141
#> 270   0.7122902082 -0.3065758680  0.213855214
#> 271   0.5249367851 -0.3992391628  0.182632745
#> 272   0.6904297723 -0.3070054620  0.180239495
#> 273   0.6816311544 -0.4163212925 -0.039952101
#> 274   0.7079470723 -0.3610009017  0.054809659
#> 275   0.5789198176 -0.8915624778 -0.481303034
#> 276   0.7541737546  0.0187041829  0.887661749
#> 277   0.5912669379 -0.2663851610  0.351040918
#> 278   0.4511050887 -0.3470161697 -0.127273192
#> 279   0.7651701689 -0.1287963737  0.699075614
#> 280   0.7640845424 -0.0196575448  0.765355435
#> 281   0.3251882231 -0.9207862919 -0.257631383
#> 282   0.9353773898  0.1790014719  1.189338709
#> 283   0.8031734114 -0.2146573356  0.451566173
#> 284   0.5970873749 -0.3835577815  0.307353771
#> 285   0.6813008547 -0.4139869542  0.279512205
#> 286   0.4561990674 -0.7295905991 -0.507929972
#> 287   0.5107362214 -0.7175435587 -0.327359272
#> 288   0.3566727139 -0.4385658328  0.809186108
#> 289   0.3634196553 -0.7090044770 -0.076622925
#> 290   0.5640185118 -0.2810665525  0.318625912
#> 291   0.5236066433 -0.2572231726  0.437416497
#> 292   0.5107362214 -0.7175435587 -0.327359272
#> 293   0.1818869437 -0.2971066031  0.335497643
#> 294   0.4650348458 -0.0038578971  0.882457555
#> 295   0.5404524973 -0.2887551269  0.524244602
#> 296   0.3396425059 -0.8992973305  0.079639646
#> 297   0.2618769214 -0.3178535112  0.076350311
#> 298   0.6571376499 -0.6216386730  0.080795598
#> 299   0.5500423383 -0.6936777151 -0.480943339
#> 300   0.3396425059 -0.8992973305  0.079639646
#> 301   0.6027901720 -0.3292005033  0.186603639
#> 302   0.6521806063 -0.5018948742  0.251868877
#> 303   0.5592406500 -0.0805991571  0.664493861
#> 304   0.5297780829 -0.5770214037  0.113435471
#> 305   0.1893224792 -0.4228326778  0.218416962
#> 306   0.5934359465 -0.5461291342  0.026508610
#> 307   0.3697055672 -0.5358847221  0.205839986
#> 308   0.7809258920 -0.0433188826  0.904293046
#> 309   0.7452143671 -0.0657351560  0.536759692
#> 310   0.5107362214 -0.7175435587 -0.327359272
#> 311   0.9627026980 -0.0385952983  0.213056168
#> 312   0.5186643291 -0.8245968671 -0.393220231
#> 313   0.7162727243 -0.1407145792  0.850823255
#> 314   0.7552566728 -0.3201711083  0.101159247
#> 315   0.6936561990 -0.2798397199  0.283327349
#> 316   0.7354159932 -0.1775566189  0.506677755
#> 317   0.5035990584 -0.0683560230  0.361424159
#> 318   0.4705070049  0.1489169763  1.034536383
#> 319   0.6263587446 -0.3657774294  0.080475824
#> 320   0.5297780829 -0.5770214037  0.113435471
#> 321   0.8087062987 -0.1592159351  0.466862002
#> 322   0.6537424531 -0.3155047364  0.070232978
#> 323   0.0309385803 -0.4859638395  0.705136363
#> 324   0.4143932226 -0.4382222556  0.445199480
#> 325   0.3457964517 -0.3379411392  0.107200061
#> 326   0.6263587446 -0.3657774294  0.080475824
#> 327   0.5353761088 -0.2337743825  0.234667222
#> 328   0.8379872574 -0.0636708756  0.724445615
#> 329   0.4615316613 -0.6799128655  0.375589167
#> 330   0.4615316613 -0.6799128655  0.375589167
#> 331   0.6992337809 -0.2501793232  0.199116827
#> 332   0.4964923238 -0.1140863890  0.196617941
#> 333   0.3176110581 -1.0315205299 -0.191458435
#> 334   0.7249660669 -0.1612808448  0.539110209
#> 335   0.2896072177 -0.2103433438  0.894540113
#> 336   0.3439950060  0.1000959001  0.942385223
#> 337   0.7714275127 -0.1012506151  0.694102324
#> 338   0.6972243187 -0.4000661609  0.605164505
#> 339   0.6194469229 -0.2725130659  0.677653604
#> 340   0.4387311824 -0.3761303477  0.086555293
#> 341   0.2660681313 -0.9977712466 -0.706348134
#> 342   0.8146037515 -0.1955339326  0.271193553
#> 343   0.2771172013 -0.3838324366  0.182452637
#> 344   0.3452685533 -0.8682725442  0.023360403
#> 345   0.5912669379 -0.2663851610  0.351040918
#> 346   0.5107362214 -0.7175435587 -0.327359272
#> 347   0.4196368511 -0.0097229098  0.483629993
#> 348   0.1859911033 -0.9520693591  0.258049648
#> 349   0.3695209071 -0.5510462635  0.223500807
#> 350   0.8300798772 -0.1215353460  0.203461418
#> 351   0.6095869478 -0.2006885115  0.295312809
#> 352   0.7466775890 -0.4205864688 -0.022566218
#> 353   0.3697055672 -0.5358847221  0.205839986
#> 354   0.5288688780 -0.6965293158 -0.553470408
#> 355   0.1831739091 -0.1075737086  0.065669371
#> 356   0.6315012280 -0.2936321734  0.051795860
#> 357   0.5235764651 -0.6396839013  0.304820316
#> 358   0.6263587446 -0.3657774294  0.080475824
#> 359  -0.0512127149 -0.5799031728  0.170317888
#> 360   0.4362444399 -0.3079784579  0.245580894
#> 361   0.5954833863 -0.2254367165  0.433767790
#> 362   0.4462779142 -0.7519690667 -0.399616516
#> 363  -0.0439440691 -0.8587059281 -0.915334043
#> 364   0.6408376957 -0.1725437950  0.338146955
#> 365   0.4332246513 -0.4306624595  0.374400778
#> 366   0.3837249878  0.1137817306  0.879541686
#> 367   0.5087101205 -0.5323985366  0.626365704
#> 368   0.5747463535 -0.6353071888  0.036348421
#> 369   0.6684053489 -0.2343678166  0.445877920
#> 370   0.3377740520 -0.6686890628  0.366247753
#> 371   0.1914522334 -0.2188826100  0.362577201
#> 372   0.7354159932 -0.1775566189  0.506677755
#> 373   0.5521112637 -0.7143361247 -0.246402103
#> 374   0.6147551515 -0.3490009854 -0.084546258
#> 375   0.5640185118 -0.2810665525  0.318625912
#> 376   0.6813008547 -0.4139869542  0.279512205
#> 377   0.7452143671 -0.0657351560  0.536759692
#> 378   0.5681181380 -0.3034795269  0.491499331
#> 379   0.5012216910 -0.4947943042  0.454068302
#> 380   0.4479388666 -0.1334450894  0.599683135
#> 381   0.5222209921 -0.4034562199  0.129986736
#> 382   0.6845084633 -0.6084237701  0.167255203
#> 383   0.7354159932 -0.1775566189  0.506677755
#> 384   0.6628263477 -0.6651010220 -0.448327329
#> 385   0.6263587446 -0.3657774294  0.080475824
#> 386   0.4907960469 -0.3395639097  0.547603664
#> 387   0.7079470723 -0.3610009017  0.054809659
#> 388   0.5592406500 -0.0805991571  0.664493861
#> 389   0.7162727243 -0.1407145792  0.850823255
#> 390   0.4697631319 -0.6919061711  0.071326039
#> 391   0.7809258920 -0.0433188826  0.904293046
#> 392   0.4100681792 -0.0146129041  0.386789853
#> 393   0.8118390992 -0.1778821450  0.481104821
#> 394   0.4824644622 -0.5649165867  0.161872823
#> 395   0.0915120340 -0.7003187135 -0.245966271
#> 396   0.5592406500 -0.0805991571  0.664493861
#> 397   0.7162727243 -0.1407145792  0.850823255
#> 398   0.6156863956 -0.3796418179  0.159697216
#> 399   0.7160144698 -0.0002193819  0.652932282
#> 400   0.7315915381 -0.3415518429  0.270006275
#> 401   0.3988217695 -0.5916182757  0.588367449
#> 402   0.4120684435 -0.3455315372  0.334233051
#> 403   0.6628263477 -0.6651010220 -0.448327329
#> 404   0.5970873749 -0.3835577815  0.307353771
#> 405   0.2927322156 -1.0785830960  0.148417576
#> 406   0.6157011823 -0.0167303839  0.823542460
#> 407   0.2660681313 -0.9977712466 -0.706348134
#> 408   0.5764118435 -0.3464549306  0.107379417
#> 409   0.4051361723 -0.6170006546  0.384274895
#> 410   0.3216747349 -0.6830281042 -0.073375146
#> 411   0.4987976331 -0.4315880502  0.203086202
#> 412   0.4406798291 -0.5333898451  0.565429433
#> 413   0.7122902082 -0.3065758680  0.213855214
#> 414   0.5671248779 -0.2626458001  0.242357762
#> 415   0.5979016092 -0.1165199855  0.380247601
#> 416   0.6492596172 -0.5364112572  0.185821030
#> 417   0.3040685242 -0.1260022860  0.226598723
#> 418   0.4949734671  0.0830026320  0.820200445
#> 419   0.3452685533 -0.8682725442  0.023360403
#> 420   0.3996154529 -0.3309332893  0.644716196
#> 421   0.4999591026 -0.3962704644  0.089793454
#> 422   0.4462779142 -0.7519690667 -0.399616516
#> 423   0.6406049902  0.0045335514  0.237723406
#> 424   0.5805417107 -0.9246624356 -0.670316163
#> 425   0.2936037570 -0.9163084540 -0.208244268
#> 426   0.4818021234 -0.0955815574  0.181566156
#> 427   0.5690798842 -0.4347965157  0.056073062
#> 428   0.0985852093 -0.7012332048  0.100532580
#> 429   0.5202420068 -0.5526890882 -0.060786945
#> 430   0.5199813217 -0.4680689782  0.064906241
#> 431   0.2514512219 -0.8846239467 -0.434700071
#> 432   0.5592406500 -0.0805991571  0.664493861
#> 433   0.5772597592 -0.5310496685 -0.011217417
#> 434   0.3449384762 -0.5727926514  0.081392601
#> 435   0.5785663253 -0.6357008414 -0.187580094
#> 436   0.6315012280 -0.2936321734  0.051795860
#> 437   0.5681181380 -0.3034795269  0.491499331
#> 438   0.0854016465 -0.2799718847  0.230310003
#> 439   0.5469899170 -0.5060702167  0.009217203
#> 440   0.6293260227 -0.1399545869  0.886314073
#> 441   0.4362444399 -0.3079784579  0.245580894
#> 442   0.5203758470 -0.6901594509 -0.403859292
#> 443   0.8118390992 -0.1778821450  0.481104821
#> 444   0.4874021470 -0.6631147391  0.141783620
#> 445   0.4146149585 -0.0748562994  0.172415700
#> 446   0.3153026376 -0.5351209295 -0.014368749
#> 447   0.5934359465 -0.5461291342  0.026508610
#> 448   0.4462779142 -0.7519690667 -0.399616516
#> 449   0.5857631912 -0.2892764392 -0.127780213
#> 450   0.3322397629 -0.1620377243  0.218565318
#> 451   0.3607612015 -0.4861047386  0.423388263
#> 452   0.7603256213 -0.0404097236  0.639798945
#> 453   0.4601694584 -0.5545053775  0.516267875
#> 454   0.4047067712 -0.6552464904  0.362430297
#> 455   0.4639915456 -0.3629787490  0.183917234
#> 456   0.3280959907 -0.0732735069  0.542715469
#> 457   0.6106988042 -0.3769071167  0.308107305
#> 458   0.5861602717 -0.3091343286  0.005318896
#> 459   0.6018598447  0.0304716883  0.498860075
#> 460   0.5202420068 -0.5526890882 -0.060786945
#> 461   0.4958002112 -0.1670755466  0.776951862
#> 462   0.7882707976 -0.2191166260  0.279975728
#> 463   0.7651701689 -0.1287963737  0.699075614
#> 464   0.6349057240 -0.2322621527  0.329547378
#> 465   0.6157011823 -0.0167303839  0.823542460
#> 466   0.5069126625 -0.3631824793  0.447047972
#> 467   0.6840509016 -0.0638423306  0.300432905
#> 468   0.6674725372  0.0207615033  0.484470910
#> 469   0.6072460262 -0.4173707971 -0.184055500
#> 470   0.5469899170 -0.5060702167  0.009217203
#> 471   0.6870921626 -0.3140007764  0.517553168
#> 472  -0.0822687073 -1.0486058339 -0.496448576
#> 473   0.8126134485  0.0110740060  0.484777355
#> 474   0.5886188278 -0.1964063865  0.264063234
#> 475   0.5360182941 -0.0929116561  0.772572629
#> 476   0.7439411737  0.1648447498  0.607090459
#> 477   0.8983032632 -0.1442563474  0.951627726
#> 478   0.6881770303 -0.3671060831  0.650232891
#> 479   0.2995705593 -0.8523282720 -0.453428284
#> 480   0.6147551515 -0.3490009854 -0.084546258
#> 481   0.6992337809 -0.2501793232  0.199116827
#> 482   0.3021588644 -0.5850086317 -0.084909256
#> 483   0.7248564365 -0.1005688158  0.419340684
#> 484   0.3877898521 -0.0191145216  0.224360295
#> 485   0.5937826288 -0.3393059054  0.202998239
#> 486   0.7170923825  0.0782960132  0.780229137
#> 487   0.3604987142 -0.6976655694 -0.205219696
#> 488   0.3101129136 -0.4987481180 -0.059717016
#> 489   0.6870921626 -0.3140007764  0.517553168
#> 490   1.0168533923 -0.0303555222  0.389002553
#> 491   0.5087101205 -0.5323985366  0.626365704
#> 492   0.5764118435 -0.3464549306  0.107379417
#> 493   0.5235764651 -0.6396839013  0.304820316
#> 494   0.5262889967  0.1005549358  0.960387273
#> 495   0.4620752557 -0.4078202438 -0.211205760
#> 496   0.6233945954 -0.1485918555  0.226475554
#> 497   0.8367897676 -0.3074373156  0.293785001
#> 498   0.6904297723 -0.3070054620  0.180239495
#> 499   0.7444548814  0.1621567902  1.236579224
#> 500   0.6902224608 -0.1659294804  0.678752772
#> 501   0.6816311544 -0.4163212925 -0.039952101
#> 502   0.6882786585 -0.2497086178  0.284512333
#> 503   0.7447310321 -0.3027233112  0.150157652
#> 504   0.1873085903 -1.2547478174 -0.482453344
#> 505   0.5970873749 -0.3835577815  0.307353771
#> 506   0.4963165193 -0.9024584988 -0.432998206
#> 507   0.5554633160 -0.4042291502  0.238831270
#> 508   0.5521112637 -0.7143361247 -0.246402103
#> 509   0.3780188812 -0.9126853039 -0.190017978
#> 510   0.1880718800 -0.6343524525 -0.672254048
#> 511   0.3427707006 -0.5911276297 -0.467189713
#> 512   0.6389428881 -0.7702318541 -0.300626560
#> 513   0.7136091806 -0.1054403540  0.823092270
#> 514   0.5087101205 -0.5323985366  0.626365704
#> 515   0.7268068618  0.0086081782  0.556680793
#> 516   0.6148690903 -0.5823673824  0.040912244
#> 517   0.7344514722 -0.2672070366  0.185346709
#> 518   0.3475331231 -0.1500855507  0.197818868
#> 519   0.6521806063 -0.5018948742  0.251868877
#> 520   0.7444548814  0.1621567902  1.236579224
#> 521   0.5115836024 -0.4606996165  0.138310253
#> 522   0.0153708811 -0.7687482697 -0.277957336
#> 523   0.6870921626 -0.3140007764  0.517553168
#> 524   0.2816422262 -1.0876130073 -0.168940956
#> 525   0.6077947216  0.0976074613  0.605894608
#> 526   0.5772597592 -0.5310496685 -0.011217417
#> 527   0.8379872574 -0.0636708756  0.724445615
#> 528   1.0020345204  0.0889399789  0.447306826
#> 529  -0.0840591669 -1.4121363806 -0.511596426
#> 530   0.7552566728 -0.3201711083  0.101159247
#> 531   0.7248564365 -0.1005688158  0.419340684
#> 532   0.5222381232 -0.5496771215 -0.174818644
#> 533   0.7268068618  0.0086081782  0.556680793
#> 534   0.5130901912 -0.6956931443 -0.387849868
#> 535   0.7706232318 -0.4977973123 -0.023642689
#> 536   0.3594595398 -0.1938955606  0.429903858
#> 537   0.6674725372  0.0207615033  0.484470910
#> 538   0.6840509016 -0.0638423306  0.300432905
#> 539   0.7354159932 -0.1775566189  0.506677755
#> 540   0.5404524973 -0.2887551269  0.524244602
#> 541   0.6936561990 -0.2798397199  0.283327349
#> 542   0.6148690903 -0.5823673824  0.040912244
#> 543   0.3151344334 -0.1620692419  0.127884566
#> 544   0.6848273176 -0.5420220699  0.021127247
#> 545   0.5313440212 -0.3719077756  0.382800342
#> 546   0.5693780876 -0.2625710545  0.607532221
#> 547   0.5496123451 -0.3832373540 -0.036795978
#> 548   0.6081198501 -0.2321900604  0.376531100
#> 549   0.4448860329 -0.3984550485  0.326957201
#> 550   0.1602790129 -0.8799665197  0.434673774
#> 551   0.2783228004 -0.4418791369  0.296150954
#> 552   0.3628294627 -0.3581881105  0.210904471
#> 553   0.6904297723 -0.3070054620  0.180239495
#> 554   0.7079470723 -0.3610009017  0.054809659
#> 555   0.5640185118 -0.2810665525  0.318625912
#> 556   0.3094658138 -0.6587683058 -0.347552654
#> 557   0.4017316985 -1.0086441567  0.162888280
#> 558   0.4902640150 -0.5854591091  0.144061312
#> 559   0.4357295817 -0.3263478693  0.529916515
#> 560   0.0571883458 -1.2380192311 -0.084854507
#> 561   0.6904297723 -0.3070054620  0.180239495
#> 562   0.5202420068 -0.5526890882 -0.060786945
#> 563   0.4061717626 -0.7842720570 -0.052289989
#> 564   0.5313440212 -0.3719077756  0.382800342
#> 565   0.6813008547 -0.4139869542  0.279512205
#> 566   0.6840509016 -0.0638423306  0.300432905
#> 567   0.9029061682  0.0168448304  0.798491292
#> 568   0.3466625889 -0.5733052570  0.554261399
#> 569   0.4444328321 -0.6957581523 -0.123427136
#> 570   0.6969570093 -0.4149651672  0.165800524
#> 571   0.7236154540 -0.2884366653  0.234493061
#> 572   0.3153026376 -0.5351209295 -0.014368749
#> 573   0.4402964466 -0.7332791382  0.130280377
#> 574   0.4069755579 -0.7050958352  0.160030040
#> 575   0.1487189717 -0.5388593775 -0.065143423
#> 576   0.3184983213 -0.0489772161  0.749017327
#> 577   0.6095869478 -0.2006885115  0.295312809
#> 578   0.5646095711 -0.0767005651  0.241448131
#> 579   0.4124736720 -0.7795627631 -0.362014511
#> 580   0.6758354680 -0.2602156059  0.224660327
#> 581   0.6251077990 -0.5332590193 -0.343268075
#> 582   0.3604987142 -0.6976655694 -0.205219696
#> 583   0.2995705593 -0.8523282720 -0.453428284
#> 584   0.5537835829 -0.0731773436  0.877521021
#> 585   0.5637367809 -0.2997372772  0.065095734
#> 586   0.4902640150 -0.5854591091  0.144061312
#> 587   0.5189921391 -0.2456186304  0.738643313
#> 588   0.4874021470 -0.6631147391  0.141783620
#> 589   0.3021588644 -0.5850086317 -0.084909256
#> 590   0.5059671584 -0.8780493780 -0.497646111
#> 591   0.5469899170 -0.5060702167  0.009217203
#> 592   0.9317357401  0.0195283087  0.387894374
#> 593   0.3538709453 -0.3188540057 -0.102352066
#> 594   0.5350188143 -0.5943631190  0.142519340
#> 595   0.5882837385 -0.6418476051 -0.018261584
#> 596   0.5554633160 -0.4042291502  0.238831270
#> 597   0.4705070049  0.1489169763  1.034536383
#> 598   0.7634529655 -0.3518688515 -0.129448955
#> 599   0.4611730751 -0.3409025597  0.373847465
#> 600   0.9029061682  0.0168448304  0.798491292
#> 601   0.8356221035 -0.0720588466  0.441885213
#> 602   0.5681181380 -0.3034795269  0.491499331
#> 603   0.5059671584 -0.8780493780 -0.497646111
#> 604   0.3423541658 -0.1700210764  0.506392080
#> 605   0.3492612996 -0.4972837584  0.151095342
#> 606   0.8146037515 -0.1955339326  0.271193553
#> 607   0.4875436133 -0.4502708055  0.261418118
#> 608   0.2015922234 -0.2180674229  0.298363819
#> 609   0.7578727507 -0.0940537694  0.393884972
#> 610   0.5882837385 -0.6418476051 -0.018261584
#> 611   0.6969570093 -0.4149651672  0.165800524
#> 612   0.2514512219 -0.8846239467 -0.434700071
#> 613   0.7249660669 -0.1612808448  0.539110209
#> 614   0.9060222281  0.0332211181  0.309558009
#> 615   0.6782648310 -0.1394527451  0.232283454
#> 616   0.5012216910 -0.4947943042  0.454068302
#> 617   0.4719819958 -0.6187990705 -0.428769462
#> 618   0.4561990674 -0.7295905991 -0.507929972
#> 619   0.2514512219 -0.8846239467 -0.434700071
#> 620   0.6870921626 -0.3140007764  0.517553168
#> 621   0.7603256213 -0.0404097236  0.639798945
#> 622   0.6778575335 -0.2329601261  0.494805328
#> 623   0.1837905554 -0.2400894449  0.364468880
#> 624   0.4487280285 -0.6416461856 -0.548265580
#> 625   0.7133112968 -0.3624525615  0.156763361
#> 626   0.5606104482 -0.5431091365 -0.047739529
#> 627   0.7101601638 -0.2982947346  0.710823914
#> 628   0.7498927555  0.0204683159  0.860562828
#> 629   0.2539081593 -0.0509909741  0.599120021
#> 630   0.3848996731 -0.6635615101 -0.388815869
#> 631   0.6680381817 -0.4190494040 -0.078995196
#> 632   0.2385955148 -0.2055399945  0.425178544
#> 633   0.2758901054 -0.5229930247  0.151458519
#> 634   0.6758354680 -0.2602156059  0.224660327
#> 635   0.6636650924 -0.1878755838  0.689062052
#> 636   0.7101601638 -0.2982947346  0.710823914
#> 637   0.5837333514 -0.3474906378  0.397398414
#> 638   0.7541737546  0.0187041829  0.887661749
#> 639   0.4439261272 -0.6700124294  0.474222054
#> 640   0.6599336068 -0.6537423877  0.004729023
#> 641   0.2182390522 -0.5015803701 -0.086604469
#> 642   0.4210121457 -0.3734485376  0.078389303
#> 643   0.6072460262 -0.4173707971 -0.184055500
#> 644   0.8331948033 -0.2560867114  0.202384057
#> 645   0.6870921626 -0.3140007764  0.517553168
#> 646   0.5934359465 -0.5461291342  0.026508610
#> 647   0.6680381817 -0.4190494040 -0.078995196
#> 648   0.2282983855 -0.0704253712  0.662014869
#> 649   0.2514512219 -0.8846239467 -0.434700071
#> 650   0.5069126625 -0.3631824793  0.447047972
#> 651   0.5603085629 -0.3604006928  0.078987476
#> 652   0.6117145688 -0.2842886940  0.158677722
#> 653   0.5045753649 -0.5232586258 -0.027909337
#> 654   0.6636650924 -0.1878755838  0.689062052
#> 655   0.4597630429  0.0313254722  1.134975388
#> 656   0.4402841165 -0.2568555635 -0.203422842
#> 657   0.6726831189  0.1134391259  0.600402640
#> 658   0.6521806063 -0.5018948742  0.251868877
#> 659   0.5114786151 -0.7348044546 -0.607429534
#> 660   0.4047067712 -0.6552464904  0.362430297
#> 661   0.7541737546  0.0187041829  0.887661749
#> 662   0.5365409940 -0.5350451143  0.277928300
#> 663   0.6636650924 -0.1878755838  0.689062052
#> 664   0.3704992975 -0.1593438096  0.471065095
#> 665   0.4444328321 -0.6957581523 -0.123427136
#> 666   0.6636650924 -0.1878755838  0.689062052
#> 667   0.5970873749 -0.3835577815  0.307353771
#> 668   0.5603085629 -0.3604006928  0.078987476
#> 669   0.6096544795 -0.5283424110 -0.253708134
#> 670   0.1275668056 -0.2546933721 -0.013117001
#> 671   0.7552566728 -0.3201711083  0.101159247
#> 672   0.4402964466 -0.7332791382  0.130280377
#> 673   0.8356221035 -0.0720588466  0.441885213
#> 674   0.4110516244 -0.4595400865  0.565370984
#> 675   0.6014037567 -0.0622859129  0.660288298
#> 676   0.6266502259 -0.2377250089  0.543124123
#> 677   0.6027901720 -0.3292005033  0.186603639
#> 678   0.5910787044 -0.1925434441  1.087473637
#> 679   0.5130901912 -0.6956931443 -0.387849868
#> 680   0.7170923825  0.0782960132  0.780229137
#> 681   0.4874021470 -0.6631147391  0.141783620
#> 682   0.7882707976 -0.2191166260  0.279975728
#> 683   0.6441846061 -0.4211754596  0.295867945
#> 684   0.7498927555  0.0204683159  0.860562828
#> 685   0.2623888785 -0.7331407961  0.108985573
#> 686   0.6250481469 -0.1025736590  0.706392940
#> 687   0.3452685533 -0.8682725442  0.023360403
#> 688   0.6680381817 -0.4190494040 -0.078995196
#> 689   0.7953399242 -0.2301559491  0.271117077
#> 690   0.2414287506 -1.0742386704 -0.581775707
#> 691   0.4667678645 -0.3058025109  0.188451683
#> 692   0.6293260227 -0.1399545869  0.886314073
#> 693   0.5117725543 -0.1754682468  0.620526971
#> 694   0.6680381817 -0.4190494040 -0.078995196
#> 695   0.5107362214 -0.7175435587 -0.327359272
#> 696   0.5899707885 -0.3074429416  0.706641821
#> 697   0.6400935060 -0.3807922507 -0.098447502
#> 698   0.6636650924 -0.1878755838  0.689062052
#> 699   0.5977612699 -0.1934951342  0.871409391
#> 700   0.6315012280 -0.2936321734  0.051795860
#> 701   0.7552566728 -0.3201711083  0.101159247
#> 702   0.9353773898  0.1790014719  1.189338709
#> 703   0.2743846394 -0.2712119722  0.087662276
#> 704   0.6492596172 -0.5364112572  0.185821030
#> 705   0.7067691217 -0.3506600000  0.016518141
#> 706   0.6812904347 -0.3787693450  0.291462506
#> 707   0.4244012943 -0.9525638457  0.162312465
#> 708   0.4110516244 -0.4595400865  0.565370984
#> 709   0.3449384762 -0.5727926514  0.081392601
#> 710   0.5045753649 -0.5232586258 -0.027909337
#> 711   0.4153134070 -1.1926562538 -0.733084088
#> 712   0.6963126347 -0.3940574979 -0.177720696
#> 713   0.3958590386 -0.3025591933  0.250660089
#> 714   0.5365409940 -0.5350451143  0.277928300
#> 715   0.5102085708 -0.3360621392 -0.167960122
#> 716   0.2556135943 -0.2479646591  0.628223395
#> 717   0.6312352895 -0.6567273300  0.038128555
#> 718   0.6904297723 -0.3070054620  0.180239495
#> 719   0.2418183923 -0.5285287149  0.149790360
#> 720   0.5059671584 -0.8780493780 -0.497646111
#> 721   0.2409362764 -0.1355345708  0.231346936
#> 722   0.6006863698 -0.3167464495  0.128717422
#> 723   0.5102085708 -0.3360621392 -0.167960122
#> 724   0.6574742157 -0.2130316864  0.883428515
#> 725   0.8087062987 -0.1592159351  0.466862002
#> 726   0.3394912834 -0.7387385184  0.050673834
#> 727   0.6531250380 -0.2789371029  0.140430161
#> 728   0.7122902082 -0.3065758680  0.213855214
#> 729   0.8586141390  0.1888355164  0.880428082
#> 730   0.2818850560 -0.8417149775 -0.343820097
#> 731   0.3250912639 -0.0429926660  0.709839306
#> 732   0.5249367851 -0.3992391628  0.182632745
#> 733   0.3304801580 -0.9003141257  0.019329430
#> 734   0.4496193483 -0.8687630562 -0.663601640
#> 735   0.5910787044 -0.1925434441  1.087473637
#> 736   0.6823849738 -0.1614579546  0.449702051
#> 737   0.0528638284 -0.1048974059  0.479742364
#> 738   0.4779236979 -0.6180979657  0.079214321
#> 739   0.4402964466 -0.7332791382  0.130280377
#> 740   0.4820811425 -0.2084744964  0.074227766
#> 741   0.2683556651 -0.5711364976  0.150996251
#> 742   0.5186643291 -0.8245968671 -0.393220231
#> 743   0.3472321981 -0.5085467542 -0.257133351
#> 744   0.4895781649 -0.9959492524 -0.526412091
#> 745   0.5606104482 -0.5431091365 -0.047739529
#> 746   0.5640185118 -0.2810665525  0.318625912
#> 747   0.4451358939 -0.1317387789  0.340268321
#> 748   0.5115836024 -0.4606996165  0.138310253
#> 749   0.5222381232 -0.5496771215 -0.174818644
#> 750   0.5632663960 -0.3669338532  0.051737548
#> 751   0.2926404832 -0.6708600036  0.417532170
#> 752   0.5621049128 -0.5993675758  0.768978752
#> 753   0.0756989214 -0.3206619180 -0.108676011
#> 754   0.7758208281 -0.4911158670 -0.122843846
#> 755   0.4939113795 -0.2782833764  0.307742953
#> 756   0.8126134485  0.0110740060  0.484777355
#> 757   0.4822975696 -0.0102782749  0.640073733
#> 758   0.6266502259 -0.2377250089  0.543124123
#> 759   0.3452685533 -0.8682725442  0.023360403
#> 760   0.2437039479 -0.8735791708 -0.333067194
#> 761   0.2698240661 -0.5999141804 -0.322478247
#> 762   0.5189921391 -0.2456186304  0.738643313
#> 763   0.2709877974 -0.3642235048  0.144074328
#> 764   0.5186643291 -0.8245968671 -0.393220231
#> 765   0.6599336068 -0.6537423877  0.004729023
#> 766   0.5369965281 -0.1904047119  0.351914696
#> 767   0.4332246513 -0.4306624595  0.374400778
#> 768   0.5747463535 -0.6353071888  0.036348421
#> 769   0.8118390992 -0.1778821450  0.481104821
#> 770   0.4164300405 -0.6122437278  0.103330667
#> 771   0.7541737546  0.0187041829  0.887661749
#> 772   0.5297780829 -0.5770214037  0.113435471
#> 773   0.4153134070 -1.1926562538 -0.733084088
#> 774   1.0020345204  0.0889399789  0.447306826
#> 775   0.7640845424 -0.0196575448  0.765355435
#> 776   0.8975819774  0.2434803772  0.765331488
#> 777   0.7236154540 -0.2884366653  0.234493061
#> 778   0.7122902082 -0.3065758680  0.213855214
#> 779   0.7640845424 -0.0196575448  0.765355435
#> 780   0.2638872067 -0.6673326654 -0.206423160
#> 781   0.8122687889 -0.2314861152  0.643083665
#> 782   0.6147551515 -0.3490009854 -0.084546258
#> 783   0.8204902163  0.1124491140  0.490788553
#> 784   0.5059671584 -0.8780493780 -0.497646111
#> 785   0.7882707976 -0.2191166260  0.279975728
#> 786   0.4824644622 -0.5649165867  0.161872823
#> 787   0.5637367809 -0.2997372772  0.065095734
#> 788   0.4824644622 -0.5649165867  0.161872823
#> 789   0.3304801580 -0.9003141257  0.019329430
#> 790   0.8379872574 -0.0636708756  0.724445615
#> 791   0.2174376366 -0.4476079868  0.070399934
#> 792   0.2842302801 -0.1729490819  0.246455178
#> 793   0.6081198501 -0.2321900604  0.376531100
#> 794   0.7452143671 -0.0657351560  0.536759692
#> 795   0.4124736720 -0.7795627631 -0.362014511
#> 796   0.6531250380 -0.2789371029  0.140430161
#> 797   0.6809904355 -0.4785510663  0.014105108
#> 798   0.2338069734 -0.6299189955 -0.154094271
#> 799   0.4402964466 -0.7332791382  0.130280377
#> 800   0.5681181380 -0.3034795269  0.491499331
#> 801   0.4639915456 -0.3629787490  0.183917234
#> 802   0.4987976331 -0.4315880502  0.203086202
#> 803   0.8379872574 -0.0636708756  0.724445615
#> 804   0.4597630429  0.0313254722  1.134975388
#> 805   0.5262889967  0.1005549358  0.960387273
#> 806   0.5367660540 -0.1240374236  0.516197627
#> 807   0.2233828637 -0.0943870710  0.308148565
#> 808   0.7354159932 -0.1775566189  0.506677755
#> 809   0.5592406500 -0.0805991571  0.664493861
#> 810   0.6851873886 -0.6141531641 -0.193676949
#> 811   0.6222989124 -0.2439703431  0.144484959
#> 812   0.5675679045 -0.3567872073  0.162467454
#> 813   0.6156863956 -0.3796418179  0.159697216
#> 814   0.7136091806 -0.1054403540  0.823092270
#> 815   0.4377399810 -0.2439567346  0.454834647
#> 816   0.5309580073  0.1037805329  0.638882380
#> 817   0.4963165193 -0.9024584988 -0.432998206
#> 818   0.5939978910 -0.0796321983  0.310827319
#> 819   0.5236066433 -0.2572231726  0.437416497
#> 820   0.5606104482 -0.5431091365 -0.047739529
#> 821   0.6778575335 -0.2329601261  0.494805328
#> 822   0.5882837385 -0.6418476051 -0.018261584
#> 823   0.5012216910 -0.4947943042  0.454068302
#> 824   0.7354159932 -0.1775566189  0.506677755
#> 825   0.6574742157 -0.2130316864  0.883428515
#> 826   0.3837249878  0.1137817306  0.879541686
#> 827  -0.0808006380 -0.1065346837  0.080048400
#> 828   0.6130700631 -1.2382233084 -1.168777248
#> 829  -0.1172288710 -0.4064899689  0.271045882
#> 830   0.6293260227 -0.1399545869  0.886314073
#> 831   0.9530659059  0.0279834277  0.613647278
#> 832   0.4987976331 -0.4315880502  0.203086202
#> 833   0.5768076554 -0.0942831962  0.646564063
#> 834   0.6963126347 -0.3940574979 -0.177720696
#> 835   0.6848273176 -0.5420220699  0.021127247
#> 836   0.4956072301 -0.3538656407  0.123767159
#> 837   0.8688149331  0.2245192842  0.544588562
#> 838   0.7067691217 -0.3506600000  0.016518141
#> 839   0.3337570768 -0.9784971099 -0.312577459
#> 840   0.4303659148 -0.1234988932  0.852333713
#> 841   0.4110516244 -0.4595400865  0.565370984
#> 842  -0.0083020090 -1.1313648453  0.121025064
#> 843   0.2940463041 -0.8499835954 -0.523294038
#> 844   0.3310482508 -0.0141187860  0.544078086
#> 845   0.5236066433 -0.2572231726  0.437416497
#> 846   0.5241841090 -0.4920953153 -0.249145739
#> 847   0.2878875537 -0.5933128970 -0.082755472
#> 848   0.7651701689 -0.1287963737  0.699075614
#> 849   0.8126134485  0.0110740060  0.484777355
#> 850   0.3292343952 -0.9222642317  0.434076986
#> 851   0.5970873749 -0.3835577815  0.307353771
#> 852   0.3337570768 -0.9784971099 -0.312577459
#> 853   0.6782174145  0.1061489842  0.360641469
#> 854   0.4351898592 -0.1689587078  0.368572207
#> 855   0.7052463216  0.0673664072  0.587289848
#> 856   0.6148690903 -0.5823673824  0.040912244
#> 857   0.3033127384 -0.4219656799 -0.201614279
#> 858   0.6936561990 -0.2798397199  0.283327349
#> 859   0.3756675727 -0.6883293307  0.449585543
#> 860   0.6148690903 -0.5823673824  0.040912244
#> 861   0.5910787044 -0.1925434441  1.087473637
#> 862   0.2580948022 -0.8929740570 -0.291967659
#> 863   0.6205055849 -0.0022902054  0.102667833
#> 864   0.3464268584 -0.2018025609  0.192930448
#> 865   0.7136091806 -0.1054403540  0.823092270
#> 866   0.4462779142 -0.7519690667 -0.399616516
#> 867   0.9353773898  0.1790014719  1.189338709
#> 868   0.5057387631 -0.1836760181  0.203735123
#> 869   0.6293260227 -0.1399545869  0.886314073
#> 870   0.6809904355 -0.4785510663  0.014105108
#> 871   0.5203758470 -0.6901594509 -0.403859292
#> 872   0.1563285247 -0.0539005675  0.860215502
#> 873   0.3427707006 -0.5911276297 -0.467189713
#> 874   0.5603085629 -0.3604006928  0.078987476
#> 875   0.7399748982 -0.4677290038 -0.171550495
#> 876   0.4760995619 -0.1751790668  0.417048952
#> 877   0.3669713907 -0.2649297383  0.394680418
#> 878   0.2414287506 -1.0742386704 -0.581775707
#> 879   0.2778644926 -1.0023382906 -0.183814326
#> 880   0.5012216910 -0.4947943042  0.454068302
#> 881   0.6628263477 -0.6651010220 -0.448327329
#> 882   0.7249660669 -0.1612808448  0.539110209
#> 883   0.7634529655 -0.3518688515 -0.129448955
#> 884   0.3467643293  0.2864894878  0.789811253
#> 885   0.4028401444  0.0761703921  0.948124431
#> 886   0.5313440212 -0.3719077756  0.382800342
#> 887   0.4191807678 -0.2422575120  0.172846745
#> 888   0.3394912834 -0.7387385184  0.050673834
#> 889   0.8367897676 -0.3074373156  0.293785001
#> 890   0.3819707368 -0.4714347125  0.065445159
#> 891   0.7892499713 -0.1126804606  0.708594072
#> 892  -0.0001899186 -0.5800156431 -0.316361799
#> 893   0.3695209071 -0.5510462635  0.223500807
#> 894   0.4662450650 -0.1493233214  0.506952053
#> 895   0.4963165193 -0.9024584988 -0.432998206
#> 896   0.2878875537 -0.5933128970 -0.082755472
#> 897   0.4462779142 -0.7519690667 -0.399616516
#> 898   0.6936561990 -0.2798397199  0.283327349
#> 899   0.7344514722 -0.2672070366  0.185346709
#> 900   0.3193390261 -0.8669373424  0.783445312
#> 901   0.5253870741 -0.2187640495  0.919970892
#> 902   0.6684053489 -0.2343678166  0.445877920
#> 903   0.8833045341 -0.1821553627  0.371552772
#> 904   0.5262889967  0.1005549358  0.960387273
#> 905   0.5521112637 -0.7143361247 -0.246402103
#> 906   0.1516632320 -1.2960334756 -0.090768072
#> 907   0.5681181380 -0.3034795269  0.491499331
#> 908   0.3958808973 -0.0348424079  0.740623617
#> 909   0.7532238093 -0.2968851873  0.731283481
#> 910   0.7136091806 -0.1054403540  0.823092270
#> 911   0.5495349605 -0.6802577312  0.223049110
#> 912   0.4260054242 -0.7575519841 -0.081290088
#> 913   0.6148690903 -0.5823673824  0.040912244
#> 914   0.4999591026 -0.3962704644  0.089793454
#> 915   0.4609536424 -0.2312079172  0.099933709
#> 916   0.5469899170 -0.5060702167  0.009217203
#> 917   0.5059671584 -0.8780493780 -0.497646111
#> 918  -0.0134652786 -1.0990419961  0.425393139
#> 919   0.6293260227 -0.1399545869  0.886314073
#> 920   0.7160144698 -0.0002193819  0.652932282
#> 921   0.4367979945 -0.5704846587 -0.151959083
#> 922   0.3839372050 -0.3334828381  0.293514161
#> 923   0.6117145688 -0.2842886940  0.158677722
#> 924   0.7236154540 -0.2884366653  0.234493061
#> 925   0.3493622947 -0.4578913642 -0.212750109
#> 926   0.6441846061 -0.4211754596  0.295867945
#> 927   0.5297780829 -0.5770214037  0.113435471
#> 928   0.5621049128 -0.5993675758  0.768978752
#> 929   0.6096544795 -0.5283424110 -0.253708134
#> 930   0.3695209071 -0.5510462635  0.223500807
#> 931   0.3618109234  0.0540793997  0.676296618
#> 932   0.6516723103 -0.3969780127  0.266931260
#> 933   0.3424254386 -0.3164879004  0.347016741
#> 934   0.2756972464  0.0465021300  0.221425702
#> 935   0.7399748982 -0.4677290038 -0.171550495
#> 936   0.5189921391 -0.2456186304  0.738643313
#> 937   0.6147551515 -0.3490009854 -0.084546258
#> 938   0.7464102362 -0.2447132481  0.620854295
#> 939   0.3756675727 -0.6883293307  0.449585543
#> 940   0.5764118435 -0.3464549306  0.107379417
#> 941   0.6882786585 -0.2497086178  0.284512333
#> 942   0.6816311544 -0.4163212925 -0.039952101
#> 943   0.4039258823  0.1322595458  0.753073504
#> 944   0.2089591268 -0.3589028764  0.684273567
#> 945   0.6969570093 -0.4149651672  0.165800524
#> 946   0.4444328321 -0.6957581523 -0.123427136
#> 947   0.4639915456 -0.3629787490  0.183917234
#> 948   0.2843010517 -0.5966383464 -0.134287170
#> 949   0.7892499713 -0.1126804606  0.708594072
#> 950   0.5069126625 -0.3631824793  0.447047972
#> 951   0.5115836024 -0.4606996165  0.138310253
#> 952   0.5463321459 -0.0326429230  0.262098820
#> 953   0.5059671584 -0.8780493780 -0.497646111
#> 954   0.2783228004 -0.4418791369  0.296150954
#> 955   0.6263587446 -0.3657774294  0.080475824
#> 956   0.1873085903 -1.2547478174 -0.482453344
#> 957   0.4416342771 -0.2724011866  0.203753768
#> 958   0.4856240908 -0.5439819968  0.248142949
#> 959   0.1170421202 -0.6102958338  0.245079405
#> 960   0.6969570093 -0.4149651672  0.165800524
#> 961   0.4061717626 -0.7842720570 -0.052289989
#> 962   0.7552566728 -0.3201711083  0.101159247
#> 963   0.4537114714 -0.2346137211 -0.088299395
#> 964   0.1901937015 -0.1934773208 -0.082456820
#> 965   0.0170784596 -0.6838351307 -0.234079241
#> 966   0.6823849738 -0.1614579546  0.449702051
#> 967   0.6217592612 -0.1661614097  0.151268796
#> 968   0.1779876738 -1.1166868640 -0.833582857
#> 969   0.4145432870 -0.8020649101 -0.681981760
#> 970   0.5496123451 -0.3832373540 -0.036795978
#> 971   0.6156863956 -0.3796418179  0.159697216
#> 972   0.6027901720 -0.3292005033  0.186603639
#> 973   0.6392855086 -0.2552195230  1.296722790
#> 974   1.0317152192  0.3160586801  1.197140688
#> 975   0.7133112968 -0.3624525615  0.156763361
#> 976   0.5797912015 -0.3931698753  0.404208285
#> 977   0.7136091806 -0.1054403540  0.823092270
#> 978   0.7603256213 -0.0404097236  0.639798945
#> 979   0.5222381232 -0.5496771215 -0.174818644
#> 980   0.5764118435 -0.3464549306  0.107379417
#> 981   0.7892499713 -0.1126804606  0.708594072
#> 982   0.5747463535 -0.6353071888  0.036348421
#> 983   0.5222381232 -0.5496771215 -0.174818644
#> 984   0.7714275127 -0.1012506151  0.694102324
#> 985   0.5199813217 -0.4680689782  0.064906241
#> 986   0.6708928939  0.0064000251  0.501907117
#> 987   0.6392855086 -0.2552195230  1.296722790
#> 988   0.6972243187 -0.4000661609  0.605164505
#> 989   0.6992337809 -0.2501793232  0.199116827
#> 990   0.6904297723 -0.3070054620  0.180239495
#> 991   0.2936037570 -0.9163084540 -0.208244268
#> 992   0.5970873749 -0.3835577815  0.307353771
#> 993   0.5155444322 -0.3392611466  0.144264578
#> 994   0.6293260227 -0.1399545869  0.886314073
#> 995   0.4719819958 -0.6187990705 -0.428769462
#> 996   0.5235764651 -0.6396839013  0.304820316
#> 997   0.7651701689 -0.1287963737  0.699075614
#> 998   0.4402964466 -0.7332791382  0.130280377
#> 999   0.6537424531 -0.3155047364  0.070232978
#> 1000  0.7603256213 -0.0404097236  0.639798945
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1  0.461 -0.827 -0.183 
#>  2      2        4       70        1  0.196 -0.700  0.0375
#>  3      3        4       70        1  0.266 -0.998 -0.706 
#>  4      4        4       70        1  0.568 -0.357  0.162 
#>  5      5        4       70        1  0.471 -0.402 -0.342 
#>  6      6        4       70        1  0.354 -0.319 -0.102 
#>  7      7        4       70        1  0.771 -0.498 -0.0236
#>  8      8        4       70        1  0.708 -0.361  0.0548
#>  9      9        4       70        1  0.291 -0.611  0.335 
#> 10     10        4       70        1  0.639 -0.203  0.680 
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  6.34  30.6 0.833 0.207 27.6  1079.       844.
#> 2      1    14     4    70     1  6.34  30.6 0.833 0.207  5.04    0.0214   154.
#> 3      2     1     4    70     1  4.87  34.8 1.04  0.140 29.0   928.      1009.
#> 4      2    14     4    70     1  4.87  34.8 1.04  0.140  9.71    0.00128  337.
#> 5      3     1     4    70     1  5.22  25.8 0.493 0.202 22.0  1385.       567.
#> 6      3    14     4    70     1  5.22  25.8 0.493 0.202  8.00    2.27     206.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#>