Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e4,n_resample=1e3)
#> $eta
#>            ETA_Cl        ETA_Vc       ETA_Ka
#> 1     0.630750326 -0.3469106943  0.259351051
#> 2     0.570824659 -0.2187431226  0.306583776
#> 3     0.495498817 -0.5775514746  0.034980529
#> 4    -0.054123147 -0.3970852415  0.013183102
#> 5    -0.417771760 -0.4747273963  0.078564341
#> 6     0.426857090 -0.1219573467  0.266063912
#> 7     0.358109799 -0.3721713026  0.594120877
#> 8     0.520184267 -0.4956750517  0.400627072
#> 9     0.765719594 -0.1362386450  0.485680756
#> 10    0.680554748 -0.5459069867  0.046953704
#> 11    0.552325584 -0.5424671720  0.040777836
#> 12    0.547818885 -0.1594581225  1.036005190
#> 13    0.469141359 -0.3720085040  0.384112689
#> 14    0.275412910 -0.5878258794 -0.341903307
#> 15    0.358825327 -0.0464915923  0.387558905
#> 16    0.775846679 -0.2469894472  0.003948734
#> 17    0.692386374 -0.3186897647  0.428494352
#> 18    0.325544245 -0.6666017280  0.504718288
#> 19    0.666826627 -0.5258683999  0.710784467
#> 20    0.449289465 -0.1786624209  0.369022512
#> 21    0.761033855 -0.1753108003  0.570167890
#> 22    0.818515833 -0.3199005678 -0.048849179
#> 23    0.658315560 -0.4617453191  0.131113581
#> 24    0.547818885 -0.1594581225  1.036005190
#> 25    0.405827763 -0.7769411518 -0.051808543
#> 26    0.587346613 -0.2408070579  0.280653009
#> 27    0.579012596 -0.1080902667  0.674712533
#> 28    0.679811362  0.0892256483  0.482040811
#> 29    0.478957764 -0.5340557166  0.252336062
#> 30    0.345677131 -0.0866193706  0.578808760
#> 31    0.608659885 -0.5198231667  0.156806859
#> 32    0.503127414 -0.6932721257 -0.289710781
#> 33    0.477434098 -0.6060305975 -0.324520458
#> 34    0.242494444  0.0465710170  0.737636504
#> 35    0.726357198 -0.1728930552 -0.052211771
#> 36   -0.006130753 -0.4109768601 -0.028250346
#> 37    0.198125694 -0.2311194362  0.485853063
#> 38    0.495853088 -0.4013736003  0.720399137
#> 39   -0.018744371 -0.4459141258  0.201135522
#> 40    0.320713012 -0.7575018805 -0.293070189
#> 41    0.568660775 -0.6371843813 -0.061004803
#> 42    0.648600254 -0.4092123108  0.408909293
#> 43    0.464946780 -0.3079662307  0.403152563
#> 44    0.550080497 -0.4410584549 -0.405144690
#> 45    0.654266083 -0.3786190786  0.198252652
#> 46    0.071673758  0.1640011057  0.358809036
#> 47    0.807304383 -0.1976143827  0.149037182
#> 48    0.765719594 -0.1362386450  0.485680756
#> 49    0.478491420 -0.7359408405 -0.170419719
#> 50    0.747388455 -0.2299008892  0.629457501
#> 51    0.522805258 -0.2056525940  0.335142380
#> 52    0.569195473 -0.0331657768  1.031632598
#> 53    0.880887839 -0.0400161573  0.348698980
#> 54    0.730856619 -0.3977310009 -0.008600514
#> 55    0.641813063 -0.0676380929  0.308671984
#> 56    0.343864760 -1.0850953714 -0.216909744
#> 57    0.185366326 -0.0874316948  0.648839019
#> 58    0.400675857 -0.0928449301  0.722376626
#> 59   -0.125928438 -0.6963983666 -0.035871105
#> 60    0.565764643 -0.4664557518 -0.066263506
#> 61    0.644506649 -0.4690901717 -0.112289571
#> 62    0.334942259 -0.6645730260  0.204704300
#> 63    0.486344761 -0.5312173139 -0.046101632
#> 64    0.470498661 -0.7860215242 -0.031421425
#> 65    0.441841720 -0.2309798550  0.635153876
#> 66    0.794131245 -0.2192560093  0.418359129
#> 67    0.495498817 -0.5775514746  0.034980529
#> 68   -0.057502571 -0.2814696239  0.305342352
#> 69   -0.073987217 -0.3509265234  0.586921139
#> 70    0.295191704 -0.8140352407 -0.344759487
#> 71    0.547216050 -0.5270740238  0.602940808
#> 72    0.719520437 -0.3747267428  0.803123970
#> 73    0.665530141  0.1197336663  0.883331912
#> 74    0.799766938 -0.0605214155  0.358723532
#> 75    0.707225363 -0.0324680183  0.693385545
#> 76    0.505989675 -0.6689697244 -0.087146129
#> 77    0.178790190 -0.4019062488 -0.123119655
#> 78    0.484977577 -0.8124994559  0.016686815
#> 79    0.638014717 -0.5908965165  0.008892999
#> 80    0.520184267 -0.4956750517  0.400627072
#> 81    0.761033855 -0.1753108003  0.570167890
#> 82    0.691714582 -0.2865414470  0.198229880
#> 83    0.501137625 -0.8048919399 -0.244850041
#> 84    0.072348442 -0.6323884741 -0.023796701
#> 85    0.621937742 -0.3342982826  0.551703667
#> 86    0.580575520 -0.5763051332  0.008445292
#> 87    0.417648087  0.0292116634  0.560321553
#> 88    0.616437850 -0.1423682869  1.236568305
#> 89    0.580575520 -0.5763051332  0.008445292
#> 90    0.520790919 -0.1699599936  0.484650913
#> 91    0.510336283 -0.4134707116  0.055228169
#> 92    0.580575520 -0.5763051332  0.008445292
#> 93    0.739336551 -0.4687573272 -0.085457919
#> 94    0.505989675 -0.6689697244 -0.087146129
#> 95    0.489711509 -0.3543949449  0.028690576
#> 96    0.742666527 -0.0903633124  0.645536967
#> 97    0.444308374 -0.7872674419 -0.730080384
#> 98    0.723557821 -0.0377581581  0.612870312
#> 99    0.654266083 -0.3786190786  0.198252652
#> 100   0.473389296 -0.3808540099  0.170317884
#> 101   0.531957645 -0.5600875976  0.461378659
#> 102   0.757995635  0.0729570624  0.732841042
#> 103   0.393891163 -0.4869378459  0.797907523
#> 104   0.156415263 -0.6307645934 -0.189795208
#> 105   0.348862578 -0.1213422939  0.462825699
#> 106   0.658315560 -0.4617453191  0.131113581
#> 107   0.583779581 -0.1337646674  0.434874569
#> 108   0.316781542 -0.8299339889 -0.234740506
#> 109   0.754346468 -0.2127313338  0.303977629
#> 110   0.329343341 -0.6902772029 -0.023930234
#> 111   0.286038934 -0.1095768019  0.951639185
#> 112   0.402373468 -0.5047439624  0.042668622
#> 113   0.250999792 -0.0355563926  0.457634107
#> 114   0.624341955 -0.5033349493 -0.102783347
#> 115   0.269896504 -0.5292367172  0.079499878
#> 116   0.485671761 -0.2187162379  0.124208376
#> 117   0.279315845 -0.5429820848  0.664222519
#> 118   0.307975318  0.0607886646  0.494804280
#> 119   0.501791973 -0.3527723836 -0.030503470
#> 120   0.606463384 -0.6659632134 -0.542254939
#> 121   0.541768367 -0.7768375118 -0.289799615
#> 122   0.461197060 -0.4519121131  0.109310227
#> 123   0.605148870 -0.5553520794  0.356091124
#> 124   0.730856619 -0.3977310009 -0.008600514
#> 125   0.621937742 -0.3342982826  0.551703667
#> 126   0.475072951 -0.4908675121  0.452746570
#> 127   0.691666359 -0.3580102252  0.101451694
#> 128   0.594714782 -0.2956859225  1.037801727
#> 129   0.549698145 -0.1291552621  0.588592974
#> 130   0.173540764 -0.5602344802 -0.143450917
#> 131   0.699979359 -0.3127497323  0.137940040
#> 132   0.275180256 -0.7897158197  0.240100894
#> 133   0.473389296 -0.3808540099  0.170317884
#> 134   0.547216050 -0.5270740238  0.602940808
#> 135   0.543062175 -0.7647464425 -0.281021828
#> 136   0.646012848 -0.4147604973  0.065567924
#> 137   0.521619792 -0.3648169202  0.593199137
#> 138   0.469141359 -0.3720085040  0.384112689
#> 139   0.680554748 -0.5459069867  0.046953704
#> 140   0.494770443 -0.6855825738 -0.399309620
#> 141   0.484977577 -0.8124994559  0.016686815
#> 142   0.465576055 -0.3913701213  0.374037777
#> 143   0.399310156 -0.8715022149 -0.107092594
#> 144   0.684094057 -0.3822165398  0.655270752
#> 145   0.200626584 -0.2130442885  0.171909219
#> 146   0.510276023 -0.3550367898  0.247163005
#> 147   0.478491420 -0.7359408405 -0.170419719
#> 148   0.670160759 -0.1403250884  0.110437374
#> 149   0.659133461  0.1196360139  1.125889936
#> 150   0.400699322 -0.2696094480  0.807980842
#> 151   0.275191569 -0.0358868356  0.227520019
#> 152   0.608659885 -0.5198231667  0.156806859
#> 153   0.569195473 -0.0331657768  1.031632598
#> 154   0.503277907 -0.0852638656  0.375672117
#> 155   0.498207096 -0.2488802371 -0.175210940
#> 156   0.505989675 -0.6689697244 -0.087146129
#> 157   0.284781090 -0.6785838538 -0.128691950
#> 158   0.524134345 -0.8351105987 -0.487963506
#> 159   0.566671467 -0.3998167796  0.687601565
#> 160   0.611719903  0.0490354603  1.147897509
#> 161   0.708796633 -0.5070151001  0.151112089
#> 162   0.581310551 -0.1205315632  0.154304812
#> 163   0.573184086 -0.6350235854  0.516305966
#> 164   0.627265249 -0.2020708830  0.641656139
#> 165   0.488048683 -0.5189840996  0.049837700
#> 166   0.402544412 -0.2558361713  0.069526278
#> 167   0.508815980 -0.7208848043  0.007270023
#> 168   0.334942259 -0.6645730260  0.204704300
#> 169   0.758969628  0.2067019742  0.615784201
#> 170   0.699416487 -0.1643944306  0.736971339
#> 171   0.684054982 -0.4985642636  0.643095629
#> 172   0.517027023 -0.1224592621  0.418971853
#> 173   0.235125643 -0.0648194510  0.155513013
#> 174   0.624577780 -0.5133122650 -0.043160935
#> 175   0.573184086 -0.6350235854  0.516305966
#> 176   0.533474727 -0.4958831149  0.290775482
#> 177   0.531026580 -0.3839951947 -0.097644821
#> 178   0.644506649 -0.4690901717 -0.112289571
#> 179   0.399567211 -0.0950235495  0.226842348
#> 180   0.558643053 -0.5879834139 -0.064830414
#> 181   0.678233932  0.0965742928  0.499385451
#> 182   0.475072951 -0.4908675121  0.452746570
#> 183   0.488048683 -0.5189840996  0.049837700
#> 184   0.394309185 -0.4004087890  0.074166125
#> 185   0.346365519 -0.3064358311  0.372820615
#> 186   0.371391374 -0.8149312967 -0.263629408
#> 187   0.389141338 -0.6833461944  0.168273780
#> 188   0.693161951 -0.1853914147  1.095439769
#> 189   0.794131245 -0.2192560093  0.418359129
#> 190   0.603574864 -0.3598764891  0.218675214
#> 191   0.552325584 -0.5424671720  0.040777836
#> 192   0.521619792 -0.3648169202  0.593199137
#> 193   0.472719492 -0.6925791872  0.101122154
#> 194   0.501137625 -0.8048919399 -0.244850041
#> 195   0.473164533 -0.0322004271  0.602406077
#> 196   0.320426362 -0.0397044712  0.468497238
#> 197   0.511352596 -0.4915707857 -0.134373264
#> 198   0.328403310  0.1385413714  0.700983088
#> 199   0.564960562 -0.4433393024 -0.118777924
#> 200   0.568660775 -0.6371843813 -0.061004803
#> 201   0.268342943 -1.1779558507 -0.551618738
#> 202   0.831603715  0.0457240086  0.640451287
#> 203   0.358109799 -0.3721713026  0.594120877
#> 204   0.713917798 -0.3207785042 -0.164687509
#> 205   0.605534915 -0.5156753429  0.168580697
#> 206   0.538691757 -0.4199017375  0.056176619
#> 207   0.358807841 -0.2245348599  0.520204210
#> 208   0.182694809 -0.3149802899  0.423159692
#> 209   0.639537878 -0.3514054189  0.313414823
#> 210   0.529777636 -0.5891031783 -0.157323610
#> 211   0.670197033  0.1052407710  0.711860677
#> 212   0.777206655  0.0776103774  0.480433093
#> 213   0.486506070 -0.3778933262  0.102684963
#> 214   0.711108447 -0.0690512510  0.224548630
#> 215   0.758715249 -0.2277070793  0.996434516
#> 216   0.241005501 -0.7097166718  0.133962429
#> 217   0.516376975 -0.2339550187  0.261320230
#> 218   0.510336283 -0.4134707116  0.055228169
#> 219   0.330304149 -0.9110684514  0.303839778
#> 220   0.653951466 -0.6384714887 -0.353552934
#> 221   0.644420351 -0.1339095856  0.558365896
#> 222   0.579012596 -0.1080902667  0.674712533
#> 223   0.399310156 -0.8715022149 -0.107092594
#> 224   0.537473594 -0.7121418161  0.112749371
#> 225   0.807304383 -0.1976143827  0.149037182
#> 226   0.704625482 -0.3957921871 -0.015665171
#> 227   0.637601130 -0.3100033348  0.425364194
#> 228   0.691714582 -0.2865414470  0.198229880
#> 229   0.844231310 -0.0487740540  0.548005716
#> 230   0.501791973 -0.3527723836 -0.030503470
#> 231   0.570824659 -0.2187431226  0.306583776
#> 232   0.633623954 -0.6045737731 -0.304291395
#> 233   0.347044642 -0.9043220672  0.035982809
#> 234   0.309947121 -0.5788288177 -0.360353697
#> 235   0.535430574 -0.5926003047 -0.197710845
#> 236   0.371391374 -0.8149312967 -0.263629408
#> 237  -0.023489655 -0.6349408301 -0.177576322
#> 238   0.208753321 -1.0027628282 -0.166614168
#> 239   0.495498817 -0.5775514746  0.034980529
#> 240   0.546501527 -0.3955329901 -0.124089750
#> 241   0.379751544 -0.3557429647  0.059234486
#> 242   0.522805258 -0.2056525940  0.335142380
#> 243   0.383769504 -0.6749125059 -0.225949823
#> 244   0.675765254 -0.4907265452  0.242046174
#> 245   0.402797755 -0.6533415746 -0.258213589
#> 246   0.552522043 -0.7322225660  0.232800211
#> 247   0.693429380 -0.4928519139 -0.301461491
#> 248   0.543062175 -0.7647464425 -0.281021828
#> 249   0.522655422 -0.3957636242  0.188850848
#> 250   0.799766938 -0.0605214155  0.358723532
#> 251   0.465576055 -0.3913701213  0.374037777
#> 252   0.801594749 -0.1897360236 -0.074992297
#> 253   0.692386374 -0.3186897647  0.428494352
#> 254   0.295097146 -0.5223590498  0.048893409
#> 255   0.709666889 -0.2740009158  0.230536948
#> 256   0.820618390 -0.3652746519  0.099767265
#> 257   0.617054576 -0.2380166450  0.377074383
#> 258   0.621937742 -0.3342982826  0.551703667
#> 259   0.185366326 -0.0874316948  0.648839019
#> 260   0.699416487 -0.1643944306  0.736971339
#> 261   0.579012596 -0.1080902667  0.674712533
#> 262   0.530649171 -0.1306434010  0.750639880
#> 263   0.347962314  0.0088953562  0.867280872
#> 264   0.479758911 -0.2738241371  0.446058749
#> 265   0.779910235 -0.3209005927  0.794422859
#> 266   0.638014717 -0.5908965165  0.008892999
#> 267   0.207116110 -0.0277882555  0.205457376
#> 268   0.503127414 -0.6932721257 -0.289710781
#> 269   0.430836868 -0.4154031711 -0.029505056
#> 270   0.580575520 -0.5763051332  0.008445292
#> 271   0.494770443 -0.6855825738 -0.399309620
#> 272   0.847237469 -0.0089514304  0.631981134
#> 273   0.533474727 -0.4958831149  0.290775482
#> 274   0.425293169 -0.7783599157  0.342950354
#> 275   0.514752925 -0.4790178929 -0.017104470
#> 276   0.492243164 -0.3252623714  0.367817200
#> 277   0.785048750 -0.1629779324  0.147511499
#> 278   0.115036132 -0.4127049809  0.134711413
#> 279   0.005494727 -0.1291122019  0.319152654
#> 280   0.426985838 -0.6191124544  0.387428179
#> 281   0.269449974 -0.0885452409  0.626923972
#> 282   0.692386374 -0.3186897647  0.428494352
#> 283   0.535430574 -0.5926003047 -0.197710845
#> 284   0.381826009 -0.0899868122  0.362999107
#> 285   0.646012848 -0.4147604973  0.065567924
#> 286   0.654430154 -0.2607641783  0.386218899
#> 287   0.689374883 -0.2347061793  0.004904187
#> 288   0.433557803 -0.4164708992 -0.019769411
#> 289   0.633623954 -0.6045737731 -0.304291395
#> 290   0.618333091 -0.0965904515  0.626088997
#> 291   0.465035930 -0.2104732714  0.444295708
#> 292   0.194836313 -0.4928086279  0.247542833
#> 293   0.522805258 -0.2056525940  0.335142380
#> 294   0.189333205 -0.7745892544 -0.276350665
#> 295   0.680554748 -0.5459069867  0.046953704
#> 296   0.212439808 -0.2412021808  0.107673100
#> 297   0.716021119 -0.5555646713 -0.560857091
#> 298   0.636219445 -0.2659242852  0.548682900
#> 299   0.201967637 -0.6239895986 -0.363488671
#> 300   0.470498661 -0.7860215242 -0.031421425
#> 301   0.364242091 -0.5130521811 -0.361116628
#> 302   0.844231310 -0.0487740540  0.548005716
#> 303   0.654266083 -0.3786190786  0.198252652
#> 304   0.568660775 -0.6371843813 -0.061004803
#> 305   0.668484401 -0.6152215258  0.106516759
#> 306   0.514752925 -0.4790178929 -0.017104470
#> 307   0.524022037 -0.3005461661  0.262257974
#> 308   0.752936081 -0.0900894469  0.520149911
#> 309   0.582872349 -0.4446120173  0.094488375
#> 310   0.682070585 -0.5305272976  0.370337608
#> 311   0.520184267 -0.4956750517  0.400627072
#> 312   0.797691870 -0.0980830245  0.294481831
#> 313   0.794131245 -0.2192560093  0.418359129
#> 314   0.444036659 -0.1725632223  0.352690368
#> 315   0.278658898 -0.7795811332  0.220617758
#> 316   0.584303674 -0.6556606403  0.444191317
#> 317   0.638014717 -0.5908965165  0.008892999
#> 318   0.380440934 -0.3413395649  0.386933139
#> 319   0.616329841 -0.1862407449 -0.010856924
#> 320   0.626860571 -0.4790717370 -0.302472104
#> 321   0.416745133 -0.4233091843  0.346972273
#> 322   0.646012848 -0.4147604973  0.065567924
#> 323   0.699416487 -0.1643944306  0.736971339
#> 324   0.572849387 -0.0746489912  0.732614569
#> 325   0.693161951 -0.1853914147  1.095439769
#> 326   0.626860571 -0.4790717370 -0.302472104
#> 327   0.638014717 -0.5908965165  0.008892999
#> 328   0.378102541 -0.9668822494  0.046675509
#> 329   0.658315560 -0.4617453191  0.131113581
#> 330   0.526760906 -0.6346257046  0.521511965
#> 331   0.637601130 -0.3100033348  0.425364194
#> 332   0.506929292 -0.3516959688  0.480852957
#> 333   0.527593421 -0.6790980885  0.480196328
#> 334   0.138855028 -0.7521611972 -0.410991938
#> 335   0.619482976 -0.5067553036  0.160837426
#> 336   0.316781542 -0.8299339889 -0.234740506
#> 337   0.747388455 -0.2299008892  0.629457501
#> 338   0.495498817 -0.5775514746  0.034980529
#> 339   0.603574864 -0.3598764891  0.218675214
#> 340   0.591681927 -0.0521393595  0.783676797
#> 341   0.576178619  0.1940371568  0.527094213
#> 342   0.488048683 -0.5189840996  0.049837700
#> 343   0.109471338 -0.4574033337  0.159471439
#> 344   0.175431954 -0.3454204533  0.836891512
#> 345   0.485605724 -0.6055016403  0.044789338
#> 346   0.014883845 -1.4521228523 -0.016827127
#> 347   0.799766938 -0.0605214155  0.358723532
#> 348   0.842193544 -0.1041899699  0.343848401
#> 349   0.383769504 -0.6749125059 -0.225949823
#> 350   0.346278204 -0.1072102523  0.520558097
#> 351   0.524134345 -0.8351105987 -0.487963506
#> 352   0.579012596 -0.1080902667  0.674712533
#> 353   0.518832391 -0.1613999607  0.672197362
#> 354   0.505989675 -0.6689697244 -0.087146129
#> 355   0.536384486 -0.4394800819  0.405079980
#> 356   0.228311018 -0.5007423370  0.054234509
#> 357   0.577120748 -0.6516643045 -0.001479730
#> 358   0.371357152 -0.6203776546 -0.465191141
#> 359   0.290258619 -0.1306330445  0.621888495
#> 360   0.716021119 -0.5555646713 -0.560857091
#> 361   0.432945115 -0.5232993267  0.421315336
#> 362   0.709666889 -0.2740009158  0.230536948
#> 363   0.329343341 -0.6902772029 -0.023930234
#> 364   0.536945712 -0.0649351622  0.563390201
#> 365   0.669772842 -0.4033058970  0.120854799
#> 366   0.319438895 -0.7253033198  0.412935038
#> 367   0.392512435 -0.4651549500  0.907835135
#> 368   0.717779999 -0.5583900535 -0.480814015
#> 369   0.605805327 -0.2247393615  0.461890869
#> 370   0.490290714 -0.2228036642  0.719159747
#> 371   0.503127414 -0.6932721257 -0.289710781
#> 372   0.707225363 -0.0324680183  0.693385545
#> 373   0.568660775 -0.6371843813 -0.061004803
#> 374   0.699416487 -0.1643944306  0.736971339
#> 375   0.389787306 -0.1833853724  0.619175221
#> 376   0.630750326 -0.3469106943  0.259351051
#> 377   0.753646289 -0.3915816041 -0.393858217
#> 378   0.387355346 -0.7129194877  0.439357374
#> 379   0.617887981 -0.2913004765 -0.067841087
#> 380   0.415878606 -0.4944832880 -0.033366470
#> 381   0.699120848  0.2508079242  0.964309027
#> 382   0.606463384 -0.6659632134 -0.542254939
#> 383   0.797742471  0.0395323470  0.709947461
#> 384   0.450854077 -0.1720008181  0.324784157
#> 385   0.501137625 -0.8048919399 -0.244850041
#> 386   0.699416487 -0.1643944306  0.736971339
#> 387   0.311998599 -0.6434425186 -0.585341336
#> 388   0.605327024 -0.1174581853  0.351113482
#> 389   0.486506070 -0.3778933262  0.102684963
#> 390   0.683608359 -0.1626286425  0.222504720
#> 391   0.533474727 -0.4958831149  0.290775482
#> 392   0.536384486 -0.4394800819  0.405079980
#> 393   0.410719649 -0.3501400916  0.041166231
#> 394   0.605534915 -0.5156753429  0.168580697
#> 395   0.616773077 -0.5625817189  0.180450838
#> 396   0.684094057 -0.3822165398  0.655270752
#> 397   0.508815980 -0.7208848043  0.007270023
#> 398   0.648600254 -0.4092123108  0.408909293
#> 399   0.624577780 -0.5133122650 -0.043160935
#> 400   0.785048750 -0.1629779324  0.147511499
#> 401   0.765719594 -0.1362386450  0.485680756
#> 402   0.564600229 -0.3547002358 -0.172301469
#> 403   0.813325953  0.1747758534  0.515330554
#> 404   0.684054982 -0.4985642636  0.643095629
#> 405  -0.286690881 -0.5488063999  0.328040810
#> 406   0.566671467 -0.3998167796  0.687601565
#> 407   0.426985838 -0.6191124544  0.387428179
#> 408   0.227211075  0.1065245600  1.238275564
#> 409  -0.152033612 -0.6000109506  0.018588539
#> 410   0.616773077 -0.5625817189  0.180450838
#> 411   0.282111024 -0.4140739383  0.766567022
#> 412   0.541665027 -0.8464010429 -0.359186766
#> 413   0.051006237 -0.3163548119  0.785712417
#> 414   0.386494568 -0.2135309636  0.064910763
#> 415   0.699979359 -0.3127497323  0.137940040
#> 416   0.580575520 -0.5763051332  0.008445292
#> 417   0.495498817 -0.5775514746  0.034980529
#> 418   0.552325584 -0.5424671720  0.040777836
#> 419   0.844231310 -0.0487740540  0.548005716
#> 420   0.621142705 -0.2236736236  0.414681218
#> 421   0.565534508 -0.1929856488 -0.111792059
#> 422   0.500693136 -0.2687209893  0.632724606
#> 423   0.524134345 -0.8351105987 -0.487963506
#> 424   0.454109029 -0.8322827610  0.214310058
#> 425   0.475072951 -0.4908675121  0.452746570
#> 426   0.541665027 -0.8464010429 -0.359186766
#> 427   0.634148647 -0.6213047264  0.106976746
#> 428   0.761033855 -0.1753108003  0.570167890
#> 429   0.523172232 -0.5380418779 -0.195021617
#> 430   0.666123114  0.0439383469  0.833267411
#> 431   0.518832391 -0.1613999607  0.672197362
#> 432   0.469141359 -0.3720085040  0.384112689
#> 433   0.416745133 -0.4233091843  0.346972273
#> 434   0.775846679 -0.2469894472  0.003948734
#> 435   0.465035930 -0.2104732714  0.444295708
#> 436   0.679975347 -0.2567608735  0.125295895
#> 437   0.831603715  0.0457240086  0.640451287
#> 438   0.654266083 -0.3786190786  0.198252652
#> 439   0.495845685 -0.7217559063 -0.518352429
#> 440   0.524134345 -0.8351105987 -0.487963506
#> 441   0.514752925 -0.4790178929 -0.017104470
#> 442   0.584687732  0.4527155916  1.062718902
#> 443   0.153628499 -0.8908120694  0.459500007
#> 444   0.628551233 -0.1239183375  0.089602801
#> 445   0.226206362 -0.4787507017 -0.181747088
#> 446  -0.086020693  0.0274094735  0.468278878
#> 447   0.552308811 -0.7422262392 -0.208287785
#> 448   0.486344761 -0.5312173139 -0.046101632
#> 449   0.254623941 -0.3889360881  0.047870328
#> 450   0.645364722 -0.4605913501  0.139344260
#> 451   0.124046049 -0.3722124972  0.509144650
#> 452   0.670197033  0.1052407710  0.711860677
#> 453   0.621142705 -0.2236736236  0.414681218
#> 454   0.659379103  0.3086675835  1.100072669
#> 455   0.654266083 -0.3786190786  0.198252652
#> 456   0.503277907 -0.0852638656  0.375672117
#> 457   0.517749759 -0.4197160957 -0.141258668
#> 458   0.693161951 -0.1853914147  1.095439769
#> 459   0.860472119 -0.0995337847  0.137105796
#> 460   0.371391374 -0.8149312967 -0.263629408
#> 461   0.322691930 -0.5966449793 -0.390665173
#> 462   0.420407666 -0.5971516521 -0.429582089
#> 463   0.707225363 -0.0324680183  0.693385545
#> 464   0.269896504 -0.5292367172  0.079499878
#> 465   0.639537878 -0.3514054189  0.313414823
#> 466   0.633623954 -0.6045737731 -0.304291395
#> 467   0.319352521 -0.1962154558  0.599518949
#> 468   0.415878606 -0.4944832880 -0.033366470
#> 469   0.444036659 -0.1725632223  0.352690368
#> 470   0.752936081 -0.0900894469  0.520149911
#> 471   0.143088961 -0.2561331388  0.577716248
#> 472   0.422094452 -0.1675585879  0.549298512
#> 473   0.503127414 -0.6932721257 -0.289710781
#> 474   0.383948713 -0.6835819168  0.321766689
#> 475   0.415878606 -0.4944832880 -0.033366470
#> 476   0.675313552 -0.3720852966  0.706400999
#> 477   0.347044642 -0.9043220672  0.035982809
#> 478   0.354104955 -0.6144633667  0.159532908
#> 479   0.004014460 -0.3147825179  0.265477038
#> 480   0.547818885 -0.1594581225  1.036005190
#> 481   0.521619792 -0.3648169202  0.593199137
#> 482   0.111654314 -0.5609125154  0.001423546
#> 483   0.630750326 -0.3469106943  0.259351051
#> 484   0.524134345 -0.8351105987 -0.487963506
#> 485   0.213618124 -1.1408631089 -0.876045959
#> 486   0.467740580 -0.4214672015  0.731209363
#> 487   0.371391374 -0.8149312967 -0.263629408
#> 488   0.547750129 -0.2448818062  0.459651368
#> 489   0.591681927 -0.0521393595  0.783676797
#> 490   0.603574864 -0.3598764891  0.218675214
#> 491   0.797691870 -0.0980830245  0.294481831
#> 492   0.785048750 -0.1629779324  0.147511499
#> 493   0.535430574 -0.5926003047 -0.197710845
#> 494   0.389141338 -0.6833461944  0.168273780
#> 495  -0.113525059  0.1248908414  0.796978327
#> 496   0.636219445 -0.2659242852  0.548682900
#> 497   0.495836604 -0.5224995883  0.380654868
#> 498   0.503127414 -0.6932721257 -0.289710781
#> 499   0.321952351 -0.3232129589  0.635474977
#> 500   0.546945743 -0.1257080800  0.294048711
#> 501   0.794131245 -0.2192560093  0.418359129
#> 502   0.291462801 -0.6236412194 -0.083747569
#> 503   0.453386681 -0.6634389517 -0.467517013
#> 504   0.616437850 -0.1423682869  1.236568305
#> 505   0.522655422 -0.3957636242  0.188850848
#> 506   0.547216050 -0.5270740238  0.602940808
#> 507   0.272377971 -0.5527550232  0.530771799
#> 508   0.704625482 -0.3957921871 -0.015665171
#> 509   0.617779376 -0.3002645851 -0.099681614
#> 510   0.503842913 -0.0137283618  0.614700125
#> 511   0.739781129  0.1501214908  1.607946479
#> 512   0.683648135 -0.1799505322 -0.104824110
#> 513   0.529777636 -0.5891031783 -0.157323610
#> 514   0.582872349 -0.4446120173  0.094488375
#> 515   0.387355346 -0.7129194877  0.439357374
#> 516   0.771307429 -0.3362236978  0.463422200
#> 517   0.459844108 -0.6621186469  0.653922769
#> 518   0.434362869 -0.5878430085  0.333271952
#> 519   0.464336921 -0.0623478163  0.303715685
#> 520   0.875729576 -0.0570343799  0.676968836
#> 521   0.885580719 -0.1359080724  0.068873154
#> 522   0.947281178 -0.1124886599  0.309095682
#> 523   0.290642489 -0.2466967357  0.383002111
#> 524   0.274705762 -0.4344212159  0.417679490
#> 525   0.541665027 -0.8464010429 -0.359186766
#> 526   0.645364722 -0.4605913501  0.139344260
#> 527   0.679975347 -0.2567608735  0.125295895
#> 528   0.638014717 -0.5908965165  0.008892999
#> 529   0.421804702 -0.8729232059 -0.138562388
#> 530   0.345139982 -0.2586839047  0.100820523
#> 531   0.723557821 -0.0377581581  0.612870312
#> 532   0.538691757 -0.4199017375  0.056176619
#> 533   0.607694105 -0.3009320650  0.329534903
#> 534   0.281386167 -0.7523285028 -0.050113506
#> 535   0.605327024 -0.1174581853  0.351113482
#> 536   0.696211013 -0.6135278263  0.131940051
#> 537   0.290437507  0.1142015340  0.926601177
#> 538   0.633623954 -0.6045737731 -0.304291395
#> 539   0.684094057 -0.3822165398  0.655270752
#> 540   0.434736267 -0.5456482476 -0.328757710
#> 541   0.392512435 -0.4651549500  0.907835135
#> 542   0.478491420 -0.7359408405 -0.170419719
#> 543   0.708796633 -0.5070151001  0.151112089
#> 544   0.505989675 -0.6689697244 -0.087146129
#> 545   0.747388455 -0.2299008892  0.629457501
#> 546   0.617054576 -0.2380166450  0.377074383
#> 547   0.503127414 -0.6932721257 -0.289710781
#> 548   0.055573767 -0.0748413925  0.572829640
#> 549   0.648600254 -0.4092123108  0.408909293
#> 550   0.537473594 -0.7121418161  0.112749371
#> 551   0.492243164 -0.3252623714  0.367817200
#> 552   0.469542452 -0.1428172479  0.085190760
#> 553   0.521619792 -0.3648169202  0.593199137
#> 554   0.431842606 -0.5971367958  0.218942999
#> 555   0.699979359 -0.3127497323  0.137940040
#> 556   0.618581060  0.2224936793  0.515332322
#> 557   0.107098709 -0.7356954327  0.008643756
#> 558   0.842193544 -0.1041899699  0.343848401
#> 559   0.439935837 -0.1149729978  0.091215750
#> 560   0.467740580 -0.4214672015  0.731209363
#> 561   0.699979359 -0.3127497323  0.137940040
#> 562   0.254431932 -0.2485085214  0.627962365
#> 563   0.318921393 -0.1182648762  0.305001564
#> 564  -0.270722957  0.1356711073  0.636689330
#> 565   0.367959246 -0.2271309697  0.379789655
#> 566   0.765719594 -0.1362386450  0.485680756
#> 567   0.731298371 -0.3121684161  0.371293456
#> 568   0.510276023 -0.3550367898  0.247163005
#> 569   0.607694105 -0.3009320650  0.329534903
#> 570   0.431097966 -0.4059126040 -0.104226981
#> 571   0.456053179 -0.1072300471  0.641175051
#> 572   0.492243164 -0.3252623714  0.367817200
#> 573   0.341417671 -0.6123203754  0.152632348
#> 574   0.253583516 -0.2207157368  0.359001292
#> 575   0.674138771 -0.4129891463 -0.056240806
#> 576   0.665063346 -0.3742920080  0.748671775
#> 577   0.427373489 -0.8940541523 -0.204252131
#> 578   0.717779999 -0.5583900535 -0.480814015
#> 579   0.427373489 -0.8940541523 -0.204252131
#> 580   0.513490605  0.2975927869  1.019841796
#> 581   0.903293407 -0.0720580610  0.736373488
#> 582   0.356865131 -0.7329502190  0.050123249
#> 583   0.109471338 -0.4574033337  0.159471439
#> 584   0.702988637 -0.5382947323  0.162188079
#> 585  -0.233477841 -0.1976966948  0.492794847
#> 586   0.699416487 -0.1643944306  0.736971339
#> 587   0.569195473 -0.0331657768  1.031632598
#> 588   0.158310734 -0.6683685714 -0.179084488
#> 589   0.898760839  0.1061687672  0.464674081
#> 590   0.559607733 -0.3345745135  0.483929287
#> 591   0.462691697 -0.4675892125  0.500925553
#> 592   0.371391374 -0.8149312967 -0.263629408
#> 593   0.765719594 -0.1362386450  0.485680756
#> 594  -0.005163282 -0.3061660467  0.188406558
#> 595   0.627265249 -0.2020708830  0.641656139
#> 596   0.654430154 -0.2607641783  0.386218899
#> 597   0.261856884 -0.3099663772  0.489270321
#> 598   0.054000259 -0.7169783481  0.221782314
#> 599   0.779556041 -0.2411662319  0.350230000
#> 600   0.402373468 -0.5047439624  0.042668622
#> 601   0.585034256 -0.1899031736  0.604609484
#> 602   0.619482976 -0.5067553036  0.160837426
#> 603   0.522655422 -0.3957636242  0.188850848
#> 604   0.669772842 -0.4033058970  0.120854799
#> 605   0.570824659 -0.2187431226  0.306583776
#> 606   0.478957764 -0.5340557166  0.252336062
#> 607   0.279882040 -0.0325085769  0.701766955
#> 608   0.558643053 -0.5879834139 -0.064830414
#> 609   0.572849387 -0.0746489912  0.732614569
#> 610   0.618333091 -0.0965904515  0.626088997
#> 611   0.761033855 -0.1753108003  0.570167890
#> 612   0.624341955 -0.5033349493 -0.102783347
#> 613   0.356865131 -0.7329502190  0.050123249
#> 614   0.617779376 -0.3002645851 -0.099681614
#> 615   0.572849387 -0.0746489912  0.732614569
#> 616   0.535430574 -0.5926003047 -0.197710845
#> 617   0.527960232 -0.2867591076  0.341165765
#> 618   0.859425046 -0.0091270256  0.096564383
#> 619   0.468713653 -0.4269561786 -0.198315942
#> 620   0.375943587 -0.5255856388  0.155884039
#> 621   0.699979359 -0.3127497323  0.137940040
#> 622   0.457154693 -0.2053050555  0.496677978
#> 623   0.533474727 -0.4958831149  0.290775482
#> 624   0.465035930 -0.2104732714  0.444295708
#> 625   0.691714582 -0.2865414470  0.198229880
#> 626   0.707225363 -0.0324680183  0.693385545
#> 627   0.865936822  0.1753830660  0.961915290
#> 628   0.616329841 -0.1862407449 -0.010856924
#> 629   0.671706545  0.1465995375  0.329153126
#> 630   0.389880286  0.1016719670  0.622372631
#> 631   0.700757912 -0.5621080556 -0.130470673
#> 632   0.608659885 -0.5198231667  0.156806859
#> 633   0.697528857 -0.4441559680  0.145073033
#> 634   0.334497249 -0.0694085745  1.034116005
#> 635   0.444851636 -0.7010967048 -0.348341508
#> 636   0.549698145 -0.1291552621  0.588592974
#> 637   0.709666889 -0.2740009158  0.230536948
#> 638   0.628551233 -0.1239183375  0.089602801
#> 639   0.454013882 -0.6121080962  0.512674179
#> 640   0.907494922 -0.2834480593 -0.101668588
#> 641   0.494770443 -0.6855825738 -0.399309620
#> 642   0.514752925 -0.4790178929 -0.017104470
#> 643   0.265540792 -0.2058666921  0.461504366
#> 644   0.138788826 -0.0079829538  0.546628620
#> 645   0.589014707  0.0710075393  0.777838021
#> 646   0.511667539 -0.5533109002 -0.284383916
#> 647   0.679975347 -0.2567608735  0.125295895
#> 648   0.844231310 -0.0487740540  0.548005716
#> 649   0.558235228  0.0285878707  0.301800060
#> 650  -0.082960600 -0.0053240869  0.342113490
#> 651   0.642450665 -0.5452313026 -0.201705302
#> 652   0.364242091 -0.5130521811 -0.361116628
#> 653   0.654266083 -0.3786190786  0.198252652
#> 654   0.289930473 -0.1706333636  0.287735049
#> 655   0.627265249 -0.2020708830  0.641656139
#> 656   0.729543529 -0.1639455075  0.124150071
#> 657   0.739781129  0.1501214908  1.607946479
#> 658   0.794131245 -0.2192560093  0.418359129
#> 659   0.754346468 -0.2127313338  0.303977629
#> 660   0.468688702 -0.3367335065 -0.147937506
#> 661   0.054000259 -0.7169783481  0.221782314
#> 662   0.486344761 -0.5312173139 -0.046101632
#> 663   0.645364722 -0.4605913501  0.139344260
#> 664   0.541665027 -0.8464010429 -0.359186766
#> 665   0.266574275 -0.4502605527  0.433896971
#> 666   0.536384486 -0.4394800819  0.405079980
#> 667   0.506929292 -0.3516959688  0.480852957
#> 668   0.669772842 -0.4033058970  0.120854799
#> 669   0.670197033  0.1052407710  0.711860677
#> 670   0.320415157 -0.7866608332 -0.341942958
#> 671   0.637601130 -0.3100033348  0.425364194
#> 672  -0.383751161 -0.1175976250  0.665600945
#> 673   0.624341955 -0.5033349493 -0.102783347
#> 674   0.793934776 -0.0215126844  0.385487981
#> 675   0.750265666 -0.5303276124  0.017387173
#> 676   0.441841720 -0.2309798550  0.635153876
#> 677   0.753646289 -0.3915816041 -0.393858217
#> 678   0.392512435 -0.4651549500  0.907835135
#> 679   0.747388455 -0.2299008892  0.629457501
#> 680   0.453056677 -0.3514706494  0.447202856
#> 681   0.378102541 -0.9668822494  0.046675509
#> 682   0.358109799 -0.3721713026  0.594120877
#> 683   0.526760906 -0.6346257046  0.521511965
#> 684   0.665063346 -0.3742920080  0.748671775
#> 685   0.654266083 -0.3786190786  0.198252652
#> 686   0.533474727 -0.4958831149  0.290775482
#> 687   0.336981303 -0.7076102979  0.025565166
#> 688   0.290506285 -0.9070337154  0.024135582
#> 689   0.485605724 -0.6055016403  0.044789338
#> 690  -0.196867465 -0.6114400095  0.157993071
#> 691   0.731606662 -0.0973454543  0.324632344
#> 692   0.485605724 -0.6055016403  0.044789338
#> 693  -0.068606617 -0.2978996259 -0.087411313
#> 694   0.642450665 -0.5452313026 -0.201705302
#> 695   0.779556041 -0.2411662319  0.350230000
#> 696   0.572849387 -0.0746489912  0.732614569
#> 697   0.371400489 -0.3551981406  0.294943799
#> 698   0.646012848 -0.4147604973  0.065567924
#> 699   0.697528857 -0.4441559680  0.145073033
#> 700   0.572640772 -0.1365033759  0.350442683
#> 701   0.777410232 -0.5925627108 -0.357624660
#> 702   0.694124464 -0.5955599723 -0.129602667
#> 703   0.475072951 -0.4908675121  0.452746570
#> 704   0.501677262 -0.5886356478  0.012460715
#> 705   0.074888292 -0.7289544198  0.005698807
#> 706   0.478024863 -0.3229358690  0.618781632
#> 707   0.552753131 -0.6447694171 -0.451488487
#> 708   0.584467947 -0.3110719897 -0.005808336
#> 709   0.537749125 -0.4247878923 -0.163145845
#> 710   0.158503228 -0.0158895562  0.466703884
#> 711   0.591681927 -0.0521393595  0.783676797
#> 712   0.742666527 -0.0903633124  0.645536967
#> 713   0.674138771 -0.4129891463 -0.056240806
#> 714   0.571693379 -0.2890570269  0.699257966
#> 715   0.664993759 -0.5337485816  0.713581521
#> 716   0.261459834 -0.7129607795 -0.070612588
#> 717   0.868319323 -0.1788784962  0.318321087
#> 718   0.743911381 -0.4606106578 -0.017075531
#> 719   0.427025374 -0.3999339978 -0.073761072
#> 720   0.389787306 -0.1833853724  0.619175221
#> 721   0.152337800  0.0861175824  0.908956532
#> 722   0.506929292 -0.3516959688  0.480852957
#> 723   0.611719903  0.0490354603  1.147897509
#> 724   0.403255388 -0.7584733247  0.575129743
#> 725   0.261604346 -0.4250191091 -0.211262710
#> 726   0.723557821 -0.0377581581  0.612870312
#> 727   0.516376975 -0.2339550187  0.261320230
#> 728   0.426985838 -0.6191124544  0.387428179
#> 729   0.631345868 -0.2532170416  0.030348408
#> 730   0.474692878 -0.2108803316  0.259086912
#> 731   0.699416487 -0.1643944306  0.736971339
#> 732   0.268780796 -0.3263337910  0.282701121
#> 733   0.752936081 -0.0900894469  0.520149911
#> 734   0.699120848  0.2508079242  0.964309027
#> 735   0.382919800 -0.3515284378  0.635761616
#> 736   0.501677262 -0.5886356478  0.012460715
#> 737   0.505989675 -0.6689697244 -0.087146129
#> 738   0.347962314  0.0088953562  0.867280872
#> 739   0.565764643 -0.4664557518 -0.066263506
#> 740   0.511352596 -0.4915707857 -0.134373264
#> 741   0.728756135 -0.4142320599 -0.357001715
#> 742   0.522655422 -0.3957636242  0.188850848
#> 743   0.506929292 -0.3516959688  0.480852957
#> 744   0.387355346 -0.7129194877  0.439357374
#> 745   0.603574864 -0.3598764891  0.218675214
#> 746   0.501137625 -0.8048919399 -0.244850041
#> 747   0.645364722 -0.4605913501  0.139344260
#> 748   0.742544731 -0.0176195541  0.405884460
#> 749   0.457154693 -0.2053050555  0.496677978
#> 750   0.316881639 -0.4680475841 -0.137556262
#> 751   0.742544731 -0.0176195541  0.405884460
#> 752   0.251519244 -0.8886610817  0.173213654
#> 753   0.616329841 -0.1862407449 -0.010856924
#> 754   0.488048683 -0.5189840996  0.049837700
#> 755   0.743911381 -0.4606106578 -0.017075531
#> 756   0.425755099 -0.6116472217  0.167825544
#> 757   0.441151275  0.1526927090  0.592606391
#> 758   0.707225363 -0.0324680183  0.693385545
#> 759   0.378373725 -0.7980279450  0.728913304
#> 760   0.511180172 -0.1843581772  0.362053183
#> 761   0.485671761 -0.2187162379  0.124208376
#> 762   0.524134345 -0.8351105987 -0.487963506
#> 763   0.664993759 -0.5337485816  0.713581521
#> 764   0.292473778 -0.4035915630  0.415610913
#> 765   0.281386167 -0.7523285028 -0.050113506
#> 766   0.680554748 -0.5459069867  0.046953704
#> 767   0.506839964 -0.1076813586  0.208696632
#> 768   0.621142705 -0.2236736236  0.414681218
#> 769   0.178790190 -0.4019062488 -0.123119655
#> 770   0.551930892 -0.5705180911  0.460435327
#> 771   0.444084756 -0.4367917206  0.432257071
#> 772   0.627343899 -0.4202702532 -0.263645357
#> 773   0.710040740 -0.1805662318  0.173032582
#> 774   0.521619792 -0.3648169202  0.593199137
#> 775   0.847237469 -0.0089514304  0.631981134
#> 776   0.638014717 -0.5908965165  0.008892999
#> 777   0.669772842 -0.4033058970  0.120854799
#> 778   0.240023443 -0.5048452963  0.248068879
#> 779   0.538691757 -0.4199017375  0.056176619
#> 780   0.484977577 -0.8124994559  0.016686815
#> 781   0.794131245 -0.2192560093  0.418359129
#> 782   0.644420351 -0.1339095856  0.558365896
#> 783   0.363538417 -0.4626745348  0.312413801
#> 784   0.240871205 -0.4246850412 -0.016405172
#> 785   0.398831694 -0.1569394612  0.579712967
#> 786   0.633623954 -0.6045737731 -0.304291395
#> 787   0.691714582 -0.2865414470  0.198229880
#> 788   0.739781129  0.1501214908  1.607946479
#> 789   0.569195473 -0.0331657768  1.031632598
#> 790   0.323756163 -0.7124829510  0.241602408
#> 791   0.558643053 -0.5879834139 -0.064830414
#> 792   0.617779376 -0.3002645851 -0.099681614
#> 793   0.742666527 -0.0903633124  0.645536967
#> 794   0.757995635  0.0729570624  0.732841042
#> 795   0.501137625 -0.8048919399 -0.244850041
#> 796   0.254492959 -0.4073041894  0.254688408
#> 797   0.636219445 -0.2659242852  0.548682900
#> 798   0.535430574 -0.5926003047 -0.197710845
#> 799   0.551930892 -0.5705180911  0.460435327
#> 800   0.531026580 -0.3839951947 -0.097644821
#> 801   0.568660775 -0.6371843813 -0.061004803
#> 802   0.495498817 -0.5775514746  0.034980529
#> 803   0.674138771 -0.4129891463 -0.056240806
#> 804   0.527593421 -0.6790980885  0.480196328
#> 805   0.521619792 -0.3648169202  0.593199137
#> 806   0.603574864 -0.3598764891  0.218675214
#> 807   0.448520883 -0.1237671721  0.642483209
#> 808   0.524854017 -0.3822697482  0.175301919
#> 809   0.320415157 -0.7866608332 -0.341942958
#> 810   0.380872202 -0.3876654281 -0.314354825
#> 811   0.713291618 -0.0509571558  0.330469875
#> 812   0.543402973  0.1934258210  0.846568172
#> 813   0.582872349 -0.4446120173  0.094488375
#> 814   0.847237469 -0.0089514304  0.631981134
#> 815   0.427373489 -0.8940541523 -0.204252131
#> 816   0.475072951 -0.4908675121  0.452746570
#> 817   0.692386374 -0.3186897647  0.428494352
#> 818   0.572640772 -0.1365033759  0.350442683
#> 819   0.362790156 -0.6740529184  0.095085660
#> 820   0.669772842 -0.4033058970  0.120854799
#> 821   0.444308374 -0.7872674419 -0.730080384
#> 822   0.762047419  0.0003828142  0.368494637
#> 823   0.502419667 -0.4178521940  0.665125051
#> 824   0.699416487 -0.1643944306  0.736971339
#> 825   0.448520883 -0.1237671721  0.642483209
#> 826   0.421804702 -0.8729232059 -0.138562388
#> 827   0.290437507  0.1142015340  0.926601177
#> 828   0.747388455 -0.2299008892  0.629457501
#> 829   0.431842606 -0.5971367958  0.218942999
#> 830   0.669772842 -0.4033058970  0.120854799
#> 831   0.653992236 -0.3326559955 -0.142232229
#> 832   0.531596269  0.0145020787  0.594870995
#> 833   0.531957645 -0.5600875976  0.461378659
#> 834   0.278692295 -0.4115517919  0.423732126
#> 835   0.465035930 -0.2104732714  0.444295708
#> 836   0.490290714 -0.2228036642  0.719159747
#> 837   0.568660775 -0.6371843813 -0.061004803
#> 838   0.478957764 -0.5340557166  0.252336062
#> 839   0.624577780 -0.5133122650 -0.043160935
#> 840   0.479758911 -0.2738241371  0.446058749
#> 841   0.529777636 -0.5891031783 -0.157323610
#> 842   0.524134345 -0.8351105987 -0.487963506
#> 843   0.467740580 -0.4214672015  0.731209363
#> 844   0.606119177 -0.1678707857  0.249973915
#> 845   0.594714782 -0.2956859225  1.037801727
#> 846   0.555257708 -0.3679693436 -0.036787770
#> 847   0.786166676 -0.1192555344  0.126910214
#> 848   0.831603715  0.0457240086  0.640451287
#> 849   0.399310156 -0.8715022149 -0.107092594
#> 850   0.339478956 -0.5656020157 -0.067049342
#> 851   0.654266083 -0.3786190786  0.198252652
#> 852   0.511667539 -0.5533109002 -0.284383916
#> 853   0.294513494 -0.8864447839  0.182930893
#> 854   0.302393079 -0.3517148571  0.184014432
#> 855   0.432945115 -0.5232993267  0.421315336
#> 856   0.371400489 -0.3551981406  0.294943799
#> 857   0.624341955 -0.5033349493 -0.102783347
#> 858   0.665063346 -0.3742920080  0.748671775
#> 859   0.444851636 -0.7010967048 -0.348341508
#> 860   0.863371475 -0.0731706021  0.547928109
#> 861   0.619482976 -0.5067553036  0.160837426
#> 862   0.304069926 -0.3937111495  0.254240019
#> 863   0.566671467 -0.3998167796  0.687601565
#> 864   0.765719594 -0.1362386450  0.485680756
#> 865   0.402797755 -0.6533415746 -0.258213589
#> 866   0.538691757 -0.4199017375  0.056176619
#> 867   0.243440742 -0.5920352794 -0.044161595
#> 868   0.707225363 -0.0324680183  0.693385545
#> 869   0.501137625 -0.8048919399 -0.244850041
#> 870   0.395072967 -0.3552089974  0.046118662
#> 871   0.380872202 -0.3876654281 -0.314354825
#> 872   0.547750129 -0.2448818062  0.459651368
#> 873   0.419836547 -0.6386619038 -0.012974204
#> 874   0.651883035  0.0161908479  0.164602958
#> 875   0.316781542 -0.8299339889 -0.234740506
#> 876   0.574285930 -0.2159723826  0.070748829
#> 877   0.425755099 -0.6116472217  0.167825544
#> 878   0.639537878 -0.3514054189  0.313414823
#> 879   0.801262739 -0.3486654854  0.446505635
#> 880   0.771307429 -0.3362236978  0.463422200
#> 881   0.801262739 -0.3486654854  0.446505635
#> 882   0.608659885 -0.5198231667  0.156806859
#> 883   0.181425055 -1.1983784346 -0.042403667
#> 884   0.523172232 -0.5380418779 -0.195021617
#> 885   0.545508401 -0.6679838469 -0.371725521
#> 886   0.532218089 -0.1323387295  0.327747351
#> 887   0.483073286 -0.7434500815  0.239735382
#> 888   0.340847924 -0.4738938914  0.253389853
#> 889   0.367682594 -1.2130299689 -0.504764258
#> 890   0.488048683 -0.5189840996  0.049837700
#> 891   0.552325584 -0.5424671720  0.040777836
#> 892   0.188529864 -1.0119256008 -0.052142597
#> 893   0.455147201 -0.0544168636  0.854748100
#> 894   0.558643053 -0.5879834139 -0.064830414
#> 895   0.674138771 -0.4129891463 -0.056240806
#> 896   0.842193544 -0.1041899699  0.343848401
#> 897   0.552522043 -0.7322225660  0.232800211
#> 898   0.362309599 -0.4206286509 -0.010039292
#> 899   0.645364722 -0.4605913501  0.139344260
#> 900   0.700757912 -0.5621080556 -0.130470673
#> 901   0.145593971 -0.7678828657  0.052437868
#> 902   0.565764643 -0.4664557518 -0.066263506
#> 903   0.599783104 -0.6283307185 -0.423473207
#> 904   0.670227244  0.1002033364  0.372333590
#> 905   0.805545140 -0.3516560983  0.211187019
#> 906   0.546945743 -0.1257080800  0.294048711
#> 907   0.707225363 -0.0324680183  0.693385545
#> 908  -0.393643337 -0.6286473815 -0.160487799
#> 909   0.259058369 -0.3941285564  0.764653733
#> 910   0.568660775 -0.6371843813 -0.061004803
#> 911   0.579012596 -0.1080902667  0.674712533
#> 912   0.617672391 -0.4109576812  0.332166684
#> 913   0.363538417 -0.4626745348  0.312413801
#> 914   0.188529864 -1.0119256008 -0.052142597
#> 915   0.541665027 -0.8464010429 -0.359186766
#> 916   0.531596269  0.0145020787  0.594870995
#> 917   0.517873584 -0.3095883697  0.827168989
#> 918   0.552753131 -0.6447694171 -0.451488487
#> 919   0.752936081 -0.0900894469  0.520149911
#> 920   0.524854017 -0.3822697482  0.175301919
#> 921   0.761033855 -0.1753108003  0.570167890
#> 922   0.630750326 -0.3469106943  0.259351051
#> 923   0.349486014 -0.0641446103  0.364352727
#> 924   0.541768367 -0.7768375118 -0.289799615
#> 925   0.541665027 -0.8464010429 -0.359186766
#> 926   0.706247464 -0.5719142102  0.105251656
#> 927   0.526760906 -0.6346257046  0.521511965
#> 928   0.582872349 -0.4446120173  0.094488375
#> 929   0.621142705 -0.2236736236  0.414681218
#> 930   0.779556041 -0.2411662319  0.350230000
#> 931   0.559607733 -0.3345745135  0.483929287
#> 932   0.594714782 -0.2956859225  1.037801727
#> 933   0.351672518 -0.3839034255  0.265871718
#> 934   0.616773077 -0.5625817189  0.180450838
#> 935   0.765719594 -0.1362386450  0.485680756
#> 936   0.607694105 -0.3009320650  0.329534903
#> 937   0.434362869 -0.5878430085  0.333271952
#> 938   0.511849097 -0.1208242392  0.210628122
#> 939   0.709666889 -0.2740009158  0.230536948
#> 940   0.605534915 -0.5156753429  0.168580697
#> 941   0.141275211 -0.5502521855  0.428341745
#> 942   0.524095152 -0.8149889133 -0.707560960
#> 943   0.638014717 -0.5908965165  0.008892999
#> 944   0.742666527 -0.0903633124  0.645536967
#> 945   0.514752925 -0.4790178929 -0.017104470
#> 946   0.470498661 -0.7860215242 -0.031421425
#> 947   0.754346468 -0.2127313338  0.303977629
#> 948   0.730856619 -0.3977310009 -0.008600514
#> 949   0.478491420 -0.7359408405 -0.170419719
#> 950   0.739781129  0.1501214908  1.607946479
#> 951  -0.080898010  0.0795886597  0.487044923
#> 952   0.658315560 -0.4617453191  0.131113581
#> 953   0.510336283 -0.4134707116  0.055228169
#> 954   0.506007720 -0.3159789537 -0.217112766
#> 955   0.495836604 -0.5224995883  0.380654868
#> 956   0.552308811 -0.7422262392 -0.208287785
#> 957   0.607694105 -0.3009320650  0.329534903
#> 958   0.699979359 -0.3127497323  0.137940040
#> 959   0.717779999 -0.5583900535 -0.480814015
#> 960   0.520790919 -0.1699599936  0.484650913
#> 961   0.529777636 -0.5891031783 -0.157323610
#> 962   0.468713653 -0.4269561786 -0.198315942
#> 963   0.847237469 -0.0089514304  0.631981134
#> 964   0.334942259 -0.6645730260  0.204704300
#> 965   0.358109799 -0.3721713026  0.594120877
#> 966   0.444308374 -0.7872674419 -0.730080384
#> 967   0.597156712 -0.3596004400  1.029754050
#> 968   0.291462801 -0.6236412194 -0.083747569
#> 969   0.645364722 -0.4605913501  0.139344260
#> 970   0.321380529 -0.4477040356  0.033649294
#> 971   0.330837691  0.3186462868  0.473719676
#> 972   0.623588532 -0.4790322859  0.291896800
#> 973   0.630895105 -0.5085899694 -0.577393465
#> 974   0.475072951 -0.4908675121  0.452746570
#> 975   0.362309599 -0.4206286509 -0.010039292
#> 976   0.777206655  0.0776103774  0.480433093
#> 977   0.526760906 -0.6346257046  0.521511965
#> 978   0.660938562 -0.4159801766  0.887947027
#> 979   0.699416487 -0.1643944306  0.736971339
#> 980   0.344633322 -0.5829736194 -0.020933826
#> 981   0.266574275 -0.4502605527  0.433896971
#> 982   0.503127414 -0.6932721257 -0.289710781
#> 983   0.752936081 -0.0900894469  0.520149911
#> 984   0.224377925 -0.1734484040  0.322919047
#> 985   0.245129118 -1.0573065565 -0.159782948
#> 986   0.358109799 -0.3721713026  0.594120877
#> 987   0.508815980 -0.7208848043  0.007270023
#> 988   0.172613434 -0.2961069996 -0.057778935
#> 989   0.785048750 -0.1629779324  0.147511499
#> 990   0.582872349 -0.4446120173  0.094488375
#> 991   0.508949138 -0.1104347427  0.248518305
#> 992   0.431952372 -0.4851050772 -0.388775429
#> 993   0.624577780 -0.5133122650 -0.043160935
#> 994   0.621937742 -0.3342982826  0.551703667
#> 995   0.585034256 -0.1899031736  0.604609484
#> 996   0.405827763 -0.7769411518 -0.051808543
#> 997   0.603574864 -0.3598764891  0.218675214
#> 998   0.693429380 -0.4928519139 -0.301461491
#> 999   0.691666359 -0.3580102252  0.101451694
#> 1000  0.665063346 -0.3742920080  0.748671775
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>  <dbl>  <dbl>
#>  1      1        4       70        1  0.631  -0.347 0.259 
#>  2      2        4       70        1  0.571  -0.219 0.307 
#>  3      3        4       70        1  0.495  -0.578 0.0350
#>  4      4        4       70        1 -0.0541 -0.397 0.0132
#>  5      5        4       70        1 -0.418  -0.475 0.0786
#>  6      6        4       70        1  0.427  -0.122 0.266 
#>  7      7        4       70        1  0.358  -0.372 0.594 
#>  8      8        4       70        1  0.520  -0.496 0.401 
#>  9      9        4       70        1  0.766  -0.136 0.486 
#> 10     10        4       70        1  0.681  -0.546 0.0470
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  7.52  49.5  1.30 0.152 23.2  770.        1150.
#> 2      1    14     4    70     1  7.52  49.5  1.30 0.152  5.67   0.0000371  281.
#> 3      2     1     4    70     1  7.08  56.2  1.36 0.126 21.2  736.        1195.
#> 4      2    14     4    70     1  7.08  56.2  1.36 0.126  6.94   0.0000157  391.
#> 5      3     1     4    70     1  6.57  39.3  1.04 0.167 25.3  930.         995.
#> 6      3    14     4    70     1  6.57  39.3  1.04 0.167  6.10   0.00132    240.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#>