Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object

A posologyr list, created by the posologyr function.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01,n_sample=1e4,n_resample=1e3)
#> 
#> $eta
#>            ETA_Cl       ETA_Vc       ETA_Ka
#> 1     0.359107011 -0.618114373  0.360089110
#> 2     0.405612847 -0.332157557  0.122328511
#> 3     0.772581363 -0.410544146 -0.119629742
#> 4     0.554712891 -0.492602078  0.035517660
#> 5     0.350836142 -0.634682581 -0.334944826
#> 6     0.445043308 -0.461319024  0.565697962
#> 7     0.568021068 -0.686554753  0.194385927
#> 8     0.273692910 -0.589765220  0.167615493
#> 9     0.550196261 -0.467194952  0.268555943
#> 10    0.697028053 -0.379919120 -0.289006286
#> 11    0.554712891 -0.492602078  0.035517660
#> 12    0.650188746 -0.251956321  0.691063030
#> 13    0.531432384 -0.098979338  0.067446499
#> 14    0.520458840 -0.298118684 -0.039692325
#> 15    0.641468298 -0.586351578  0.002195579
#> 16    0.692784408  0.101940638  0.792039536
#> 17    0.636588883 -0.367147004  0.332888301
#> 18    0.341706409 -0.461240508  0.102232027
#> 19    0.361448891 -0.659559950 -0.149409502
#> 20    0.437797086 -0.750549802  0.437933777
#> 21    0.647841321 -0.495814904  0.243180436
#> 22    0.892696933  0.134120012  0.759068339
#> 23    0.618803282 -0.146067545  0.337977161
#> 24    0.455510590 -0.443394941  0.120144318
#> 25    0.786005013  0.278199833  0.776203744
#> 26    0.619745681 -0.382856381  0.380307466
#> 27    0.364526129 -0.658997470 -0.107311090
#> 28    0.751186111 -0.279388642  0.929987870
#> 29    0.747220493 -0.277083036  0.253456943
#> 30    0.616428287 -0.294859622  0.298831140
#> 31    0.544386367 -0.561198431 -0.054121386
#> 32    0.577361767 -0.114379879  0.504797772
#> 33    0.037953434 -0.383923644 -0.032975629
#> 34    0.518708465 -0.346224398  0.193310755
#> 35    0.569199020 -0.456009964  0.200174201
#> 36    0.662504780 -0.599814717 -0.145540849
#> 37    0.726256023 -0.387765479  0.209754528
#> 38    0.448256927 -0.077700023  0.635086065
#> 39    0.371327016 -1.051612733 -0.420001671
#> 40    0.759678442  0.014107045  0.539186143
#> 41    0.728393620 -0.100756295  0.043202737
#> 42    0.384263586 -0.787192132  0.045723354
#> 43    0.596449855 -0.345931606  0.687200434
#> 44    0.772581363 -0.410544146 -0.119629742
#> 45    0.242151807 -0.341845436  0.151135284
#> 46    0.651197741 -0.380709294  0.103554970
#> 47    0.726256023 -0.387765479  0.209754528
#> 48    0.824731978  0.005315237  0.417440586
#> 49    0.564700102 -0.417930604  0.087843692
#> 50    0.557231242 -0.243845781  0.311296538
#> 51    0.160313241 -0.625530806  0.246436789
#> 52    0.580396873 -0.358603753 -0.182723748
#> 53    0.603847514 -0.484299635  0.072151406
#> 54    0.652285798 -0.314496718  0.508986084
#> 55    0.566263516 -0.321066399  0.716030685
#> 56    0.678737563  0.042346377  0.670978897
#> 57    0.504740508 -0.461227713 -0.037220993
#> 58    0.330684625 -0.238622996  0.032205339
#> 59   -0.006456895 -0.938441087 -0.358542195
#> 60    0.190497542 -0.828647682  0.594153676
#> 61    0.736207913 -0.323027188  0.163841116
#> 62    0.637493224 -0.011420123  0.171803807
#> 63    0.554712891 -0.492602078  0.035517660
#> 64    0.052406166 -1.123757972 -0.713042663
#> 65    0.604785289 -0.852039873 -0.416829879
#> 66    0.418364114 -0.612902938  0.177840304
#> 67    0.908321559  0.122150446  0.457880025
#> 68    0.612737134 -0.563561449  0.772534921
#> 69    0.507251922 -0.397334264  0.178983283
#> 70    0.647841321 -0.495814904  0.243180436
#> 71    0.620933237 -0.020848379  0.636024509
#> 72    0.742652318 -0.317031223  0.090969160
#> 73    0.714802733 -0.329736360  0.133694863
#> 74    0.769626837 -0.286006102  0.097688657
#> 75    0.632262505 -0.325892971  0.366164462
#> 76    0.382826031 -0.833825423 -0.078399477
#> 77    0.692373591 -0.464377793  0.304429339
#> 78    0.535726349 -0.437725063  0.590694876
#> 79    0.314725419 -0.363541529  0.369474038
#> 80    0.651191244 -0.329776855  0.227197648
#> 81    0.214239376 -1.041174259 -0.082006661
#> 82    0.418364114 -0.612902938  0.177840304
#> 83    0.396651396 -0.950167245 -0.220993672
#> 84    0.652285798 -0.314496718  0.508986084
#> 85    0.264796727 -0.311097041  0.020678072
#> 86    0.650188746 -0.251956321  0.691063030
#> 87    0.619745681 -0.382856381  0.380307466
#> 88    0.605164309 -0.798562124 -0.313214329
#> 89    0.676196717 -0.487057769  0.098249404
#> 90    0.476617235 -0.263397988  0.073535123
#> 91    0.838441200  0.048988481  0.567287570
#> 92    0.310721397 -0.630023790  0.230186128
#> 93    0.403706274 -0.406755366  0.398168263
#> 94    0.742238197 -0.046079827  0.275613625
#> 95    0.717770647 -0.343957306 -0.131808554
#> 96    0.493903926 -0.198434873  1.155167866
#> 97    0.835527563  0.052845257  0.661870554
#> 98    0.388703085 -0.893027588 -0.646910457
#> 99    0.348352921 -0.147640703  0.119777550
#> 100   0.668154798 -0.120610106  0.464841725
#> 101   0.544386367 -0.561198431 -0.054121386
#> 102   0.854273693 -0.202395800  0.010827977
#> 103   0.626706253 -0.472921613 -0.495626763
#> 104   0.780603394 -0.007893495  0.575764960
#> 105  -0.197740274 -0.690929396  0.576765052
#> 106   0.589991563 -0.248154969  0.065723159
#> 107   0.566263516 -0.321066399  0.716030685
#> 108   0.418965199 -0.372623263  0.409786047
#> 109   0.755604057  0.118326035  0.749573113
#> 110   0.460920142 -0.705479171  0.309014455
#> 111   0.260759871 -0.503069865  0.179010294
#> 112   0.347430319 -0.343887031  0.073663381
#> 113   0.669959507  0.174010600  0.806114380
#> 114   0.661081114 -0.194233032  0.768084728
#> 115   0.593184541 -0.555507690  0.094833097
#> 116   0.824761749 -0.221795040  0.029473302
#> 117   0.288486592 -1.029424799 -0.701186579
#> 118   0.896224924 -0.196703080  0.288170957
#> 119   0.237160979 -0.595703813 -0.006718105
#> 120   0.422037975 -0.068687507  0.368754786
#> 121   0.743655627 -0.296830382  0.214225366
#> 122   0.417515562 -0.431096083  0.011867171
#> 123   0.634261350 -0.189401737  0.472713220
#> 124   0.982183275 -0.160858326 -0.034023941
#> 125   0.105915382 -0.880998184  0.270921065
#> 126   0.088889197 -0.217432613  0.290094336
#> 127   0.460920142 -0.705479171  0.309014455
#> 128   0.392657096 -0.775740297 -0.368812365
#> 129   0.518708465 -0.346224398  0.193310755
#> 130   0.392657096 -0.775740297 -0.368812365
#> 131   0.396411746 -0.482937388  0.608849000
#> 132   0.337051102 -0.977446753 -0.336058628
#> 133   0.437643794 -0.480776265  0.859336310
#> 134   0.221325180 -0.425893293  0.541503772
#> 135   0.647278910 -0.628365226 -0.079829707
#> 136   0.572775973 -0.526138963  0.407734522
#> 137   0.196760663 -1.012654550 -0.547675369
#> 138   0.494492191 -0.199039795  0.802604145
#> 139   0.480706860 -0.809346403 -0.319370723
#> 140   0.637855034 -0.204684307  0.695384272
#> 141   0.569189364 -0.470908925  0.048369276
#> 142   0.432693844 -0.642528151  0.399330419
#> 143   0.467170968 -0.177366391  0.378751025
#> 144   0.488945533 -0.676959626 -0.338296716
#> 145   0.751186111 -0.279388642  0.929987870
#> 146   0.690107509 -0.316707138  0.021469792
#> 147   0.384263586 -0.787192132  0.045723354
#> 148   0.196222560 -0.377494786 -0.182445943
#> 149   0.568617645 -0.493922142  0.170307169
#> 150   0.552408507 -0.265715860  0.238693378
#> 151   0.705754733 -0.116450703  0.406209492
#> 152   0.546402937 -0.250185983  0.486902492
#> 153   0.566136851 -0.601046252 -0.070622559
#> 154   0.686497386 -0.148311019  0.344182869
#> 155   0.414512133 -0.578557939  0.627922067
#> 156   0.568617645 -0.493922142  0.170307169
#> 157   0.678737563  0.042346377  0.670978897
#> 158   0.396411746 -0.482937388  0.608849000
#> 159   0.490466109 -0.362320855  0.720169031
#> 160   0.593184541 -0.555507690  0.094833097
#> 161   0.318378066 -0.680455665  0.259361301
#> 162   0.581446450 -0.382977350  0.279866558
#> 163   0.595524689 -0.706378017 -0.275404321
#> 164   0.697650358 -0.166965585  0.553993305
#> 165   0.487842891 -0.637355656 -0.328356290
#> 166   0.532839556 -0.322344379  1.084798405
#> 167   0.542910453 -0.444003011  0.029091222
#> 168   0.682822465 -0.593025177  0.085009880
#> 169   0.004587988 -0.380388922  0.237928133
#> 170   0.467133629 -0.387488708  0.695967299
#> 171   0.020656143  0.021266543  0.422695053
#> 172   0.667434051  0.063254802  0.599418854
#> 173   0.435528954  0.307691310  1.220484032
#> 174   0.487545152 -0.159935430  0.639487006
#> 175   0.613406896 -0.368254794  0.412278120
#> 176   0.412951205 -0.464308829  0.600419770
#> 177  -0.232354034 -0.053807327  0.308397641
#> 178   0.171576798 -0.433966967 -0.275470239
#> 179   0.911599325  0.062531484  0.499163545
#> 180  -0.391554363 -0.226126861  0.057141337
#> 181   0.414512133 -0.578557939  0.627922067
#> 182   0.523808082 -0.234221509  0.272075849
#> 183   0.578282410 -0.164108386  0.424013991
#> 184   0.273692910 -0.589765220  0.167615493
#> 185   0.371327016 -1.051612733 -0.420001671
#> 186   0.258149103 -0.353395847  0.486394153
#> 187   0.660642457 -0.275410170  0.232847336
#> 188   0.449297075 -0.563575028  0.199417655
#> 189   0.321532968 -0.304604467  1.026940397
#> 190   0.923884541  0.438572906  0.602315648
#> 191   0.680909282 -0.285679737  0.429883635
#> 192   0.451776825 -0.688813204 -0.162750049
#> 193   0.392699146 -0.095381113  0.712449638
#> 194   0.538264362 -0.526015532 -0.084690119
#> 195   0.264790192  0.252667525  0.466224222
#> 196   0.160041983 -0.282706359  0.255886813
#> 197   0.264904015 -0.403001002  0.288918054
#> 198   0.065460615 -0.480436462 -0.061415641
#> 199   0.578282410 -0.164108386  0.424013991
#> 200   0.288486592 -1.029424799 -0.701186579
#> 201   0.494841900 -0.684760027  0.136315910
#> 202   0.323459509 -0.783984215  0.312793738
#> 203   0.449987082 -0.725068271  0.085636813
#> 204   0.538537128 -0.502669518  0.013714627
#> 205   0.621426593 -0.456127742  0.265427435
#> 206   0.388386385 -0.252087197  0.407731343
#> 207   0.538537128 -0.502669518  0.013714627
#> 208   0.560098049 -0.473806096 -0.263006742
#> 209   0.647841321 -0.495814904  0.243180436
#> 210   0.344339873 -0.748104615  0.716603945
#> 211   0.699839724 -0.415079056 -0.157507415
#> 212   0.708828980 -0.431014259  0.378384615
#> 213   0.755604057  0.118326035  0.749573113
#> 214   0.415122875 -0.806799110 -0.493957324
#> 215   0.211419887 -0.774104475 -0.070565356
#> 216  -0.256740242 -0.428628909  0.131256856
#> 217   0.420870913 -0.293581853  0.867929940
#> 218   0.649081548 -0.565986317 -0.110490920
#> 219   0.893271320 -0.143353134  0.128763061
#> 220   0.629757427 -0.613445028 -0.443548172
#> 221   0.595524689 -0.706378017 -0.275404321
#> 222   0.649081548 -0.565986317 -0.110490920
#> 223   0.552978971 -0.748670001 -0.444800773
#> 224   0.749278280 -0.103147592  0.268221614
#> 225   0.759704578  0.035233302  0.478204536
#> 226   0.381997426 -0.404177007  0.069886710
#> 227   0.505800747 -0.477341753 -0.086041101
#> 228   0.297591427 -0.157047454  0.836235300
#> 229   0.388177623 -0.698444133 -0.051094223
#> 230   0.422037975 -0.068687507  0.368754786
#> 231   0.487196498 -0.778519253 -0.122325253
#> 232   0.633446127 -0.929266034 -0.896214573
#> 233   0.726256023 -0.387765479  0.209754528
#> 234   0.236075623 -0.361719187  0.325062731
#> 235   0.446002428 -0.760664784 -0.257580101
#> 236   0.504153281 -0.222510122  0.476950783
#> 237   0.751186111 -0.279388642  0.929987870
#> 238   0.513107019 -0.843619466 -0.516638465
#> 239   0.581969807 -0.213869439  0.213976708
#> 240   0.666073411 -0.554468199 -0.295331521
#> 241   0.366074817 -0.512885760  0.157638610
#> 242   0.487660038 -0.536367780  0.331839534
#> 243   0.527692342 -0.186747798  0.595727361
#> 244   0.480706860 -0.809346403 -0.319370723
#> 245   0.797975861  0.029476846  0.619378614
#> 246   0.302664357 -0.685124059  0.312179449
#> 247   0.581873682 -0.325833103  0.277747212
#> 248   0.403603345 -0.551464596  0.119066215
#> 249   0.641468298 -0.586351578  0.002195579
#> 250   0.467133629 -0.387488708  0.695967299
#> 251   0.630210230 -0.390718496  0.332359555
#> 252   0.656907071 -0.558353056  0.604380426
#> 253   0.537852439 -0.409655584  0.099894270
#> 254   0.692373591 -0.464377793  0.304429339
#> 255   0.630210230 -0.390718496  0.332359555
#> 256   0.660642457 -0.275410170  0.232847336
#> 257   0.433149786 -0.268292474  0.016633264
#> 258   0.745655264 -0.576937077 -0.369517995
#> 259   0.692373591 -0.464377793  0.304429339
#> 260   0.418364114 -0.612902938  0.177840304
#> 261   0.496743938 -0.415185930  0.530922171
#> 262   0.759678442  0.014107045  0.539186143
#> 263   0.600351236 -0.314022287  0.913098403
#> 264   0.494492191 -0.199039795  0.802604145
#> 265   0.499253629 -0.734777952 -0.460496345
#> 266   0.787763098 -0.355938850 -0.134863636
#> 267   0.347100927 -0.683570485  0.269304990
#> 268   0.538228799 -0.410740772 -0.029457761
#> 269   0.466927804 -0.397937736  0.398761223
#> 270   0.528326180 -0.472422657  0.092099736
#> 271   0.475204049 -0.835595209  0.162364414
#> 272   0.555188316 -0.499876429  0.156198864
#> 273   0.552978971 -0.748670001 -0.444800773
#> 274   0.838245948 -0.109864766  0.518414343
#> 275   0.634261350 -0.189401737  0.472713220
#> 276   0.633327451 -0.482225443  0.326358301
#> 277   0.651191244 -0.329776855  0.227197648
#> 278   0.378320740 -0.855739764 -0.358141140
#> 279   0.502601579 -0.369866990  0.206177540
#> 280   0.449193209 -0.056686086  0.131702675
#> 281   0.316009270 -0.486371707  0.459378337
#> 282   0.455028544 -0.250589033  0.555951813
#> 283   0.560514396 -0.728360311 -0.255712605
#> 284   0.386308252 -0.640843118 -0.266101348
#> 285   0.574562021 -0.649990434 -0.438556444
#> 286  -0.080960735 -0.346037775  0.602333525
#> 287   0.574761805 -0.704538366 -0.338203621
#> 288   0.480706860 -0.809346403 -0.319370723
#> 289   0.559958942 -0.447793417 -0.089755247
#> 290   0.118240448 -0.439287972  0.116231328
#> 291   0.379648732 -0.350675491  0.543387146
#> 292   0.759704578  0.035233302  0.478204536
#> 293   0.386320429  0.087981917  0.584250584
#> 294   0.532345097 -0.292247693  0.068786001
#> 295   0.448174174 -0.101402563  0.213904369
#> 296   0.650188746 -0.251956321  0.691063030
#> 297   0.480859670 -0.643638165  0.453588671
#> 298   0.476309167 -0.124081394  0.341630790
#> 299   0.650188746 -0.251956321  0.691063030
#> 300   0.487744513 -0.796196707 -0.170366113
#> 301   0.460920142 -0.705479171  0.309014455
#> 302   0.581873682 -0.325833103  0.277747212
#> 303   0.487660038 -0.536367780  0.331839534
#> 304   0.751186111 -0.279388642  0.929987870
#> 305   0.484198162 -0.450827522  0.099120513
#> 306   0.608808159 -0.558237660  0.332305205
#> 307   0.225039271 -0.529007005  0.186427041
#> 308   0.144368616 -0.464723148 -0.013782793
#> 309   0.475204049 -0.835595209  0.162364414
#> 310   0.021231604 -1.189688489 -0.477542125
#> 311   0.637855034 -0.204684307  0.695384272
#> 312   0.384263586 -0.787192132  0.045723354
#> 313   0.632262505 -0.325892971  0.366164462
#> 314   0.418739185 -0.681259921  0.104188727
#> 315   0.431248072 -0.516045813 -0.296242998
#> 316   0.292382938 -1.220047038 -0.296729169
#> 317   0.791198883 -0.319338982 -0.003488426
#> 318   0.496743938 -0.415185930  0.530922171
#> 319   0.361318522 -0.057314729  0.499705272
#> 320   0.412890372 -0.352614554  0.724436514
#> 321   0.382826031 -0.833825423 -0.078399477
#> 322   0.797975861  0.029476846  0.619378614
#> 323   0.341841822 -0.898717435 -0.196427866
#> 324   0.353353072 -0.894859068  0.087457231
#> 325   0.520458840 -0.298118684 -0.039692325
#> 326   0.577361767 -0.114379879  0.504797772
#> 327   0.105272428 -0.651070767 -0.277655655
#> 328   0.697650358 -0.166965585  0.553993305
#> 329   0.494841900 -0.684760027  0.136315910
#> 330   0.316703951 -0.094522595  0.739490077
#> 331   0.116153418 -0.194677901  0.334611924
#> 332   0.497274758 -0.647184681 -0.056799484
#> 333   0.418364114 -0.612902938  0.177840304
#> 334   0.537852439 -0.409655584  0.099894270
#> 335   0.162977407  0.022066081  0.953833288
#> 336   0.295554807 -0.335128116  0.517936850
#> 337   0.379706848 -0.584410585 -0.137615817
#> 338   0.347430319 -0.343887031  0.073663381
#> 339   0.626179353 -0.114478038  0.347735407
#> 340   0.305847179 -0.837000263 -0.470460643
#> 341   0.676196717 -0.487057769  0.098249404
#> 342   0.736207913 -0.323027188  0.163841116
#> 343   0.317340300 -0.476345783  0.230519913
#> 344   0.552998713 -0.307179314  0.308874448
#> 345   0.075513262 -0.884846242 -0.115199084
#> 346   0.606159708 -0.710704246 -0.032628773
#> 347   0.686497386 -0.148311019  0.344182869
#> 348   0.315719496 -0.261375555  0.354680925
#> 349   0.595524689 -0.706378017 -0.275404321
#> 350   0.600351236 -0.314022287  0.913098403
#> 351   0.285078057 -1.045133276 -0.707816067
#> 352   0.574296576 -0.395634276  0.735745828
#> 353   0.651197741 -0.380709294  0.103554970
#> 354   0.560514396 -0.728360311 -0.255712605
#> 355   0.728568004 -0.385075510  0.158547135
#> 356   0.743639298 -0.272477917  1.048098989
#> 357   0.473691781 -0.270451427  0.306308017
#> 358   0.535726349 -0.437725063  0.590694876
#> 359   0.511134813 -0.244365674  0.269864769
#> 360   0.612737134 -0.563561449  0.772534921
#> 361   0.210002183 -0.311574776  0.359608265
#> 362   0.417515562 -0.431096083  0.011867171
#> 363   0.660642457 -0.275410170  0.232847336
#> 364   0.569199020 -0.456009964  0.200174201
#> 365   0.644287716 -0.052170933  0.800251562
#> 366   0.310721397 -0.630023790  0.230186128
#> 367   0.574562021 -0.649990434 -0.438556444
#> 368   0.647278910 -0.628365226 -0.079829707
#> 369   0.353395300 -0.149859827  0.219367896
#> 370   0.373474142 -0.215976427  0.064898262
#> 371   0.335190453 -0.351546432  0.121127620
#> 372   0.546402937 -0.250185983  0.486902492
#> 373   0.581969807 -0.213869439  0.213976708
#> 374   0.615233899 -0.059305362  0.423593828
#> 375   0.714802733 -0.329736360  0.133694863
#> 376   0.454047796 -0.214597378  0.156267408
#> 377   0.487660038 -0.536367780  0.331839534
#> 378   0.629951966 -0.129384995  0.578053501
#> 379   0.477483081 -0.621487755  0.021623006
#> 380   0.638968312 -0.197718092 -0.130966406
#> 381   0.737705572 -0.066932008  0.576935289
#> 382   0.616428287 -0.294859622  0.298831140
#> 383   0.453455972 -0.208088893  1.011234955
#> 384   0.477549214 -0.236374669  0.340636176
#> 385   0.058820038 -0.532756450  0.088081735
#> 386   0.499253629 -0.734777952 -0.460496345
#> 387   0.856551611 -0.115652920  0.396471085
#> 388   0.540095134 -0.622503435 -0.103610176
#> 389   0.426668794 -0.659787162 -0.151462879
#> 390   0.559542998 -0.471917053 -0.227514472
#> 391   0.554712891 -0.492602078  0.035517660
#> 392   0.477483081 -0.621487755  0.021623006
#> 393   0.342989311 -1.129623692 -0.542354232
#> 394   0.539792421 -0.404856017  0.371884150
#> 395   0.732616569 -0.094330505  0.562816454
#> 396   0.660060577 -0.226780277  0.569781150
#> 397   0.335190453 -0.351546432  0.121127620
#> 398   0.956957848  0.029603777  0.853966551
#> 399   0.619745681 -0.382856381  0.380307466
#> 400   0.717171732  0.101702396  0.738915443
#> 401   0.687624692 -0.387794135 -0.087599433
#> 402   0.736207913 -0.323027188  0.163841116
#> 403   0.157504482 -0.702927510 -0.475346130
#> 404   0.403603345 -0.551464596  0.119066215
#> 405   0.699713377 -0.410809444 -0.346154997
#> 406   0.465194068 -0.603542560  0.595175965
#> 407   0.581969807 -0.213869439  0.213976708
#> 408   0.586888622 -0.573141368  0.442859483
#> 409   0.513107019 -0.843619466 -0.516638465
#> 410   0.707569929  0.117010079  0.256591710
#> 411   0.563736866 -0.106306322  0.609053868
#> 412   0.684054886  0.040848218  0.492697936
#> 413   0.449297075 -0.563575028  0.199417655
#> 414   0.759678442  0.014107045  0.539186143
#> 415   0.550196261 -0.467194952  0.268555943
#> 416   0.717171732  0.101702396  0.738915443
#> 417   0.660060577 -0.226780277  0.569781150
#> 418   0.737850403 -0.030731797  0.290846156
#> 419   0.273592432 -0.289104219  0.006760626
#> 420   0.560593985 -0.489185325  0.339346336
#> 421   0.538264362 -0.526015532 -0.084690119
#> 422   0.856551611 -0.115652920  0.396471085
#> 423   0.460236943 -0.324839445  0.676198638
#> 424   0.256241587 -0.617099695  0.237808385
#> 425   0.574562021 -0.649990434 -0.438556444
#> 426   0.345099495 -0.627379937  0.166966820
#> 427   0.687624692 -0.387794135 -0.087599433
#> 428   0.544709723 -0.175698641  0.784642018
#> 429   0.560593985 -0.489185325  0.339346336
#> 430   0.708901282 -0.630350939 -0.598217327
#> 431   0.584170710 -0.793375572 -0.153686240
#> 432   0.173215801 -0.457870869  0.594974619
#> 433   0.603669793 -0.013765739  0.406548840
#> 434   0.525949174 -0.602774045 -0.469427040
#> 435   0.465194068 -0.603542560  0.595175965
#> 436   0.651197741 -0.380709294  0.103554970
#> 437   0.569199020 -0.456009964  0.200174201
#> 438   0.278634602 -0.381800476 -0.045523353
#> 439   0.774491826 -0.134119475  0.899600072
#> 440   0.529530987 -0.424205772  0.515162221
#> 441   0.349490951 -0.422808978  0.303682270
#> 442   0.662504780 -0.599814717 -0.145540849
#> 443   0.621426593 -0.456127742  0.265427435
#> 444   0.446415536 -0.424818648  0.007327336
#> 445   0.347430319 -0.343887031  0.073663381
#> 446   0.580030202 -0.190164957  0.470545486
#> 447   0.535726349 -0.437725063  0.590694876
#> 448   0.453407241 -0.082184402  0.468615956
#> 449   0.507251922 -0.397334264  0.178983283
#> 450   0.493615282 -0.658645844  0.368343445
#> 451   0.304922741 -0.551367544 -0.063722024
#> 452  -0.436199576 -0.766889274 -0.166499267
#> 453   0.357883710 -0.672460407 -0.058998488
#> 454   0.633327451 -0.482225443  0.326358301
#> 455   0.371327016 -1.051612733 -0.420001671
#> 456   0.337051102 -0.977446753 -0.336058628
#> 457   0.630210230 -0.390718496  0.332359555
#> 458   0.388710140 -0.703631408  0.715928640
#> 459   0.279388278 -0.071576086  0.548089253
#> 460   0.751186111 -0.279388642  0.929987870
#> 461   0.515483908  0.045124250  0.539556631
#> 462   0.581446450 -0.382977350  0.279866558
#> 463   0.781651349  0.039065115  0.474424991
#> 464   0.650188746 -0.251956321  0.691063030
#> 465   0.439613899 -0.436102170  1.371669588
#> 466   0.394347165 -0.384350361  0.247749035
#> 467   0.568617645 -0.493922142  0.170307169
#> 468   0.301317855 -0.139216771  0.219285364
#> 469   0.649031015 -0.362371165  0.304897882
#> 470   0.680909282 -0.285679737  0.429883635
#> 471   0.554712891 -0.492602078  0.035517660
#> 472   0.593184541 -0.555507690  0.094833097
#> 473   0.791328949 -0.190179249  0.502791718
#> 474   0.661081114 -0.194233032  0.768084728
#> 475   0.487196498 -0.778519253 -0.122325253
#> 476   0.578282410 -0.164108386  0.424013991
#> 477   0.451776825 -0.688813204 -0.162750049
#> 478   0.629757427 -0.613445028 -0.443548172
#> 479   0.566136851 -0.601046252 -0.070622559
#> 480   0.493698342 -0.197129078  0.733571611
#> 481   0.107475002 -0.819538414 -0.534228135
#> 482   0.418320733 -0.479552150  0.387855249
#> 483   0.661669030 -0.653052532 -0.568606719
#> 484   0.345875673 -0.175987456  0.561146161
#> 485   0.824761749 -0.221795040  0.029473302
#> 486   0.644287716 -0.052170933  0.800251562
#> 487   0.388710140 -0.703631408  0.715928640
#> 488   0.426546040 -0.642041953 -0.063800483
#> 489   0.728568004 -0.385075510  0.158547135
#> 490   0.619745681 -0.382856381  0.380307466
#> 491   0.487196498 -0.778519253 -0.122325253
#> 492   0.726256023 -0.387765479  0.209754528
#> 493   0.351465716 -0.914421775  0.086485695
#> 494   0.176894807 -0.301230609  0.025040273
#> 495   0.678737563  0.042346377  0.670978897
#> 496   0.468339337 -0.662511565  0.318505081
#> 497   0.743655627 -0.296830382  0.214225366
#> 498   0.400612253 -0.387419517  0.465241669
#> 499   0.524850236 -0.218462453  0.501276182
#> 500   0.759678442  0.014107045  0.539186143
#> 501   0.093856141 -0.574153927 -0.164085740
#> 502   0.466927804 -0.397937736  0.398761223
#> 503   0.368096940 -0.085703395  0.293855696
#> 504   0.497133070 -0.009662606  0.199026335
#> 505   0.255264699 -0.526657167 -0.353685546
#> 506   0.344454530 -0.294293955  0.936436691
#> 507   0.571197165 -0.285121601  0.287426757
#> 508   0.101992480 -1.130218338 -0.406661713
#> 509   0.100426195  0.227201345  0.527025153
#> 510   0.557500477 -0.177158434  0.282154159
#> 511   0.754573550 -0.182502280  0.222213400
#> 512   0.650188746 -0.251956321  0.691063030
#> 513   0.603578490 -0.091561160  0.398468560
#> 514   0.537852439 -0.409655584  0.099894270
#> 515   0.702102269  0.140720290  0.706826138
#> 516   0.020500734 -1.288012869  0.142680496
#> 517   0.387782805 -0.213313744  0.964111036
#> 518   0.680971636 -0.090509229  0.670039922
#> 519   0.399194438 -0.507987506  0.130180824
#> 520   0.489740951 -0.370226310 -0.022541344
#> 521   0.481801284 -0.605959391  0.115419056
#> 522   0.613406896 -0.368254794  0.412278120
#> 523   0.450032772 -0.402978035  0.347839414
#> 524   0.680909282 -0.285679737  0.429883635
#> 525   0.446614011  0.188885515  0.652688789
#> 526   0.559542998 -0.471917053 -0.227514472
#> 527   0.478493012 -0.158953284  0.037090269
#> 528   0.241470237  0.142332430  0.696594313
#> 529   0.680909282 -0.285679737  0.429883635
#> 530   0.751186111 -0.279388642  0.929987870
#> 531   0.671440682 -0.097528013  0.738602189
#> 532   0.353353072 -0.894859068  0.087457231
#> 533   0.306998511 -0.691793533  0.155722784
#> 534   0.668154798 -0.120610106  0.464841725
#> 535   0.497274758 -0.647184681 -0.056799484
#> 536   0.202747381 -0.498245679 -0.062767101
#> 537   0.452484913 -0.131219131  0.266440649
#> 538   0.332005038 -0.736576502  0.757274712
#> 539   0.555188316 -0.499876429  0.156198864
#> 540   0.570333616 -0.694894766  0.634558469
#> 541   0.529530987 -0.424205772  0.515162221
#> 542   0.657996384 -0.426908445  0.187970777
#> 543   0.407212953 -0.188295865  0.187582151
#> 544   0.564971066 -0.354737460  0.350134913
#> 545   0.406091630 -0.658634603  0.525648061
#> 546   0.451776825 -0.688813204 -0.162750049
#> 547   0.535726349 -0.437725063  0.590694876
#> 548   0.386308252 -0.640843118 -0.266101348
#> 549   0.361448891 -0.659559950 -0.149409502
#> 550   0.542910453 -0.444003011  0.029091222
#> 551   0.566136851 -0.601046252 -0.070622559
#> 552   0.678904053 -0.264975346  0.024413123
#> 553   0.419970414 -0.387348761  0.018155655
#> 554   0.552998713 -0.307179314  0.308874448
#> 555   0.477270167 -0.281732313  0.276118632
#> 556   0.684588771 -0.570363301 -0.458364322
#> 557   0.747220493 -0.277083036  0.253456943
#> 558   0.629757427 -0.613445028 -0.443548172
#> 559   0.520428675 -0.485722574  0.010430402
#> 560   0.482112866 -0.192698532  0.718076156
#> 561   0.513107019 -0.843619466 -0.516638465
#> 562   0.446415536 -0.424818648  0.007327336
#> 563   0.647841321 -0.495814904  0.243180436
#> 564   0.636588883 -0.367147004  0.332888301
#> 565   0.565374817 -0.350930998 -0.010792926
#> 566   0.699839724 -0.415079056 -0.157507415
#> 567   0.496299682 -0.856923786 -0.112527605
#> 568   0.435822151 -0.196822719  0.555537535
#> 569   0.538245221 -0.211044197  0.270053461
#> 570   0.483804593 -0.419051854 -0.112647699
#> 571   0.603847514 -0.484299635  0.072151406
#> 572   0.697650358 -0.166965585  0.553993305
#> 573   0.634261350 -0.189401737  0.472713220
#> 574   0.643663409  0.058266034  0.720325174
#> 575   0.791198883 -0.319338982 -0.003488426
#> 576   0.616428287 -0.294859622  0.298831140
#> 577   0.600267519  0.048503959  0.399490880
#> 578   0.634261350 -0.189401737  0.472713220
#> 579   0.680971636 -0.090509229  0.670039922
#> 580   0.546402937 -0.250185983  0.486902492
#> 581   0.418364114 -0.612902938  0.177840304
#> 582   0.626706253 -0.472921613 -0.495626763
#> 583   0.418276040 -0.291993376  0.086604012
#> 584   0.497274758 -0.647184681 -0.056799484
#> 585   0.523882473 -0.309711622 -0.017831593
#> 586   0.797821587 -0.307930403 -0.080621676
#> 587   0.574761805 -0.704538366 -0.338203621
#> 588   0.759785248 -0.305436759  0.495872340
#> 589   0.437050777 -0.050710536  0.426617165
#> 590   0.487196498 -0.778519253 -0.122325253
#> 591   0.494841900 -0.684760027  0.136315910
#> 592   0.662504780 -0.599814717 -0.145540849
#> 593   0.577361767 -0.114379879  0.504797772
#> 594   0.463894710 -0.024145334  0.359655140
#> 595   0.399217621 -0.750223547  0.242580645
#> 596   0.715474231 -0.345556767  0.165225251
#> 597   0.616428287 -0.294859622  0.298831140
#> 598   0.824731978  0.005315237  0.417440586
#> 599   0.920700480  0.197291646  1.112597676
#> 600   0.005665662 -0.205603187  0.464832831
#> 601   0.310850685 -0.085146727  0.488891508
#> 602   0.680909282 -0.285679737  0.429883635
#> 603   0.568617645 -0.493922142  0.170307169
#> 604   0.540095134 -0.622503435 -0.103610176
#> 605   0.552998713 -0.307179314  0.308874448
#> 606   0.838441200  0.048988481  0.567287570
#> 607   0.320632022 -0.223360305  1.010907321
#> 608   0.726256023 -0.387765479  0.209754528
#> 609   0.671252199 -0.260952408  0.226754580
#> 610   0.795126267  0.112688785  0.738051185
#> 611   0.736207913 -0.323027188  0.163841116
#> 612   0.662504780 -0.599814717 -0.145540849
#> 613   0.697650358 -0.166965585  0.553993305
#> 614   0.339240643 -0.677864835  0.581458579
#> 615   0.528326180 -0.472422657  0.092099736
#> 616   0.660060577 -0.226780277  0.569781150
#> 617   0.568617645 -0.493922142  0.170307169
#> 618   0.552978971 -0.748670001 -0.444800773
#> 619   0.405612847 -0.332157557  0.122328511
#> 620   0.469367941 -0.616997605  0.005258646
#> 621   0.379367996 -0.202376100  0.269842009
#> 622   0.263123522 -0.926845777 -0.582758158
#> 623   0.501680514  0.017130562  0.628196915
#> 624  -0.116033643 -0.871365514 -0.018899438
#> 625   0.345875673 -0.175987456  0.561146161
#> 626   0.492914163 -0.272890850  0.217704633
#> 627   0.224360183 -0.357541435  0.706148426
#> 628   0.791198883 -0.319338982 -0.003488426
#> 629   0.464972559 -0.645542740  0.326321741
#> 630   0.666073411 -0.554468199 -0.295331521
#> 631   0.371327016 -1.051612733 -0.420001671
#> 632   0.207662734 -0.709747519  0.173024986
#> 633   0.632262505 -0.325892971  0.366164462
#> 634   0.513107019 -0.843619466 -0.516638465
#> 635   0.323910006 -0.098107768  0.546712194
#> 636   0.678737563  0.042346377  0.670978897
#> 637   0.650188746 -0.251956321  0.691063030
#> 638   0.743639298 -0.272477917  1.048098989
#> 639   0.797821587 -0.307930403 -0.080621676
#> 640   0.548982107 -0.182761692  0.241480713
#> 641   0.586888622 -0.573141368  0.442859483
#> 642   0.652285798 -0.314496718  0.508986084
#> 643   0.736207913 -0.323027188  0.163841116
#> 644   0.280031626 -0.367258361  0.093312376
#> 645   0.214239376 -1.041174259 -0.082006661
#> 646   0.616428287 -0.294859622  0.298831140
#> 647   0.755033753  0.099210534  0.643516054
#> 648   0.625354751 -0.074604014  0.089213170
#> 649   0.581446450 -0.382977350  0.279866558
#> 650   0.415122875 -0.806799110 -0.493957324
#> 651   0.487660038 -0.536367780  0.331839534
#> 652   0.497274758 -0.647184681 -0.056799484
#> 653   0.651191244 -0.329776855  0.227197648
#> 654   0.608808159 -0.558237660  0.332305205
#> 655   0.469367941 -0.616997605  0.005258646
#> 656   0.552978971 -0.748670001 -0.444800773
#> 657   0.466158596 -0.373491485  0.010418729
#> 658   0.481801284 -0.605959391  0.115419056
#> 659   0.128140643 -0.696018613 -0.276920787
#> 660   0.774491826 -0.134119475  0.899600072
#> 661   0.538537128 -0.502669518  0.013714627
#> 662   0.560593985 -0.489185325  0.339346336
#> 663   0.288667656 -0.804160613 -0.286618561
#> 664   0.648771684 -0.211731617  0.209384629
#> 665   0.552978971 -0.748670001 -0.444800773
#> 666   0.743655627 -0.296830382  0.214225366
#> 667   0.527692342 -0.186747798  0.595727361
#> 668   0.647841321 -0.495814904  0.243180436
#> 669   0.422827769  0.181143761  1.235589343
#> 670   0.303706926 -0.169715349  0.083144593
#> 671   0.260985074 -0.514846241 -0.091548422
#> 672   0.163540916 -0.369935577  0.600977930
#> 673   0.652285798 -0.314496718  0.508986084
#> 674   0.559958942 -0.447793417 -0.089755247
#> 675   0.500414614 -0.660079776  0.628816137
#> 676   0.361448891 -0.659559950 -0.149409502
#> 677   0.342989311 -1.129623692 -0.542354232
#> 678   0.392657096 -0.775740297 -0.368812365
#> 679   0.315808941 -0.373266238  0.149019266
#> 680   0.504153281 -0.222510122  0.476950783
#> 681   0.568617645 -0.493922142  0.170307169
#> 682   0.471425342 -1.098681981 -0.578815541
#> 683   0.075627613 -0.322508977  0.413711283
#> 684   0.417515562 -0.431096083  0.011867171
#> 685   0.310274261 -0.084657087  0.535053949
#> 686   0.574761805 -0.704538366 -0.338203621
#> 687   0.530862872 -0.467045382 -0.042447710
#> 688   0.671440682 -0.097528013  0.738602189
#> 689   0.464972559 -0.645542740  0.326321741
#> 690   0.665667806 -0.440990508  1.101227928
#> 691   0.566136851 -0.601046252 -0.070622559
#> 692   0.680971636 -0.090509229  0.670039922
#> 693   0.554712891 -0.492602078  0.035517660
#> 694   0.497274758 -0.647184681 -0.056799484
#> 695   0.742652318 -0.317031223  0.090969160
#> 696   0.324871099 -0.084843718  0.810798150
#> 697   0.789942417 -0.458799369  0.249364296
#> 698   0.513107019 -0.843619466 -0.516638465
#> 699   0.305847179 -0.837000263 -0.470460643
#> 700   0.728568004 -0.385075510  0.158547135
#> 701   0.295394902 -0.295871081  0.172625718
#> 702   0.426668794 -0.659787162 -0.151462879
#> 703   0.406928358 -0.363426143  0.482645110
#> 704   0.378320740 -0.855739764 -0.358141140
#> 705   0.791198883 -0.319338982 -0.003488426
#> 706   0.563736866 -0.106306322  0.609053868
#> 707   0.572775973 -0.526138963  0.407734522
#> 708   0.496299682 -0.856923786 -0.112527605
#> 709   0.678737563  0.042346377  0.670978897
#> 710   0.589138990 -0.086038677  0.384991945
#> 711   0.557231242 -0.243845781  0.311296538
#> 712   0.649031015 -0.362371165  0.304897882
#> 713   0.544386367 -0.561198431 -0.054121386
#> 714   0.644287716 -0.052170933  0.800251562
#> 715   0.671252199 -0.260952408  0.226754580
#> 716   0.568617645 -0.493922142  0.170307169
#> 717   0.694274849  0.372969197  0.862450096
#> 718   0.233371626 -0.371577219  0.482528793
#> 719   0.535726349 -0.437725063  0.590694876
#> 720   0.058578337  0.184170538  1.001525251
#> 721   0.512636087 -0.096323609  0.145647377
#> 722   0.170250287 -1.206242029 -0.464626218
#> 723   0.222078435 -0.604932162 -0.047949370
#> 724   0.280957358 -0.282594405  0.378967934
#> 725   0.732616569 -0.094330505  0.562816454
#> 726   0.566263516 -0.321066399  0.716030685
#> 727   0.611515665 -0.193493956  0.249005007
#> 728   0.508189520 -0.326636087  0.034723405
#> 729   0.317340300 -0.476345783  0.230519913
#> 730   0.497958172 -0.280185807 -0.058702181
#> 731   0.279426349 -0.801636099 -0.288515917
#> 732   0.557231242 -0.243845781  0.311296538
#> 733   0.108599460 -0.980854034  0.044820674
#> 734   0.603847514 -0.484299635  0.072151406
#> 735   0.650188746 -0.251956321  0.691063030
#> 736   0.457412546 -0.165323156  0.391783285
#> 737   0.412916565 -0.615252121  0.247783318
#> 738   0.209465704 -0.734829489 -0.074246653
#> 739   0.837654669 -0.239848426  0.039930897
#> 740   0.581446450 -0.382977350  0.279866558
#> 741   0.838245948 -0.109864766  0.518414343
#> 742   0.434661477 -0.783980818 -0.101539245
#> 743   0.469782813 -0.380794301  0.529866058
#> 744   0.396651396 -0.950167245 -0.220993672
#> 745   0.200532967 -0.725258490 -0.080695773
#> 746  -0.400003556 -0.581360266  0.306317082
#> 747   0.544386367 -0.561198431 -0.054121386
#> 748   0.450032772 -0.402978035  0.347839414
#> 749   0.747220493 -0.277083036  0.253456943
#> 750   0.451776825 -0.688813204 -0.162750049
#> 751   0.743655627 -0.296830382  0.214225366
#> 752   0.487660038 -0.536367780  0.331839534
#> 753   0.477270167 -0.281732313  0.276118632
#> 754   0.378320740 -0.855739764 -0.358141140
#> 755   0.518708465 -0.346224398  0.193310755
#> 756   0.412890372 -0.352614554  0.724436514
#> 757   0.564971066 -0.354737460  0.350134913
#> 758   0.824731978  0.005315237  0.417440586
#> 759   0.490466109 -0.362320855  0.720169031
#> 760   0.331055544 -0.473272029  0.405278548
#> 761   0.477483081 -0.621487755  0.021623006
#> 762   0.546402937 -0.250185983  0.486902492
#> 763   0.418739185 -0.681259921  0.104188727
#> 764   0.571197165 -0.285121601  0.287426757
#> 765   0.804460678 -0.146019528  0.196660993
#> 766   0.838245948 -0.109864766  0.518414343
#> 767   0.569189364 -0.470908925  0.048369276
#> 768   0.394245632 -0.123117373  0.139782314
#> 769   0.480706860 -0.809346403 -0.319370723
#> 770   0.625394498 -0.338457175  0.486775557
#> 771   0.555188316 -0.499876429  0.156198864
#> 772   0.487196498 -0.778519253 -0.122325253
#> 773   0.754573550 -0.182502280  0.222213400
#> 774   0.111762758 -0.219077961  0.156221727
#> 775   0.468834335 -0.428925313 -0.106372515
#> 776   0.604524377 -0.413010345 -0.202983714
#> 777   0.737705572 -0.066932008  0.576935289
#> 778   0.316538695  0.045013015  0.805124964
#> 779   0.568021068 -0.686554753  0.194385927
#> 780   0.369067893 -0.550206258  0.211208218
#> 781   0.604451519 -0.221990933 -0.213737579
#> 782   0.560593985 -0.489185325  0.339346336
#> 783   0.163742987 -0.669862273 -0.270042659
#> 784   0.270775372 -0.682402723 -0.059933919
#> 785   0.613406896 -0.368254794  0.412278120
#> 786   0.468339337 -0.662511565  0.318505081
#> 787   0.759785248 -0.305436759  0.495872340
#> 788   0.513107019 -0.843619466 -0.516638465
#> 789   0.667434051  0.063254802  0.599418854
#> 790   0.157001455 -0.704237543  0.297762826
#> 791   0.572786233 -0.508873991 -0.558204678
#> 792   0.262345129 -0.318663373  0.559538139
#> 793   0.501321879 -0.840644885 -0.905702452
#> 794   0.491432146 -0.195413015  0.077040518
#> 795   0.743655627 -0.296830382  0.214225366
#> 796   0.538537128 -0.502669518  0.013714627
#> 797   0.184588539 -0.535319144 -0.043221402
#> 798   0.025922157 -0.326815160  0.023481216
#> 799   0.390621045 -0.322276763  0.350097628
#> 800   0.774491826 -0.134119475  0.899600072
#> 801   0.805472344 -0.301693946 -0.129726161
#> 802   0.487660038 -0.536367780  0.331839534
#> 803   0.518373380 -0.561757120  0.787969601
#> 804   0.538264362 -0.526015532 -0.084690119
#> 805   0.489740951 -0.370226310 -0.022541344
#> 806   0.629951966 -0.129384995  0.578053501
#> 807   0.574296576 -0.395634276  0.735745828
#> 808   0.649031015 -0.362371165  0.304897882
#> 809   0.526011255  0.013021776  0.527215503
#> 810   0.361223789 -0.202753169  0.722503528
#> 811   0.457895280 -0.602367687  0.517539710
#> 812   0.538537128 -0.502669518  0.013714627
#> 813   0.539792421 -0.404856017  0.371884150
#> 814   0.538264362 -0.526015532 -0.084690119
#> 815   0.732616569 -0.094330505  0.562816454
#> 816   0.699713377 -0.410809444 -0.346154997
#> 817   0.316009270 -0.486371707  0.459378337
#> 818   0.449987082 -0.725068271  0.085636813
#> 819   0.932591510  0.248857059  0.659908395
#> 820   0.434459270 -0.420610196  0.622370122
#> 821   0.418320733 -0.479552150  0.387855249
#> 822   0.538537128 -0.502669518  0.013714627
#> 823   0.755033753  0.099210534  0.643516054
#> 824   0.442288312 -0.092689837  0.378662865
#> 825   0.496299682 -0.856923786 -0.112527605
#> 826  -0.150278964  0.166780793  0.547726960
#> 827   0.692784408  0.101940638  0.792039536
#> 828   0.288486592 -1.029424799 -0.701186579
#> 829   0.539792421 -0.404856017  0.371884150
#> 830   0.747220493 -0.277083036  0.253456943
#> 831   0.446933187 -0.476159823 -0.182629390
#> 832  -0.561492914  0.279469254  0.544455209
#> 833   0.487196498 -0.778519253 -0.122325253
#> 834   0.379367996 -0.202376100  0.269842009
#> 835   0.569189364 -0.470908925  0.048369276
#> 836   0.457895280 -0.602367687  0.517539710
#> 837   0.504740508 -0.461227713 -0.037220993
#> 838   0.759785248 -0.305436759  0.495872340
#> 839   0.649031015 -0.362371165  0.304897882
#> 840   0.234760536 -0.473061276 -0.085263884
#> 841   0.473782097  0.141276275  0.625225819
#> 842   0.791328949 -0.190179249  0.502791718
#> 843   0.487744513 -0.796196707 -0.170366113
#> 844   0.077739459 -0.540138871  0.631201236
#> 845   0.059740300 -0.118108746  0.355083365
#> 846   0.612737134 -0.563561449  0.772534921
#> 847   0.739787692 -0.343179102  0.468798315
#> 848   0.616428287 -0.294859622  0.298831140
#> 849   0.651197741 -0.380709294  0.103554970
#> 850   0.069581661 -0.213454652  0.300181443
#> 851   0.581969807 -0.213869439  0.213976708
#> 852   0.446002428 -0.760664784 -0.257580101
#> 853   0.642831539 -0.474504051  0.047721296
#> 854   0.657996384 -0.426908445  0.187970777
#> 855   0.426546040 -0.642041953 -0.063800483
#> 856   0.550196261 -0.467194952  0.268555943
#> 857   0.283207185 -0.132646203  0.922409257
#> 858   0.477549214 -0.236374669  0.340636176
#> 859   0.574074565 -0.154813183  0.277269689
#> 860   0.649031015 -0.362371165  0.304897882
#> 861   0.264904015 -0.403001002  0.288918054
#> 862   0.659198712 -0.562084896  0.018129401
#> 863   0.409566189 -0.749993414 -0.093926551
#> 864   0.419696002 -0.733914647  0.319097213
#> 865   0.593184541 -0.555507690  0.094833097
#> 866   0.418739185 -0.681259921  0.104188727
#> 867   0.490466109 -0.362320855  0.720169031
#> 868   0.116153418 -0.194677901  0.334611924
#> 869   0.692746874 -0.046366092  0.277715337
#> 870   0.316647265 -0.868871806  0.141878922
#> 871   0.788841977 -0.204248156 -0.051788671
#> 872   0.659198712 -0.562084896  0.018129401
#> 873   0.557500477 -0.177158434  0.282154159
#> 874   0.619745681 -0.382856381  0.380307466
#> 875   0.424135946 -0.369061797 -0.333373479
#> 876   0.595524689 -0.706378017 -0.275404321
#> 877   0.200542586 -0.704264537  0.427237274
#> 878   0.668154798 -0.120610106  0.464841725
#> 879   0.755058884  0.169756422  0.944153469
#> 880   0.470248868 -0.485226351  0.573990678
#> 881   0.652285798 -0.314496718  0.508986084
#> 882   0.488564722 -0.313464003 -0.102113270
#> 883   0.737705572 -0.066932008  0.576935289
#> 884   0.429458988  0.224097944  0.502534396
#> 885   0.532345097 -0.292247693  0.068786001
#> 886   0.615233899 -0.059305362  0.423593828
#> 887   0.524850236 -0.218462453  0.501276182
#> 888   0.019882537 -0.351048935  0.627961453
#> 889   0.390081963 -0.664184812  0.441518426
#> 890   0.426546040 -0.642041953 -0.063800483
#> 891   0.754573550 -0.182502280  0.222213400
#> 892   0.563736866 -0.106306322  0.609053868
#> 893   0.655151823 -0.546895439  0.149477584
#> 894   0.388703085 -0.893027588 -0.646910457
#> 895   0.473691781 -0.270451427  0.306308017
#> 896   0.438097083 -0.261881703  0.334800819
#> 897   0.516297270 -0.310819931  0.216656293
#> 898   0.322831326 -0.439519278  0.373083636
#> 899   0.838441200  0.048988481  0.567287570
#> 900   0.392657096 -0.775740297 -0.368812365
#> 901   0.469677238 -0.506649198 -0.096877534
#> 902   0.736207913 -0.323027188  0.163841116
#> 903   0.833085625 -0.008548886  0.532816154
#> 904   0.660060577 -0.226780277  0.569781150
#> 905   0.817268126 -0.204540556 -0.006094052
#> 906   0.649031015 -0.362371165  0.304897882
#> 907   0.550196261 -0.467194952  0.268555943
#> 908   0.664692633 -0.455056125  0.282939128
#> 909   0.619745681 -0.382856381  0.380307466
#> 910   0.503998428 -0.402986435  0.379395660
#> 911   0.503998428 -0.402986435  0.379395660
#> 912   0.521996358 -0.740870094  0.215177714
#> 913   0.367787054 -1.117634225 -0.416910071
#> 914   0.651197741 -0.380709294  0.103554970
#> 915   0.306624884 -0.236752049  0.558060142
#> 916   0.500079919 -0.639802204 -0.453422012
#> 917   0.616428287 -0.294859622  0.298831140
#> 918   0.540095134 -0.622503435 -0.103610176
#> 919   0.595524689 -0.706378017 -0.275404321
#> 920   0.749350156 -0.123255644  0.120080561
#> 921   0.446933187 -0.476159823 -0.182629390
#> 922   0.073347540 -0.347286964  0.430704409
#> 923   0.503998428 -0.402986435  0.379395660
#> 924   0.649031015 -0.362371165  0.304897882
#> 925   0.494492191 -0.199039795  0.802604145
#> 926   0.568617645 -0.493922142  0.170307169
#> 927   0.502601579 -0.369866990  0.206177540
#> 928   0.791328949 -0.190179249  0.502791718
#> 929   0.390818897 -0.689830780 -0.016187121
#> 930   0.737705572 -0.066932008  0.576935289
#> 931   0.691956301  0.198657817  0.693913631
#> 932   0.163170258 -0.764992415  0.276428751
#> 933   0.378320740 -0.855739764 -0.358141140
#> 934   0.477549214 -0.236374669  0.340636176
#> 935   0.535726349 -0.437725063  0.590694876
#> 936   0.496299682 -0.856923786 -0.112527605
#> 937   0.620933237 -0.020848379  0.636024509
#> 938   0.480859670 -0.643638165  0.453588671
#> 939   0.633327451 -0.482225443  0.326358301
#> 940   0.564700102 -0.417930604  0.087843692
#> 941   0.302664357 -0.685124059  0.312179449
#> 942   0.518708465 -0.346224398  0.193310755
#> 943   0.613406896 -0.368254794  0.412278120
#> 944   0.595524689 -0.706378017 -0.275404321
#> 945   0.692373591 -0.464377793  0.304429339
#> 946   0.540095134 -0.622503435 -0.103610176
#> 947   0.636588883 -0.367147004  0.332888301
#> 948   0.647841321 -0.495814904  0.243180436
#> 949   0.667434051  0.063254802  0.599418854
#> 950   0.550196261 -0.467194952  0.268555943
#> 951   0.483804593 -0.419051854 -0.112647699
#> 952   0.419696002 -0.733914647  0.319097213
#> 953   0.505800747 -0.477341753 -0.086041101
#> 954   0.540095134 -0.622503435 -0.103610176
#> 955   0.838245948 -0.109864766  0.518414343
#> 956   0.404206442 -0.424564263  0.248998148
#> 957   0.471870793 -0.233655096  0.157335769
#> 958   0.564700102 -0.417930604  0.087843692
#> 959   0.555188316 -0.499876429  0.156198864
#> 960   0.516297270 -0.310819931  0.216656293
#> 961   0.487744513 -0.796196707 -0.170366113
#> 962   0.616428287 -0.294859622  0.298831140
#> 963   0.804460678 -0.146019528  0.196660993
#> 964   0.690899987 -0.369485983 -0.253575121
#> 965   0.742652318 -0.317031223  0.090969160
#> 966   0.726380584  0.159473909  0.715725490
#> 967   0.787840576 -0.088607004  0.245772921
#> 968   0.530862872 -0.467045382 -0.042447710
#> 969   0.694166704 -0.590330062 -0.002983548
#> 970   0.661081114 -0.194233032  0.768084728
#> 971   0.728568004 -0.385075510  0.158547135
#> 972   0.290597257 -0.658212288 -0.022543843
#> 973   0.255795310 -0.058106774  0.319851791
#> 974   0.572775973 -0.526138963  0.407734522
#> 975   0.671440682 -0.097528013  0.738602189
#> 976   0.568617645 -0.493922142  0.170307169
#> 977   0.542910453 -0.444003011  0.029091222
#> 978   0.342989311 -1.129623692 -0.542354232
#> 979   0.466900800 -0.248138638  0.644988533
#> 980   0.396411746 -0.482937388  0.608849000
#> 981   0.620933237 -0.020848379  0.636024509
#> 982   0.566136851 -0.601046252 -0.070622559
#> 983   0.436568938 -0.273773635 -0.173032059
#> 984   0.371327016 -1.051612733 -0.420001671
#> 985   0.371327016 -1.051612733 -0.420001671
#> 986   0.372278108 -0.713676296  0.098307465
#> 987   0.354130714 -0.598107154  0.295103646
#> 988   0.581446450 -0.382977350  0.279866558
#> 989   0.451776825 -0.688813204 -0.162750049
#> 990   0.686497386 -0.148311019  0.344182869
#> 991   0.522222047  0.019227849  0.273328924
#> 992   0.504740508 -0.461227713 -0.037220993
#> 993   0.418364114 -0.612902938  0.177840304
#> 994   0.737705572 -0.066932008  0.576935289
#> 995  -0.070839264  0.131886524  0.814225261
#> 996   0.504740508 -0.461227713 -0.037220993
#> 997   0.518708465 -0.346224398  0.193310755
#> 998   0.460920142 -0.705479171  0.309014455
#> 999   0.368349383 -0.611589905  0.540648886
#> 1000  0.482112866 -0.192698532  0.718076156
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1  0.359 -0.618  0.360 
#>  2      2        4       70        1  0.406 -0.332  0.122 
#>  3      3        4       70        1  0.773 -0.411 -0.120 
#>  4      4        4       70        1  0.555 -0.493  0.0355
#>  5      5        4       70        1  0.351 -0.635 -0.335 
#>  6      6        4       70        1  0.445 -0.461  0.566 
#>  7      7        4       70        1  0.568 -0.687  0.194 
#>  8      8        4       70        1  0.274 -0.590  0.168 
#>  9      9        4       70        1  0.550 -0.467  0.269 
#> 10     10        4       70        1  0.697 -0.380 -0.289 
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  5.73  37.7 1.43  0.152 32.3      6.97e+2 1217.
#> 2      1    14     4    70     1  5.73  37.7 1.43  0.152  7.35     5.63e-6  277.
#> 3      2     1     4    70     1  6.00  50.2 1.13  0.120 21.4      8.68e+2 1074.
#> 4      2    14     4    70     1  6.00  50.2 1.13  0.120  8.61     3.62e-4  433.
#> 5      3     1     4    70     1  8.66  46.4 0.887 0.187 19.2      1.04e+3  890.
#> 6      3    14     4    70     1  8.66  46.4 0.887 0.187  4.20     1.01e-2  195.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>