Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object
A posologyr list, created by the posologyr function.

n_sample
Number of samples from the S-step

n_resample
Number of samples from the R-step

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)
#> 
#> $eta
#>             ETA_ke         ETA_V
#> 1    -0.1315531190  0.0692130684
#> 2    -0.1868752240  0.1088307798
#> 3    -0.0566388148  0.0126950411
#> 4    -0.0362144212 -0.0262765107
#> 5    -0.0849465230  0.0827757278
#> 6    -0.0948014889  0.0267272516
#> 7    -0.1498054804  0.0255469917
#> 8    -0.0907903891  0.0722442756
#> 9    -0.1135423459  0.0312670064
#> 10   -0.0871911963  0.0401257032
#> 11   -0.0189316275 -0.0414468786
#> 12   -0.0453284078  0.1014076894
#> 13   -0.0380312290  0.1104543543
#> 14   -0.1660630720  0.1383677606
#> 15   -0.0889799122  0.1152873703
#> 16   -0.2006780419  0.1421736796
#> 17   -0.0335198167  0.0134508695
#> 18   -0.0832277174  0.0964927760
#> 19   -0.0494509040 -0.1395361112
#> 20   -0.0904695951  0.0261855090
#> 21   -0.1365981578  0.1017941788
#> 22   -0.1006121879  0.1188250077
#> 23   -0.1514654601  0.1738508628
#> 24   -0.0551489844  0.0257702540
#> 25    0.0288366579  0.0028355094
#> 26    0.0116322606  0.0039094704
#> 27   -0.0611179686  0.0903729812
#> 28   -0.1983984785 -0.0321086847
#> 29   -0.1841231834  0.1105554919
#> 30   -0.1234222600  0.1313684684
#> 31   -0.1226441409  0.0963186288
#> 32   -0.1752311048  0.1419074585
#> 33   -0.0316622512 -0.1132445479
#> 34   -0.1366872841  0.1958553177
#> 35   -0.1140548772  0.1135123119
#> 36   -0.1058216179  0.0274301406
#> 37   -0.3039212451  0.2065712088
#> 38   -0.0373360441  0.0250690039
#> 39   -0.1695179847  0.0546303867
#> 40   -0.1374125229  0.0954183086
#> 41   -0.0571385324  0.0033484640
#> 42   -0.1215783767  0.0974905139
#> 43   -0.0820232153  0.0733732491
#> 44   -0.1436893784  0.1466844324
#> 45   -0.0274122775  0.0092515616
#> 46   -0.0626950776 -0.0343080800
#> 47    0.0039447900 -0.0478582823
#> 48   -0.1133862373  0.1465913025
#> 49   -0.0561631680  0.0745145252
#> 50   -0.1755654827  0.0298515676
#> 51   -0.0578599787  0.0188387370
#> 52   -0.2030897576  0.0927645218
#> 53   -0.0026925967  0.0782953606
#> 54   -0.1227956651 -0.0270780398
#> 55   -0.1161954834  0.0176311617
#> 56   -0.0671857757  0.1111894774
#> 57   -0.0815153745  0.0794713849
#> 58   -0.0657515869  0.0736194225
#> 59   -0.1234222600  0.1313684684
#> 60   -0.1008499771 -0.0856492280
#> 61   -0.0459272663 -0.0508366408
#> 62   -0.0722744137  0.0449893989
#> 63   -0.1240732952  0.0668130903
#> 64   -0.0173884570  0.0379755604
#> 65   -0.2289556454  0.0694968988
#> 66   -0.1072377102  0.1521160836
#> 67   -0.1274482076  0.1910209894
#> 68   -0.0168964753  0.0466071486
#> 69   -0.1074979045  0.1074805476
#> 70   -0.1260578747  0.1541019492
#> 71   -0.0148975668 -0.0659958581
#> 72   -0.1032834689  0.0335498703
#> 73   -0.1400061456  0.0973118840
#> 74   -0.0794666582  0.0892254246
#> 75   -0.1530553737  0.1410048393
#> 76   -0.1008000708  0.0735385553
#> 77   -0.0838148405 -0.0016714345
#> 78   -0.1270985294  0.0915449049
#> 79   -0.1728611128  0.0859198800
#> 80   -0.0753919716  0.1425932981
#> 81   -0.0595831872 -0.0020984581
#> 82   -0.1486400732  0.1095859368
#> 83   -0.1399664520  0.1512500056
#> 84   -0.1719884718  0.0829470786
#> 85   -0.0798510344 -0.0714996856
#> 86   -0.2082585229 -0.0136667689
#> 87   -0.2227405300  0.2706176443
#> 88   -0.0600998300  0.0013991655
#> 89   -0.1805990074  0.0512693731
#> 90   -0.0889799122  0.1152873703
#> 91   -0.0502349008  0.1173460386
#> 92   -0.1203503057  0.1715626906
#> 93   -0.0209615007  0.0439099675
#> 94   -0.0108799258  0.0275677128
#> 95   -0.0899936177  0.0399798319
#> 96   -0.0988256292  0.0673152870
#> 97   -0.1022874907  0.0164076628
#> 98   -0.0713294190 -0.0619834803
#> 99   -0.0587871865  0.0120801631
#> 100  -0.1289946738  0.0372131441
#> 101  -0.2291593033  0.1899048918
#> 102  -0.1069092793  0.1190478837
#> 103  -0.1152572112  0.1671165195
#> 104  -0.1465362875  0.1087972560
#> 105  -0.1610061279  0.0923610849
#> 106   0.0007068138  0.0194223833
#> 107  -0.0454487720  0.0555881185
#> 108  -0.0809633692  0.1056004981
#> 109  -0.0932816367  0.0003017728
#> 110  -0.0749133308  0.1693332601
#> 111  -0.0412810424  0.0004674091
#> 112  -0.0224567617 -0.0285927270
#> 113  -0.0814831053  0.0850374168
#> 114  -0.1270985294  0.0915449049
#> 115  -0.0417003882  0.0662881126
#> 116  -0.2330300602 -0.0099891584
#> 117  -0.0399507294  0.0560121791
#> 118   0.0224140286 -0.0036936533
#> 119  -0.1550447583  0.2367887493
#> 120  -0.1481294369  0.1699581799
#> 121  -0.0791283315 -0.0305619294
#> 122  -0.0571385324  0.0033484640
#> 123  -0.0904695951  0.0261855090
#> 124  -0.1113225505  0.1436362995
#> 125  -0.1196984147  0.1099680494
#> 126  -0.0866838684 -0.0111493562
#> 127  -0.2288508167  0.0579067393
#> 128  -0.1722774286  0.2562322060
#> 129  -0.1067799745  0.0945458600
#> 130  -0.1239445717  0.0292358374
#> 131  -0.1274482076  0.1910209894
#> 132  -0.1242186706  0.1614200886
#> 133  -0.2512643824  0.1436641387
#> 134  -0.0574365003  0.0326240645
#> 135  -0.2103905269  0.0408510682
#> 136  -0.1185725676  0.1615227865
#> 137  -0.0940046067  0.0335383843
#> 138  -0.0889799122  0.1152873703
#> 139  -0.1020132523  0.0168620100
#> 140   0.0111791925  0.0216559461
#> 141  -0.1050672551 -0.1169060094
#> 142  -0.0786469492 -0.0109866981
#> 143  -0.2851018638  0.2592936254
#> 144  -0.0904695951  0.0261855090
#> 145  -0.1311471421  0.0809113840
#> 146  -0.0919422310  0.0431605933
#> 147  -0.1098358931  0.0909079322
#> 148  -0.0962879035  0.2519326879
#> 149  -0.2180618121  0.0916197989
#> 150  -0.0832277174  0.0964927760
#> 151  -0.0665908314  0.1250043577
#> 152  -0.0505379489  0.0308621670
#> 153  -0.2811946758  0.1166324569
#> 154  -0.0539668818  0.1111734489
#> 155  -0.2047149866 -0.0384734273
#> 156  -0.0534632906 -0.0609077240
#> 157  -0.0562974122  0.0178584490
#> 158  -0.0592622352  0.1113000500
#> 159  -0.0721079511  0.0783544482
#> 160  -0.0968836113  0.0687829262
#> 161  -0.1002679261 -0.0015774417
#> 162  -0.2192351104  0.0412124359
#> 163  -0.0245350248 -0.0930846474
#> 164  -0.1285758004  0.1032452399
#> 165   0.0473360861 -0.0799678367
#> 166  -0.1128359586  0.1206974326
#> 167  -0.0833881534  0.1095392345
#> 168  -0.2288508167  0.0579067393
#> 169  -0.0489028064  0.0156448726
#> 170   0.0162092698 -0.0666894529
#> 171  -0.0649990463  0.1203754745
#> 172  -0.1061159815  0.0200460299
#> 173  -0.0724307584  0.0988336650
#> 174  -0.1158333151  0.0270882684
#> 175  -0.1693551859  0.1198311745
#> 176  -0.1704165334  0.1183373666
#> 177  -0.0245160197 -0.0238528913
#> 178  -0.0748176621  0.0718336914
#> 179  -0.0940046067  0.0335383843
#> 180  -0.1855100988  0.2050268881
#> 181  -0.1457712482  0.0687223794
#> 182  -0.0252221106 -0.0080916701
#> 183  -0.2764726852  0.1447440817
#> 184  -0.0814097262  0.1739803175
#> 185  -0.0515820516  0.1808935389
#> 186  -0.1695179847  0.0546303867
#> 187  -0.1575916665  0.0805544170
#> 188  -0.0866506031  0.0817901171
#> 189  -0.0350855806 -0.0453998355
#> 190  -0.1426600636  0.0732414373
#> 191  -0.1831452389  0.2036167290
#> 192  -0.1399664520  0.1512500056
#> 193  -0.0936575577 -0.0034098821
#> 194  -0.1113225505  0.1436362995
#> 195  -0.2798640126  0.1335046912
#> 196  -0.0371408545  0.0339873299
#> 197  -0.1089484870  0.1841881054
#> 198  -0.2065500710  0.0857204916
#> 199  -0.1210745785  0.1253975458
#> 200   0.0334958707 -0.0469316913
#> 201  -0.0574365003  0.0326240645
#> 202  -0.2046019851  0.0946300996
#> 203  -0.1001106105  0.2651514151
#> 204  -0.0482923489  0.2496868199
#> 205  -0.1901661802  0.1857737961
#> 206  -0.1070464054  0.1427157276
#> 207  -0.1704625423  0.1610499270
#> 208   0.0523451900  0.0184035298
#> 209  -0.0437620582 -0.0055091172
#> 210   0.0384712841  0.0062362721
#> 211  -0.0335198167  0.0134508695
#> 212  -0.1939733454  0.0928203813
#> 213  -0.1181871154  0.1242262591
#> 214  -0.2501719659  0.1759732183
#> 215  -0.1359278797  0.1501657546
#> 216  -0.1513896097  0.1227658618
#> 217  -0.0797723165  0.0001586398
#> 218  -0.1214354503  0.2039815035
#> 219  -0.1575916665  0.0805544170
#> 220  -0.1171983630  0.0662949418
#> 221  -0.0410674926  0.0731194619
#> 222  -0.1435818055  0.0875729589
#> 223  -0.1191086777 -0.0156058110
#> 224  -0.1072377102  0.1521160836
#> 225  -0.0834210214  0.0053236161
#> 226  -0.1559506518  0.0664525802
#> 227  -0.0608942858  0.1417175910
#> 228  -0.0412906521  0.0157430320
#> 229  -0.1038791266  0.0806167993
#> 230  -0.0749752150  0.0977752787
#> 231  -0.0871379482  0.0317591060
#> 232  -0.1752311048  0.1419074585
#> 233  -0.1929988419  0.1458240827
#> 234  -0.0645411660 -0.0755087747
#> 235  -0.1704165334  0.1183373666
#> 236  -0.1514654601  0.1738508628
#> 237  -0.1069425951  0.0052885851
#> 238  -0.1187267634  0.0813068313
#> 239  -0.0628744576  0.0135257289
#> 240  -0.1530701140  0.1228000863
#> 241  -0.0832277174  0.0964927760
#> 242  -0.0046392540  0.0117079073
#> 243  -0.1215783767  0.0974905139
#> 244  -0.1042955504  0.0489196235
#> 245  -0.0883772556  0.0192581449
#> 246  -0.0418439351 -0.0347890517
#> 247   0.0459544893  0.0200565758
#> 248  -0.0417003882  0.0662881126
#> 249  -0.0378752738 -0.0395032775
#> 250  -0.1289946738  0.0372131441
#> 251  -0.1891739896  0.1030572891
#> 252  -0.2089852089  0.2255721965
#> 253  -0.1652677933  0.1855590316
#> 254  -0.0832277174  0.0964927760
#> 255  -0.0173884570  0.0379755604
#> 256  -0.2059778698  0.1163983327
#> 257  -0.1205990353 -0.0113701500
#> 258  -0.1392332959  0.0119732337
#> 259  -0.0463739542  0.0474225546
#> 260  -0.1406811435  0.1080059982
#> 261  -0.0352854992 -0.0351659000
#> 262  -0.1588598875  0.2732897414
#> 263   0.0671281429 -0.0875917238
#> 264  -0.1180755979  0.0370491815
#> 265  -0.0275070058 -0.1117468896
#> 266  -0.1645164181  0.0750029139
#> 267  -0.0849465230  0.0827757278
#> 268  -0.0493195667 -0.0501094617
#> 269  -0.1726342605  0.0553168986
#> 270  -0.1032834689  0.0335498703
#> 271  -0.0834667534 -0.0576210907
#> 272  -0.0432717177 -0.0349674933
#> 273  -0.1831452389  0.2036167290
#> 274  -0.0489028064  0.0156448726
#> 275  -0.1113225505  0.1436362995
#> 276  -0.0238966633  0.0499120244
#> 277  -0.0246692405  0.0099121457
#> 278  -0.0555554584 -0.0937265706
#> 279  -0.1426600636  0.0732414373
#> 280  -0.1070464054  0.1427157276
#> 281  -0.1481294369  0.1699581799
#> 282  -0.1369527379  0.1038298242
#> 283  -0.2555493270  0.0748098893
#> 284  -0.0529974556  0.1088107902
#> 285  -0.0273389268  0.0778381172
#> 286  -0.1552803934  0.1694862788
#> 287  -0.0354334955  0.1397629122
#> 288  -0.0562974122  0.0178584490
#> 289  -0.0760691364  0.0346746907
#> 290  -0.0769358972  0.0272601223
#> 291  -0.1872438051  0.1345015283
#> 292  -0.0619627884 -0.0152263582
#> 293  -0.0671857757  0.1111894774
#> 294  -0.1218826393 -0.0841916133
#> 295  -0.0290526654 -0.0277983802
#> 296  -0.0944586739  0.1021581504
#> 297  -0.1100328086  0.1627102642
#> 298  -0.1629240511  0.0374281734
#> 299  -0.0848451386 -0.0082327034
#> 300  -0.0665951846  0.0380157427
#> 301  -0.1627506290  0.1360335429
#> 302  -0.3122233076  0.0439600239
#> 303  -0.0866142254  0.1687369862
#> 304  -0.1111288867  0.0558966338
#> 305  -0.0987110874  0.0877565632
#> 306  -0.0827832251  0.0322228309
#> 307  -0.1034873979  0.0469944027
#> 308  -0.1226441409  0.0963186288
#> 309  -0.0608354109  0.0819213878
#> 310  -0.0883772556  0.0192581449
#> 311  -0.1058976731  0.0705157838
#> 312  -0.1049524436 -0.0246614662
#> 313  -0.0518050527 -0.1454201804
#> 314  -0.1277586944  0.2144880670
#> 315  -0.1303868426  0.1019928088
#> 316  -0.0680474614  0.0186945518
#> 317  -0.0515326118  0.0475720136
#> 318  -0.1204183296 -0.0476003289
#> 319  -0.0348231651  0.0027803105
#> 320  -0.1743462215  0.0328845206
#> 321  -0.0192118416 -0.0288394039
#> 322  -0.0702754579  0.0008070362
#> 323  -0.1008082725  0.0804360598
#> 324  -0.2033634205  0.1414361018
#> 325  -0.1984189159  0.2130087481
#> 326  -0.2632089549  0.1670278893
#> 327  -0.0218212579  0.0051513987
#> 328  -0.2103905269  0.0408510682
#> 329  -0.1987555212  0.1397689318
#> 330  -0.0571385324  0.0033484640
#> 331  -0.1270985294  0.0915449049
#> 332  -0.2587370939  0.1355640735
#> 333  -0.0761088124  0.0609726565
#> 334  -0.0796083728  0.0696295749
#> 335  -0.1872438051  0.1345015283
#> 336  -0.0426402146 -0.0162256011
#> 337  -0.0054583771 -0.0270612920
#> 338  -0.0583271814 -0.0695731639
#> 339  -0.0855199163  0.0223326615
#> 340  -0.1721398089  0.1768468425
#> 341  -0.0613957482  0.0993968289
#> 342  -0.1133862373  0.1465913025
#> 343  -0.1902118285  0.1219774412
#> 344  -0.0824915716  0.0037616635
#> 345  -0.0759829843 -0.0435086825
#> 346  -0.1242186706  0.1614200886
#> 347  -0.1679291942  0.0270890334
#> 348  -0.2479510561  0.2436397475
#> 349  -0.1001178585  0.0706867488
#> 350  -0.1684274834  0.1865930922
#> 351  -0.0937956031 -0.0175668569
#> 352   0.0081159844  0.0281235519
#> 353  -0.1550916738  0.1022963167
#> 354  -0.1472682167  0.0343051197
#> 355  -0.0930479600 -0.0487309338
#> 356  -0.1552803934  0.1694862788
#> 357  -0.0724307584  0.0988336650
#> 358  -0.1111288867  0.0558966338
#> 359  -0.0583627507  0.0282200937
#> 360  -0.0657231316  0.0275837296
#> 361  -0.0671857757  0.1111894774
#> 362   0.0046690140 -0.0729650265
#> 363  -0.0546419049  0.1268619588
#> 364  -0.0855796683  0.0289525029
#> 365  -0.0351662117 -0.0243774261
#> 366  -0.1366872841  0.1958553177
#> 367  -0.1477582959 -0.0238808593
#> 368  -0.2153669326  0.0290648036
#> 369  -0.0990797585 -0.0034228263
#> 370  -0.2335690496  0.1475962401
#> 371  -0.1040146031 -0.0485375275
#> 372  -0.1070464054  0.1427157276
#> 373  -0.2501719659  0.1759732183
#> 374  -0.1242219263  0.0184420577
#> 375  -0.1073143322  0.1744416031
#> 376  -0.1855755719  0.0681282018
#> 377  -0.1524009563  0.0953902847
#> 378  -0.1340016562  0.0643077885
#> 379  -0.2418044849  0.1145371451
#> 380  -0.1451944927  0.1019524389
#> 381  -0.1440358260  0.1054436883
#> 382  -0.1461305874  0.1286021164
#> 383  -0.1185725676  0.1615227865
#> 384  -0.0348231651  0.0027803105
#> 385  -0.1193405881  0.0495699898
#> 386  -0.1872438051  0.1345015283
#> 387  -0.1325434214  0.0348050030
#> 388  -0.0732844616 -0.0403467336
#> 389  -0.2486509144  0.1147702610
#> 390  -0.1679291942  0.0270890334
#> 391  -0.0527241184  0.0467061080
#> 392  -0.2218485185  0.0242822191
#> 393  -0.1250884067  0.0753892503
#> 394  -0.0907629628  0.0464937105
#> 395  -0.0286022801  0.0254246182
#> 396  -0.2278592579  0.1540218661
#> 397  -0.2055316641  0.1299716953
#> 398  -0.1901611325  0.1563008866
#> 399  -0.1230795492  0.0690979201
#> 400  -0.1872593792  0.0718659307
#> 401  -0.0582432024  0.0204600279
#> 402  -0.0441289141  0.0602403425
#> 403  -0.1177776070 -0.0397005354
#> 404  -0.1115600858  0.0380730605
#> 405  -0.0683510077  0.0066181315
#> 406  -0.1482401704  0.0128574777
#> 407  -0.0827832251  0.0322228309
#> 408  -0.0917455748  0.0055130500
#> 409  -0.1478480514  0.0892531391
#> 410  -0.0820232153  0.0733732491
#> 411  -0.1583282752  0.0616762250
#> 412  -0.0609393948 -0.0093391225
#> 413  -0.1089475669  0.0660913589
#> 414  -0.1017048307  0.1118037362
#> 415  -0.0946516437  0.1220815956
#> 416  -0.0224787997 -0.1598373108
#> 417  -0.0251721290  0.0108593022
#> 418  -0.1481294369  0.1699581799
#> 419  -0.1246982518  0.0632597755
#> 420  -0.2227405300  0.2706176443
#> 421  -0.0246692405  0.0099121457
#> 422  -0.0871379482  0.0317591060
#> 423  -0.0911342027  0.0882158358
#> 424  -0.0739844941 -0.0815930073
#> 425  -0.0866821121  0.0069657811
#> 426  -0.0624437836 -0.0687978634
#> 427  -0.0972847949  0.0035277603
#> 428  -0.0994471089  0.2137630338
#> 429  -0.1017048307  0.1118037362
#> 430  -0.0644491939  0.0671267431
#> 431  -0.1500815500  0.1472197020
#> 432  -0.0647676825  0.1274898681
#> 433  -0.2526727076  0.1099482699
#> 434  -0.0522278430  0.0012722737
#> 435  -0.1413817792  0.0447723743
#> 436  -0.0875788980 -0.0029597148
#> 437  -0.1295999424  0.1600586184
#> 438  -0.0988256292  0.0673152870
#> 439  -0.0428358671 -0.0204712243
#> 440  -0.1088145169 -0.0351093881
#> 441  -0.1072377102  0.1521160836
#> 442  -0.1011305041  0.1199395946
#> 443  -0.1343384189  0.0464739623
#> 444  -0.1252247749  0.1494051018
#> 445  -0.0138911397 -0.0417726541
#> 446  -0.1239315701  0.0153655607
#> 447  -0.1459300344  0.0045080294
#> 448  -0.1275336848  0.2046950966
#> 449  -0.1375841192  0.1676441299
#> 450  -0.0889423009  0.1321541934
#> 451  -0.1311947724  0.0768690065
#> 452  -0.0418041363  0.0296260301
#> 453  -0.0687643794 -0.0173474234
#> 454  -0.1246982518  0.0632597755
#> 455  -0.1655812949 -0.0132759470
#> 456  -0.1170479401 -0.1277409127
#> 457  -0.0555694126  0.0859331694
#> 458  -0.0003827257  0.0584114557
#> 459  -0.0660442797  0.0575188262
#> 460  -0.0940046067  0.0335383843
#> 461  -0.1006121879  0.1188250077
#> 462  -0.1017048307  0.1118037362
#> 463  -0.1440358260  0.1054436883
#> 464  -0.2486020672  0.0328841852
#> 465  -0.0219250753  0.1004353715
#> 466  -0.1556413699  0.2174508285
#> 467  -0.0875836493  0.1699467527
#> 468  -0.1568432741  0.0297105966
#> 469  -0.1521695446 -0.0254469177
#> 470  -0.1185123001  0.1067771622
#> 471  -0.0592622352  0.1113000500
#> 472  -0.1286628043 -0.0526873642
#> 473  -0.0553380704  0.0551347928
#> 474  -0.1526069336  0.0928271309
#> 475  -0.0889799122  0.1152873703
#> 476  -0.1001178585  0.0706867488
#> 477  -0.1058216179  0.0274301406
#> 478  -0.0651098627  0.0632558902
#> 479  -0.1869152610  0.0995502324
#> 480  -0.0824415371 -0.0356271030
#> 481  -0.1627640596  0.1661700968
#> 482  -0.1177519240  0.0024494600
#> 483  -0.2565568739  0.1471077871
#> 484   0.0193560821  0.0255649028
#> 485  -0.0750045252 -0.0471098965
#> 486  -0.1014314394  0.0395588996
#> 487  -0.1957730752  0.1436942512
#> 488  -0.2850348159  0.1706888695
#> 489  -0.0922165129  0.2250223492
#> 490  -0.2024519562  0.0071194149
#> 491   0.0184180166  0.0620375824
#> 492  -0.1738283432  0.0810392552
#> 493  -0.1584696366  0.2156101724
#> 494  -0.1855100988  0.2050268881
#> 495  -0.1341563498 -0.0251040274
#> 496  -0.0983510025  0.0384383555
#> 497  -0.0781627295  0.0577941945
#> 498  -0.0706175155  0.0130247002
#> 499  -0.0787690916  0.1962728560
#> 500  -0.2256332363  0.1072995235
#> 501  -0.0547980712  0.1048030057
#> 502  -0.1311947724  0.0768690065
#> 503   0.0216018854  0.0093297365
#> 504  -0.0468654843 -0.0024317154
#> 505  -0.0209615007  0.0439099675
#> 506  -0.0335024061  0.0937246745
#> 507  -0.0938359687  0.0766663931
#> 508  -0.0858704852 -0.0132140388
#> 509  -0.1385801921 -0.0280993431
#> 510  -0.1469417772  0.0553099411
#> 511  -0.1186372523 -0.0151385190
#> 512  -0.2501719659  0.1759732183
#> 513  -0.0437620582 -0.0055091172
#> 514  -0.1441819261 -0.0094103152
#> 515  -0.1160523561 -0.0088190479
#> 516  -0.1293143272  0.1659873720
#> 517  -0.0251859276  0.0934708115
#> 518  -0.0630926707  0.0821274694
#> 519  -0.0619627884 -0.0152263582
#> 520  -0.0593638587 -0.0280477229
#> 521  -0.1870631431  0.0207425145
#> 522  -0.1494728509  0.0457917474
#> 523  -0.2843661566  0.1940543891
#> 524  -0.1107727974  0.0742193521
#> 525  -0.0275070058 -0.1117468896
#> 526  -0.0907903891  0.0722442756
#> 527  -0.1036681717 -0.0177513632
#> 528  -0.0470852190  0.0955759797
#> 529  -0.1227066830 -0.0238215404
#> 530  -0.1077914254  0.0704938794
#> 531  -0.1556413699  0.2174508285
#> 532  -0.2077123005 -0.0024718873
#> 533  -0.0361726729 -0.0168602291
#> 534  -0.0548398839  0.1076011745
#> 535  -0.1250884067  0.0753892503
#> 536  -0.1987555212  0.1397689318
#> 537  -0.1541723184  0.0905672955
#> 538  -0.1563174307  0.1127161295
#> 539  -0.0279243771  0.0544909933
#> 540  -0.2045192296  0.0767807300
#> 541  -0.0686484816  0.0318516421
#> 542  -0.0757502635 -0.0623088754
#> 543  -0.1457712482  0.0687223794
#> 544  -0.0204737489  0.0022251511
#> 545   0.0118178855  0.0076921895
#> 546  -0.1475763357  0.1902699007
#> 547  -0.0832277174  0.0964927760
#> 548  -0.0326494366 -0.0499756728
#> 549  -0.1684337719  0.0724152384
#> 550  -0.2055316641  0.1299716953
#> 551  -0.1693852500  0.1656943415
#> 552  -0.0923295059 -0.0205777554
#> 553  -0.0644491939  0.0671267431
#> 554  -0.0907629628  0.0464937105
#> 555  -0.1858022910  0.0726518422
#> 556  -0.2172755832  0.0932543391
#> 557  -0.1426600636  0.0732414373
#> 558  -0.0766824889  0.1612701770
#> 559  -0.0995619096  0.0969633453
#> 560  -0.1758938920 -0.0060189335
#> 561  -0.1467391823  0.2937478581
#> 562  -0.1115600858  0.0380730605
#> 563  -0.1256269616  0.0193778379
#> 564  -0.1328337256  0.0612222222
#> 565  -0.1250884067  0.0753892503
#> 566  -0.1190415446 -0.0219060911
#> 567  -0.0907069451 -0.1019051638
#> 568  -0.1870631431  0.0207425145
#> 569  -0.1412001461 -0.0058507102
#> 570  -0.1550916738  0.1022963167
#> 571  -0.0391346120  0.0618004035
#> 572  -0.1384565635  0.1980699262
#> 573  -0.0861719313  0.0360803550
#> 574  -0.0325434699  0.0357584629
#> 575  -0.1040099689  0.1038795786
#> 576  -0.0919422310  0.0431605933
#> 577   0.0735016276 -0.1201420633
#> 578  -0.2081961359  0.1871495451
#> 579  -0.1309678743 -0.0188033387
#> 580  -0.1145893533  0.1068480555
#> 581  -0.0148845771 -0.0476316454
#> 582  -0.3101560364  0.1457381222
#> 583  -0.1112728253  0.1418815408
#> 584  -0.1311947724  0.0768690065
#> 585  -0.0861719313  0.0360803550
#> 586  -0.0372923642  0.0117634917
#> 587  -0.0273389268  0.0778381172
#> 588  -0.1530553737  0.1410048393
#> 589  -0.0749752150  0.0977752787
#> 590  -0.0837109186  0.1196293853
#> 591  -0.1754492491  0.1075540703
#> 592  -0.1688894286  0.1466197182
#> 593   0.0204969662  0.0352007670
#> 594  -0.1557135754  0.0640746192
#> 595  -0.1462876585  0.0747729450
#> 596  -0.0530932218  0.0476053333
#> 597  -0.0996964384  0.1558104617
#> 598  -0.1263200564  0.0444532573
#> 599  -0.0999853483  0.1379241538
#> 600  -0.1201723509  0.1070836727
#> 601  -0.0429945591  0.1613286178
#> 602  -0.2056964761  0.0168377673
#> 603  -0.1500815500  0.1472197020
#> 604  -0.1426600636  0.0732414373
#> 605  -0.0611894195  0.0705784788
#> 606  -0.1688894286  0.1466197182
#> 607  -0.1653534455 -0.0213678238
#> 608  -0.0853058583  0.0405169890
#> 609  -0.0762068187 -0.0228890145
#> 610  -0.2300598101  0.2057392483
#> 611  -0.1017093855  0.1008201095
#> 612  -0.0959920249  0.1189928728
#> 613  -0.0074701693  0.0455340136
#> 614  -0.0962848632  0.0453421147
#> 615  -0.1900143327  0.1342497063
#> 616  -0.2055316641  0.1299716953
#> 617  -0.0327207237 -0.1373649712
#> 618  -0.0173884570  0.0379755604
#> 619  -0.1135423459  0.0312670064
#> 620  -0.1660630720  0.1383677606
#> 621  -0.1570968041  0.0334810698
#> 622  -0.2081961359  0.1871495451
#> 623  -0.1295826545  0.0043068252
#> 624  -0.0659154462  0.0433350067
#> 625  -0.1500265701  0.0459800685
#> 626  -0.1237828367 -0.0133361820
#> 627  -0.1269540494  0.1389540454
#> 628  -0.0908729363 -0.0369121807
#> 629  -0.1616411129  0.0991069206
#> 630  -0.0275119296  0.0519375351
#> 631  -0.1554352958  0.1098316303
#> 632  -0.0610134821 -0.0153240725
#> 633  -0.0920721914 -0.0216421988
#> 634  -0.0478698575  0.0332699422
#> 635  -0.2526727076  0.1099482699
#> 636  -0.1632998959  0.0177352332
#> 637  -0.2420651920 -0.0364245195
#> 638   0.0170634986  0.0841354959
#> 639  -0.1341323402  0.1293838902
#> 640  -0.0416406125  0.1256741129
#> 641  -0.0713202371  0.0798131374
#> 642  -0.1214354503  0.2039815035
#> 643  -0.1242605853  0.1164847896
#> 644  -0.1762428604  0.0963800959
#> 645  -0.0359985221 -0.0984299630
#> 646  -0.0632610337  0.0910919447
#> 647  -0.1807869756 -0.0183075853
#> 648  -0.0983510025  0.0384383555
#> 649  -0.1815320522  0.1466600963
#> 650  -0.0875555463  0.2031563476
#> 651  -0.0854372506  0.0135622777
#> 652  -0.0976144027  0.0083576689
#> 653  -0.0453114275 -0.0888623971
#> 654  -0.0683558060  0.0852047701
#> 655  -0.1725073528  0.0518941460
#> 656  -0.1746812863  0.0611969920
#> 657  -0.0959920249  0.1189928728
#> 658  -0.0422532890  0.0636700539
#> 659  -0.0702065379  0.1155477950
#> 660  -0.1742530124  0.2062951387
#> 661  -0.1152572112  0.1671165195
#> 662  -0.0815460268  0.0350757179
#> 663  -0.1089484870  0.1841881054
#> 664  -0.1755279898  0.0992433093
#> 665  -0.0917866641 -0.0530897576
#> 666  -0.0553640250 -0.0040515523
#> 667  -0.1446562936 -0.0049719922
#> 668  -0.0973545320  0.1058704322
#> 669  -0.1502548198  0.0684515994
#> 670  -0.1072377102  0.1521160836
#> 671  -0.1083905201  0.0352758042
#> 672  -0.1369527379  0.1038298242
#> 673  -0.0569796010  0.0193024634
#> 674  -0.1884527047  0.1555685050
#> 675  -0.1468011776 -0.0397217537
#> 676  -0.0264768706 -0.0891491318
#> 677  -0.1094764564  0.1618721324
#> 678  -0.0459272663 -0.0508366408
#> 679  -0.1405815294  0.0304877876
#> 680  -0.1797195758  0.0362627648
#> 681  -0.0394827568  0.0442941803
#> 682  -0.0574365003  0.0326240645
#> 683  -0.1145893533  0.1068480555
#> 684  -0.1074979045  0.1074805476
#> 685   0.0273333889 -0.0523725440
#> 686  -0.1783177926  0.1463241502
#> 687  -0.0610199927 -0.0213418350
#> 688  -0.1035513685  0.0745569054
#> 689  -0.1359744658  0.1960411079
#> 690  -0.0411178723 -0.0181158607
#> 691  -0.0901908859  0.0263471748
#> 692  -0.0626950776 -0.0343080800
#> 693  -0.1496971595  0.0902889629
#> 694  -0.0657515869  0.0736194225
#> 695  -0.0076502781 -0.0117835278
#> 696  -0.1426600636  0.0732414373
#> 697  -0.0006858215 -0.0137834439
#> 698  -0.1623920061  0.0998955409
#> 699  -0.1240732952  0.0668130903
#> 700  -0.0716928678 -0.1132404449
#> 701  -0.1210745785  0.1253975458
#> 702  -0.1359744658  0.1960411079
#> 703  -0.3412874651  0.1309854102
#> 704  -0.0502098257  0.0651938298
#> 705  -0.0855559528  0.0436203707
#> 706  -0.0785464216  0.0282576483
#> 707   0.0113069091 -0.0400163393
#> 708  -0.0903419515 -0.0209921786
#> 709  -0.0519543488 -0.0756675970
#> 710  -0.0979422262  0.0557743457
#> 711  -0.1451944927  0.1019524389
#> 712  -0.1187267634  0.0813068313
#> 713  -0.1688894286  0.1466197182
#> 714  -0.1465362875  0.1087972560
#> 715  -0.0604467729  0.1048436029
#> 716  -0.1242605853  0.1164847896
#> 717  -0.0885487683  0.0949680455
#> 718  -0.0794666582  0.0892254246
#> 719  -0.1777570816  0.1186466593
#> 720  -0.0894200690  0.0489300260
#> 721  -0.0251721290  0.0108593022
#> 722  -0.2568268212  0.1361622145
#> 723   0.0278371549 -0.1124193970
#> 724  -0.1638902356  0.0550275400
#> 725  -0.1719884718  0.0829470786
#> 726  -0.2889662942  0.0889723509
#> 727  -0.1120395911  0.1036867975
#> 728  -0.1471800888  0.1372885573
#> 729  -0.0009919880 -0.0907384559
#> 730  -0.0355829703  0.0597606002
#> 731  -0.2180939142  0.1077789146
#> 732  -0.0855559528  0.0436203707
#> 733   0.0452840815 -0.1341606618
#> 734  -0.0922477254 -0.0008749234
#> 735  -0.0997972657  0.0589711172
#> 736  -0.1526069336  0.0928271309
#> 737  -0.0449738170  0.1496076874
#> 738  -0.0883772556  0.0192581449
#> 739  -0.1951502377  0.2342907499
#> 740  -0.2084500535  0.0862404943
#> 741  -0.1108022047  0.0864375373
#> 742  -0.1226441409  0.0963186288
#> 743  -0.0814097262  0.1739803175
#> 744  -0.1551023875 -0.0352978155
#> 745  -0.0561631680  0.0745145252
#> 746  -0.1185123001  0.1067771622
#> 747  -0.0747982486  0.0560197202
#> 748  -0.1727205956  0.0312499110
#> 749  -0.0711253846  0.1614305841
#> 750   0.0008346948 -0.0481148200
#> 751  -0.2225799635  0.0906051765
#> 752  -0.1442860185  0.0891245768
#> 753  -0.0908729363 -0.0369121807
#> 754  -0.1001178585  0.0706867488
#> 755  -0.2627542187  0.2176755993
#> 756  -0.1054336843 -0.0190541561
#> 757  -0.0995253021  0.0480003635
#> 758  -0.2020822540 -0.0331712078
#> 759  -0.1737702386  0.0953037176
#> 760  -0.0974655435 -0.0421142263
#> 761  -0.1742530124  0.2062951387
#> 762  -0.0968836113  0.0687829262
#> 763  -0.0632632489 -0.1129486823
#> 764  -0.0609393948 -0.0093391225
#> 765  -0.1138735083  0.0134624498
#> 766  -0.1743462215  0.0328845206
#> 767  -0.1436893784  0.1466844324
#> 768  -0.1375841192  0.1676441299
#> 769  -0.0718577623 -0.0030983227
#> 770  -0.1722774286  0.2562322060
#> 771  -0.0616649663 -0.0646287276
#> 772  -0.0111833782  0.0276625953
#> 773  -0.1385373993  0.0352335582
#> 774  -0.0534499988 -0.0179501827
#> 775  -0.0485479177  0.1485936882
#> 776  -0.1541723184  0.0905672955
#> 777  -0.0618529893  0.1466797641
#> 778  -0.0593580052  0.0467099832
#> 779  -0.0968836113  0.0687829262
#> 780  -0.2248551477  0.0708904505
#> 781  -0.1619375643  0.0118327122
#> 782  -0.0168964753  0.0466071486
#> 783  -0.0829064429  0.0147337510
#> 784  -0.0349465136  0.0690215406
#> 785  -0.0553119959 -0.0310562548
#> 786  -0.1325939976  0.0450947125
#> 787  -0.1467391823  0.2937478581
#> 788  -0.0469343378  0.1020158516
#> 789  -0.0960426105  0.0178683281
#> 790  -0.0377468866 -0.0284679791
#> 791  -0.0099437552  0.0633343922
#> 792  -0.0859617135  0.1052071574
#> 793  -0.0665411474  0.1037307846
#> 794  -0.1448558309  0.0074534834
#> 795   0.0166167489 -0.0498020219
#> 796  -0.1362675268  0.1324458107
#> 797  -0.0515326118  0.0475720136
#> 798  -0.1697973932  0.0500363967
#> 799  -0.0634768108 -0.0651214241
#> 800  -0.0704016594 -0.0310976998
#> 801  -0.0706870133  0.2212719176
#> 802  -0.0168160928  0.1162438961
#> 803  -0.0919422310  0.0431605933
#> 804  -0.0173078673  0.1172721110
#> 805  -0.1694114002  0.0923866203
#> 806  -0.0871911963  0.0401257032
#> 807  -0.1226441409  0.0963186288
#> 808  -0.0912872883  0.0354779730
#> 809  -0.0905326196  0.0692220265
#> 810  -0.1600994102 -0.0983582652
#> 811  -0.1625725992  0.1976829896
#> 812  -0.1984189159  0.2130087481
#> 813  -0.2406234900  0.2274772967
#> 814  -0.0351241996 -0.1275173617
#> 815  -0.1274846951  0.0504727900
#> 816  -0.0387594731  0.0648237093
#> 817  -0.1554749142  0.0429049631
#> 818  -0.1499077949  0.0755741786
#> 819  -0.1253132042  0.1573129179
#> 820  -0.2256356525  0.1682897784
#> 821  -0.0661477140  0.0877455589
#> 822  -0.1957730752  0.1436942512
#> 823  -0.2312447405 -0.0152841246
#> 824  -0.1325939976  0.0450947125
#> 825  -0.0034430026  0.0535545241
#> 826  -0.1362675268  0.1324458107
#> 827  -0.0719763377  0.0260113229
#> 828  -0.1275420918  0.0584422120
#> 829   0.0197374705  0.0043453619
#> 830  -0.0647676825  0.1274898681
#> 831  -0.0847312117 -0.0632046964
#> 832  -0.1777570816  0.1186466593
#> 833  -0.0553657371 -0.0788905241
#> 834  -0.0516595588 -0.0217175048
#> 835  -0.1210745785  0.1253975458
#> 836  -0.0886140693  0.0363764574
#> 837  -0.1325434214  0.0348050030
#> 838   0.0069295145 -0.0600467090
#> 839  -0.1746812863  0.0611969920
#> 840  -0.1556413699  0.2174508285
#> 841  -0.0454487720  0.0555881185
#> 842  -0.0871911963  0.0401257032
#> 843  -0.0937956031 -0.0175668569
#> 844  -0.1089475669  0.0660913589
#> 845  -0.1549119440  0.1751790739
#> 846  -0.2278592579  0.1540218661
#> 847  -0.1392332959  0.0119732337
#> 848  -0.1665634953  0.1672505997
#> 849  -0.1060309549  0.0073295716
#> 850  -0.1901661802  0.1857737961
#> 851  -0.0994471089  0.2137630338
#> 852  -0.1777570816  0.1186466593
#> 853  -0.0069162947 -0.0276016170
#> 854  -0.1694114002  0.0923866203
#> 855  -0.0866821121  0.0069657811
#> 856  -0.2038525577  0.1756274632
#> 857  -0.1486400732  0.1095859368
#> 858  -0.0883772556  0.0192581449
#> 859  -0.2238309088  0.1238504139
#> 860  -0.1608362315 -0.0133249080
#> 861  -0.0561631680  0.0745145252
#> 862  -0.0682815548  0.0456085644
#> 863  -0.0553380704  0.0551347928
#> 864  -0.1186581627  0.1394373938
#> 865  -0.0619627884 -0.0152263582
#> 866  -0.0761461557  0.0420101735
#> 867  -0.1679831622 -0.0083600916
#> 868  -0.1393430817  0.0975752105
#> 869  -0.1475055549  0.1627192890
#> 870  -0.1141606863  0.0148215668
#> 871  -0.0987110874  0.0877565632
#> 872  -0.1679291942  0.0270890334
#> 873  -0.1467391823  0.2937478581
#> 874  -0.0355829703  0.0597606002
#> 875  -0.0502098257  0.0651938298
#> 876  -0.0724307584  0.0988336650
#> 877  -0.1070464054  0.1427157276
#> 878  -0.1089475669  0.0660913589
#> 879  -0.1046143218 -0.0804853664
#> 880  -0.0159775686  0.0454574373
#> 881  -0.1722774286  0.2562322060
#> 882  -0.0972845598  0.0364818381
#> 883  -0.1011305041  0.1199395946
#> 884  -0.1439911408  0.1068205608
#> 885  -0.1530553737  0.1410048393
#> 886  -0.1134880077 -0.0134253465
#> 887  -0.0881491507  0.1604807559
#> 888  -0.0721491779 -0.0238237667
#> 889  -0.1530701140  0.1228000863
#> 890  -0.0425195001  0.0440060822
#> 891  -0.1167370885  0.1163578349
#> 892   0.0015530785  0.0286480922
#> 893  -0.1270985294  0.0915449049
#> 894   0.0268648111  0.0553980055
#> 895  -0.1655812949 -0.0132759470
#> 896  -0.0568550975  0.0869381351
#> 897  -0.0886980377  0.0007501571
#> 898  -0.1187267634  0.0813068313
#> 899  -0.0854372506  0.0135622777
#> 900  -0.1483277784  0.1774924174
#> 901  -0.1267813413  0.1069853734
#> 902  -0.0833881534  0.1095392345
#> 903  -0.0617306544 -0.0467220825
#> 904  -0.1675992433  0.1131885252
#> 905  -0.1119266015  0.0608475295
#> 906  -0.1468011776 -0.0397217537
#> 907  -0.0866506031  0.0817901171
#> 908  -0.1234222600  0.1313684684
#> 909  -0.0681613891  0.0648805863
#> 910  -0.0145059074 -0.1575417282
#> 911  -0.0412745345  0.0214779860
#> 912  -0.0848451386 -0.0082327034
#> 913  -0.0818630907  0.0811033929
#> 914  -0.0886140693  0.0363764574
#> 915  -0.0864225544 -0.0763639192
#> 916  -0.0452359973 -0.0172738020
#> 917  -0.0798243748  0.0912100587
#> 918  -0.1370573877  0.0628734247
#> 919  -0.1135542757  0.0752922015
#> 920  -0.0410674926  0.0731194619
#> 921  -0.2845446825  0.1853037323
#> 922  -0.1131556039  0.0312212670
#> 923  -0.1509187532 -0.0064543348
#> 924  -0.1430795430  0.1215184656
#> 925  -0.0990797585 -0.0034228263
#> 926  -0.1442583763  0.0945190516
#> 927  -0.0979422262  0.0557743457
#> 928  -0.1032641351  0.0121849875
#> 929  -0.1475055549  0.1627192890
#> 930  -0.1528108188  0.0169469616
#> 931  -0.0957723707  0.0453532123
#> 932  -0.0906020945  0.0662025203
#> 933  -0.0192118416 -0.0288394039
#> 934  -0.0604467729  0.1048436029
#> 935  -0.1439981042  0.0713102766
#> 936  -0.1522074133 -0.0155181554
#> 937  -0.1413817792  0.0447723743
#> 938  -0.1675992433  0.1131885252
#> 939  -0.2045192296  0.0767807300
#> 940  -0.1194986975  0.0976476768
#> 941  -0.2951760586  0.2185640974
#> 942  -0.2510904797  0.2665200930
#> 943  -0.0946516437  0.1220815956
#> 944  -0.3372934811  0.0494121230
#> 945  -0.1120395911  0.1036867975
#> 946  -0.1461305874  0.1286021164
#> 947  -0.1970355151  0.1434507869
#> 948  -0.1270985294  0.0915449049
#> 949  -0.0907629628  0.0464937105
#> 950  -0.1530553737  0.1410048393
#> 951  -0.1100328086  0.1627102642
#> 952   0.0007381768 -0.0355860680
#> 953  -0.1325939976  0.0450947125
#> 954   0.1127456101 -0.0557235924
#> 955  -0.1133862373  0.1465913025
#> 956  -0.1178042794  0.0922341852
#> 957  -0.1269540494  0.1389540454
#> 958  -0.1332790948  0.0924846504
#> 959  -0.0081529780  0.0481279367
#> 960  -0.1120030720  0.1620390927
#> 961  -0.0567895163  0.0789083800
#> 962  -0.0493195667 -0.0501094617
#> 963  -0.0547980712  0.1048030057
#> 964  -0.0555694126  0.0859331694
#> 965  -0.0665951846  0.0380157427
#> 966  -0.0661041560  0.0676545965
#> 967  -0.1754492491  0.1075540703
#> 968  -0.1069945477  0.0960909237
#> 969  -0.1116017787  0.0923486771
#> 970  -0.2605889556  0.1331366672
#> 971   0.0436326928 -0.0034695762
#> 972  -0.0831852550  0.1308970529
#> 973  -0.1374125229  0.0954183086
#> 974  -0.0909712025 -0.0542741517
#> 975  -0.0412810424  0.0004674091
#> 976  -0.1404984360  0.1152126435
#> 977  -0.0604467729  0.1048436029
#> 978  -0.1187267634  0.0813068313
#> 979  -0.0589274685 -0.0947045921
#> 980  -0.0466695954  0.0808562910
#> 981  -0.0441289141  0.0602403425
#> 982  -0.1210745785  0.1253975458
#> 983  -0.1061159815  0.0200460299
#> 984  -0.0259938454  0.1576879336
#> 985  -0.1522571262  0.0670319542
#> 986  -0.1015457929  0.0904164595
#> 987  -0.1002679261 -0.0015774417
#> 988  -0.0652183286 -0.0032993372
#> 989  -0.1316442566  0.0517158681
#> 990  -0.2089446492  0.0992568301
#> 991   0.0506774981  0.0066900816
#> 992  -0.0148724940  0.0157350929
#> 993  -0.1469417772  0.0553099411
#> 994  -0.1965745807  0.0254863236
#> 995  -0.1475055549  0.1627192890
#> 996  -0.1011305041  0.1199395946
#> 997  -0.0987048009  0.0643454408
#> 998  -0.0805246257  0.0319182235
#> 999  -0.1210020866  0.0688872351
#> 1000 -0.0082636581 -0.0436913121
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.132   0.0692
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.187   0.109 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.0566  0.0127
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0362 -0.0263
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.0849  0.0828
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0948  0.0267
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.150   0.0255
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0908  0.0722
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.114   0.0313
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.0872  0.0401
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.218  20.9 0.041  0.12 79.1  1650. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.218  20.9 0.041  0.12  5.61  117. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.207  21.7 0.041  0.12 76.7  1665. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.207  21.7 0.041  0.12  6.11  133. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.235  19.7 0.041  0.12 82.6  1629. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.235  19.7 0.041  0.12  4.97   98.1
#> # … with 1,994 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>