Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 1e+05,
  n_resample = 10000,
  return_model = TRUE
)

Arguments

object
A posologyr list, created by the posologyr function.

n_sample
Number of samples from the S-step

n_resample
Number of samples from the R-step

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)

#> 

#> $eta
#>             ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1     4.068828e-02 -0.0952374677       0       0
#> 2    -1.879016e-01  0.1324035331       0       0
#> 3    -1.263841e-01  0.0189892697       0       0
#> 4    -1.268188e-01  0.0301568219       0       0
#> 5    -8.689782e-02  0.0788028246       0       0
#> 6    -2.802471e-01  0.3713649918       0       0
#> 7    -2.881873e-02  0.0049391972       0       0
#> 8    -1.814134e-01  0.0513403470       0       0
#> 9    -2.089062e-02  0.0171080859       0       0
#> 10   -1.130388e-02  0.0089429058       0       0
#> 11   -1.186049e-01  0.1172626953       0       0
#> 12   -1.545197e-01  0.1648945599       0       0
#> 13   -1.365304e-03  0.0311668855       0       0
#> 14    4.495806e-03  0.1311194771       0       0
#> 15   -9.508663e-02  0.0922446669       0       0
#> 16    5.586699e-02 -0.1603480700       0       0
#> 17   -1.084245e-01  0.1102760821       0       0
#> 18   -4.381243e-02 -0.0397039238       0       0
#> 19   -1.060046e-02 -0.0424641437       0       0
#> 20   -6.095630e-02 -0.1635645186       0       0
#> 21   -1.790067e-01  0.1244942710       0       0
#> 22   -1.082324e-01  0.0938814619       0       0
#> 23   -1.840729e-01  0.2893523702       0       0
#> 24   -3.571303e-01  0.2338764714       0       0
#> 25   -1.443353e-01  0.1381424121       0       0
#> 26   -3.779737e-01  0.1185593310       0       0
#> 27   -3.607142e-02 -0.0639497730       0       0
#> 28   -1.417515e-01  0.1857063159       0       0
#> 29   -3.081467e-01  0.1040743961       0       0
#> 30   -2.019616e-01  0.1225261327       0       0
#> 31   -2.570105e-01  0.1206080523       0       0
#> 32   -5.987941e-02 -0.0880089913       0       0
#> 33   -1.216721e-01  0.3193997419       0       0
#> 34   -1.433592e-01 -0.0808133723       0       0
#> 35   -2.443516e-01  0.2594840910       0       0
#> 36   -1.507595e-01  0.0848495623       0       0
#> 37   -1.980147e-02 -0.0503087123       0       0
#> 38   -2.081274e-01  0.0010240621       0       0
#> 39   -5.457910e-02 -0.0070851345       0       0
#> 40   -1.038249e-01  0.1213348247       0       0
#> 41   -8.625527e-02 -0.0015121654       0       0
#> 42    2.610572e-04 -0.0203889137       0       0
#> 43   -8.616907e-02  0.0539620392       0       0
#> 44   -7.346182e-02 -0.0936019550       0       0
#> 45   -3.203354e-01  0.0674185391       0       0
#> 46   -1.715461e-01  0.0049228425       0       0
#> 47   -3.931001e-02  0.1410194398       0       0
#> 48    5.692819e-02 -0.2513433225       0       0
#> 49   -9.665364e-02 -0.0667250516       0       0
#> 50   -8.242559e-02  0.0850033009       0       0
#> 51   -1.894252e-01  0.0830809692       0       0
#> 52   -2.339568e-01  0.0304314699       0       0
#> 53   -1.306238e-01 -0.0002296813       0       0
#> 54   -2.968626e-02 -0.0083062248       0       0
#> 55   -1.084245e-01  0.1102760821       0       0
#> 56    5.445477e-02 -0.0192885094       0       0
#> 57   -1.098124e-01 -0.0135246066       0       0
#> 58   -5.892658e-02 -0.1107769210       0       0
#> 59   -2.272338e-01  0.1927464124       0       0
#> 60   -5.908750e-02 -0.0482377345       0       0
#> 61   -9.467286e-02 -0.0677231389       0       0
#> 62   -2.544299e-01  0.1744399880       0       0
#> 63   -2.289830e-01  0.1165902173       0       0
#> 64   -1.438573e-01  0.1020676849       0       0
#> 65   -2.264892e-01  0.0268864766       0       0
#> 66   -1.624502e-01  0.2513177452       0       0
#> 67   -7.789609e-02  0.0623996782       0       0
#> 68   -8.889932e-02  0.1117230183       0       0
#> 69   -1.510709e-01  0.2971956514       0       0
#> 70   -4.550655e-02  0.1731537499       0       0
#> 71   -1.264988e-01  0.0405044742       0       0
#> 72    4.167356e-02 -0.0152799314       0       0
#> 73   -6.963644e-02  0.0850666934       0       0
#> 74   -1.438573e-01  0.1020676849       0       0
#> 75   -1.640885e-01  0.0682322657       0       0
#> 76   -1.628324e-01  0.1835723315       0       0
#> 77   -7.565316e-02  0.1494885854       0       0
#> 78   -7.084913e-02  0.0194678679       0       0
#> 79   -7.777336e-02  0.0402330663       0       0
#> 80   -1.623191e-01  0.1746943024       0       0
#> 81   -4.879126e-02 -0.0877748572       0       0
#> 82    2.048489e-02 -0.1437962304       0       0
#> 83   -1.539648e-01  0.1395807106       0       0
#> 84   -2.510501e-01 -0.0738909413       0       0
#> 85   -2.838016e-01  0.0960617352       0       0
#> 86   -2.443516e-01  0.2594840910       0       0
#> 87   -5.796374e-02  0.1818688360       0       0
#> 88   -3.864305e-02 -0.1162834114       0       0
#> 89   -1.813295e-02  0.0062960097       0       0
#> 90   -5.416717e-02  0.2136048695       0       0
#> 91   -7.817593e-02 -0.0600618041       0       0
#> 92   -1.814547e-01  0.1058679380       0       0
#> 93   -1.311689e-01 -0.1091110854       0       0
#> 94   -1.984115e-02 -0.1549945873       0       0
#> 95   -2.274168e-01  0.2558291152       0       0
#> 96   -1.736707e-01  0.0904147123       0       0
#> 97    5.323904e-02 -0.0955479486       0       0
#> 98   -2.021222e-01  0.1414977262       0       0
#> 99   -8.462138e-02  0.1279969001       0       0
#> 100  -4.590112e-02 -0.0029071849       0       0
#> 101  -1.018541e-01  0.0836112342       0       0
#> 102  -1.046558e-01  0.1849801428       0       0
#> 103  -3.952839e-02 -0.0630098002       0       0
#> 104  -1.757805e-01  0.1432468786       0       0
#> 105  -1.729013e-01  0.0402005880       0       0
#> 106  -4.550655e-02  0.1731537499       0       0
#> 107  -1.227325e-02 -0.1187651526       0       0
#> 108  -1.516763e-01  0.1249516823       0       0
#> 109  -1.624502e-01  0.2513177452       0       0
#> 110  -1.377557e-01 -0.0496492477       0       0
#> 111   2.610572e-04 -0.0203889137       0       0
#> 112  -1.789478e-02 -0.0359818986       0       0
#> 113  -1.842760e-01  0.1411124225       0       0
#> 114   7.032965e-02  0.0280440853       0       0
#> 115  -3.791621e-02 -0.0810156117       0       0
#> 116  -1.759868e-01  0.1990409176       0       0
#> 117   1.691383e-02 -0.1026434174       0       0
#> 118  -1.749912e-01  0.1227803421       0       0
#> 119  -4.571744e-02 -0.0841483827       0       0
#> 120  -9.477112e-02  0.0275051159       0       0
#> 121  -2.398647e-01  0.2221437401       0       0
#> 122  -1.510501e-02 -0.0580380308       0       0
#> 123  -1.644799e-01  0.1301315001       0       0
#> 124  -1.020916e-01  0.0357771670       0       0
#> 125  -1.206369e-01  0.2851685279       0       0
#> 126  -7.245337e-02 -0.0691449845       0       0
#> 127  -8.616907e-02  0.0539620392       0       0
#> 128  -1.735890e-02  0.0384634870       0       0
#> 129  -2.398647e-01  0.2221437401       0       0
#> 130  -6.937435e-02  0.0160441443       0       0
#> 131  -1.059572e-01  0.0818790330       0       0
#> 132   4.495806e-03  0.1311194771       0       0
#> 133  -2.070258e-01  0.1727539366       0       0
#> 134  -8.927172e-02 -0.0010439125       0       0
#> 135  -8.662593e-02  0.0177037904       0       0
#> 136  -2.929754e-02 -0.0206091833       0       0
#> 137  -2.049575e-01  0.1126319426       0       0
#> 138  -3.000382e-01  0.1544930996       0       0
#> 139  -3.212976e-01  0.0118614022       0       0
#> 140  -1.627940e-01  0.0339601800       0       0
#> 141  -1.618151e-01 -0.0128850496       0       0
#> 142  -1.068240e-01  0.0447850798       0       0
#> 143  -2.604626e-02  0.0923937775       0       0
#> 144  -3.782733e-01  0.3327009653       0       0
#> 145  -1.112416e-01  0.1086287159       0       0
#> 146  -3.669622e-01  0.1741172567       0       0
#> 147  -1.175040e-01 -0.2105142799       0       0
#> 148  -3.817274e-02  0.0032788150       0       0
#> 149  -1.438573e-01  0.1020676849       0       0
#> 150  -1.545197e-01  0.1648945599       0       0
#> 151  -7.429072e-02  0.0626376974       0       0
#> 152  -7.709225e-02 -0.0090524248       0       0
#> 153  -1.873608e-01  0.1041443924       0       0
#> 154  -1.757903e-01 -0.0054441911       0       0
#> 155  -2.400373e-02 -0.0499748803       0       0
#> 156  -2.306082e-01  0.1103074781       0       0
#> 157  -1.967438e-01  0.3056698714       0       0
#> 158  -1.929354e-01 -0.0139634302       0       0
#> 159  -1.181783e-02 -0.0579719313       0       0
#> 160  -2.240155e-02 -0.1175613112       0       0
#> 161  -1.143793e-01 -0.0214380686       0       0
#> 162  -1.777319e-01  0.2049568150       0       0
#> 163  -1.813295e-02  0.0062960097       0       0
#> 164  -1.189840e-01 -0.0164982448       0       0
#> 165  -1.967438e-01  0.3056698714       0       0
#> 166  -4.093052e-02  0.0473789405       0       0
#> 167  -2.687007e-01  0.1043134739       0       0
#> 168  -5.962776e-02 -0.0033454715       0       0
#> 169  -1.216721e-01  0.3193997419       0       0
#> 170  -8.368903e-02  0.1171915773       0       0
#> 171  -1.023288e-01  0.1694650905       0       0
#> 172   1.750600e-02  0.0704853826       0       0
#> 173  -1.790067e-01  0.1244942710       0       0
#> 174  -1.663323e-01  0.1793861412       0       0
#> 175  -8.280144e-02 -0.0385907197       0       0
#> 176  -7.153593e-02  0.1937988357       0       0
#> 177  -5.758087e-03  0.0200892359       0       0
#> 178  -5.299451e-02  0.0180034515       0       0
#> 179  -1.988216e-01  0.1352218878       0       0
#> 180  -6.231183e-02 -0.0161663496       0       0
#> 181  -1.756213e-01  0.1582275221       0       0
#> 182  -1.017616e-01  0.0010343717       0       0
#> 183  -2.909247e-02 -0.1125566500       0       0
#> 184  -6.689157e-02  0.0695064829       0       0
#> 185  -8.852719e-02 -0.2246715611       0       0
#> 186  -4.877371e-02  0.1064699584       0       0
#> 187  -8.631230e-02  0.0516017946       0       0
#> 188  -2.044229e-01 -0.0049578306       0       0
#> 189  -8.608769e-02  0.1779798713       0       0
#> 190  -2.574763e-01  0.2540735859       0       0
#> 191  -1.845611e-01  0.2352257178       0       0
#> 192  -9.812820e-02 -0.0921923518       0       0
#> 193  -2.086019e-02  0.1259578005       0       0
#> 194  -2.424900e-01  0.0203538049       0       0
#> 195  -1.820353e-01  0.1781600174       0       0
#> 196  -1.777319e-01  0.2049568150       0       0
#> 197  -6.832807e-02  0.0633998591       0       0
#> 198  -6.032031e-02 -0.0713252874       0       0
#> 199  -8.476417e-02  0.0949822683       0       0
#> 200  -2.060851e-01  0.0896862534       0       0
#> 201  -1.368534e-01  0.0625056198       0       0
#> 202  -2.720276e-01  0.0597702905       0       0
#> 203  -3.165023e-02 -0.1306859479       0       0
#> 204  -2.133432e-01  0.2485182647       0       0
#> 205  -1.967438e-01  0.3056698714       0       0
#> 206  -1.625921e-01  0.1147894068       0       0
#> 207  -4.877371e-02  0.1064699584       0       0
#> 208  -1.757066e-01  0.1390329253       0       0
#> 209  -2.057926e-01  0.2759767481       0       0
#> 210  -2.802471e-01  0.3713649918       0       0
#> 211  -6.677466e-02  0.1485316400       0       0
#> 212  -1.521791e-01  0.1196101823       0       0
#> 213  -1.316482e-02 -0.0405965626       0       0
#> 214  -1.077966e-01 -0.0694409763       0       0
#> 215  -5.708892e-02 -0.0190390493       0       0
#> 216  -1.348696e-01 -0.0024063006       0       0
#> 217  -4.781319e-02 -0.0376638693       0       0
#> 218  -1.724326e-01  0.0467898539       0       0
#> 219  -3.535254e-01  0.2299753038       0       0
#> 220  -2.280321e-01  0.1764683584       0       0
#> 221  -1.268188e-01  0.0301568219       0       0
#> 222  -2.223222e-01  0.1476685399       0       0
#> 223  -1.054301e-01  0.0324233899       0       0
#> 224  -2.633017e-02  0.1416504265       0       0
#> 225  -1.631819e-01  0.0776987246       0       0
#> 226  -2.268453e-01  0.0462345545       0       0
#> 227  -1.684091e-01  0.1092441438       0       0
#> 228  -7.921618e-02  0.1241897380       0       0
#> 229  -6.146078e-02  0.0586251867       0       0
#> 230  -2.054677e-01  0.2172835432       0       0
#> 231  -6.569837e-02  0.1760848617       0       0
#> 232  -1.004874e-01  0.1348248535       0       0
#> 233  -5.369685e-02  0.0189085204       0       0
#> 234  -9.267311e-02  0.1998430067       0       0
#> 235  -1.592305e-01  0.1660385255       0       0
#> 236  -8.108319e-02  0.0041295606       0       0
#> 237  -8.460073e-02 -0.0645967368       0       0
#> 238  -1.466788e-01  0.3299473700       0       0
#> 239  -1.591361e-02  0.0740739781       0       0
#> 240  -7.416657e-02 -0.1403829767       0       0
#> 241  -1.017616e-01  0.0010343717       0       0
#> 242  -1.592305e-01  0.1660385255       0       0
#> 243  -1.574581e-01  0.1293898535       0       0
#> 244  -2.957507e-02  0.0105967618       0       0
#> 245  -2.048534e-01  0.1446286508       0       0
#> 246  -3.386237e-01  0.2684013123       0       0
#> 247  -1.499132e-01  0.0849708399       0       0
#> 248  -2.802471e-01  0.3713649918       0       0
#> 249  -1.187726e-01  0.0035917612       0       0
#> 250  -3.769237e-02  0.0035275295       0       0
#> 251  -5.632422e-02 -0.1326303049       0       0
#> 252  -1.709595e-01  0.0343472644       0       0
#> 253  -1.539648e-01  0.1395807106       0       0
#> 254  -1.688153e-01 -0.1161638144       0       0
#> 255   2.974804e-02  0.0171953080       0       0
#> 256  -5.965536e-02 -0.0995723070       0       0
#> 257  -1.946648e-02 -0.0623416192       0       0
#> 258  -1.216721e-01  0.3193997419       0       0
#> 259  -4.750017e-01  0.0795800965       0       0
#> 260  -2.433851e-02  0.1066489123       0       0
#> 261  -3.388658e-02  0.0357818009       0       0
#> 262  -1.266561e-01  0.2300449351       0       0
#> 263  -6.877969e-02  0.2376668410       0       0
#> 264  -1.857357e-01  0.2775356688       0       0
#> 265  -1.213587e-03  0.0118124017       0       0
#> 266  -1.277940e-01  0.0217315363       0       0
#> 267  -5.362495e-02  0.1069887864       0       0
#> 268  -1.624502e-01  0.2513177452       0       0
#> 269  -2.947887e-01 -0.0217908338       0       0
#> 270  -2.700247e-01  0.1798340745       0       0
#> 271  -2.236681e-01  0.0207568428       0       0
#> 272  -9.612677e-02  0.1884162783       0       0
#> 273  -1.095532e-01  0.0613773140       0       0
#> 274  -1.679384e-01  0.1075735326       0       0
#> 275  -1.628324e-01  0.1835723315       0       0
#> 276  -6.633044e-02  0.0302448925       0       0
#> 277  -3.956585e-02 -0.1130002482       0       0
#> 278  -2.075214e-01  0.2421530842       0       0
#> 279  -2.717518e-01  0.1773708610       0       0
#> 280  -1.086221e-03  0.0927485385       0       0
#> 281   5.332316e-02 -0.0522080567       0       0
#> 282  -2.385024e-01  0.2526715122       0       0
#> 283  -1.971930e-01  0.1333877412       0       0
#> 284   3.316799e-03 -0.0595205046       0       0
#> 285  -1.283811e-01  0.1161497164       0       0
#> 286  -9.535028e-02  0.1442501759       0       0
#> 287  -3.585779e-01  0.1027217836       0       0
#> 288  -8.323213e-03  0.0187620344       0       0
#> 289   1.385669e-02  0.0492199933       0       0
#> 290  -1.832298e-01  0.2142279880       0       0
#> 291  -2.272338e-01  0.1927464124       0       0
#> 292  -1.139667e-01 -0.0256892215       0       0
#> 293  -1.014253e-01 -0.0954882312       0       0
#> 294  -2.037850e-01 -0.0262155638       0       0
#> 295  -7.683593e-02 -0.1306836677       0       0
#> 296  -1.627940e-01  0.0339601800       0       0
#> 297  -2.764255e-02 -0.0327704832       0       0
#> 298  -3.842095e-02  0.0235023301       0       0
#> 299  -5.448343e-02  0.1721607523       0       0
#> 300  -4.420411e-02  0.0024827132       0       0
#> 301  -5.581359e-02 -0.0010291858       0       0
#> 302  -8.775501e-02 -0.0300196069       0       0
#> 303  -1.979903e-01 -0.0243650484       0       0
#> 304  -2.419281e-01  0.1628256354       0       0
#> 305  -2.700247e-01  0.1798340745       0       0
#> 306  -3.332455e-02  0.1852523587       0       0
#> 307  -7.389025e-02  0.0683693785       0       0
#> 308  -3.024065e-02  0.0055501323       0       0
#> 309   2.819664e-02 -0.1929719319       0       0
#> 310  -1.779876e-02 -0.0383904791       0       0
#> 311  -1.510709e-01  0.2971956514       0       0
#> 312  -1.062234e-01  0.0778502585       0       0
#> 313  -4.059085e-02  0.0469523027       0       0
#> 314  -1.984447e-01  0.0837422619       0       0
#> 315  -2.454342e-01  0.0417722034       0       0
#> 316  -9.923665e-02 -0.0159832527       0       0
#> 317  -1.760947e-01  0.1353528873       0       0
#> 318  -1.514197e-01  0.0817832937       0       0
#> 319  -9.781452e-02  0.0606882045       0       0
#> 320  -8.387885e-02  0.0747495398       0       0
#> 321  -5.546761e-02  0.0656400156       0       0
#> 322  -1.826995e-01  0.0027092620       0       0
#> 323  -1.323000e-01  0.0455593812       0       0
#> 324  -2.802471e-01  0.3713649918       0       0
#> 325   1.680169e-02 -0.0860222240       0       0
#> 326  -2.495301e-03  0.0849611545       0       0
#> 327  -4.024798e-02  0.1189755318       0       0
#> 328  -3.091008e-01  0.2004981326       0       0
#> 329  -1.484728e-01  0.3274836243       0       0
#> 330  -1.499048e-01  0.1163289787       0       0
#> 331  -1.684537e-01  0.1693757395       0       0
#> 332   6.882680e-02 -0.1094585369       0       0
#> 333  -1.877961e-01 -0.0023948600       0       0
#> 334  -1.988216e-01  0.1352218878       0       0
#> 335  -5.786335e-02 -0.1075973533       0       0
#> 336  -2.591517e-01  0.1147716722       0       0
#> 337  -1.376757e-01  0.0243514236       0       0
#> 338  -1.710731e-01  0.0844845005       0       0
#> 339  -5.369685e-02  0.0189085204       0       0
#> 340  -1.699715e-02 -0.0113634080       0       0
#> 341  -1.521534e-01  0.0436720063       0       0
#> 342  -1.431458e-01 -0.0502384787       0       0
#> 343  -9.940327e-02  0.1571106935       0       0
#> 344  -1.791383e-01  0.1355327248       0       0
#> 345  -1.336546e-01  0.2966594426       0       0
#> 346  -1.713108e-01  0.1263910140       0       0
#> 347  -1.007820e-01 -0.0395210418       0       0
#> 348   1.227957e-03 -0.0683388641       0       0
#> 349  -2.492524e-01  0.0931791404       0       0
#> 350  -7.253856e-02  0.0742412219       0       0
#> 351  -8.643999e-02 -0.1132044944       0       0
#> 352  -2.187064e-01  0.1604129002       0       0
#> 353  -6.362138e-02  0.1291960547       0       0
#> 354  -7.881423e-02  0.0578379713       0       0
#> 355  -1.342231e-01  0.0330371934       0       0
#> 356  -1.684287e-02 -0.0400134101       0       0
#> 357  -1.843130e-02  0.1058151602       0       0
#> 358  -7.154681e-02  0.1373567715       0       0
#> 359  -1.724735e-01  0.1331414811       0       0
#> 360  -1.050501e-01  0.0988801026       0       0
#> 361  -2.176681e-01 -0.0343413159       0       0
#> 362  -7.745906e-02  0.2549041830       0       0
#> 363   2.918479e-02 -0.1241001708       0       0
#> 364  -2.152475e-01  0.1247595393       0       0
#> 365  -1.019731e-01 -0.0174403248       0       0
#> 366  -1.605132e-03  0.1073176013       0       0
#> 367  -3.810682e-02  0.0103791270       0       0
#> 368  -3.149431e-01  0.2058339369       0       0
#> 369  -2.478016e-01  0.0929262977       0       0
#> 370  -2.016586e-01  0.1065758397       0       0
#> 371  -1.814761e-01 -0.0356409801       0       0
#> 372  -1.007779e-01  0.0437969110       0       0
#> 373  -2.274168e-01  0.2558291152       0       0
#> 374  -2.964109e-01  0.2539135549       0       0
#> 375  -3.555597e-01  0.0701557346       0       0
#> 376   7.039209e-03 -0.0481198945       0       0
#> 377  -1.550224e-01 -0.0287123711       0       0
#> 378  -6.467788e-02  0.1054552531       0       0
#> 379  -2.675288e-02 -0.0211040305       0       0
#> 380  -1.820353e-01  0.1781600174       0       0
#> 381  -3.564747e-03  0.1966408140       0       0
#> 382  -6.652071e-02  0.0897780877       0       0
#> 383  -1.574581e-01  0.1293898535       0       0
#> 384  -2.616023e-01 -0.0800955192       0       0
#> 385  -1.985565e-01  0.2746844096       0       0
#> 386  -3.772815e-02 -0.1626743740       0       0
#> 387  -2.364231e-01  0.1712235445       0       0
#> 388  -4.513672e-02 -0.0842708586       0       0
#> 389  -1.855366e-01  0.2422995460       0       0
#> 390  -3.886466e-02  0.0101866094       0       0
#> 391  -2.930059e-01  0.3137968932       0       0
#> 392  -2.419281e-01  0.1628256354       0       0
#> 393  -1.545197e-01  0.1648945599       0       0
#> 394  -8.242559e-02  0.0850033009       0       0
#> 395  -1.391362e-01 -0.0187661837       0       0
#> 396  -2.547665e-02  0.0617780634       0       0
#> 397   4.387929e-02 -0.1051723582       0       0
#> 398  -2.497839e-01  0.2735576502       0       0
#> 399  -3.011049e-02 -0.0026845732       0       0
#> 400  -7.284575e-02  0.0670306309       0       0
#> 401  -2.126789e-01  0.2792615940       0       0
#> 402  -7.354675e-02  0.0023956118       0       0
#> 403  -1.175574e-01  0.0714492181       0       0
#> 404  -8.668481e-02 -0.0156368335       0       0
#> 405  -8.185514e-02  0.1638473384       0       0
#> 406  -1.163410e-01  0.0816861711       0       0
#> 407  -2.714858e-01  0.1992411294       0       0
#> 408  -2.102573e-01  0.0813832838       0       0
#> 409  -4.228679e-02  0.1007023551       0       0
#> 410  -1.574123e-01  0.1455389619       0       0
#> 411  -1.060384e-01  0.2039224514       0       0
#> 412  -3.032364e-01  0.0356909396       0       0
#> 413  -1.503026e-01  0.0757653392       0       0
#> 414  -6.644328e-02  0.1215453142       0       0
#> 415   1.094438e-01 -0.0989326428       0       0
#> 416  -8.445059e-02  0.1190380501       0       0
#> 417  -7.513917e-02  0.0186219206       0       0
#> 418  -4.260653e-01  0.3167975652       0       0
#> 419  -6.107280e-02  0.0871352795       0       0
#> 420  -1.777319e-01  0.2049568150       0       0
#> 421  -2.129613e-01  0.0985815961       0       0
#> 422  -2.517035e-02  0.0122342763       0       0
#> 423   1.095089e-02 -0.0584197250       0       0
#> 424  -8.301949e-02  0.0418945916       0       0
#> 425  -1.165306e-01 -0.0220061683       0       0
#> 426  -2.070258e-01  0.1727539366       0       0
#> 427  -1.463057e-01  0.1660706100       0       0
#> 428  -1.967438e-01  0.3056698714       0       0
#> 429  -2.591517e-01  0.1147716722       0       0
#> 430  -9.736906e-02  0.1748790713       0       0
#> 431  -1.015589e-01  0.1439736086       0       0
#> 432  -6.666085e-03  0.1341701636       0       0
#> 433  -1.803915e-02  0.0617223582       0       0
#> 434  -2.226163e-01  0.1250144220       0       0
#> 435   8.173245e-03 -0.0258059512       0       0
#> 436  -1.319675e-01  0.0988891957       0       0
#> 437  -6.177162e-02 -0.0641518831       0       0
#> 438  -7.989909e-02 -0.1207332063       0       0
#> 439  -1.394327e-01 -0.0069716444       0       0
#> 440   2.162314e-04  0.0056630496       0       0
#> 441  -2.389833e-01  0.1014894292       0       0
#> 442  -2.449050e-01  0.0921535877       0       0
#> 443  -1.348696e-01 -0.0024063006       0       0
#> 444  -2.165540e-01  0.0934299760       0       0
#> 445  -1.283811e-01  0.1161497164       0       0
#> 446   1.536755e-02 -0.0314665881       0       0
#> 447  -7.737216e-02  0.1133938651       0       0
#> 448  -8.631973e-02  0.1200401305       0       0
#> 449  -1.868384e-01  0.2320386076       0       0
#> 450  -1.624502e-01  0.2513177452       0       0
#> 451  -8.042505e-02  0.1066007804       0       0
#> 452  -1.319380e-01  0.0104155194       0       0
#> 453  -9.091687e-02  0.1460988401       0       0
#> 454  -6.531097e-02  0.0957365067       0       0
#> 455  -1.134412e-01  0.1336676429       0       0
#> 456  -7.042694e-02 -0.0229133637       0       0
#> 457   4.952173e-03 -0.1392453312       0       0
#> 458  -2.458797e-01  0.0611520211       0       0
#> 459  -1.321274e-01  0.0006626474       0       0
#> 460  -1.727988e-02 -0.0892882021       0       0
#> 461  -1.665733e-01  0.1656026918       0       0
#> 462  -1.738744e-02 -0.0551662437       0       0
#> 463  -3.799501e-02  0.0915658418       0       0
#> 464  -1.335474e-01  0.0658527774       0       0
#> 465  -3.149431e-01  0.2058339369       0       0
#> 466  -5.259968e-02 -0.1447972491       0       0
#> 467   2.383035e-02  0.0667070135       0       0
#> 468  -3.454640e-02  0.1206488479       0       0
#> 469  -1.456362e-01  0.0263451665       0       0
#> 470  -6.644328e-02  0.1215453142       0       0
#> 471  -6.832807e-02  0.0633998591       0       0
#> 472  -3.988727e-02  0.2214840482       0       0
#> 473  -9.143457e-02  0.0115290822       0       0
#> 474  -2.659504e-01  0.0246813321       0       0
#> 475  -9.804059e-02  0.0485742392       0       0
#> 476  -1.749692e-01  0.0170379962       0       0
#> 477  -1.900011e-01  0.1593040818       0       0
#> 478  -1.499048e-01  0.1163289787       0       0
#> 479  -1.651830e-01 -0.0041594936       0       0
#> 480  -1.018541e-01  0.0836112342       0       0
#> 481  -2.274168e-01  0.2558291152       0       0
#> 482  -2.125631e-01  0.2306881493       0       0
#> 483  -1.007597e-01  0.1912339940       0       0
#> 484  -1.499048e-01  0.1163289787       0       0
#> 485   2.710408e-02 -0.0949565433       0       0
#> 486  -2.095514e-01  0.1525260349       0       0
#> 487  -1.159154e-01  0.2759911629       0       0
#> 488  -1.380500e-01 -0.1183544457       0       0
#> 489  -9.416918e-02 -0.0819438371       0       0
#> 490  -6.574937e-04 -0.1307592561       0       0
#> 491  -2.411495e-01  0.1554876409       0       0
#> 492  -1.086794e-01  0.0760123356       0       0
#> 493  -2.388225e-01  0.0122778791       0       0
#> 494  -1.681102e-01  0.0489558214       0       0
#> 495  -2.086019e-02  0.1259578005       0       0
#> 496  -3.245422e-02  0.0658113320       0       0
#> 497  -9.644819e-02  0.0586075829       0       0
#> 498  -8.688216e-02 -0.0297549775       0       0
#> 499   6.371825e-02  0.0465344208       0       0
#> 500  -9.871290e-02  0.0470171335       0       0
#> 501  -3.102343e-01  0.1514564382       0       0
#> 502  -5.868961e-02 -0.0803215241       0       0
#> 503  -4.111313e-02  0.1061916165       0       0
#> 504  -8.360826e-02  0.0414479445       0       0
#> 505  -6.529418e-02  0.1641817550       0       0
#> 506  -2.710301e-01  0.0034041163       0       0
#> 507   1.197670e-02  0.1137052409       0       0
#> 508  -4.252582e-02  0.0333574608       0       0
#> 509  -1.181164e-01  0.0728641988       0       0
#> 510  -7.407028e-02  0.0510587297       0       0
#> 511   2.336111e-02  0.0290998321       0       0
#> 512  -2.192976e-01  0.0064055495       0       0
#> 513  -1.367232e-01 -0.1000148216       0       0
#> 514   6.202578e-04  0.0904268340       0       0
#> 515  -1.229135e-01  0.0674858364       0       0
#> 516   7.861369e-02 -0.1787345101       0       0
#> 517  -1.381353e-01  0.0172190207       0       0
#> 518  -6.290206e-02  0.1390986814       0       0
#> 519  -1.778199e-01  0.0545634254       0       0
#> 520  -2.213269e-01  0.0416013288       0       0
#> 521  -8.573997e-02 -0.0124687763       0       0
#> 522  -1.883220e-01  0.1241804701       0       0
#> 523  -1.574330e-01  0.0131828274       0       0
#> 524  -1.534654e-01  0.0081489890       0       0
#> 525  -1.663323e-01  0.1793861412       0       0
#> 526  -1.038249e-01  0.1213348247       0       0
#> 527  -1.831324e-01  0.0111204168       0       0
#> 528  -1.108586e-01  0.0463595228       0       0
#> 529  -1.905981e-01 -0.0021796156       0       0
#> 530  -7.701812e-02  0.1188676894       0       0
#> 531  -1.953619e-01  0.1375857780       0       0
#> 532  -2.599932e-01  0.2546086977       0       0
#> 533  -7.921618e-02  0.1241897380       0       0
#> 534   9.480763e-02 -0.1950786684       0       0
#> 535  -1.574581e-01  0.1293898535       0       0
#> 536   3.775087e-02 -0.0604571567       0       0
#> 537  -8.258918e-02  0.0725330617       0       0
#> 538  -3.440342e-01  0.2435401668       0       0
#> 539  -1.659185e-01  0.1689015534       0       0
#> 540  -7.513917e-02  0.0186219206       0       0
#> 541  -4.857054e-02  0.1326504902       0       0
#> 542  -2.500729e-01  0.2993730880       0       0
#> 543  -1.777319e-01  0.2049568150       0       0
#> 544  -2.909247e-02 -0.1125566500       0       0
#> 545  -2.799440e-01 -0.0381078587       0       0
#> 546  -8.188672e-02  0.0450400546       0       0
#> 547  -4.696140e-02 -0.1481552383       0       0
#> 548  -4.784494e-02  0.0207253479       0       0
#> 549  -6.915050e-02  0.2528377044       0       0
#> 550  -2.636902e-01  0.1133096884       0       0
#> 551  -1.189732e-01  0.1058772666       0       0
#> 552  -2.400373e-02 -0.0499748803       0       0
#> 553  -2.749196e-01  0.1006882986       0       0
#> 554  -1.848208e-01  0.1721437831       0       0
#> 555  -1.841884e-01  0.0594953248       0       0
#> 556  -9.206967e-02 -0.0330429009       0       0
#> 557   1.185649e-02  0.0430535974       0       0
#> 558  -1.656303e-01 -0.0826851290       0       0
#> 559  -4.366711e-01  0.0361828791       0       0
#> 560  -1.769373e-01  0.0215531121       0       0
#> 561  -1.545197e-01  0.1648945599       0       0
#> 562  -7.429072e-02  0.0626376974       0       0
#> 563  -1.462061e-01  0.0943132095       0       0
#> 564  -2.979051e-02  0.0233956176       0       0
#> 565  -2.904401e-01  0.1969269912       0       0
#> 566  -5.765560e-02 -0.0335504317       0       0
#> 567  -9.165950e-02  0.1261470045       0       0
#> 568  -3.269023e-01  0.0994144196       0       0
#> 569  -5.637163e-02 -0.0406049368       0       0
#> 570  -1.550264e-02  0.0753496976       0       0
#> 571  -3.320666e-01  0.0080434282       0       0
#> 572  -9.508704e-02  0.1246247906       0       0
#> 573  -2.714858e-01  0.1992411294       0       0
#> 574  -2.075214e-01  0.2421530842       0       0
#> 575  -8.739831e-02 -0.0930050148       0       0
#> 576  -2.776663e-01  0.2053543912       0       0
#> 577  -5.613204e-02 -0.0409358444       0       0
#> 578  -1.278518e-01  0.1283187020       0       0
#> 579  -1.252770e-01 -0.0922352251       0       0
#> 580  -1.391435e-01  0.1192098321       0       0
#> 581  -1.618151e-01 -0.0128850496       0       0
#> 582  -2.385024e-01  0.2526715122       0       0
#> 583  -1.206369e-01  0.2851685279       0       0
#> 584  -1.082324e-01  0.0938814619       0       0
#> 585  -1.279229e-01 -0.0245986050       0       0
#> 586  -8.258759e-04 -0.1005143298       0       0
#> 587  -1.663146e-02  0.1286131597       0       0
#> 588  -1.494217e-01  0.0435951132       0       0
#> 589  -1.096309e-01 -0.0063409648       0       0
#> 590  -1.842760e-01  0.1411124225       0       0
#> 591  -3.799501e-02  0.0915658418       0       0
#> 592  -1.386532e-02  0.0443823175       0       0
#> 593  -3.782733e-01  0.3327009653       0       0
#> 594  -3.048543e-01  0.1266432850       0       0
#> 595  -1.400395e-01 -0.0077895311       0       0
#> 596  -3.931001e-02  0.1410194398       0       0
#> 597  -1.318267e-01 -0.0744245876       0       0
#> 598  -2.309271e-01 -0.0883581950       0       0
#> 599  -2.930059e-01  0.3137968932       0       0
#> 600  -1.255651e-01  0.0939007149       0       0
#> 601  -4.435171e-02 -0.0241976382       0       0
#> 602  -2.212296e-01  0.0890988115       0       0
#> 603  -1.021756e-01  0.1097660039       0       0
#> 604  -1.759868e-01  0.1990409176       0       0
#> 605   3.851918e-02  0.1495318224       0       0
#> 606  -8.608769e-02  0.1779798713       0       0
#> 607  -1.478856e-01 -0.0219983363       0       0
#> 608  -1.336546e-01  0.2966594426       0       0
#> 609  -3.577244e-01  0.0142802165       0       0
#> 610   1.782881e-02  0.0909432192       0       0
#> 611  -5.645446e-02  0.0560159905       0       0
#> 612  -8.537466e-02  0.0168718109       0       0
#> 613   5.660262e-02 -0.1418119363       0       0
#> 614  -1.913367e-01  0.1043956399       0       0
#> 615  -6.467788e-02  0.1054552531       0       0
#> 616   4.479787e-02 -0.0868428852       0       0
#> 617  -4.195170e-02 -0.0225954694       0       0
#> 618  -1.334730e-01 -0.1948845663       0       0
#> 619   6.076043e-02 -0.1931677261       0       0
#> 620  -9.781452e-02  0.0606882045       0       0
#> 621  -1.875395e-01  0.1124073815       0       0
#> 622  -1.221162e-01  0.0147617332       0       0
#> 623  -3.187296e-02 -0.0921577153       0       0
#> 624  -6.893167e-02  0.0549527041       0       0
#> 625  -3.388658e-02  0.0357818009       0       0
#> 626  -9.432498e-02 -0.0656030173       0       0
#> 627  -1.658220e-01  0.0939246154       0       0
#> 628  -4.686489e-02 -0.0786192051       0       0
#> 629  -8.884485e-02  0.1129248670       0       0
#> 630  -1.488258e-01  0.1415397097       0       0
#> 631  -2.509468e-01 -0.0244672650       0       0
#> 632  -1.437535e-01  0.0864820010       0       0
#> 633  -5.890062e-02  0.1609144059       0       0
#> 634  -6.341211e-03 -0.0372606178       0       0
#> 635  -5.138954e-02  0.0454052546       0       0
#> 636  -6.470050e-02  0.0021239515       0       0
#> 637  -1.624502e-01  0.2513177452       0       0
#> 638  -9.002321e-02  0.2059683837       0       0
#> 639  -3.066554e-01  0.2443637177       0       0
#> 640  -3.386237e-01  0.2684013123       0       0
#> 641  -4.045320e-02 -0.1397276619       0       0
#> 642  -2.665714e-01  0.2546361695       0       0
#> 643  -1.237670e-01  0.0637614809       0       0
#> 644  -2.949100e-01  0.0572680884       0       0
#> 645  -6.381924e-02  0.0370208093       0       0
#> 646  -1.553032e-01  0.1189687025       0       0
#> 647  -3.419087e-01 -0.1528411402       0       0
#> 648  -6.185532e-02  0.1173048924       0       0
#> 649  -2.802471e-01  0.3713649918       0       0
#> 650  -8.501769e-03 -0.0591609398       0       0
#> 651  -2.659504e-01  0.0246813321       0       0
#> 652   3.391834e-02 -0.0905035377       0       0
#> 653   2.500831e-03 -0.0094045504       0       0
#> 654  -5.732902e-02  0.0930634068       0       0
#> 655  -9.508704e-02  0.1246247906       0       0
#> 656  -1.120320e-01  0.0509729233       0       0
#> 657   2.080936e-02  0.0315377442       0       0
#> 658  -1.212554e-02  0.0499426718       0       0
#> 659  -1.516763e-01  0.1249516823       0       0
#> 660  -7.496476e-02  0.1739850464       0       0
#> 661  -4.576205e-02 -0.2161741065       0       0
#> 662  -1.210765e-01  0.1063627836       0       0
#> 663  -2.378939e-01  0.2010988024       0       0
#> 664  -1.442699e-01 -0.0096651934       0       0
#> 665  -1.335474e-01  0.0658527774       0       0
#> 666  -9.141049e-03 -0.0209662448       0       0
#> 667  -1.339382e-01 -0.0408460737       0       0
#> 668   3.542378e-02  0.0578311706       0       0
#> 669  -1.662989e-01  0.0185990105       0       0
#> 670  -2.781265e-02 -0.1840542763       0       0
#> 671  -1.751508e-01  0.0116369654       0       0
#> 672  -6.990088e-02  0.0623988406       0       0
#> 673  -1.081443e-01  0.0891143077       0       0
#> 674  -1.810442e-01  0.3369571373       0       0
#> 675  -1.060384e-01  0.2039224514       0       0
#> 676  -9.923665e-02 -0.0159832527       0       0
#> 677  -4.822701e-02 -0.0784192714       0       0
#> 678  -1.790067e-01  0.1244942710       0       0
#> 679  -9.804059e-02  0.0485742392       0       0
#> 680  -1.387880e-01 -0.2053830011       0       0
#> 681  -2.760113e-02 -0.0350707586       0       0
#> 682  -9.244213e-02  0.1841397400       0       0
#> 683   1.180840e-02 -0.0919085430       0       0
#> 684  -6.884242e-03 -0.2030634593       0       0
#> 685  -3.186252e-01  0.3315286589       0       0
#> 686  -1.508078e-01  0.0433291485       0       0
#> 687  -1.608195e-01  0.0516131763       0       0
#> 688  -3.783060e-01  0.1707674110       0       0
#> 689  -1.857357e-01  0.2775356688       0       0
#> 690  -1.172512e-01  0.0734165020       0       0
#> 691  -1.108586e-01  0.0463595228       0       0
#> 692  -9.214256e-02  0.1322606543       0       0
#> 693  -1.567406e-01  0.0203798543       0       0
#> 694  -8.609655e-02  0.0273614713       0       0
#> 695  -1.853487e-01 -0.0416516991       0       0
#> 696   2.184776e-02 -0.1154068458       0       0
#> 697  -1.177679e-01 -0.0780686988       0       0
#> 698  -8.048076e-02 -0.0580030540       0       0
#> 699  -1.054301e-01  0.0324233899       0       0
#> 700  -1.471141e-01  0.1017503674       0       0
#> 701  -3.987628e-02  0.0795573510       0       0
#> 702  -2.930059e-01  0.3137968932       0       0
#> 703  -7.042694e-02 -0.0229133637       0       0
#> 704  -8.371820e-03 -0.2302221212       0       0
#> 705  -2.964109e-01  0.2539135549       0       0
#> 706  -3.849204e-01  0.2234380033       0       0
#> 707  -1.583257e-01  0.1266901704       0       0
#> 708  -2.819550e-01 -0.0389610463       0       0
#> 709   4.535738e-02 -0.1097389810       0       0
#> 710  -9.758092e-02  0.0460308937       0       0
#> 711  -1.046558e-01  0.1849801428       0       0
#> 712  -1.829513e-01 -0.1186320711       0       0
#> 713  -2.957507e-02  0.0105967618       0       0
#> 714  -7.752637e-02  0.1074816289       0       0
#> 715  -7.557041e-02  0.0296323218       0       0
#> 716  -1.510709e-01  0.2971956514       0       0
#> 717  -8.288334e-05 -0.1429648013       0       0
#> 718   4.659730e-03 -0.0288542246       0       0
#> 719  -4.766433e-02  0.0404182295       0       0
#> 720  -2.350495e-02  0.0313949240       0       0
#> 721  -1.019282e-01  0.1045569303       0       0
#> 722  -2.049575e-01  0.1126319426       0       0
#> 723  -6.959924e-02 -0.0736991663       0       0
#> 724  -2.413392e-01  0.0278531067       0       0
#> 725  -1.414962e-01  0.0297746641       0       0
#> 726  -2.556918e-01 -0.0045183384       0       0
#> 727  -1.286817e-01  0.2732124901       0       0
#> 728  -1.072688e-01 -0.0103845316       0       0
#> 729  -1.848208e-01  0.1721437831       0       0
#> 730  -2.063992e-01  0.1043008191       0       0
#> 731  -3.329934e-02 -0.0728997482       0       0
#> 732  -3.912921e-02 -0.2098791517       0       0
#> 733  -2.059083e-01  0.0622013851       0       0
#> 734  -1.784949e-01  0.0031307233       0       0
#> 735  -2.562715e-01  0.1805882190       0       0
#> 736  -2.240779e-01  0.1940164203       0       0
#> 737  -1.480492e-02  0.0242374453       0       0
#> 738  -1.835516e-01  0.0839299524       0       0
#> 739  -2.603296e-01  0.3317180981       0       0
#> 740  -2.430792e-02  0.0424219104       0       0
#> 741  -7.737216e-02  0.1133938651       0       0
#> 742   8.213618e-03  0.0159032500       0       0
#> 743   1.540068e-02 -0.0897631847       0       0
#> 744  -3.316311e-01  0.0069055011       0       0
#> 745  -3.104761e-02 -0.1732785845       0       0
#> 746  -1.599353e-01 -0.0343929239       0       0
#> 747  -7.921023e-03  0.0123874510       0       0
#> 748  -1.448071e-01  0.1326750534       0       0
#> 749  -1.921658e-01  0.0805839379       0       0
#> 750  -2.016586e-01  0.1065758397       0       0
#> 751  -3.137579e-01 -0.0622149284       0       0
#> 752  -1.901994e-01  0.0485110972       0       0
#> 753  -6.735713e-02  0.0533242215       0       0
#> 754   1.971416e-02 -0.0985359012       0       0
#> 755  -3.403500e-01  0.0481090689       0       0
#> 756  -6.467788e-02  0.1054552531       0       0
#> 757  -1.508769e-01  0.1263600843       0       0
#> 758  -1.112416e-01  0.1086287159       0       0
#> 759  -1.073784e-01  0.0963045666       0       0
#> 760  -2.583968e-01  0.0876246842       0       0
#> 761  -1.932655e-01  0.0506020770       0       0
#> 762  -1.855366e-01  0.2422995460       0       0
#> 763  -1.060384e-01  0.2039224514       0       0
#> 764   2.380658e-03  0.0317958818       0       0
#> 765  -1.012525e-01 -0.0009022275       0       0
#> 766  -3.931001e-02  0.1410194398       0       0
#> 767  -6.235762e-02 -0.1521866682       0       0
#> 768  -1.206438e-02  0.0143172096       0       0
#> 769   5.906024e-02 -0.0355553963       0       0
#> 770  -5.839442e-02  0.1051982730       0       0
#> 771   4.048144e-02 -0.0301096174       0       0
#> 772  -5.839442e-02  0.1051982730       0       0
#> 773  -9.923665e-02 -0.0159832527       0       0
#> 774  -1.065888e-01 -0.0982085665       0       0
#> 775  -6.383969e-02  0.0439170446       0       0
#> 776  -1.838479e-01  0.2310259526       0       0
#> 777  -1.274825e-01  0.1816519214       0       0
#> 778  -1.613506e-01  0.0746854173       0       0
#> 779  -2.163374e-01  0.1965435357       0       0
#> 780  -1.159154e-01  0.2759911629       0       0
#> 781  -5.212490e-03  0.0521418357       0       0
#> 782  -6.414099e-02 -0.1119277693       0       0
#> 783  -5.364293e-02  0.1336991193       0       0
#> 784  -7.709225e-02 -0.0090524248       0       0
#> 785  -1.161379e-02 -0.0287661336       0       0
#> 786  -1.462061e-01  0.0943132095       0       0
#> 787  -6.845037e-02  0.0604533739       0       0
#> 788  -1.625921e-01  0.1147894068       0       0
#> 789  -1.081246e-01  0.2037629120       0       0
#> 790  -2.478016e-01  0.0929262977       0       0
#> 791  -1.065594e-01  0.1430035857       0       0
#> 792  -8.559517e-02 -0.1219962796       0       0
#> 793  -2.230828e-01  0.1320200268       0       0
#> 794  -2.984904e-01  0.1930832352       0       0
#> 795  -4.857054e-02  0.1326504902       0       0
#> 796  -2.714858e-01  0.1992411294       0       0
#> 797  -1.791067e-01  0.0099915823       0       0
#> 798  -1.656303e-01 -0.0826851290       0       0
#> 799  -8.122828e-03 -0.1224950069       0       0
#> 800  -1.751548e-01  0.1537434013       0       0
#> 801  -1.278289e-01  0.1146070843       0       0
#> 802  -1.510709e-01  0.2971956514       0       0
#> 803  -5.711585e-02  0.0876659389       0       0
#> 804  -1.651969e-01  0.1171212684       0       0
#> 805  -3.576781e-02  0.0437945637       0       0
#> 806  -1.593013e-01  0.1203210123       0       0
#> 807  -1.269119e-02  0.0810736757       0       0
#> 808  -6.362138e-02  0.1291960547       0       0
#> 809  -6.552190e-02  0.1569664956       0       0
#> 810  -2.093244e-01 -0.0054978801       0       0
#> 811  -1.756992e-01  0.0515388940       0       0
#> 812  -1.507730e-01  0.0937579371       0       0
#> 813  -6.491954e-02  0.0844732780       0       0
#> 814  -3.614064e-01  0.0140350001       0       0
#> 815   2.184776e-02 -0.1154068458       0       0
#> 816  -3.224376e-01  0.2283106709       0       0
#> 817  -1.095532e-01  0.0613773140       0       0
#> 818  -1.896068e-01 -0.0138324006       0       0
#> 819  -2.280321e-01  0.1764683584       0       0
#> 820  -1.442259e-01  0.1586673067       0       0
#> 821  -1.890531e-01  0.1214741848       0       0
#> 822  -3.742828e-01  0.1302014592       0       0
#> 823  -1.685856e-01 -0.0177685654       0       0
#> 824  -1.873608e-01  0.1041443924       0       0
#> 825  -1.181783e-02 -0.0579719313       0       0
#> 826  -1.184111e-01  0.0279311542       0       0
#> 827  -1.048252e-01  0.1589016287       0       0
#> 828  -1.677332e-01  0.0673121138       0       0
#> 829  -3.558288e-02  0.0048213144       0       0
#> 830  -1.120354e-01  0.1019870297       0       0
#> 831  -2.567310e-01  0.0339692029       0       0
#> 832  -1.851326e-01  0.1447318231       0       0
#> 833  -9.535028e-02  0.1442501759       0       0
#> 834  -1.636524e-01  0.0147106124       0       0
#> 835  -1.624502e-01  0.2513177452       0       0
#> 836  -2.327477e-02  0.1475446413       0       0
#> 837  -7.038320e-02 -0.0110076248       0       0
#> 838  -1.007597e-01  0.1912339940       0       0
#> 839  -1.248055e-01  0.1673828593       0       0
#> 840  -2.081040e-01  0.1497295684       0       0
#> 841  -1.206369e-01  0.2851685279       0       0
#> 842  -2.136411e-01 -0.0336105680       0       0
#> 843  -1.970368e-01  0.1583413415       0       0
#> 844  -4.934833e-02  0.0279590308       0       0
#> 845  -1.755184e-01  0.1376436854       0       0
#> 846  -1.596410e-01  0.0738371995       0       0
#> 847  -7.454772e-02 -0.0108319775       0       0
#> 848  -3.024065e-02  0.0055501323       0       0
#> 849  -3.000382e-01  0.1544930996       0       0
#> 850  -1.134412e-01  0.1336676429       0       0
#> 851   4.942523e-02 -0.0683314965       0       0
#> 852  -2.247865e-02 -0.0120073483       0       0
#> 853  -8.388664e-02 -0.0623340108       0       0
#> 854  -2.949100e-01  0.0572680884       0       0
#> 855  -1.301852e-01  0.0734005559       0       0
#> 856  -1.436682e-01  0.1268431440       0       0
#> 857  -1.466788e-01  0.3299473700       0       0
#> 858  -2.367141e-01  0.0036307700       0       0
#> 859   5.327467e-03 -0.0415183483       0       0
#> 860  -3.492413e-01 -0.0032042145       0       0
#> 861  -9.737694e-02  0.1108815783       0       0
#> 862  -7.458420e-02  0.0016663802       0       0
#> 863   2.178165e-02  0.0789592850       0       0
#> 864   3.798201e-02  0.0882409156       0       0
#> 865  -7.737216e-02  0.1133938651       0       0
#> 866  -2.054781e-02  0.0901164515       0       0
#> 867  -1.840729e-01  0.2893523702       0       0
#> 868  -1.189732e-01  0.1058772666       0       0
#> 869  -2.132327e-01  0.0552085540       0       0
#> 870   1.697558e-02  0.0551556759       0       0
#> 871  -2.508043e-01  0.1272466067       0       0
#> 872  -7.502773e-02  0.0322183530       0       0
#> 873  -2.930059e-01  0.3137968932       0       0
#> 874  -6.177162e-02 -0.0641518831       0       0
#> 875   2.485878e-02 -0.0218265373       0       0
#> 876  -4.850747e-02 -0.0422962442       0       0
#> 877  -1.171839e-01  0.2888087538       0       0
#> 878  -8.967054e-02 -0.0887510875       0       0
#> 879  -1.832298e-01  0.2142279880       0       0
#> 880  -2.083566e-01  0.1061666306       0       0
#> 881  -4.252582e-02  0.0333574608       0       0
#> 882  -1.159648e-01  0.0759586862       0       0
#> 883  -2.511382e-01 -0.0263740446       0       0
#> 884  -1.160695e-01 -0.0667408267       0       0
#> 885  -3.332455e-02  0.1852523587       0       0
#> 886  -8.609655e-02  0.0273614713       0       0
#> 887  -1.719256e-01  0.0275528729       0       0
#> 888  -2.231863e-01  0.0234931559       0       0
#> 889  -7.375745e-02  0.1101604294       0       0
#> 890  -1.484728e-01  0.3274836243       0       0
#> 891  -1.582966e-01  0.0907309702       0       0
#> 892  -2.041676e-01 -0.0763266001       0       0
#> 893  -8.462138e-02  0.1279969001       0       0
#> 894  -1.017765e-01  0.0852719808       0       0
#> 895  -1.814547e-01  0.1058679380       0       0
#> 896  -2.149830e-01  0.1398899303       0       0
#> 897  -5.839442e-02  0.1051982730       0       0
#> 898  -2.108900e-01  0.0876940586       0       0
#> 899  -9.937270e-02 -0.0716745837       0       0
#> 900  -2.817826e-01  0.1985549418       0       0
#> 901  -3.893068e-02 -0.0623461315       0       0
#> 902  -2.078567e-02  0.1494876754       0       0
#> 903  -1.879016e-01  0.1324035331       0       0
#> 904  -1.306771e-02  0.0008514075       0       0
#> 905  -6.695643e-02 -0.0191614138       0       0
#> 906  -2.899856e-01  0.1554681967       0       0
#> 907  -7.378462e-02 -0.1428535471       0       0
#> 908  -1.387174e-01  0.0540047789       0       0
#> 909  -3.371283e-02 -0.0015727362       0       0
#> 910  -4.192210e-02  0.0688877598       0       0
#> 911   3.170603e-02 -0.0057628268       0       0
#> 912  -1.206369e-01  0.2851685279       0       0
#> 913  -6.531551e-02  0.0481820855       0       0
#> 914  -1.855366e-01  0.2422995460       0       0
#> 915  -7.811184e-02 -0.1325399305       0       0
#> 916  -6.798269e-02  0.3069999072       0       0
#> 917  -3.845319e-01  0.2068201768       0       0
#> 918  -1.389230e-01 -0.0469473236       0       0
#> 919  -1.704587e-01 -0.1159099560       0       0
#> 920  -3.129200e-02 -0.0551146194       0       0
#> 921  -3.646825e-03  0.0734612837       0       0
#> 922  -1.007597e-01  0.1912339940       0       0
#> 923  -1.177679e-01 -0.0780686988       0       0
#> 924  -1.896068e-01 -0.0138324006       0       0
#> 925   5.579559e-02 -0.0345145668       0       0
#> 926  -1.653823e-01  0.1535450714       0       0
#> 927  -7.164301e-02 -0.0904438708       0       0
#> 928  -6.814680e-03  0.0354963889       0       0
#> 929  -1.263841e-01  0.0189892697       0       0
#> 930  -7.942435e-02  0.0914381053       0       0
#> 931  -1.364075e-01  0.0620332774       0       0
#> 932  -1.015589e-01  0.1439736086       0       0
#> 933  -3.650354e-02  0.1180871566       0       0
#> 934   3.171505e-02 -0.0763939173       0       0
#> 935  -7.580629e-03  0.0579855253       0       0
#> 936  -8.405650e-02 -0.1766297254       0       0
#> 937  -4.111313e-02  0.1061916165       0       0
#> 938  -1.832298e-01  0.2142279880       0       0
#> 939  -1.553385e-01  0.0273729881       0       0
#> 940  -4.311911e-02  0.0581109554       0       0
#> 941  -1.736881e-01  0.1695693635       0       0
#> 942  -1.955268e-01  0.1683274315       0       0
#> 943  -1.842760e-01  0.1411124225       0       0
#> 944  -1.171839e-01  0.2888087538       0       0
#> 945  -1.825651e-01  0.0866217517       0       0
#> 946  -2.013807e-01  0.2383112659       0       0
#> 947  -9.739264e-02  0.0446813398       0       0
#> 948  -7.453940e-02 -0.0528223572       0       0
#> 949  -7.711254e-02 -0.0006879567       0       0
#> 950  -1.814805e-01 -0.0491556463       0       0
#> 951  -2.443516e-01  0.2594840910       0       0
#> 952  -1.636524e-01  0.0147106124       0       0
#> 953  -2.125631e-01  0.2306881493       0       0
#> 954  -1.021756e-01  0.1097660039       0       0
#> 955  -1.276751e-01  0.0495562343       0       0
#> 956  -1.947037e-02 -0.1184191129       0       0
#> 957  -4.931779e-02 -0.0815121645       0       0
#> 958  -6.381924e-02  0.0370208093       0       0
#> 959  -1.101971e-01  0.0356633498       0       0
#> 960  -6.644251e-02  0.0274477580       0       0
#> 961  -9.141049e-03 -0.0209662448       0       0
#> 962  -2.129341e-01  0.1501156853       0       0
#> 963  -3.128347e-01  0.1269213628       0       0
#> 964  -2.242633e-01  0.2918133805       0       0
#> 965  -1.021756e-01  0.1097660039       0       0
#> 966   3.002463e-02  0.0700676751       0       0
#> 967  -4.124360e-02  0.0785495976       0       0
#> 968  -2.385024e-01  0.2526715122       0       0
#> 969  -2.122913e-02 -0.0647947425       0       0
#> 970  -4.795474e-02  0.0546677161       0       0
#> 971  -1.185272e-01  0.2232687428       0       0
#> 972  -1.556121e-01  0.0578039618       0       0
#> 973  -1.308802e-01  0.1848874952       0       0
#> 974  -1.829510e-01  0.0572680553       0       0
#> 975  -4.132862e-02  0.0014717214       0       0
#> 976  -2.899856e-01  0.1554681967       0       0
#> 977  -1.781290e-01  0.1231686461       0       0
#> 978   4.086533e-03  0.0378691884       0       0
#> 979  -1.679384e-01  0.1075735326       0       0
#> 980  -1.665015e-01  0.1980319075       0       0
#> 981  -1.171839e-01  0.2888087538       0       0
#> 982  -1.611685e-01  0.1946855020       0       0
#> 983  -8.491363e-02  0.1112240101       0       0
#> 984  -2.252975e-01  0.1171813792       0       0
#> 985  -6.724208e-02  0.0245379639       0       0
#> 986  -1.605132e-03  0.1073176013       0       0
#> 987  -1.479234e-01  0.1345812530       0       0
#> 988  -2.445547e-01  0.1556510196       0       0
#> 989  -2.179038e-01  0.1578038636       0       0
#> 990   6.455911e-02 -0.0037266411       0       0
#> 991   1.504646e-03  0.1810396412       0       0
#> 992  -1.565690e-01  0.1154040945       0       0
#> 993  -1.592305e-01  0.1660385255       0       0
#> 994  -7.638691e-02  0.0927579882       0       0
#> 995  -5.684483e-02 -0.0169724453       0       0
#> 996  -1.320651e-01  0.0363690210       0       0
#> 997  -7.513917e-02  0.0186219206       0       0
#> 998  -1.283811e-01  0.1161497164       0       0
#> 999  -1.814547e-01  0.1058679380       0       0
#> 1000 -3.395063e-01  0.1360600619       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12  0.0407 -0.0952 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.188   0.132  
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.126   0.0190 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.127   0.0302 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.0869  0.0788 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.280   0.371  
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0288  0.00494
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.181   0.0513 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0209  0.0171 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.0113  0.00894
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.259  17.7 0.041  0.12 90.4  1600. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.259  17.7 0.041  0.12  4.35   77.0
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.206  22.2 0.041  0.12 74.9  1665. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.206  22.2 0.041  0.12  5.98  133. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.219  19.8 0.041  0.12 83.1  1649. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.219  19.8 0.041  0.12  5.83  116. 
#> # … with 1,994 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>