Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object
A posologyr list, created by the posologyr function.

n_sample
Number of samples from the S-step

n_resample
Number of samples from the R-step

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)

#> 

#> $eta
#>             ETA_ke         ETA_V
#> 1    -1.528108e-01  0.0169469616
#> 2    -6.631472e-02 -0.0892642480
#> 3    -1.000767e-01  0.0327053682
#> 4    -8.202322e-02  0.0733732491
#> 5    -1.901611e-01  0.1563008866
#> 6    -4.529166e-02  0.0617049508
#> 7    -1.463569e-01  0.2665664879
#> 8    -3.048945e-01  0.1162788237
#> 9    -6.887731e-02 -0.0986242457
#> 10   -6.835101e-02  0.0066181315
#> 11   -1.583283e-01  0.0616762250
#> 12   -1.051144e-01  0.0895003952
#> 13   -1.519950e-01  0.1313111389
#> 14   -1.036570e-01  0.2383258797
#> 15   -2.218348e-01  0.1974216022
#> 16    3.714378e-02 -0.0119892051
#> 17   -9.897267e-02  0.1327982977
#> 18   -2.695441e-01  0.0963188259
#> 19   -2.296822e-01  0.0920953028
#> 20   -5.367948e-02  0.1683755823
#> 21    2.438082e-02  0.0494444076
#> 22   -3.512420e-02 -0.1275173617
#> 23   -9.793250e-02 -0.1727798779
#> 24   -6.196279e-02 -0.0152263582
#> 25   -6.247951e-03 -0.0162566214
#> 26   -2.603373e-02  0.2284219217
#> 27   -1.052774e-01  0.0129059468
#> 28   -2.826614e-02  0.0330574687
#> 29   -2.065510e-01  0.0248607692
#> 30   -1.839025e-01  0.0399526903
#> 31   -3.875947e-02  0.0648237093
#> 32   -1.831452e-01  0.2036167290
#> 33   -7.494208e-02 -0.1045178549
#> 34   -9.201715e-02 -0.0129669255
#> 35   -8.246111e-03 -0.0309270724
#> 36    6.229984e-02  0.0343741658
#> 37   -1.400061e-01  0.0973118840
#> 38   -5.237926e-02  0.0577879470
#> 39   -9.506661e-02 -0.0322078051
#> 40   -7.912833e-02 -0.0305619294
#> 41   -4.327172e-02 -0.0349674933
#> 42   -6.897539e-02  0.0680817306
#> 43   -3.187184e-02 -0.1181348253
#> 44   -8.395340e-02  0.0440665585
#> 45   -1.526193e-01  0.0646783370
#> 46   -5.396688e-02  0.1111734489
#> 47   -3.443973e-02 -0.1406371028
#> 48   -9.871109e-02  0.0877565632
#> 49   -8.668211e-02  0.0069657811
#> 50   -1.126605e-01 -0.0348452914
#> 51   -6.119380e-02  0.1235708689
#> 52   -7.185776e-02 -0.0030983227
#> 53   -5.681013e-02 -0.0941065854
#> 54   -7.185776e-02 -0.0030983227
#> 55   -2.278593e-01  0.1540218661
#> 56   -3.301554e-02 -0.1076348824
#> 57   -1.145758e-01  0.1042232441
#> 58   -1.830971e-01 -0.0325740827
#> 59   -2.379663e-01  0.1077118935
#> 60    1.806135e-02  0.0445929693
#> 61   -2.836786e-01  0.2832071487
#> 62    6.397671e-03  0.0575074830
#> 63   -1.226441e-01  0.0963186288
#> 64   -1.684338e-01  0.0724152384
#> 65   -1.743462e-01  0.0328845206
#> 66   -1.901662e-01  0.1857737961
#> 67   -2.850654e-01  0.2504002684
#> 68   -2.195004e-02 -0.0787907985
#> 69   -1.346793e-01  0.0234972091
#> 70   -2.362116e-01  0.1164390301
#> 71   -9.476067e-02 -0.1191961690
#> 72   -2.335690e-01  0.1475962401
#> 73   -7.382680e-04 -0.0255688514
#> 74    2.412541e-02  0.0846111558
#> 75   -4.268741e-02 -0.0360871441
#> 76   -8.111015e-02  0.0793557910
#> 77   -3.928564e-02  0.0085378064
#> 78   -2.260971e-01  0.1753001149
#> 79   -8.944789e-02 -0.0213736156
#> 80   -9.867648e-02  0.0959093433
#> 81   -1.391155e-01 -0.0404000540
#> 82   -2.096150e-02  0.0439099675
#> 83   -9.604261e-02  0.0178683281
#> 84   -4.395931e-02 -0.0027189584
#> 85   -2.280233e-01 -0.0209970157
#> 86    1.660815e-02 -0.1229443556
#> 87   -2.006780e-01  0.1421736796
#> 88   -1.656576e-01  0.0368047246
#> 89    2.872035e-02 -0.0143737965
#> 90   -2.290658e-01  0.1117544941
#> 91   -1.446221e-01  0.1678402761
#> 92   -1.132906e-02 -0.0959525117
#> 93   -1.170772e-01  0.0206686310
#> 94   -5.579184e-02  0.1407578709
#> 95   -9.628790e-02  0.2519326879
#> 96   -9.465164e-02  0.1220815956
#> 97   -9.870480e-02  0.0643454408
#> 98   -1.630309e-01 -0.0653380569
#> 99   -1.503168e-01  0.1166515182
#> 100  -9.794223e-02  0.0557743457
#> 101  -3.786024e-02  0.0508675258
#> 102  -2.082585e-01 -0.0136667689
#> 103  -9.307075e-02  0.1546311585
#> 104  -2.964674e-02  0.0746978860
#> 105  -9.628790e-02  0.2519326879
#> 106  -2.604950e-01  0.0132732067
#> 107  -2.190733e-01  0.2218969619
#> 108  -8.661423e-02  0.1687369862
#> 109   2.075747e-02 -0.0462238792
#> 110  -2.928237e-01  0.1582629514
#> 111   3.537475e-03 -0.1376184248
#> 112  -1.441819e-01 -0.0094103152
#> 113  -1.882083e-01  0.1738661308
#> 114  -1.038791e-01  0.0806167993
#> 115  -2.373755e-01  0.1585447785
#> 116  -3.617267e-02 -0.0168602291
#> 117  -6.093631e-02 -0.0226594097
#> 118  -7.864695e-02 -0.0109866981
#> 119  -1.317008e-01  0.0459943108
#> 120  -6.233213e-02  0.0606077743
#> 121   1.017431e-02  0.0678517539
#> 122  -9.422564e-02 -0.0413777249
#> 123  -1.553461e-01  0.0907065449
#> 124  -1.582858e-01  0.0546172505
#> 125   2.321051e-02  0.0570236081
#> 126   3.166111e-02 -0.0945156081
#> 127  -4.848568e-02  0.0163011460
#> 128  -2.479511e-01  0.2436397475
#> 129  -1.386433e-01  0.1094110583
#> 130  -1.959952e-01  0.0264563413
#> 131   4.123848e-02  0.0358138158
#> 132  -1.260579e-01  0.1541019492
#> 133  -6.835101e-02  0.0066181315
#> 134  -9.060209e-02  0.0662025203
#> 135   4.588600e-02 -0.0952710571
#> 136   6.363356e-02 -0.1131470300
#> 137  -1.869153e-01  0.0995502324
#> 138  -1.309646e-01  0.2619166687
#> 139  -2.218348e-01  0.1974216022
#> 140  -1.441819e-01 -0.0094103152
#> 141  -1.646290e-01  0.0268908800
#> 142  -1.340017e-01  0.0643077885
#> 143  -2.300598e-01  0.2057392483
#> 144  -1.791392e-01 -0.0158096813
#> 145  -1.026586e-01  0.0235095425
#> 146  -3.536198e-02  0.1462245005
#> 147  -1.026586e-01  0.0235095425
#> 148  -1.089476e-01  0.0660913589
#> 149  -8.652406e-02 -0.0474198784
#> 150  -2.819053e-01  0.0656865393
#> 151  -5.479192e-02 -0.1149812684
#> 152  -1.917771e-01  0.0508095480
#> 153  -1.401469e-01  0.0275542783
#> 154  -1.718533e-01  0.1563341210
#> 155  -1.721398e-01  0.1768468425
#> 156  -1.600994e-01 -0.0983582652
#> 157  -1.556414e-01  0.2174508285
#> 158  -8.202322e-02  0.0733732491
#> 159  -3.163525e-02 -0.0768310965
#> 160  -4.488281e-01  0.1240414027
#> 161  -1.735649e-01  0.1364635335
#> 162  -1.032840e-01  0.0038412462
#> 163  -9.422564e-02 -0.0413777249
#> 164  -1.855756e-01  0.0681282018
#> 165  -9.113420e-02  0.0882158358
#> 166   1.357964e-02 -0.0573281122
#> 167  -1.771861e-01  0.1629321312
#> 168  -1.599015e-02 -0.1887724957
#> 169  -1.818529e-01  0.2061114212
#> 170  -6.476768e-02  0.1274898681
#> 171  -1.027640e-01 -0.0101298544
#> 172  -1.924438e-01  0.0129516377
#> 173  -8.020841e-02  0.1723448342
#> 174  -1.015458e-01  0.0904164595
#> 175  -3.536198e-02  0.1462245005
#> 176  -6.659083e-02  0.1250043577
#> 177  -8.894230e-02  0.1321541934
#> 178  -1.962866e-01  0.1407040543
#> 179  -1.277587e-01  0.2144880670
#> 180  -2.236116e-01  0.1621439210
#> 181  -1.623920e-01  0.0998955409
#> 182  -3.714085e-02  0.0339873299
#> 183   1.708644e-02 -0.2025340960
#> 184  -1.161955e-01  0.0176311617
#> 185  -1.668026e-01  0.1626805459
#> 186  -1.203503e-01  0.1715626906
#> 187  -1.665847e-01  0.0973393352
#> 188  -6.089429e-02  0.1417175910
#> 189  -5.831817e-02 -0.0478669429
#> 190  -8.186309e-02  0.0811033929
#> 191  -1.388999e-01 -0.0134657164
#> 192  -2.036621e-01  0.1744875631
#> 193  -1.515700e-01 -0.1582771839
#> 194   6.046493e-02  0.0063314789
#> 195   5.067750e-02  0.0066900816
#> 196  -2.552489e-02 -0.1449445808
#> 197  -2.251428e-01  0.1379591123
#> 198  -1.461306e-01  0.1286021164
#> 199  -1.234223e-01  0.1313684684
#> 200  -1.466029e-01 -0.1389330770
#> 201  -3.338554e-01  0.0479633100
#> 202  -2.820164e-01  0.3487927175
#> 203  -4.936338e-01  0.0273950345
#> 204  -3.075022e-01  0.2830544036
#> 205  -2.385962e-01  0.1380729450
#> 206  -3.047488e-02  0.1694945365
#> 207  -7.928834e-03  0.0164442376
#> 208  -1.763429e-01 -0.0254327739
#> 209  -1.033998e-01  0.1658851730
#> 210  -2.832465e-02  0.0211453406
#> 211   2.109341e-02 -0.0489854636
#> 212  -9.383597e-02  0.0766663931
#> 213  -2.951761e-01  0.2185640974
#> 214  -3.714085e-02  0.0339873299
#> 215  -8.926497e-03  0.1418550335
#> 216  -7.668249e-02  0.1612701770
#> 217  -1.625726e-01  0.1976829896
#> 218  -2.365535e-01 -0.0162757990
#> 219  -8.052463e-02  0.0319182235
#> 220  -7.977232e-02  0.0001586398
#> 221  -1.679967e-01  0.0414960028
#> 222  -3.603475e-01  0.0987031023
#> 223  -1.194987e-01  0.0976476768
#> 224  -2.083517e-01  0.2102622421
#> 225  -2.373755e-01  0.1585447785
#> 226  -1.113448e-01  0.1680794786
#> 227  -2.451602e-02 -0.0238528913
#> 228  -3.589136e-01  0.1797428821
#> 229  -1.595261e-01  0.1369800853
#> 230  -2.447094e-02 -0.0676092737
#> 231  -2.595724e-01  0.2729482509
#> 232  -1.008001e-01  0.0735385553
#> 233  -1.018990e-01  0.1323924576
#> 234  -1.675844e-01  0.1028847746
#> 235  -2.048482e-01  0.0492207322
#> 236  -1.395796e-01 -0.0567828603
#> 237   6.738994e-02 -0.0918655425
#> 238  -1.115601e-01  0.0380730605
#> 239  -1.233636e-01  0.1574468614
#> 240  -3.983861e-02 -0.0615055366
#> 241  -5.483988e-02  0.1076011745
#> 242  -1.032243e-01  0.0086049939
#> 243  -2.008699e-01 -0.0249797403
#> 244  -2.311543e-01 -0.0321608519
#> 245  -1.319086e-01 -0.1362596059
#> 246  -1.051144e-01  0.0895003952
#> 247  -3.098655e-01  0.3118734442
#> 248  -9.577237e-02  0.0453532123
#> 249  -1.018990e-01  0.1323924576
#> 250  -2.510905e-01  0.2665200930
#> 251  -1.463569e-01  0.2665664879
#> 252  -3.075022e-01  0.2830544036
#> 253   1.202275e-02  0.0340602901
#> 254  -6.842080e-02  0.1923957258
#> 255  -2.145100e-01  0.2197182881
#> 256  -2.123730e-01 -0.0985159783
#> 257  -2.243657e-01  0.0104035154
#> 258  -2.071594e-01  0.1461335322
#> 259  -1.399665e-01  0.1512500056
#> 260  -1.929988e-01  0.1458240827
#> 261  -2.466432e-02  0.1285378637
#> 262  -1.451945e-01  0.1019524389
#> 263  -1.951956e-01  0.0366943344
#> 264  -1.776878e-01  0.0880976090
#> 265   2.792868e-02  0.1336352934
#> 266  -1.818508e-02 -0.0787113966
#> 267  -1.348001e-01  0.2502100830
#> 268  -1.625726e-01  0.1976829896
#> 269  -1.509188e-01 -0.0064543348
#> 270  -1.100660e-01 -0.0243893323
#> 271  -1.722774e-01  0.2562322060
#> 272  -5.552767e-02 -0.0654864082
#> 273  -2.146545e-01  0.0682534041
#> 274  -1.951502e-01  0.2342907499
#> 275  -1.762842e-01  0.0924153080
#> 276  -1.730787e-02  0.1172721110
#> 277  -2.373755e-01  0.1585447785
#> 278  -2.478939e-01  0.2124112066
#> 279  -2.300598e-01  0.2057392483
#> 280  -1.716009e-02 -0.0446253753
#> 281  -9.569362e-02 -0.0865459709
#> 282   6.573615e-02  0.0494658579
#> 283  -5.034960e-02  0.0291176356
#> 284  -2.333137e-01  0.1407648740
#> 285  -9.855279e-02 -0.0326882872
#> 286  -1.554898e-01  0.0942250240
#> 287  -2.342763e-01  0.1096607035
#> 288  -1.524005e-01  0.0844484043
#> 289  -1.116018e-01  0.0923486771
#> 290  -1.087993e-02  0.0275677128
#> 291  -1.818134e-01 -0.0611501311
#> 292  -1.274743e-02 -0.0645605054
#> 293  -7.614616e-02  0.0420101735
#> 294  -2.055317e-01  0.1299716953
#> 295  -1.026897e-01  0.1415597582
#> 296  -7.207006e-02 -0.0767498396
#> 297  -3.479493e-01 -0.0175644464
#> 298  -1.832122e-01 -0.1210336576
#> 299  -3.528550e-02 -0.0351659000
#> 300  -1.032203e-01  0.2399075594
#> 301  -5.632643e-02  0.0069128166
#> 302  -7.993105e-02  0.0688264167
#> 303  -1.948085e-01  0.0308037586
#> 304  -1.115601e-01  0.0380730605
#> 305  -4.579368e-02 -0.0495983518
#> 306  -6.835581e-02  0.0852047701
#> 307  -7.913449e-02  0.1406638857
#> 308  -1.116983e-02 -0.0144408108
#> 309  -6.334311e-02  0.2134563460
#> 310  -1.303868e-01  0.1019928088
#> 311  -6.033232e-02  0.0578810451
#> 312  -8.674044e-02  0.1302926196
#> 313  -3.056671e-01  0.0260311274
#> 314  -1.350720e-01  0.0374490654
#> 315  -1.401469e-01  0.0275542783
#> 316  -8.244154e-02 -0.0356271030
#> 317   8.346948e-04 -0.0481148200
#> 318  -1.771861e-01  0.1629321312
#> 319  -1.069426e-01  0.0052885851
#> 320  -1.008083e-01  0.0804360598
#> 321  -8.322772e-02  0.0964927760
#> 322  -1.831255e-01  0.1419657397
#> 323  -1.959952e-01  0.0264563413
#> 324  -1.668616e-01  0.0503420745
#> 325  -6.093631e-02 -0.0226594097
#> 326  -3.977148e-01  0.0759018560
#> 327  -1.953680e-01  0.2588364806
#> 328  -6.385662e-02  0.0440973024
#> 329  -1.364156e-01  0.1215158078
#> 330  -2.055778e-02 -0.0293378227
#> 331  -2.869698e-01  0.1054932445
#> 332  -9.224773e-02 -0.0008749234
#> 333  -1.332791e-01  0.0924846504
#> 334  -1.224491e-01  0.0886765667
#> 335  -1.957731e-01  0.1436942512
#> 336   2.715225e-02 -0.1158848158
#> 337  -3.187184e-02 -0.1181348253
#> 338  -1.522571e-01  0.0670319542
#> 339  -1.293143e-01  0.1659873720
#> 340   1.413555e-02 -0.0117148510
#> 341  -2.158170e-01 -0.1131686104
#> 342  -1.267813e-01  0.1069853734
#> 343  -1.501814e-01  0.0830717321
#> 344  -1.465363e-01  0.1087972560
#> 345  -1.002470e-01  0.0276522269
#> 346  -1.557136e-01  0.0640746192
#> 347  -2.406323e-01  0.0628958925
#> 348  -3.580235e-02  0.1768370105
#> 349  -1.849132e-01 -0.0047070181
#> 350  -5.003897e-02  0.0717044221
#> 351  -2.170455e-01  0.0385279922
#> 352  -2.102032e-01  0.0527710449
#> 353   4.047008e-02 -0.0048529428
#> 354   1.100064e-02 -0.0446451422
#> 355  -6.454117e-02 -0.0755087747
#> 356  -7.491333e-02  0.1693332601
#> 357  -1.191087e-01 -0.0156058110
#> 358   4.128573e-02  0.0919723543
#> 359  -1.094765e-01  0.1618721324
#> 360  -9.870086e-02  0.2698987833
#> 361  -1.868752e-01  0.1088307798
#> 362  -9.907976e-02 -0.0034228263
#> 363  -8.897991e-02  0.1152873703
#> 364  -8.713795e-02  0.0317591060
#> 365  -8.318526e-02  0.1308970529
#> 366  -5.136980e-02 -0.0402889226
#> 367  -7.693590e-02  0.0272601223
#> 368  -6.214910e-02 -0.0593725392
#> 369  -4.257475e-02  0.0854641175
#> 370   2.404349e-02  0.1402776961
#> 371  -1.048598e-01 -0.0759569676
#> 372  -5.878719e-02  0.0120801631
#> 373  -1.285758e-01  0.1032452399
#> 374   3.069231e-02  0.0191039370
#> 375  -1.080864e-01 -0.0577335994
#> 376  -9.577237e-02  0.0453532123
#> 377  -1.665847e-01  0.0973393352
#> 378  -3.676101e-02  0.0767884716
#> 379  -2.182126e-02  0.0051513987
#> 380  -2.245676e-02 -0.0285927270
#> 381  -2.055778e-02 -0.0293378227
#> 382  -1.588599e-01  0.2732897414
#> 383  -1.074979e-01  0.1074805476
#> 384  -7.287645e-02  0.0303392927
#> 385  -1.267813e-01  0.1069853734
#> 386  -6.044677e-02  0.1048436029
#> 387  -1.401469e-01  0.0275542783
#> 388  -1.042222e-01  0.1526533901
#> 389  -1.332791e-01  0.0924846504
#> 390  -6.455096e-03  0.1502193884
#> 391  -4.034587e-02 -0.0751668007
#> 392  -4.718985e-01  0.1034757405
#> 393  -2.042253e-01 -0.0843511477
#> 394  -2.807965e-02  0.1404231928
#> 395  -9.835100e-02  0.0384383555
#> 396  -1.907344e-01  0.0665775544
#> 397  -8.942007e-02  0.0489300260
#> 398  -6.196279e-02 -0.0152263582
#> 399   1.512009e-02 -0.0195498478
#> 400  -1.061286e-01 -0.0051202043
#> 401  -3.948276e-02  0.0442941803
#> 402  -5.632643e-02  0.0069128166
#> 403  -1.556414e-01  0.2174508285
#> 404  -1.762429e-01  0.0963800959
#> 405  -1.152572e-01  0.1671165195
#> 406  -1.350720e-01  0.0374490654
#> 407  -6.341689e-02 -0.1015228534
#> 408  -9.800110e-02  0.2938333105
#> 409  -3.210205e-01  0.2517653381
#> 410  -2.153664e-01 -0.1373551448
#> 411  -5.666863e-01 -0.0526863524
#> 412  -2.291593e-01  0.1899048918
#> 413  -1.730787e-02  0.1172721110
#> 414  -7.112538e-02  0.1614305841
#> 415  -2.576694e-01  0.1503266689
#> 416  -1.519561e-02  0.0921620649
#> 417  -5.835848e-02 -0.0251011718
#> 418   2.245158e-02 -0.0005839678
#> 419  -3.909781e-02  0.1378854073
#> 420  -5.881424e-02  0.0708742337
#> 421  -2.260971e-01  0.1753001149
#> 422  -1.783094e-02  0.0487568810
#> 423  -6.897465e-02  0.0583460186
#> 424  -1.113226e-01  0.1436362995
#> 425  -1.006122e-01  0.1188250077
#> 426  -1.818529e-01  0.2061114212
#> 427  -8.245107e-02  0.1644413477
#> 428  -9.174557e-02  0.0055130500
#> 429  -7.389314e-02  0.2224384646
#> 430  -9.907976e-02 -0.0034228263
#> 431  -1.500266e-01  0.0459800685
#> 432  -1.805990e-01  0.0512693731
#> 433  -4.180414e-02  0.0296260301
#> 434   8.050219e-04  0.0858014014
#> 435  -2.160460e-02 -0.0585370146
#> 436  -4.170039e-02  0.0662881126
#> 437  -1.879085e-01  0.0181028988
#> 438  -3.254527e-01  0.1130001512
#> 439  -3.747622e-01  0.1445672376
#> 440  -2.874718e-01  0.0103534660
#> 441  -1.845945e-01  0.0830532172
#> 442  -1.575917e-01  0.0805544170
#> 443  -1.193620e-01  0.2301641975
#> 444  -1.077914e-01  0.0704938794
#> 445   3.727304e-02 -0.0603777302
#> 446  -4.376206e-02 -0.0055091172
#> 447  -1.848125e-01  0.1986437762
#> 448  -1.831452e-01  0.2036167290
#> 449  -7.112538e-02  0.1614305841
#> 450  -1.013193e-02 -0.1833788433
#> 451  -5.268714e-02  0.0279751480
#> 452  -2.089852e-01  0.2255721965
#> 453  -1.951395e-01  0.1915900349
#> 454  -4.446662e-03 -0.0922174731
#> 455   1.512009e-02 -0.0195498478
#> 456  -1.070464e-01  0.1427157276
#> 457  -9.588727e-02 -0.0853432254
#> 458  -3.343724e-01 -0.0088701001
#> 459  -1.227485e-02  0.0737147439
#> 460   1.214281e-02 -0.0039320129
#> 461  -4.034587e-02 -0.0751668007
#> 462  -9.053262e-02  0.0692220265
#> 463  -2.503156e-02  0.0065389539
#> 464  -3.447222e-01  0.1390628577
#> 465  -1.642297e-01  0.0168158182
#> 466  -1.245007e-01 -0.0068846443
#> 467  -1.078426e-01  0.0103795424
#> 468  -2.309927e-01  0.0606564191
#> 469  -1.475763e-01  0.1902699007
#> 470  -2.280233e-01 -0.0209970157
#> 471   1.663915e-02  0.0101781942
#> 472  -4.529166e-02  0.0617049508
#> 473  -9.644352e-02 -0.0091731157
#> 474  -1.260579e-01  0.1541019492
#> 475  -1.318147e-01 -0.0602479299
#> 476   1.597769e-02  0.0456450571
#> 477  -1.242219e-01  0.0184420577
#> 478  -2.043373e-01  0.1870581748
#> 479   2.047825e-02 -0.0569016047
#> 480  -5.522543e-02 -0.0357041947
#> 481  -7.764920e-02  0.1047974491
#> 482  -3.092160e-02  0.0619974459
#> 483  -9.465164e-02  0.1220815956
#> 484  -3.737224e-02  0.1510499746
#> 485  -1.191087e-01 -0.0156058110
#> 486  -1.855101e-01  0.2050268881
#> 487  -2.000885e-01 -0.0486816843
#> 488  -1.530554e-01  0.1410048393
#> 489  -2.848266e-02 -0.0464449689
#> 490  -7.243076e-02  0.0988336650
#> 491   3.286210e-02 -0.0848768625
#> 492  -9.053262e-02  0.0692220265
#> 493   3.137737e-02 -0.2042525304
#> 494  -1.164699e-01  0.0174395527
#> 495  -1.399665e-01  0.1512500056
#> 496   2.351008e-02  0.0598033467
#> 497  -1.001106e-01  0.2651514151
#> 498   1.331114e-01 -0.2986873374
#> 499  -3.249574e-01  0.1223170856
#> 500  -1.656576e-01  0.0368047246
#> 501  -1.042956e-01  0.0489196235
#> 502  -2.713299e-01  0.1851133870
#> 503  -1.062543e-01  0.1118813970
#> 504  -8.177286e-02  0.0224744427
#> 505  -1.026897e-01  0.1415597582
#> 506  -3.324873e-01  0.1257655696
#> 507  -6.214910e-02 -0.0593725392
#> 508  -4.613908e-02  0.0824933263
#> 509  -2.291593e-01  0.1899048918
#> 510  -1.972432e-01  0.0843104836
#> 511   1.078267e-02  0.0212022819
#> 512  -1.483278e-01  0.1774924174
#> 513  -1.034358e-01  0.1702725528
#> 514  -2.342763e-01  0.1096607035
#> 515  -1.107049e-01  0.2291288428
#> 516  -1.250884e-01  0.0753892503
#> 517  -1.291187e-01  0.1456568972
#> 518  -1.738283e-01  0.0810392552
#> 519  -1.627641e-01  0.1661700968
#> 520  -1.929988e-01  0.1458240827
#> 521  -1.513896e-01  0.1227658618
#> 522  -1.348001e-01  0.2502100830
#> 523  -4.053728e-02  0.1855598873
#> 524  -3.796128e-02  0.0420350013
#> 525  -1.844001e-01  0.1366541999
#> 526  -1.848125e-01  0.1986437762
#> 527  -1.113448e-01  0.1680794786
#> 528  -1.226441e-01  0.0963186288
#> 529  -3.529213e-01  0.3267103687
#> 530  -6.173065e-02 -0.0467220825
#> 531  -2.721435e-01  0.0564942407
#> 532  -1.496972e-01  0.0902889629
#> 533  -1.984189e-01  0.2130087481
#> 534  -1.286596e-01 -0.0246469519
#> 535  -2.845447e-01  0.1853037323
#> 536  -8.494652e-02  0.0827757278
#> 537  -4.299456e-02  0.1613286178
#> 538   1.111983e-02 -0.0055086137
#> 539   8.442147e-02 -0.2452331870
#> 540  -1.762842e-01  0.0924153080
#> 541  -1.721398e-01  0.1768468425
#> 542  -1.234915e-02 -0.1389696522
#> 543  -2.007487e-01  0.0395632006
#> 544  -7.405934e-02 -0.0422613115
#> 545  -2.737290e-01 -0.0211793704
#> 546  -1.141572e-01 -0.0451058159
#> 547  -9.215994e-02  0.0788557030
#> 548  -1.006122e-01  0.1188250077
#> 549  -1.675423e-01  0.1058657580
#> 550   1.746439e-02 -0.0029825979
#> 551  -3.440211e-02  0.0441697468
#> 552  -1.668026e-01  0.1626805459
#> 553  -8.661423e-02  0.1687369862
#> 554  -1.870631e-01  0.0207425145
#> 555  -1.776878e-01  0.0880976090
#> 556  -1.185123e-01  0.1067771622
#> 557  -1.652678e-01  0.1855590316
#> 558  -6.191191e-02  0.0332493298
#> 559  -1.386433e-01  0.1094110583
#> 560   3.537475e-03 -0.1376184248
#> 561  -1.446487e-01 -0.0778977663
#> 562  -1.036570e-01  0.2383258797
#> 563  -1.250884e-01  0.0753892503
#> 564  -8.318526e-02  0.1308970529
#> 565  -1.346793e-01  0.0234972091
#> 566   2.176866e-02  0.1909034442
#> 567   6.291996e-02 -0.1285049293
#> 568  -1.411168e-01 -0.0608363615
#> 569  -1.660631e-01  0.1383677606
#> 570   1.994569e-02  0.0389772834
#> 571  -1.761362e-01  0.0621724534
#> 572  -9.144053e-02  0.0923429402
#> 573   4.825478e-02 -0.0747214918
#> 574   1.130691e-02 -0.0400163393
#> 575  -1.174568e-01  0.0297145952
#> 576  -1.242187e-01  0.1614200886
#> 577  -1.234151e-01  0.0467494985
#> 578  -9.940854e-02  0.0648002510
#> 579  -1.001106e-01  0.2651514151
#> 580  -1.684275e-01  0.1865930922
#> 581  -1.204183e-01 -0.0476003289
#> 582  -1.716009e-02 -0.0446253753
#> 583  -2.623616e-01 -0.1710680312
#> 584   4.475831e-02  0.0175501933
#> 585  -1.463569e-01  0.2665664879
#> 586  -3.858403e-02 -0.0247394342
#> 587  -3.559229e-02  0.1754795650
#> 588   9.723661e-02 -0.0970263044
#> 589  -9.388752e-02  0.2084066826
#> 590  -1.567513e-01  0.0125916984
#> 591  -2.036621e-01  0.1744875631
#> 592  -7.097211e-02  0.0083166718
#> 593   5.548451e-03 -0.1562003149
#> 594  -1.295999e-01  0.1600586184
#> 595   5.178905e-02 -0.1287501715
#> 596  -1.042222e-01  0.1526533901
#> 597  -1.203503e-01  0.1715626906
#> 598  -1.384566e-01  0.1980699262
#> 599  -1.386929e-01  0.0239619725
#> 600  -1.439911e-01  0.1068205608
#> 601  -2.234608e-01  0.0007683698
#> 602  -5.492274e-02 -0.0162452907
#> 603  -1.218826e-01 -0.0841916133
#> 604  -1.454307e-01  0.3174613974
#> 605  -1.953680e-01  0.2588364806
#> 606  -7.790998e-02 -0.0851572101
#> 607  -7.505941e-02  0.1512582625
#> 608  -6.198298e-02 -0.0921522403
#> 609  -1.135543e-01  0.0752922015
#> 610  -4.283587e-02 -0.0204712243
#> 611  -2.910003e-03  0.0619057540
#> 612  -1.831452e-01  0.2036167290
#> 613  -4.529166e-02  0.0617049508
#> 614  -8.202322e-02  0.0733732491
#> 615  -1.901662e-01  0.1857737961
#> 616  -1.060839e-01 -0.0386926837
#> 617  -1.366873e-01  0.1958553177
#> 618  -1.762262e-01  0.0196800385
#> 619  -5.584122e-02  0.0952025275
#> 620  -5.003897e-02  0.0717044221
#> 621  -3.210477e-01  0.2325813170
#> 622  -1.828946e-01  0.0840193024
#> 623  -3.858403e-02 -0.0247394342
#> 624  -1.629607e-01  0.1446403310
#> 625  -6.093631e-02 -0.0226594097
#> 626  -2.773134e-01  0.1960927150
#> 627  -1.440358e-01  0.1054436883
#> 628  -2.260971e-01  0.1753001149
#> 629  -2.108079e-01  0.1661930467
#> 630  -9.174557e-02  0.0055130500
#> 631  -1.762842e-01  0.0924153080
#> 632  -2.165554e-01  0.0589765454
#> 633  -1.098359e-01  0.0909079322
#> 634  -1.302271e-01  0.0617448701
#> 635  -1.552804e-01  0.1694862788
#> 636  -1.625726e-01  0.1976829896
#> 637  -2.951761e-01  0.2185640974
#> 638  -2.027493e-01 -0.0581659611
#> 639  -3.380134e-01  0.0794550155
#> 640  -3.189941e-01  0.1611386177
#> 641  -1.032641e-01  0.0121849875
#> 642  -3.901096e-01  0.2006012251
#> 643  -1.325228e-01  0.1770560804
#> 644  -2.176383e-01  0.0322033558
#> 645  -1.984189e-01  0.2130087481
#> 646  -2.360201e-01 -0.0918229865
#> 647  -1.514497e-01 -0.0157869556
#> 648  -1.530554e-01  0.1410048393
#> 649  -9.593453e-02 -0.0641647600
#> 650  -4.466485e-02  0.1274745809
#> 651  -2.466432e-02  0.1285378637
#> 652  -1.509229e-01  0.0718619851
#> 653   3.755349e-02  0.0668939333
#> 654  -1.370574e-01  0.0628734247
#> 655  -2.256357e-01  0.1682897784
#> 656  -2.664770e-01  0.1818007355
#> 657  -4.890281e-02  0.0156448726
#> 658  -2.036621e-01  0.1744875631
#> 659  -6.720519e-03 -0.0354288819
#> 660   4.782841e-02 -0.1350961916
#> 661  -1.949429e-01 -0.0572353064
#> 662  -4.264021e-02 -0.0162256011
#> 663  -6.763936e-02  0.0769698434
#> 664  -9.681499e-02  0.0147236412
#> 665  -3.529213e-01  0.3267103687
#> 666  -1.559507e-01  0.0664525802
#> 667  -8.377173e-02 -0.0199115488
#> 668  -8.322772e-02  0.0964927760
#> 669  -1.439449e-01  0.0315836936
#> 670  -1.462877e-01  0.0747729450
#> 671  -1.598301e-02 -0.0945468862
#> 672  -2.956749e-02 -0.0408045730
#> 673  -3.301554e-02 -0.1076348824
#> 674  -2.951761e-01  0.2185640974
#> 675   6.625488e-02 -0.1011156477
#> 676  -6.661888e-02  0.1976225321
#> 677  -7.707840e-02  0.1576991593
#> 678  -1.984189e-01  0.2130087481
#> 679  -9.046960e-02  0.0261855090
#> 680  -2.389666e-02  0.0499120244
#> 681   1.413555e-02 -0.0117148510
#> 682  -2.138266e-01  0.1230833066
#> 683  -1.440358e-01  0.1054436883
#> 684  -1.069945e-01  0.0960909237
#> 685  -9.240166e-02 -0.1088564713
#> 686  -3.392532e-03 -0.0386602161
#> 687  -4.088852e-02 -0.1329722391
#> 688  -9.307075e-02  0.1546311585
#> 689  -1.308381e-01 -0.0471492911
#> 690  -1.665847e-01  0.0973393352
#> 691  -1.386929e-01  0.0239619725
#> 692  -7.616672e-02  0.1970669238
#> 693  -1.521754e-01 -0.0394879217
#> 694  -1.932795e-02 -0.1570321344
#> 695  -9.447790e-02  0.2936418335
#> 696  -1.083905e-01  0.0352758042
#> 697  -1.069945e-01  0.0960909237
#> 698   1.019219e-02  0.0761928243
#> 699  -2.746965e-01  0.0852212309
#> 700  -1.722774e-01  0.2562322060
#> 701  -9.644352e-02 -0.0091731157
#> 702  -2.664770e-01  0.1818007355
#> 703  -5.685510e-02  0.0869381351
#> 704  -5.949125e-02  0.0449709934
#> 705  -1.403480e-01 -0.0752831430
#> 706   3.232093e-02 -0.0593744683
#> 707  -1.118262e-01 -0.1733008353
#> 708  -3.559229e-02  0.1754795650
#> 709  -2.815427e-01  0.1616371432
#> 710  -1.467392e-01  0.2937478581
#> 711  -2.036621e-01  0.1744875631
#> 712  -8.098511e-02  0.0273229673
#> 713  -9.428883e-02  0.1075578311
#> 714  -1.777571e-01  0.1186466593
#> 715   4.185820e-02 -0.0110896451
#> 716   1.208892e-02  0.0380079141
#> 717  -6.499905e-02  0.1203754745
#> 718  -9.830192e-02  0.1226098894
#> 719  -1.660631e-01  0.1383677606
#> 720   2.253814e-03  0.0951951447
#> 721   8.740523e-03 -0.0177419688
#> 722  -2.860228e-02  0.0254246182
#> 723  -1.944703e-01 -0.0226416036
#> 724  -1.845945e-01  0.0830532172
#> 725  -2.089446e-01  0.0992568301
#> 726  -5.158636e-02  0.1575135479
#> 727  -1.722774e-01  0.2562322060
#> 728   5.164229e-02  0.0648131792
#> 729  -1.409885e-01 -0.1201583806
#> 730  -1.595261e-01  0.1369800853
#> 731  -4.942441e-02 -0.0409986254
#> 732  -6.760797e-02  0.1534125184
#> 733  -3.983861e-02 -0.0615055366
#> 734  -1.664568e-01  0.0453803469
#> 735  -9.746554e-02 -0.0421142263
#> 736  -1.911194e-01 -0.0027319477
#> 737  -1.652678e-01  0.1855590316
#> 738  -1.277587e-01  0.2144880670
#> 739  -1.226441e-01  0.0963186288
#> 740  -1.274444e-01 -0.0082712692
#> 741  -8.894230e-02  0.1321541934
#> 742  -1.552804e-01  0.1694862788
#> 743  -7.020654e-02  0.1155477950
#> 744  -2.934328e-01  0.1356854364
#> 745  -7.112538e-02  0.1614305841
#> 746  -1.164937e-01  0.1244682874
#> 747  -2.153431e-01  0.1114858918
#> 748  -3.858403e-02 -0.0247394342
#> 749   1.663915e-02  0.0101781942
#> 750  -6.661888e-02  0.1976225321
#> 751  -3.322372e-02  0.0840242481
#> 752  -5.345000e-02 -0.0179501827
#> 753  -3.528550e-02 -0.0351659000
#> 754  -1.242606e-01  0.1164847896
#> 755  -3.449355e-02  0.1788401352
#> 756  -2.256357e-01  0.1682897784
#> 757  -1.721398e-01  0.1768468425
#> 758  -6.835101e-02  0.0066181315
#> 759  -3.589136e-01  0.1797428821
#> 760  -1.460700e-02 -0.0651435407
#> 761  -4.569823e-02  0.0386762319
#> 762  -6.284034e-02 -0.0209695330
#> 763  -3.210205e-01  0.2517653381
#> 764  -3.729236e-02  0.0117634917
#> 765  -1.553461e-01  0.0907065449
#> 766  -1.100660e-01 -0.0243893323
#> 767  -1.193620e-01  0.2301641975
#> 768   1.248144e-02 -0.1354458076
#> 769  -2.479511e-01  0.2436397475
#> 770  -2.848266e-02 -0.0464449689
#> 771  -1.496972e-01  0.0902889629
#> 772  -6.973619e-02 -0.0110953615
#> 773  -1.098939e-02 -0.0140021769
#> 774  -1.199788e-01 -0.0041969827
#> 775  -1.760213e-01  0.1395226901
#> 776  -1.018990e-01  0.1323924576
#> 777  -3.578370e-02  0.0895748365
#> 778  -9.071455e-02  0.1668102801
#> 779  -4.106749e-02  0.0731194619
#> 780  -2.604972e-01  0.1951633734
#> 781  -1.556414e-01  0.2174508285
#> 782  -1.588599e-01  0.2732897414
#> 783   4.726886e-02 -0.1195349901
#> 784  -1.032641e-01  0.0121849875
#> 785  -1.385374e-01  0.0352335582
#> 786  -4.880996e-02  0.1531877480
#> 787  -1.902331e-01 -0.0179757441
#> 788  -7.083852e-02  0.1968137513
#> 789  -2.070920e-01  0.1963216589
#> 790  -7.299967e-02  0.0743247372
#> 791  -5.832718e-02 -0.0695731639
#> 792  -2.977048e-01 -0.0074568573
#> 793  -1.015458e-01  0.0904164595
#> 794  -1.293143e-01  0.1659873720
#> 795  -2.038526e-01  0.1756274632
#> 796  -9.071140e-02 -0.1066216179
#> 797  -1.074979e-01  0.1074805476
#> 798  -1.646290e-01  0.0268908800
#> 799  -4.069249e-01  0.0493438124
#> 800  -3.829152e-01 -0.0070273262
#> 801  -7.928834e-03  0.0164442376
#> 802  -1.652678e-01  0.1855590316
#> 803  -8.758365e-02  0.1699467527
#> 804  -2.268376e-02  0.1187577400
#> 805  -2.055020e-01  0.0734363929
#> 806  -8.736510e-02  0.1123461410
#> 807  -2.648643e-01  0.1350063286
#> 808  -1.390479e-01 -0.0971120986
#> 809  -1.322208e-01 -0.1897728084
#> 810  -7.620682e-02 -0.0228890145
#> 811  -1.505202e-01  0.1468555292
#> 812  -6.254599e-02  0.0998981019
#> 813  -4.795412e-02  0.0135307905
#> 814   3.596077e-02  0.0464082957
#> 815  -2.449660e-01  0.0253604945
#> 816  -1.120031e-01  0.1620390927
#> 817  -1.089476e-01  0.0660913589
#> 818   5.566001e-02 -0.1886751464
#> 819  -1.267984e-02  0.0407856169
#> 820  -1.194987e-01  0.0976476768
#> 821  -1.965746e-01  0.0254863236
#> 822   4.656160e-02 -0.1957213928
#> 823  -1.524010e-01  0.0953902847
#> 824  -7.132024e-02  0.0798131374
#> 825  -8.245107e-02  0.1644413477
#> 826  -7.479825e-02  0.0560197202
#> 827  -1.777571e-01  0.1186466593
#> 828  -8.194357e-02 -0.0219107881
#> 829  -3.929207e-02 -0.0238962810
#> 830   7.381768e-04 -0.0355860680
#> 831  -1.550448e-01  0.2367887493
#> 832  -1.588599e-01  0.2732897414
#> 833  -2.565754e-02 -0.1431956791
#> 834  -1.675423e-01  0.1058657580
#> 835  -4.034587e-02 -0.0751668007
#> 836  -3.197864e-01  0.0537889543
#> 837  -1.091473e-01 -0.1024396724
#> 838  -3.745516e-02  0.1050665203
#> 839  -2.627542e-01  0.2176755993
#> 840  -1.913115e-01 -0.0353221730
#> 841  -4.034587e-02 -0.0751668007
#> 842  -2.125526e-01  0.1075113537
#> 843  -4.629493e-02  0.1044578365
#> 844   2.502530e-02 -0.0305606623
#> 845  -1.466029e-01 -0.1389330770
#> 846  -2.845703e-01  0.1237955808
#> 847  -1.777571e-01  0.1186466593
#> 848  -9.023575e-02  0.0422219152
#> 849   5.547134e-05 -0.1760173950
#> 850  -5.881424e-02  0.0708742337
#> 851  -7.877467e-02 -0.1905062088
#> 852  -1.054782e-01 -0.0788435739
#> 853  -4.579368e-02 -0.0495983518
#> 854  -8.719120e-02  0.0401257032
#> 855  -2.486509e-01  0.1147702610
#> 856  -9.598758e-02  0.2131723029
#> 857  -1.439911e-01  0.1068205608
#> 858  -1.962866e-01  0.1407040543
#> 859  -1.483786e-01  0.3424241478
#> 860  -2.342763e-01  0.1096607035
#> 861  -9.654390e-03 -0.0898815771
#> 862  -1.615683e-01 -0.0578046791
#> 863  -9.155075e-02  0.1609524533
#> 864  -1.508409e-01  0.1610519969
#> 865  -1.178846e-01  0.0295329432
#> 866  -2.273897e-01 -0.0163808350
#> 867  -1.089485e-01  0.1841881054
#> 868  -1.185123e-01  0.1067771622
#> 869  -1.791499e-01  0.0362100709
#> 870  -2.851019e-01  0.2592936254
#> 871  -9.969644e-02  0.1558104617
#> 872  -3.345705e-02  0.1242757438
#> 873  -2.059588e-02  0.0611581286
#> 874  -2.043373e-01  0.1870581748
#> 875  -5.811011e-02 -0.0184657384
#> 876  -4.343324e-02 -0.0798533955
#> 877  -5.942544e-02 -0.0317467202
#> 878  -1.559032e-01 -0.0432805194
#> 879  -1.000767e-01  0.0327053682
#> 880   7.900094e-03 -0.0942175981
#> 881  -7.876909e-02  0.1962728560
#> 882  -2.242928e-01 -0.0366021950
#> 883  -9.741528e-02  0.1082068841
#> 884  -7.985103e-02 -0.0714996856
#> 885   9.290579e-02  0.0379610244
#> 886  -1.002679e-01 -0.0015774417
#> 887  -2.870563e-01  0.1972254470
#> 888  -8.657907e-02  0.1185238544
#> 889   5.272599e-02  0.0198076914
#> 890  -7.302706e-02  0.0769983053
#> 891  -1.595261e-01  0.1369800853
#> 892  -8.861407e-02  0.0363764574
#> 893  -2.526727e-01  0.1099482699
#> 894  -2.193777e-01  0.0713814966
#> 895  -2.038048e-02  0.0347276951
#> 896  -1.732210e-01 -0.0417718659
#> 897  -1.921184e-02 -0.0288394039
#> 898  -3.517476e-02  0.0293173371
#> 899  -1.054337e-01 -0.0190541561
#> 900  -3.195449e-02  0.1567536362
#> 901  -1.462877e-01  0.0747729450
#> 902   4.488887e-02 -0.0940081777
#> 903  -2.806299e-01  0.1723455923
#> 904  -1.187206e-01 -0.1005106412
#> 905  -1.107049e-01  0.2291288428
#> 906  -1.957723e-02  0.0062149203
#> 907  -2.055317e-01  0.1299716953
#> 908  -1.783178e-01  0.1463241502
#> 909  -1.550448e-01  0.2367887493
#> 910  -1.953680e-01  0.2588364806
#> 911  -7.816273e-02  0.0577941945
#> 912  -1.939733e-01  0.0928203813
#> 913   8.103848e-02 -0.1185435542
#> 914  -6.706852e-02 -0.0949381149
#> 915  -7.668249e-02  0.1612701770
#> 916  -2.664770e-01  0.1818007355
#> 917  -1.471801e-01  0.1372885573
#> 918  -7.940679e-02 -0.1071325445
#> 919  -1.316443e-01  0.0517158681
#> 920  -8.713795e-02  0.0317591060
#> 921  -1.665635e-01  0.1672505997
#> 922  -1.403480e-01 -0.0752831430
#> 923  -4.686548e-02 -0.0024317154
#> 924  -1.120396e-01  0.1036867975
#> 925  -1.718533e-01  0.1563341210
#> 926  -1.441187e-01  0.0902527980
#> 927  -9.599202e-02  0.1189928728
#> 928  -8.244154e-02 -0.0356271030
#> 929  -3.745516e-02  0.1050665203
#> 930  -4.160511e-02  0.0947360027
#> 931  -5.685510e-02  0.0869381351
#> 932  -2.675582e-02 -0.0453650769
#> 933  -1.054337e-01 -0.0190541561
#> 934  -1.493626e-01 -0.1042337314
#> 935  -4.967802e-02 -0.0849270487
#> 936  -3.065806e-02  0.0286401987
#> 937  -1.744584e-01  0.1193552617
#> 938  -1.895291e-01 -0.0372266212
#> 939  -6.876438e-02 -0.0173474234
#> 940  -1.599733e-01 -0.0288561205
#> 941  -1.665847e-01  0.0973393352
#> 942  -1.185109e-01  0.0037671369
#> 943  -1.235853e-02 -0.0025119405
#> 944  -1.215784e-01  0.0974905139
#> 945  -2.010017e-01 -0.0126747202
#> 946  -1.274743e-02 -0.0645605054
#> 947  -2.734344e-01  0.0829978443
#> 948  -8.555595e-02  0.0436203707
#> 949  -4.025948e-01  0.0480426545
#> 950  -1.741719e-01  0.0899977416
#> 951  -1.776878e-01  0.0880976090
#> 952  -1.648811e-01 -0.0244717825
#> 953  -6.864848e-02  0.0318516421
#> 954  -2.541980e-01  0.3345270969
#> 955  -2.382062e-01 -0.0216037151
#> 956  -5.479320e-02  0.1352022666
#> 957   2.406232e-02 -0.0247764659
#> 958  -2.195259e-01  0.1398852389
#> 959  -1.804588e-01  0.0191785871
#> 960  -1.026897e-01  0.1415597582
#> 961  -1.901611e-01  0.1563008866
#> 962  -1.131556e-01  0.0312212670
#> 963   5.780911e-02 -0.1886929505
#> 964  -1.776677e-01  0.0227107016
#> 965  -2.059779e-01  0.1163983327
#> 966  -2.604071e-01 -0.0680016485
#> 967  -3.065110e-02  0.0081943018
#> 968   5.263855e-02 -0.0345689339
#> 969  -1.941844e-01  0.2352837472
#> 970  -2.275250e-01  0.0465986906
#> 971  -1.993730e-01  0.1884590691
#> 972  -7.876909e-02  0.1962728560
#> 973  -1.239044e-01 -0.0358272718
#> 974  -2.624541e-02 -0.0963650065
#> 975  -7.097211e-02  0.0083166718
#> 976  -1.112728e-01  0.1418815408
#> 977  -8.668387e-02 -0.0111493562
#> 978  -1.308318e-01 -0.1097296193
#> 979  -1.107049e-01  0.2291288428
#> 980  -2.391257e-01  0.0574170487
#> 981  -2.270787e-01 -0.0621433855
#> 982  -1.055580e-02 -0.0064224938
#> 983   4.863578e-02 -0.0035997567
#> 984  -8.245107e-02  0.1644413477
#> 985  -9.871109e-02  0.0877565632
#> 986  -1.442860e-01  0.0891245768
#> 987   1.541333e-02 -0.1884390792
#> 988  -6.842080e-02  0.1923957258
#> 989  -1.692918e-01  0.1495218477
#> 990  -3.821733e-02  0.0955446529
#> 991  -2.278942e-01  0.0196620081
#> 992  -1.858023e-01  0.0726518422
#> 993  -1.483278e-01  0.1774924174
#> 994  -6.659518e-02  0.0380157427
#> 995  -7.764920e-02  0.1047974491
#> 996  -1.704625e-01  0.1610499270
#> 997  -1.056023e-02  0.1102072874
#> 998  -6.894423e-02  0.1065730124
#> 999  -1.215784e-01  0.0974905139
#> 1000 -1.036570e-01  0.2383258797
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.153   0.0169 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.0663 -0.0893 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.100   0.0327 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0820  0.0734 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.190   0.156  
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0453  0.0617 
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.146   0.267  
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.305   0.116  
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0689 -0.0986 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.0684  0.00662
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.214  19.8 0.041  0.12 83.6  1656.
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.214  19.8 0.041  0.12  6.20  123.
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.233  17.8 0.041  0.12 91.7  1632.
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.233  17.8 0.041  0.12  5.64  100.
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.225  20.1 0.041  0.12 81.6  1642.
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.225  20.1 0.041  0.12  5.41  109.
#> # … with 1,994 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>