Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object

A posologyr list, created by the posologyr function.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01,n_sample=1e4,n_resample=1e3)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1     0.22271891  0.160921516  0.533092958
#> 2     0.33441230 -0.792574549 -0.025187028
#> 3     0.57774734 -0.422713712  0.106462061
#> 4     0.57580879 -0.629296662 -0.072648075
#> 5     0.59099320 -0.509730049  0.401155411
#> 6     0.76209833 -0.219890349  0.504488770
#> 7     0.68801949 -0.235439893  0.156779989
#> 8     0.19719802 -0.447782520  0.191080947
#> 9     0.77740142 -0.478899716 -0.132217813
#> 10    0.43244426 -0.531772863  0.183229522
#> 11    0.78403924 -0.176646206  0.300057300
#> 12    0.47025283 -0.365488955 -0.228626652
#> 13    0.33441230 -0.792574549 -0.025187028
#> 14    0.59304735 -0.243286872  0.167643687
#> 15    0.31016826 -0.397653498  0.221229070
#> 16    0.57195175 -0.177782486  0.951797712
#> 17    0.54518657 -0.380966345  0.036069184
#> 18    0.47600984 -0.649861550  0.120920415
#> 19    0.22593314 -1.028179844 -0.322215004
#> 20    0.39780777 -0.791037371 -0.162024458
#> 21    0.36899633 -0.571135845  0.443488626
#> 22    0.90060377  0.113823871  0.886071899
#> 23    0.63451362 -0.157105153  0.515831966
#> 24    0.47162121 -0.821644887 -0.325428335
#> 25    0.56973624 -0.473123776  0.071389363
#> 26    0.75337489 -0.134004424  0.658113525
#> 27    0.43161435 -0.710027158 -0.062169150
#> 28    0.43438710 -0.695618480 -0.038173630
#> 29    0.69697848  0.016350757  0.788096150
#> 30    0.71784457 -0.250012730  0.319825369
#> 31    0.60956851 -0.415292316  0.192131074
#> 32    0.43785127 -1.001398957 -0.863318535
#> 33    0.74353332 -0.280496475  0.008893078
#> 34    0.53531482 -0.714146172 -0.146778964
#> 35    0.47011792 -0.645162044 -0.025825770
#> 36    0.39448044 -0.195420995  1.167242830
#> 37    0.10035179 -0.408852030  0.332733063
#> 38    0.09212035 -0.521954699 -0.086147862
#> 39    0.63080615 -0.433882056 -0.066951295
#> 40    0.05712371 -0.271516248  0.542083162
#> 41    0.77335089 -0.422619967 -0.077239781
#> 42    0.03736604 -0.076119570  0.802537203
#> 43    0.63668531 -0.446139960 -0.099075024
#> 44    0.50262843 -0.550564164  0.194935102
#> 45    0.69480817 -0.371830367 -0.116372563
#> 46    0.63476976 -0.461787433 -0.078905679
#> 47    0.30667815 -0.259829281  0.541199976
#> 48    0.59099320 -0.509730049  0.401155411
#> 49    0.47986990 -0.760386207 -0.366083390
#> 50    0.26337629 -0.031733914  0.701192123
#> 51    0.61145958 -0.657806745 -0.187504103
#> 52    0.55071236 -0.124750546  0.165718122
#> 53    0.40790793 -0.290384990  0.444105117
#> 54    0.65289658 -0.324650286  0.280673258
#> 55    0.59236409 -0.188568205  0.212891760
#> 56    0.57774734 -0.422713712  0.106462061
#> 57    0.22613189 -0.386718629 -0.033007632
#> 58    0.64359955 -0.498224057  0.380419922
#> 59    0.68430009 -0.268334493  0.354966165
#> 60    0.68936903 -0.006487272  0.340493277
#> 61    0.57195175 -0.177782486  0.951797712
#> 62    0.56659226 -0.349077435  0.098837947
#> 63    0.64809788 -0.556890647 -0.218689257
#> 64    0.63549033 -0.389321097 -0.272556099
#> 65    0.65811015  0.096719696  0.748908651
#> 66    0.42274748 -0.026185452  0.549303743
#> 67    0.33345352 -0.265313541  0.390607939
#> 68    0.45479324 -0.579153313  0.184361575
#> 69    0.27458938 -0.793649344 -0.078308256
#> 70    0.55675322 -0.502790213  0.288888253
#> 71    0.67233770 -0.186577851  0.713714035
#> 72    0.71930818 -0.215642105  0.258831014
#> 73    0.54157047 -0.567557604  0.231729723
#> 74    0.44008850  0.005851398  0.804104063
#> 75    0.43487975 -0.627726746  0.075577680
#> 76    0.72621790 -0.317610372 -0.122924362
#> 77    0.13959323 -0.596338057  0.270022896
#> 78    0.29991156 -0.927221653 -0.464468104
#> 79    0.58020169 -0.603806814  0.087008614
#> 80    0.43690209 -0.209391228  0.262314390
#> 81    0.46601257 -0.571447886 -0.211384708
#> 82    0.06249888  0.326053512  0.760135462
#> 83    0.12338319 -0.235113562  0.597606681
#> 84    0.47011792 -0.645162044 -0.025825770
#> 85    0.39262313 -0.674586923  0.400223008
#> 86    0.77711143 -0.343425841  0.386221758
#> 87    0.39012890 -0.328438415  0.006599815
#> 88    0.47162121 -0.821644887 -0.325428335
#> 89    0.81293397 -0.222745732  0.545014504
#> 90    0.55675322 -0.502790213  0.288888253
#> 91    0.76209833 -0.219890349  0.504488770
#> 92    0.57774734 -0.422713712  0.106462061
#> 93    0.52225016 -0.425304427 -0.110638439
#> 94    0.53222987 -0.593493588  0.236077143
#> 95    0.79657498 -0.306695742  0.034228062
#> 96    0.39196750 -0.790654013  0.344915114
#> 97    0.37334403 -1.075498481 -0.358215260
#> 98    0.55882039 -0.531616246  0.608807144
#> 99    0.46394825 -0.929746855 -0.454318341
#> 100   0.52451482 -0.528343337  0.179548280
#> 101   0.39514071 -0.451573396  0.096597599
#> 102   0.67741777  0.105706286  0.523212763
#> 103   0.42968639 -0.311785502  0.075772463
#> 104   0.47419507 -0.495760400 -0.113534219
#> 105   0.58032901 -0.624467554  0.015084688
#> 106   0.49936330 -0.237292402  0.301252020
#> 107   0.46954969 -0.258241155  0.353266519
#> 108   0.68430009 -0.268334493  0.354966165
#> 109   0.58020169 -0.603806814  0.087008614
#> 110   0.63921282 -0.543571179 -0.101302897
#> 111   0.66071976 -0.417526332  0.268656562
#> 112   0.58457041 -0.362562283  0.444015055
#> 113   0.55187186 -0.488333657 -0.211425066
#> 114   0.29230382 -0.934545772  0.121229790
#> 115   0.34355223 -0.582343690  0.096233562
#> 116   0.37715840 -0.190812315  0.533973479
#> 117   0.14094738 -0.847194113 -0.551760073
#> 118   0.49447853 -0.341027962  0.464029232
#> 119   0.64022307 -0.354636472  0.160793954
#> 120   0.20593736 -0.898865294  0.158528942
#> 121   1.03127091  0.187833806  1.392170483
#> 122   0.14606178 -0.173866177  0.342830851
#> 123   0.55302700 -0.637914160  0.079587366
#> 124   0.43785127 -1.001398957 -0.863318535
#> 125   0.54544736 -0.120184943  0.654104548
#> 126   0.65796870 -0.608261375 -0.181451503
#> 127   0.55861273 -0.665663766 -0.005716252
#> 128   0.59264304 -0.643020069 -0.077283486
#> 129   0.64483094 -0.335371426  0.102272825
#> 130   0.48920040 -0.672646062  0.168923703
#> 131   0.46259920 -0.617846549  0.197179188
#> 132   0.48853262 -0.550175315  0.121996031
#> 133   0.50894773 -0.347561805  0.363000532
#> 134   0.56551248 -0.402378787 -0.048887896
#> 135  -0.27689884 -0.275839190  0.162162731
#> 136   0.62170516 -0.358789161 -0.450831239
#> 137   0.81168578 -0.165327520  0.213256683
#> 138   0.79022485  0.254688266  0.509186545
#> 139   0.29318907 -0.878733272 -0.318699698
#> 140   0.56256112 -0.438349054  0.436749854
#> 141   0.38360952 -0.406660483  0.054588472
#> 142   0.55696157 -0.345743743  0.498074128
#> 143   0.58020169 -0.603806814  0.087008614
#> 144   0.68801949 -0.235439893  0.156779989
#> 145   0.54518657 -0.380966345  0.036069184
#> 146   0.90060377  0.113823871  0.886071899
#> 147   0.71284966 -0.184603944  0.313687884
#> 148   0.75907924  0.021543333  0.908443128
#> 149   0.31594181 -0.773337221  0.017692294
#> 150   0.37736448 -0.670266721 -0.131909079
#> 151   0.51682515 -0.111747769  1.033771044
#> 152   0.81293397 -0.222745732  0.545014504
#> 153   0.43032315 -0.251860607  0.973898527
#> 154   0.79503661  0.184800797  0.739140891
#> 155   0.45433358 -0.300979352 -0.271848504
#> 156   0.67233770 -0.186577851  0.713714035
#> 157   0.76209833 -0.219890349  0.504488770
#> 158   0.57430835 -0.112154278  0.420404962
#> 159   0.76097435 -0.316887821  0.196729799
#> 160   0.91451667  0.068416245  0.658184052
#> 161   0.38434042 -0.158741093  0.404268763
#> 162   0.65811015  0.096719696  0.748908651
#> 163   0.71758764 -0.296279457  0.130185287
#> 164   0.57774734 -0.422713712  0.106462061
#> 165   0.55270865 -0.199445792  0.379444169
#> 166   0.47956197  0.003473349  0.585617792
#> 167   0.75337489 -0.134004424  0.658113525
#> 168   0.47407350 -0.067920214  0.499850945
#> 169   0.54544736 -0.120184943  0.654104548
#> 170   0.44206403 -0.117694333  0.594801792
#> 171   0.37736448 -0.670266721 -0.131909079
#> 172   0.87850661 -0.186234945  0.077049555
#> 173   0.56119848 -0.419359563  0.636055664
#> 174   0.54975464 -0.387957052  0.107710101
#> 175   0.11520465 -0.155131610  0.060030359
#> 176   0.58020169 -0.603806814  0.087008614
#> 177   0.16152271 -0.045188017  0.636938434
#> 178   0.43785762 -0.469677685  1.260898606
#> 179   0.66952258 -0.124155918  0.452703391
#> 180   0.57267802 -0.476955404  0.286743738
#> 181   0.59264304 -0.643020069 -0.077283486
#> 182   0.64725795 -0.223764263  0.670036895
#> 183   0.38121549 -0.167557329  0.585892993
#> 184   0.71284966 -0.184603944  0.313687884
#> 185   0.37648950 -0.821084808 -0.240689711
#> 186   0.52610511 -0.410377516 -0.120994350
#> 187   0.56659226 -0.349077435  0.098837947
#> 188   0.47276563 -0.314104286  0.036047329
#> 189   0.64483094 -0.335371426  0.102272825
#> 190   0.48418803 -0.386735983  0.083369047
#> 191   0.63451362 -0.157105153  0.515831966
#> 192   0.57330107 -0.361481788  0.493499216
#> 193   0.67634332 -0.213150598  0.497780568
#> 194   0.57840668 -0.532500719  0.128807108
#> 195   0.37134361 -0.368870465  0.542339618
#> 196   0.47986990 -0.760386207 -0.366083390
#> 197   0.35676441 -0.598157845 -0.199685798
#> 198   0.47698705 -0.517100924  0.071624256
#> 199   0.25032023 -0.331034685  0.042813305
#> 200   0.57267802 -0.476955404  0.286743738
#> 201   0.71512579 -0.392608569  0.041057824
#> 202   0.55187186 -0.488333657 -0.211425066
#> 203   0.47244770 -0.474954112  0.351556010
#> 204   0.71834459 -0.293863481  0.542048029
#> 205   0.35329713 -1.028439732  0.048156325
#> 206   0.75907924  0.021543333  0.908443128
#> 207   0.91289416  0.133311545  0.940469971
#> 208   0.22265247 -0.717846702 -0.087341000
#> 209   0.61515370 -0.132078301  1.051741453
#> 210   0.24066087 -0.266038284  0.405203368
#> 211   0.38625234 -0.173989915  0.217121673
#> 212   0.42173401 -0.743606298 -0.543204325
#> 213   0.62011758 -0.608478964  0.472211205
#> 214   0.43259247 -0.667096819 -0.089974522
#> 215   0.63972196 -0.152214377  0.789514749
#> 216   0.68327232 -0.140921499  0.404592562
#> 217   0.64483094 -0.335371426  0.102272825
#> 218   0.61515370 -0.132078301  1.051741453
#> 219   0.46954969 -0.258241155  0.353266519
#> 220   0.54544736 -0.120184943  0.654104548
#> 221   0.80057541 -0.058131248  0.209457426
#> 222   0.67634332 -0.213150598  0.497780568
#> 223   0.52610511 -0.410377516 -0.120994350
#> 224   0.65289658 -0.324650286  0.280673258
#> 225   0.63972196 -0.152214377  0.789514749
#> 226   0.39625496 -0.942123209 -0.311252784
#> 227   0.25163530 -0.993335506 -0.105940185
#> 228   0.47011792 -0.645162044 -0.025825770
#> 229   0.55644269 -0.155697224  0.295934079
#> 230   0.59996541 -0.229701037  0.491160197
#> 231   0.56659226 -0.349077435  0.098837947
#> 232   0.65410430 -0.429468728 -0.134088466
#> 233   0.52451482 -0.528343337  0.179548280
#> 234   0.71512579 -0.392608569  0.041057824
#> 235   0.27590400 -0.743225452 -0.454430364
#> 236   0.54024124  0.066003374  0.114442031
#> 237   0.35471474 -0.468109410  0.267712101
#> 238   0.43790109 -0.382331300 -0.114096056
#> 239   0.46601257 -0.571447886 -0.211384708
#> 240   0.58283365 -0.514847651 -0.040043751
#> 241   0.49710198 -0.258009118  0.719389142
#> 242   0.27316409 -0.525810933  0.066879931
#> 243   0.79657498 -0.306695742  0.034228062
#> 244   0.54458731 -0.680380144  0.662311455
#> 245   0.57268800 -0.430149340 -0.208548251
#> 246   0.28618227 -0.565749214  0.088221902
#> 247   0.71758764 -0.296279457  0.130185287
#> 248   0.46958169 -0.956443428 -0.455650474
#> 249   0.49710198 -0.258009118  0.719389142
#> 250   0.44327092 -0.817196840 -0.020900759
#> 251   0.77330919 -0.302877949 -0.012989957
#> 252  -0.03039750 -0.044236675  0.443676291
#> 253   0.45675744 -0.916232859 -0.270910641
#> 254   0.68327232 -0.140921499  0.404592562
#> 255   0.36899633 -0.571135845  0.443488626
#> 256   0.64725795 -0.223764263  0.670036895
#> 257   0.47698705 -0.517100924  0.071624256
#> 258   0.57774734 -0.422713712  0.106462061
#> 259   0.46213953 -0.305588982  0.134858945
#> 260   0.77330919 -0.302877949 -0.012989957
#> 261   0.73864601 -0.188752016  0.015631478
#> 262   0.50652403 -0.318427812  0.664705298
#> 263   0.65913213 -0.596382395  0.147896098
#> 264   0.48920040 -0.672646062  0.168923703
#> 265   0.83770446 -0.267650634  0.152742196
#> 266   0.59010962 -0.089581912  0.481940997
#> 267   0.46406265 -0.020306572  0.764496352
#> 268   0.52991534 -0.256713528  0.481578355
#> 269   0.66071976 -0.417526332  0.268656562
#> 270   0.35471474 -0.468109410  0.267712101
#> 271   0.36441755 -0.313102195  0.634437746
#> 272   0.55675322 -0.502790213  0.288888253
#> 273   0.16047620 -0.403641084 -0.129576775
#> 274   0.51117679 -0.465470310 -0.042061493
#> 275   0.28030781 -0.450134039  0.332394957
#> 276   0.11939191 -0.357101012  0.035343282
#> 277   0.47600984 -0.649861550  0.120920415
#> 278   0.30650911 -0.749722944  0.044791843
#> 279   0.60956851 -0.415292316  0.192131074
#> 280   0.28618227 -0.565749214  0.088221902
#> 281   0.30650911 -0.749722944  0.044791843
#> 282   0.47602416 -0.417497806  0.440876680
#> 283   0.34779916 -0.064904691  0.555874342
#> 284   0.56651428  0.048258526  0.409666639
#> 285  -0.06817215 -0.612169400  0.435160426
#> 286   0.58629667 -0.245281709  0.626989164
#> 287   0.57195175 -0.177782486  0.951797712
#> 288   0.59099320 -0.509730049  0.401155411
#> 289   0.47698705 -0.517100924  0.071624256
#> 290   0.66111906 -0.483695929 -0.099917250
#> 291   0.63476976 -0.461787433 -0.078905679
#> 292   0.49216429 -0.399277612  0.161909658
#> 293   0.75109916 -0.379168608 -0.009246565
#> 294   0.64022307 -0.354636472  0.160793954
#> 295   0.69615847 -0.155441330  0.312666571
#> 296   0.46259920 -0.617846549  0.197179188
#> 297   0.30427038 -0.311808473  0.134560817
#> 298   0.47011792 -0.645162044 -0.025825770
#> 299   0.87555058  0.096843987  0.562812913
#> 300   0.46394825 -0.929746855 -0.454318341
#> 301   0.68430253 -0.238218647  0.922582400
#> 302   0.33555056 -0.103621727 -0.113008779
#> 303   0.62012313 -0.355316823  0.074574922
#> 304   0.66071976 -0.417526332  0.268656562
#> 305   0.76097435 -0.316887821  0.196729799
#> 306   0.71317996  0.083021280  1.115230301
#> 307   0.68430253 -0.238218647  0.922582400
#> 308   0.76209833 -0.219890349  0.504488770
#> 309   0.39181865 -0.291717452  0.094051444
#> 310   0.36899633 -0.571135845  0.443488626
#> 311   0.51658366 -0.391286009  0.136916131
#> 312   0.72232382 -0.167543169  0.055345669
#> 313   0.40354890 -0.258650644  0.214308319
#> 314   0.57430835 -0.112154278  0.420404962
#> 315   0.50262843 -0.550564164  0.194935102
#> 316   0.67233770 -0.186577851  0.713714035
#> 317   0.46394825 -0.929746855 -0.454318341
#> 318   0.64809788 -0.556890647 -0.218689257
#> 319   0.39288946 -0.890564520 -0.518282878
#> 320   0.47986990 -0.760386207 -0.366083390
#> 321   0.23239937  0.106082804  1.016495776
#> 322   0.58020169 -0.603806814  0.087008614
#> 323   0.76462745 -0.167744437  0.057837894
#> 324   0.57840668 -0.532500719  0.128807108
#> 325   0.07022441 -0.194975760  0.610830192
#> 326   0.42173401 -0.743606298 -0.543204325
#> 327   0.31661163 -0.486235263  0.336378825
#> 328   0.54703965 -0.606149284 -0.133122521
#> 329   0.47419507 -0.495760400 -0.113534219
#> 330   0.63972196 -0.152214377  0.789514749
#> 331   0.54518657 -0.380966345  0.036069184
#> 332   0.72646141 -0.282317011  0.278773846
#> 333   0.42016359 -0.008316260  0.157460724
#> 334   0.87555058  0.096843987  0.562812913
#> 335   0.43161435 -0.710027158 -0.062169150
#> 336   0.75770961 -0.233285638  0.448499151
#> 337   0.27739813 -0.290561668  0.075776608
#> 338   0.75109916 -0.379168608 -0.009246565
#> 339   0.58820338 -0.435097255  0.549801478
#> 340   0.55270865 -0.199445792  0.379444169
#> 341   0.56256112 -0.438349054  0.436749854
#> 342   0.71371528 -0.062231650  0.421349901
#> 343   0.71163199  0.195897661  0.543083612
#> 344   0.51682515 -0.111747769  1.033771044
#> 345   0.89356440 -0.139755348  0.436646924
#> 346   0.46047436 -0.534002462  0.127174289
#> 347   0.59264304 -0.643020069 -0.077283486
#> 348  -0.12429401 -0.192666214  0.496508781
#> 349   0.76409533 -0.444993207 -0.124697732
#> 350   0.46689499 -0.756077304  0.315994700
#> 351   0.14861169 -0.835191131 -0.614509802
#> 352   0.56861522 -0.192362679  0.363518411
#> 353   0.64323797  0.017120264  0.409867722
#> 354   0.75141004 -0.066180262  0.159221830
#> 355   0.48418803 -0.386735983  0.083369047
#> 356   0.49282117 -0.347876804  1.020450937
#> 357   0.70180418 -0.012797798  0.345250431
#> 358   0.10853511 -0.262797247  0.509685255
#> 359   0.78403924 -0.176646206  0.300057300
#> 360   0.24675300 -0.728213312  0.488013791
#> 361   0.48920040 -0.672646062  0.168923703
#> 362   0.45518856 -0.156414690  0.263384636
#> 363   0.58283365 -0.514847651 -0.040043751
#> 364   0.78688795 -0.334402444  0.521988264
#> 365   0.67163365 -0.251072627  0.423738670
#> 366   0.57748947  0.328521444  0.628732134
#> 367   0.58457041 -0.362562283  0.444015055
#> 368   0.44545673 -0.169478087  0.505520627
#> 369   0.72034640 -0.323082265  0.531781319
#> 370   0.58820338 -0.435097255  0.549801478
#> 371   0.52735213 -0.060513969  0.446345914
#> 372   0.25923603 -0.122459751  0.331760540
#> 373   0.64899125 -0.721076444 -0.053963699
#> 374   0.64483094 -0.335371426  0.102272825
#> 375   0.68293901  0.016331546  0.847263656
#> 376   0.76209833 -0.219890349  0.504488770
#> 377   0.64022307 -0.354636472  0.160793954
#> 378   0.78688795 -0.334402444  0.521988264
#> 379   0.12648692 -0.419459871  1.130496894
#> 380   0.88097956 -0.160398591  0.133378274
#> 381   0.59294348 -0.545061773  0.603934939
#> 382   0.58714591 -0.330632653  0.079989797
#> 383   0.76209833 -0.219890349  0.504488770
#> 384   0.51288575 -0.452284658 -0.051766783
#> 385   0.52752207 -0.537450682  0.309318051
#> 386   0.54528468 -0.095604206  0.571655134
#> 387   0.25163530 -0.993335506 -0.105940185
#> 388   0.71284966 -0.184603944  0.313687884
#> 389   0.31092189 -0.500200825 -0.192893469
#> 390   0.91289416  0.133311545  0.940469971
#> 391   0.71522917 -0.229912006  0.371059034
#> 392   0.38940579 -1.037832122 -0.105680671
#> 393   0.87555058  0.096843987  0.562812913
#> 394  -0.17946632 -0.050391041  0.409393627
#> 395   0.61145958 -0.657806745 -0.187504103
#> 396   0.71834459 -0.293863481  0.542048029
#> 397   0.46972111 -0.897492211 -0.484119160
#> 398   0.22908919 -0.470578636  0.152473596
#> 399   0.49600399 -0.248711885  0.252423202
#> 400   0.30038608 -0.241919049  0.239174334
#> 401   0.79657498 -0.306695742  0.034228062
#> 402   0.71317996  0.083021280  1.115230301
#> 403   0.58283365 -0.514847651 -0.040043751
#> 404   0.75337489 -0.134004424  0.658113525
#> 405   0.96383389 -0.003602274  0.330036095
#> 406   0.59236409 -0.188568205  0.212891760
#> 407   0.31177987 -0.248766786  0.521199855
#> 408   0.56861522 -0.192362679  0.363518411
#> 409  -0.20425606 -1.448961321 -0.482492184
#> 410   0.42974496 -0.789098419 -0.131103557
#> 411   0.43625465 -0.665240472 -0.316636744
#> 412   0.75337489 -0.134004424  0.658113525
#> 413   0.49387531 -0.152032419  1.041978230
#> 414   0.56926968 -0.144393914  0.345397245
#> 415   0.27316409 -0.525810933  0.066879931
#> 416   0.12679528 -0.705185213 -0.110089372
#> 417   0.70182764 -0.141998014  0.294043413
#> 418   0.71512579 -0.392608569  0.041057824
#> 419   0.68293901  0.016331546  0.847263656
#> 420   0.37758865 -0.488060851  0.403390904
#> 421   0.57580879 -0.629296662 -0.072648075
#> 422   0.57811153 -0.162257392  0.484212521
#> 423   0.21969862 -1.330160642 -0.464780667
#> 424   0.71284966 -0.184603944  0.313687884
#> 425   0.53902693 -0.372958740 -0.218059421
#> 426   0.58669046 -0.058600178  0.184339988
#> 427   0.49481058 -0.361552064  0.049055573
#> 428   0.03847533 -0.107607790  0.059489291
#> 429   0.47011792 -0.645162044 -0.025825770
#> 430   0.25934604 -1.041536724  0.242449156
#> 431   0.56861522 -0.192362679  0.363518411
#> 432   0.69798144  0.095558578  0.452426173
#> 433   0.41793133 -0.876126151  0.384402711
#> 434   0.51117679 -0.465470310 -0.042061493
#> 435   0.67634332 -0.213150598  0.497780568
#> 436   0.35583969  0.024289324  0.419735840
#> 437   0.63921282 -0.543571179 -0.101302897
#> 438   0.61822102 -0.320625834  0.567746523
#> 439   0.43868660 -0.681677123  0.247838490
#> 440   0.71500272 -0.068816017  0.575371494
#> 441   0.33384097 -0.350102558  0.164422289
#> 442   0.30667815 -0.259829281  0.541199976
#> 443   0.66111906 -0.483695929 -0.099917250
#> 444   0.68430009 -0.268334493  0.354966165
#> 445   0.63080615 -0.433882056 -0.066951295
#> 446   0.43757459 -0.589723442 -0.312514357
#> 447   0.50383771 -0.552130188 -0.265489627
#> 448   0.15776397 -0.844233437 -0.128553822
#> 449   0.67634332 -0.213150598  0.497780568
#> 450   0.44327092 -0.817196840 -0.020900759
#> 451   0.45184312 -0.508160647 -0.257921374
#> 452   0.25749227 -0.563708790  0.333757071
#> 453   0.19363433 -0.301374890  0.375180938
#> 454   0.65410430 -0.429468728 -0.134088466
#> 455   0.32838063 -0.617033382  0.661501249
#> 456   0.27859769 -1.184631259 -0.566494589
#> 457   0.64373993 -0.596069112 -0.273828833
#> 458   0.76097435 -0.316887821  0.196729799
#> 459   0.43790109 -0.382331300 -0.114096056
#> 460   0.38647342 -0.420201826  0.584980539
#> 461   0.34233867 -0.507018795  0.072223005
#> 462   0.63972196 -0.152214377  0.789514749
#> 463   0.63080615 -0.433882056 -0.066951295
#> 464   0.71500272 -0.068816017  0.575371494
#> 465   0.69611602 -0.035072958  0.412689894
#> 466   0.39695632 -0.598843114  0.571089679
#> 467   0.60284391 -0.677039947 -0.075330545
#> 468   0.52898206 -0.838295715  0.145037092
#> 469   0.71578684 -0.193093111  0.516328788
#> 470   0.37100792 -0.712878300 -0.619544523
#> 471   0.53004986  0.045889955  0.771199611
#> 472   0.55696157 -0.345743743  0.498074128
#> 473   0.19860259 -1.018595411 -0.554796328
#> 474   0.33832448 -0.481236988 -0.318671226
#> 475   0.77330919 -0.302877949 -0.012989957
#> 476   0.44945331 -0.006251418  0.931498937
#> 477   0.22187844 -0.828679241 -0.406269295
#> 478   0.47698705 -0.517100924  0.071624256
#> 479   0.50156880 -0.107066882  1.135785156
#> 480   0.71930818 -0.215642105  0.258831014
#> 481   0.81407851  0.130771776  0.563117792
#> 482   0.71317996  0.083021280  1.115230301
#> 483   0.69415258 -0.437189519  0.485669464
#> 484   0.71317996  0.083021280  1.115230301
#> 485   0.50544752 -0.338639291  0.362077115
#> 486   0.33416895 -0.711657580 -0.564310372
#> 487   0.75072295 -0.383525458 -0.113818376
#> 488   0.30667815 -0.259829281  0.541199976
#> 489   0.67163365 -0.251072627  0.423738670
#> 490   0.64483094 -0.335371426  0.102272825
#> 491   0.43785127 -1.001398957 -0.863318535
#> 492   0.65913213 -0.596382395  0.147896098
#> 493   0.59574238 -0.485546335 -0.197310207
#> 494   0.66827430 -0.705454018 -0.190099772
#> 495   0.71784457 -0.250012730  0.319825369
#> 496   0.61515370 -0.132078301  1.051741453
#> 497   0.55675322 -0.502790213  0.288888253
#> 498   0.46277669 -0.369118171 -0.109293860
#> 499   0.31016826 -0.397653498  0.221229070
#> 500   0.67050740 -0.073420132  0.788064633
#> 501   0.46958169 -0.956443428 -0.455650474
#> 502   0.47419507 -0.495760400 -0.113534219
#> 503   0.67634332 -0.213150598  0.497780568
#> 504   0.33345352 -0.265313541  0.390607939
#> 505   0.42488287 -0.755077935 -0.531772465
#> 506   0.33108254  0.156689790  1.187620476
#> 507   0.45675744 -0.916232859 -0.270910641
#> 508   0.52225016 -0.425304427 -0.110638439
#> 509   0.21969862 -1.330160642 -0.464780667
#> 510   0.75109916 -0.379168608 -0.009246565
#> 511   0.56256112 -0.438349054  0.436749854
#> 512   0.66071976 -0.417526332  0.268656562
#> 513   0.47162121 -0.821644887 -0.325428335
#> 514   0.68025411 -0.354019431 -0.044255771
#> 515   0.51136815 -0.232715785  0.743280697
#> 516   0.17123604 -0.480184936  0.109695720
#> 517   0.71930818 -0.215642105  0.258831014
#> 518   0.64483094 -0.335371426  0.102272825
#> 519   0.65289658 -0.324650286  0.280673258
#> 520   0.54975464 -0.387957052  0.107710101
#> 521   0.64022307 -0.354636472  0.160793954
#> 522   0.06177458 -0.144696514  0.534155878
#> 523   0.59574238 -0.485546335 -0.197310207
#> 524   0.48275473 -0.539036215  0.256160294
#> 525   0.69065516 -0.003363997  0.728425623
#> 526   0.35471698 -0.169693515  0.293700588
#> 527   0.43438710 -0.695618480 -0.038173630
#> 528   0.59996541 -0.229701037  0.491160197
#> 529   0.50262843 -0.550564164  0.194935102
#> 530   0.75109916 -0.379168608 -0.009246565
#> 531   0.34507378 -0.132811052  0.986697625
#> 532   0.31717070 -1.168613435 -0.107511141
#> 533   0.36625597 -0.580782714  0.189035971
#> 534   0.61145958 -0.657806745 -0.187504103
#> 535   0.65736803 -0.009040059  0.397658597
#> 536   0.56236755 -0.286685796  0.178536808
#> 537   0.49600345 -0.195478143  0.466324040
#> 538   0.89303394  0.144278928  0.373406975
#> 539   0.43520155 -0.331132535  0.025509668
#> 540   0.57774734 -0.422713712  0.106462061
#> 541   0.04070794 -0.130344183  0.071433448
#> 542   0.54856457 -0.453721433  0.120459613
#> 543   0.46972111 -0.897492211 -0.484119160
#> 544   0.65646221 -0.108370807  0.960610835
#> 545   0.43259247 -0.667096819 -0.089974522
#> 546   0.27991205 -0.550905648 -0.067478251
#> 547   0.57811153 -0.162257392  0.484212521
#> 548   0.25163530 -0.993335506 -0.105940185
#> 549   0.76097435 -0.316887821  0.196729799
#> 550   0.64022307 -0.354636472  0.160793954
#> 551   0.43487975 -0.627726746  0.075577680
#> 552   0.54856457 -0.453721433  0.120459613
#> 553   0.67634332 -0.213150598  0.497780568
#> 554   0.74576691 -0.647994093 -0.583264709
#> 555   0.15792075 -0.589118069  0.144491985
#> 556   0.43868660 -0.681677123  0.247838490
#> 557   0.30708722 -0.843222796  1.360596365
#> 558   0.38940579 -1.037832122 -0.105680671
#> 559   0.54544736 -0.120184943  0.654104548
#> 560   0.35890246 -0.628539435  0.565791422
#> 561   0.57330107 -0.361481788  0.493499216
#> 562   0.55187186 -0.488333657 -0.211425066
#> 563   0.45308302 -0.352560933  0.216457523
#> 564   0.65289658 -0.324650286  0.280673258
#> 565   0.71930818 -0.215642105  0.258831014
#> 566   0.68293901  0.016331546  0.847263656
#> 567   0.57268800 -0.430149340 -0.208548251
#> 568   0.47581818 -0.153881540  0.299032961
#> 569   0.64483094 -0.335371426  0.102272825
#> 570   0.37942080 -0.667280600  0.932501460
#> 571   0.66952258 -0.124155918  0.452703391
#> 572   0.56551248 -0.402378787 -0.048887896
#> 573   0.39780777 -0.791037371 -0.162024458
#> 574   0.57267802 -0.476955404  0.286743738
#> 575   0.63972196 -0.152214377  0.789514749
#> 576   0.75907924  0.021543333  0.908443128
#> 577   0.69615847 -0.155441330  0.312666571
#> 578   0.30427038 -0.311808473  0.134560817
#> 579   0.39288946 -0.890564520 -0.518282878
#> 580   0.58714591 -0.330632653  0.079989797
#> 581  -0.10036167 -0.015997351  0.331571192
#> 582   0.73864601 -0.188752016  0.015631478
#> 583   0.57840668 -0.532500719  0.128807108
#> 584   0.45213407 -0.798091905  0.206511467
#> 585   0.46689499 -0.756077304  0.315994700
#> 586   0.33806935 -0.541028167 -0.161728557
#> 587   0.50533877 -0.489437083  0.580394143
#> 588   0.38647342 -0.420201826  0.584980539
#> 589   0.45184312 -0.508160647 -0.257921374
#> 590   0.37248210 -0.746402845 -0.577253343
#> 591   0.37943431 -0.292233890  0.354837794
#> 592   0.67233770 -0.186577851  0.713714035
#> 593   0.53456803 -0.113093666  0.282894881
#> 594   0.76209833 -0.219890349  0.504488770
#> 595   0.71317996  0.083021280  1.115230301
#> 596   0.54518657 -0.380966345  0.036069184
#> 597   0.46958169 -0.956443428 -0.455650474
#> 598   0.25253877 -0.200162158  0.316779243
#> 599   0.43622197 -0.631417735  0.081500339
#> 600   0.90060377  0.113823871  0.886071899
#> 601   0.83770446 -0.267650634  0.152742196
#> 602   0.68430253 -0.238218647  0.922582400
#> 603   0.75907924  0.021543333  0.908443128
#> 604   0.26337629 -0.031733914  0.701192123
#> 605   0.19339376 -1.236451500 -0.072569147
#> 606   0.35952778 -0.234787513  0.645209431
#> 607   0.71930818 -0.215642105  0.258831014
#> 608   0.32476040 -0.705230011  0.486254129
#> 609   0.63476976 -0.461787433 -0.078905679
#> 610   0.55644269 -0.155697224  0.295934079
#> 611   0.29699583 -0.522864301  0.215301761
#> 612   0.68675261 -0.388672715  0.390155338
#> 613   0.83770446 -0.267650634  0.152742196
#> 614   0.65410430 -0.429468728 -0.134088466
#> 615   0.79776519 -0.089208358  0.719353198
#> 616   0.25618007 -0.288614871  0.220055023
#> 617   0.39762454 -0.569232938 -0.159419131
#> 618   0.66930699 -0.167533412  0.323060814
#> 619   0.58020169 -0.603806814  0.087008614
#> 620   0.68801949 -0.235439893  0.156779989
#> 621   0.68430253 -0.238218647  0.922582400
#> 622   0.57774734 -0.422713712  0.106462061
#> 623   0.70786941 -0.021800084  0.546553509
#> 624   0.54703965 -0.606149284 -0.133122521
#> 625   0.68675261 -0.388672715  0.390155338
#> 626   0.54856457 -0.453721433  0.120459613
#> 627   0.90060377  0.113823871  0.886071899
#> 628   0.65736803 -0.009040059  0.397658597
#> 629   0.51175429 -0.029348056  0.806844624
#> 630   0.06616480 -0.797177450 -0.330438263
#> 631   0.21709652 -0.138762279  0.187465219
#> 632   0.57580879 -0.629296662 -0.072648075
#> 633   0.16163966 -0.808137306 -0.330787637
#> 634   0.60284391 -0.677039947 -0.075330545
#> 635   0.67741777  0.105706286  0.523212763
#> 636   0.39005114 -0.816445761 -0.015949478
#> 637   0.63043877 -0.242187516  0.139804532
#> 638   0.22187844 -0.828679241 -0.406269295
#> 639   0.29231957 -0.618693959 -0.447618410
#> 640   0.55187186 -0.488333657 -0.211425066
#> 641   0.54544736 -0.120184943  0.654104548
#> 642   0.13721739 -0.080851309  0.386151196
#> 643   0.68430253 -0.238218647  0.922582400
#> 644   0.76409533 -0.444993207 -0.124697732
#> 645   0.79657498 -0.306695742  0.034228062
#> 646   0.52225016 -0.425304427 -0.110638439
#> 647   0.55093743 -0.662098478  0.393707766
#> 648   0.80530568  0.297176816  0.904035042
#> 649   0.65410430 -0.429468728 -0.134088466
#> 650   0.39288946 -0.890564520 -0.518282878
#> 651   0.37736448 -0.670266721 -0.131909079
#> 652   0.45479324 -0.579153313  0.184361575
#> 653   0.68327232 -0.140921499  0.404592562
#> 654   0.57970343 -0.547477263  0.569961503
#> 655   0.83462962 -0.159930858 -0.008496159
#> 656   0.64001865 -0.281831741  0.872785438
#> 657   0.56562150 -0.301740449  0.097243811
#> 658   0.54703965 -0.606149284 -0.133122521
#> 659   0.68293901  0.016331546  0.847263656
#> 660   0.39288946 -0.890564520 -0.518282878
#> 661   0.49387531 -0.152032419  1.041978230
#> 662   0.42153197 -0.766738328  0.280308409
#> 663   0.68025411 -0.354019431 -0.044255771
#> 664   0.71317996  0.083021280  1.115230301
#> 665   0.68801949 -0.235439893  0.156779989
#> 666   0.79776519 -0.089208358  0.719353198
#> 667   0.47986990 -0.760386207 -0.366083390
#> 668   0.39675578 -0.476453990  0.030532427
#> 669   0.54157047 -0.567557604  0.231729723
#> 670   0.50330762 -0.610824323  0.161677963
#> 671   0.67050740 -0.073420132  0.788064633
#> 672   0.43487975 -0.627726746  0.075577680
#> 673   0.30731355 -0.132645724  0.289529508
#> 674   0.45812200 -1.137169403 -0.726931103
#> 675   0.50156880 -0.107066882  1.135785156
#> 676   0.44150553  0.351656177  0.949427186
#> 677   0.64483094 -0.335371426  0.102272825
#> 678   0.76580182 -0.402494983  0.366592748
#> 679  -0.12030287 -0.551137898  0.507133065
#> 680   0.72646141 -0.282317011  0.278773846
#> 681   0.84043138 -0.336296515  0.784734083
#> 682   0.55302700 -0.637914160  0.079587366
#> 683   0.38951566 -0.314311223  0.667572513
#> 684   0.61145958 -0.657806745 -0.187504103
#> 685   0.35740045 -0.847538129 -0.430229544
#> 686   0.61145958 -0.657806745 -0.187504103
#> 687   0.63921282 -0.543571179 -0.101302897
#> 688   0.22724805 -0.571266262  0.023573897
#> 689   0.49481058 -0.361552064  0.049055573
#> 690   0.49447853 -0.341027962  0.464029232
#> 691   0.46601257 -0.571447886 -0.211384708
#> 692   0.29230382 -0.934545772  0.121229790
#> 693   0.47698705 -0.517100924  0.071624256
#> 694   0.20785958 -0.676906263 -0.078315972
#> 695   0.54703965 -0.606149284 -0.133122521
#> 696   0.12857869 -0.619090005  0.091326603
#> 697   0.61145958 -0.657806745 -0.187504103
#> 698   0.86835416 -0.223868943 -0.093783936
#> 699   0.35890246 -0.628539435  0.565791422
#> 700   0.37664630 -1.299438542 -0.672735147
#> 701   0.56861522 -0.192362679  0.363518411
#> 702   0.57580879 -0.629296662 -0.072648075
#> 703  -0.53720121 -0.257794775  0.447378874
#> 704   0.76097435 -0.316887821  0.196729799
#> 705   0.39288946 -0.890564520 -0.518282878
#> 706   0.63080615 -0.433882056 -0.066951295
#> 707   0.55675322 -0.502790213  0.288888253
#> 708   0.57840668 -0.532500719  0.128807108
#> 709   0.46972111 -0.897492211 -0.484119160
#> 710   0.79657498 -0.306695742  0.034228062
#> 711   0.54133620 -0.411951430 -0.274248908
#> 712   0.42974496 -0.789098419 -0.131103557
#> 713   0.37334403 -1.075498481 -0.358215260
#> 714   0.59304735 -0.243286872  0.167643687
#> 715   0.57330107 -0.361481788  0.493499216
#> 716   0.45675744 -0.916232859 -0.270910641
#> 717   0.54703965 -0.606149284 -0.133122521
#> 718   0.78403924 -0.176646206  0.300057300
#> 719   0.39211412 -0.162660807  0.281169949
#> 720   0.67163365 -0.251072627  0.423738670
#> 721   0.43438710 -0.695618480 -0.038173630
#> 722   0.65796870 -0.608261375 -0.181451503
#> 723   0.58283365 -0.514847651 -0.040043751
#> 724   0.25985776 -0.375179600  0.097001643
#> 725   0.64215800 -0.344621410  0.355787762
#> 726   0.50977514 -0.521304395  0.186438796
#> 727   0.78030276 -0.432338642 -0.196249278
#> 728   0.16821889  0.023443431  0.958105191
#> 729   0.43622197 -0.631417735  0.081500339
#> 730   0.62318346  0.014447399  0.252910415
#> 731   0.58714591 -0.330632653  0.079989797
#> 732   0.43161435 -0.710027158 -0.062169150
#> 733   0.35143967 -0.413728473  0.473169638
#> 734   0.49600345 -0.195478143  0.466324040
#> 735   0.48920040 -0.672646062  0.168923703
#> 736   0.64373993 -0.596069112 -0.273828833
#> 737   0.39625496 -0.942123209 -0.311252784
#> 738   0.33384097 -0.350102558  0.164422289
#> 739   0.53081042 -0.368908558  0.090494427
#> 740   0.40273744  0.120957012  0.991928079
#> 741   0.65811015  0.096719696  0.748908651
#> 742   0.49569255 -0.464062180  0.781662108
#> 743   0.13959323 -0.596338057  0.270022896
#> 744   0.58283365 -0.514847651 -0.040043751
#> 745  -0.02549068 -0.580596566 -0.313842269
#> 746   0.75770961 -0.233285638  0.448499151
#> 747   0.43438710 -0.695618480 -0.038173630
#> 748   0.46394825 -0.929746855 -0.454318341
#> 749   0.46958169 -0.956443428 -0.455650474
#> 750   0.63972196 -0.152214377  0.789514749
#> 751   0.66952258 -0.124155918  0.452703391
#> 752   0.75141004 -0.066180262  0.159221830
#> 753   0.64809788 -0.556890647 -0.218689257
#> 754   0.65289658 -0.324650286  0.280673258
#> 755   0.54528468 -0.095604206  0.571655134
#> 756   0.58283365 -0.514847651 -0.040043751
#> 757   0.60777349 -0.046531110  0.240022863
#> 758   0.57840668 -0.532500719  0.128807108
#> 759   0.29028318 -0.004955282  0.412656769
#> 760   0.50533877 -0.489437083  0.580394143
#> 761   0.17662110 -0.040167748  0.508118306
#> 762   0.44649889 -0.318702586 -0.017974812
#> 763   0.52451482 -0.528343337  0.179548280
#> 764   0.19908661 -0.534155741  0.114893778
#> 765   0.52225016 -0.425304427 -0.110638439
#> 766   0.05133018 -0.895148324  0.585008899
#> 767   0.53800375 -0.374013940  0.063381857
#> 768   0.29231957 -0.618693959 -0.447618410
#> 769   0.68430009 -0.268334493  0.354966165
#> 770   0.71578684 -0.193093111  0.516328788
#> 771   0.77711143 -0.343425841  0.386221758
#> 772   0.40447783 -0.106524170  0.499778563
#> 773   0.58714591 -0.330632653  0.079989797
#> 774   0.64373993 -0.596069112 -0.273828833
#> 775   0.82198086  0.033573494  0.697593606
#> 776   0.40354890 -0.258650644  0.214308319
#> 777   0.43625465 -0.665240472 -0.316636744
#> 778   0.51003076 -0.452298890 -0.063586764
#> 779   0.71784457 -0.250012730  0.319825369
#> 780   0.59010962 -0.089581912  0.481940997
#> 781   0.26259646 -0.676121269  0.449918225
#> 782   0.71284966 -0.184603944  0.313687884
#> 783   0.59264304 -0.643020069 -0.077283486
#> 784   0.56704085 -0.607856355  0.651949325
#> 785   0.68293901  0.016331546  0.847263656
#> 786   0.49825864 -0.158516792  0.471289485
#> 787   0.42159674 -0.481701610  0.242475107
#> 788   0.46234297 -0.222769186  0.359454615
#> 789   0.38415125 -0.173459112  0.155226091
#> 790   0.63080615 -0.433882056 -0.066951295
#> 791   0.74463351 -0.035233149  0.691500984
#> 792   0.47956197  0.003473349  0.585617792
#> 793   0.65410430 -0.429468728 -0.134088466
#> 794   0.47698705 -0.517100924  0.071624256
#> 795   0.65289658 -0.324650286  0.280673258
#> 796   0.67233770 -0.186577851  0.713714035
#> 797   0.73806330 -0.325036353  0.114193725
#> 798   0.55675322 -0.502790213  0.288888253
#> 799   0.59264304 -0.643020069 -0.077283486
#> 800   0.58283365 -0.514847651 -0.040043751
#> 801   0.57580879 -0.629296662 -0.072648075
#> 802   0.58714591 -0.330632653  0.079989797
#> 803   0.41122455 -0.396907475 -0.061428964
#> 804   0.46047436 -0.534002462  0.127174289
#> 805   0.75337489 -0.134004424  0.658113525
#> 806   0.25669990  0.042136559  0.446888304
#> 807   0.26041460 -0.791947885 -0.432185223
#> 808   0.43034749 -0.200980047  0.362898353
#> 809   0.30650911 -0.749722944  0.044791843
#> 810   0.49369611  0.110062521  0.914251955
#> 811   0.38651478 -0.197148577  0.201993677
#> 812   0.34355223 -0.582343690  0.096233562
#> 813   0.71500272 -0.068816017  0.575371494
#> 814   0.61822102 -0.320625834  0.567746523
#> 815   0.57774734 -0.422713712  0.106462061
#> 816   0.52991534 -0.256713528  0.481578355
#> 817   0.57195175 -0.177782486  0.951797712
#> 818   0.55586272 -0.219595455  0.742476553
#> 819   0.45213407 -0.798091905  0.206511467
#> 820   0.48650025 -0.528883036 -0.052790990
#> 821   0.57580879 -0.629296662 -0.072648075
#> 822   0.52451482 -0.528343337  0.179548280
#> 823   0.68675261 -0.388672715  0.390155338
#> 824   0.56973624 -0.473123776  0.071389363
#> 825   0.54518657 -0.380966345  0.036069184
#> 826   0.41995911 -0.139783055  0.489848075
#> 827   0.66952258 -0.124155918  0.452703391
#> 828   0.49936330 -0.237292402  0.301252020
#> 829   0.47162121 -0.821644887 -0.325428335
#> 830   0.48328897 -0.118391615  1.024067430
#> 831   0.77041650  0.050261066  0.967789436
#> 832   0.49387531 -0.152032419  1.041978230
#> 833   0.46972111 -0.897492211 -0.484119160
#> 834   0.64215800 -0.344621410  0.355787762
#> 835   0.65796870 -0.608261375 -0.181451503
#> 836   0.71834459 -0.293863481  0.542048029
#> 837   0.50990318 -0.698597229  0.102583072
#> 838   0.41953741 -0.412974324  0.767841081
#> 839   0.27970464 -0.465537776 -0.244744163
#> 840   0.60517481 -0.197544791  0.406056648
#> 841   0.40042664  0.144317299  1.055595497
#> 842   0.43034749 -0.200980047  0.362898353
#> 843   0.33806935 -0.541028167 -0.161728557
#> 844   0.46708747 -0.349465690 -0.240702069
#> 845   0.50017107 -0.217422031  0.431588944
#> 846   0.22417640 -0.277055339  0.991647385
#> 847   0.71284966 -0.184603944  0.313687884
#> 848   0.53081042 -0.368908558  0.090494427
#> 849   0.73806330 -0.325036353  0.114193725
#> 850   0.42895681 -0.265989271  0.098539566
#> 851   0.57774734 -0.422713712  0.106462061
#> 852   0.29523720 -0.108948031  0.631507315
#> 853   0.62881653 -0.339974281  0.528638313
#> 854   0.47162121 -0.821644887 -0.325428335
#> 855   0.57044058 -0.096211841  0.445334400
#> 856   0.43785127 -1.001398957 -0.863318535
#> 857   0.57330107 -0.361481788  0.493499216
#> 858   0.22265247 -0.717846702 -0.087341000
#> 859   0.74353332 -0.280496475  0.008893078
#> 860   0.42231816 -0.899874430  0.041903034
#> 861   0.42895681 -0.265989271  0.098539566
#> 862   0.35663930 -0.521288583 -0.011415356
#> 863   0.60634932 -0.527089374 -0.570241375
#> 864   0.57774734 -0.422713712  0.106462061
#> 865   0.54518657 -0.380966345  0.036069184
#> 866   0.18444883 -0.365161353  0.202429567
#> 867   0.34710392 -0.394727001  0.851075612
#> 868   0.72621790 -0.317610372 -0.122924362
#> 869   0.56236755 -0.286685796  0.178536808
#> 870   0.50977514 -0.521304395  0.186438796
#> 871   0.39005114 -0.816445761 -0.015949478
#> 872   0.56014255 -0.313519940  0.242825208
#> 873   0.43868660 -0.681677123  0.247838490
#> 874   0.42974496 -0.789098419 -0.131103557
#> 875   0.50544752 -0.338639291  0.362077115
#> 876   0.51003076 -0.452298890 -0.063586764
#> 877   0.43997846 -0.146277497  0.222243094
#> 878   0.69480817 -0.371830367 -0.116372563
#> 879   0.52988091 -0.827126368 -0.840275587
#> 880   0.50383771 -0.552130188 -0.265489627
#> 881   0.52610511 -0.410377516 -0.120994350
#> 882   0.42771507 -0.535187043  0.940898070
#> 883   0.60484730 -0.232010740  0.366408858
#> 884   0.58629667 -0.245281709  0.626989164
#> 885   0.26367091 -0.585109365  0.078172072
#> 886   0.47419507 -0.495760400 -0.113534219
#> 887   0.71784457 -0.250012730  0.319825369
#> 888   0.42974496 -0.789098419 -0.131103557
#> 889   0.58283365 -0.514847651 -0.040043751
#> 890   0.64373993 -0.596069112 -0.273828833
#> 891   0.52138323 -0.094461490  0.446361826
#> 892   0.52225016 -0.425304427 -0.110638439
#> 893   0.47268814 -0.292368047  0.536131681
#> 894   0.32163126 -0.311015907  0.716652681
#> 895   0.10939910 -0.618901664  0.107163345
#> 896   0.66654043  0.364795683  1.672918356
#> 897   0.60588271 -0.652598471 -0.262764616
#> 898   0.43487975 -0.627726746  0.075577680
#> 899   0.66071976 -0.417526332  0.268656562
#> 900   0.36987856 -0.350028818  0.401906193
#> 901   0.89960151 -0.117351208  0.229470634
#> 902   0.55696157 -0.345743743  0.498074128
#> 903   0.64022307 -0.354636472  0.160793954
#> 904   0.71758764 -0.296279457  0.130185287
#> 905   0.43997846 -0.146277497  0.222243094
#> 906   0.65646221 -0.108370807  0.960610835
#> 907   0.73806330 -0.325036353  0.114193725
#> 908   0.43790109 -0.382331300 -0.114096056
#> 909   0.49300504  0.013385072  0.555208700
#> 910   0.63043877 -0.242187516  0.139804532
#> 911  -0.11421031 -0.504083349  0.283262439
#> 912   0.55870119 -0.113186040  0.421389265
#> 913   0.62881653 -0.339974281  0.528638313
#> 914   0.63041235 -0.596911215  0.285976616
#> 915   0.30043989  0.167858992  0.880050787
#> 916   0.42924399 -0.670544567  0.345849088
#> 917   0.71371528 -0.062231650  0.421349901
#> 918   0.41534155  0.203166055  1.510956442
#> 919   0.02805566 -0.212210041  0.255791449
#> 920   0.74463351 -0.035233149  0.691500984
#> 921   0.49936330 -0.237292402  0.301252020
#> 922   0.56973624 -0.473123776  0.071389363
#> 923   0.49600399 -0.248711885  0.252423202
#> 924   0.25749227 -0.563708790  0.333757071
#> 925   0.50330762 -0.610824323  0.161677963
#> 926   0.29699583 -0.522864301  0.215301761
#> 927  -0.06811814 -0.523867700 -0.257880195
#> 928   0.58457041 -0.362562283  0.444015055
#> 929   0.63476976 -0.461787433 -0.078905679
#> 930   0.72646141 -0.282317011  0.278773846
#> 931   0.55187186 -0.488333657 -0.211425066
#> 932   0.32201343 -0.636170234 -0.086063732
#> 933   0.55270865 -0.199445792  0.379444169
#> 934   0.66930699 -0.167533412  0.323060814
#> 935   0.63668531 -0.446139960 -0.099075024
#> 936   0.58093830 -0.660030002  0.024042179
#> 937   0.44045380 -0.343509896  0.255688397
#> 938   0.37726727 -0.229281553  0.095008997
#> 939   0.53800375 -0.374013940  0.063381857
#> 940   0.05863841 -0.028481511  0.597634161
#> 941   0.72621790 -0.317610372 -0.122924362
#> 942   0.69611602 -0.035072958  0.412689894
#> 943   0.79776519 -0.089208358  0.719353198
#> 944   0.11164582 -1.397474925 -0.514582997
#> 945   0.60588271 -0.652598471 -0.262764616
#> 946   0.71512579 -0.392608569  0.041057824
#> 947   0.64725795 -0.223764263  0.670036895
#> 948   0.67634332 -0.213150598  0.497780568
#> 949   0.54759526 -0.415451132 -0.321186910
#> 950   0.21522210 -0.829147370  0.061696727
#> 951   0.60358955  0.236384782  1.080679615
#> 952   0.29084708 -0.010287924  0.560446682
#> 953   0.60484730 -0.232010740  0.366408858
#> 954   0.58629667 -0.245281709  0.626989164
#> 955   0.37605704 -0.575241290 -0.125588190
#> 956   0.43487975 -0.627726746  0.075577680
#> 957   0.04118033 -0.351601424  0.287650544
#> 958   0.43244426 -0.531772863  0.183229522
#> 959   0.57774734 -0.422713712  0.106462061
#> 960   0.71642211 -0.328857738  0.468559552
#> 961   0.25439635 -1.136657236 -0.275097903
#> 962   0.35775924 -0.779285172 -0.187503152
#> 963   0.51288575 -0.452284658 -0.051766783
#> 964   0.50977514 -0.521304395  0.186438796
#> 965   0.56562150 -0.301740449  0.097243811
#> 966   0.71834459 -0.293863481  0.542048029
#> 967   0.71642211 -0.328857738  0.468559552
#> 968   0.63043877 -0.242187516  0.139804532
#> 969   0.48275473 -0.539036215  0.256160294
#> 970   0.76097435 -0.316887821  0.196729799
#> 971   0.57198230 -0.544794075  0.592863488
#> 972   0.43785762 -0.469677685  1.260898606
#> 973   0.71784457 -0.250012730  0.319825369
#> 974   0.58032901 -0.624467554  0.015084688
#> 975   0.44327092 -0.817196840 -0.020900759
#> 976   0.71834459 -0.293863481  0.542048029
#> 977   0.76097435 -0.316887821  0.196729799
#> 978   0.49825864 -0.158516792  0.471289485
#> 979   0.47602416 -0.417497806  0.440876680
#> 980   0.60517481 -0.197544791  0.406056648
#> 981   0.43757459 -0.589723442 -0.312514357
#> 982  -0.08600931 -0.691821796 -0.080353048
#> 983   0.47941544 -0.384996136  1.184850415
#> 984   0.57580879 -0.629296662 -0.072648075
#> 985   0.75337489 -0.134004424  0.658113525
#> 986   0.56236755 -0.286685796  0.178536808
#> 987   0.24661997 -0.383877420  0.285384767
#> 988   0.68293901  0.016331546  0.847263656
#> 989   0.45433358 -0.300979352 -0.271848504
#> 990   0.39695632 -0.598843114  0.571089679
#> 991   0.30009126 -0.040882011  1.240073204
#> 992   0.16158352 -0.329990825  0.202140864
#> 993   0.56236755 -0.286685796  0.178536808
#> 994   0.49825864 -0.158516792  0.471289485
#> 995   0.56119848 -0.419359563  0.636055664
#> 996   0.51476904 -0.414323380  0.032749381
#> 997   0.64022307 -0.354636472  0.160793954
#> 998   0.56236755 -0.286685796  0.178536808
#> 999   0.72646141 -0.282317011  0.278773846
#> 1000  0.55675322 -0.502790213  0.288888253
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1  0.223  0.161  0.533 
#>  2      2        4       70        1  0.334 -0.793 -0.0252
#>  3      3        4       70        1  0.578 -0.423  0.106 
#>  4      4        4       70        1  0.576 -0.629 -0.0726
#>  5      5        4       70        1  0.591 -0.510  0.401 
#>  6      6        4       70        1  0.762 -0.220  0.504 
#>  7      7        4       70        1  0.688 -0.235  0.157 
#>  8      8        4       70        1  0.197 -0.448  0.191 
#>  9      9        4       70        1  0.777 -0.479 -0.132 
#> 10     10        4       70        1  0.432 -0.532  0.183 
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  5.00  82.2 1.70  0.0608 16.9     5.74e+2 1386.
#> 2      1    14     4    70     1  5.00  82.2 1.70  0.0608 10.9     1.38e-7  899.
#> 3      2     1     4    70     1  5.59  31.7 0.975 0.176  30.1     9.72e+2  953.
#> 4      2    14     4    70     1  5.59  31.7 0.975 0.176   6.82    3.04e-3  216.
#> 5      3     1     4    70     1  7.13  45.9 1.11  0.155  22.8     8.80e+2 1047.
#> 6      3    14     4    70     1  7.13  45.9 1.11  0.155   5.98    4.62e-4  274.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#>