Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object

A posologyr list, created by the posologyr function.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01,n_sample=1e4,n_resample=1e3)
#> $eta
#>           ETA_Cl       ETA_Vc        ETA_Ka
#> 1     0.74689174 -0.046377313  0.8145792005
#> 2     0.77529400 -0.315800762 -0.2475432168
#> 3     0.68247263 -0.197359269  0.2674037168
#> 4     0.38357233 -0.230568574  0.6919618483
#> 5     0.45968366 -0.274386670  0.0098288480
#> 6     0.39636164 -0.816034391  0.1242167965
#> 7     0.17630092 -0.506108172 -0.0100774422
#> 8     0.15608877 -0.811970053 -0.5387191583
#> 9     0.62468095 -0.580171964 -0.1923786213
#> 10    0.69179649 -0.135363618  0.6367796057
#> 11    0.61445001 -0.499566077 -0.2774436370
#> 12    0.73104654 -0.207175394 -0.0697522826
#> 13    0.54608599 -0.136689444  0.0871073626
#> 14    0.43753208 -0.794898805 -0.3078075334
#> 15    0.58147434 -0.286637249  0.8008210177
#> 16    0.54203072 -0.284514266  0.2841830218
#> 17    0.44042502 -0.588830620 -0.1131725629
#> 18    0.24142390 -0.289797173  0.3279969200
#> 19    0.49248029 -0.551793977  0.6155890113
#> 20    0.35505777 -0.427676094  0.1274407589
#> 21    0.58683810 -0.590760141 -0.2989846496
#> 22    0.69845368 -0.330286865  0.2331217467
#> 23    0.31368014 -0.526661551  0.3550542246
#> 24    0.44630636 -0.750918368 -0.0069742798
#> 25    0.52806895 -0.241068197  0.1109710872
#> 26    0.65806881 -0.253295255  0.0483571540
#> 27    0.36123421 -0.353001557  0.2872487968
#> 28    0.72281764 -0.255544422 -0.0841290516
#> 29    0.44410752 -0.552683680  0.3979812091
#> 30    0.52973542 -0.441271942 -0.0971984542
#> 31    0.41532091 -0.487260319 -0.3184517763
#> 32    0.41983777 -0.804415818 -0.3618482191
#> 33    0.55820707 -0.646319758 -0.1598267403
#> 34    0.69741447 -0.456599949  0.3570067323
#> 35    0.35918897 -0.605635502 -0.4019817152
#> 36    0.66509082 -0.536663724 -0.0529563153
#> 37    0.67583005  0.213942172  0.9579147732
#> 38    0.65581111 -0.255229561  0.4208370226
#> 39    0.64171599 -0.461242662  0.6573480371
#> 40    0.32768856 -0.122876337  0.6659634082
#> 41    0.40964936 -0.329622897  0.0446560641
#> 42    0.53958291 -0.017863697  0.5817641254
#> 43    0.46256995 -0.437468361  0.4669031666
#> 44    0.17820935 -0.214337954 -0.0271498930
#> 45    0.77076946 -0.252813499  0.2769267202
#> 46    0.71696885  0.032035381  0.4475448719
#> 47    0.10776009 -0.652897855  0.1858465676
#> 48    0.53018556 -0.667533597  0.5926073265
#> 49    0.11369126  0.029885450  0.6069953950
#> 50    0.85857101 -0.156628025  0.7205738837
#> 51    0.13035853 -0.363549760  0.7013757666
#> 52    0.45671501 -0.519742487  0.1377234784
#> 53    0.47848990 -0.273984758  0.5435843554
#> 54    0.16718153 -0.779664375 -0.2189115298
#> 55    0.72449050 -0.272773496 -0.2195261369
#> 56    0.59786709 -0.237560134  0.1017409626
#> 57    0.47912944 -0.714055403  0.1536937162
#> 58    0.52806895 -0.241068197  0.1109710872
#> 59    0.76904297 -0.160051214  0.2771064424
#> 60    0.33584952 -0.451844318  0.2717205423
#> 61    0.70977693 -0.310383551  0.7017925862
#> 62    0.83288490 -0.301899502  0.7603006973
#> 63    0.71028578 -0.455337921  0.4015318188
#> 64    0.47937896 -0.385091729  0.0078402063
#> 65    0.46757427 -0.521108611  0.8312887467
#> 66    0.37508641 -0.592376876 -0.0679279773
#> 67    0.42619326 -0.146013583  0.6291503784
#> 68    0.66754043 -0.121183939  0.4697668492
#> 69    0.42579078 -0.233912150  0.4624864450
#> 70    0.46436078 -0.380078894  0.2323239842
#> 71    0.54203072 -0.284514266  0.2841830218
#> 72    0.63883979 -0.294633612 -0.0545249330
#> 73    0.74537906 -0.541980900  0.1339822962
#> 74    0.93148377  0.051714644  0.4838531996
#> 75    0.42101065 -0.498685957  0.2554591847
#> 76    0.62035632  0.134884165  0.6788419809
#> 77    0.44753006 -0.538684930  1.0815146257
#> 78    0.59296366 -0.265018598  0.3396277072
#> 79    0.35995513 -0.734159324  0.1227920574
#> 80    0.13796739 -0.965399004  0.1115662589
#> 81    0.71185515  0.140897552  0.3135447136
#> 82    0.63456368  0.090259175  0.7443703479
#> 83    0.21554359 -0.088038325  0.2946438247
#> 84    0.49504548 -0.324230669  0.3043826978
#> 85    0.32884842 -0.509250807 -0.3552872700
#> 86    0.34386744 -0.644605034 -0.4971604583
#> 87    0.31933982 -0.660146187  0.3247310159
#> 88    0.43753208 -0.794898805 -0.3078075334
#> 89    0.09148870 -0.345752776  0.1846206826
#> 90    0.57349238 -0.655247959 -0.1680784296
#> 91    0.41983777 -0.804415818 -0.3618482191
#> 92    0.73023843 -0.191571791  0.1395903787
#> 93    0.82004740 -0.085312108  0.4844790870
#> 94    0.52973542 -0.441271942 -0.0971984542
#> 95    0.23821341 -0.684405295 -0.2251562992
#> 96    0.40907590 -0.831980860 -0.3363745945
#> 97    0.43543772 -0.663905441 -0.3961282661
#> 98    0.30825531 -0.239708456  0.4408498717
#> 99    0.42619326 -0.146013583  0.6291503784
#> 100  -0.24821535 -0.542093732  0.1733737969
#> 101   0.43543772 -0.663905441 -0.3961282661
#> 102   0.56644591 -0.683220199 -0.3461736987
#> 103   0.43250261 -0.392466596 -0.1028040586
#> 104   0.67251307 -0.246740055  0.3849273558
#> 105   0.27081263 -0.318375071 -0.2304739049
#> 106   0.32272474 -0.140855078  0.7179639651
#> 107   0.44238191 -0.424504966 -0.2802741292
#> 108   0.63712130 -0.238939202  0.6521093917
#> 109   0.63456368  0.090259175  0.7443703479
#> 110   0.67704915 -0.605029949 -0.4467158886
#> 111   0.57349238 -0.655247959 -0.1680784296
#> 112   0.74689174 -0.046377313  0.8145792005
#> 113   0.71028578 -0.455337921  0.4015318188
#> 114   0.58683810 -0.590760141 -0.2989846496
#> 115   0.52629379 -0.181849074  0.1127380847
#> 116   0.45724645 -0.517725344  0.2040453476
#> 117   0.76904297 -0.160051214  0.2771064424
#> 118   0.68264515 -0.331270775  0.2405168442
#> 119   0.28077362 -0.620590142 -0.3953819331
#> 120   0.44212184 -0.229372534  0.1587576264
#> 121  -0.09177362 -0.123491242  0.5595867134
#> 122   0.31368014 -0.526661551  0.3550542246
#> 123   0.66112718 -0.445022252 -0.1040301594
#> 124   0.53080990 -0.384566301  0.4005459349
#> 125   0.54702393 -0.296816949  0.0439343426
#> 126   0.56644591 -0.683220199 -0.3461736987
#> 127   0.40964936 -0.329622897  0.0446560641
#> 128   0.33699409 -0.654880076 -0.0216754266
#> 129   0.50222953 -0.082707806  0.1855486456
#> 130   0.75018723 -0.187125605  0.0576567887
#> 131   0.63258128 -0.518639117  0.4378879676
#> 132   0.58147434 -0.286637249  0.8008210177
#> 133   0.76904297 -0.160051214  0.2771064424
#> 134   0.59250294 -0.618189314 -0.2248972977
#> 135   0.53080990 -0.384566301  0.4005459349
#> 136   0.57382598 -0.823103148 -0.4855075456
#> 137   0.61427768 -0.360210681  0.3510780741
#> 138   0.30679282 -0.358683963  0.0674864203
#> 139   0.52329004 -0.548106823 -0.4751994747
#> 140   0.69179649 -0.135363618  0.6367796057
#> 141   0.92256693 -0.135669749  0.0039493040
#> 142   0.24966987 -0.229437432  0.8134062143
#> 143   0.46509909 -0.035846960  0.4692617058
#> 144   0.51353261 -0.749915386 -0.2669797904
#> 145   0.38098049 -0.171943013  0.2264845161
#> 146   0.68264515 -0.331270775  0.2405168442
#> 147   0.58196607  0.127083360  0.4737129241
#> 148   0.62265315 -0.517326542  0.2862996481
#> 149   0.91736331  0.132446874  0.5762341191
#> 150   0.39759009 -0.586423078  0.6584404874
#> 151   0.52088852 -0.491688763  0.6460260370
#> 152   0.29744959 -0.532413488 -0.1143364566
#> 153   0.24258991 -0.785858840  0.0089909581
#> 154   0.41970576 -0.151597181 -0.0261813042
#> 155   0.32272474 -0.140855078  0.7179639651
#> 156   0.58683810 -0.590760141 -0.2989846496
#> 157   0.68174656 -0.088777230  0.6843814912
#> 158   0.63430272 -0.300336843  0.1374133536
#> 159   0.37919268 -0.673342831  0.0705044648
#> 160   0.17340614 -0.630388976 -0.4396514130
#> 161   0.43753208 -0.794898805 -0.3078075334
#> 162   0.55820707 -0.646319758 -0.1598267403
#> 163   0.39694830 -0.757745042  0.2865092909
#> 164   0.49569253 -0.136214712  0.2037082113
#> 165   0.51968142  0.173944231  0.8341875591
#> 166   0.61770433 -0.431952312 -0.0499418228
#> 167   0.45284898 -0.230956711  0.4903659719
#> 168   0.31490385 -0.446353367  0.2838823739
#> 169   0.75297902 -0.246301214  0.0468612617
#> 170   0.52279396 -0.332365425 -0.0149178731
#> 171  -0.31952684 -0.147687216  0.9995452371
#> 172   0.35027948 -0.158504352  0.5541752992
#> 173   0.70440114 -0.590053691 -0.3362860268
#> 174   0.51828561 -0.372742120  0.6154481562
#> 175   0.51680669 -0.699913448  0.0408541512
#> 176   0.62188545 -0.667312469 -0.3344145828
#> 177   0.63043450 -0.665417450 -0.0718435605
#> 178   0.48545203 -0.502827049  0.2488282115
#> 179   0.80326198 -0.156283800  0.0832953067
#> 180   0.51680669 -0.699913448  0.0408541512
#> 181   0.26198183 -0.448519142  0.1176719594
#> 182   0.44366870 -0.699360419 -0.2256445540
#> 183   0.47753228 -0.482116028  0.2887315354
#> 184   0.64315878 -0.592159549  0.4536222151
#> 185   0.73141898 -0.026568937  0.6392492112
#> 186   0.38499263 -0.586447598  0.2621594360
#> 187   0.63712130 -0.238939202  0.6521093917
#> 188   0.70977693 -0.310383551  0.7017925862
#> 189   0.61427768 -0.360210681  0.3510780741
#> 190   0.33389657 -0.392610189  0.1854091130
#> 191   0.49907998 -0.342225743 -0.0889760869
#> 192   0.66082456 -0.429019540  0.7510064734
#> 193   0.41336349 -1.052730795 -0.6332462762
#> 194   0.66509082 -0.536663724 -0.0529563153
#> 195   0.41897550 -0.348354474  0.1695914977
#> 196   0.59250294 -0.618189314 -0.2248972977
#> 197   0.55602076 -0.778296571 -0.4039044867
#> 198   0.42059814 -0.253600737  0.2312439105
#> 199   0.55854437 -0.472500561  0.1208543570
#> 200   0.29762359 -0.470462256 -0.3568409310
#> 201   0.42613585 -0.511146944  0.4183233909
#> 202   0.52236848 -0.578739697  0.4272298197
#> 203   0.66106754  0.100464993  0.8966364569
#> 204   0.58073789 -0.625663860 -0.2037501953
#> 205   0.49799980 -0.424349633  0.1084196524
#> 206   0.54633852 -0.420086920  0.0016254666
#> 207   0.27079955 -1.132369041 -0.1203475740
#> 208   0.58483354 -0.348170354  0.8640330804
#> 209   0.65938607 -0.647289921 -0.1935366886
#> 210   0.38347078 -0.589137034 -0.0090350983
#> 211   0.26349567 -0.374352038  0.1596631445
#> 212   0.33699409 -0.654880076 -0.0216754266
#> 213   0.61427768 -0.360210681  0.3510780741
#> 214   0.44956898 -0.565145630  0.0391269454
#> 215   0.60197658 -0.097272623  0.8634771861
#> 216   0.48545203 -0.502827049  0.2488282115
#> 217   0.44630636 -0.750918368 -0.0069742798
#> 218   0.56639112 -0.416447320 -0.3795595931
#> 219   0.45914224 -0.349800853  0.8324760462
#> 220   0.50107153 -0.149462199  0.4239839521
#> 221   0.46436078 -0.380078894  0.2323239842
#> 222   0.42041791 -0.584504378  0.0824251928
#> 223   0.59344407 -0.325838814  0.3407141707
#> 224   0.60104328 -0.534232020  0.5117290982
#> 225   0.70286802 -0.328452816  0.6315761738
#> 226   0.71696885  0.032035381  0.4475448719
#> 227   0.63883979 -0.294633612 -0.0545249330
#> 228   0.57382598 -0.823103148 -0.4855075456
#> 229   0.70851556 -0.276933464  0.5389992742
#> 230   0.57382598 -0.823103148 -0.4855075456
#> 231   0.38784719 -0.717383090  0.2926883076
#> 232   0.19202042 -0.826055468 -0.0614114477
#> 233   0.51424615 -0.576625358  0.4106730234
#> 234   0.58683810 -0.590760141 -0.2989846496
#> 235   0.45845772 -0.550780988  0.2894405838
#> 236   0.62188545 -0.667312469 -0.3344145828
#> 237   0.31230699 -0.212656559  0.0996756727
#> 238   0.42619326 -0.146013583  0.6291503784
#> 239   0.66754043 -0.121183939  0.4697668492
#> 240   0.28709723 -0.244333531  0.1522902652
#> 241   0.71028578 -0.455337921  0.4015318188
#> 242   0.51356862 -0.515182708  0.1720116185
#> 243   0.62750404 -0.367786797  0.8049520137
#> 244   0.48459395  0.135600963  0.5215019070
#> 245   0.61864615 -0.285542433  0.3485217047
#> 246   0.66845854 -0.211990824  0.5455203280
#> 247   0.69179649 -0.135363618  0.6367796057
#> 248   0.55102707 -0.627786498  1.0968063374
#> 249   0.30897609 -0.227588175  0.9703493161
#> 250   0.69179649 -0.135363618  0.6367796057
#> 251   0.58114911 -0.190652501  0.2321163542
#> 252   0.61427768 -0.360210681  0.3510780741
#> 253   0.32910393 -0.225556884  0.1659741667
#> 254   0.64690436 -0.443321818 -0.1354354997
#> 255   0.43656833 -0.702918167 -0.4153535748
#> 256   0.46574431 -0.501593164  0.0208127942
#> 257   0.56638464 -0.152129269  0.0526817230
#> 258   0.59268270 -0.023315489  0.7502454201
#> 259   0.69399612 -0.414747132 -0.2511274292
#> 260   0.57635206 -0.713242571 -0.1709322154
#> 261   0.66754043 -0.121183939  0.4697668492
#> 262   0.69845368 -0.330286865  0.2331217467
#> 263   0.57349238 -0.655247959 -0.1680784296
#> 264   0.58683810 -0.590760141 -0.2989846496
#> 265   0.65938607 -0.647289921 -0.1935366886
#> 266   0.58288775 -0.251107890  0.0386281094
#> 267   0.32074108 -0.304796831  0.4065581756
#> 268   0.51918460 -0.411445411  0.3660778289
#> 269   0.47235264 -0.466634480  0.0312608888
#> 270   0.26224629 -1.027078630 -0.4000552127
#> 271   0.47326543 -0.243998934  1.1581667177
#> 272   0.53863873 -0.259109121  0.4708994146
#> 273   0.55996177 -0.611314941 -0.0053410996
#> 274   0.80656368 -0.004106104  0.1703093425
#> 275   0.55820707 -0.646319758 -0.1598267403
#> 276   0.68174656 -0.088777230  0.6843814912
#> 277   0.47469122 -0.412817197  0.0450344375
#> 278   0.57349238 -0.655247959 -0.1680784296
#> 279   0.43274769 -0.288378212  0.8296402467
#> 280   0.76904297 -0.160051214  0.2771064424
#> 281   0.41926193 -0.818351828  0.4485356792
#> 282   0.64573097 -0.222954847  0.0910116505
#> 283   0.22020322 -0.149274746  0.5412184755
#> 284   0.21299276 -0.320626269  0.3145045516
#> 285   0.79318612 -0.265532942  0.2388998556
#> 286   0.42101065 -0.498685957  0.2554591847
#> 287   0.57635206 -0.713242571 -0.1709322154
#> 288   0.66330688 -0.361861155  0.1691612358
#> 289   0.53469475 -0.084122627  0.6660449640
#> 290   0.29875549  0.042737666  0.3524744596
#> 291   0.47763186 -0.619641635 -0.2564566870
#> 292  -0.02376127 -0.405397233 -0.1903134708
#> 293   0.65310389 -0.165705458  0.2490033063
#> 294   0.66150335 -0.192789324  0.3807427047
#> 295   0.68264515 -0.331270775  0.2405168442
#> 296   0.52693852 -0.574746951 -0.2686515480
#> 297   0.67251307 -0.246740055  0.3849273558
#> 298   0.55553434 -0.066120383  0.1522004571
#> 299   0.08452932 -0.714824359  0.0763648578
#> 300   0.57518604 -0.599335319 -0.2754663703
#> 301   0.63456368  0.090259175  0.7443703479
#> 302  -0.03842220 -0.410986141  0.3786585376
#> 303   0.55989202 -0.147769861  0.5663240205
#> 304   0.76677408 -0.335819119  0.0306446745
#> 305   0.69399612 -0.414747132 -0.2511274292
#> 306   0.67020913 -0.441910240 -0.0844783968
#> 307   0.52182952  0.005454811  0.4270779249
#> 308   0.42877703 -0.077596833  0.2979475831
#> 309   0.33608834 -0.843921391 -0.5142773918
#> 310   0.57635206 -0.713242571 -0.1709322154
#> 311   0.53951581 -0.356420991  0.3263492303
#> 312   0.34685812 -0.480692947 -0.2796400747
#> 313   0.52580151 -0.663577768  0.2140434521
#> 314   0.36970222 -0.506914662 -0.0018881371
#> 315   0.15760568 -0.105413801  0.4770480827
#> 316   0.43239448 -0.646537362  0.3489184472
#> 317   0.53255235 -0.628631127  0.5217400023
#> 318   0.48736718 -0.122600243  0.1952433875
#> 319   0.47912944 -0.714055403  0.1536937162
#> 320   0.61273156 -0.247283837  0.2859387279
#> 321   0.68354344 -0.201675991  0.7445557963
#> 322   0.20258149 -0.795304562 -0.4354404786
#> 323   0.63712130 -0.238939202  0.6521093917
#> 324   0.35328547  0.242193294  0.8744979392
#> 325   0.12049902 -0.521056781  0.1962177514
#> 326  -0.02571309 -0.208815000  0.0016585516
#> 327   0.51918460 -0.411445411  0.3660778289
#> 328   0.69179649 -0.135363618  0.6367796057
#> 329   0.67459540 -0.462056485  0.1984093989
#> 330   0.83288490 -0.301899502  0.7603006973
#> 331   0.62123318 -0.278218629  0.5478232193
#> 332   0.70851556 -0.276933464  0.5389992742
#> 333   0.22393647 -0.222273272  0.1632802430
#> 334   0.38011358 -0.896335588 -0.2438200383
#> 335   0.82004740 -0.085312108  0.4844790870
#> 336   0.40175914 -0.696506564 -0.3071778640
#> 337   0.61273156 -0.247283837  0.2859387279
#> 338   0.68174656 -0.088777230  0.6843814912
#> 339   0.43508749 -0.196389034  0.2813376374
#> 340   0.32852910 -0.164793267  1.0617625368
#> 341   0.64212932 -0.445821488  0.1516983977
#> 342   0.26198183 -0.448519142  0.1176719594
#> 343   0.44956898 -0.565145630  0.0391269454
#> 344   0.44366870 -0.699360419 -0.2256445540
#> 345   0.68354344 -0.201675991  0.7445557963
#> 346   0.41926193 -0.818351828  0.4485356792
#> 347   0.55820707 -0.646319758 -0.1598267403
#> 348   0.47326543 -0.243998934  1.1581667177
#> 349   0.31223555 -1.017515820  0.1282144969
#> 350   0.82869417 -0.110442233  0.0409835654
#> 351   0.45845772 -0.550780988  0.2894405838
#> 352   0.65549675  0.040036538  0.4482764068
#> 353   0.60104328 -0.534232020  0.5117290982
#> 354   0.61864615 -0.285542433  0.3485217047
#> 355  -0.10698646 -0.034349803  0.1574965027
#> 356   0.38784719 -0.717383090  0.2926883076
#> 357   0.57062233  0.056039120  0.4453038029
#> 358   0.25044844 -0.461116710  0.1339655718
#> 359   0.41971058 -0.506023512  0.1048050507
#> 360   0.35995513 -0.734159324  0.1227920574
#> 361   0.32074108 -0.304796831  0.4065581756
#> 362   0.86286469 -0.211183227  0.0277705398
#> 363   0.45671501 -0.519742487  0.1377234784
#> 364   0.35995513 -0.734159324  0.1227920574
#> 365   0.53257261 -0.170131456  0.0986629626
#> 366   0.33637780 -0.146978079  0.5017773677
#> 367   0.46266510 -0.160586563  0.2628545603
#> 368   0.66330688 -0.361861155  0.1691612358
#> 369   0.38011358 -0.896335588 -0.2438200383
#> 370   0.73447193 -0.274042958  0.2618683684
#> 371   0.68354344 -0.201675991  0.7445557963
#> 372   0.15170465 -1.465708682 -0.5689957212
#> 373   0.61427768 -0.360210681  0.3510780741
#> 374   0.50132479 -0.215669729  0.2744433193
#> 375   0.42041791 -0.584504378  0.0824251928
#> 376   0.35995513 -0.734159324  0.1227920574
#> 377   0.41025062 -0.560151636  0.0679375527
#> 378   0.73141898 -0.026568937  0.6392492112
#> 379   0.63430272 -0.300336843  0.1374133536
#> 380   0.47165344 -0.045103353  0.7864822783
#> 381   0.57382598 -0.823103148 -0.4855075456
#> 382   0.48078828 -0.066301557  0.6977623492
#> 383   0.51353261 -0.749915386 -0.2669797904
#> 384   0.31803919 -0.429160386  0.2497179658
#> 385   0.67459540 -0.462056485  0.1984093989
#> 386   0.83288490 -0.301899502  0.7603006973
#> 387   0.45576202 -0.779783886  0.1646449204
#> 388   0.70286802 -0.328452816  0.6315761738
#> 389   0.24711486 -0.946676477 -0.0784234946
#> 390   0.61273156 -0.247283837  0.2859387279
#> 391  -0.07697670 -0.568769213  0.4557752874
#> 392   0.94781920  0.125765604  0.8102971615
#> 393   0.54687660 -0.514449989 -0.1213013334
#> 394   0.43598906 -0.607843793  0.6433279939
#> 395   0.67321820 -0.575470458 -0.0596643103
#> 396   0.91607686 -0.106047936  0.3961074671
#> 397   0.51828561 -0.372742120  0.6154481562
#> 398   0.57382598 -0.823103148 -0.4855075456
#> 399   0.43598906 -0.607843793  0.6433279939
#> 400   0.26330541  0.019258186  0.3304315723
#> 401   0.61770433 -0.431952312 -0.0499418228
#> 402   0.54552924 -0.017152528  0.0267816151
#> 403   0.55898460 -0.144649284  0.2188104278
#> 404   0.61427768 -0.360210681  0.3510780741
#> 405   0.46781526 -0.765661371  0.3320947566
#> 406   0.64351101 -0.537909105  0.6877194278
#> 407   0.54392344  0.038094554  0.1980670245
#> 408   0.60197658 -0.097272623  0.8634771861
#> 409   0.29823338 -0.588308485  0.0993981268
#> 410   0.47723337 -0.444763927  0.2743816095
#> 411   0.40905340 -0.693946292 -0.2521328424
#> 412   0.17696354 -0.517685761  0.1566499629
#> 413   0.62123318 -0.278218629  0.5478232193
#> 414   0.73490192 -0.038866548  0.5704604339
#> 415   0.22949452 -0.352034376  0.1425878641
#> 416   0.42613585 -0.511146944  0.4183233909
#> 417   0.50132479 -0.215669729  0.2744433193
#> 418   0.52236848 -0.578739697  0.4272298197
#> 419   0.59422521 -0.361307024 -0.0949291618
#> 420   0.55746934 -0.157483342  0.0179986610
#> 421   0.78898285  0.001012107  1.1072669224
#> 422   0.42769875 -0.435864791  0.1403957364
#> 423   0.93605744  0.367796898  0.6657790692
#> 424   0.39686550 -0.824983571 -0.3301133510
#> 425   0.36295230 -0.452312011 -0.0701577140
#> 426   0.57635206 -0.713242571 -0.1709322154
#> 427   0.51424615 -0.576625358  0.4106730234
#> 428   0.37919268 -0.673342831  0.0705044648
#> 429   0.68174656 -0.088777230  0.6843814912
#> 430   0.46160313 -0.517014574  0.0462562023
#> 431   0.34298893 -0.369546034  0.0533367795
#> 432   0.52783293  0.044758572  0.8165751329
#> 433   0.26349567 -0.374352038  0.1596631445
#> 434   0.62265315 -0.517326542  0.2862996481
#> 435   0.37183359 -0.543611170 -0.1717381915
#> 436   0.17943923  0.047565311  0.4991534908
#> 437   0.75486787 -0.134019233  0.0511483112
#> 438   0.43753208 -0.794898805 -0.3078075334
#> 439   0.25250747 -0.429057982  0.2753746127
#> 440   0.44366870 -0.699360419 -0.2256445540
#> 441   0.42613585 -0.511146944  0.4183233909
#> 442   0.73490192 -0.038866548  0.5704604339
#> 443   0.55854437 -0.472500561  0.1208543570
#> 444   0.51828561 -0.372742120  0.6154481562
#> 445   0.52580151 -0.663577768  0.2140434521
#> 446   0.46434655 -0.508728884 -0.0807735858
#> 447   0.47723337 -0.444763927  0.2743816095
#> 448   0.59786709 -0.237560134  0.1017409626
#> 449   0.63290124 -0.621108146  0.2897731128
#> 450   0.15170465 -1.465708682 -0.5689957212
#> 451   0.71028578 -0.455337921  0.4015318188
#> 452   0.53255235 -0.628631127  0.5217400023
#> 453   0.69339097 -0.175328853  0.1857408590
#> 454   0.49185713 -0.447493575  0.0394659427
#> 455   0.47912944 -0.714055403  0.1536937162
#> 456   0.72576638 -0.320053223  0.3373322991
#> 457   0.51680669 -0.699913448  0.0408541512
#> 458   0.37742390 -0.331126943  0.2611003488
#> 459   0.34913046 -0.148880532  0.4693474359
#> 460   0.34650643 -0.081428992  0.4967122504
#> 461  -0.26819533 -0.409901745  0.2861031329
#> 462   0.49248029 -0.551793977  0.6155890113
#> 463   0.83133877  0.210384589  0.6721096242
#> 464   0.56794149 -0.267882744 -0.0413174951
#> 465   0.43656833 -0.702918167 -0.4153535748
#> 466   0.57349238 -0.655247959 -0.1680784296
#> 467   0.28924109 -0.200999565  0.5281087178
#> 468   0.47753228 -0.482116028  0.2887315354
#> 469   0.47469122 -0.412817197  0.0450344375
#> 470   0.50813148 -0.710966503  0.4477073570
#> 471   0.74689174 -0.046377313  0.8145792005
#> 472   0.52973542 -0.441271942 -0.0971984542
#> 473   0.47723337 -0.444763927  0.2743816095
#> 474   0.26800234 -0.113976416  0.3249553803
#> 475   0.36468780 -1.002136177 -0.1139896640
#> 476   0.31803919 -0.429160386  0.2497179658
#> 477   0.68247263 -0.197359269  0.2674037168
#> 478   0.38499263 -0.586447598  0.2621594360
#> 479   0.45576202 -0.779783886  0.1646449204
#> 480   0.48066557 -0.967009978 -0.1955611077
#> 481   0.53080990 -0.384566301  0.4005459349
#> 482   0.49471769 -0.514029461  0.2923724888
#> 483   0.73447193 -0.274042958  0.2618683684
#> 484   0.65144645 -0.401622154 -0.1986687437
#> 485   0.57518604 -0.599335319 -0.2754663703
#> 486   0.52580151 -0.663577768  0.2140434521
#> 487   0.62123318 -0.278218629  0.5478232193
#> 488   0.38589681 -1.016852343 -0.1277949725
#> 489   0.34525839 -0.572479164 -0.4636761067
#> 490   0.33608834 -0.843921391 -0.5142773918
#> 491   0.56644591 -0.683220199 -0.3461736987
#> 492   0.66754043 -0.121183939  0.4697668492
#> 493   0.42976339 -0.643868568  0.1219951515
#> 494   0.67251307 -0.246740055  0.3849273558
#> 495   0.52693852 -0.574746951 -0.2686515480
#> 496   0.51335376 -0.519438680 -0.4788592444
#> 497   0.64315878 -0.592159549  0.4536222151
#> 498   0.42978086 -0.540191528 -0.0764402248
#> 499   0.60113457  0.012241337  0.6189382202
#> 500   0.63290124 -0.621108146  0.2897731128
#> 501   0.51353261 -0.749915386 -0.2669797904
#> 502   0.20642555 -0.965015600 -0.6007139636
#> 503   0.53018556 -0.667533597  0.5926073265
#> 504   0.44665420 -0.108525701  0.8192888071
#> 505   0.46781526 -0.765661371  0.3320947566
#> 506   0.19321985 -1.006000480 -0.5838569131
#> 507   0.78629522 -0.099915866  0.4576807613
#> 508   0.74302857 -0.100777960  0.3089047508
#> 509   0.68247263 -0.197359269  0.2674037168
#> 510   0.59344407 -0.325838814  0.3407141707
#> 511   0.66183233 -0.129254919  0.9552748530
#> 512   0.00332510 -0.472013945 -0.2454043388
#> 513   0.42430971 -0.196756360  0.8247728854
#> 514   0.58147434 -0.286637249  0.8008210177
#> 515   0.45914224 -0.349800853  0.8324760462
#> 516  -0.04924151 -0.944065908 -0.4851977404
#> 517   0.45692775 -1.111199332 -0.5608200762
#> 518   0.55589491 -0.758932604  0.2361324616
#> 519   0.77076946 -0.252813499  0.2769267202
#> 520   0.56282873 -0.497474046  0.7222891994
#> 521   0.21499735 -0.156529558  0.4304317015
#> 522   0.44410752 -0.552683680  0.3979812091
#> 523   0.47165344 -0.045103353  0.7864822783
#> 524   0.35329762 -0.491704065  0.0818226498
#> 525   0.26224629 -1.027078630 -0.4000552127
#> 526   0.45041553 -0.210241064  0.1084850959
#> 527   0.65271652 -0.230404329  0.4568551594
#> 528   0.10357920 -0.166567484  0.3815103227
#> 529   0.57378714 -0.318594150 -0.1411584972
#> 530   0.40905340 -0.693946292 -0.2521328424
#> 531   0.34966327 -0.917275013  0.7778681282
#> 532   0.59344407 -0.325838814  0.3407141707
#> 533  -0.35275148  0.008484225  0.9927128029
#> 534   0.78898285  0.001012107  1.1072669224
#> 535   0.37826356 -0.485962272 -0.0768727288
#> 536   0.61427768 -0.360210681  0.3510780741
#> 537   0.70851556 -0.276933464  0.5389992742
#> 538   0.33637780 -0.146978079  0.5017773677
#> 539   0.64212932 -0.445821488  0.1516983977
#> 540   0.36597463 -0.881160192 -0.5723294763
#> 541   0.38305304 -0.088565508  0.5005878994
#> 542   0.62018837 -0.143520761  0.9493424876
#> 543   0.32964500 -0.530507939  0.3266617801
#> 544   0.51943934 -0.364658698  0.4439684375
#> 545   0.68264515 -0.331270775  0.2405168442
#> 546   0.78898285  0.001012107  1.1072669224
#> 547   0.55989202 -0.147769861  0.5663240205
#> 548   0.60197658 -0.097272623  0.8634771861
#> 549   0.36597463 -0.881160192 -0.5723294763
#> 550   0.67020913 -0.441910240 -0.0844783968
#> 551   0.13963459 -1.098423234 -0.0079046334
#> 552   0.57630922 -0.119613742  0.3419928758
#> 553   0.63430272 -0.300336843  0.1374133536
#> 554   0.60113457  0.012241337  0.6189382202
#> 555   0.36845394 -0.385477827 -0.0942709208
#> 556   0.75949517 -0.255752142  0.0314206999
#> 557   0.30845855 -0.310321059  0.4380804914
#> 558   0.40907590 -0.831980860 -0.3363745945
#> 559   0.65271652 -0.230404329  0.4568551594
#> 560   0.52332269  0.226358321  0.8910578032
#> 561   0.43656833 -0.702918167 -0.4153535748
#> 562   0.39339440 -0.328650296  0.0925544423
#> 563   0.22158697 -0.856743826 -0.0121874619
#> 564   0.33500838 -0.571341379  0.4896415286
#> 565   0.40245756 -0.490873937  0.2051734550
#> 566   0.47915843 -0.083330636  0.8978075604
#> 567   0.36593953 -0.654600127  0.1476434261
#> 568   0.33637780 -0.146978079  0.5017773677
#> 569   0.45945189 -0.453165957 -0.0257440577
#> 570   0.30686201 -0.540057638  0.6968969493
#> 571   0.48545203 -0.502827049  0.2488282115
#> 572   0.48078828 -0.066301557  0.6977623492
#> 573   0.69845368 -0.330286865  0.2331217467
#> 574   0.66509082 -0.536663724 -0.0529563153
#> 575   0.68264515 -0.331270775  0.2405168442
#> 576   0.52329004 -0.548106823 -0.4751994747
#> 577   0.49917180 -0.687066292  0.4265629648
#> 578   0.66655901 -0.582659933  0.6144140489
#> 579   0.59296366 -0.265018598  0.3396277072
#> 580   0.31071617 -0.434640797 -0.2390642015
#> 581   0.21871995  0.073832405  0.7165231124
#> 582   0.57630922 -0.119613742  0.3419928758
#> 583   0.70851556 -0.276933464  0.5389992742
#> 584   0.40728438 -0.958941886 -0.7961588507
#> 585   0.55996177 -0.611314941 -0.0053410996
#> 586   0.65271652 -0.230404329  0.4568551594
#> 587   0.08945413 -0.183026431  0.3459971424
#> 588   0.30825531 -0.239708456  0.4408498717
#> 589   0.34806268 -0.593593655 -0.0005243815
#> 590   0.36531025 -0.318101707  0.1336921403
#> 591   0.77076946 -0.252813499  0.2769267202
#> 592   0.53257261 -0.170131456  0.0986629626
#> 593   0.69179649 -0.135363618  0.6367796057
#> 594   0.57349238 -0.655247959 -0.1680784296
#> 595   0.56565035 -0.484283904 -0.3725226911
#> 596   0.56282873 -0.497474046  0.7222891994
#> 597   0.62750404 -0.367786797  0.8049520137
#> 598   0.45845772 -0.550780988  0.2894405838
#> 599   0.43954515 -0.416393071 -0.0398707563
#> 600   0.83288490 -0.301899502  0.7603006973
#> 601   0.72449050 -0.272773496 -0.2195261369
#> 602   0.60734656 -0.483109968 -0.2345446795
#> 603   0.63258128 -0.518639117  0.4378879676
#> 604   0.39694830 -0.757745042  0.2865092909
#> 605   0.50107153 -0.149462199  0.4239839521
#> 606   0.65310389 -0.165705458  0.2490033063
#> 607   0.49248029 -0.551793977  0.6155890113
#> 608   0.73490192 -0.038866548  0.5704604339
#> 609   0.28924109 -0.200999565  0.5281087178
#> 610   0.41415191 -0.279916062  0.1594191001
#> 611   0.57378714 -0.318594150 -0.1411584972
#> 612   0.73447193 -0.274042958  0.2618683684
#> 613   0.52888389 -0.304879449  0.3684323365
#> 614   0.61770433 -0.431952312 -0.0499418228
#> 615   0.70286802 -0.328452816  0.6315761738
#> 616   0.42976339 -0.643868568  0.1219951515
#> 617   0.47763186 -0.619641635 -0.2564566870
#> 618   0.44042502 -0.588830620 -0.1131725629
#> 619   0.58120899 -0.165629713  0.6842217541
#> 620   0.47235264 -0.466634480  0.0312608888
#> 621   0.51038623 -0.330604925  0.1546399143
#> 622   0.86162403 -0.137965288  0.6970469209
#> 623   0.51353261 -0.749915386 -0.2669797904
#> 624   0.43656833 -0.702918167 -0.4153535748
#> 625   0.59912816 -0.568239990 -0.2021803373
#> 626  -0.25254669 -0.061482459  0.6856658591
#> 627   0.23256414 -0.966840323 -0.6212378190
#> 628   0.73141898 -0.026568937  0.6392492112
#> 629   0.67459540 -0.462056485  0.1984093989
#> 630   0.46438469 -0.199977499  0.8188155010
#> 631   0.40905340 -0.693946292 -0.2521328424
#> 632   0.24648875 -1.111267874  0.2863504167
#> 633   0.22924188 -0.668577405 -0.1171977021
#> 634   0.52806895 -0.241068197  0.1109710872
#> 635   0.67583005  0.213942172  0.9579147732
#> 636   0.48601798 -0.472001227  0.6322807185
#> 637   0.60104328 -0.534232020  0.5117290982
#> 638   0.62750404 -0.367786797  0.8049520137
#> 639   0.42976339 -0.643868568  0.1219951515
#> 640   0.51707955 -0.259296417 -0.2423957832
#> 641   0.45611935 -0.374384106  0.3524557026
#> 642   0.59116428 -0.076659873  0.2104712342
#> 643   0.62188545 -0.667312469 -0.3344145828
#> 644   0.67251307 -0.246740055  0.3849273558
#> 645   0.43508749 -0.196389034  0.2813376374
#> 646   0.57635206 -0.713242571 -0.1709322154
#> 647   0.69179649 -0.135363618  0.6367796057
#> 648   0.65183443 -0.229082285  0.1500651871
#> 649   0.33608834 -0.843921391 -0.5142773918
#> 650   0.58120899 -0.165629713  0.6842217541
#> 651   0.58647183 -0.672798593  0.1514869765
#> 652   0.23256414 -0.966840323 -0.6212378190
#> 653   0.41336349 -1.052730795 -0.6332462762
#> 654   0.49388451 -0.728046564  0.5462357241
#> 655   0.47912944 -0.714055403  0.1536937162
#> 656   0.71651034 -0.425081408 -0.1152228260
#> 657   0.42930150 -0.731740496 -0.5071578555
#> 658   0.70286802 -0.328452816  0.6315761738
#> 659   0.67020913 -0.441910240 -0.0844783968
#> 660   0.53951581 -0.356420991  0.3263492303
#> 661   0.60113457  0.012241337  0.6189382202
#> 662   0.69179649 -0.135363618  0.6367796057
#> 663   0.38499263 -0.586447598  0.2621594360
#> 664   0.61922027 -0.592560414  0.3006645509
#> 665   0.19321985 -1.006000480 -0.5838569131
#> 666   0.53080990 -0.384566301  0.4005459349
#> 667   0.57518604 -0.599335319 -0.2754663703
#> 668   0.22516784 -0.210271791  0.2964773945
#> 669   0.43289908 -0.198669310  0.7221434197
#> 670   0.38991515 -0.213194911  0.1808761032
#> 671   0.52510251 -0.188609925  0.2463381929
#> 672   0.69179649 -0.135363618  0.6367796057
#> 673   0.38011358 -0.896335588 -0.2438200383
#> 674   0.69845368 -0.330286865  0.2331217467
#> 675   0.52592723 -0.109785084  0.9945960122
#> 676   0.49471769 -0.514029461  0.2923724888
#> 677   0.52162548 -0.185368002  0.2817287169
#> 678   0.67251307 -0.246740055  0.3849273558
#> 679   0.47848990 -0.273984758  0.5435843554
#> 680   0.22409499 -0.695907972  0.3389225017
#> 681   0.35027948 -0.158504352  0.5541752992
#> 682   0.55820707 -0.646319758 -0.1598267403
#> 683   0.45914224 -0.349800853  0.8324760462
#> 684   0.40907590 -0.831980860 -0.3363745945
#> 685   0.58073789 -0.625663860 -0.2037501953
#> 686   0.76904297 -0.160051214  0.2771064424
#> 687   0.05754491 -0.019603197  0.6632473651
#> 688   0.74076906  0.015441073  0.4282430822
#> 689   0.67197237 -0.494489806 -0.3972073265
#> 690   0.46746389 -0.500978711  0.2314276293
#> 691   0.57382598 -0.823103148 -0.4855075456
#> 692   0.59250294 -0.618189314 -0.2248972977
#> 693   0.51424615 -0.576625358  0.4106730234
#> 694   0.56952143 -0.399401365  0.3936602166
#> 695   0.82695519  0.113936874  0.7159921016
#> 696   0.19441331 -0.888706644  0.1221067768
#> 697   0.38029121 -0.210460319  0.2680926635
#> 698   0.53863873 -0.259109121  0.4708994146
#> 699   0.38357233 -0.230568574  0.6919618483
#> 700   0.46434655 -0.508728884 -0.0807735858
#> 701   0.62188545 -0.667312469 -0.3344145828
#> 702   0.51356862 -0.515182708  0.1720116185
#> 703   0.46073781 -0.508579262 -0.0354347778
#> 704   0.48601798 -0.472001227  0.6322807185
#> 705   0.41336349 -1.052730795 -0.6332462762
#> 706   0.52735862 -0.128242015  0.5627272508
#> 707   0.46574431 -0.501593164  0.0208127942
#> 708   0.55898460 -0.144649284  0.2188104278
#> 709   0.32507885  0.054981660  0.8130035278
#> 710   0.62468095 -0.580171964 -0.1923786213
#> 711   0.53469475 -0.084122627  0.6660449640
#> 712   0.73228360 -0.325164312  0.1891457233
#> 713   0.47235264 -0.466634480  0.0312608888
#> 714   0.43508749 -0.196389034  0.2813376374
#> 715   0.66330688 -0.361861155  0.1691612358
#> 716   0.61854494 -0.674162440 -0.3572291476
#> 717   0.41362520 -0.467663064 -0.1741187617
#> 718   0.66315344 -0.343686635  0.3662197864
#> 719   0.43753208 -0.794898805 -0.3078075334
#> 720   0.76904297 -0.160051214  0.2771064424
#> 721   0.49342340 -0.485285635  0.1846447826
#> 722   0.42112737 -0.348979758  0.1424409934
#> 723   0.77076946 -0.252813499  0.2769267202
#> 724   0.32786466 -0.428110788  0.1943186748
#> 725   0.58073789 -0.625663860 -0.2037501953
#> 726   0.50887063 -0.263287401  0.8703694132
#> 727   0.67020913 -0.441910240 -0.0844783968
#> 728  -0.49415636 -0.316049481  0.3751173689
#> 729   0.45157833 -0.690747068  0.1197098546
#> 730   0.40890341 -0.634516267 -0.3976323361
#> 731   0.41971058 -0.506023512  0.1048050507
#> 732   0.51335376 -0.519438680 -0.4788592444
#> 733   0.64212932 -0.445821488  0.1516983977
#> 734   0.65581111 -0.255229561  0.4208370226
#> 735   0.52973542 -0.441271942 -0.0971984542
#> 736   0.67583005  0.213942172  0.9579147732
#> 737   0.73228360 -0.325164312  0.1891457233
#> 738   0.67251307 -0.246740055  0.3849273558
#> 739   0.63430272 -0.300336843  0.1374133536
#> 740   0.62468095 -0.580171964 -0.1923786213
#> 741   0.42978086 -0.540191528 -0.0764402248
#> 742   0.68174656 -0.088777230  0.6843814912
#> 743   0.60027977 -0.139182262  0.2078853110
#> 744   0.55820707 -0.646319758 -0.1598267403
#> 745   0.80326198 -0.156283800  0.0832953067
#> 746   0.28176851 -0.568348274 -0.0412891484
#> 747   0.44030223 -0.036766210  0.7489682926
#> 748   0.55620053 -0.234701189 -0.0151571311
#> 749   0.43598906 -0.607843793  0.6433279939
#> 750   0.63456368  0.090259175  0.7443703479
#> 751   0.15170465 -1.465708682 -0.5689957212
#> 752   0.66082456 -0.429019540  0.7510064734
#> 753   0.14534481 -0.925316731 -0.1634671235
#> 754   0.41025062 -0.560151636  0.0679375527
#> 755   0.31689400 -0.616606245 -0.0910392823
#> 756   0.57518604 -0.599335319 -0.2754663703
#> 757   0.67758453 -0.372224024 -0.0325482973
#> 758   0.61445001 -0.499566077 -0.2774436370
#> 759   0.40772712 -0.347778723 -0.2208212145
#> 760   0.53835772 -0.696184039  0.1109625806
#> 761   0.51918460 -0.411445411  0.3660778289
#> 762   0.27674268 -0.302675221  0.3437770304
#> 763   0.67197237 -0.494489806 -0.3972073265
#> 764   0.67961406  0.362509710  1.3214598978
#> 765   0.63215097 -0.529462819 -0.4945847455
#> 766  -0.04394610 -1.086673016 -0.2104630690
#> 767   0.54998842 -0.094139319  0.8320065538
#> 768   0.46434655 -0.508728884 -0.0807735858
#> 769   0.52182952  0.005454811  0.4270779249
#> 770   0.68174656 -0.088777230  0.6843814912
#> 771   0.09311741 -0.453542461  0.1699841862
#> 772   0.59542100 -0.001113551  0.6126137672
#> 773   0.66150335 -0.192789324  0.3807427047
#> 774   0.50887063 -0.263287401  0.8703694132
#> 775   0.86364241 -0.373500059 -0.0704657692
#> 776   0.58114911 -0.190652501  0.2321163542
#> 777   0.63258128 -0.518639117  0.4378879676
#> 778   0.65183443 -0.229082285  0.1500651871
#> 779   0.74076906  0.015441073  0.4282430822
#> 780   0.66112718 -0.445022252 -0.1040301594
#> 781   0.18766060 -1.033107670 -0.6072628141
#> 782   0.40334569 -0.965492078 -0.5131127457
#> 783   0.73447193 -0.274042958  0.2618683684
#> 784   0.53951581 -0.356420991  0.3263492303
#> 785   0.65938607 -0.647289921 -0.1935366886
#> 786  -0.70288356  0.065357348  0.5574463191
#> 787   0.82004740 -0.085312108  0.4844790870
#> 788   0.46266510 -0.160586563  0.2628545603
#> 789   0.69845368 -0.330286865  0.2331217467
#> 790   0.63712130 -0.238939202  0.6521093917
#> 791   0.68180984 -0.361863457 -0.0149816138
#> 792   0.14568110 -0.687496379 -0.1911031397
#> 793   0.36728556 -0.274688182  0.1525805271
#> 794   0.66183233 -0.129254919  0.9552748530
#> 795   0.73141898 -0.026568937  0.6392492112
#> 796   0.77141088 -0.172584178  0.0223218495
#> 797   0.44481863 -0.185067250  0.8487202249
#> 798   0.58972129  0.009086453  0.2820798081
#> 799   0.52182952  0.005454811  0.4270779249
#> 800   0.66106754  0.100464993  0.8966364569
#> 801   0.44178041 -0.345810474  0.2439199598
#> 802   0.59344407 -0.325838814  0.3407141707
#> 803   0.45945189 -0.453165957 -0.0257440577
#> 804   0.66150335 -0.192789324  0.3807427047
#> 805   0.68264515 -0.331270775  0.2405168442
#> 806   0.20904317  0.058818152  0.5473078656
#> 807   0.50107153 -0.149462199  0.4239839521
#> 808   0.72224186 -0.257351534 -0.0263991409
#> 809   0.66112718 -0.445022252 -0.1040301594
#> 810   0.65884031 -0.198144396  0.0388580452
#> 811   0.59786709 -0.237560134  0.1017409626
#> 812   0.45845772 -0.550780988  0.2894405838
#> 813   0.64690436 -0.443321818 -0.1354354997
#> 814   0.55996177 -0.611314941 -0.0053410996
#> 815   0.58144095 -0.283853892  0.1607819970
#> 816   0.44042502 -0.588830620 -0.1131725629
#> 817   0.73447193 -0.274042958  0.2618683684
#> 818   0.54203072 -0.284514266  0.2841830218
#> 819   0.42374374 -0.910877066 -0.2458211309
#> 820   0.68264515 -0.331270775  0.2405168442
#> 821   0.66928048 -0.160880796  0.3613130535
#> 822   0.62018837 -0.143520761  0.9493424876
#> 823   0.60484750 -0.276688792  0.1983620796
#> 824   0.47235264 -0.466634480  0.0312608888
#> 825   0.47753228 -0.482116028  0.2887315354
#> 826   0.22376153 -0.516406081 -0.0248529068
#> 827   0.57349238 -0.655247959 -0.1680784296
#> 828   0.76082638 -0.337575300  0.4338458301
#> 829  -0.24616530 -0.455427750  0.6834472272
#> 830   0.84537657 -0.239021985 -0.0959132892
#> 831   0.48545203 -0.502827049  0.2488282115
#> 832   0.62018837 -0.143520761  0.9493424876
#> 833   0.45723661 -0.410422143 -0.1166285591
#> 834   0.46996962 -0.233200888  0.8135257791
#> 835   0.75975709 -0.431284798  0.0573373635
#> 836   0.44410752 -0.552683680  0.3979812091
#> 837   0.38722652 -0.253963293  0.1168916300
#> 838   0.74834779 -0.113059170  0.3508632565
#> 839   0.63395563 -0.355399521  0.8961107275
#> 840   0.61287648 -0.099785917  0.3113042432
#> 841   0.46271274 -0.359606200 -0.1990240936
#> 842   0.51353261 -0.749915386 -0.2669797904
#> 843   0.64041802 -0.025780051  0.2687034623
#> 844   0.35971714 -0.218759163  0.2651943815
#> 845   0.44042502 -0.588830620 -0.1131725629
#> 846   0.22949452 -0.352034376  0.1425878641
#> 847   0.94081112  0.060831142  0.6777098994
#> 848   0.73228360 -0.325164312  0.1891457233
#> 849   0.53255235 -0.628631127  0.5217400023
#> 850   0.51718515 -0.178534226  1.1188431388
#> 851   0.74537906 -0.541980900  0.1339822962
#> 852   0.47277085 -0.267270801  1.0611101545
#> 853   0.23942977 -0.529774887 -0.3303106246
#> 854   0.07686125  0.317806792  1.1073418294
#> 855   0.47763186 -0.619641635 -0.2564566870
#> 856   0.60807988 -0.509715306 -0.1364509292
#> 857   0.57518604 -0.599335319 -0.2754663703
#> 858   0.54998842 -0.094139319  0.8320065538
#> 859   0.46618405 -0.697019169 -0.4474541247
#> 860   0.73490192 -0.038866548  0.5704604339
#> 861   0.06907425 -0.194239162  0.3929290097
#> 862   0.55602076 -0.778296571 -0.4039044867
#> 863   0.44705470 -0.215198171  0.1370369890
#> 864   0.74371885 -0.173186878  0.4532774658
#> 865   0.68247263 -0.197359269  0.2674037168
#> 866   0.66845854 -0.211990824  0.5455203280
#> 867   0.53255235 -0.628631127  0.5217400023
#> 868   0.61770433 -0.431952312 -0.0499418228
#> 869  -0.22626625 -0.574435952 -0.4456757025
#> 870   0.43239448 -0.646537362  0.3489184472
#> 871   0.31933982 -0.660146187  0.3247310159
#> 872   0.38011358 -0.896335588 -0.2438200383
#> 873   0.61864615 -0.285542433  0.3485217047
#> 874   0.46073781 -0.508579262 -0.0354347778
#> 875   0.73228360 -0.325164312  0.1891457233
#> 876   0.40287405 -0.792498542 -0.6175040523
#> 877   0.54687660 -0.514449989 -0.1213013334
#> 878   0.60104328 -0.534232020  0.5117290982
#> 879   0.44410752 -0.552683680  0.3979812091
#> 880   0.59268270 -0.023315489  0.7502454201
#> 881   0.39070795 -0.572178428 -0.4214799472
#> 882   0.71651034 -0.425081408 -0.1152228260
#> 883   0.61286654 -0.601388066  1.0428023672
#> 884   0.31035284 -0.397226378  0.1941247829
#> 885   0.73228360 -0.325164312  0.1891457233
#> 886   0.32056860 -0.034333722  1.3376281278
#> 887   0.61854494 -0.674162440 -0.3572291476
#> 888   0.78629522 -0.099915866  0.4576807613
#> 889   0.45157833 -0.690747068  0.1197098546
#> 890   0.62750404 -0.367786797  0.8049520137
#> 891   0.46574431 -0.501593164  0.0208127942
#> 892   0.23821341 -0.684405295 -0.2251562992
#> 893   0.17630092 -0.506108172 -0.0100774422
#> 894   0.53625391 -0.112431078  0.5032458060
#> 895   0.66811447  0.156523054  0.8663051699
#> 896   0.44178041 -0.345810474  0.2439199598
#> 897   0.31105570 -0.060562465  0.8476865897
#> 898   0.52693852 -0.574746951 -0.2686515480
#> 899   0.62188545 -0.667312469 -0.3344145828
#> 900   0.60807988 -0.509715306 -0.1364509292
#> 901   0.62123318 -0.278218629  0.5478232193
#> 902   0.41971058 -0.506023512  0.1048050507
#> 903   0.48860467 -0.301449578  0.4730922367
#> 904   0.47753228 -0.482116028  0.2887315354
#> 905   0.65446865 -0.180552892  0.1899478759
#> 906   0.67894002 -0.392269137 -0.3127634974
#> 907   0.65581111 -0.255229561  0.4208370226
#> 908   0.32074108 -0.304796831  0.4065581756
#> 909   0.60807988 -0.509715306 -0.1364509292
#> 910   0.58147434 -0.286637249  0.8008210177
#> 911   0.82004740 -0.085312108  0.4844790870
#> 912   0.49471769 -0.514029461  0.2923724888
#> 913   0.73104654 -0.207175394 -0.0697522826
#> 914   0.32786466 -0.428110788  0.1943186748
#> 915   0.44233341 -0.686676889  0.1299440210
#> 916   0.35925676 -0.645028032  0.4134734080
#> 917   0.66928048 -0.160880796  0.3613130535
#> 918   0.54687660 -0.514449989 -0.1213013334
#> 919   0.36181968 -0.345252662  0.1712234178
#> 920   0.37432686 -0.073152915  0.7712594214
#> 921   0.34475365 -0.555275862  0.0958381558
#> 922   0.73447193 -0.274042958  0.2618683684
#> 923   0.52735862 -0.128242015  0.5627272508
#> 924   0.63712130 -0.238939202  0.6521093917
#> 925   0.67459540 -0.462056485  0.1984093989
#> 926   0.51828561 -0.372742120  0.6154481562
#> 927   0.51918460 -0.411445411  0.3660778289
#> 928   0.53080990 -0.384566301  0.4005459349
#> 929   0.66315344 -0.343686635  0.3662197864
#> 930   0.01935248 -0.369214498 -0.0744015043
#> 931   0.61273156 -0.247283837  0.2859387279
#> 932   0.51680669 -0.699913448  0.0408541512
#> 933   0.55854437 -0.472500561  0.1208543570
#> 934   0.94081112  0.060831142  0.6777098994
#> 935   0.73490192 -0.038866548  0.5704604339
#> 936   0.66315344 -0.343686635  0.3662197864
#> 937   0.56952143 -0.399401365  0.3936602166
#> 938   0.61864615 -0.285542433  0.3485217047
#> 939   0.52592723 -0.109785084  0.9945960122
#> 940   0.52236848 -0.578739697  0.4272298197
#> 941   0.29744959 -0.532413488 -0.1143364566
#> 942   0.69159927  0.279152372  0.7508346866
#> 943   0.54473693 -0.105484446  0.3985105680
#> 944   0.78629522 -0.099915866  0.4576807613
#> 945   0.49569253 -0.136214712  0.2037082113
#> 946   0.39858234 -0.632559034 -0.0039399641
#> 947   0.95059943  0.122311446  0.4439480826
#> 948   0.20642555 -0.965015600 -0.6007139636
#> 949   0.33093249 -0.204614919  0.2307376478
#> 950   0.36652052 -0.431620185  0.0402645690
#> 951   0.77076946 -0.252813499  0.2769267202
#> 952   0.69555273 -0.469519514 -0.3187226131
#> 953   0.61770433 -0.431952312 -0.0499418228
#> 954   0.54634499 -0.111890652  0.5142059945
#> 955   0.53320358 -0.262588643  1.2378985638
#> 956   0.14228647 -1.269296241 -0.5676428803
#> 957   0.70851556 -0.276933464  0.5389992742
#> 958   0.59344407 -0.325838814  0.3407141707
#> 959   0.66928048 -0.160880796  0.3613130535
#> 960   0.42374374 -0.910877066 -0.2458211309
#> 961   0.51356862 -0.515182708  0.1720116185
#> 962   0.35399943 -0.606767587  0.0823793042
#> 963   0.67251307 -0.246740055  0.3849273558
#> 964   0.60315736 -0.347693841  1.2835391627
#> 965   0.44665420 -0.108525701  0.8192888071
#> 966   0.42536820 -0.539123204  0.6256927416
#> 967   0.22830249 -0.284908812  0.2858114906
#> 968  -0.02256512 -0.328384309  0.1368193065
#> 969  -0.14541636 -0.071040382  0.2859989555
#> 970   0.17193364 -0.843104126 -0.1497062009
#> 971   0.41415191 -0.279916062  0.1594191001
#> 972   0.32964500 -0.530507939  0.3266617801
#> 973   0.46436078 -0.380078894  0.2323239842
#> 974   0.42769875 -0.435864791  0.1403957364
#> 975   0.60807988 -0.509715306 -0.1364509292
#> 976   0.91607686 -0.106047936  0.3961074671
#> 977   0.79806929 -0.359968485 -0.2159391388
#> 978   0.56794149 -0.267882744 -0.0413174951
#> 979   0.45945189 -0.453165957 -0.0257440577
#> 980   0.24779203 -0.306454824  0.2254912407
#> 981   0.58144095 -0.283853892  0.1607819970
#> 982   0.41336349 -1.052730795 -0.6332462762
#> 983   0.15608877 -0.811970053 -0.5387191583
#> 984   0.53951581 -0.356420991  0.3263492303
#> 985   0.52592723 -0.109785084  0.9945960122
#> 986   0.37919268 -0.673342831  0.0705044648
#> 987   0.59264471 -0.741609085  0.0125900592
#> 988   0.65183443 -0.229082285  0.1500651871
#> 989   0.78629522 -0.099915866  0.4576807613
#> 990   0.36597463 -0.881160192 -0.5723294763
#> 991   0.65938607 -0.647289921 -0.1935366886
#> 992   0.65271652 -0.230404329  0.4568551594
#> 993   0.70977693 -0.310383551  0.7017925862
#> 994   0.44042502 -0.588830620 -0.1131725629
#> 995   0.61854494 -0.674162440 -0.3572291476
#> 996   0.48948467 -0.526692942  0.3503202568
#> 997   0.53255235 -0.628631127  0.5217400023
#> 998   0.26967049 -0.446200365  0.2455337768
#> 999   0.55620053 -0.234701189 -0.0151571311
#> 1000  0.59422521 -0.361307024 -0.0949291618
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>    <dbl>
#>  1      1        4       70        1  0.747 -0.0464  0.815  
#>  2      2        4       70        1  0.775 -0.316  -0.248  
#>  3      3        4       70        1  0.682 -0.197   0.267  
#>  4      4        4       70        1  0.384 -0.231   0.692  
#>  5      5        4       70        1  0.460 -0.274   0.00983
#>  6      6        4       70        1  0.396 -0.816   0.124  
#>  7      7        4       70        1  0.176 -0.506  -0.0101 
#>  8      8        4       70        1  0.156 -0.812  -0.539  
#>  9      9        4       70        1  0.625 -0.580  -0.192  
#> 10     10        4       70        1  0.692 -0.135   0.637  
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc     depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>     <dbl> <dbl>
#> 1      1     1     4    70     1  8.44  66.8 2.26  0.126 22.7   3.88e+ 2 1517.
#> 2      1    14     4    70     1  8.44  66.8 2.26  0.126  5.58 -2.53e-10  373.
#> 3      2     1     4    70     1  8.68  51.0 0.781 0.170 16.0   1.12e+ 3  819.
#> 4      2    14     4    70     1  8.68  51.0 0.781 0.170  4.83  4.38e- 2  246.
#> 5      3     1     4    70     1  7.92  57.5 1.31  0.138 20.2   7.64e+ 2 1163.
#> 6      3    14     4    70     1  7.92  57.5 1.31  0.138  5.86  3.21e- 5  336.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#>