Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e3,n_resample=1e2)
#> $eta
#>          ETA_Cl      ETA_Vc      ETA_Ka
#> 1    0.44852256 -0.29386695  0.25508142
#> 2    0.62575211 -0.12598185  0.39581785
#> 3    0.59394319 -0.61283467  0.44309473
#> 4    0.40840374 -0.11424231  0.88523822
#> 5    0.39724995 -0.81841411  0.03467140
#> 6    0.39724995 -0.81841411  0.03467140
#> 7    0.63383635 -0.09208788  0.75313256
#> 8    0.56503241 -0.38163429  0.57870438
#> 9    0.39724995 -0.81841411  0.03467140
#> 10   0.63383635 -0.09208788  0.75313256
#> 11   0.68819786 -0.43456690  0.60967953
#> 12   0.70134619 -0.25923790  0.36886223
#> 13   0.57930142 -0.19223329  0.36675144
#> 14   0.48137467 -0.64077946  0.37626331
#> 15   0.62575211 -0.12598185  0.39581785
#> 16   0.70134619 -0.25923790  0.36886223
#> 17   0.48137467 -0.64077946  0.37626331
#> 18   0.56600985 -0.62603555  0.27337490
#> 19   0.53665236 -0.51802076  0.20000937
#> 20   0.45547826 -0.17371863  0.17395510
#> 21   0.66424372 -0.24149717  0.47234145
#> 22   0.57040379 -0.01801166  0.57500526
#> 23   0.55291123 -0.28819509  0.20794277
#> 24   0.67968027 -0.21978421 -0.09310346
#> 25   0.52795909 -0.54265204  0.44086921
#> 26   0.70134619 -0.25923790  0.36886223
#> 27   0.47328909 -0.58699755 -0.04404405
#> 28   0.52795909 -0.54265204  0.44086921
#> 29   0.45793620 -0.19267881  0.41387686
#> 30   0.25327703  0.03343085  0.50612767
#> 31   0.52674334 -0.85568939 -0.49246147
#> 32   0.50199137 -0.62194414 -0.24232686
#> 33   0.59394319 -0.61283467  0.44309473
#> 34   0.31373832 -0.09809849  0.57244114
#> 35   0.18632569 -0.89355390  0.19705018
#> 36   0.63383635 -0.09208788  0.75313256
#> 37   0.53665236 -0.51802076  0.20000937
#> 38   0.45722782 -0.71688383  0.07553748
#> 39  -0.05336188 -0.69167479  0.14448127
#> 40   0.45793620 -0.19267881  0.41387686
#> 41   0.51192193 -0.27347692  0.10202829
#> 42   0.63493108 -0.37178185  0.16556913
#> 43   0.30917621 -0.14303649  0.80466715
#> 44   0.38975992 -0.57378467  0.13013143
#> 45   0.82675130 -0.28968964  0.69288199
#> 46   0.47328909 -0.58699755 -0.04404405
#> 47   0.63383635 -0.09208788  0.75313256
#> 48   0.47328909 -0.58699755 -0.04404405
#> 49   0.63844449 -0.62540066 -0.51541830
#> 50   0.68819786 -0.43456690  0.60967953
#> 51   0.55291123 -0.28819509  0.20794277
#> 52   0.51192193 -0.27347692  0.10202829
#> 53   0.39724995 -0.81841411  0.03467140
#> 54   0.63493108 -0.37178185  0.16556913
#> 55   0.44992593 -0.32358339  0.03937631
#> 56   0.48137467 -0.64077946  0.37626331
#> 57   0.53665236 -0.51802076  0.20000937
#> 58   0.52795909 -0.54265204  0.44086921
#> 59   0.11813435 -0.18505916  0.93527226
#> 60   0.25020257 -0.74575927  0.09959267
#> 61   0.52842062 -0.15474910  0.34863771
#> 62   0.69634207 -0.07639143  0.66612093
#> 63   0.52674334 -0.85568939 -0.49246147
#> 64   0.49828543 -0.31856056  0.27567836
#> 65   0.42622606 -0.54409046  0.44775164
#> 66   0.15901163 -1.18927775 -0.57918034
#> 67   0.17490942 -1.10956435 -0.45372110
#> 68   0.69634207 -0.07639143  0.66612093
#> 69   0.14302569 -1.10172538 -0.52271669
#> 70   0.52674334 -0.85568939 -0.49246147
#> 71   0.48033027 -0.64682978 -0.02260275
#> 72   0.15901163 -1.18927775 -0.57918034
#> 73   0.62575211 -0.12598185  0.39581785
#> 74   0.49828543 -0.31856056  0.27567836
#> 75   0.36795919 -0.75276468 -0.06179420
#> 76   0.69634207 -0.07639143  0.66612093
#> 77   0.54905702 -0.14764846  0.04712659
#> 78   0.44852256 -0.29386695  0.25508142
#> 79   0.32465620 -0.39749854 -0.07578372
#> 80   0.63493108 -0.37178185  0.16556913
#> 81   0.16422777 -0.25200611  0.13971027
#> 82   0.45722782 -0.71688383  0.07553748
#> 83   0.57040379 -0.01801166  0.57500526
#> 84   0.47328909 -0.58699755 -0.04404405
#> 85   0.67968027 -0.21978421 -0.09310346
#> 86   0.45722782 -0.71688383  0.07553748
#> 87   0.42474687  0.14418013  0.59865979
#> 88   0.69634207 -0.07639143  0.66612093
#> 89   0.52674334 -0.85568939 -0.49246147
#> 90   0.42607558 -0.89032711  0.86664143
#> 91   0.45793620 -0.19267881  0.41387686
#> 92   0.51192193 -0.27347692  0.10202829
#> 93   0.69634207 -0.07639143  0.66612093
#> 94   0.68819786 -0.43456690  0.60967953
#> 95   0.63493108 -0.37178185  0.16556913
#> 96   0.63493108 -0.37178185  0.16556913
#> 97   0.47874096 -0.35202249 -0.06691333
#> 98   0.47874096 -0.35202249 -0.06691333
#> 99   0.36795919 -0.75276468 -0.06179420
#> 100  0.69634207 -0.07639143  0.66612093
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1  0.449 -0.294  0.255 
#>  2      2        4       70        1  0.626 -0.126  0.396 
#>  3      3        4       70        1  0.594 -0.613  0.443 
#>  4      4        4       70        1  0.408 -0.114  0.885 
#>  5      5        4       70        1  0.397 -0.818  0.0347
#>  6      6        4       70        1  0.397 -0.818  0.0347
#>  7      7        4       70        1  0.634 -0.0921 0.753 
#>  8      8        4       70        1  0.565 -0.382  0.579 
#>  9      9        4       70        1  0.397 -0.818  0.0347
#> 10     10        4       70        1  0.634 -0.0921 0.753 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  6.26  52.2  1.29 0.120 22.3      7.73e+2 1163.
#> 2      1    14     4    70     1  6.26  52.2  1.29 0.120  8.11     4.00e-5  423.
#> 3      2     1     4    70     1  7.48  61.7  1.49 0.121 20.4      6.72e+2 1258.
#> 4      2    14     4    70     1  7.48  61.7  1.49 0.121  6.67     2.75e-6  412.
#> 5      3     1     4    70     1  7.24  37.9  1.56 0.191 32.9      6.38e+2 1249.
#> 6      3    14     4    70     1  7.24  37.9  1.56 0.191  4.35     1.03e-6  165.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>