Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e4,n_resample=1e3)
#> $eta
#>             ETA_Cl        ETA_Vc       ETA_Ka
#> 1     0.2852706593 -0.7362776816 -0.400417391
#> 2     0.3394036389 -0.9618620094 -0.529977458
#> 3     0.8035919039 -0.2700850673  0.406780480
#> 4     0.6750166909 -0.0089994803  1.029404799
#> 5     0.4682236516 -0.6110707445 -0.298070241
#> 6     0.5603886390 -0.7128409421 -0.499113938
#> 7     0.5393431897 -0.6952058358 -0.101491150
#> 8     0.2057607693 -1.1422175214 -0.321910953
#> 9     0.1685488779  0.2593287144  0.968984282
#> 10    0.6750166909 -0.0089994803  1.029404799
#> 11    0.6886193656 -0.1359456839  0.332723710
#> 12    0.3720555968 -0.4954469211  0.327433336
#> 13    0.3273808098 -0.4789252570  0.594276593
#> 14    0.4760573521 -0.5781506531  0.002424496
#> 15    0.5339662086 -0.5547411643 -0.053606088
#> 16    0.1825802089 -0.3142558795  0.016955745
#> 17    0.5969374205  0.1114353728  0.692552825
#> 18    0.7363084171 -0.3578733712  0.211604109
#> 19    0.5281331091  0.0753946414  0.783164566
#> 20    0.5412051355 -0.4069637129  0.168800560
#> 21    0.5685695014 -0.6600854461  0.203285139
#> 22    0.5998127557 -0.5562235057  0.240571454
#> 23    0.3473417140 -0.5019526879 -0.021721026
#> 24    0.4514432516 -0.1646137774  0.223040354
#> 25    0.5695268578 -0.5878518944  0.055316082
#> 26    0.4869369442 -0.6483198389 -0.055597940
#> 27    0.8192849638 -0.0988214241  0.322648194
#> 28    0.7162084075 -0.3102480553  0.045184720
#> 29    0.4696308629 -0.3311784915  0.579710925
#> 30    0.1705322975 -0.1781978998  0.007762943
#> 31    0.9196827735  0.0066097657  0.304504401
#> 32    0.6118952377 -0.3378488093  0.442344335
#> 33    0.5455244416 -0.2955400583  0.138215479
#> 34    0.7408890566 -0.3508074064  0.144613759
#> 35    0.5453378859 -0.7288587614 -0.086115523
#> 36    0.2968229859 -0.9398413399 -0.374008128
#> 37    0.1517627561 -0.0386119157  0.390331203
#> 38    0.5000431348 -0.3830180801  0.101331134
#> 39    0.2207171434 -0.1567778212  0.609569420
#> 40    0.7066825123 -0.2838685166  0.108879210
#> 41   -0.0251613103 -0.7628314557 -0.327957901
#> 42    0.6873848770  0.2890419335  0.968018068
#> 43    0.8642959456  0.0486905588  0.950655162
#> 44    0.4233143556 -0.5935703105 -0.520646639
#> 45    0.4627676361 -0.1202113317  0.452418914
#> 46    0.6210965139  0.3494808140  1.565287826
#> 47    0.6531345044 -0.5016929550 -0.040291997
#> 48    0.4351980203 -0.1965224458  0.709355156
#> 49    0.4553942684 -0.8537614662 -0.125882944
#> 50    0.6998083868 -0.2291333088  0.295685435
#> 51    0.3001391366 -0.5369265245 -0.232608675
#> 52    0.6860690351 -0.0702435709  0.289950219
#> 53    0.7511122500 -0.2074193049  0.489367009
#> 54    0.6150808370 -0.5238562767  0.257365615
#> 55    0.5504524797 -0.7053095951 -0.299875336
#> 56    0.3741764924 -0.2357190411  0.996785549
#> 57    0.5982871028 -0.5379908450  0.276593459
#> 58    0.3894827716  0.0448044299  0.740668013
#> 59    0.6788737888 -0.4659663463  0.060461962
#> 60    0.3971202179  0.0641184204  0.685400345
#> 61    0.4627676361 -0.1202113317  0.452418914
#> 62    0.3870496521 -0.4335232210  0.201116517
#> 63    0.6390802861  0.0428583982  0.969644676
#> 64    0.8440032513  0.2049313264  1.059534051
#> 65    0.7535859285 -0.2738703004  0.267589472
#> 66    0.2999303908 -0.4187707755  0.261467480
#> 67    0.6830207182 -0.7340166631 -0.696763804
#> 68    0.7014742625  0.2322150742  0.852809385
#> 69    0.8369679009  0.0530205578  0.519416219
#> 70    0.7604628835 -0.4574423793 -0.139094486
#> 71    0.1639024184  0.0148592663  0.694029211
#> 72    0.7728673113  0.1304022482  1.032643884
#> 73    0.4403478353 -0.4748823127 -0.020608811
#> 74    0.5951906572 -0.3050537908  0.054423749
#> 75    0.5281242049 -0.4608060120  0.084935797
#> 76    0.2327549686 -0.1717085671  0.281174287
#> 77    0.5877669944 -0.2395677554  0.778132023
#> 78    0.0977965949 -0.6467467106  0.578810723
#> 79    0.8407655267 -0.0979668276  1.111071718
#> 80    0.5799867686 -0.6579623889 -0.172411031
#> 81    0.6627130591 -0.2934735254  0.384453148
#> 82    0.5814730760 -0.2498598412  0.382395578
#> 83    0.5525076671 -0.6810873697 -0.152299920
#> 84    0.2429289218 -0.6725946678 -0.006647113
#> 85    0.5391846205 -0.7174033030 -0.025594060
#> 86    0.1851531535  0.0810307739  0.769993944
#> 87    0.5291002550 -0.4489777412  0.623927720
#> 88    0.8232018580 -0.0051218626  0.782433057
#> 89    0.1920670985 -0.6201249020 -0.014906847
#> 90    0.5969374205  0.1114353728  0.692552825
#> 91    0.6660111640 -0.1697013758  1.011927759
#> 92    0.6956106630 -0.3670402080  0.144781471
#> 93    0.1320145344 -0.3420909853  0.352879102
#> 94    0.8231117050 -0.0883379250  0.407171124
#> 95    0.5139634076 -0.4104620956  0.154777553
#> 96    0.5838075560 -0.6287640836 -0.404894381
#> 97    0.4365715082 -0.3014904378  0.345112336
#> 98    0.3444543564 -0.2288611226  0.249239793
#> 99    0.3599669860  0.0332337489  1.039563356
#> 100   0.5910724797 -0.4260331119 -0.264763971
#> 101   0.7019818542 -0.4263125341  0.256955080
#> 102   0.4899241114 -0.4455822410  0.284259884
#> 103   0.5312233214 -0.6458712396  0.102473171
#> 104   0.5525076671 -0.6810873697 -0.152299920
#> 105   0.4025335892 -0.5806713971  0.031949388
#> 106   0.6097714172 -0.3442291847  0.264689536
#> 107   0.6886193656 -0.1359456839  0.332723710
#> 108   0.6115525812 -0.1057183121  0.218130272
#> 109   0.6531345044 -0.5016929550 -0.040291997
#> 110   0.5393431897 -0.6952058358 -0.101491150
#> 111   0.8041367959  0.0841328847  0.219604026
#> 112   0.2239042665 -0.2914225492  0.598381614
#> 113   0.5998301523 -0.3247869073  0.505965392
#> 114   0.2097880671 -0.5259732350  0.072113309
#> 115   0.5412592910 -0.5221638577  0.996839908
#> 116   0.2648973198 -0.2079407118  0.414784805
#> 117   0.5412592910 -0.5221638577  0.996839908
#> 118   0.2755660500 -0.8671739256 -0.139597968
#> 119   0.3411213166 -0.7182881882 -0.212990965
#> 120   0.4669017352 -0.6765303525 -0.182865403
#> 121   0.3870948471 -0.4514233667  0.110156438
#> 122   0.2575220456 -0.7147002488  0.121005761
#> 123   0.6489448615 -0.2264088114  0.953738205
#> 124   0.5453378859 -0.7288587614 -0.086115523
#> 125   0.5504524797 -0.7053095951 -0.299875336
#> 126   0.2656507680 -0.7547063839  0.655905222
#> 127   0.3394036389 -0.9618620094 -0.529977458
#> 128   0.6750166909 -0.0089994803  1.029404799
#> 129   0.5868943481 -0.1415866715  0.220591216
#> 130   0.4810060762 -0.3692849963  0.289283635
#> 131   0.7636930391 -0.0430404608  0.189951134
#> 132   0.7753085246 -0.3062136184 -0.021202704
#> 133   0.4979648902 -0.2104491199  0.566378861
#> 134   0.4365715082 -0.3014904378  0.345112336
#> 135   0.5165948964 -0.4682456414  0.419465853
#> 136   0.4056700265  0.0537298371  0.119001712
#> 137   0.6150808370 -0.5238562767  0.257365615
#> 138   0.5462259745 -0.3478552055  0.380420054
#> 139   0.2707258870 -0.9031605763 -0.187615352
#> 140   0.3741764924 -0.2357190411  0.996785549
#> 141   0.6454816059 -0.0327136624  0.294480006
#> 142   0.0859484706 -0.5593335382  0.464885949
#> 143   0.7567408467 -0.1839804420  0.797065268
#> 144   0.4749891271 -0.4094230498 -0.084942427
#> 145   0.5307787319 -0.3153362569  0.238584941
#> 146   0.6098232444 -0.4963677953  0.132364734
#> 147   0.4270618647 -0.7864028689  0.055775936
#> 148   0.5046632077 -0.2164942850  0.397390717
#> 149   0.4767486413 -0.7267912408 -0.132796639
#> 150   0.5453378859 -0.7288587614 -0.086115523
#> 151   0.5113725485 -0.5581640745  0.007409800
#> 152   0.6949354617 -0.2547585747  0.177664288
#> 153   0.1764624704 -0.2103445121  0.434047160
#> 154   0.4669017352 -0.6765303525 -0.182865403
#> 155   0.7019818542 -0.4263125341  0.256955080
#> 156   0.6778908091 -0.2605755501  0.395395440
#> 157   0.4412855886  0.1711959856  0.880669846
#> 158   0.4426388073 -0.2716502700  0.691920481
#> 159   0.2664113307 -1.2392818287 -0.821460962
#> 160   0.7511122500 -0.2074193049  0.489367009
#> 161   0.3655677826 -0.5084344265  0.387544146
#> 162   0.5286304475 -0.3382188483  0.610657882
#> 163   0.3960400038 -0.6823618268 -0.428293029
#> 164   0.2202877391 -0.3713170281  0.668527444
#> 165   0.8741198065 -0.0050432055  0.493590103
#> 166   0.6024582943 -0.4265938622  0.033154418
#> 167   0.4866324100 -0.6713679466  0.243756943
#> 168   0.6917484046 -0.4805386354  0.119357670
#> 169   0.3583394765 -0.4837805462 -0.089752991
#> 170  -0.1041000262 -0.4506808989  0.046233834
#> 171   0.5391846205 -0.7174033030 -0.025594060
#> 172   0.8221773497 -0.1083321822  0.302212966
#> 173   0.6221917359 -0.6413850966  0.596951328
#> 174   0.3871738947 -0.2553089147  0.332571781
#> 175   0.5865508858 -0.6574367708  0.002837785
#> 176   0.7019818542 -0.4263125341  0.256955080
#> 177   0.3679628961 -0.2008560333  0.739633660
#> 178   0.6492511747 -0.4130905466  0.268513581
#> 179   0.0663658682 -0.0761530295  0.410092220
#> 180   0.1727767501 -1.1440702899 -0.526326155
#> 181   0.5312233214 -0.6458712396  0.102473171
#> 182   0.4149393953 -0.4329601829  0.281213391
#> 183   0.8813194129 -0.0243435542  0.281905708
#> 184   0.4882573270 -0.5620599341 -0.141375799
#> 185   0.6459682918 -0.6068233584  0.074107215
#> 186   0.5685695014 -0.6600854461  0.203285139
#> 187   0.2049056939 -0.3241880007  0.459280816
#> 188   0.7829491150 -0.2073734649  0.205880290
#> 189   0.5027707367 -0.0785064804  0.184853552
#> 190   0.5998127557 -0.5562235057  0.240571454
#> 191   0.5753663818 -0.2603862912  0.462921127
#> 192   0.3231200699 -0.6431707380 -0.049503516
#> 193   0.6886193656 -0.1359456839  0.332723710
#> 194   0.4350737013 -0.4620314726  0.148899683
#> 195   0.6289180556 -0.2921010684  0.322461065
#> 196   0.0681724693 -0.6291680144 -0.185420705
#> 197   0.8231117050 -0.0883379250  0.407171124
#> 198   0.6899945273 -0.3199031945  0.287050335
#> 199   0.2450918283 -0.2879429530  0.250291350
#> 200   0.3807289267 -0.6670016923 -0.097413980
#> 201   0.4760573521 -0.5781506531  0.002424496
#> 202   0.6173380239 -0.7433425450 -0.677989197
#> 203   0.2409461901  0.0130566070  0.689938005
#> 204   0.3956913804 -0.3066387527  0.575853895
#> 205   0.8244497339 -0.3468416564  0.173057413
#> 206   0.7802567214 -0.0062038951  0.355529469
#> 207   0.5542207555 -0.3494229153 -0.149382433
#> 208   0.5139634076 -0.4104620956  0.154777553
#> 209   0.5409012400 -0.4725427872  0.275851699
#> 210   0.7201375334 -0.2237380814  0.400165813
#> 211   0.6013389577 -0.5572225161 -0.152714704
#> 212   0.7221060719 -0.4001456808  0.042215186
#> 213   0.5355870437 -0.2230431688  0.550563362
#> 214   0.3561978354 -0.3063143313 -0.026571900
#> 215   0.0092315105 -0.1790172550  0.213995583
#> 216   0.1538210943 -0.3708211624 -0.076827745
#> 217   0.5120481735 -0.6647493258  0.375467170
#> 218   0.7749543771 -0.1413367401  0.151078277
#> 219   0.3389763881 -0.8879766095  0.341509524
#> 220   0.5982871028 -0.5379908450  0.276593459
#> 221   0.3551006842 -0.6062177443 -0.056107269
#> 222   0.5331707820 -0.3751243428 -0.153140414
#> 223   0.0917677629 -0.4998489357  0.481756421
#> 224   0.3394036389 -0.9618620094 -0.529977458
#> 225   0.2969845332 -0.5982946067  0.380502435
#> 226   0.1795203420 -0.1273103333  0.169029060
#> 227   0.8232018580 -0.0051218626  0.782433057
#> 228   0.6492511747 -0.4130905466  0.268513581
#> 229   0.5286304475 -0.3382188483  0.610657882
#> 230   0.5509338420 -0.4925303132  0.173526199
#> 231   0.4680001624 -1.0888056137 -1.012629483
#> 232   0.1090754942 -1.2256704877 -0.293234898
#> 233   0.4317847763 -0.5800060413 -0.085484515
#> 234   0.7728673113  0.1304022482  1.032643884
#> 235   0.6098232444 -0.4963677953  0.132364734
#> 236   0.5092180342 -0.4555032557  0.198161956
#> 237   0.8451278537 -0.0217657167  0.588384902
#> 238   0.8221773497 -0.1083321822  0.302212966
#> 239   0.8741198065 -0.0050432055  0.493590103
#> 240   0.5631346110 -0.4265181078 -0.156864780
#> 241  -0.3356376154 -0.0042445220  0.249956853
#> 242   0.4372699104 -0.5993380899  0.036382920
#> 243   0.5380083456  0.0735248909  0.734216712
#> 244   0.7538144473 -0.3985421957 -0.100245093
#> 245   0.4553942684 -0.8537614662 -0.125882944
#> 246   0.3405021051 -0.7778911539 -0.136965933
#> 247   0.3523933077 -0.7779375522 -0.390157989
#> 248   0.6917484046 -0.4805386354  0.119357670
#> 249   0.3289842440 -0.8148337944  0.522373702
#> 250   0.8231117050 -0.0883379250  0.407171124
#> 251   0.7139838836 -0.1002783194  0.278353252
#> 252   0.5663369965  0.0749594834  0.730810551
#> 253   0.4372699104 -0.5993380899  0.036382920
#> 254   0.4270618647 -0.7864028689  0.055775936
#> 255   0.4021069074 -0.4998419657  0.102506930
#> 256   0.4879336161 -0.5720389984 -0.410609373
#> 257   0.4808855809 -0.6296626378 -0.093341767
#> 258   0.1380317512 -0.0010449735  0.889468185
#> 259   0.2763718555 -0.8824287633 -0.157195238
#> 260   0.5814730760 -0.2498598412  0.382395578
#> 261   0.4420111192  0.1222667340  0.671565110
#> 262   0.1485606375 -0.2775314916  0.448913549
#> 263   0.4334139941 -0.5339667358  0.094592809
#> 264   0.5801381172  0.0380458633  0.718161092
#> 265   0.4908689904 -0.5717412600  0.602153411
#> 266   0.5455244416 -0.2955400583  0.138215479
#> 267   0.7168000480 -0.0336768657  0.505029417
#> 268   0.8743378392 -0.2231175782 -0.059993401
#> 269   0.6483000164 -0.0180245496  0.580878287
#> 270   0.2451614001 -0.3694479508  0.105638997
#> 271   0.8228730134 -0.2145837837  0.068819682
#> 272   0.7937088968 -0.2705754262 -0.227042764
#> 273   0.9346060001  0.0315651717  0.432642716
#> 274   0.6140695229 -0.1853145643 -0.116316431
#> 275   0.7162084075 -0.3102480553  0.045184720
#> 276   0.3405021051 -0.7778911539 -0.136965933
#> 277   0.4207373093 -0.4225682983  0.811681049
#> 278   0.5281242049 -0.4608060120  0.084935797
#> 279   0.8192849638 -0.0988214241  0.322648194
#> 280   0.3913979105 -0.4530335334 -0.108327513
#> 281   0.5640823430 -0.0317995576  0.522263078
#> 282   0.1761424800 -0.1910175216  0.457006725
#> 283   0.5616941139 -0.2809527294  0.297091597
#> 284   0.6151905445  0.0354839028  0.705502299
#> 285   0.4270618647 -0.7864028689  0.055775936
#> 286   0.5502640938 -0.1713876717  0.202353535
#> 287   0.7182753473 -0.1832780738  0.237011548
#> 288   0.6258670219 -0.2893416904  0.011604489
#> 289   0.5982871028 -0.5379908450  0.276593459
#> 290   0.4899241114 -0.4455822410  0.284259884
#> 291   0.4696308629 -0.3311784915  0.579710925
#> 292   0.2694489658 -0.6245659760  0.157494230
#> 293   0.3289842440 -0.8148337944  0.522373702
#> 294   0.5695268578 -0.5878518944  0.055316082
#> 295   0.6899945273 -0.3199031945  0.287050335
#> 296   0.6228821150 -0.1469846316  0.295643930
#> 297   0.6561454083 -0.3946028697  0.575828072
#> 298   0.6312056732 -0.2317543107  0.237063505
#> 299   0.4749716340 -0.7588452582 -0.403061187
#> 300   0.5509338420 -0.4925303132  0.173526199
#> 301   0.4291829137 -0.3547387719  0.172605790
#> 302   0.5380083456  0.0735248909  0.734216712
#> 303   0.4812613330 -0.6536061045 -0.322431793
#> 304   0.6298949207 -0.2637484389  0.299146500
#> 305   0.4808855809 -0.6296626378 -0.093341767
#> 306   0.3432412951 -0.4176900706  0.379162537
#> 307   0.5421219512 -0.2849931939  1.010593465
#> 308   0.4106746878 -0.5283939005  0.048956458
#> 309   0.5503969738 -0.6434676114 -0.325483521
#> 310   0.8407655267 -0.0979668276  1.111071718
#> 311   0.8231117050 -0.0883379250  0.407171124
#> 312   0.4741898656 -0.4697732272 -0.175090762
#> 313   0.2555786556 -0.2831547652  0.194211843
#> 314   0.4420111192  0.1222667340  0.671565110
#> 315   0.4010551109 -0.6034339754  0.695587893
#> 316   0.5286304475 -0.3382188483  0.610657882
#> 317   0.2720747094 -0.6779240149 -0.052694137
#> 318   0.6329342360 -0.6101258455 -0.400007225
#> 319   0.5577770179 -0.4415338454  0.596948975
#> 320   0.5222739460 -0.0356983785  0.674097111
#> 321   0.6150808370 -0.5238562767  0.257365615
#> 322   0.5331707820 -0.3751243428 -0.153140414
#> 323   0.7022531310 -0.3296464955  0.201861408
#> 324   0.7802567214 -0.0062038951  0.355529469
#> 325   0.5695268578 -0.5878518944  0.055316082
#> 326   0.8232018580 -0.0051218626  0.782433057
#> 327   0.3984124808 -0.3479137723  0.255697777
#> 328   0.1950052253 -1.0548521891 -0.310798602
#> 329   0.6370567593 -0.3562208347  0.622517463
#> 330   0.2852570716  0.0262538479  0.589408928
#> 331   0.6952700708 -0.0967894896  0.261714054
#> 332   0.5000431348 -0.3830180801  0.101331134
#> 333   0.7543840870 -0.2342060386  0.239516320
#> 334   0.7408890566 -0.3508074064  0.144613759
#> 335   0.0591484607 -1.2145919744  0.070769459
#> 336   0.6128015001 -0.0485548264  0.626500298
#> 337   0.5409012400 -0.4725427872  0.275851699
#> 338   0.6873848770  0.2890419335  0.968018068
#> 339   0.3192149641 -0.8007012802  0.134870525
#> 340   0.1649806457 -0.3360475519  0.131285471
#> 341   0.6097714172 -0.3442291847  0.264689536
#> 342   0.5504533430 -0.2656944844  0.494806648
#> 343   0.2835039664 -1.0729852674 -0.092121723
#> 344   0.5286304475 -0.3382188483  0.610657882
#> 345   0.4415052096 -0.7296365902 -0.345061377
#> 346   0.4267370877 -0.4262818047 -0.168006989
#> 347   0.7543840870 -0.2342060386  0.239516320
#> 348   0.4779949835 -0.2866284796  0.142283562
#> 349   0.6788737888 -0.4659663463  0.060461962
#> 350   0.5525076671 -0.6810873697 -0.152299920
#> 351   0.7708921928  0.0841070116  0.613791686
#> 352   0.2035952065 -0.2848993584  0.225170983
#> 353   0.7364532598 -0.1464409723  0.270350213
#> 354   0.5520053582 -0.3563949264 -0.054141076
#> 355   0.6483000164 -0.0180245496  0.580878287
#> 356   0.8642959456  0.0486905588  0.950655162
#> 357   0.5370934957 -0.3877607583  0.500147890
#> 358   0.5838075560 -0.6287640836 -0.404894381
#> 359   0.4403478353 -0.4748823127 -0.020608811
#> 360   0.7330962967 -0.4096119065 -0.074171488
#> 361   0.6438459466 -0.6408896268 -0.653287757
#> 362   0.5426995551 -0.1390382389  0.681949733
#> 363   0.5339662086 -0.5547411643 -0.053606088
#> 364   0.6492511747 -0.4130905466  0.268513581
#> 365   0.5838075560 -0.6287640836 -0.404894381
#> 366   0.2755660500 -0.8671739256 -0.139597968
#> 367   0.5139634076 -0.4104620956  0.154777553
#> 368   0.3125198002 -0.0201932471  0.507993863
#> 369   0.5504533430 -0.2656944844  0.494806648
#> 370  -0.4648196871  0.0431729521  0.467266620
#> 371   0.8440032513  0.2049313264  1.059534051
#> 372   0.3725976481 -0.5368313621  0.285548506
#> 373   0.3597872806 -0.3457979924  0.912948522
#> 374   0.7201375334 -0.2237380814  0.400165813
#> 375   0.5462259745 -0.3478552055  0.380420054
#> 376   0.6735074440 -0.4689964595 -0.037719389
#> 377   0.4838999153 -0.8085144032  0.050494587
#> 378   0.7812910265 -0.0995265976  0.589430982
#> 379   0.6279025082 -0.4088492432  0.647852099
#> 380   0.3439465119 -0.7545846371  0.079739234
#> 381   0.2608808007 -1.1653043811 -0.430018195
#> 382   0.3394036389 -0.9618620094 -0.529977458
#> 383  -0.0744577660  0.0627658355  0.414388074
#> 384   0.4010551109 -0.6034339754  0.695587893
#> 385   0.4760573521 -0.5781506531  0.002424496
#> 386   0.8256878044 -0.1665142281  1.218173256
#> 387   0.8451278537 -0.0217657167  0.588384902
#> 388   0.0965204479 -0.7386178362 -0.270308214
#> 389   0.4233916672 -0.6810951638  0.023722108
#> 390   0.4749716340 -0.7588452582 -0.403061187
#> 391   0.6839983775 -0.2491422973 -0.283655765
#> 392   0.4391031146 -0.3165867860  0.078469083
#> 393   0.7039652550 -0.2483784510  0.025640208
#> 394   0.8197458939 -0.0952069635  0.407658000
#> 395   0.5871048912 -0.5330464481  0.223001074
#> 396   0.4921826152 -0.5842873666 -0.244331378
#> 397   0.7535859285 -0.2738703004  0.267589472
#> 398   0.3325007676 -0.1090035551  0.157704628
#> 399   0.3439465119 -0.7545846371  0.079739234
#> 400   0.6098232444 -0.4963677953  0.132364734
#> 401   0.6289180556 -0.2921010684  0.322461065
#> 402   0.2438363922 -0.1143277755  0.337977720
#> 403   0.6660364700 -0.3803391488 -0.013155003
#> 404   0.5814730760 -0.2498598412  0.382395578
#> 405   0.3389763881 -0.8879766095  0.341509524
#> 406   0.8197458939 -0.0952069635  0.407658000
#> 407   0.5525076671 -0.6810873697 -0.152299920
#> 408   0.4350737013 -0.4620314726  0.148899683
#> 409   0.4417221683 -0.5342528579  0.170149017
#> 410   0.0168748268 -0.5180027313  0.321572657
#> 411   0.4555105267 -0.2880392323  0.195010723
#> 412   0.5443383926 -0.2701865648  0.075954036
#> 413   0.5801381172  0.0380458633  0.718161092
#> 414   0.7839368618  0.1084585162  0.517799204
#> 415   0.8035919039 -0.2700850673  0.406780480
#> 416   0.4458251991 -0.6751493487 -0.217083916
#> 417   0.8868792649  0.1366057150  0.318864284
#> 418   0.4921826152 -0.5842873666 -0.244331378
#> 419   0.4317847763 -0.5800060413 -0.085484515
#> 420   0.8244497339 -0.3468416564  0.173057413
#> 421   0.6489448615 -0.2264088114  0.953738205
#> 422   0.6634097304  0.0048345518  0.634571632
#> 423   0.7812910265 -0.0995265976  0.589430982
#> 424   0.4553942684 -0.8537614662 -0.125882944
#> 425   0.7107967801  0.0193665200  0.877304501
#> 426   0.4534595290 -0.3080273788  0.916586991
#> 427   0.7330962967 -0.4096119065 -0.074171488
#> 428   0.1380317512 -0.0010449735  0.889468185
#> 429   0.6298949207 -0.2637484389  0.299146500
#> 430   0.4053205627 -0.8771367064 -0.603381281
#> 431   0.6233007915 -0.2701789629  0.383681879
#> 432   0.3807289267 -0.6670016923 -0.097413980
#> 433   0.5402240875 -0.3614207287  0.552627895
#> 434   0.6150808370 -0.5238562767  0.257365615
#> 435   0.2226254591 -0.0011342184  0.881199026
#> 436   0.5393431897 -0.6952058358 -0.101491150
#> 437   0.7112459196 -0.0488965425  0.733810757
#> 438   0.1258608971  0.1502491747  0.685534542
#> 439   0.3538693209 -0.2657266236  0.052462812
#> 440   0.4749716340 -0.7588452582 -0.403061187
#> 441   0.4869369442 -0.6483198389 -0.055597940
#> 442   0.5426995551 -0.1390382389  0.681949733
#> 443   0.2986412624 -0.6300936318  0.106750391
#> 444   0.6097714172 -0.3442291847  0.264689536
#> 445   0.5165780679 -0.8986232572 -0.324055764
#> 446   0.7567408467 -0.1839804420  0.797065268
#> 447   0.5253747746 -0.5815683001  0.142478986
#> 448   0.4207373093 -0.4225682983  0.811681049
#> 449   0.3328875328 -0.6843926047  0.666283113
#> 450   0.6849327223 -0.3391657827  0.058574921
#> 451   0.4720981796 -0.1825247624  1.197058825
#> 452   0.5370934957 -0.3877607583  0.500147890
#> 453   0.5126231444 -0.8643629820 -0.161475391
#> 454   0.3230912711 -0.6619084121  0.335801885
#> 455   0.2846902098  0.0942387772  0.724467934
#> 456   0.4159141592 -0.5018378385  0.090907066
#> 457   0.2429289218 -0.6725946678 -0.006647113
#> 458   0.3322525192 -0.2848303639  0.286287716
#> 459   0.0766195624 -1.1290124856 -0.516210061
#> 460   0.4894334191 -0.6731728246  0.438565872
#> 461   0.8451278537 -0.0217657167  0.588384902
#> 462   0.9156492780 -0.0938471423  0.706470249
#> 463   0.6298949207 -0.2637484389  0.299146500
#> 464   0.7162084075 -0.3102480553  0.045184720
#> 465   0.2968229859 -0.9398413399 -0.374008128
#> 466   0.4233916672 -0.6810951638  0.023722108
#> 467   0.7728673113  0.1304022482  1.032643884
#> 468   0.3807289267 -0.6670016923 -0.097413980
#> 469   0.5222739460 -0.0356983785  0.674097111
#> 470   0.4417221683 -0.5342528579  0.170149017
#> 471   0.5307787319 -0.3153362569  0.238584941
#> 472   0.6233007915 -0.2701789629  0.383681879
#> 473   0.5595108058 -0.5041474188  0.230603597
#> 474   0.5466179788 -0.2127759761 -0.106219760
#> 475   0.4866324100 -0.6713679466  0.243756943
#> 476   0.5142834015 -0.6739173158  0.798090663
#> 477   0.5998301523 -0.3247869073  0.505965392
#> 478   0.7363084171 -0.3578733712  0.211604109
#> 479   0.6390802861  0.0428583982  0.969644676
#> 480   0.7014742625  0.2322150742  0.852809385
#> 481   0.5664066291 -0.2397275618  0.482918870
#> 482   0.5408025407 -0.4516931188  0.605553373
#> 483   0.5504533430 -0.2656944844  0.494806648
#> 484   0.5027707367 -0.0785064804  0.184853552
#> 485   0.6118952377 -0.3378488093  0.442344335
#> 486   0.6100286010 -0.1031145195  0.658934379
#> 487   0.4484419150 -0.7728907483 -0.429257461
#> 488   0.7107967801  0.0193665200  0.877304501
#> 489   0.4869369442 -0.6483198389 -0.055597940
#> 490   0.3774209509 -0.6517649812 -0.326782429
#> 491   0.7594888787 -0.1932041944  0.170562056
#> 492   0.7450301619 -0.1645220235  0.120303167
#> 493   0.5404354986 -0.3991711251 -0.342549597
#> 494   0.3447410456 -0.6960819597  0.109687491
#> 495   0.4767486413 -0.7267912408 -0.132796639
#> 496   0.8245433447 -0.1677040627  0.595087473
#> 497   0.4559065812 -0.3984150509  0.249072398
#> 498   0.6561454083 -0.3946028697  0.575828072
#> 499  -0.1758386465 -0.5100702067 -0.111660963
#> 500   0.2961965564 -0.7874860931  0.159604121
#> 501   0.6298949207 -0.2637484389  0.299146500
#> 502   0.5603886390 -0.7128409421 -0.499113938
#> 503   0.3356636614 -0.3818718448  0.625785348
#> 504   0.2419558257 -1.0805902234  0.015586364
#> 505   0.4936170722 -0.5606983880  0.114251334
#> 506   0.3228397142 -0.5140563743  0.103280594
#> 507   0.4385622872 -0.5898842054  0.089112044
#> 508   0.5799867686 -0.6579623889 -0.172411031
#> 509   0.4808855809 -0.6296626378 -0.093341767
#> 510   0.7538144473 -0.3985421957 -0.100245093
#> 511   0.6952700708 -0.0967894896  0.261714054
#> 512   0.6455285283 -0.3439687679 -0.170030148
#> 513   0.4810060762 -0.3692849963  0.289283635
#> 514   0.6621483307  0.1483624419  0.733993611
#> 515   0.6441485727 -0.6527791060 -0.314105784
#> 516   0.7535136009 -0.2609014669  0.584230994
#> 517   0.2422013214 -0.2666015738  0.214730120
#> 518   0.6561454083 -0.3946028697  0.575828072
#> 519   1.0195716902  0.1136329612  0.921300237
#> 520   0.2212053961 -0.4467228721  0.491449540
#> 521   0.5525076671 -0.6810873697 -0.152299920
#> 522   0.7310167530 -0.5720353176 -0.542583377
#> 523   0.2707258870 -0.9031605763 -0.187615352
#> 524   0.1065965703 -0.1213320709  0.355696603
#> 525   0.6842812611 -0.4574227275  0.322580133
#> 526   0.2952118923 -0.8368882754  0.153292343
#> 527   0.7535859285 -0.2738703004  0.267589472
#> 528   0.3141857292 -0.2952357994  0.246342850
#> 529   0.7022531310 -0.3296464955  0.201861408
#> 530   0.4426388073 -0.2716502700  0.691920481
#> 531   0.2133869761 -0.1231267936  0.696858512
#> 532   0.4979648902 -0.2104491199  0.566378861
#> 533   0.1239289247 -1.3557501352 -0.444101201
#> 534   0.5838075560 -0.6287640836 -0.404894381
#> 535   0.3599669860  0.0332337489  1.039563356
#> 536   0.5665896096 -0.8056926158 -0.326298352
#> 537   0.4669017352 -0.6765303525 -0.182865403
#> 538   0.6639757626 -0.4621253589 -0.163040821
#> 539   0.6561454083 -0.3946028697  0.575828072
#> 540   0.4553030700 -0.5084231342  0.032017814
#> 541   0.4892951844 -0.4473525667  0.389627544
#> 542   0.7543840870 -0.2342060386  0.239516320
#> 543   0.4322404467 -0.5211763609  0.204418991
#> 544   0.3394036389 -0.9618620094 -0.529977458
#> 545   0.3530489101 -0.6151597590  0.081134956
#> 546   0.6316104326 -0.6355756741  0.685785110
#> 547   0.0039393966 -0.0430459705  0.648585084
#> 548   0.6375788045 -0.2743785411  0.246959657
#> 549   0.7829491150 -0.2073734649  0.205880290
#> 550   0.4372699104 -0.5993380899  0.036382920
#> 551   0.6492511747 -0.4130905466  0.268513581
#> 552   0.6608463899 -0.4362716971 -0.009607615
#> 553   0.5390881722 -0.3216374628  0.586521133
#> 554   0.5727931303 -0.1259755641  0.615306895
#> 555   0.5235194991 -0.1957509569  0.840489397
#> 556   0.4149393953 -0.4329601829  0.281213391
#> 557   0.0399619615 -0.9148266743  0.440229612
#> 558   0.6627130591 -0.2934735254  0.384453148
#> 559   0.4882573270 -0.5620599341 -0.141375799
#> 560   0.8041367959  0.0841328847  0.219604026
#> 561   0.5229490975 -0.4311186278  0.360036567
#> 562   0.8232018580 -0.0051218626  0.782433057
#> 563   0.5391846205 -0.7174033030 -0.025594060
#> 564   0.7067267828  0.1727620248  0.597284168
#> 565   0.4484419150 -0.7728907483 -0.429257461
#> 566   0.6024582943 -0.4265938622  0.033154418
#> 567   0.5000533979 -0.0032017922  0.230985161
#> 568   0.5799867686 -0.6579623889 -0.172411031
#> 569   0.5998127557 -0.5562235057  0.240571454
#> 570   0.6258670219 -0.2893416904  0.011604489
#> 571   0.3879251400 -0.6736019077  0.128256027
#> 572   0.7725155705  0.0788370588  0.418061338
#> 573   0.5822612624 -0.0840406249  0.531125094
#> 574   0.8598452768 -0.3238542747  0.202389867
#> 575   0.5951906572 -0.3050537908  0.054423749
#> 576   0.7408890566 -0.3508074064  0.144613759
#> 577   0.8973494515 -0.0804563835  0.423523272
#> 578   0.8319277699  0.0756030425  0.406263537
#> 579   0.4741898656 -0.4697732272 -0.175090762
#> 580   0.2251483645  0.1738922448  0.818038029
#> 581   0.5865508858 -0.6574367708  0.002837785
#> 582   0.4316003730 -0.3248095426  0.501039295
#> 583   0.2304734629  0.2877488975  1.081704259
#> 584   0.6233007915 -0.2701789629  0.383681879
#> 585   0.7182753473 -0.1832780738  0.237011548
#> 586   0.5588167805  0.1694945312  0.452596398
#> 587   0.4892951844 -0.4473525667  0.389627544
#> 588   0.6370567593 -0.3562208347  0.622517463
#> 589   0.5426995551 -0.1390382389  0.681949733
#> 590   0.5370934957 -0.3877607583  0.500147890
#> 591   0.5509338420 -0.4925303132  0.173526199
#> 592   0.4760573521 -0.5781506531  0.002424496
#> 593   0.4669108065 -0.5888083447  0.073714336
#> 594   0.5370934957 -0.3877607583  0.500147890
#> 595   0.8407655267 -0.0979668276  1.111071718
#> 596   0.6312056732 -0.2317543107  0.237063505
#> 597   0.6406621951 -0.2150417509  0.451423864
#> 598   0.2409461901  0.0130566070  0.689938005
#> 599   0.6161778737 -0.0978911275  1.152163947
#> 600   0.7363084171 -0.3578733712  0.211604109
#> 601   0.5426995551 -0.1390382389  0.681949733
#> 602   0.1853476765 -0.7139195790 -0.406760505
#> 603   0.5455244416 -0.2955400583  0.138215479
#> 604   0.1592516330 -0.5947641495 -0.038454311
#> 605  -0.1289814250 -0.8800215902 -0.103702605
#> 606   0.5393431897 -0.6952058358 -0.101491150
#> 607   0.6279025082 -0.4088492432  0.647852099
#> 608   0.4385622872 -0.5898842054  0.089112044
#> 609   0.8183177074 -0.2993331926  0.451961132
#> 610   0.3322525192 -0.2848303639  0.286287716
#> 611   0.2693757619 -0.4001062877  0.425725566
#> 612   0.3762641208 -0.1322944937  0.272032536
#> 613   0.5113725485 -0.5581640745  0.007409800
#> 614   0.2355841986 -0.3048117400  0.410140045
#> 615   0.6903942696  0.0715345168  0.395903772
#> 616   0.3446587394 -0.8858029514  0.567803275
#> 617   0.6375788045 -0.2743785411  0.246959657
#> 618   0.7067267828  0.1727620248  0.597284168
#> 619   0.5664066291 -0.2397275618  0.482918870
#> 620   0.6627130591 -0.2934735254  0.384453148
#> 621   0.1816854971  0.0410115000  0.284997431
#> 622   0.5339662086 -0.5547411643 -0.053606088
#> 623   0.3671484068 -0.3410153819  0.450161823
#> 624   0.3864801486 -0.2325714677  0.332159396
#> 625   0.1730397790 -0.6707703794 -0.073940253
#> 626   0.7022531310 -0.3296464955  0.201861408
#> 627   0.5462259745 -0.3478552055  0.380420054
#> 628   0.4365715082 -0.3014904378  0.345112336
#> 629   0.8109542957 -0.0721907924  0.200683766
#> 630   0.1783878074 -0.9215754553  0.045002070
#> 631   0.4318323065 -0.7183506111  0.048778932
#> 632   0.5445971052 -0.5367932850 -0.398885626
#> 633   0.4403478353 -0.4748823127 -0.020608811
#> 634   0.5453378859 -0.7288587614 -0.086115523
#> 635   0.1086958066 -0.5045148961  0.111667422
#> 636   0.5393431897 -0.6952058358 -0.101491150
#> 637   0.5871048912 -0.5330464481  0.223001074
#> 638   0.5577770179 -0.4415338454  0.596948975
#> 639   0.6098232444 -0.4963677953  0.132364734
#> 640   0.4458251991 -0.6751493487 -0.217083916
#> 641  -0.3710027036 -0.2143888467  0.432959905
#> 642   0.5126231444 -0.8643629820 -0.161475391
#> 643   0.4316582207 -0.8260090951  0.035895596
#> 644   0.6315361228 -0.6112551703 -0.137122185
#> 645   0.2472550271 -0.2941245122  0.567422705
#> 646   0.7812910265 -0.0995265976  0.589430982
#> 647   0.4894334191 -0.6731728246  0.438565872
#> 648   0.5871048912 -0.5330464481  0.223001074
#> 649   0.8673850266 -0.2238824486  0.422306384
#> 650   0.4767486413 -0.7267912408 -0.132796639
#> 651   0.5998127557 -0.5562235057  0.240571454
#> 652   0.8651802057  0.0292043526  0.317090603
#> 653   0.2602602324 -0.9384030860 -0.214418357
#> 654   0.7201375334 -0.2237380814  0.400165813
#> 655   0.5603886390 -0.7128409421 -0.499113938
#> 656   0.8231117050 -0.0883379250  0.407171124
#> 657   0.3499918296 -0.8190976694 -0.212063154
#> 658   0.4808855809 -0.6296626378 -0.093341767
#> 659   0.0980022833 -0.4825658510 -0.194707118
#> 660   0.4760573521 -0.5781506531  0.002424496
#> 661   0.8056501400 -0.1018358047  0.573230623
#> 662   0.5165780679 -0.8986232572 -0.324055764
#> 663   0.0292660010 -0.3962318636  0.159513809
#> 664   0.6233007915 -0.2701789629  0.383681879
#> 665   0.4749716340 -0.7588452582 -0.403061187
#> 666   0.5192856489 -0.2986942638  0.037600047
#> 667   0.5868315224 -0.7756947655 -0.132930861
#> 668   0.6884465072  0.0204141890  0.630078946
#> 669   0.4322404467 -0.5211763609  0.204418991
#> 670   0.4484419150 -0.7728907483 -0.429257461
#> 671   0.5455750830 -0.1272802746  0.147504793
#> 672   0.5113725485 -0.5581640745  0.007409800
#> 673   0.2460895332 -0.4931566522 -0.109767908
#> 674   0.7560829849 -0.4074008120  0.166870950
#> 675   0.6406621951 -0.2150417509  0.451423864
#> 676   0.6094518864 -0.6544628996  0.514965019
#> 677   0.4605155779 -0.5349268712  0.094618037
#> 678   0.4317847763 -0.5800060413 -0.085484515
#> 679   0.4441106862 -0.5599085817  0.074802490
#> 680   0.6406621951 -0.2150417509  0.451423864
#> 681   0.4213308458 -0.3224579983  0.295242746
#> 682   0.4441106862 -0.5599085817  0.074802490
#> 683   0.6998083868 -0.2291333088  0.295685435
#> 684   0.4598954780 -0.0134512126  0.587310181
#> 685   0.5642438711 -0.4720023432  0.064612163
#> 686   0.6100286010 -0.1031145195  0.658934379
#> 687   0.6483000164 -0.0180245496  0.580878287
#> 688   0.5727931303 -0.1259755641  0.615306895
#> 689   0.6998420070 -0.0378224168  0.385663925
#> 690   0.6298949207 -0.2637484389  0.299146500
#> 691   0.4441254342 -0.3712848509  0.181352472
#> 692   0.4908689904 -0.5717412600  0.602153411
#> 693   0.2932659207 -0.3387262933  0.334626303
#> 694   0.6492511747 -0.4130905466  0.268513581
#> 695   0.2429289218 -0.6725946678 -0.006647113
#> 696   0.5355870437 -0.2230431688  0.550563362
#> 697   0.5754852064 -0.2424219063 -0.147869824
#> 698   0.4752690807  0.2272917345  0.705742011
#> 699   0.4652897841  0.0393420228  0.893432422
#> 700   0.7568792045 -0.2089862729  0.508502204
#> 701   0.5871048912 -0.5330464481  0.223001074
#> 702   0.4553030700 -0.5084231342  0.032017814
#> 703   0.3447410456 -0.6960819597  0.109687491
#> 704   0.5256927334 -0.2757948507  0.173882014
#> 705   0.4316582207 -0.8260090951  0.035895596
#> 706   0.5503969738 -0.6434676114 -0.325483521
#> 707   0.3741764924 -0.2357190411  0.996785549
#> 708   0.7053409487 -0.5496642320  0.012924202
#> 709   0.4518940286 -0.5257828535 -0.050234810
#> 710   0.2926737390 -0.4408925064  0.057869654
#> 711   0.4765364810 -0.2388316479  0.606719604
#> 712   0.5603531829 -0.4884897060 -0.345666641
#> 713   0.8451278537 -0.0217657167  0.588384902
#> 714   0.4270527571 -0.4221709247  0.229543921
#> 715   0.6842812611 -0.4574227275  0.322580133
#> 716   0.2924872957 -0.1420559589  0.412739620
#> 717   0.4808855809 -0.6296626378 -0.093341767
#> 718   0.0360805046 -0.0503007927  1.071842434
#> 719   0.3878429001 -0.6867403293  0.068375436
#> 720   0.4808855809 -0.6296626378 -0.093341767
#> 721   0.5868315224 -0.7756947655 -0.132930861
#> 722   0.8231117050 -0.0883379250  0.407171124
#> 723   0.0459694351 -0.9425217602  0.093113793
#> 724   0.7053409487 -0.5496642320  0.012924202
#> 725   0.4270618647 -0.7864028689  0.055775936
#> 726   0.6150808370 -0.5238562767  0.257365615
#> 727   0.5695268578 -0.5878518944  0.055316082
#> 728   0.7535859285 -0.2738703004  0.267589472
#> 729   0.6214110790  0.1154499052  0.366804108
#> 730   0.2912183940 -0.7016654103  0.783954998
#> 731   0.8642959456  0.0486905588  0.950655162
#> 732   0.4765364810 -0.2388316479  0.606719604
#> 733  -0.4648196871  0.0431729521  0.467266620
#> 734   0.4669108065 -0.5888083447  0.073714336
#> 735   0.5252864378 -0.3221236519  0.227611993
#> 736   0.6210965139  0.3494808140  1.565287826
#> 737   0.5402240875 -0.3614207287  0.552627895
#> 738   0.7133351684 -0.0314645922  0.288057257
#> 739   0.1224840724 -0.4902802181 -0.044019485
#> 740   0.2737087486 -0.6416114233 -0.429166150
#> 741   0.6778908091 -0.2605755501  0.395395440
#> 742   0.7728673113  0.1304022482  1.032643884
#> 743   0.5910724797 -0.4260331119 -0.264763971
#> 744   0.4532168403 -0.4806761091 -0.181659240
#> 745   0.7019818542 -0.4263125341  0.256955080
#> 746   0.6489448615 -0.2264088114  0.953738205
#> 747   0.4533629359 -0.0540363925  0.636081284
#> 748   0.6233007915 -0.2701789629  0.383681879
#> 749   0.4339669912 -0.4448804194 -0.150474285
#> 750   0.3245312661 -0.3010149105  1.089602030
#> 751   0.3499918296 -0.8190976694 -0.212063154
#> 752   0.5113725485 -0.5581640745  0.007409800
#> 753   0.6660364700 -0.3803391488 -0.013155003
#> 754   0.6006072568 -0.1201841256  0.307794636
#> 755   0.5951906572 -0.3050537908  0.054423749
#> 756   0.5046632077 -0.2164942850  0.397390717
#> 757   0.5504533430 -0.2656944844  0.494806648
#> 758   0.3618811094 -0.8388368881  0.527505160
#> 759   0.6644648072 -0.0897974991  0.193704909
#> 760   0.5998127557 -0.5562235057  0.240571454
#> 761   0.7749543771 -0.1413367401  0.151078277
#> 762   0.5502640938 -0.1713876717  0.202353535
#> 763   0.5814730760 -0.2498598412  0.382395578
#> 764   0.2489227139 -0.2864588743  0.316447783
#> 765   0.4605155779 -0.5349268712  0.094618037
#> 766   0.4053205627 -0.8771367064 -0.603381281
#> 767   0.6627130591 -0.2934735254  0.384453148
#> 768   0.8232018580 -0.0051218626  0.782433057
#> 769   0.4063989562 -0.2021134470  0.947376837
#> 770   0.5291002550 -0.4489777412  0.623927720
#> 771   0.7350780247  0.1851969313  0.891813623
#> 772   0.5509338420 -0.4925303132  0.173526199
#> 773   0.3115167305 -0.1090017968  0.244166075
#> 774   0.3394036389 -0.9618620094 -0.529977458
#> 775   0.6315361228 -0.6112551703 -0.137122185
#> 776   0.3499918296 -0.8190976694 -0.212063154
#> 777   0.2400965382 -0.4563534918  0.026114072
#> 778   0.5509338420 -0.4925303132  0.173526199
#> 779   0.5393431897 -0.6952058358 -0.101491150
#> 780   0.4322404467 -0.5211763609  0.204418991
#> 781   0.1568488531 -1.1018376774  0.511246446
#> 782   0.5462259745 -0.3478552055  0.380420054
#> 783  -0.0140352191 -0.9430820098 -0.183557071
#> 784   0.5749985509  0.2446075164  0.969636875
#> 785   0.6441485727 -0.6527791060 -0.314105784
#> 786   0.5814730760 -0.2498598412  0.382395578
#> 787   0.5729753923 -0.2439071508  0.250575054
#> 788   0.4007693494 -0.5880584200 -0.317004497
#> 789   0.3499918296 -0.8190976694 -0.212063154
#> 790   0.5642438711 -0.4720023432  0.064612163
#> 791   0.6202592326  0.1586888455  0.424758560
#> 792   0.3336811786 -0.6841954452 -0.431358929
#> 793   0.3720555968 -0.4954469211  0.327433336
#> 794   0.2073685269 -0.5886389965  0.817313494
#> 795   0.5445559043 -0.5950224539 -0.195028408
#> 796   0.1257830928 -1.0390242311  0.253424236
#> 797   0.3038946732 -1.2449083654 -0.491736205
#> 798   0.8231117050 -0.0883379250  0.407171124
#> 799   0.6314385777 -0.0903990937  0.384604247
#> 800   0.2755660500 -0.8671739256 -0.139597968
#> 801   0.4257581440 -0.7363171956  0.488568365
#> 802   0.6370567593 -0.3562208347  0.622517463
#> 803   0.5799867686 -0.6579623889 -0.172411031
#> 804   0.4415052096 -0.7296365902 -0.345061377
#> 805   0.4658212848 -0.3296347472  0.319346837
#> 806   0.8407655267 -0.0979668276  1.111071718
#> 807  -0.2925219724 -0.5276434953  0.089276656
#> 808   0.5339662086 -0.5547411643 -0.053606088
#> 809   0.3042783940 -0.4215895820  0.241035699
#> 810   0.3913979105 -0.4530335334 -0.108327513
#> 811   0.3137028134 -0.4970737862 -0.081218643
#> 812   0.8221773497 -0.1083321822  0.302212966
#> 813   0.1783878074 -0.9215754553  0.045002070
#> 814   0.4741898656 -0.4697732272 -0.175090762
#> 815   0.5165780679 -0.8986232572 -0.324055764
#> 816   0.6315361228 -0.6112551703 -0.137122185
#> 817   0.3264901142 -0.9769646501  0.106072391
#> 818   0.6289180556 -0.2921010684  0.322461065
#> 819   0.4878015820 -0.9160557723 -0.069990643
#> 820   0.5925099207 -0.1784528052  0.207412547
#> 821   0.1154927228 -0.2381972384  0.434117648
#> 822   0.7350780247  0.1851969313  0.891813623
#> 823   0.5503969738 -0.6434676114 -0.325483521
#> 824   0.8232018580 -0.0051218626  0.782433057
#> 825   0.3291689092 -0.4418652289 -0.003782644
#> 826   0.2156448219 -0.8597260413 -0.042908649
#> 827   0.6778908091 -0.2605755501  0.395395440
#> 828   0.6434198868 -0.5097439627 -0.303344316
#> 829   0.7086032188  0.1112300537  0.171653719
#> 830   0.0807087281 -0.1132595285  0.322023705
#> 831   0.6098232444 -0.4963677953  0.132364734
#> 832   0.4539444138 -0.2996342811  0.274184359
#> 833   0.5640823430 -0.0317995576  0.522263078
#> 834   0.6008362150  0.0983042694  0.945802216
#> 835   0.7363084171 -0.3578733712  0.211604109
#> 836   0.2555786556 -0.2831547652  0.194211843
#> 837   0.4063989562 -0.2021134470  0.947376837
#> 838   0.5455244416 -0.2955400583  0.138215479
#> 839   0.6483000164 -0.0180245496  0.580878287
#> 840   0.0963713061 -0.5827034036 -0.282727861
#> 841   0.3394036389 -0.9618620094 -0.529977458
#> 842   0.6489448615 -0.2264088114  0.953738205
#> 843   0.5165780679 -0.8986232572 -0.324055764
#> 844   0.5391846205 -0.7174033030 -0.025594060
#> 845   0.5664066291 -0.2397275618  0.482918870
#> 846   0.2239042665 -0.2914225492  0.598381614
#> 847   0.0725356427 -0.4917441608 -0.193814355
#> 848   0.7511122500 -0.2074193049  0.489367009
#> 849   0.4978965851 -0.0334292781  0.659219220
#> 850   0.7364532598 -0.1464409723  0.270350213
#> 851   0.6406621951 -0.2150417509  0.451423864
#> 852   0.7082845778 -0.6100595701 -0.305407689
#> 853   0.6492511747 -0.4130905466  0.268513581
#> 854   0.4892951844 -0.4473525667  0.389627544
#> 855   0.2737087486 -0.6416114233 -0.429166150
#> 856   0.5432964287 -0.4372545671 -0.140761264
#> 857   0.7112459196 -0.0488965425  0.733810757
#> 858   0.6329342360 -0.6101258455 -0.400007225
#> 859   0.3796470479 -0.6819817917  0.497092308
#> 860   0.6899945273 -0.3199031945  0.287050335
#> 861   0.8407655267 -0.0979668276  1.111071718
#> 862   0.7568792045 -0.2089862729  0.508502204
#> 863   0.2664113307 -1.2392818287 -0.821460962
#> 864   0.6550425888  0.0185897527  0.262481917
#> 865   0.4415052096 -0.7296365902 -0.345061377
#> 866   0.7162084075 -0.3102480553  0.045184720
#> 867   0.6158809866 -0.2033984758  0.128452627
#> 868   0.8632331257  0.0055937126  1.152214282
#> 869   0.4417221683 -0.5342528579  0.170149017
#> 870  -0.0002034854 -0.3625617202  0.205659134
#> 871   0.6483000164 -0.0180245496  0.580878287
#> 872   0.5695268578 -0.5878518944  0.055316082
#> 873   0.5412051355 -0.4069637129  0.168800560
#> 874   0.5139634076 -0.4104620956  0.154777553
#> 875   0.4087395769 -0.5960471161  0.105414174
#> 876   0.6233007915 -0.2701789629  0.383681879
#> 877   0.2836634474 -0.5469590331  0.327244842
#> 878   0.3394036389 -0.9618620094 -0.529977458
#> 879   0.5509338420 -0.4925303132  0.173526199
#> 880   0.5291002550 -0.4489777412  0.623927720
#> 881   0.5603886390 -0.7128409421 -0.499113938
#> 882   0.5006481327  0.0007427915  0.847312325
#> 883   0.6508047851 -0.1584514733  0.170648907
#> 884   0.4316582207 -0.8260090951  0.035895596
#> 885   0.1835662592 -0.7103281403 -0.411924506
#> 886   0.8522516240  0.1064399318  0.659873009
#> 887   0.4084431278 -0.6278077760 -0.471747050
#> 888   0.5339662086 -0.5547411643 -0.053606088
#> 889   0.8156384159 -0.0061373889  0.325794542
#> 890   0.6899945273 -0.3199031945  0.287050335
#> 891   0.2094417715 -1.2077303679 -0.550942651
#> 892   0.6849327223 -0.3391657827  0.058574921
#> 893   0.1853476765 -0.7139195790 -0.406760505
#> 894   0.5281242049 -0.4608060120  0.084935797
#> 895   0.5814730760 -0.2498598412  0.382395578
#> 896   0.3617911581 -0.0665274585  0.576482551
#> 897   0.4087001103 -0.5464503463 -0.143087578
#> 898   0.4128762340 -0.2901343839  0.467580674
#> 899   0.4339669912 -0.4448804194 -0.150474285
#> 900   0.6735074440 -0.4689964595 -0.037719389
#> 901   0.4181609385 -0.6198452926 -0.318883601
#> 902   0.4484419150 -0.7728907483 -0.429257461
#> 903   0.3987818316 -0.2764829342  0.205324645
#> 904   0.9037212282 -0.2048572267  0.238291767
#> 905   0.5165948964 -0.4682456414  0.419465853
#> 906   0.4557725616 -0.2874837786  0.425580354
#> 907   0.6899945273 -0.3199031945  0.287050335
#> 908   0.6998083868 -0.2291333088  0.295685435
#> 909   0.5504524797 -0.7053095951 -0.299875336
#> 910   0.6842812611 -0.4574227275  0.322580133
#> 911   0.5421219512 -0.2849931939  1.010593465
#> 912   0.5470424820 -0.1665151967  0.145528264
#> 913   0.4969245864 -0.4389667200 -0.449306344
#> 914   0.5000431348 -0.3830180801  0.101331134
#> 915   0.1754030590 -1.3501123413 -0.470200896
#> 916   0.7543840870 -0.2342060386  0.239516320
#> 917   0.2948736894 -0.4012457210  0.365204190
#> 918   0.4767486413 -0.7267912408 -0.132796639
#> 919   0.4869369442 -0.6483198389 -0.055597940
#> 920   0.7098007007 -0.2579103383  0.106668831
#> 921   0.6492511747 -0.4130905466  0.268513581
#> 922   0.4658212848 -0.3296347472  0.319346837
#> 923   0.6549892231  0.1548989779  0.890171288
#> 924   0.6526654150 -0.3537695095  1.087012954
#> 925   0.7162084075 -0.3102480553  0.045184720
#> 926   0.5488924028 -0.4388275047 -0.149812122
#> 927   0.4869369442 -0.6483198389 -0.055597940
#> 928   0.2608808007 -1.1653043811 -0.430018195
#> 929   0.5640823430 -0.0317995576  0.522263078
#> 930   0.3551006842 -0.6062177443 -0.056107269
#> 931   0.7535859285 -0.2738703004  0.267589472
#> 932   0.5520053582 -0.3563949264 -0.054141076
#> 933   0.2113334395 -0.1710710301  0.364501982
#> 934   0.6998420070 -0.0378224168  0.385663925
#> 935  -0.3238055065 -0.2646291662  0.325155916
#> 936   0.2491986184 -0.7726210285 -0.138161549
#> 937   0.7110745348 -0.6983697450 -0.489709902
#> 938   0.7168000480 -0.0336768657  0.505029417
#> 939   0.8522516240  0.1064399318  0.659873009
#> 940   0.6227374943 -0.2168854951  0.187543098
#> 941   0.8232018580 -0.0051218626  0.782433057
#> 942   0.6098232444 -0.4963677953  0.132364734
#> 943   0.6115525812 -0.1057183121  0.218130272
#> 944   0.3679628961 -0.2008560333  0.739633660
#> 945   0.5291002550 -0.4489777412  0.623927720
#> 946   0.4553942684 -0.8537614662 -0.125882944
#> 947   0.5235194991 -0.1957509569  0.840489397
#> 948   0.6778908091 -0.2605755501  0.395395440
#> 949   0.2655205791 -0.5304823734  0.272139995
#> 950   0.5504524797 -0.7053095951 -0.299875336
#> 951   0.6329342360 -0.6101258455 -0.400007225
#> 952   0.3629469635 -0.2810448857  0.852312028
#> 953   0.3499918296 -0.8190976694 -0.212063154
#> 954   0.3601234917 -0.5394360605  0.480074203
#> 955   0.5433965699 -0.4464294157 -0.241145627
#> 956   0.7538144473 -0.3985421957 -0.100245093
#> 957   0.5969094350 -0.2292333333 -0.064628299
#> 958   0.2665054723 -0.6723683079 -0.181904543
#> 959   0.6388672319 -0.1362676141  0.119108183
#> 960   0.5715617911 -0.1156509481  0.472290844
#> 961   0.2852706593 -0.7362776816 -0.400417391
#> 962   0.5253747746 -0.5815683001  0.142478986
#> 963   0.4933059008 -0.6115630032  0.447209091
#> 964   0.5982871028 -0.5379908450  0.276593459
#> 965  -0.4493365086 -0.3979186909 -0.166272439
#> 966   0.6150808370 -0.5238562767  0.257365615
#> 967   0.6788737888 -0.4659663463  0.060461962
#> 968   0.5616941139 -0.2809527294  0.297091597
#> 969   0.5421219512 -0.2849931939  1.010593465
#> 970   0.4937559543 -0.1129487623  0.415081505
#> 971   0.7567408467 -0.1839804420  0.797065268
#> 972   0.0128249332 -0.3767260647  0.386346272
#> 973   0.7399693979 -0.0672539739  0.693195300
#> 974   0.5253747746 -0.5815683001  0.142478986
#> 975   0.3705406211 -1.0743384768 -0.866084247
#> 976   0.5691954434 -0.1147738410  0.305273215
#> 977   0.5912577336  0.0149429477  0.387597564
#> 978   0.5092180342 -0.4555032557  0.198161956
#> 979   0.7567408467 -0.1839804420  0.797065268
#> 980   0.5000431348 -0.3830180801  0.101331134
#> 981   0.6660111640 -0.1697013758  1.011927759
#> 982   0.7594888787 -0.1932041944  0.170562056
#> 983   0.7829491150 -0.2073734649  0.205880290
#> 984   0.5000431348 -0.3830180801  0.101331134
#> 985   0.7149297841 -0.4347092248  0.540800663
#> 986   0.3341324139 -0.1288283825  0.382159757
#> 987   0.6899945273 -0.3199031945  0.287050335
#> 988   0.6375788045 -0.2743785411  0.246959657
#> 989   0.8642959456  0.0486905588  0.950655162
#> 990   0.2083315292 -0.5210125725  0.465852148
#> 991   0.8109542957 -0.0721907924  0.200683766
#> 992   0.4707340518 -0.0239501763  0.633268482
#> 993   0.8245433447 -0.1677040627  0.595087473
#> 994   0.4197952063 -0.6004583673 -0.359973091
#> 995   0.4810060762 -0.3692849963  0.289283635
#> 996   0.9366517362 -0.0544731896  0.188005267
#> 997   0.3725976481 -0.5368313621  0.285548506
#> 998   0.5091875852  0.0796868924  0.755563987
#> 999   0.5698604385 -0.0380164609  0.266038254
#> 1000  0.5664066291 -0.2397275618  0.482918870
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl   ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>    <dbl>  <dbl>
#>  1      1        4       70        1  0.285 -0.736   -0.400
#>  2      2        4       70        1  0.339 -0.962   -0.530
#>  3      3        4       70        1  0.804 -0.270    0.407
#>  4      4        4       70        1  0.675 -0.00900  1.03 
#>  5      5        4       70        1  0.468 -0.611   -0.298
#>  6      6        4       70        1  0.560 -0.713   -0.499
#>  7      7        4       70        1  0.539 -0.695   -0.101
#>  8      8        4       70        1  0.206 -1.14    -0.322
#>  9      9        4       70        1  0.169  0.259    0.969
#> 10     10        4       70        1  0.675 -0.00900  1.03 
#> # ℹ 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  5.32  33.5 0.670 0.159 21.9      1.22e+3  734.
#> 2      1    14     4    70     1  5.32  33.5 0.670 0.159  8.81     2.00e-1  295.
#> 3      2     1     4    70     1  5.62  26.8 0.589 0.210 24.3      1.29e+3  651.
#> 4      2    14     4    70     1  5.62  26.8 0.589 0.210  6.45     6.13e-1  172.
#> 5      3     1     4    70     1  8.93  53.4 1.50  0.167 23.2      6.64e+2 1239.
#> 6      3    14     4    70     1  8.93  53.4 1.50  0.167  4.23     2.20e-6  226.
#> # ℹ 1,994 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>