Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e4,n_resample=1e3)
#> $eta
#>            ETA_Cl       ETA_Vc       ETA_Ka
#> 1     0.053361947 -0.341528010  0.272716434
#> 2     0.472635119 -0.741095989 -0.065169694
#> 3     0.547767291 -0.252866561  0.315929795
#> 4     0.427554073 -0.427261280 -0.207783326
#> 5     0.506709496 -0.407749182  0.208198039
#> 6     0.473718372 -0.634364088  0.390832268
#> 7     0.677071458 -0.315218867 -0.080501084
#> 8     0.471280262 -0.398442741  0.183141512
#> 9     0.475733949 -0.430083081  0.388328364
#> 10    0.661202136 -0.680030606  0.057358509
#> 11    0.301939590 -0.451233351 -0.057097246
#> 12    0.481825251 -0.278347628  0.510698225
#> 13    0.484163383 -0.683967121 -0.212298154
#> 14    0.589598619 -0.092072509  0.672610627
#> 15    0.535955679 -0.545647015  0.304434435
#> 16    0.218451435 -0.988179629 -0.720029509
#> 17   -0.096552342 -0.660876263  0.048506590
#> 18    0.687669758 -0.392069758 -0.034096251
#> 19    0.317370524 -0.589813855 -0.089744284
#> 20    0.107786818 -0.564921908 -0.131944556
#> 21    0.783150396 -0.295291183  0.834707934
#> 22    0.157456683 -0.489386737  0.155004154
#> 23    0.406810705 -0.367531479  0.263775982
#> 24    0.692086002 -0.182265677  0.171684228
#> 25    0.589598619 -0.092072509  0.672610627
#> 26    0.334230867 -1.020516294 -0.370572330
#> 27    0.689130649 -0.195956714  0.087406622
#> 28    0.576465526 -0.458554733  0.155482021
#> 29    0.529452248 -0.555899306  0.763986170
#> 30    0.612731324 -0.443474655 -0.163805448
#> 31    0.400756831 -0.382086982  0.292303221
#> 32    0.471476238 -0.580054590  0.266056851
#> 33    0.315053319 -0.680594069 -0.256442697
#> 34    0.335255736 -0.640096142  0.630368503
#> 35    0.558266275 -0.386087482  0.225396611
#> 36    0.716004876 -0.384440020  0.115428619
#> 37    0.548660108  0.092601328  0.547028356
#> 38    0.733318039 -0.163650427  0.408582344
#> 39    0.380307211 -0.609581641  0.061237774
#> 40    0.301539429 -0.745204876 -0.269036319
#> 41    0.484566110 -0.405492822  0.977401340
#> 42    0.556903902 -0.485405870  0.241898507
#> 43    0.562511425 -0.356322437  0.288030749
#> 44    0.077704099 -0.036111811  0.403223067
#> 45    0.715930805 -0.237557854  0.363175260
#> 46    0.491249441 -0.619117865 -0.335577399
#> 47    0.514948977 -0.357027182  0.263416702
#> 48    0.507848878 -0.355178174  0.323898763
#> 49    0.473487588 -0.144163895 -0.039827853
#> 50    0.491141896 -0.148664363  1.096080857
#> 51    0.728733459 -0.191632228  0.356046547
#> 52    0.588852335 -0.706873567 -0.033828700
#> 53    0.629594377 -0.445773583  0.019999732
#> 54    0.332607497 -1.046989074 -0.527555013
#> 55    0.479253547 -0.631312672 -0.124256575
#> 56    0.548727519 -0.886313488 -0.397433073
#> 57   -0.108165798 -1.130836312 -0.232829522
#> 58    0.231261487 -1.318258864 -0.722582189
#> 59    0.624863356 -0.521893908  0.061672540
#> 60    0.490904809 -0.194093349  0.251047274
#> 61    0.621260720 -0.187152603  0.451752569
#> 62    0.578365466 -0.599803319 -0.036326004
#> 63    0.717335386 -0.377390453  0.122611949
#> 64    0.710690610 -0.265988309  0.126670153
#> 65    0.470964779 -0.427066564  0.628711777
#> 66    0.569785952 -0.621264822 -0.100719148
#> 67    0.502455462 -0.403643383  0.384652424
#> 68    0.471476238 -0.580054590  0.266056851
#> 69    0.623497556 -0.396417475  0.696783722
#> 70    0.834327375  0.003089501  1.147415026
#> 71    0.629594377 -0.445773583  0.019999732
#> 72    0.256901024 -0.452492316  0.384762561
#> 73    0.645599064 -0.329772791  0.346767560
#> 74    0.622366278 -0.335431213  0.403639193
#> 75    0.694959630 -0.528384725  0.265240817
#> 76    0.789275797 -0.255221390  0.353113702
#> 77    0.559870859 -0.315033682  0.108386483
#> 78    0.717335386 -0.377390453  0.122611949
#> 79    0.325794322 -0.944885676  0.245836337
#> 80    0.662363097  0.261644741  0.659833936
#> 81    0.574054588 -0.784169471 -0.657284581
#> 82    0.645599064 -0.329772791  0.346767560
#> 83    0.587469224 -0.732797490  0.257845126
#> 84    0.518468014 -0.673856603 -0.098651458
#> 85    0.014090597  0.193510875  0.852483770
#> 86    0.114336734 -0.735527454 -0.286902720
#> 87    0.755968248 -0.054718061  0.905220257
#> 88    0.544970701 -0.195342460  0.838750626
#> 89    0.510980283 -0.706409093  0.263288682
#> 90    0.731531337 -0.239265173  0.055947802
#> 91    0.376703656 -0.960130847 -0.411772439
#> 92    0.441954981 -0.650621254  0.403221059
#> 93    0.698046169 -0.302546502 -0.062953201
#> 94    0.583052118 -0.353427005 -0.004608063
#> 95    0.393658070 -0.268394274  0.824067307
#> 96    0.599714064 -0.320343662  0.642805664
#> 97    0.623497556 -0.396417475  0.696783722
#> 98    0.514948977 -0.357027182  0.263416702
#> 99    0.569785952 -0.621264822 -0.100719148
#> 100   0.510980283 -0.706409093  0.263288682
#> 101   0.743438104  0.024968877  0.365781110
#> 102   0.630924343 -0.171855369  0.096669511
#> 103   0.567565910 -0.438839169  0.443262041
#> 104   0.466263602 -0.608598545  0.129954751
#> 105   0.600129004 -0.333581899  0.822580654
#> 106   0.847767307 -0.184590438  0.343495171
#> 107   0.466966894 -0.450769334  0.597882629
#> 108   0.435359731 -0.159653757  0.220402666
#> 109   0.809632793  0.227085555  0.772847324
#> 110   0.510492860 -0.593411316 -0.091549530
#> 111   0.839119601 -0.140498340  0.200506988
#> 112   0.511553028 -0.835461208 -0.150221163
#> 113   0.789275797 -0.255221390  0.353113702
#> 114   0.258265294 -0.883009165 -0.162975047
#> 115   0.421893816 -0.548447869  0.496571998
#> 116   0.508237061 -0.362357103  0.036013563
#> 117   0.729413802 -0.410456101  0.460523291
#> 118   0.521447561 -0.176033927 -0.126055612
#> 119   0.648965822 -0.464801613  0.097044803
#> 120   0.544970701 -0.195342460  0.838750626
#> 121   0.645599064 -0.329772791  0.346767560
#> 122   0.535956380 -0.434812332  0.236953028
#> 123   0.209451047 -0.555185715 -0.195261914
#> 124   0.466263602 -0.608598545  0.129954751
#> 125   0.276460066 -0.298863011  0.364951265
#> 126   0.834327375  0.003089501  1.147415026
#> 127   0.879524497 -0.187972247  0.783060320
#> 128   0.490729004 -0.098758016  0.717582804
#> 129   0.579882148 -0.055661136  0.445297457
#> 130   0.370534555 -0.621058776  0.068934840
#> 131  -0.029585885 -0.312168804  0.127931712
#> 132   0.696549882 -0.062358466  0.352068233
#> 133   0.510492860 -0.593411316 -0.091549530
#> 134   0.571143478 -0.789874450 -0.304266860
#> 135   0.512079663 -0.003084482  0.790080717
#> 136   0.508237061 -0.362357103  0.036013563
#> 137   0.598228647 -0.242856119  0.743549155
#> 138   0.688360130 -0.099866427  0.913214611
#> 139   0.735016918  0.118077848  0.841383580
#> 140   0.510492860 -0.593411316 -0.091549530
#> 141   0.506709496 -0.407749182  0.208198039
#> 142   0.544970701 -0.195342460  0.838750626
#> 143   0.797736381  0.076459083  0.930059374
#> 144   0.811233687 -0.211153316 -0.054900740
#> 145   0.697667616 -0.319408497  0.172218174
#> 146   0.708471597 -0.434017570  0.044575575
#> 147  -0.147296865  0.115836637  1.026045591
#> 148   0.582287905 -0.476709698  0.292474390
#> 149   0.555151567 -0.229736738  0.260140351
#> 150   0.610839146 -0.344448344 -0.192998995
#> 151   0.645599064 -0.329772791  0.346767560
#> 152   0.616586516 -0.208932157  0.149071199
#> 153   0.692911672 -0.337044200 -0.336792577
#> 154   0.614465979 -0.233084209  0.259892060
#> 155   0.513524877 -0.089290801  0.923979968
#> 156   0.589598619 -0.092072509  0.672610627
#> 157   0.639820441 -0.433482798  0.222440517
#> 158   0.446760644 -0.346601330 -0.195219391
#> 159   0.055776158 -1.040653510 -0.616433877
#> 160   0.452365749 -0.257819561  0.516648044
#> 161   0.466006845 -0.093816586  0.617231900
#> 162   0.698046169 -0.302546502 -0.062953201
#> 163   0.723155972 -0.300950803  0.059150575
#> 164   0.315327858 -0.225966703  0.163587907
#> 165   0.688360130 -0.099866427  0.913214611
#> 166   0.605281145 -0.407340100  0.057707445
#> 167   0.247575511 -0.682561809 -0.255694476
#> 168   0.332741103 -0.410690728  0.239215478
#> 169   0.588852335 -0.706873567 -0.033828700
#> 170   0.147453897 -0.336422237  0.189150714
#> 171   0.525779881 -0.017422543  0.567784855
#> 172   0.576718428 -0.576632318  0.500241471
#> 173   0.624863356 -0.521893908  0.061672540
#> 174   0.730567439 -0.295025880  0.765974509
#> 175   0.343013968 -0.904065209 -0.536534668
#> 176   0.578365466 -0.599803319 -0.036326004
#> 177   0.054167321 -0.147999771  0.438078071
#> 178   1.045027706  0.030573383 -0.097943834
#> 179   0.194827665 -0.837501931 -0.192582325
#> 180   0.559870859 -0.315033682  0.108386483
#> 181   0.544970701 -0.195342460  0.838750626
#> 182   0.561492706 -0.417621880  0.470676139
#> 183   0.540823621 -0.059611886  0.871581919
#> 184   0.391869681 -0.525408580  0.214565814
#> 185   0.309148851 -0.226393647 -0.022847620
#> 186   0.182623794 -0.019760369  0.316785608
#> 187   0.467967563 -0.411117446  0.379598545
#> 188   0.527930852 -0.642134104 -0.312493262
#> 189   0.392852885 -0.214356119  0.246423163
#> 190   0.650229171 -0.470913879  0.149853296
#> 191   0.427235164 -0.206225061  0.184503104
#> 192   0.182237576 -0.449763188  0.116701011
#> 193   0.728733459 -0.191632228  0.356046547
#> 194   0.484163383 -0.683967121 -0.212298154
#> 195   0.508708283 -0.608508875  0.068836900
#> 196   0.698046169 -0.302546502 -0.062953201
#> 197   0.576104943 -0.372661203  0.336900280
#> 198   0.744905644 -0.513819546  0.410610861
#> 199   0.358533256 -0.626925914 -0.092360138
#> 200   0.466966894 -0.450769334  0.597882629
#> 201   0.412626006 -0.854079059  0.110529091
#> 202   0.476064907 -0.047138996  0.393806538
#> 203   0.460325012 -0.636086494  0.170382173
#> 204   0.671062866 -0.019067388  0.532598792
#> 205   0.466263602 -0.608598545  0.129954751
#> 206   0.350233404 -0.812797803 -0.357055318
#> 207   0.784828547 -0.181482827  0.368051536
#> 208   0.604914554 -0.430453022  0.121489044
#> 209   0.657735722 -0.050564790  0.606703525
#> 210   0.472635119 -0.741095989 -0.065169694
#> 211   0.699435607 -0.399497951  0.281178641
#> 212   0.598228647 -0.242856119  0.743549155
#> 213   0.596686529 -0.394023374  0.159731250
#> 214   0.404564444  0.212267147  0.756628842
#> 215   0.499855276 -0.484013087 -0.051464503
#> 216   0.231691364 -1.077290737 -0.929412416
#> 217   0.167191609 -0.544180738 -0.022092221
#> 218   0.681178741 -0.388469117  0.779885222
#> 219   0.514059474 -0.202810476  0.732146349
#> 220   0.596686529 -0.394023374  0.159731250
#> 221   0.705511383 -0.404472349  0.364539591
#> 222   0.490729004 -0.098758016  0.717582804
#> 223   0.510492860 -0.593411316 -0.091549530
#> 224   0.514948977 -0.357027182  0.263416702
#> 225   0.764459140 -0.423797969  0.192808912
#> 226   0.576465526 -0.458554733  0.155482021
#> 227   0.729413802 -0.410456101  0.460523291
#> 228   0.676170513 -0.236108985  0.590526780
#> 229   0.709750286 -0.148070238  0.161922362
#> 230   0.460325012 -0.636086494  0.170382173
#> 231   0.598228647 -0.242856119  0.743549155
#> 232   0.671031021 -0.536799695 -0.359795997
#> 233   0.472384450 -0.437704232 -0.096382829
#> 234   0.396870399 -0.125637947  0.657506465
#> 235   0.358995813 -0.361103778 -0.019359913
#> 236   0.372925415 -0.516322216  1.217648361
#> 237   0.728733459 -0.191632228  0.356046547
#> 238   0.441793964  0.132788640  1.709588496
#> 239   0.336992772 -0.436122573  0.579505911
#> 240   0.296525970 -0.846507709 -0.296010446
#> 241   0.631921169 -0.138149197 -0.004667471
#> 242   0.592334381 -0.163333996  0.877664243
#> 243   0.672962145 -0.356552151  0.143689040
#> 244   0.544852673 -0.201059071  0.282673685
#> 245   0.740476664 -0.240240562  0.340960638
#> 246   0.747043503  0.131917263  0.538480679
#> 247   0.429614681 -0.873288170 -0.659640195
#> 248   0.406076685 -0.803118279  0.178220999
#> 249   0.839119601 -0.140498340  0.200506988
#> 250   0.784828547 -0.181482827  0.368051536
#> 251   0.435042153 -0.197214804  0.264368375
#> 252   0.472635119 -0.741095989 -0.065169694
#> 253   0.540006601  0.081043539  0.699828673
#> 254   0.637760262 -0.463082478 -0.174278026
#> 255   0.476125117 -0.184527422  0.323170327
#> 256   0.525495963 -0.672309950  0.421487499
#> 257   0.565361095 -0.271472045  0.708351955
#> 258   0.598228647 -0.242856119  0.743549155
#> 259   0.365875100 -0.648339519 -0.017638770
#> 260   0.602069420  0.010863627  0.329227766
#> 261   0.720645162 -0.213133721  0.350275947
#> 262   0.231623573 -0.602792648 -0.464393941
#> 263   0.723155972 -0.300950803  0.059150575
#> 264   0.419356917 -0.373089486  0.068529232
#> 265   0.506709496 -0.407749182  0.208198039
#> 266   0.602816859 -0.266935884  0.390825846
#> 267   0.599714064 -0.320343662  0.642805664
#> 268   0.518091657 -0.951815867 -0.840680159
#> 269   0.504829261 -0.446141009  0.240062411
#> 270   0.365840289 -0.545704913 -0.253872487
#> 271   0.676170513 -0.236108985  0.590526780
#> 272   0.647002797  0.064058279  0.689308374
#> 273   0.539359271 -0.152804971  0.275555286
#> 274   0.472635119 -0.741095989 -0.065169694
#> 275   0.535955679 -0.545647015  0.304434435
#> 276   0.576465526 -0.458554733  0.155482021
#> 277   0.629594377 -0.445773583  0.019999732
#> 278   0.809168558 -0.182118445  0.527461958
#> 279   0.552615951 -0.362418612  0.505508252
#> 280   0.289171219 -0.308671111  0.354308636
#> 281   0.495103241 -0.305228496  0.497507336
#> 282   0.332607497 -1.046989074 -0.527555013
#> 283   0.514902499 -0.730449233  0.003881155
#> 284   0.723155972 -0.300950803  0.059150575
#> 285   0.307554913 -0.894263353  0.019010154
#> 286   0.416725398 -0.467509417 -0.352511346
#> 287   0.316881820 -0.937514363 -0.126895461
#> 288   0.359028009 -0.334448148  0.343437791
#> 289   0.322574514 -0.781564328 -0.244781980
#> 290   0.555281262 -0.627854559  0.094655927
#> 291   0.688360130 -0.099866427  0.913214611
#> 292   0.676170513 -0.236108985  0.590526780
#> 293   0.311863916 -0.762926569 -0.549696871
#> 294   0.600129004 -0.333581899  0.822580654
#> 295   0.641675574 -0.389087053  0.136973743
#> 296   0.212362266 -0.591679213 -0.058180327
#> 297   0.418357219 -0.237598425  0.452734390
#> 298   0.388406678 -0.330889421  0.536865732
#> 299   0.440609071 -0.589217348 -0.055963256
#> 300   0.697667616 -0.319408497  0.172218174
#> 301   0.510492860 -0.593411316 -0.091549530
#> 302   0.576718428 -0.576632318  0.500241471
#> 303   0.427677094 -0.251846583  0.260910765
#> 304   0.511553028 -0.835461208 -0.150221163
#> 305   0.293688682 -0.423787072  0.348544982
#> 306   0.506205102  0.237682890  1.386998486
#> 307   0.357337514  0.088079965  0.761451807
#> 308   0.373642622 -0.229046082 -0.048165667
#> 309   0.688360130 -0.099866427  0.913214611
#> 310   0.612535986 -0.498093881  0.219012511
#> 311   0.276460066 -0.298863011  0.364951265
#> 312   0.306317147 -0.779163078 -0.068343588
#> 313   0.809168558 -0.182118445  0.527461958
#> 314   0.571143478 -0.789874450 -0.304266860
#> 315   0.546986215 -0.179558305  0.446572855
#> 316   0.527930852 -0.642134104 -0.312493262
#> 317   0.634009533 -0.311181860 -0.186812202
#> 318   0.353306193 -0.678115794  0.219901679
#> 319   0.350233404 -0.812797803 -0.357055318
#> 320   0.032024125 -0.607830152 -0.105633665
#> 321   0.576718428 -0.576632318  0.500241471
#> 322   0.778149979 -0.395180726 -0.097588309
#> 323   0.200300982 -1.124933595  0.085054136
#> 324   0.585181508 -0.212724118  0.469474897
#> 325   0.556903902 -0.485405870  0.241898507
#> 326   0.470650421 -0.651280300  0.650359067
#> 327   0.648965822 -0.464801613  0.097044803
#> 328   0.468822054 -0.267783733  0.729192618
#> 329   0.521866863 -0.446301960  0.636558007
#> 330   0.471476238 -0.580054590  0.266056851
#> 331   0.476589232 -0.538176738  0.587041646
#> 332   0.700327528 -0.069490751  0.318927033
#> 333   0.629031344 -0.122244631  0.259316874
#> 334   0.657735722 -0.050564790  0.606703525
#> 335  -0.046906556 -0.251308295  0.498287854
#> 336   0.267897067 -0.723829640 -0.630964353
#> 337  -0.179584742 -0.402858924  0.130355230
#> 338   0.479808283 -0.567765962  0.430450818
#> 339   0.560513031 -0.107610642  0.230048245
#> 340   0.146629488 -0.226181610  0.156317464
#> 341   0.668319042 -0.580434479  0.341338668
#> 342   0.552097416 -0.415673567 -0.148176102
#> 343   0.727922625 -0.258244945  0.379665600
#> 344   0.136789496 -0.926945321  0.284904121
#> 345   0.385649574 -0.634303092  0.166856528
#> 346   0.834327375  0.003089501  1.147415026
#> 347   0.227479565  0.052633747  0.887440762
#> 348   0.500604043 -0.404078445  0.208327626
#> 349   0.490729004 -0.098758016  0.717582804
#> 350   0.740476664 -0.240240562  0.340960638
#> 351   0.472635119 -0.741095989 -0.065169694
#> 352   0.509827971 -0.767730392 -0.221408754
#> 353   0.509827971 -0.767730392 -0.221408754
#> 354   0.484163383 -0.683967121 -0.212298154
#> 355   0.556781770 -0.589792861 -0.030183776
#> 356   0.714007954 -0.343944299 -0.193361697
#> 357   0.755968248 -0.054718061  0.905220257
#> 358   0.700215520 -0.265836296  0.548589752
#> 359   0.449246489 -0.508004189 -0.367382339
#> 360  -0.273001929  0.033273919  0.366556599
#> 361   0.534276746 -0.305805446  0.113855860
#> 362   0.700215520 -0.265836296  0.548589752
#> 363   0.339561245 -0.481475513  0.685007992
#> 364   0.682983800 -0.391370515  0.494427575
#> 365   0.722371857 -0.448832372  0.384556017
#> 366   0.565361095 -0.271472045  0.708351955
#> 367   0.808023833 -0.295245448  0.699697855
#> 368   0.582287905 -0.476709698  0.292474390
#> 369   0.577923689 -0.041950190  0.573787818
#> 370   0.807614553 -0.379683972 -0.125451523
#> 371   0.859317396  0.004521422  0.772519333
#> 372   0.565998245 -0.590595674  0.277119306
#> 373   0.604914554 -0.430453022  0.121489044
#> 374   0.543519249 -0.278542291  0.187995020
#> 375   0.513793782 -0.335532237 -0.275652804
#> 376   0.833257464 -0.332282630  0.217797054
#> 377   0.835577465  0.029923946  0.738029529
#> 378   0.564953501 -0.317927715  0.009957633
#> 379   0.522022621 -0.571328718 -0.043149585
#> 380   0.697667616 -0.319408497  0.172218174
#> 381   0.316881820 -0.937514363 -0.126895461
#> 382   0.512934443 -0.195307927  0.221388465
#> 383   0.730567439 -0.295025880  0.765974509
#> 384   0.612535986 -0.498093881  0.219012511
#> 385   0.472635119 -0.741095989 -0.065169694
#> 386   0.719930716 -1.038870722 -1.119500825
#> 387   0.459019500 -0.469119771  0.195234914
#> 388   0.887926577  0.209335224  1.050885869
#> 389   0.024080582 -0.798139871 -0.429664846
#> 390   0.671062866 -0.019067388  0.532598792
#> 391   0.317678108 -1.057286613  0.115078217
#> 392   0.442541099 -0.360970510 -0.037964873
#> 393   0.434625771 -0.318096663  0.259433562
#> 394   0.334230867 -1.020516294 -0.370572330
#> 395   0.665555608 -0.275442257  0.116165197
#> 396   0.535956380 -0.434812332  0.236953028
#> 397   0.621260720 -0.187152603  0.451752569
#> 398   0.605281145 -0.407340100  0.057707445
#> 399   0.768561092 -0.244044848  0.390640504
#> 400   0.423264381 -0.251435460  0.265689248
#> 401   0.624099658  0.046744993  0.818829239
#> 402   0.619338856 -0.344318103  0.221888251
#> 403   0.535955679 -0.545647015  0.304434435
#> 404   0.700021389 -0.072968528  0.378302430
#> 405   0.334230867 -1.020516294 -0.370572330
#> 406   0.558266275 -0.386087482  0.225396611
#> 407   0.446734986 -0.275965224 -0.282870194
#> 408   0.476589232 -0.538176738  0.587041646
#> 409   0.612731324 -0.443474655 -0.163805448
#> 410   0.451300983 -0.097812095  0.305870001
#> 411   0.723343385 -0.474385132 -0.299240162
#> 412   0.629594377 -0.445773583  0.019999732
#> 413   0.521866863 -0.446301960  0.636558007
#> 414   0.343390983 -0.761629249  0.087067463
#> 415   0.700215520 -0.265836296  0.548589752
#> 416   0.418357219 -0.237598425  0.452734390
#> 417   0.484163383 -0.683967121 -0.212298154
#> 418   0.535956380 -0.434812332  0.236953028
#> 419   0.444038045 -0.758281659 -0.404794716
#> 420   0.078487827  0.149588300  0.546683231
#> 421   0.470650421 -0.651280300  0.650359067
#> 422   0.538070694 -0.342691512 -0.021002916
#> 423   0.613075617 -0.444652929  0.232309140
#> 424   0.720645162 -0.213133721  0.350275947
#> 425   0.100617949 -0.414121970  0.415878956
#> 426   0.466188312 -0.766927344 -0.602049064
#> 427   0.675796567 -0.430551247  0.221680113
#> 428   0.091491846 -0.206097629  0.162318960
#> 429   0.676170513 -0.236108985  0.590526780
#> 430   0.544302651 -0.482235053  0.725393796
#> 431   0.259678654 -0.576295143  0.082814601
#> 432   0.437909011 -0.586549619  0.080130717
#> 433   0.565998245 -0.590595674  0.277119306
#> 434   0.376056385 -0.583072594 -0.244077886
#> 435   0.688578313 -0.533861375 -0.180760969
#> 436   0.466263602 -0.608598545  0.129954751
#> 437   0.349941781 -0.846605164 -0.477556415
#> 438   0.316023335 -0.927091595  0.420584222
#> 439   0.558266275 -0.386087482  0.225396611
#> 440   0.809168558 -0.182118445  0.527461958
#> 441   0.793193152  0.183653027  0.383461617
#> 442   0.429614681 -0.873288170 -0.659640195
#> 443   0.585003985 -0.322544573  0.415537414
#> 444   0.716004876 -0.384440020  0.115428619
#> 445   0.488058458 -0.411372438  0.640136871
#> 446   0.699435607 -0.399497951  0.281178641
#> 447   0.379081533 -0.432984533 -0.086335395
#> 448   0.494878175 -0.638036859  0.158228036
#> 449   0.473354710 -0.317362887 -0.139026165
#> 450   0.507340089 -0.572900813 -0.007396039
#> 451   0.752456564 -0.122279383  0.507632494
#> 452   0.367708417 -1.040187139 -0.344119524
#> 453   0.578365466 -0.599803319 -0.036326004
#> 454   0.733318039 -0.163650427  0.408582344
#> 455   0.733318039 -0.163650427  0.408582344
#> 456   0.412626006 -0.854079059  0.110529091
#> 457   0.797736381  0.076459083  0.930059374
#> 458   0.755968248 -0.054718061  0.905220257
#> 459   0.585181508 -0.212724118  0.469474897
#> 460   0.455446153 -0.571801322 -0.085274649
#> 461   0.598228647 -0.242856119  0.743549155
#> 462   0.483018171 -0.210780588  0.180175293
#> 463   0.733318039 -0.163650427  0.408582344
#> 464   0.733318039 -0.163650427  0.408582344
#> 465   0.053674949 -0.150704925  0.062999382
#> 466   0.611287809 -0.074537871  0.562030942
#> 467   0.368014350 -0.439881924  0.125900776
#> 468   0.316922248 -0.519906517 -0.243462244
#> 469   0.416019700 -0.169376340 -0.012119084
#> 470   0.723343385 -0.474385132 -0.299240162
#> 471   0.508708283 -0.608508875  0.068836900
#> 472   0.841124219  0.460930327  1.037344738
#> 473   0.306317147 -0.779163078 -0.068343588
#> 474   0.218451435 -0.988179629 -0.720029509
#> 475   0.600129004 -0.333581899  0.822580654
#> 476   0.269922977 -0.828719890 -0.214810195
#> 477   0.783150396 -0.295291183  0.834707934
#> 478   0.523956348 -0.516210995  0.062567672
#> 479   0.218451435 -0.988179629 -0.720029509
#> 480   0.186042496 -0.909015860  0.567679932
#> 481   0.510311041 -0.466439155 -0.131264857
#> 482   0.784828547 -0.181482827  0.368051536
#> 483   0.169971647 -0.324976756  0.493222840
#> 484   0.576465526 -0.458554733  0.155482021
#> 485   0.619338856 -0.344318103  0.221888251
#> 486   0.420712811 -0.244753342  0.274763379
#> 487   0.512079663 -0.003084482  0.790080717
#> 488   0.750012789 -0.291515525  0.962228906
#> 489   0.500355728 -0.572267898  0.728430771
#> 490   0.610382814 -0.296682521  0.359992963
#> 491   0.504389276 -0.322054270  0.184951136
#> 492   0.233140458 -0.626381844  0.209017380
#> 493   0.332607497 -1.046989074 -0.527555013
#> 494   0.169736366  0.120406092  0.981683392
#> 495   0.418357219 -0.237598425  0.452734390
#> 496   0.373407186 -0.394281632  0.634821266
#> 497   0.615218606 -0.476257104  0.151293659
#> 498   0.560513031 -0.107610642  0.230048245
#> 499   0.525809277 -0.165208780  0.576561043
#> 500   0.743693053 -0.096255842  0.231684450
#> 501   0.824699448  0.020275786  0.725201443
#> 502   0.522770226  0.152189452  0.843810821
#> 503   0.582287905 -0.476709698  0.292474390
#> 504   0.456739875 -0.430906372  0.405546058
#> 505   0.365577527 -0.679465752  0.177177938
#> 506   0.567910417 -0.466958694  0.383460810
#> 507   0.697667616 -0.319408497  0.172218174
#> 508   0.418357219 -0.237598425  0.452734390
#> 509   0.540867992 -0.405106415 -0.160866253
#> 510   0.456739875 -0.430906372  0.405546058
#> 511   0.468822054 -0.267783733  0.729192618
#> 512   0.483575187 -0.303292806  0.127432911
#> 513   0.393820434 -0.344561152  0.618696437
#> 514   0.561492706 -0.417621880  0.470676139
#> 515   0.796855406 -0.043444130  0.255648461
#> 516   0.574290379 -0.135392155  0.327476841
#> 517   0.527930852 -0.642134104 -0.312493262
#> 518   0.290305925 -0.122413238  0.541278321
#> 519   0.111943767 -0.310003968  0.318940736
#> 520   0.475078596 -0.442158459  0.553034074
#> 521   0.365778637 -0.727410419  0.008370239
#> 522   0.621301685 -0.466595896  0.523933133
#> 523   0.555281262 -0.627854559  0.094655927
#> 524   0.797736381  0.076459083  0.930059374
#> 525   0.500604043 -0.404078445  0.208327626
#> 526   0.382874735 -0.381056257  0.424939073
#> 527   0.008157348 -0.429907975  0.710039919
#> 528   0.495103241 -0.305228496  0.497507336
#> 529   0.450601057 -0.408527983  1.226906372
#> 530   0.554008293 -0.178611534  0.408941088
#> 531   0.134943503 -0.093193646  0.259071435
#> 532   0.822812871 -0.121361285  0.538382153
#> 533   0.398645118 -0.662916922  0.358079243
#> 534   0.429614681 -0.873288170 -0.659640195
#> 535   0.826460962 -0.192167871  0.368703125
#> 536   0.421973058 -0.261465724  0.051871276
#> 537   0.367708417 -1.040187139 -0.344119524
#> 538   0.226919562 -0.407479147  0.339193348
#> 539   0.735016918  0.118077848  0.841383580
#> 540   0.676170513 -0.236108985  0.590526780
#> 541   0.612731324 -0.443474655 -0.163805448
#> 542   0.615218606 -0.476257104  0.151293659
#> 543   0.362507180 -0.516982697  0.993093395
#> 544   0.562511425 -0.356322437  0.288030749
#> 545   0.641675574 -0.389087053  0.136973743
#> 546   0.475733949 -0.430083081  0.388328364
#> 547   0.033760677 -0.009299800  0.687162657
#> 548   0.657735722 -0.050564790  0.606703525
#> 549   0.841753836 -0.171923203  0.672306395
#> 550   0.715930805 -0.237557854  0.363175260
#> 551   0.479253547 -0.631312672 -0.124256575
#> 552   0.392484585 -0.842021613  0.574612912
#> 553   0.619338856 -0.344318103  0.221888251
#> 554   0.522022621 -0.571328718 -0.043149585
#> 555   0.534107132 -0.660837763  0.074836247
#> 556   0.368263665 -0.477792450  0.351461428
#> 557   0.160488814 -0.707106391 -0.454145885
#> 558   0.418920588 -0.274821529  0.502431468
#> 559   0.605281145 -0.407340100  0.057707445
#> 560   0.510980283 -0.706409093  0.263288682
#> 561   0.797736381  0.076459083  0.930059374
#> 562   0.376703656 -0.960130847 -0.411772439
#> 563   0.835577465  0.029923946  0.738029529
#> 564   0.716004876 -0.384440020  0.115428619
#> 565   0.207212804 -0.399747330  0.128105892
#> 566   0.647833972 -0.407775388  0.124895207
#> 567   0.612535986 -0.498093881  0.219012511
#> 568   0.479253547 -0.631312672 -0.124256575
#> 569   0.690332408 -0.588329765 -0.148003413
#> 570   0.632529436 -0.177515740  0.291739404
#> 571   0.473063444 -0.150367301  0.427446229
#> 572   0.523956348 -0.516210995  0.062567672
#> 573   0.472551853  0.211890689  0.877518983
#> 574   0.518468014 -0.673856603 -0.098651458
#> 575   0.642840298 -0.526412092  0.029857572
#> 576   0.700021389 -0.072968528  0.378302430
#> 577   0.510492860 -0.593411316 -0.091549530
#> 578   0.665555608 -0.275442257  0.116165197
#> 579   0.592334381 -0.163333996  0.877664243
#> 580   0.725438931 -0.502762561  0.102740349
#> 581   0.835577465  0.029923946  0.738029529
#> 582   0.437909011 -0.586549619  0.080130717
#> 583   0.548727519 -0.886313488 -0.397433073
#> 584   0.598228647 -0.242856119  0.743549155
#> 585   0.630166508 -0.585304844 -0.229130906
#> 586   0.423779297 -0.398168718  0.207204750
#> 587   0.583052118 -0.353427005 -0.004608063
#> 588   0.504389276 -0.322054270  0.184951136
#> 589   0.520427397 -0.115675609  0.860248444
#> 590   0.692993526 -0.371863347  0.052656686
#> 591   0.583052118 -0.353427005 -0.004608063
#> 592   0.184943385 -0.860120675 -0.416183465
#> 593   0.479253547 -0.631312672 -0.124256575
#> 594   0.688643373 -0.675495967 -0.397777981
#> 595   0.887926577  0.209335224  1.050885869
#> 596   0.745394591 -0.422966236  0.320391863
#> 597   0.587469224 -0.732797490  0.257845126
#> 598   0.406810705 -0.367531479  0.263775982
#> 599   0.538070694 -0.342691512 -0.021002916
#> 600   0.632529436 -0.177515740  0.291739404
#> 601   0.514902499 -0.730449233  0.003881155
#> 602   0.859317396  0.004521422  0.772519333
#> 603   0.613075617 -0.444652929  0.232309140
#> 604   0.308320513 -0.399710386  0.683235697
#> 605   0.514902499 -0.730449233  0.003881155
#> 606   0.472635119 -0.741095989 -0.065169694
#> 607   0.715930805 -0.237557854  0.363175260
#> 608   0.412626006 -0.854079059  0.110529091
#> 609   0.456739875 -0.430906372  0.405546058
#> 610   0.346526015 -0.352801752 -0.109622133
#> 611   0.467967563 -0.411117446  0.379598545
#> 612   0.431937737 -0.534882943 -0.268501866
#> 613   0.466263602 -0.608598545  0.129954751
#> 614   0.845752299 -0.120030114 -0.001917168
#> 615   0.032648850 -0.265610417  0.261840238
#> 616   0.720645162 -0.213133721  0.350275947
#> 617   0.468822054 -0.267783733  0.729192618
#> 618   0.444038045 -0.758281659 -0.404794716
#> 619   0.372925415 -0.516322216  1.217648361
#> 620   0.583052118 -0.353427005 -0.004608063
#> 621   0.722371857 -0.448832372  0.384556017
#> 622   0.700215520 -0.265836296  0.548589752
#> 623   0.585181508 -0.212724118  0.469474897
#> 624   0.272301844 -0.416801563  0.007418677
#> 625   0.630166508 -0.585304844 -0.229130906
#> 626   0.539388004 -0.604041645  0.547580623
#> 627   0.442000466 -0.472232420  0.285179941
#> 628   0.561492706 -0.417621880  0.470676139
#> 629   0.191525016  0.200034791  1.035396030
#> 630   0.499780649 -0.111140872  0.725720339
#> 631   0.435359731 -0.159653757  0.220402666
#> 632   0.381752219 -0.350697228  0.362931889
#> 633   0.277458878 -0.422219688  0.157985215
#> 634   0.061292761 -0.652024891  0.375360370
#> 635   0.621301685 -0.466595896  0.523933133
#> 636   0.172390977 -0.332700970  0.328481498
#> 637   0.924279816  0.227623800  0.576857384
#> 638   0.544852673 -0.201059071  0.282673685
#> 639   0.423264381 -0.251435460  0.265689248
#> 640   0.360547196 -0.284514910  0.415864902
#> 641   0.841753836 -0.171923203  0.672306395
#> 642   0.604501795 -0.481854343  0.388075926
#> 643   0.714440727 -0.274713040  0.087363327
#> 644  -0.054763160 -0.884045695 -0.393336372
#> 645   0.598228647 -0.242856119  0.743549155
#> 646   0.720645162 -0.213133721  0.350275947
#> 647   0.571143478 -0.789874450 -0.304266860
#> 648   0.709033064  0.167972021  0.415604061
#> 649   0.567565910 -0.438839169  0.443262041
#> 650   0.574054588 -0.784169471 -0.657284581
#> 651   0.768632119 -0.312756676 -0.076224627
#> 652   0.470015717 -0.550362898  0.397087374
#> 653   0.639948998 -0.455886196  0.180194112
#> 654   0.820256421 -0.412173238 -0.310068627
#> 655   0.642840298 -0.526412092  0.029857572
#> 656   0.491249441 -0.619117865 -0.335577399
#> 657   1.025186420 -0.044311291  0.698540608
#> 658   0.784828547 -0.181482827  0.368051536
#> 659   0.367708417 -1.040187139 -0.344119524
#> 660   0.534276746 -0.305805446  0.113855860
#> 661   0.423309985 -0.548422721  0.247683663
#> 662   0.726208714 -0.234145478  0.662839623
#> 663   0.463579631  0.070414228  0.410924279
#> 664   0.207787020 -0.231872833  0.245768835
#> 665   0.692993526 -0.371863347  0.052656686
#> 666   0.296525970 -0.846507709 -0.296010446
#> 667   0.231261487 -1.318258864 -0.722582189
#> 668   0.752456564 -0.122279383  0.507632494
#> 669   0.592334381 -0.163333996  0.877664243
#> 670   0.700021389 -0.072968528  0.378302430
#> 671   0.197394269 -0.296624439  0.136442911
#> 672   0.576104943 -0.372661203  0.336900280
#> 673   0.225286918 -0.808489978  0.105160740
#> 674  -0.116892464 -0.699106590  0.352706064
#> 675   0.456456259 -0.327331153  0.291013162
#> 676   0.672962145 -0.356552151  0.143689040
#> 677   0.423309985 -0.548422721  0.247683663
#> 678   0.447325419 -0.596427314  0.265652347
#> 679   0.483575187 -0.303292806  0.127432911
#> 680   0.642840298 -0.526412092  0.029857572
#> 681   0.826460962 -0.192167871  0.368703125
#> 682   0.692993526 -0.371863347  0.052656686
#> 683   0.269922977 -0.828719890 -0.214810195
#> 684   0.556903902 -0.485405870  0.241898507
#> 685   0.555151567 -0.229736738  0.260140351
#> 686   0.671062866 -0.019067388  0.532598792
#> 687   0.808023833 -0.295245448  0.699697855
#> 688   0.494878175 -0.638036859  0.158228036
#> 689   0.379636693 -0.674381863 -0.024557742
#> 690   0.677071458 -0.315218867 -0.080501084
#> 691   0.834327375  0.003089501  1.147415026
#> 692   0.567565910 -0.438839169  0.443262041
#> 693   0.391869681 -0.525408580  0.214565814
#> 694   0.316881820 -0.937514363 -0.126895461
#> 695   0.758190805 -0.374048089 -0.008281636
#> 696   0.538413914 -0.452983215  0.121437459
#> 697   0.579501539 -0.576453922 -0.004296412
#> 698   0.511576822 -0.421925205  0.479305741
#> 699   0.750012789 -0.291515525  0.962228906
#> 700   0.735016918  0.118077848  0.841383580
#> 701   0.472367998 -0.361098105  0.070294961
#> 702   0.624863356 -0.521893908  0.061672540
#> 703   0.728336946 -0.312502487  0.265202746
#> 704   0.381752219 -0.350697228  0.362931889
#> 705   0.520981751 -0.405529730 -0.273933302
#> 706   0.510980283 -0.706409093  0.263288682
#> 707   0.727922625 -0.258244945  0.379665600
#> 708   0.499780649 -0.111140872  0.725720339
#> 709   0.718323484 -0.687856613 -0.363431815
#> 710   0.543519249 -0.278542291  0.187995020
#> 711   0.227800349 -0.516922718 -0.213762906
#> 712   0.527930852 -0.642134104 -0.312493262
#> 713   0.577923689 -0.041950190  0.573787818
#> 714   0.312323777 -0.541156108  0.179625693
#> 715   0.826331253  0.207589230  0.711321060
#> 716   0.610382814 -0.296682521  0.359992963
#> 717   0.388406678 -0.330889421  0.536865732
#> 718   0.452770046 -0.131069827  0.401623021
#> 719   0.481879523 -0.725773643 -0.562010325
#> 720   0.479253547 -0.631312672 -0.124256575
#> 721   0.650229171 -0.470913879  0.149853296
#> 722   0.346732525 -0.541635042 -0.456455135
#> 723   0.499780649 -0.111140872  0.725720339
#> 724   0.298726633 -0.261059695  0.204744111
#> 725   0.382498094 -0.137115890  0.448746061
#> 726   0.539388004 -0.604041645  0.547580623
#> 727   0.664071943 -0.254055027  0.365480118
#> 728   0.470015717 -0.550362898  0.397087374
#> 729   0.491249441 -0.619117865 -0.335577399
#> 730   0.599714064 -0.320343662  0.642805664
#> 731   0.283011785 -0.209339791  0.437448807
#> 732   0.576465526 -0.458554733  0.155482021
#> 733   0.687669758 -0.392069758 -0.034096251
#> 734   0.714007954 -0.343944299 -0.193361697
#> 735   0.425776651 -0.883309780  0.009196556
#> 736   0.624863356 -0.521893908  0.061672540
#> 737   0.720645162 -0.213133721  0.350275947
#> 738   0.687669758 -0.392069758 -0.034096251
#> 739   0.241514692 -0.241047264  0.837107505
#> 740   0.676170513 -0.236108985  0.590526780
#> 741   0.613075617 -0.444652929  0.232309140
#> 742   0.412626006 -0.854079059  0.110529091
#> 743   0.172497331 -0.114595352  0.350195119
#> 744   0.748974697 -0.142336711  0.641035225
#> 745   0.473718372 -0.634364088  0.390832268
#> 746   0.523956348 -0.516210995  0.062567672
#> 747   0.720645162 -0.213133721  0.350275947
#> 748   0.571143478 -0.789874450 -0.304266860
#> 749   0.713809963 -0.102810982  0.177555263
#> 750   0.437909011 -0.586549619  0.080130717
#> 751   0.752456564 -0.122279383  0.507632494
#> 752   0.714867193 -0.207816501  0.049908915
#> 753   0.398041981  0.007830623  0.383046020
#> 754   0.495394939 -0.025729001  0.908683151
#> 755   0.623497556 -0.396417475  0.696783722
#> 756   0.510492860 -0.593411316 -0.091549530
#> 757   0.728336946 -0.312502487  0.265202746
#> 758   0.525495963 -0.672309950  0.421487499
#> 759   0.370534555 -0.621058776  0.068934840
#> 760   0.688360130 -0.099866427  0.913214611
#> 761   0.521799972  0.040381006  0.219542700
#> 762   0.623497556 -0.396417475  0.696783722
#> 763   0.535956380 -0.434812332  0.236953028
#> 764   0.691843660 -0.149003989  0.517732612
#> 765   0.472635119 -0.741095989 -0.065169694
#> 766   0.598228647 -0.242856119  0.743549155
#> 767   0.518091657 -0.951815867 -0.840680159
#> 768   0.387271110 -0.184906946  0.423275514
#> 769   0.690332408 -0.588329765 -0.148003413
#> 770   0.341773038 -0.122077435  0.434152201
#> 771   0.516446890 -0.080704320  0.296904345
#> 772   0.700021389 -0.072968528  0.378302430
#> 773   0.714294602  0.105261215  0.338558939
#> 774   0.491141896 -0.148664363  1.096080857
#> 775   0.202997187 -0.531674264 -0.239631772
#> 776   0.128849286 -0.583480591 -0.112269526
#> 777   0.555281262 -0.627854559  0.094655927
#> 778   0.321765216 -0.672243499  0.422703270
#> 779   0.521866863 -0.446301960  0.636558007
#> 780   0.621301685 -0.466595896  0.523933133
#> 781   0.615218606 -0.476257104  0.151293659
#> 782   0.506741057 -0.475489700  0.291290802
#> 783   0.755968248 -0.054718061  0.905220257
#> 784   0.755376804 -0.455685489 -0.266456510
#> 785   0.465160311 -0.381218919  0.279208336
#> 786   0.613075617 -0.444652929  0.232309140
#> 787  -0.393501103 -0.185324802  0.508011346
#> 788   0.665555608 -0.275442257  0.116165197
#> 789   0.538413914 -0.452983215  0.121437459
#> 790   0.510492860 -0.593411316 -0.091549530
#> 791   0.304039471  0.086196728  0.821211898
#> 792   0.565361095 -0.271472045  0.708351955
#> 793   0.552615951 -0.362418612  0.505508252
#> 794   0.148629226 -0.932379678 -0.663514322
#> 795   0.569785952 -0.621264822 -0.100719148
#> 796   0.459019500 -0.469119771  0.195234914
#> 797   0.376703656 -0.960130847 -0.411772439
#> 798   0.220505787 -0.589075190  0.074217477
#> 799   0.382874735 -0.381056257  0.424939073
#> 800   0.639551554  0.038642360  0.534834415
#> 801   0.223640372 -0.794952219  0.109512202
#> 802   0.748974697 -0.142336711  0.641035225
#> 803   0.547767291 -0.252866561  0.315929795
#> 804   0.022293067 -0.431278839  0.079519785
#> 805   0.447325419 -0.596427314  0.265652347
#> 806   0.555151567 -0.229736738  0.260140351
#> 807   0.941379734  0.015502295  0.176201834
#> 808   0.605281145 -0.407340100  0.057707445
#> 809   0.556781770 -0.589792861 -0.030183776
#> 810   0.375918896 -0.582732326  0.180065912
#> 811   0.481825251 -0.278347628  0.510698225
#> 812   0.435854454 -0.116056683 -0.014336449
#> 813   0.733318039 -0.163650427  0.408582344
#> 814   0.565998245 -0.590595674  0.277119306
#> 815   0.476589232 -0.538176738  0.587041646
#> 816   0.385250696 -0.109451656  0.785412760
#> 817   0.473718372 -0.634364088  0.390832268
#> 818   0.784828547 -0.181482827  0.368051536
#> 819   0.392003951 -0.645258834  0.316601908
#> 820   0.745394591 -0.422966236  0.320391863
#> 821   0.441954981 -0.650621254  0.403221059
#> 822   0.788441024 -0.246306887 -0.021677325
#> 823   0.748974697 -0.142336711  0.641035225
#> 824   0.834327375  0.003089501  1.147415026
#> 825   0.427035519 -0.295241111  0.836167373
#> 826   0.579501539 -0.576453922 -0.004296412
#> 827   0.612535986 -0.498093881  0.219012511
#> 828   0.602069420  0.010863627  0.329227766
#> 829   0.650229171 -0.470913879  0.149853296
#> 830   0.688578313 -0.533861375 -0.180760969
#> 831   0.209047241 -0.717188535 -0.000926498
#> 832   0.472635119 -0.741095989 -0.065169694
#> 833   0.079429787 -0.105829029  0.830537503
#> 834   0.622366278 -0.335431213  0.403639193
#> 835   0.727922625 -0.258244945  0.379665600
#> 836   0.507340089 -0.572900813 -0.007396039
#> 837   0.510576138 -0.411497170  0.482700513
#> 838   0.538413914 -0.452983215  0.121437459
#> 839   0.538413914 -0.452983215  0.121437459
#> 840   0.073208889 -0.903384768  0.131203120
#> 841   0.571143478 -0.789874450 -0.304266860
#> 842   0.592334381 -0.163333996  0.877664243
#> 843   0.280794694 -0.268952234 -0.014857137
#> 844   0.612731324 -0.443474655 -0.163805448
#> 845   0.413910972 -0.471283908 -0.079588721
#> 846   0.723155972 -0.300950803  0.059150575
#> 847   0.084521218 -0.658609057  0.153609589
#> 848   0.879524497 -0.187972247  0.783060320
#> 849   0.961029945 -0.075419443  0.479274008
#> 850   0.471280262 -0.398442741  0.183141512
#> 851   0.411576128 -0.319105557  0.337399304
#> 852   0.521866863 -0.446301960  0.636558007
#> 853   0.622366278 -0.335431213  0.403639193
#> 854   0.382271563 -0.685880550  0.776625670
#> 855   0.258334716 -0.051116307  0.759634072
#> 856   0.225286918 -0.808489978  0.105160740
#> 857   0.132108252 -0.298354829  0.541038780
#> 858   0.619338856 -0.344318103  0.221888251
#> 859   0.293688682 -0.423787072  0.348544982
#> 860   0.392003951 -0.645258834  0.316601908
#> 861   0.558266275 -0.386087482  0.225396611
#> 862   0.612535986 -0.498093881  0.219012511
#> 863   0.730567439 -0.295025880  0.765974509
#> 864   0.447325419 -0.596427314  0.265652347
#> 865   0.688578313 -0.533861375 -0.180760969
#> 866   0.632529436 -0.177515740  0.291739404
#> 867   0.720645162 -0.213133721  0.350275947
#> 868   0.176891589 -0.626664172  0.180735997
#> 869   0.664071943 -0.254055027  0.365480118
#> 870   0.381675425 -0.729614212  0.510084125
#> 871   0.567565910 -0.438839169  0.443262041
#> 872   0.245416044 -0.372612013  0.227923976
#> 873   0.252327808 -0.076266358  0.605026221
#> 874   0.723155972 -0.300950803  0.059150575
#> 875   0.824699448  0.020275786  0.725201443
#> 876   0.442000466 -0.472232420  0.285179941
#> 877   0.717335386 -0.377390453  0.122611949
#> 878   0.543005058  0.053674514  0.557457601
#> 879   0.502455462 -0.403643383  0.384652424
#> 880   0.165467692 -0.398091174  0.631132507
#> 881   0.604914554 -0.430453022  0.121489044
#> 882   0.475078596 -0.442158459  0.553034074
#> 883   0.859317396  0.004521422  0.772519333
#> 884   0.715930805 -0.237557854  0.363175260
#> 885   0.534107132 -0.660837763  0.074836247
#> 886   0.510311041 -0.466439155 -0.131264857
#> 887   0.516795062 -0.227749624  0.127542966
#> 888   0.571143478 -0.789874450 -0.304266860
#> 889   0.647833972 -0.407775388  0.124895207
#> 890   0.477475537 -0.832702361  0.313374501
#> 891   0.315193394 -0.646860972 -0.181449237
#> 892   0.588852335 -0.706873567 -0.033828700
#> 893   0.509827971 -0.767730392 -0.221408754
#> 894   0.244032379 -0.570471911 -0.236023371
#> 895   0.280464679 -0.287256124  0.389440005
#> 896   0.160578284 -0.448283463  0.095775971
#> 897   0.506741057 -0.475489700  0.291290802
#> 898   0.523956348 -0.516210995  0.062567672
#> 899   0.278351763 -0.232766745  0.074816815
#> 900   0.484566110 -0.405492822  0.977401340
#> 901   0.768632119 -0.312756676 -0.076224627
#> 902   0.431740816 -0.318525086  0.332178538
#> 903   0.713809963 -0.102810982  0.177555263
#> 904   0.675796567 -0.430551247  0.221680113
#> 905   0.589598619 -0.092072509  0.672610627
#> 906   0.565998245 -0.590595674  0.277119306
#> 907   0.507848878 -0.355178174  0.323898763
#> 908   0.552097416 -0.415673567 -0.148176102
#> 909   0.567565910 -0.438839169  0.443262041
#> 910   0.402903802 -0.308895368  0.079414316
#> 911   0.613561376 -0.487479761  0.580808671
#> 912   0.486387034 -0.715811910  0.076780911
#> 913   0.616818447 -0.164776247  0.251826418
#> 914   0.556781770 -0.589792861 -0.030183776
#> 915   0.327467520 -0.473941038  0.110974350
#> 916   0.726208714 -0.234145478  0.662839623
#> 917   0.521866863 -0.446301960  0.636558007
#> 918   0.468822054 -0.267783733  0.729192618
#> 919   0.367116702 -0.539983085  0.276614960
#> 920   0.392003951 -0.645258834  0.316601908
#> 921   0.822812871 -0.121361285  0.538382153
#> 922   0.783150396 -0.295291183  0.834707934
#> 923   0.740476664 -0.240240562  0.340960638
#> 924   0.233140458 -0.626381844  0.209017380
#> 925   0.471476238 -0.580054590  0.266056851
#> 926   0.738385095 -0.138641119 -0.149610857
#> 927   0.499855276 -0.484013087 -0.051464503
#> 928   0.546951563 -0.529394668 -0.289255890
#> 929   0.514948977 -0.357027182  0.263416702
#> 930   0.824699448  0.020275786  0.725201443
#> 931   0.836733143 -0.079676232  0.572541752
#> 932   0.686463552 -0.232947559  0.284655128
#> 933   0.859317396  0.004521422  0.772519333
#> 934   0.500355728 -0.572267898  0.728430771
#> 935   0.511576822 -0.421925205  0.479305741
#> 936   0.507848878 -0.355178174  0.323898763
#> 937   0.315193394 -0.646860972 -0.181449237
#> 938   0.700215520 -0.265836296  0.548589752
#> 939   0.365577527 -0.679465752  0.177177938
#> 940   0.639948998 -0.455886196  0.180194112
#> 941   0.406594214 -0.758677403  0.688148438
#> 942   0.334230867 -1.020516294 -0.370572330
#> 943   0.641675574 -0.389087053  0.136973743
#> 944   0.311823587 -0.352834469  0.353420944
#> 945   0.797736381  0.076459083  0.930059374
#> 946   0.227479565  0.052633747  0.887440762
#> 947   0.029212761 -0.380813147 -0.109417571
#> 948   0.546951563 -0.529394668 -0.289255890
#> 949   0.582287905 -0.476709698  0.292474390
#> 950   0.142523492 -0.404696508  0.086903939
#> 951   0.569785952 -0.621264822 -0.100719148
#> 952   0.350233404 -0.812797803 -0.357055318
#> 953   0.335255736 -0.640096142  0.630368503
#> 954   0.694959630 -0.528384725  0.265240817
#> 955   0.479808283 -0.567765962  0.430450818
#> 956   0.471476238 -0.580054590  0.266056851
#> 957   0.484566110 -0.405492822  0.977401340
#> 958   0.513049634 -0.250925344  0.706483663
#> 959  -0.046167308 -0.161052196  0.823403812
#> 960   0.456456259 -0.327331153  0.291013162
#> 961   0.708471597 -0.434017570  0.044575575
#> 962   0.710690610 -0.265988309  0.126670153
#> 963   0.639820441 -0.433482798  0.222440517
#> 964   0.605281145 -0.407340100  0.057707445
#> 965   0.316881820 -0.937514363 -0.126895461
#> 966   0.599714064 -0.320343662  0.642805664
#> 967   0.752456564 -0.122279383  0.507632494
#> 968   0.585003985 -0.322544573  0.415537414
#> 969   0.616384412 -0.610128554 -0.447558076
#> 970   0.834327375  0.003089501  1.147415026
#> 971   0.629594377 -0.445773583  0.019999732
#> 972   0.639948998 -0.455886196  0.180194112
#> 973   0.325794322 -0.944885676  0.245836337
#> 974   0.639948998 -0.455886196  0.180194112
#> 975   0.562257868 -0.429251312  0.512527023
#> 976   0.697667616 -0.319408497  0.172218174
#> 977   0.479253547 -0.631312672 -0.124256575
#> 978   0.508904208 -0.629117382  0.324563279
#> 979   0.365875100 -0.648339519 -0.017638770
#> 980   0.169879874 -0.434223862  0.359206906
#> 981   0.460325012 -0.636086494  0.170382173
#> 982   0.624863356 -0.521893908  0.061672540
#> 983   0.710690610 -0.265988309  0.126670153
#> 984   0.727730989  0.200437257  0.837362377
#> 985   0.292604032 -0.681914654  0.488478857
#> 986   0.605152262 -0.347594511 -0.167196298
#> 987   0.616384412 -0.610128554 -0.447558076
#> 988   0.280464679 -0.287256124  0.389440005
#> 989   0.647833972 -0.407775388  0.124895207
#> 990   0.420712811 -0.244753342  0.274763379
#> 991   0.365875100 -0.648339519 -0.017638770
#> 992   0.460325012 -0.636086494  0.170382173
#> 993   0.540026713 -0.195595743  0.110930565
#> 994   0.612535986 -0.498093881  0.219012511
#> 995   0.486708898 -0.100765705  0.563038118
#> 996   0.697667616 -0.319408497  0.172218174
#> 997   0.699435607 -0.399497951  0.281178641
#> 998   0.720645162 -0.213133721  0.350275947
#> 999   0.176391708 -0.797258980 -0.421709202
#> 1000  0.725438931 -0.502762561  0.102740349
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1 0.0534 -0.342  0.273 
#>  2      2        4       70        1 0.473  -0.741 -0.0652
#>  3      3        4       70        1 0.548  -0.253  0.316 
#>  4      4        4       70        1 0.428  -0.427 -0.208 
#>  5      5        4       70        1 0.507  -0.408  0.208 
#>  6      6        4       70        1 0.474  -0.634  0.391 
#>  7      7        4       70        1 0.677  -0.315 -0.0805
#>  8      8        4       70        1 0.471  -0.398  0.183 
#>  9      9        4       70        1 0.476  -0.430  0.388 
#> 10     10        4       70        1 0.661  -0.680  0.0574
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  4.22  49.7 1.31  0.0848 24.0     7.60e+2 1194.
#> 2      1    14     4    70     1  4.22  49.7 1.31  0.0848 13.4     2.92e-5  666.
#> 3      2     1     4    70     1  6.42  33.4 0.937 0.192  27.6     1.00e+3  922.
#> 4      2    14     4    70     1  6.42  33.4 0.937 0.192   5.36    5.13e-3  179.
#> 5      3     1     4    70     1  6.92  54.4 1.37  0.127  22.1     7.29e+2 1201.
#> 6      3    14     4    70     1  6.92  54.4 1.37  0.127   7.05    1.32e-5  383.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#>