Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object
A posologyr list, created by the posologyr function.

n_sample
Number of samples from the S-step

n_resample
Number of samples from the R-step

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)
#> 
#> $eta
#>            ETA_ke         ETA_V
#> 1    -0.136416847  2.806144e-02
#> 2    -0.055343770  2.302994e-01
#> 3    -0.159941804 -6.862780e-02
#> 4    -0.090974984  8.845723e-02
#> 5    -0.134724977  1.036587e-01
#> 6     0.036042464 -4.495473e-02
#> 7    -0.160584480 -2.291309e-02
#> 8    -0.365002978  1.881017e-01
#> 9    -0.097952048 -3.527794e-02
#> 10   -0.185007458  3.112959e-02
#> 11   -0.204391718  1.179522e-01
#> 12   -0.213111506  2.109108e-01
#> 13   -0.006404233  1.340769e-01
#> 14   -0.155145452  4.488405e-02
#> 15   -0.136965725  4.994071e-02
#> 16   -0.117588976  1.478690e-01
#> 17   -0.050600220  1.112757e-01
#> 18   -0.153177018 -7.523170e-02
#> 19   -0.084888649  4.656541e-02
#> 20   -0.140614950  2.929463e-02
#> 21   -0.182345932  8.241573e-02
#> 22   -0.117265002 -4.661459e-02
#> 23   -0.099185878 -1.249978e-01
#> 24   -0.028905699 -5.155363e-02
#> 25   -0.192242096  6.921441e-02
#> 26   -0.074078815  1.166657e-01
#> 27   -0.118765352  1.036479e-01
#> 28   -0.126495988  1.150475e-01
#> 29   -0.059329326  4.161865e-03
#> 30   -0.102242591  2.445067e-01
#> 31   -0.180148340  1.250629e-01
#> 32   -0.151890875  9.811596e-02
#> 33   -0.045572663  8.530681e-02
#> 34   -0.088335258  1.248247e-02
#> 35   -0.108367651  2.828584e-02
#> 36   -0.176255336  2.503983e-01
#> 37    0.016101785  9.132402e-02
#> 38   -0.261709634  1.516500e-01
#> 39   -0.091777826  7.393123e-02
#> 40   -0.098497372  1.125699e-01
#> 41   -0.138583337 -1.267248e-02
#> 42   -0.150910005  1.388200e-01
#> 43   -0.138082830 -1.249218e-01
#> 44   -0.107291683  1.569305e-01
#> 45   -0.206296218  1.316650e-01
#> 46   -0.308024413 -4.919064e-02
#> 47    0.043906649 -2.783202e-02
#> 48   -0.165093382  1.731383e-02
#> 49   -0.092434093 -3.735786e-02
#> 50   -0.063042918 -6.191913e-02
#> 51   -0.159526412  1.137679e-01
#> 52   -0.021023364 -9.117228e-02
#> 53   -0.005993242  2.872994e-02
#> 54   -0.197065014  3.518804e-01
#> 55   -0.080126701  1.222433e-01
#> 56   -0.152926874  1.610750e-01
#> 57   -0.110308765 -8.332284e-02
#> 58   -0.102306382  1.581471e-02
#> 59   -0.368729162  2.704967e-01
#> 60   -0.090665393  1.796478e-01
#> 61   -0.086525368  7.061947e-02
#> 62   -0.060301678  2.195316e-02
#> 63   -0.144506018  7.105209e-02
#> 64   -0.179571796 -5.996003e-02
#> 65   -0.106069867  2.885839e-01
#> 66   -0.161955743  1.568081e-01
#> 67   -0.074584002 -3.303179e-02
#> 68    0.056301142 -1.423463e-01
#> 69   -0.122448239  8.790013e-02
#> 70   -0.115146629  1.496136e-01
#> 71   -0.028064963  1.221437e-01
#> 72   -0.151890875  9.811596e-02
#> 73   -0.023879784  1.074415e-01
#> 74   -0.135274866  6.977983e-02
#> 75   -0.161288426 -1.670018e-01
#> 76   -0.081780805  7.415847e-02
#> 77   -0.117273523  4.419776e-02
#> 78   -0.056767066 -8.098243e-02
#> 79   -0.198238651  1.382508e-01
#> 80   -0.102882413  2.445717e-02
#> 81   -0.080821462 -1.017045e-01
#> 82   -0.131130718  1.315268e-01
#> 83   -0.113505485  1.334465e-01
#> 84    0.049496696 -8.541769e-02
#> 85   -0.021312794 -3.716493e-02
#> 86   -0.153977717  1.210764e-01
#> 87   -0.122740674  2.280852e-01
#> 88   -0.050950317 -4.719334e-03
#> 89   -0.124550238  5.454116e-02
#> 90   -0.153506351  7.665740e-02
#> 91   -0.180103611  2.378392e-01
#> 92   -0.081086645 -8.775028e-03
#> 93   -0.290931573  2.505404e-01
#> 94   -0.110585821  1.056314e-01
#> 95   -0.156871876 -7.796018e-02
#> 96   -0.214783097  3.482514e-01
#> 97    0.000917054  5.244553e-02
#> 98   -0.138834185  9.605948e-02
#> 99   -0.158525643 -1.280337e-01
#> 100   0.038473406 -1.610755e-01
#> 101  -0.109863760 -1.170131e-02
#> 102  -0.261612992  1.589764e-01
#> 103  -0.136501236  1.284437e-01
#> 104  -0.165174164 -3.892578e-02
#> 105  -0.002266775 -2.037522e-01
#> 106  -0.135599782  7.278769e-02
#> 107  -0.102346167  2.949320e-02
#> 108  -0.061673150 -9.038063e-02
#> 109  -0.072032980 -8.153586e-02
#> 110  -0.043064631  1.238556e-01
#> 111  -0.130291065  7.411598e-02
#> 112  -0.180103611  2.378392e-01
#> 113  -0.230029106  1.202738e-01
#> 114  -0.150715860  1.148552e-01
#> 115  -0.008548013  5.950966e-02
#> 116  -0.061723263  1.446975e-01
#> 117  -0.136501236  1.284437e-01
#> 118  -0.142003216 -1.800580e-02
#> 119  -0.005333200 -6.605126e-02
#> 120  -0.059839248  9.063875e-02
#> 121  -0.106630699  1.333769e-01
#> 122  -0.046970227  1.695076e-01
#> 123  -0.105867839  1.082545e-01
#> 124  -0.145733642 -2.063561e-02
#> 125  -0.168718358  1.134621e-01
#> 126  -0.156636505  1.717248e-01
#> 127  -0.055762810  7.026853e-02
#> 128  -0.138888509  1.629163e-02
#> 129   0.045081910 -2.616608e-02
#> 130  -0.138249203  9.088687e-02
#> 131  -0.226625470  1.213558e-01
#> 132  -0.103214055 -7.828869e-02
#> 133  -0.118443477  3.987262e-04
#> 134  -0.212837016  7.011470e-02
#> 135  -0.289078748  3.015543e-01
#> 136  -0.156871876 -7.796018e-02
#> 137  -0.223897927  2.063675e-01
#> 138  -0.033235486  8.726331e-02
#> 139  -0.173771079  6.887001e-02
#> 140  -0.168775659  1.112184e-01
#> 141   0.050940866 -5.256540e-02
#> 142  -0.127506603  1.744108e-01
#> 143  -0.284339454  6.208554e-02
#> 144  -0.127142657  1.231932e-01
#> 145  -0.097580655  4.852428e-02
#> 146  -0.061843071  7.153678e-02
#> 147  -0.205054038  1.821941e-01
#> 148  -0.155991122  2.430178e-01
#> 149  -0.189120391  1.223467e-01
#> 150  -0.203211913  1.304329e-01
#> 151  -0.030351155  7.791249e-02
#> 152  -0.012774496  1.126636e-01
#> 153  -0.317530547  2.022291e-01
#> 154  -0.136348612  1.578510e-02
#> 155   0.014283360  2.841816e-02
#> 156  -0.047940413  1.632476e-01
#> 157  -0.067011841  1.305297e-01
#> 158  -0.169833056  1.155643e-01
#> 159  -0.040210672  2.953802e-01
#> 160  -0.119538528  2.193947e-02
#> 161  -0.123324228 -1.525769e-02
#> 162  -0.049603864  1.748092e-02
#> 163  -0.061871180 -6.143823e-02
#> 164  -0.179434240  1.481341e-01
#> 165  -0.123992541  2.047964e-01
#> 166  -0.112157592  4.317242e-02
#> 167  -0.038980077  4.681823e-02
#> 168  -0.061116120 -2.634106e-02
#> 169  -0.261612992  1.589764e-01
#> 170  -0.058660483  8.696098e-02
#> 171  -0.235087662  2.078541e-01
#> 172   0.041398309  7.651104e-02
#> 173  -0.205976709  1.620773e-01
#> 174   0.010128762 -4.788945e-02
#> 175  -0.140568264  5.068587e-02
#> 176  -0.183686498  1.318146e-02
#> 177   0.014306040 -1.100169e-01
#> 178   0.010598127  6.816721e-02
#> 179   0.069405663 -7.864319e-02
#> 180  -0.290817288  2.772469e-01
#> 181  -0.204263907  2.791464e-01
#> 182  -0.114722580  1.766795e-01
#> 183  -0.060805319  4.210092e-02
#> 184  -0.170978603  2.041582e-02
#> 185  -0.053910394  7.449422e-02
#> 186  -0.150406058  1.823521e-01
#> 187  -0.042821612 -3.814443e-02
#> 188  -0.112384161  2.455104e-02
#> 189  -0.060320194 -2.511922e-01
#> 190  -0.102306382  1.581471e-02
#> 191  -0.091790112  1.131074e-01
#> 192   0.014792173 -1.298808e-01
#> 193   0.030656472  8.524173e-02
#> 194  -0.065891923  6.594403e-02
#> 195  -0.211769512 -3.702378e-02
#> 196  -0.174160343  2.667920e-02
#> 197  -0.034308762  1.567431e-01
#> 198  -0.246012617  1.816715e-01
#> 199  -0.117561491  1.285658e-01
#> 200  -0.123301167  7.619196e-02
#> 201  -0.197586715  9.569096e-02
#> 202  -0.189921706  9.681597e-02
#> 203  -0.127142657  1.231932e-01
#> 204  -0.279840911  8.691202e-02
#> 205  -0.293200317  1.656159e-01
#> 206  -0.161383569 -6.098139e-02
#> 207  -0.213111506  2.109108e-01
#> 208  -0.035882751 -7.589601e-02
#> 209  -0.006531553 -4.406341e-02
#> 210  -0.188694425  1.647364e-01
#> 211  -0.152454640 -1.401280e-02
#> 212   0.021381827  2.195254e-02
#> 213  -0.169115344  1.711143e-01
#> 214  -0.161753191  1.085508e-01
#> 215  -0.029986180  4.413480e-02
#> 216  -0.130291065  7.411598e-02
#> 217  -0.144250876  1.038940e-01
#> 218  -0.061723263  1.446975e-01
#> 219  -0.135165757  1.421171e-01
#> 220  -0.122053602  2.397408e-01
#> 221  -0.237009438  7.274575e-02
#> 222  -0.035846621 -8.363197e-02
#> 223   0.078022453 -2.231149e-01
#> 224  -0.167414375  1.584456e-01
#> 225  -0.115080560  1.616763e-01
#> 226  -0.182666608  2.797295e-01
#> 227  -0.155997668  9.809752e-02
#> 228  -0.144216430  9.754534e-02
#> 229  -0.069759156 -9.359004e-02
#> 230  -0.147114580  1.546334e-01
#> 231  -0.220493831  1.305238e-01
#> 232  -0.197065014  3.518804e-01
#> 233   0.040857152  4.966763e-03
#> 234  -0.061843071  7.153678e-02
#> 235  -0.110622668  1.317009e-01
#> 236  -0.229773826  1.571854e-01
#> 237  -0.091513647 -9.721192e-03
#> 238  -0.098497372  1.125699e-01
#> 239  -0.132227744 -1.013013e-02
#> 240  -0.124869441  1.339104e-01
#> 241  -0.182218481  7.063451e-02
#> 242  -0.085498958 -7.658516e-02
#> 243  -0.139579871  4.121141e-02
#> 244  -0.195397184  2.368954e-01
#> 245  -0.064415445  1.062420e-01
#> 246  -0.062162540  9.611799e-02
#> 247   0.033537300 -1.473920e-01
#> 248  -0.221201077  2.737923e-01
#> 249  -0.243489251  6.772102e-02
#> 250  -0.074288445  1.152670e-01
#> 251  -0.036417917  4.113774e-02
#> 252  -0.312209040  1.829299e-01
#> 253  -0.336687448  3.347039e-01
#> 254  -0.131943252  2.244560e-01
#> 255  -0.050807613  8.743525e-02
#> 256  -0.098872570  3.041481e-01
#> 257  -0.259250993 -3.784563e-02
#> 258  -0.047903344 -6.910291e-02
#> 259  -0.144034586  3.936408e-01
#> 260  -0.052314722 -2.845926e-02
#> 261  -0.117775265  1.640641e-01
#> 262  -0.062202615  8.640920e-02
#> 263   0.006225375  1.334680e-01
#> 264  -0.056071198  1.757982e-03
#> 265  -0.137546389  6.089948e-02
#> 266  -0.064900383 -5.696109e-02
#> 267  -0.182666608  2.797295e-01
#> 268   0.060278782 -3.869119e-02
#> 269  -0.016842604 -2.413526e-01
#> 270  -0.112310138  6.135399e-02
#> 271   0.014439204  1.174382e-01
#> 272  -0.098283299 -4.597322e-02
#> 273  -0.297033291  1.636225e-01
#> 274  -0.027306394  1.606794e-02
#> 275  -0.246679615 -5.215043e-06
#> 276  -0.194669348  2.120359e-02
#> 277  -0.422804749  7.715882e-02
#> 278  -0.290931573  2.505404e-01
#> 279  -0.100438145  6.643765e-02
#> 280  -0.082250267  1.562315e-01
#> 281  -0.102381316 -1.117109e-01
#> 282  -0.304080400  1.318268e-01
#> 283  -0.041611487  4.576503e-02
#> 284   0.061816991 -2.148244e-01
#> 285  -0.074761222 -8.774765e-02
#> 286  -0.288478167  2.073537e-01
#> 287  -0.147228731  2.033709e-01
#> 288  -0.132138941  3.081622e-02
#> 289  -0.123132762  1.399448e-01
#> 290  -0.076478260 -1.271790e-01
#> 291  -0.160206393  1.471459e-01
#> 292  -0.132481238  7.700629e-04
#> 293  -0.143999397 -5.429970e-02
#> 294   0.026383835 -1.071742e-01
#> 295  -0.057217248  2.164069e-01
#> 296  -0.136501236  1.284437e-01
#> 297  -0.165870153  5.144530e-02
#> 298  -0.153506351  7.665740e-02
#> 299  -0.107755558  4.376983e-02
#> 300   0.078603866 -4.085753e-02
#> 301  -0.093108093  2.114322e-01
#> 302   0.029514178 -2.627587e-02
#> 303   0.021624944  9.364562e-03
#> 304   0.066977470 -1.585628e-01
#> 305  -0.102056965  4.700187e-02
#> 306  -0.114022594  5.445072e-02
#> 307   0.119614556 -2.080953e-01
#> 308  -0.204537617 -7.986605e-04
#> 309  -0.340788496  2.895757e-01
#> 310  -0.117588976  1.478690e-01
#> 311  -0.146686985 -8.825295e-02
#> 312  -0.027385593  8.904485e-02
#> 313  -0.084592039 -1.276061e-01
#> 314   0.106876737 -2.426291e-01
#> 315   0.040564893 -8.102635e-02
#> 316   0.030477174 -6.203982e-02
#> 317  -0.221201077  2.737923e-01
#> 318  -0.100180404 -8.732202e-02
#> 319  -0.109074396  2.593979e-01
#> 320  -0.042840038  4.399009e-02
#> 321  -0.361804341  1.461529e-01
#> 322  -0.220493831  1.305238e-01
#> 323  -0.112132348  6.351389e-02
#> 324  -0.030308390 -3.327168e-03
#> 325  -0.135723110  4.156042e-02
#> 326  -0.038783722  1.980792e-02
#> 327  -0.116220309 -9.566936e-02
#> 328  -0.114969839  5.210812e-02
#> 329  -0.106926272 -4.063467e-02
#> 330  -0.047071749 -1.095944e-01
#> 331  -0.356398105  6.710908e-02
#> 332  -0.162320176  1.215592e-01
#> 333  -0.127195163  1.649140e-01
#> 334  -0.147422227  8.030333e-02
#> 335  -0.166206888  1.509489e-01
#> 336  -0.027759099  6.268725e-03
#> 337  -0.251384505  1.973516e-01
#> 338  -0.094338950 -5.703132e-02
#> 339  -0.295098455  1.698762e-01
#> 340  -0.221201077  2.737923e-01
#> 341  -0.081780805  7.415847e-02
#> 342  -0.028231186  9.927214e-02
#> 343   0.014752549 -8.768163e-02
#> 344  -0.117998191  1.447594e-01
#> 345  -0.144034586  3.936408e-01
#> 346  -0.080552026  2.329876e-02
#> 347  -0.197065014  3.518804e-01
#> 348  -0.114022594  5.445072e-02
#> 349  -0.175163947  2.071270e-01
#> 350   0.029872285 -2.416383e-02
#> 351  -0.251384505  1.973516e-01
#> 352  -0.122767537 -2.176306e-02
#> 353  -0.084077182  1.669448e-01
#> 354  -0.166704801  3.402290e-01
#> 355   0.027479316 -1.698963e-01
#> 356  -0.179684873  5.910760e-02
#> 357  -0.130691907  1.253269e-01
#> 358  -0.216097270  1.295356e-01
#> 359  -0.105333328  1.077902e-01
#> 360  -0.156636505  1.717248e-01
#> 361  -0.108223687  1.302002e-02
#> 362  -0.144034586  3.936408e-01
#> 363  -0.211451943  6.042183e-02
#> 364  -0.111765977  1.094153e-01
#> 365  -0.047071691 -8.093258e-02
#> 366  -0.037439844  1.715992e-01
#> 367  -0.005713389  3.904352e-02
#> 368  -0.282617507  1.243188e-01
#> 369  -0.107823730  4.664474e-02
#> 370  -0.220134217  3.515691e-02
#> 371  -0.039448863 -2.749602e-02
#> 372  -0.003556180  5.171987e-02
#> 373  -0.272642431  2.358195e-01
#> 374  -0.076454424  4.836104e-02
#> 375  -0.055277086 -1.066765e-01
#> 376  -0.159509428  1.453735e-01
#> 377  -0.089562135 -1.116998e-01
#> 378  -0.112310138  6.135399e-02
#> 379  -0.147228731  2.033709e-01
#> 380  -0.102487984  3.836998e-02
#> 381   0.006719976 -2.162031e-01
#> 382  -0.262800033  1.727784e-01
#> 383  -0.123132762  1.399448e-01
#> 384   0.111649606  8.371069e-02
#> 385   0.031892147 -6.468376e-02
#> 386  -0.164112792  1.554407e-01
#> 387  -0.046970227  1.695076e-01
#> 388   0.049185339 -1.857507e-01
#> 389  -0.277834216  1.448218e-01
#> 390  -0.125286444  6.492052e-02
#> 391  -0.095998551  1.058303e-01
#> 392  -0.064104612 -1.325531e-02
#> 393  -0.077324385  1.521081e-01
#> 394  -0.084410034  1.335448e-01
#> 395  -0.168139534  8.112958e-02
#> 396  -0.129037536 -1.092376e-02
#> 397  -0.057217248  2.164069e-01
#> 398  -0.107140617 -4.404824e-03
#> 399  -0.095319532  1.813589e-01
#> 400  -0.177007673  1.699076e-01
#> 401  -0.143622108 -9.640780e-02
#> 402  -0.158801518  1.738117e-01
#> 403  -0.145450551  6.974801e-02
#> 404  -0.140174280  2.028968e-01
#> 405  -0.073827195  1.905955e-01
#> 406  -0.164112792  1.554407e-01
#> 407  -0.030308390 -3.327168e-03
#> 408  -0.139579871  4.121141e-02
#> 409  -0.050919880  6.463660e-02
#> 410  -0.047656826 -2.762939e-02
#> 411  -0.102234784  1.047808e-01
#> 412  -0.092822271  6.245633e-02
#> 413  -0.073755615 -1.641215e-01
#> 414  -0.114584830  1.531802e-01
#> 415  -0.190583473  1.751609e-01
#> 416  -0.095304385  6.441219e-02
#> 417  -0.230029106  1.202738e-01
#> 418  -0.167526039 -3.542755e-02
#> 419  -0.036093220  1.144756e-01
#> 420   0.025996119  3.873391e-02
#> 421  -0.205724452  7.850203e-02
#> 422  -0.076136124 -3.415757e-02
#> 423  -0.237221675  1.040900e-02
#> 424  -0.131536835  1.891371e-03
#> 425  -0.093557388  1.197106e-01
#> 426  -0.043109300  5.563053e-02
#> 427  -0.033034730  1.374037e-01
#> 428  -0.205054038  1.821941e-01
#> 429   0.005869815 -1.269427e-02
#> 430  -0.070893800 -1.546119e-01
#> 431  -0.014695767 -1.726973e-01
#> 432  -0.214046477  9.281694e-02
#> 433  -0.131943252  2.244560e-01
#> 434  -0.007921778 -1.150472e-02
#> 435  -0.205856428  2.434417e-01
#> 436  -0.137951923  1.481095e-01
#> 437  -0.210329335  3.219829e-02
#> 438  -0.016679927  4.694709e-02
#> 439  -0.156636505  1.717248e-01
#> 440  -0.155377114  2.007688e-01
#> 441  -0.106630699  1.333769e-01
#> 442  -0.233346720  1.796793e-01
#> 443  -0.102306382  1.581471e-02
#> 444  -0.221201077  2.737923e-01
#> 445  -0.057633941  3.747328e-02
#> 446  -0.086998758  1.371217e-02
#> 447  -0.121591904  1.032490e-01
#> 448  -0.117122399  5.520258e-02
#> 449   0.006306715 -1.085347e-01
#> 450  -0.155377114  2.007688e-01
#> 451  -0.330498479  8.117358e-02
#> 452  -0.051591939  7.243490e-02
#> 453  -0.124013829  1.191296e-01
#> 454  -0.030144428  3.841901e-02
#> 455  -0.132787932  1.031050e-01
#> 456  -0.122448239  8.790013e-02
#> 457  -0.079317901  1.271973e-02
#> 458  -0.293544040  1.074450e-02
#> 459   0.075649982 -6.579464e-02
#> 460  -0.134211772 -6.438197e-02
#> 461  -0.083285738  6.868187e-02
#> 462  -0.155991122  2.430178e-01
#> 463  -0.004181066  1.873957e-02
#> 464  -0.213266725  3.246294e-01
#> 465  -0.032544191  8.067710e-02
#> 466  -0.233346720  1.796793e-01
#> 467  -0.115440645 -1.155746e-02
#> 468  -0.197065014  3.518804e-01
#> 469  -0.125407785  2.140501e-01
#> 470  -0.003834067  8.188364e-02
#> 471  -0.159509428  1.453735e-01
#> 472  -0.065032799 -1.163325e-01
#> 473  -0.086525368  7.061947e-02
#> 474  -0.420008956 -1.540248e-02
#> 475  -0.174992378  2.483018e-01
#> 476  -0.173405223  1.689669e-01
#> 477  -0.253267878  1.404586e-01
#> 478   0.044227008 -3.986527e-02
#> 479  -0.102277728 -9.385693e-03
#> 480  -0.081780805  7.415847e-02
#> 481  -0.173749862 -5.275001e-02
#> 482  -0.217998995 -5.549854e-02
#> 483  -0.206296218  1.316650e-01
#> 484  -0.155917032  1.288019e-01
#> 485  -0.023264830  6.289029e-02
#> 486  -0.136247326 -6.281802e-02
#> 487  -0.036581044  9.189467e-03
#> 488  -0.122643655 -1.641155e-02
#> 489  -0.107304815  1.391746e-01
#> 490  -0.188245696 -2.216917e-02
#> 491  -0.356168104  1.470459e-01
#> 492  -0.119758084  7.909860e-02
#> 493   0.040857152  4.966763e-03
#> 494  -0.265403492  2.043736e-01
#> 495  -0.124869441  1.339104e-01
#> 496  -0.129037536 -1.092376e-02
#> 497  -0.093437267  2.458677e-02
#> 498  -0.247813374  1.614563e-01
#> 499  -0.143278031  3.222596e-01
#> 500  -0.285839078  1.318331e-01
#> 501  -0.215001436  6.272245e-02
#> 502  -0.026992538 -1.391960e-01
#> 503  -0.059909861 -6.369996e-02
#> 504  -0.165870153  5.144530e-02
#> 505  -0.228888546  1.630602e-01
#> 506  -0.039749419  6.275735e-02
#> 507  -0.144653116  9.401629e-02
#> 508  -0.254224698  1.706033e-01
#> 509  -0.082428431  6.921090e-02
#> 510  -0.163949298  1.284391e-01
#> 511  -0.176809308  4.647908e-02
#> 512   0.078380835 -2.095793e-01
#> 513  -0.013461059  7.005451e-02
#> 514  -0.020365621 -1.597141e-03
#> 515  -0.230968424  9.386545e-02
#> 516  -0.161077977  1.327592e-01
#> 517  -0.073848505 -8.777449e-03
#> 518  -0.158311848  1.023410e-01
#> 519  -0.061299964  1.122191e-01
#> 520  -0.084405574 -2.287387e-02
#> 521   0.035788321 -1.670471e-01
#> 522  -0.091777826  7.393123e-02
#> 523  -0.205054038  1.821941e-01
#> 524  -0.221753620  3.890289e-02
#> 525  -0.169833056  1.155643e-01
#> 526  -0.070181923 -7.021806e-02
#> 527  -0.087477304  4.623845e-02
#> 528  -0.040454614 -1.136771e-02
#> 529   0.020821813  9.461316e-02
#> 530  -0.039256584 -1.643288e-02
#> 531  -0.196424939  2.751042e-01
#> 532   0.017332845 -4.576630e-02
#> 533   0.049496696 -8.541769e-02
#> 534  -0.155377114  2.007688e-01
#> 535  -0.207247988  3.125315e-02
#> 536  -0.107304815  1.391746e-01
#> 537  -0.216696786  8.673635e-02
#> 538  -0.256429159 -2.921317e-03
#> 539  -0.323909610  6.579864e-02
#> 540   0.007203304 -7.192786e-02
#> 541  -0.135688213  7.386761e-02
#> 542   0.056043355 -1.998742e-02
#> 543  -0.175721082 -8.821014e-02
#> 544  -0.040116190 -1.364892e-02
#> 545  -0.142144342 -3.501080e-02
#> 546  -0.148222401 -3.417780e-02
#> 547   0.020250452 -9.323304e-02
#> 548  -0.321280502  2.117350e-01
#> 549  -0.093720365  6.074378e-02
#> 550  -0.065425380 -8.758753e-03
#> 551  -0.075032164  1.363747e-01
#> 552  -0.036093220  1.144756e-01
#> 553  -0.051175303 -4.406206e-02
#> 554   0.010128762 -4.788945e-02
#> 555  -0.336687448  3.347039e-01
#> 556  -0.085390447 -1.681838e-01
#> 557  -0.245584578  1.136250e-01
#> 558  -0.167414375  1.584456e-01
#> 559  -0.063893598  8.036114e-02
#> 560  -0.091450541 -1.789928e-02
#> 561  -0.139515968 -1.205650e-03
#> 562  -0.122493956  6.044340e-02
#> 563  -0.003901286 -1.095569e-02
#> 564  -0.123992541  2.047964e-01
#> 565  -0.036817667 -2.392401e-02
#> 566  -0.122607785 -1.090164e-01
#> 567  -0.169700515  2.468213e-01
#> 568  -0.409647008  1.849880e-01
#> 569   0.001067459 -2.119892e-02
#> 570  -0.241055982 -8.804520e-02
#> 571  -0.095319532  1.813589e-01
#> 572  -0.169833056  1.155643e-01
#> 573  -0.105204221  7.895377e-02
#> 574  -0.079085017 -1.160912e-02
#> 575  -0.194475599  6.131939e-03
#> 576   0.048491488 -6.288495e-02
#> 577  -0.058951078  7.628452e-03
#> 578  -0.111681507  2.203772e-01
#> 579  -0.101088691  1.845586e-01
#> 580  -0.365002978  1.881017e-01
#> 581  -0.108083988  4.467514e-02
#> 582  -0.038679871 -2.492786e-02
#> 583   0.035299765  4.172932e-02
#> 584  -0.272642431  2.358195e-01
#> 585  -0.122155667 -5.008441e-02
#> 586  -0.057633941  3.747328e-02
#> 587  -0.186024331  7.563250e-02
#> 588  -0.223352606  7.471005e-03
#> 589  -0.055971493  3.050389e-03
#> 590  -0.188876933  2.405552e-01
#> 591  -0.115885125  4.879264e-02
#> 592   0.010270724  6.493057e-02
#> 593  -0.112132348  6.351389e-02
#> 594  -0.054234634  4.146973e-02
#> 595  -0.118351146  2.168795e-01
#> 596  -0.003551459 -1.148866e-01
#> 597  -0.333772617  1.692127e-01
#> 598  -0.155377114  2.007688e-01
#> 599  -0.178014347  1.870972e-02
#> 600  -0.056071198  1.757982e-03
#> 601  -0.050782225 -4.479724e-02
#> 602  -0.117122399  5.520258e-02
#> 603  -0.214783097  3.482514e-01
#> 604  -0.072177800  4.123706e-02
#> 605  -0.144693472  7.048575e-02
#> 606  -0.243969281  1.485415e-01
#> 607  -0.007739412 -1.293093e-01
#> 608  -0.108829940  1.236684e-01
#> 609  -0.249538488  1.081256e-01
#> 610  -0.217964178  1.619173e-01
#> 611  -0.147228731  2.033709e-01
#> 612   0.055685998  9.666434e-02
#> 613  -0.065771261  1.478543e-01
#> 614   0.016104384 -2.236293e-02
#> 615  -0.223175328 -1.647636e-02
#> 616  -0.047300090  4.749822e-02
#> 617  -0.319714868  3.206349e-01
#> 618  -0.090051273  1.042915e-01
#> 619  -0.028705614 -1.014883e-01
#> 620  -0.029158180  2.353459e-02
#> 621  -0.083285738  6.868187e-02
#> 622  -0.084927317  1.038068e-01
#> 623  -0.201297537  8.885540e-02
#> 624  -0.188465297  1.385958e-01
#> 625  -0.180456903  1.657058e-01
#> 626  -0.045295696  6.033115e-02
#> 627  -0.063800076 -5.172157e-02
#> 628  -0.136389928  2.665742e-02
#> 629  -0.270051589  2.873600e-01
#> 630  -0.084410034  1.335448e-01
#> 631  -0.123992541  2.047964e-01
#> 632   0.041444683 -4.604684e-02
#> 633  -0.197595673  1.244083e-01
#> 634  -0.273450765  2.847353e-01
#> 635  -0.205914864  5.897160e-02
#> 636  -0.136503858  5.474144e-02
#> 637  -0.074470343  2.083916e-01
#> 638  -0.175640374  5.432805e-02
#> 639  -0.005289124  2.879836e-02
#> 640  -0.214825027  6.135841e-02
#> 641  -0.065016076 -6.194926e-02
#> 642  -0.155917032  1.288019e-01
#> 643  -0.154281997 -1.525912e-02
#> 644  -0.106224324 -4.549332e-02
#> 645  -0.084652474  9.270609e-03
#> 646  -0.059329326  4.161865e-03
#> 647  -0.232556153  1.128709e-01
#> 648  -0.146968069  1.399418e-01
#> 649  -0.095865805  1.917129e-01
#> 650   0.079958542 -1.411339e-01
#> 651  -0.146524208  3.939852e-02
#> 652  -0.057870459  2.801381e-03
#> 653  -0.064847615  1.168600e-01
#> 654  -0.130810753 -7.924440e-02
#> 655  -0.145555366  1.592410e-01
#> 656  -0.042183427 -3.710129e-02
#> 657  -0.048518582 -3.869305e-02
#> 658  -0.235529362  7.301504e-02
#> 659  -0.136144758  1.675535e-01
#> 660  -0.055515167  3.926439e-02
#> 661  -0.256551754  1.715821e-01
#> 662  -0.019300600 -9.251513e-06
#> 663  -0.179358051  1.374879e-01
#> 664  -0.169833056  1.155643e-01
#> 665   0.006307451 -1.028017e-01
#> 666  -0.084840851  2.741721e-01
#> 667  -0.032151438  1.209883e-01
#> 668  -0.103424861  9.041553e-02
#> 669   0.029408696 -2.728242e-02
#> 670   0.083160329 -4.542517e-02
#> 671   0.009554846  2.223102e-02
#> 672   0.001943527  1.115521e-01
#> 673  -0.195397184  2.368954e-01
#> 674  -0.169833056  1.155643e-01
#> 675  -0.023286326  2.819598e-02
#> 676  -0.114969839  5.210812e-02
#> 677  -0.093534658  1.492190e-01
#> 678  -0.178188060  6.525248e-03
#> 679   0.019697286 -1.049489e-01
#> 680  -0.118130028  1.446007e-01
#> 681  -0.091450541 -1.789928e-02
#> 682  -0.087885770 -2.049277e-02
#> 683  -0.137637125  1.712057e-02
#> 684  -0.123992541  2.047964e-01
#> 685  -0.115643347  1.492963e-01
#> 686  -0.063682458 -6.774979e-02
#> 687  -0.106069867  2.885839e-01
#> 688  -0.248950820  4.961805e-02
#> 689   0.043000083 -2.693473e-02
#> 690  -0.144688136  5.676602e-02
#> 691  -0.159623623  1.157548e-01
#> 692  -0.007775570  2.621408e-02
#> 693  -0.217964178  1.619173e-01
#> 694  -0.046970227  1.695076e-01
#> 695  -0.159872842  7.571233e-02
#> 696  -0.054495685  2.322819e-02
#> 697  -0.080126701  1.222433e-01
#> 698  -0.224813108  1.426602e-02
#> 699  -0.064808635 -5.953104e-02
#> 700  -0.195397184  2.368954e-01
#> 701  -0.085149146 -7.538983e-02
#> 702  -0.138441705  9.309163e-02
#> 703  -0.189921706  9.681597e-02
#> 704  -0.089122058  1.846067e-02
#> 705  -0.297832538  2.419871e-01
#> 706  -0.169115344  1.711143e-01
#> 707  -0.333866317  1.295530e-01
#> 708  -0.040480659  1.321667e-01
#> 709  -0.051591939  7.243490e-02
#> 710  -0.107495318 -1.294244e-01
#> 711  -0.174996408 -8.692032e-02
#> 712  -0.046938533 -8.278202e-02
#> 713  -0.139583052 -4.087408e-02
#> 714  -0.016679927  4.694709e-02
#> 715  -0.131943252  2.244560e-01
#> 716   0.021140594 -4.958318e-02
#> 717  -0.088413004 -1.938493e-02
#> 718  -0.042022290 -1.813702e-02
#> 719  -0.336687448  3.347039e-01
#> 720  -0.108934297  1.035793e-01
#> 721  -0.003551459 -1.148866e-01
#> 722  -0.221201077  2.737923e-01
#> 723  -0.165249761 -8.003525e-02
#> 724  -0.201175689  7.690528e-02
#> 725  -0.105972275  1.835861e-01
#> 726  -0.098283299 -4.597322e-02
#> 727  -0.144034586  3.936408e-01
#> 728  -0.067726381 -2.854537e-02
#> 729  -0.028746165 -6.226204e-02
#> 730  -0.138888509  1.629163e-02
#> 731  -0.090594734  9.222188e-02
#> 732  -0.197586715  9.569096e-02
#> 733  -0.084927317  1.038068e-01
#> 734   0.030998355 -4.143826e-02
#> 735  -0.143605314 -5.542679e-02
#> 736  -0.081246273 -1.317291e-01
#> 737  -0.062877583 -8.867021e-02
#> 738  -0.214783097  3.482514e-01
#> 739  -0.139579871  4.121141e-02
#> 740  -0.286077887  1.287143e-01
#> 741  -0.046809262  5.631696e-02
#> 742  -0.233266068  1.958109e-01
#> 743  -0.155991122  2.430178e-01
#> 744  -0.228888546  1.630602e-01
#> 745  -0.064735336  2.958734e-02
#> 746  -0.139974630  6.903562e-02
#> 747  -0.140739896  9.829566e-02
#> 748  -0.066347640  1.390034e-01
#> 749  -0.118130028  1.446007e-01
#> 750  -0.176255336  2.503983e-01
#> 751  -0.047431790  1.005151e-01
#> 752  -0.152518227  2.125040e-01
#> 753  -0.110263624  7.293871e-03
#> 754  -0.093935579  4.896881e-03
#> 755  -0.060947416  5.696546e-02
#> 756  -0.160288560  9.049468e-02
#> 757  -0.129329024  5.681902e-02
#> 758  -0.150250859  1.117759e-01
#> 759  -0.118362512  6.967142e-02
#> 760  -0.188898182 -3.890969e-02
#> 761  -0.171312297  2.869095e-02
#> 762  -0.176255336  2.503983e-01
#> 763   0.003296241 -4.593494e-02
#> 764  -0.105290070  1.225021e-01
#> 765  -0.083188054  7.173423e-02
#> 766  -0.077434656  1.173934e-01
#> 767  -0.193138170  5.032517e-02
#> 768  -0.232495713  1.662518e-01
#> 769  -0.278214504  1.438056e-01
#> 770  -0.262822357  1.487553e-01
#> 771  -0.150715860  1.148552e-01
#> 772  -0.225527725 -7.662139e-02
#> 773  -0.063555137  1.767538e-01
#> 774   0.050581527  8.253081e-02
#> 775  -0.043881227 -1.385204e-01
#> 776  -0.321710030  1.610108e-01
#> 777  -0.356370341  1.433268e-01
#> 778  -0.331891024 -2.153437e-02
#> 779  -0.051564139 -1.801340e-02
#> 780  -0.126668290  3.540311e-02
#> 781  -0.084420821  5.527029e-02
#> 782  -0.176652446  2.183280e-01
#> 783  -0.126635866  1.015011e-01
#> 784   0.064896196  5.121799e-02
#> 785  -0.037413646  9.121693e-02
#> 786  -0.177007673  1.699076e-01
#> 787  -0.179876702  1.162829e-01
#> 788  -0.054328437 -5.848159e-02
#> 789  -0.036096306  2.213366e-01
#> 790  -0.098185695 -1.287814e-02
#> 791  -0.173962379  1.951777e-01
#> 792  -0.159509428  1.453735e-01
#> 793  -0.054495685  2.322819e-02
#> 794  -0.155991122  2.430178e-01
#> 795  -0.128201039  1.704341e-01
#> 796  -0.021883825 -6.373295e-02
#> 797  -0.164703042  7.488696e-03
#> 798  -0.177007673  1.699076e-01
#> 799  -0.289264531  2.846707e-02
#> 800  -0.044490577 -2.661283e-02
#> 801  -0.045494126  5.173514e-02
#> 802  -0.161077977  1.327592e-01
#> 803  -0.038980077  4.681823e-02
#> 804  -0.178141622  1.332356e-01
#> 805  -0.088892845  2.360661e-02
#> 806  -0.030963820 -8.639361e-02
#> 807  -0.188465297  1.385958e-01
#> 808  -0.007921778 -1.150472e-02
#> 809  -0.201696823 -3.172599e-02
#> 810  -0.149577781  2.368909e-01
#> 811  -0.075382078  1.130811e-01
#> 812  -0.082134139  3.195803e-01
#> 813  -0.185243332  4.250090e-02
#> 814  -0.111765977  1.094153e-01
#> 815  -0.048696362  3.097324e-02
#> 816  -0.011150916  8.828350e-03
#> 817  -0.106078302  9.563040e-02
#> 818  -0.205976709  1.620773e-01
#> 819  -0.106224324 -4.549332e-02
#> 820  -0.106069867  2.885839e-01
#> 821  -0.117122399  5.520258e-02
#> 822  -0.040594023 -3.156245e-02
#> 823  -0.077324385  1.521081e-01
#> 824  -0.097465787  1.226563e-01
#> 825  -0.250036986  5.717121e-02
#> 826  -0.147417368  1.022189e-01
#> 827   0.042604178  1.814104e-02
#> 828  -0.173962379  1.951777e-01
#> 829  -0.051833898 -2.030138e-01
#> 830  -0.064747647 -9.099299e-04
#> 831  -0.106630699  1.333769e-01
#> 832  -0.016047343 -6.599967e-02
#> 833  -0.052314722 -2.845926e-02
#> 834  -0.220131435 -4.954912e-03
#> 835  -0.109491756 -4.479355e-02
#> 836  -0.127204832  3.721875e-02
#> 837  -0.121558262  7.928325e-02
#> 838  -0.235529362  7.301504e-02
#> 839  -0.076265620  6.080268e-02
#> 840   0.033555455 -5.818824e-02
#> 841   0.020821813  9.461316e-02
#> 842  -0.099741686  8.625766e-02
#> 843  -0.047074952  1.679329e-01
#> 844  -0.271558598  1.280661e-01
#> 845  -0.221201077  2.737923e-01
#> 846  -0.106630699  1.333769e-01
#> 847  -0.136348612  1.578510e-02
#> 848  -0.139244048  1.922860e-01
#> 849  -0.067097208 -9.568137e-02
#> 850  -0.166206888  1.509489e-01
#> 851   0.045712852 -7.955905e-02
#> 852  -0.117588976  1.478690e-01
#> 853  -0.062713849 -1.104629e-01
#> 854  -0.132227744 -1.013013e-02
#> 855  -0.244408365  2.232503e-02
#> 856  -0.063333984 -1.638269e-02
#> 857  -0.032549255  1.161630e-02
#> 858  -0.185955808  1.274251e-01
#> 859  -0.136144758  1.675535e-01
#> 860  -0.009399847 -5.577855e-02
#> 861  -0.094799615 -2.725674e-02
#> 862  -0.053661059  2.020965e-01
#> 863  -0.237221675  1.040900e-02
#> 864  -0.052003566 -3.019954e-02
#> 865   0.005467537  4.400102e-02
#> 866  -0.118130028  1.446007e-01
#> 867  -0.028705614 -1.014883e-01
#> 868   0.004500410  1.433826e-02
#> 869   0.079133838 -3.382443e-02
#> 870  -0.183059095  7.898026e-02
#> 871  -0.183686498  1.318146e-02
#> 872  -0.029986180  4.413480e-02
#> 873  -0.245718262  1.832393e-01
#> 874  -0.154281997 -1.525912e-02
#> 875  -0.102857228  3.323109e-02
#> 876  -0.143504787  1.704739e-01
#> 877  -0.046935065  1.414386e-02
#> 878  -0.209793965  5.201284e-02
#> 879  -0.281445532  1.988005e-01
#> 880  -0.182666608  2.797295e-01
#> 881  -0.056451630  5.319161e-02
#> 882  -0.178141622  1.332356e-01
#> 883  -0.141394747  1.593563e-01
#> 884  -0.083037337 -1.447711e-01
#> 885  -0.052822444  5.931384e-02
#> 886  -0.199247331  1.811605e-01
#> 887  -0.168319934  2.208140e-01
#> 888  -0.063495183 -8.557727e-03
#> 889  -0.110263624  7.293871e-03
#> 890  -0.059217170  1.318455e-02
#> 891  -0.136501236  1.284437e-01
#> 892  -0.106981791  8.228498e-02
#> 893  -0.251384505  1.973516e-01
#> 894  -0.138050047  1.121578e-01
#> 895  -0.178188060  6.525248e-03
#> 896  -0.200419491  1.120160e-01
#> 897  -0.209217573 -3.958249e-02
#> 898  -0.137092765  1.257755e-02
#> 899  -0.153700194  6.092904e-02
#> 900  -0.043628754  1.412320e-01
#> 901  -0.287319124 -2.454192e-02
#> 902  -0.075382078  1.130811e-01
#> 903  -0.080568767  3.078767e-02
#> 904  -0.009380036 -2.226562e-01
#> 905  -0.194055312 -2.745277e-02
#> 906  -0.029986180  4.413480e-02
#> 907  -0.254224698  1.706033e-01
#> 908  -0.374391596  1.080910e-01
#> 909  -0.077481351  2.725002e-02
#> 910  -0.084840851  2.741721e-01
#> 911  -0.123132762  1.399448e-01
#> 912   0.004863331  5.293407e-02
#> 913  -0.159872842  7.571233e-02
#> 914   0.029872285 -2.416383e-02
#> 915  -0.130691907  1.253269e-01
#> 916  -0.131130718  1.315268e-01
#> 917  -0.158306384  7.729718e-02
#> 918  -0.245718262  1.832393e-01
#> 919  -0.137546389  6.089948e-02
#> 920  -0.153513251  2.487648e-01
#> 921  -0.159509428  1.453735e-01
#> 922  -0.214620946  1.222817e-01
#> 923  -0.040807876 -1.986337e-01
#> 924  -0.083067042 -2.314919e-01
#> 925  -0.118351146  2.168795e-01
#> 926  -0.185081864  8.314349e-02
#> 927  -0.152929084 -6.013412e-02
#> 928  -0.173725845 -5.742271e-02
#> 929  -0.246012617  1.816715e-01
#> 930  -0.091394041  1.560251e-01
#> 931   0.001120231 -8.184070e-02
#> 932  -0.075492246  1.395425e-01
#> 933  -0.155852538  3.952466e-02
#> 934   0.031208298  8.709046e-02
#> 935  -0.093495555 -2.941409e-02
#> 936  -0.127195163  1.649140e-01
#> 937  -0.168224181  9.756200e-02
#> 938  -0.185029849  1.627447e-01
#> 939  -0.122053602  2.397408e-01
#> 940  -0.187972365  1.165721e-01
#> 941  -0.100775460 -4.467808e-02
#> 942  -0.176180522  1.149577e-02
#> 943  -0.198238651  1.382508e-01
#> 944  -0.183456470  2.904784e-03
#> 945  -0.083740199  6.820579e-02
#> 946  -0.102056965  4.700187e-02
#> 947  -0.012722376 -5.745566e-02
#> 948  -0.119493140  1.017348e-01
#> 949  -0.182666608  2.797295e-01
#> 950   0.071425404 -1.383018e-01
#> 951  -0.121913218 -1.126443e-01
#> 952  -0.014701961  6.664136e-02
#> 953  -0.168796320 -1.749877e-02
#> 954  -0.126668290  3.540311e-02
#> 955  -0.077920398 -1.474907e-01
#> 956   0.007859629  5.776791e-02
#> 957  -0.125696144  4.058639e-02
#> 958  -0.147541442 -5.727607e-02
#> 959  -0.228447597  1.084883e-01
#> 960  -0.031157445 -1.410246e-01
#> 961  -0.076028906 -6.168426e-03
#> 962  -0.137951923  1.481095e-01
#> 963  -0.043459932  3.069472e-02
#> 964  -0.117273523  4.419776e-02
#> 965  -0.140185949  1.168152e-01
#> 966  -0.037255088 -1.130232e-01
#> 967  -0.069034203  1.183415e-01
#> 968  -0.062325280 -1.688168e-01
#> 969  -0.106503617 -1.471524e-02
#> 970  -0.006481609 -1.681149e-01
#> 971  -0.096412794 -4.372072e-03
#> 972  -0.127639143  9.760048e-02
#> 973  -0.091790112  1.131074e-01
#> 974  -0.391713329  1.915398e-01
#> 975  -0.111681507  2.203772e-01
#> 976  -0.251384505  1.973516e-01
#> 977  -0.178884809  9.340305e-04
#> 978  -0.139703106  9.847858e-02
#> 979  -0.106078302  9.563040e-02
#> 980  -0.103155913  1.936587e-01
#> 981  -0.087059801  9.348378e-02
#> 982  -0.067171083  7.041907e-02
#> 983  -0.030543284 -4.721930e-02
#> 984  -0.279056482  1.400367e-01
#> 985  -0.555897139  6.835353e-03
#> 986  -0.163722570  1.522599e-03
#> 987   0.021360890  6.062700e-02
#> 988  -0.023307507 -4.942500e-02
#> 989  -0.134791199  1.102451e-01
#> 990  -0.027385593  8.904485e-02
#> 991  -0.054822849 -2.425549e-02
#> 992  -0.124013829  1.191296e-01
#> 993  -0.005766274 -1.144296e-02
#> 994  -0.120489293  1.797128e-01
#> 995  -0.215788323  6.156344e-02
#> 996  -0.159856417  1.659310e-01
#> 997  -0.174418882  1.098952e-01
#> 998  -0.026351466  1.498811e-02
#> 999  -0.182345932  8.241573e-02
#> 1000 -0.138239332  2.464605e-01
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.136   0.0281
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.0553  0.230 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.160  -0.0686
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0910  0.0885
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.135   0.104 
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.0360 -0.0450
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.161  -0.0229
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.365   0.188 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0980 -0.0353
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.185   0.0311
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.217  20.0 0.041  0.12 82.5  1651. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.217  20.0 0.041  0.12  5.91  118. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.236  24.5 0.041  0.12 66.5  1629. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.236  24.5 0.041  0.12  3.99   97.8
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.212  18.2 0.041  0.12 91.2  1658. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.212  18.2 0.041  0.12  6.87  125. 
#> # … with 1,994 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>