Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

Usage

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- function() {
  ini({
    THETA_Cl <- 4.0
    THETA_Vc <- 70.0
    THETA_Ka <- 1.0
    ETA_Cl ~ 0.2
    ETA_Vc ~ 0.2
    ETA_Ka ~ 0.2
    prop.sd <- sqrt(0.05)
  })
  model({
    TVCl <- THETA_Cl
    TVVc <- THETA_Vc
    TVKa <- THETA_Ka

    Cl <- TVCl*exp(ETA_Cl)
    Vc <- TVVc*exp(ETA_Vc)
    Ka <- TVKa*exp(ETA_Ka)

    K20 <- Cl/Vc
    Cc <- centr/Vc

    d/dt(depot) = -Ka*depot
    d/dt(centr) = Ka*depot - K20*centr
    Cc ~ prop(prop.sd)
  })
}
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e3,n_resample=1e2)
#>  
#>  
#>  
#>  
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.46161848 -0.716414881  0.614527562
#> 2    0.52148246 -0.668239394 -0.295070271
#> 3    0.62953829 -0.035668729  0.853513874
#> 4    0.69590671 -0.283513665  0.509930085
#> 5    0.50147609 -1.029637839 -0.621775927
#> 6    0.49000963 -0.095834145  0.577837796
#> 7    0.80066132 -0.051989362  0.560248504
#> 8    0.63936787 -0.286972190  0.336041029
#> 9    0.78109649 -0.047955417  0.397608759
#> 10   0.67679514 -0.343639280  0.243496741
#> 11   0.63936787 -0.286972190  0.336041029
#> 12  -0.32395932 -0.085994280  0.517265600
#> 13   0.80066132 -0.051989362  0.560248504
#> 14   0.60507403  0.270746466  1.426091847
#> 15   0.70589549  0.326374015  1.241231425
#> 16   0.14990322 -0.902217106 -0.093434080
#> 17   0.19718678 -0.352419760  0.044054866
#> 18   0.67679514 -0.343639280  0.243496741
#> 19   0.79220198 -0.209018428  0.651123420
#> 20   0.39845486 -0.583596609 -0.291951889
#> 21   0.39827282 -0.312986024  0.425138019
#> 22   0.53161867 -0.844469274 -0.665155530
#> 23   0.80066132 -0.051989362  0.560248504
#> 24   0.37213313 -0.792808796 -0.262432455
#> 25   0.93790898 -0.050599256  0.493316269
#> 26   0.51040712 -0.269260549 -0.112023368
#> 27   0.36447696 -0.497440651  0.473464037
#> 28   0.36447696 -0.497440651  0.473464037
#> 29   0.12590474 -0.022469370  0.797883506
#> 30   0.79220198 -0.209018428  0.651123420
#> 31   0.47447411 -0.298108306 -0.121546765
#> 32   0.78229265 -0.376666334  0.459688153
#> 33   0.16630386 -0.300950761  0.538725292
#> 34   0.37213313 -0.792808796 -0.262432455
#> 35   0.39827282 -0.312986024  0.425138019
#> 36   0.40607742 -0.085998986  0.488732340
#> 37   0.44243266  0.105975602  0.968274428
#> 38   0.40607742 -0.085998986  0.488732340
#> 39   0.37213313 -0.792808796 -0.262432455
#> 40   0.57036658 -0.525414643  0.195565089
#> 41   0.79220198 -0.209018428  0.651123420
#> 42   0.41965191 -0.470263417  0.387193782
#> 43   0.41976488 -0.580546047  0.307176745
#> 44   0.52148246 -0.668239394 -0.295070271
#> 45   0.78109649 -0.047955417  0.397608759
#> 46   0.57059717 -0.291818295  0.117313670
#> 47   0.77547337  0.151922192  0.342736220
#> 48   0.67541201 -0.255922793  0.152039879
#> 49   0.93790898 -0.050599256  0.493316269
#> 50   0.56914542  0.214967421  0.692684071
#> 51   0.63936787 -0.286972190  0.336041029
#> 52   0.57036658 -0.525414643  0.195565089
#> 53   0.69590671 -0.283513665  0.509930085
#> 54   0.14990322 -0.902217106 -0.093434080
#> 55   0.62953829 -0.035668729  0.853513874
#> 56   0.80066132 -0.051989362  0.560248504
#> 57   0.39827282 -0.312986024  0.425138019
#> 58   0.15613558 -0.122753405  0.058341313
#> 59   0.87912662 -0.037608066  0.203510757
#> 60   0.41754683 -0.383638099  0.197262274
#> 61   0.49000963 -0.095834145  0.577837796
#> 62   0.34159502 -0.701549967 -0.315132899
#> 63   0.46161848 -0.716414881  0.614527562
#> 64   0.29021307  0.109017338  0.672013377
#> 65  -0.05383047  0.006687033  0.483670322
#> 66   0.36446593 -0.327182532 -0.010387411
#> 67   0.70589549  0.326374015  1.241231425
#> 68   0.79220198 -0.209018428  0.651123420
#> 69   0.49000963 -0.095834145  0.577837796
#> 70   0.36775106 -0.584814035 -0.203512166
#> 71   0.36775106 -0.584814035 -0.203512166
#> 72   0.41976488 -0.580546047  0.307176745
#> 73   0.55135452 -0.156399508  0.293838236
#> 74   0.35170607 -0.773528781 -0.106449398
#> 75   0.37213313 -0.792808796 -0.262432455
#> 76   0.57059717 -0.291818295  0.117313670
#> 77   0.63936787 -0.286972190  0.336041029
#> 78   0.28861106 -0.604615595  0.418534879
#> 79   0.69590671 -0.283513665  0.509930085
#> 80   0.78829119 -0.293509629  0.123613089
#> 81   0.36104585 -1.050722649 -0.227844833
#> 82   0.78229265 -0.376666334  0.459688153
#> 83   0.57036658 -0.525414643  0.195565089
#> 84   0.53941782 -0.644512133  0.862252763
#> 85   0.44984492 -0.259441665  0.638764520
#> 86   0.57036658 -0.525414643  0.195565089
#> 87   0.53161867 -0.844469274 -0.665155530
#> 88   0.13557690 -0.300215648  0.393477546
#> 89   0.63936787 -0.286972190  0.336041029
#> 90   0.67679514 -0.343639280  0.243496741
#> 91   0.23422745 -0.552345193 -0.007780037
#> 92   0.52148246 -0.668239394 -0.295070271
#> 93   0.52148246 -0.668239394 -0.295070271
#> 94   0.79220198 -0.209018428  0.651123420
#> 95   0.60507403  0.270746466  1.426091847
#> 96   0.57036658 -0.525414643  0.195565089
#> 97   0.78829119 -0.293509629  0.123613089
#> 98   0.80066132 -0.051989362  0.560248504
#> 99   0.80066132 -0.051989362  0.560248504
#> 100  0.87694988 -0.056807607  0.902475597
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 8
#>    sim.id THETA_Cl THETA_Vc THETA_Ka prop.sd ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>  <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1   0.224  0.462 -0.716   0.615
#>  2      2        4       70        1   0.224  0.521 -0.668  -0.295
#>  3      3        4       70        1   0.224  0.630 -0.0357  0.854
#>  4      4        4       70        1   0.224  0.696 -0.284   0.510
#>  5      5        4       70        1   0.224  0.501 -1.03   -0.622
#>  6      6        4       70        1   0.224  0.490 -0.0958  0.578
#>  7      7        4       70        1   0.224  0.801 -0.0520  0.560
#>  8      8        4       70        1   0.224  0.639 -0.287   0.336
#>  9      9        4       70        1   0.224  0.781 -0.0480  0.398
#> 10     10        4       70        1   0.224  0.677 -0.344   0.243
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 14
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20  rxCc    Cc     depot
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>     <dbl>
#> 1      1     1     4    70     1  6.35  34.2 1.85  0.186 39.8  39.8   5.18e+ 2
#> 2      1    14     4    70     1  6.35  34.2 1.85  0.186  5.07  5.07  1.88e- 8
#> 3      2     1     4    70     1  6.74  35.9 0.744 0.188 21.9  21.9   1.15e+ 3
#> 4      2    14     4    70     1  6.74  35.9 0.744 0.188  5.64  5.64  7.21e- 2
#> 5      3     1     4    70     1  7.51  67.5 2.35  0.111 22.9  22.9   3.64e+ 2
#> 6      3    14     4    70     1  7.51  67.5 2.35  0.111  6.74  6.74 -6.23e-10
#> # ℹ 194 more rows
#> # ℹ 2 more variables: centr <dbl>, AUC <dbl>
#>