Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e3,n_resample=1e2)
#> $eta
#>          ETA_Cl      ETA_Vc       ETA_Ka
#> 1    0.42530093 -0.43044603  0.111009998
#> 2   -0.24032419 -0.05673346  0.402213513
#> 3    0.51207168 -0.70158960 -0.206502127
#> 4    0.34904159 -0.53200569  0.174021023
#> 5    0.46462871  0.03688321  0.351156969
#> 6    0.32864874 -0.77408591  0.094060291
#> 7    0.19053914 -0.94957196 -0.533786493
#> 8    0.76843950 -0.17570246  0.352856584
#> 9    0.24940522 -0.88532009  0.078713416
#> 10   0.51207168 -0.70158960 -0.206502127
#> 11   0.58697076 -0.01901534  0.121221766
#> 12   0.85022438 -0.21275494  0.694201571
#> 13   0.82505892 -0.19273144  0.322190382
#> 14   0.82505892 -0.19273144  0.322190382
#> 15   0.46853908 -0.21054844  0.304699457
#> 16   0.24554549 -0.41814548  0.132955397
#> 17   0.73048089 -0.16451293  0.290259654
#> 18   0.51207168 -0.70158960 -0.206502127
#> 19   0.71184910 -0.13231454  0.195372393
#> 20   0.32779695 -0.38789297  0.801806495
#> 21   0.31596580 -0.76788563  0.750746908
#> 22   0.82505892 -0.19273144  0.322190382
#> 23   0.74658274  0.14702607  0.424880083
#> 24   0.51207168 -0.70158960 -0.206502127
#> 25   0.76843950 -0.17570246  0.352856584
#> 26   0.27615714 -0.53154848 -0.181792893
#> 27   0.53334812 -0.68475624 -0.008201034
#> 28   0.73048089 -0.16451293  0.290259654
#> 29   0.73048089 -0.16451293  0.290259654
#> 30   0.54643668 -0.22370847  0.580592046
#> 31   0.70161787 -0.02665225  0.972600433
#> 32   0.30167522 -0.71903998 -0.132059402
#> 33   0.85022438 -0.21275494  0.694201571
#> 34   0.71184910 -0.13231454  0.195372393
#> 35   0.85022438 -0.21275494  0.694201571
#> 36   0.02534419 -0.65301398 -0.255639469
#> 37   0.54643668 -0.22370847  0.580592046
#> 38   0.11368197 -0.42531274 -0.214010542
#> 39   0.82505892 -0.19273144  0.322190382
#> 40   0.54377436  0.01575050  0.678160973
#> 41   0.34019074 -0.59705470 -0.311056586
#> 42   0.49753098 -0.41787129  0.420027125
#> 43   0.39996587 -0.11948754  0.398152113
#> 44   0.32573307 -0.60102198  0.256436174
#> 45   0.76843950 -0.17570246  0.352856584
#> 46   0.39654445 -0.36224479  0.996806392
#> 47   0.22091731  0.20317626  1.324350833
#> 48   0.42530093 -0.43044603  0.111009998
#> 49   0.35963967 -0.18684316  0.119451108
#> 50   0.11108275 -0.91727049  0.238391199
#> 51   0.76843950 -0.17570246  0.352856584
#> 52   0.34019074 -0.59705470 -0.311056586
#> 53   0.49753098 -0.41787129  0.420027125
#> 54   0.45236060  0.01008067  0.930404030
#> 55   0.51207168 -0.70158960 -0.206502127
#> 56   0.62013958 -0.19269950  0.292239848
#> 57   0.53236716 -0.37808380  0.253845025
#> 58   0.49636796  0.19335557  0.604811384
#> 59   0.24554549 -0.41814548  0.132955397
#> 60   0.31596580 -0.76788563  0.750746908
#> 61   0.76843950 -0.17570246  0.352856584
#> 62   0.29996745 -0.74932692 -0.042747302
#> 63   0.76843950 -0.17570246  0.352856584
#> 64   0.82505892 -0.19273144  0.322190382
#> 65   0.42811737 -0.16516572  1.125325409
#> 66   0.42530093 -0.43044603  0.111009998
#> 67   0.51207168 -0.70158960 -0.206502127
#> 68   0.34088648 -0.62587675  0.284986759
#> 69   0.53334812 -0.68475624 -0.008201034
#> 70   0.76843950 -0.17570246  0.352856584
#> 71   0.49753098 -0.41787129  0.420027125
#> 72   0.73048089 -0.16451293  0.290259654
#> 73   0.51207168 -0.70158960 -0.206502127
#> 74   0.70161787 -0.02665225  0.972600433
#> 75   0.53236716 -0.37808380  0.253845025
#> 76   0.24940522 -0.88532009  0.078713416
#> 77   0.29996745 -0.74932692 -0.042747302
#> 78   0.53334812 -0.68475624 -0.008201034
#> 79   0.32573307 -0.60102198  0.256436174
#> 80   0.54377436  0.01575050  0.678160973
#> 81   0.85022438 -0.21275494  0.694201571
#> 82   0.54643668 -0.22370847  0.580592046
#> 83   0.53334812 -0.68475624 -0.008201034
#> 84   0.29996745 -0.74932692 -0.042747302
#> 85   0.70161787 -0.02665225  0.972600433
#> 86   0.51207168 -0.70158960 -0.206502127
#> 87   0.07406204 -0.46827045  0.178052550
#> 88   0.49753098 -0.41787129  0.420027125
#> 89   0.71184910 -0.13231454  0.195372393
#> 90   0.53334812 -0.68475624 -0.008201034
#> 91   0.53986029 -0.48694656 -0.304516414
#> 92   0.51207168 -0.70158960 -0.206502127
#> 93   0.62776319 -0.58847604 -0.350708014
#> 94   0.82505892 -0.19273144  0.322190382
#> 95   0.49753098 -0.41787129  0.420027125
#> 96   0.53334812 -0.68475624 -0.008201034
#> 97   0.53334812 -0.68475624 -0.008201034
#> 98   0.62013958 -0.19269950  0.292239848
#> 99   0.54643668 -0.22370847  0.580592046
#> 100  0.62776319 -0.58847604 -0.350708014
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.425 -0.430   0.111 
#>  2      2        4       70        1 -0.240 -0.0567  0.402 
#>  3      3        4       70        1  0.512 -0.702  -0.207 
#>  4      4        4       70        1  0.349 -0.532   0.174 
#>  5      5        4       70        1  0.465  0.0369  0.351 
#>  6      6        4       70        1  0.329 -0.774   0.0941
#>  7      7        4       70        1  0.191 -0.950  -0.534 
#>  8      8        4       70        1  0.768 -0.176   0.353 
#>  9      9        4       70        1  0.249 -0.885   0.0787
#> 10     10        4       70        1  0.512 -0.702  -0.207 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  6.12  45.5 1.12  0.134  23.3     8.76e+2 1060.
#> 2      1    14     4    70     1  6.12  45.5 1.12  0.134   7.86    4.31e-4  358.
#> 3      2     1     4    70     1  3.15  66.1 1.50  0.0476 19.7     6.67e+2 1304.
#> 4      2    14     4    70     1  3.15  66.1 1.50  0.0476 16.2     2.41e-6 1074.
#> 5      3     1     4    70     1  6.67  34.7 0.813 0.192  24.1     1.09e+3  835.
#> 6      3    14     4    70     1  6.67  34.7 0.813 0.192   5.36    2.80e-2  186.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>