Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 1e+05,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object
A posologyr list, created by the posologyr function.

n_sample
Number of samples from the S-step

n_resample
Number of samples from the R-step

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)

#> 

#> $eta
#>             ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1    -1.071013e-01  2.229117e-01       0       0
#> 2    -1.245007e-01 -6.884644e-03       0       0
#> 3    -3.075022e-01  2.830544e-01       0       0
#> 4    -2.445877e-02 -3.095641e-03       0       0
#> 5    -8.883687e-02 -7.283173e-02       0       0
#> 6     1.541333e-02 -1.884391e-01       0       0
#> 7    -1.210746e-01  1.253975e-01       0       0
#> 8     1.101125e-01 -1.152763e-03       0       0
#> 9    -2.809081e-02 -7.349015e-03       0       0
#> 10   -1.549700e-01  3.663476e-02       0       0
#> 11   -3.883097e-02  1.843268e-01       0       0
#> 12   -6.294038e-02 -5.843322e-02       0       0
#> 13   -1.074979e-01  1.074805e-01       0       0
#> 14   -1.116018e-01  9.234868e-02       0       0
#> 15   -1.462217e-01  1.688936e-01       0       0
#> 16   -1.363995e-01  4.490384e-03       0       0
#> 17   -1.098939e-02 -1.400218e-02       0       0
#> 18   -7.982437e-02  9.121006e-02       0       0
#> 19   -1.093062e-01  6.706099e-02       0       0
#> 20   -2.242150e-01  4.602700e-02       0       0
#> 21   -1.181909e-01 -5.227330e-03       0       0
#> 22   -2.305222e-01  1.520354e-01       0       0
#> 23   -1.107049e-01  2.291288e-01       0       0
#> 24   -1.471401e-02  4.382063e-03       0       0
#> 25   -9.979727e-02  5.897112e-02       0       0
#> 26   -2.038526e-01  1.756275e-01       0       0
#> 27   -8.245107e-02  1.644413e-01       0       0
#> 28   -2.820164e-01  3.487927e-01       0       0
#> 29    9.692489e-02 -1.949000e-01       0       0
#> 30   -1.231958e-01  1.815153e-01       0       0
#> 31    4.900214e-03 -3.647425e-02       0       0
#> 32   -7.960837e-02  6.962957e-02       0       0
#> 33   -5.345000e-02 -1.795018e-02       0       0
#> 34   -1.322408e-01 -9.465327e-02       0       0
#> 35   -1.018990e-01  1.323925e-01       0       0
#> 36   -3.069584e-01  2.463572e-01       0       0
#> 37   -1.704625e-01  1.610499e-01       0       0
#> 38   -9.592222e-02  1.262002e-01       0       0
#> 39   -1.597690e-01 -1.193388e-01       0       0
#> 40   -1.613930e-01  1.267356e-01       0       0
#> 41   -1.471401e-02  4.382063e-03       0       0
#> 42   -1.049524e-01 -2.466147e-02       0       0
#> 43   -6.574668e-02  1.113584e-01       0       0
#> 44   -3.532128e-02 -2.162001e-02       0       0
#> 45   -2.478939e-01  2.124112e-01       0       0
#> 46   -1.115601e-01  3.807306e-02       0       0
#> 47   -1.380048e-01 -6.788219e-02       0       0
#> 48   -7.764920e-02  1.047974e-01       0       0
#> 49   -1.953680e-01  2.588365e-01       0       0
#> 50    5.839070e-02 -1.259972e-01       0       0
#> 51   -7.864695e-02 -1.098670e-02       0       0
#> 52   -3.033241e-01  1.291940e-02       0       0
#> 53   -1.494729e-01  4.579175e-02       0       0
#> 54   -2.450734e-01  1.420808e-01       0       0
#> 55   -2.305222e-01  1.520354e-01       0       0
#> 56   -4.184394e-02 -3.478905e-02       0       0
#> 57   -1.215784e-01  9.749051e-02       0       0
#> 58   -1.454826e-01  3.737101e-01       0       0
#> 59   -9.969644e-02  1.558105e-01       0       0
#> 60   -5.178396e-02 -1.414181e-01       0       0
#> 61    1.406587e-03  1.136987e-01       0       0
#> 62   -1.185786e-01  2.061009e-01       0       0
#> 63   -1.561482e-01  2.055800e-01       0       0
#> 64   -1.320261e-01  8.479667e-02       0       0
#> 65   -2.390061e-01  9.645004e-02       0       0
#> 66   -5.868304e-02 -2.522566e-01       0       0
#> 67   -5.223420e-02  1.406037e-02       0       0
#> 68   -2.820164e-01  3.487927e-01       0       0
#> 69   -1.667200e-01  8.073526e-02       0       0
#> 70   -2.059779e-01  1.163983e-01       0       0
#> 71   -1.629607e-01  1.446403e-01       0       0
#> 72   -1.187268e-01  8.130683e-02       0       0
#> 73   -1.119266e-01  6.084753e-02       0       0
#> 74   -4.023503e-02  8.110438e-02       0       0
#> 75   -1.949851e-01  7.947352e-02       0       0
#> 76   -4.815337e-02  4.683223e-03       0       0
#> 77   -1.293143e-01  1.659874e-01       0       0
#> 78    7.068138e-04  1.942238e-02       0       0
#> 79   -2.510905e-01  2.665201e-01       0       0
#> 80   -1.116018e-01  9.234868e-02       0       0
#> 81    1.889716e-02 -5.864101e-02       0       0
#> 82   -1.590318e-01  2.922195e-02       0       0
#> 83   -2.823817e-01  6.961292e-02       0       0
#> 84   -1.984189e-01  2.130087e-01       0       0
#> 85   -3.039212e-01  2.065712e-01       0       0
#> 86   -1.980745e-01 -1.032387e-02       0       0
#> 87   -2.056661e-01  1.498923e-01       0       0
#> 88   -4.942441e-02 -4.099863e-02       0       0
#> 89   -8.674044e-02  1.302926e-01       0       0
#> 90   -2.446009e-01  1.019159e-01       0       0
#> 91   -3.422362e-01  5.521591e-02       0       0
#> 92   -1.057452e-01 -2.525270e-02       0       0
#> 93   -2.809081e-02 -7.349015e-03       0       0
#> 94   -4.014652e-02 -8.290136e-02       0       0
#> 95   -3.440211e-02  4.416975e-02       0       0
#> 96   -3.163525e-02 -7.683110e-02       0       0
#> 97    2.097584e-02 -4.568083e-02       0       0
#> 98   -2.954703e-01  1.060865e-01       0       0
#> 99   -3.248641e-02 -1.469376e-01       0       0
#> 100  -5.073744e-01  1.037757e-01       0       0
#> 101  -1.463569e-01  2.665665e-01       0       0
#> 102  -2.888788e-01  6.774301e-02       0       0
#> 103  -1.001106e-01  2.651514e-01       0       0
#> 104  -2.769948e-01  2.947201e-01       0       0
#> 105  -1.990941e-02 -8.684683e-02       0       0
#> 106  -2.522211e-02 -8.091670e-03       0       0
#> 107  -3.007723e-02  4.693378e-02       0       0
#> 108  -6.334311e-02  2.134563e-01       0       0
#> 109  -1.224491e-01  8.867657e-02       0       0
#> 110  -1.452774e-01  2.830481e-02       0       0
#> 111  -1.802472e-01  6.118488e-02       0       0
#> 112  -8.395340e-02  4.406656e-02       0       0
#> 113   8.050219e-04  8.580140e-02       0       0
#> 114  -1.741719e-01  8.999774e-02       0       0
#> 115  -7.626616e-03  1.657327e-01       0       0
#> 116  -2.419076e-02 -1.015439e-01       0       0
#> 117  -1.985128e-01  5.293871e-02       0       0
#> 118  -1.113448e-01  1.680795e-01       0       0
#> 119  -7.360177e-02 -5.655499e-02       0       0
#> 120  -5.176568e-02 -1.366543e-01       0       0
#> 121  -1.436894e-01  1.466844e-01       0       0
#> 122  -1.570968e-01  3.348107e-02       0       0
#> 123   6.076612e-02  8.805914e-02       0       0
#> 124  -9.034195e-02 -2.099218e-02       0       0
#> 125   1.078267e-02  2.120228e-02       0       0
#> 126  -6.077685e-02 -7.990646e-02       0       0
#> 127   2.073308e-02 -6.438070e-03       0       0
#> 128  -1.556414e-01  2.174508e-01       0       0
#> 129  -2.269241e-02  6.495856e-02       0       0
#> 130  -1.987901e-01 -3.706072e-02       0       0
#> 131  -1.186582e-01  1.394374e-01       0       0
#> 132  -1.618631e-01  5.893161e-02       0       0
#> 133  -1.164699e-01  1.743955e-02       0       0
#> 134  -2.813388e-01  1.725624e-01       0       0
#> 135  -1.550448e-01  2.367887e-01       0       0
#> 136  -1.051144e-01  8.950040e-02       0       0
#> 137  -1.475056e-01  1.627193e-01       0       0
#> 138  -1.568433e-01  2.971060e-02       0       0
#> 139  -2.850348e-01  1.706889e-01       0       0
#> 140  -4.727634e-03  2.200083e-02       0       0
#> 141  -1.069945e-01  9.609092e-02       0       0
#> 142  -2.650617e-02  1.874435e-01       0       0
#> 143  -1.032840e-01  3.841246e-03       0       0
#> 144  -1.017048e-01  1.118037e-01       0       0
#> 145  -5.997218e-02  7.184330e-02       0       0
#> 146  -6.173065e-02 -4.672208e-02       0       0
#> 147  -1.071013e-01  2.229117e-01       0       0
#> 148  -4.014652e-02 -8.290136e-02       0       0
#> 149  -1.855756e-01  6.812820e-02       0       0
#> 150  -3.758364e-02  3.647614e-02       0       0
#> 151  -1.887000e-01  9.610088e-02       0       0
#> 152  -5.949125e-02  4.497099e-02       0       0
#> 153  -7.948251e-02  5.006184e-02       0       0
#> 154  -9.800110e-02  2.938333e-01       0       0
#> 155  -1.660631e-01  1.383678e-01       0       0
#> 156  -1.293143e-01  1.659874e-01       0       0
#> 157  -8.395340e-02  4.406656e-02       0       0
#> 158  -1.081883e-01  2.538144e-02       0       0
#> 159  -2.282250e-01  2.568921e-01       0       0
#> 160   6.397671e-03  5.750748e-02       0       0
#> 161  -1.098647e-01  2.110530e-01       0       0
#> 162  -1.901662e-01  1.857738e-01       0       0
#> 163  -1.561482e-01  2.055800e-01       0       0
#> 164  -7.940679e-02 -1.071325e-01       0       0
#> 165  -1.141607e-01  1.482157e-02       0       0
#> 166  -4.532841e-02  1.014077e-01       0       0
#> 167  -2.462399e-01  2.589371e-02       0       0
#> 168  -2.193777e-01  7.138150e-02       0       0
#> 169  -2.300598e-01  2.057392e-01       0       0
#> 170  -1.026586e-01  2.350954e-02       0       0
#> 171  -9.328164e-02  3.017728e-04       0       0
#> 172  -1.418468e-01  5.899018e-02       0       0
#> 173  -4.639254e-03  1.170791e-02       0       0
#> 174  -1.487249e-02  1.573509e-02       0       0
#> 175  -8.618071e-02 -4.799655e-02       0       0
#> 176  -5.713853e-02  3.348464e-03       0       0
#> 177  -3.275802e-01  2.176464e-01       0       0
#> 178  -2.602159e-02  1.703110e-02       0       0
#> 179  -3.498609e-02 -3.562299e-02       0       0
#> 180  -9.741528e-02  1.082069e-01       0       0
#> 181  -4.376206e-02 -5.509117e-03       0       0
#> 182  -1.935905e-01  6.728995e-02       0       0
#> 183  -4.686548e-02 -2.431715e-03       0       0
#> 184  -1.561482e-01  2.055800e-01       0       0
#> 185   1.541333e-02 -1.884391e-01       0       0
#> 186  -1.171984e-01  6.629494e-02       0       0
#> 187  -9.379560e-02 -1.756686e-02       0       0
#> 188  -2.937716e-03 -7.950223e-02       0       0
#> 189  -4.469508e-02 -1.179898e-01       0       0
#> 190  -1.120031e-01  1.620391e-01       0       0
#> 191  -2.848266e-02 -4.644497e-02       0       0
#> 192  -5.248694e-02  1.268891e-02       0       0
#> 193  -1.446935e-01  7.112818e-03       0       0
#> 194  -1.275337e-01  2.046951e-01       0       0
#> 195  -2.831037e-02 -1.677454e-02       0       0
#> 196  -3.774689e-02 -2.846798e-02       0       0
#> 197  -7.539197e-02  1.425933e-01       0       0
#> 198  -8.674044e-02  1.302926e-01       0       0
#> 199  -1.116018e-01  9.234868e-02       0       0
#> 200  -5.309322e-02  4.760533e-02       0       0
#> 201  -2.071594e-01  1.461335e-01       0       0
#> 202  -1.855101e-01  2.050269e-01       0       0
#> 203  -4.225329e-02  6.367005e-02       0       0
#> 204  -6.973619e-02 -1.109536e-02       0       0
#> 205  -4.327172e-02 -3.496749e-02       0       0
#> 206  -1.519561e-02  9.216206e-02       0       0
#> 207  -2.411226e-01  1.554679e-01       0       0
#> 208  -7.912833e-02 -3.056193e-02       0       0
#> 209  -1.967902e-01 -8.070795e-02       0       0
#> 210  -7.610881e-02  6.097266e-02       0       0
#> 211  -5.367948e-02  1.683756e-01       0       0
#> 212   4.995712e-02  5.750917e-02       0       0
#> 213  -4.078265e-02  1.199820e-01       0       0
#> 214  -1.140549e-01  1.135123e-01       0       0
#> 215  -2.347214e-01  9.304436e-02       0       0
#> 216  -2.927033e-01 -8.202941e-02       0       0
#> 217  -1.762429e-01  9.638010e-02       0       0
#> 218  -2.820164e-01  3.487927e-01       0       0
#> 219  -6.842080e-02  1.923957e-01       0       0
#> 220  -1.719645e-01  1.184220e-01       0       0
#> 221  -2.450498e-01 -4.342990e-02       0       0
#> 222  -6.659083e-02  1.250044e-01       0       0
#> 223  -3.569561e-01  1.257026e-02       0       0
#> 224  -8.894230e-02  1.321542e-01       0       0
#> 225  -2.845447e-01  1.853037e-01       0       0
#> 226  -2.813951e-01 -3.208762e-02       0       0
#> 227  -1.233374e-01  1.592197e-01       0       0
#> 228  -1.842062e-02  2.063433e-01       0       0
#> 229  -1.171984e-01  6.629494e-02       0       0
#> 230  -1.098647e-01  2.110530e-01       0       0
#> 231  -1.072377e-01  1.521161e-01       0       0
#> 232   8.339250e-02 -1.097930e-01       0       0
#> 233  -2.591991e-02 -1.073247e-01       0       0
#> 234  -2.289556e-01  6.949690e-02       0       0
#> 235   1.117919e-02  2.165595e-02       0       0
#> 236  -2.480546e-01  1.044993e-01       0       0
#> 237  -2.006780e-01  1.421737e-01       0       0
#> 238   3.349587e-02 -4.693169e-02       0       0
#> 239  -1.667200e-01  8.073526e-02       0       0
#> 240  -7.197634e-02  2.601132e-02       0       0
#> 241  -7.757603e-02 -1.136465e-01       0       0
#> 242  -1.317008e-01  4.599431e-02       0       0
#> 243  -6.327848e-02  2.161244e-03       0       0
#> 244  -8.530586e-02  4.051699e-02       0       0
#> 245  -2.215766e-01  1.403517e-01       0       0
#> 246   8.442147e-02 -2.452332e-01       0       0
#> 247  -2.261585e-01  2.553585e-03       0       0
#> 248  -4.360608e-01  2.096156e-01       0       0
#> 249  -1.490424e-02  1.071442e-01       0       0
#> 250  -6.347681e-02 -6.512142e-02       0       0
#> 251  -1.089485e-01  1.841881e-01       0       0
#> 252  -1.332791e-01  9.248465e-02       0       0
#> 253  -1.873266e-01  2.391952e-02       0       0
#> 254  -4.497382e-02  1.496077e-01       0       0
#> 255  -1.987555e-01  1.397689e-01       0       0
#> 256  -2.144739e-01  2.282823e-02       0       0
#> 257  -1.140054e-01  3.573768e-02       0       0
#> 258  -1.309646e-01  2.619167e-01       0       0
#> 259  -1.126605e-01 -3.484529e-02       0       0
#> 260  -8.894230e-02  1.321542e-01       0       0
#> 261   4.825478e-02 -7.472149e-02       0       0
#> 262  -8.318526e-02  1.308971e-01       0       0
#> 263  -1.582858e-01  5.461725e-02       0       0
#> 264  -9.260277e-02 -1.234026e-01       0       0
#> 265  -3.797717e-01  1.850886e-01       0       0
#> 266  -2.006780e-01  1.421737e-01       0       0
#> 267  -6.198298e-02 -9.215224e-02       0       0
#> 268  -3.075022e-01  2.830544e-01       0       0
#> 269  -8.665060e-02  8.179012e-02       0       0
#> 270  -2.035689e-01 -5.175202e-02       0       0
#> 271  -1.819759e-01  8.281190e-03       0       0
#> 272  -7.616672e-02  1.970669e-01       0       0
#> 273  -3.163525e-02 -7.683110e-02       0       0
#> 274  -4.009354e-02  2.595673e-02       0       0
#> 275  -1.984189e-01  2.130087e-01       0       0
#> 276  -1.457712e-01  6.872238e-02       0       0
#> 277  -2.148088e-01  2.268093e-01       0       0
#> 278  -1.060839e-01 -3.869268e-02       0       0
#> 279  -2.190618e-01 -9.010168e-02       0       0
#> 280  -9.599202e-02  1.189929e-01       0       0
#> 281  -1.204183e-01 -4.760033e-02       0       0
#> 282  -1.452774e-01  2.830481e-02       0       0
#> 283  -1.181909e-01 -5.227330e-03       0       0
#> 284  -1.361058e-01 -6.454316e-02       0       0
#> 285  -3.758364e-02  3.647614e-02       0       0
#> 286  -6.309267e-02  8.212747e-02       0       0
#> 287  -1.832828e-01 -4.948628e-02       0       0
#> 288  -1.227485e-02  7.371474e-02       0       0
#> 289  -5.531200e-02 -3.105625e-02       0       0
#> 290  -1.499078e-01  7.557418e-02       0       0
#> 291  -1.069051e-01  2.367351e-01       0       0
#> 292  -9.060209e-02  6.620252e-02       0       0
#> 293  -7.779394e-02  4.032452e-02       0       0
#> 294  -3.113010e-01 -5.736720e-03       0       0
#> 295  -2.851019e-01  2.592936e-01       0       0
#> 296  -2.462399e-01  2.589371e-02       0       0
#> 297  -1.360303e-02  1.744365e-01       0       0
#> 298  -2.823691e-01  1.004096e-01       0       0
#> 299   3.694984e-02  8.234622e-02       0       0
#> 300  -2.084907e-02 -9.947175e-02       0       0
#> 301  -1.844001e-01  1.366542e-01       0       0
#> 302  -1.549700e-01  3.663476e-02       0       0
#> 303  -3.606425e-01  2.060112e-01       0       0
#> 304  -3.514297e-02  1.716990e-01       0       0
#> 305  -1.256668e-01 -1.268770e-01       0       0
#> 306  -1.161955e-01  1.763116e-02       0       0
#> 307  -2.180939e-01  1.077789e-01       0       0
#> 308  -2.280139e-02  1.031940e-02       0       0
#> 309   4.846096e-02 -1.809478e-01       0       0
#> 310  -1.058878e-02  1.183270e-02       0       0
#> 311  -1.384566e-01  1.980699e-01       0       0
#> 312  -1.864322e-01 -2.313554e-02       0       0
#> 313  -2.478939e-01  2.124112e-01       0       0
#> 314  -1.550917e-01  1.022963e-01       0       0
#> 315  -1.887000e-01  9.610088e-02       0       0
#> 316  -1.362675e-01  1.324458e-01       0       0
#> 317  -2.169034e-02  7.604377e-03       0       0
#> 318  -2.595724e-01  2.729483e-01       0       0
#> 319  -4.170039e-02  6.628811e-02       0       0
#> 320  -4.466485e-02  1.274746e-01       0       0
#> 321  -8.861407e-02  3.637646e-02       0       0
#> 322  -2.104377e-01  5.359241e-02       0       0
#> 323  -1.440358e-01  1.054437e-01       0       0
#> 324  -1.995140e-01 -2.420611e-02       0       0
#> 325  -1.586982e-02  2.861482e-02       0       0
#> 326  -9.577237e-02  4.535321e-02       0       0
#> 327  -2.227405e-01  2.706176e-01       0       0
#> 328  -5.201464e-03 -1.086840e-01       0       0
#> 329  -1.325681e-01 -7.496160e-02       0       0
#> 330  -2.588886e-01 -2.834354e-02       0       0
#> 331  -5.237926e-02  5.778795e-02       0       0
#> 332  -2.292269e-01  7.836695e-02       0       0
#> 333  -2.780351e-01  9.822389e-02       0       0
#> 334  -6.610416e-02  6.765460e-02       0       0
#> 335  -2.033634e-01  1.414361e-01       0       0
#> 336  -1.989598e-01  4.105077e-02       0       0
#> 337   2.999462e-02 -1.531730e-01       0       0
#> 338  -1.832598e-01  1.316754e-02       0       0
#> 339  -1.370574e-01  6.287342e-02       0       0
#> 340   2.552810e-02 -1.453588e-01       0       0
#> 341  -8.484239e-02 -1.005713e-01       0       0
#> 342  -1.146059e-01  7.348112e-03       0       0
#> 343   3.151783e-02 -2.392021e-01       0       0
#> 344  -1.583283e-01  6.167622e-02       0       0
#> 345  -3.558853e-02  3.827284e-02       0       0
#> 346  -8.069220e-02  2.281898e-01       0       0
#> 347  -8.202783e-02  6.382774e-02       0       0
#> 348  -8.111015e-02  7.935579e-02       0       0
#> 349  -7.227299e-02 -9.368168e-02       0       0
#> 350  -1.458658e-01 -1.473757e-01       0       0
#> 351   2.046983e-03 -1.966566e-02       0       0
#> 352  -1.454307e-01  3.174614e-01       0       0
#> 353  -1.929988e-01  1.458241e-01       0       0
#> 354  -2.595724e-01  2.729483e-01       0       0
#> 355  -1.126605e-01 -3.484529e-02       0       0
#> 356   4.900119e-03  2.267548e-02       0       0
#> 357  -9.696620e-02  1.404324e-01       0       0
#> 358  -3.798732e-02  2.002049e-03       0       0
#> 359  -2.279349e-01 -5.840875e-02       0       0
#> 360  -2.524787e-02 -9.330925e-02       0       0
#> 361  -5.251006e-02  2.058303e-01       0       0
#> 362  -1.675423e-01  1.058658e-01       0       0
#> 363  -3.183534e-01  2.392046e-01       0       0
#> 364  -1.483786e-01  3.424241e-01       0       0
#> 365  -6.628454e-02  6.954589e-02       0       0
#> 366  -1.683383e-01  7.580286e-02       0       0
#> 367  -2.629448e-01  1.042409e-01       0       0
#> 368  -3.774689e-02 -2.846798e-02       0       0
#> 369  -1.178043e-01  9.223419e-02       0       0
#> 370  -5.523234e-02 -1.241675e-01       0       0
#> 371  -2.037509e-01 -6.523691e-02       0       0
#> 372  -1.568433e-01  2.971060e-02       0       0
#> 373  -5.678952e-02  7.890838e-02       0       0
#> 374  -1.033998e-01  1.658852e-01       0       0
#> 375  -1.120987e-01 -8.931799e-02       0       0
#> 376   3.007578e-02  8.617464e-02       0       0
#> 377  -1.033998e-01  1.658852e-01       0       0
#> 378  -2.832465e-02  2.114534e-02       0       0
#> 379  -1.193620e-01  2.301642e-01       0       0
#> 380  -9.371554e-02  5.067968e-02       0       0
#> 381  -5.881424e-02  7.087423e-02       0       0
#> 382  -1.541723e-01  9.056730e-02       0       0
#> 383  -2.059588e-02  6.115813e-02       0       0
#> 384  -8.894230e-02  1.321542e-01       0       0
#> 385  -4.020452e-01  3.731740e-02       0       0
#> 386  -5.013035e-02  1.737342e-01       0       0
#> 387  -1.284615e-01  2.903699e-02       0       0
#> 388  -1.478481e-01  8.925314e-02       0       0
#> 389  -2.290658e-01  1.117545e-01       0       0
#> 390  -7.824180e-02 -6.550701e-02       0       0
#> 391  -4.992901e-02 -1.322421e-02       0       0
#> 392   1.200306e-04 -7.297353e-02       0       0
#> 393  -1.125912e-01 -1.776009e-01       0       0
#> 394  -1.958946e-01  2.354100e-01       0       0
#> 395  -1.454826e-01  3.737101e-01       0       0
#> 396  -2.604972e-01  1.951634e-01       0       0
#> 397   4.103387e-03  5.010853e-02       0       0
#> 398  -6.720519e-03 -3.542888e-02       0       0
#> 399  -3.477018e-02  3.061039e-02       0       0
#> 400   6.861577e-02 -1.571741e-01       0       0
#> 401  -2.322474e-01  1.553192e-01       0       0
#> 402  -1.940007e-01  1.296300e-01       0       0
#> 403  -1.435818e-01  8.757296e-02       0       0
#> 404  -1.512006e-01  8.935850e-02       0       0
#> 405  -1.872094e-01  5.999511e-02       0       0
#> 406  -9.997435e-02  1.547286e-01       0       0
#> 407  -1.974950e-01  7.673662e-02       0       0
#> 408  -2.629448e-01  1.042409e-01       0       0
#> 409  -1.471801e-01  1.372886e-01       0       0
#> 410  -1.454826e-01  3.737101e-01       0       0
#> 411  -4.734738e-02 -6.707166e-02       0       0
#> 412  -1.504080e-01 -1.947756e-02       0       0
#> 413  -4.851101e-02 -1.844366e-02       0       0
#> 414  -1.271000e-01  3.775993e-02       0       0
#> 415  -2.367586e-01  9.746292e-02       0       0
#> 416  -2.873902e-02  1.148384e-05       0       0
#> 417  -1.088791e-01 -1.256327e-02       0       0
#> 418  -8.661423e-02  1.687370e-01       0       0
#> 419  -1.033998e-01  1.658852e-01       0       0
#> 420  -5.222784e-02  1.272274e-03       0       0
#> 421  -2.042253e-01 -8.435115e-02       0       0
#> 422  -1.205990e-01 -1.137015e-02       0       0
#> 423  -1.017476e-01  2.545039e-01       0       0
#> 424  -3.840766e-01  1.293913e-01       0       0
#> 425  -2.782696e-01  3.502764e-02       0       0
#> 426  -1.196551e-01 -6.489095e-02       0       0
#> 427  -1.366662e-01  1.733684e-01       0       0
#> 428  -1.525748e-01  3.297514e-02       0       0
#> 429  -1.483278e-01  1.774924e-01       0       0
#> 430  -3.130749e-01  1.632014e-01       0       0
#> 431  -8.371092e-02  1.196294e-01       0       0
#> 432  -9.746554e-02 -4.211423e-02       0       0
#> 433  -6.740764e-02  5.360772e-02       0       0
#> 434  -4.335559e-03  1.205730e-01       0       0
#> 435   8.346948e-04 -4.811482e-02       0       0
#> 436  -1.818508e-02 -7.871140e-02       0       0
#> 437  -2.651887e-02 -1.094990e-01       0       0
#> 438  -2.548342e-01 -3.804698e-02       0       0
#> 439  -1.754475e-01  7.801359e-02       0       0
#> 440   3.201121e-02 -1.065548e-01       0       0
#> 441  -1.033998e-01  1.658852e-01       0       0
#> 442  -6.760797e-02  1.534125e-01       0       0
#> 443  -7.497522e-02  9.777528e-02       0       0
#> 444  -2.090629e-01  1.587842e-01       0       0
#> 445  -1.001179e-01  7.068675e-02       0       0
#> 446  -1.459300e-01  4.508029e-03       0       0
#> 447  -1.113226e-01  1.436363e-01       0       0
#> 448  -1.331516e-01  3.761979e-02       0       0
#> 449  -2.375432e-01 -4.592540e-03       0       0
#> 450  -1.611591e-01 -2.181157e-01       0       0
#> 451  -4.928333e-02 -2.474291e-02       0       0
#> 452  -3.189941e-01  1.611386e-01       0       0
#> 453  -8.713795e-02  3.175911e-02       0       0
#> 454  -1.688894e-01  1.466197e-01       0       0
#> 455  -2.510905e-01  2.665201e-01       0       0
#> 456  -1.552804e-01  1.694863e-01       0       0
#> 457  -1.291187e-01  1.456569e-01       0       0
#> 458  -1.646823e-01  1.106045e-01       0       0
#> 459  -4.544877e-02  5.558812e-02       0       0
#> 460  -7.764920e-02  1.047974e-01       0       0
#> 461  -1.819229e-01 -1.115089e-01       0       0
#> 462  -2.396907e-05  3.396716e-02       0       0
#> 463   7.828502e-02  2.190728e-02       0       0
#> 464  -6.118104e-02 -1.224470e-01       0       0
#> 465  -1.450117e-01  2.793909e-02       0       0
#> 466  -3.528550e-02 -3.516590e-02       0       0
#> 467  -1.719885e-01  8.294708e-02       0       0
#> 468  -3.282454e-01  4.932946e-02       0       0
#> 469  -8.584483e-02 -1.298860e-01       0       0
#> 470  -2.524787e-02 -9.330925e-02       0       0
#> 471  -1.021400e-01  1.268930e-01       0       0
#> 472  -7.610806e-03  5.650584e-02       0       0
#> 473  -9.023575e-02  4.222192e-02       0       0
#> 474  -5.475677e-02  8.959691e-02       0       0
#> 475  -1.048598e-01 -7.595697e-02       0       0
#> 476  -9.447790e-02  2.936418e-01       0       0
#> 477  -2.103905e-01  4.085107e-02       0       0
#> 478  -1.032641e-01  1.218499e-02       0       0
#> 479   1.033121e-02 -7.059716e-02       0       0
#> 480  -1.561482e-01  2.055800e-01       0       0
#> 481  -6.326103e-02  9.109194e-02       0       0
#> 482  -2.420652e-01 -3.642452e-02       0       0
#> 483  -1.882534e-01  5.483862e-02       0       0
#> 484  -1.372885e-01  6.475808e-02       0       0
#> 485  -2.068264e-01  4.425988e-02       0       0
#> 486  -1.113448e-01  1.680795e-01       0       0
#> 487  -1.399665e-01  1.512500e-01       0       0
#> 488  -2.059588e-02  6.115813e-02       0       0
#> 489  -1.556886e-01 -1.074232e-01       0       0
#> 490  -9.782132e-02 -6.185547e-02       0       0
#> 491  -4.410620e-02 -2.193945e-02       0       0
#> 492  -1.646069e-01  1.029859e-01       0       0
#> 493  -1.552804e-01  1.694863e-01       0       0
#> 494  -6.326103e-02  9.109194e-02       0       0
#> 495  -4.880996e-02  1.531877e-01       0       0
#> 496  -1.332791e-01  9.248465e-02       0       0
#> 497  -1.042222e-01  1.526534e-01       0       0
#> 498  -3.449355e-02  1.788401e-01       0       0
#> 499  -6.253110e-02  1.449739e-01       0       0
#> 500  -9.379560e-02 -1.756686e-02       0       0
#> 501  -1.951395e-01  1.915900e-01       0       0
#> 502  -7.539197e-02  1.425933e-01       0       0
#> 503  -8.845629e-03  3.094721e-01       0       0
#> 504  -2.820527e-02 -6.386269e-02       0       0
#> 505  -7.876909e-02  1.962729e-01       0       0
#> 506  -8.661423e-02  1.687370e-01       0       0
#> 507  -1.079409e-01 -5.988259e-02       0       0
#> 508  -1.269453e-01  9.379116e-02       0       0
#> 509  -1.718533e-01  1.563341e-01       0       0
#> 510  -1.524312e-01 -3.983826e-02       0       0
#> 511  -6.044677e-02  1.048436e-01       0       0
#> 512  -1.030940e-01  2.661686e-02       0       0
#> 513  -1.652678e-01  1.855590e-01       0       0
#> 514   2.080572e-02 -1.027746e-01       0       0
#> 515  -2.510905e-01  2.665201e-01       0       0
#> 516  -1.689648e-02  4.660715e-02       0       0
#> 517  -1.965746e-01  2.548632e-02       0       0
#> 518   2.045059e-02 -1.020424e-01       0       0
#> 519  -4.666960e-02  8.085629e-02       0       0
#> 520  -1.884527e-01  1.555685e-01       0       0
#> 521  -1.870174e-01 -7.249358e-02       0       0
#> 522  -1.953220e-01 -2.844725e-03       0       0
#> 523  -1.799429e-01 -1.491469e-01       0       0
#> 524   6.395716e-02 -1.550646e-01       0       0
#> 525  -5.667219e-02  3.652389e-02       0       0
#> 526  -1.070464e-01  1.427157e-01       0       0
#> 527  -1.655881e-01 -8.130102e-03       0       0
#> 528  -8.833063e-02  1.781981e-01       0       0
#> 529  -8.278323e-02  3.222283e-02       0       0
#> 530  -1.855101e-01  2.050269e-01       0       0
#> 531  -1.018990e-01  1.323925e-01       0       0
#> 532  -2.323771e-01  8.416330e-02       0       0
#> 533  -2.123730e-01 -9.851598e-02       0       0
#> 534  -1.017476e-01  2.545039e-01       0       0
#> 535  -9.870086e-02  2.698988e-01       0       0
#> 536   7.049466e-03 -1.248165e-01       0       0
#> 537  -2.065501e-01  8.572049e-02       0       0
#> 538   2.323412e-02  4.525115e-02       0       0
#> 539  -1.949851e-01  7.947352e-02       0       0
#> 540  -1.098647e-01  2.110530e-01       0       0
#> 541   2.696792e-02  6.418629e-02       0       0
#> 542  -5.527101e-02 -9.095893e-02       0       0
#> 543  -4.376206e-02 -5.509117e-03       0       0
#> 544  -1.074979e-01  1.074805e-01       0       0
#> 545  -4.391934e-02  1.499343e-01       0       0
#> 546  -1.460476e-01 -6.996776e-03       0       0
#> 547   1.690579e-02 -8.134055e-02       0       0
#> 548  -8.194357e-02 -2.191079e-02       0       0
#> 549  -5.251006e-02  2.058303e-01       0       0
#> 550  -1.735649e-01  1.364635e-01       0       0
#> 551  -8.193251e-04  1.241975e-03       0       0
#> 552  -8.318526e-02  1.308971e-01       0       0
#> 553  -9.830192e-02  1.226099e-01       0       0
#> 554  -2.431671e-02  2.711017e-02       0       0
#> 555  -2.994254e-01  1.471608e-01       0       0
#> 556  -2.116928e-01  1.649250e-01       0       0
#> 557  -1.916868e-01  5.521804e-03       0       0
#> 558  -1.242219e-01  1.844206e-02       0       0
#> 559  -5.195435e-02 -7.566760e-02       0       0
#> 560  -2.217218e-01  1.052278e-01       0       0
#> 561  -3.689964e-01  2.034427e-01       0       0
#> 562  -5.248694e-02  1.268891e-02       0       0
#> 563  -2.648643e-01  1.350063e-01       0       0
#> 564  -3.617267e-02 -1.686023e-02       0       0
#> 565  -1.984189e-01  2.130087e-01       0       0
#> 566  -1.309646e-01  2.619167e-01       0       0
#> 567  -8.219655e-02 -6.082023e-02       0       0
#> 568  -1.120987e-01 -8.931799e-02       0       0
#> 569  -2.183530e-01  5.971879e-02       0       0
#> 570  -2.848266e-02 -4.644497e-02       0       0
#> 571  -2.595724e-01  2.729483e-01       0       0
#> 572  -2.038526e-01  1.756275e-01       0       0
#> 573  -3.361330e-01  1.587681e-01       0       0
#> 574  -3.941512e-02 -9.003536e-02       0       0
#> 575  -3.444869e-02 -5.335376e-02       0       0
#> 576  -2.385962e-01  1.380729e-01       0       0
#> 577  -1.619385e-01  4.890952e-02       0       0
#> 578  -1.530554e-01  1.410048e-01       0       0
#> 579  -2.424756e-01 -7.490091e-02       0       0
#> 580  -2.322474e-01  1.553192e-01       0       0
#> 581  -2.498681e-03 -2.046479e-02       0       0
#> 582  -1.227485e-02  7.371474e-02       0       0
#> 583  -8.318526e-02  1.308971e-01       0       0
#> 584  -1.233636e-01  1.574469e-01       0       0
#> 585  -4.170039e-02  6.628811e-02       0       0
#> 586  -2.556718e-01  1.194597e-01       0       0
#> 587  -1.285046e-01  7.280252e-02       0       0
#> 588  -5.997218e-02  7.184330e-02       0       0
#> 589  -1.762262e-01  1.968004e-02       0       0
#> 590  -2.170316e-01 -6.676830e-02       0       0
#> 591   4.917331e-02 -1.159446e-01       0       0
#> 592  -1.287591e-01 -1.398191e-02       0       0
#> 593  -1.227067e-01 -2.382154e-02       0       0
#> 594  -6.760797e-02  1.534125e-01       0       0
#> 595  -1.350720e-01  3.744907e-02       0       0
#> 596  -1.034358e-01  1.702726e-01       0       0
#> 597  -4.851101e-02 -1.844366e-02       0       0
#> 598  -8.992427e-03 -4.520657e-02       0       0
#> 599  -1.224491e-01  8.867657e-02       0       0
#> 600  -2.034222e-01  7.729788e-02       0       0
#> 601  -6.214910e-02 -5.937254e-02       0       0
#> 602  -3.556671e-02  7.961314e-02       0       0
#> 603  -1.759502e-02  1.794875e-01       0       0
#> 604  -1.199788e-01 -4.196983e-03       0       0
#> 605  -2.145100e-01  2.197183e-01       0       0
#> 606  -1.553461e-01  9.070654e-02       0       0
#> 607  -8.668211e-02  6.965781e-03       0       0
#> 608  -7.747969e-02  1.823341e-01       0       0
#> 609  -1.409885e-01 -1.201584e-01       0       0
#> 610   4.012339e-02 -1.032338e-01       0       0
#> 611  -2.587371e-01  1.355641e-01       0       0
#> 612  -5.584122e-02  9.520253e-02       0       0
#> 613  -8.111015e-02  7.935579e-02       0       0
#> 614  -1.561482e-01  2.055800e-01       0       0
#> 615  -2.447031e-01  1.037243e-01       0       0
#> 616  -1.880800e-01  6.911710e-02       0       0
#> 617  -1.325434e-01  3.480500e-02       0       0
#> 618  -4.078265e-02  1.199820e-01       0       0
#> 619  -1.906718e-01 -1.071438e-02       0       0
#> 620  -9.155075e-02  1.609525e-01       0       0
#> 621  -6.929709e-03 -1.514999e-01       0       0
#> 622  -1.040100e-01  1.038796e-01       0       0
#> 623  -8.530586e-02  4.051699e-02       0       0
#> 624  -5.178396e-02 -1.414181e-01       0       0
#> 625  -2.170455e-01  3.852799e-02       0       0
#> 626  -3.444869e-02 -5.335376e-02       0       0
#> 627  -1.524010e-01  9.539028e-02       0       0
#> 628  -1.845723e-02 -3.230555e-02       0       0
#> 629  -1.404984e-01  1.152126e-01       0       0
#> 630  -1.285046e-01  7.280252e-02       0       0
#> 631  -1.529588e-01  1.987396e-02       0       0
#> 632  -6.720519e-03 -3.542888e-02       0       0
#> 633   8.271816e-03  3.297992e-02       0       0
#> 634  -1.384566e-01  1.980699e-01       0       0
#> 635  -1.940007e-01  1.296300e-01       0       0
#> 636  -1.738373e-01 -6.555167e-02       0       0
#> 637  -2.619179e-01 -1.171660e-01       0       0
#> 638  -4.887144e-02 -1.665417e-01       0       0
#> 639   1.962954e-03  1.405118e-01       0       0
#> 640  -3.737224e-02  1.510500e-01       0       0
#> 641  -9.981586e-02 -6.942829e-02       0       0
#> 642  -2.873357e-02 -1.032512e-01       0       0
#> 643  -1.286596e-01 -2.464695e-02       0       0
#> 644  -1.185726e-01  1.615228e-01       0       0
#> 645  -1.686721e-01  3.320970e-02       0       0
#> 646   6.012099e-02 -1.173622e-01       0       0
#> 647  -2.711409e-01  1.146531e-01       0       0
#> 648  -1.938214e-01  1.185755e-01       0       0
#> 649   1.569758e-02 -7.089878e-02       0       0
#> 650  -2.256357e-01  1.682898e-01       0       0
#> 651  -1.818529e-01  2.061114e-01       0       0
#> 652  -4.639254e-03  1.170791e-02       0       0
#> 653  -8.718381e-03  2.020067e-01       0       0
#> 654  -3.592946e-01  1.469678e-01       0       0
#> 655  -4.469508e-02 -1.179898e-01       0       0
#> 656   1.796520e-02 -2.058493e-03       0       0
#> 657  -3.345611e-02 -6.613252e-03       0       0
#> 658  -1.870174e-01 -7.249358e-02       0       0
#> 659  -5.936760e-02  2.408218e-01       0       0
#> 660  -2.227405e-01  2.706176e-01       0       0
#> 661  -1.530554e-01  1.410048e-01       0       0
#> 662  -1.021400e-01  1.268930e-01       0       0
#> 663  -1.762429e-01  9.638010e-02       0       0
#> 664  -1.089485e-01  1.841881e-01       0       0
#> 665  -2.282250e-01  2.568921e-01       0       0
#> 666  -4.268741e-02 -3.608714e-02       0       0
#> 667   4.475831e-02  1.755019e-02       0       0
#> 668  -6.139575e-02  9.939683e-02       0       0
#> 669  -7.227441e-02  4.498940e-02       0       0
#> 670  -1.556414e-01  2.174508e-01       0       0
#> 671  -2.499526e-01  2.950748e-01       0       0
#> 672  -1.463569e-01  2.665665e-01       0       0
#> 673  -1.242187e-01  1.614201e-01       0       0
#> 674  -1.008001e-01  7.353856e-02       0       0
#> 675  -1.058977e-01  7.051578e-02       0       0
#> 676  -1.679292e-01  2.708903e-02       0       0
#> 677  -1.026264e-01 -1.015143e-01       0       0
#> 678  -1.675992e-01  1.131885e-01       0       0
#> 679  -4.213685e-01  4.425934e-02       0       0
#> 680  -1.483786e-01  3.424241e-01       0       0
#> 681  -1.728611e-01  8.591988e-02       0       0
#> 682  -1.042216e-01  7.910040e-02       0       0
#> 683  -5.848216e-02  2.603107e-01       0       0
#> 684  -3.267092e-02 -1.741715e-01       0       0
#> 685  -4.613908e-02  8.249333e-02       0       0
#> 686  -1.481294e-01  1.699582e-01       0       0
#> 687  -9.761440e-02  8.357669e-03       0       0
#> 688  -4.192576e-02  5.194969e-02       0       0
#> 689  -6.568412e-02 -8.878597e-02       0       0
#> 690  -1.055435e-01 -4.014110e-02       0       0
#> 691  -9.992036e-02 -1.315029e-02       0       0
#> 692  -2.499132e-02  7.645891e-02       0       0
#> 693  -5.311574e-02  3.942543e-02       0       0
#> 694  -4.023503e-02  8.110438e-02       0       0
#> 695  -1.893163e-02 -4.144688e-02       0       0
#> 696  -1.066614e-01 -2.478033e-01       0       0
#> 697  -2.889663e-01  8.897235e-02       0       0
#> 698  -5.685510e-02  8.693814e-02       0       0
#> 699  -3.751744e-01  2.091340e-01       0       0
#> 700  -2.896438e-02  9.918740e-02       0       0
#> 701  -1.848125e-01  1.986438e-01       0       0
#> 702  -5.949125e-02  4.497099e-02       0       0
#> 703   4.925701e-02  2.546101e-03       0       0
#> 704  -1.134132e-01 -3.610224e-02       0       0
#> 705  -1.128360e-01  1.206974e-01       0       0
#> 706  -5.150038e-02 -1.374019e-01       0       0
#> 707   4.206118e-02 -7.378384e-02       0       0
#> 708   5.690251e-02 -8.058963e-02       0       0
#> 709  -1.237013e-01  8.392337e-02       0       0
#> 710  -9.998535e-02  1.379242e-01       0       0
#> 711  -6.718578e-02  1.111895e-01       0       0
#> 712  -9.269048e-02  2.037047e-01       0       0
#> 713  -1.167371e-01  1.163578e-01       0       0
#> 714   7.068138e-04  1.942238e-02       0       0
#> 715  -1.738283e-01  8.103926e-02       0       0
#> 716  -2.006780e-01  1.421737e-01       0       0
#> 717  -1.948789e-01 -5.812579e-02       0       0
#> 718  -3.020737e-01  6.480810e-02       0       0
#> 719  -1.616601e-01 -3.002691e-02       0       0
#> 720  -8.849635e-02 -1.812943e-02       0       0
#> 721  -4.340078e-01  6.957839e-03       0       0
#> 722  -4.410620e-02 -2.193945e-02       0       0
#> 723  -1.668026e-01  1.626805e-01       0       0
#> 724  -1.118338e-02  2.766260e-02       0       0
#> 725  -1.588599e-01  2.732897e-01       0       0
#> 726  -5.367948e-02  1.683756e-01       0       0
#> 727  -2.038048e-02  3.472770e-02       0       0
#> 728  -9.071455e-02  1.668103e-01       0       0
#> 729  -9.207219e-02 -2.164220e-02       0       0
#> 730  -1.588599e-01  2.732897e-01       0       0
#> 731  -1.500266e-01  4.598007e-02       0       0
#> 732  -1.831452e-01  2.036167e-01       0       0
#> 733  -1.214355e-01  2.039815e-01       0       0
#> 734  -8.395514e-02  1.517619e-01       0       0
#> 735  -5.136980e-02 -4.028892e-02       0       0
#> 736  -8.814915e-02  1.604808e-01       0       0
#> 737  -1.970791e-01  3.141614e-02       0       0
#> 738  -1.852941e-01  8.529734e-02       0       0
#> 739  -5.223420e-02  1.406037e-02       0       0
#> 740  -6.101348e-02 -1.532407e-02       0       0
#> 741  -1.824691e-01  1.059232e-02       0       0
#> 742  -3.498609e-02 -3.562299e-02       0       0
#> 743  -1.938214e-01  1.185755e-01       0       0
#> 744  -6.274133e-02 -5.569271e-02       0       0
#> 745  -2.780351e-01  9.822389e-02       0       0
#> 746  -1.215784e-01  9.749051e-02       0       0
#> 747  -2.691501e-02 -1.747433e-01       0       0
#> 748  -1.418468e-01  5.899018e-02       0       0
#> 749  -7.223548e-02  3.405826e-02       0       0
#> 750  -3.686967e-02 -4.141720e-02       0       0
#> 751  -3.885123e-02  9.655857e-02       0       0
#> 752  -8.111015e-02  7.935579e-02       0       0
#> 753  -1.275337e-01  2.046951e-01       0       0
#> 754  -9.021986e-02 -4.470820e-02       0       0
#> 755  -5.278616e-02 -1.224689e-02       0       0
#> 756  -2.153431e-01  1.114859e-01       0       0
#> 757  -6.779178e-02  8.267082e-02       0       0
#> 758  -6.804746e-02  1.869455e-02       0       0
#> 759  -6.510986e-02  6.325589e-02       0       0
#> 760  -6.309267e-02  8.212747e-02       0       0
#> 761  -1.157532e-01  2.205349e-02       0       0
#> 762  -4.170039e-02  6.628811e-02       0       0
#> 763  -1.135794e-01  2.023293e-02       0       0
#> 764  -7.112538e-02  1.614306e-01       0       0
#> 765  -7.876909e-02  1.962729e-01       0       0
#> 766  -4.992901e-02 -1.322421e-02       0       0
#> 767  -1.030940e-01  2.661686e-02       0       0
#> 768  -8.523044e-02  1.659878e-01       0       0
#> 769  -4.544877e-02  5.558812e-02       0       0
#> 770  -2.976451e-02  1.764411e-02       0       0
#> 771  -9.741528e-02  1.082069e-01       0       0
#> 772  -1.604218e-01 -9.650855e-02       0       0
#> 773   3.349587e-02 -4.693169e-02       0       0
#> 774  -2.132141e-02  2.551731e-01       0       0
#> 775  -8.596171e-02  1.052072e-01       0       0
#> 776  -1.808645e-01  2.001230e-02       0       0
#> 777  -1.869153e-01  9.955023e-02       0       0
#> 778  -1.990941e-02 -8.684683e-02       0       0
#> 779  -1.668616e-01  5.034207e-02       0       0
#> 780  -1.346793e-01  2.349721e-02       0       0
#> 781  -9.746554e-02 -4.211423e-02       0       0
#> 782  -2.036621e-01  1.744876e-01       0       0
#> 783  -1.274482e-01  1.910210e-01       0       0
#> 784  -7.210795e-02  7.835445e-02       0       0
#> 785  -6.847404e-03 -2.061933e-02       0       0
#> 786  -2.518593e-02  9.347081e-02       0       0
#> 787  -3.167971e-02  2.996948e-02       0       0
#> 788  -1.448558e-01  7.453483e-03       0       0
#> 789  -9.447555e-02  4.866249e-02       0       0
#> 790  -3.053550e-01  9.577169e-02       0       0
#> 791  -6.548081e-02 -2.183665e-02       0       0
#> 792  -1.855756e-01  6.812820e-02       0       0
#> 793  -4.632419e-02  1.790053e-01       0       0
#> 794  -2.417537e-02  5.430883e-02       0       0
#> 795  -1.033998e-01  1.658852e-01       0       0
#> 796  -1.481294e-01  1.699582e-01       0       0
#> 797  -4.666960e-02  8.085629e-02       0       0
#> 798  -1.929988e-01  1.458241e-01       0       0
#> 799   1.258693e-02 -1.982937e-01       0       0
#> 800  -4.037146e-02 -4.247154e-02       0       0
#> 801  -1.308331e-01  2.703716e-02       0       0
#> 802  -8.278323e-02  3.222283e-02       0       0
#> 803  -1.507993e-01  1.029303e-01       0       0
#> 804  -2.316798e-02 -1.047958e-01       0       0
#> 805   4.900119e-03  2.267548e-02       0       0
#> 806  -3.953004e-02  6.593920e-02       0       0
#> 807  -5.616317e-02  7.451453e-02       0       0
#> 808  -1.157532e-01  2.205349e-02       0       0
#> 809  -1.704165e-01  1.183374e-01       0       0
#> 810  -1.442584e-01  9.451905e-02       0       0
#> 811   5.681135e-02  5.636399e-03       0       0
#> 812  -2.744365e-01 -1.912454e-02       0       0
#> 813  -2.288726e-01 -7.706264e-02       0       0
#> 814  -1.686721e-01  3.320970e-02       0       0
#> 815  -4.257221e-01  6.752488e-02       0       0
#> 816  -1.061286e-01 -5.120204e-03       0       0
#> 817  -2.411226e-01  1.554679e-01       0       0
#> 818  -3.287940e-01  1.174124e-01       0       0
#> 819  -4.509830e-02 -3.023431e-03       0       0
#> 820  -4.533629e-02  4.540416e-02       0       0
#> 821  -7.124468e-02  2.938240e-02       0       0
#> 822  -7.814413e-02  1.838305e-02       0       0
#> 823  -1.616411e-01  9.910692e-02       0       0
#> 824  -2.863016e-02 -9.889053e-02       0       0
#> 825  -1.093062e-01  6.706099e-02       0       0
#> 826  -5.191814e-02 -7.835181e-02       0       0
#> 827  -5.667219e-02  3.652389e-02       0       0
#> 828  -2.242935e-01 -1.742842e-01       0       0
#> 829  -4.376206e-02 -5.509117e-03       0       0
#> 830  -1.332791e-01  9.248465e-02       0       0
#> 831  -8.338815e-02  1.095392e-01       0       0
#> 832  -2.871905e-01  1.613681e-01       0       0
#> 833  -3.364987e-01  1.434695e-01       0       0
#> 834  -2.218348e-01  1.974216e-01       0       0
#> 835  -8.593583e-02 -8.211966e-02       0       0
#> 836  -1.325940e-01  4.509471e-02       0       0
#> 837  -8.728867e-02 -2.349857e-02       0       0
#> 838  -3.097073e-02 -2.358816e-01       0       0
#> 839  -1.442860e-01  8.912458e-02       0       0
#> 840  -1.231958e-01  1.815153e-01       0       0
#> 841  -8.596171e-02  1.052072e-01       0       0
#> 842  -8.371092e-02  1.196294e-01       0       0
#> 843  -2.604950e-01  1.327321e-02       0       0
#> 844  -3.130749e-01  1.632014e-01       0       0
#> 845  -3.494651e-02  6.902154e-02       0       0
#> 846  -1.449826e-01  3.469586e-02       0       0
#> 847  -4.795412e-02  1.353079e-02       0       0
#> 848   2.265998e-02  1.755738e-01       0       0
#> 849  -1.061286e-01 -5.120204e-03       0       0
#> 850  -2.038526e-01  1.756275e-01       0       0
#> 851  -3.959397e-01  1.125984e-01       0       0
#> 852  -9.599202e-02  1.189929e-01       0       0
#> 853  -7.922124e-02 -1.130084e-01       0       0
#> 854  -1.921184e-02 -2.883940e-02       0       0
#> 855  -1.401875e-01  3.891683e-02       0       0
#> 856  -1.099098e-03 -1.666637e-01       0       0
#> 857  -5.929716e-02  1.127732e-01       0       0
#> 858  -1.089485e-01  1.841881e-01       0       0
#> 859  -5.216326e-03  4.806331e-02       0       0
#> 860   3.843561e-03 -6.170726e-02       0       0
#> 861  -8.202783e-02  6.382774e-02       0       0
#> 862  -2.411700e-01 -2.256454e-02       0       0
#> 863  -2.721435e-01  5.649424e-02       0       0
#> 864  -1.465363e-01  1.087973e-01       0       0
#> 865  -1.073143e-01  1.744416e-01       0       0
#> 866  -1.492733e-01  6.273898e-02       0       0
#> 867  -5.835848e-02 -2.510117e-02       0       0
#> 868  -2.218348e-01  1.974216e-01       0       0
#> 869  -1.411353e-01 -3.375989e-02       0       0
#> 870   1.062603e-02 -4.120657e-02       0       0
#> 871  -2.030898e-01  9.276452e-02       0       0
#> 872  -1.006122e-01  1.188250e-01       0       0
#> 873  -1.131556e-01  3.122127e-02       0       0
#> 874  -1.776677e-01  2.271070e-02       0       0
#> 875  -1.242187e-01  1.614201e-01       0       0
#> 876  -6.847404e-03 -2.061933e-02       0       0
#> 877  -1.042956e-01  4.891962e-02       0       0
#> 878  -1.754475e-01  7.801359e-02       0       0
#> 879  -1.032203e-01  2.399076e-01       0       0
#> 880  -1.034358e-01  1.702726e-01       0       0
#> 881  -3.039212e-01  2.065712e-01       0       0
#> 882   5.163260e-02 -9.097254e-02       0       0
#> 883  -1.818529e-01  2.061114e-01       0       0
#> 884  -4.155533e-02  2.872316e-02       0       0
#> 885  -2.103905e-01  4.085107e-02       0       0
#> 886  -1.619376e-01  1.183271e-02       0       0
#> 887  -7.470169e-03  4.553401e-02       0       0
#> 888  -8.543725e-02  1.356228e-02       0       0
#> 889  -8.642255e-02 -7.636392e-02       0       0
#> 890  -8.755555e-02  2.031563e-01       0       0
#> 891  -3.823884e-01 -3.673146e-02       0       0
#> 892  -2.118336e-01  5.656121e-02       0       0
#> 893   6.819389e-03  2.426579e-03       0       0
#> 894  -2.994358e-01  5.059713e-02       0       0
#> 895  -1.325228e-01  1.770561e-01       0       0
#> 896  -2.417537e-02  5.430883e-02       0       0
#> 897  -4.543383e-02 -1.785380e-02       0       0
#> 898  -1.020133e-01  1.686201e-02       0       0
#> 899  -1.178043e-01  9.223419e-02       0       0
#> 900  -1.491736e-01  3.821676e-03       0       0
#> 901  -2.305222e-01  1.520354e-01       0       0
#> 902   5.349228e-02 -7.565167e-03       0       0
#> 903  -9.696620e-02  1.404324e-01       0       0
#> 904  -1.436012e-01 -5.851633e-02       0       0
#> 905  -2.138266e-01  1.230833e-01       0       0
#> 906  -1.664073e-01  8.990038e-02       0       0
#> 907  -1.791499e-01  3.621007e-02       0       0
#> 908  -2.137171e-01 -7.727445e-02       0       0
#> 909  -8.455745e-02  3.332612e-01       0       0
#> 910   3.119803e-02 -3.747380e-02       0       0
#> 911  -7.993105e-02  6.882642e-02       0       0
#> 912  -3.075022e-01  2.830544e-01       0       0
#> 913  -2.004582e-01  1.427084e-01       0       0
#> 914  -2.533419e-02 -1.375024e-02       0       0
#> 915  -1.341323e-01  1.293839e-01       0       0
#> 916  -1.901662e-01  1.857738e-01       0       0
#> 917  -1.241231e-01  2.454663e-02       0       0
#> 918  -1.844001e-01  1.366542e-01       0       0
#> 919  -1.646823e-01  1.106045e-01       0       0
#> 920  -6.937522e-02 -1.379389e-01       0       0
#> 921  -1.215784e-01  9.749051e-02       0       0
#> 922  -2.246054e-01  4.351603e-03       0       0
#> 923  -1.917852e-01  8.585248e-02       0       0
#> 924  -9.598758e-02  2.131723e-01       0       0
#> 925  -1.364156e-01  1.215158e-01       0       0
#> 926  -6.713391e-02 -1.933005e-01       0       0
#> 927  -1.100328e-01  1.627103e-01       0       0
#> 928  -3.175733e-01  1.629016e-01       0       0
#> 929  -5.800739e-02 -3.152346e-01       0       0
#> 930  -1.810031e-01 -7.424177e-02       0       0
#> 931  -1.445126e-01 -9.697841e-02       0       0
#> 932  -1.505202e-01  1.468555e-01       0       0
#> 933  -4.121646e-02  8.002962e-02       0       0
#> 934  -1.969556e-01  9.132332e-02       0       0
#> 935  -1.040100e-01  1.038796e-01       0       0
#> 936  -3.130749e-01  1.632014e-01       0       0
#> 937   4.791838e-03 -3.970603e-02       0       0
#> 938  -2.358249e-01  8.891268e-02       0       0
#> 939  -3.282178e-01  5.840315e-02       0       0
#> 940  -1.944703e-01 -2.264160e-02       0       0
#> 941  -2.769948e-01  2.947201e-01       0       0
#> 942  -2.099400e-01  1.244996e-02       0       0
#> 943  -1.687698e-01 -5.383990e-02       0       0
#> 944   5.169070e-02 -5.698880e-02       0       0
#> 945  -6.920081e-02 -1.613956e-02       0       0
#> 946  -6.843266e-02  1.580943e-02       0       0
#> 947  -1.467392e-01  2.937479e-01       0       0
#> 948  -9.870480e-02  6.434544e-02       0       0
#> 949  -4.312929e-01  3.384632e-01       0       0
#> 950   2.807348e-02  9.831828e-02       0       0
#> 951  -9.155075e-02  1.609525e-01       0       0
#> 952  -2.675582e-02 -4.536508e-02       0       0
#> 953  -2.871905e-01  1.613681e-01       0       0
#> 954  -8.622448e-02  1.409788e-01       0       0
#> 955  -1.684275e-01  1.865931e-01       0       0
#> 956  -9.696620e-02  1.404324e-01       0       0
#> 957  -1.181871e-01  1.242263e-01       0       0
#> 958  -1.550917e-01  1.022963e-01       0       0
#> 959  -2.392310e-02 -1.206077e-01       0       0
#> 960  -1.518736e-02 -4.548844e-02       0       0
#> 961  -1.848125e-01  1.986438e-01       0       0
#> 962  -5.929716e-02  1.127732e-01       0       0
#> 963  -1.252248e-01  1.494051e-01       0       0
#> 964  -9.969644e-02  1.558105e-01       0       0
#> 965  -2.905101e-01  9.727500e-02       0       0
#> 966  -1.052774e-01  1.290595e-02       0       0
#> 967  -2.031159e-01  6.613721e-02       0       0
#> 968  -1.270985e-01  9.154490e-02       0       0
#> 969  -2.227405e-01  2.706176e-01       0       0
#> 970  -1.980745e-01 -1.032387e-02       0       0
#> 971  -1.475056e-01  1.627193e-01       0       0
#> 972  -1.559032e-01 -4.328052e-02       0       0
#> 973   3.415888e-02  1.845152e-01       0       0
#> 974  -9.019089e-02  2.634717e-02       0       0
#> 975   6.320482e-02  9.436021e-03       0       0
#> 976  -9.070695e-02 -1.019052e-01       0       0
#> 977  -1.001106e-01  2.651514e-01       0       0
#> 978  -2.068264e-01  4.425988e-02       0       0
#> 979  -2.813388e-01  1.725624e-01       0       0
#> 980  -7.928834e-03  1.644424e-02       0       0
#> 981  -1.051116e-01 -7.659274e-02       0       0
#> 982  -2.617642e-01  1.491048e-01       0       0
#> 983  -8.154603e-02  3.507572e-02       0       0
#> 984  -2.703134e-02  9.188755e-02       0       0
#> 985  -1.027640e-01 -1.012985e-02       0       0
#> 986   4.625862e-03  1.006139e-01       0       0
#> 987  -1.128360e-01  1.206974e-01       0       0
#> 988  -2.130434e-01  1.102313e-02       0       0
#> 989  -7.598298e-02 -4.350868e-02       0       0
#> 990  -5.936760e-02  2.408218e-01       0       0
#> 991  -1.489757e-02 -6.599586e-02       0       0
#> 992  -2.237024e-02 -8.309359e-04       0       0
#> 993  -2.218348e-01  1.974216e-01       0       0
#> 994  -2.006780e-01  1.421737e-01       0       0
#> 995  -1.215784e-01  9.749051e-02       0       0
#> 996  -1.145894e-01  1.068481e-01       0       0
#> 997  -1.036682e-01 -1.775136e-02       0       0
#> 998  -1.074979e-01  1.074805e-01       0       0
#> 999  -7.020654e-02  1.155478e-01       0       0
#> 1000 -1.901611e-01  1.563009e-01       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.107   0.223  
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.125  -0.00688
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.308   0.283  
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0245 -0.00310
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.0888 -0.0728 
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.0154 -0.188  
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.121   0.125  
#>  8      8     0.21      80    65    19.8     0.041      0.12  0.110  -0.00115
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.0281 -0.00735
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.155   0.0366 
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.224  24.3 0.041  0.12 67.6  1643.
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.224  24.3 0.041  0.12  4.55  111.
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.220  19.3 0.041  0.12 85.3  1648.
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.220  19.3 0.041  0.12  5.96  115.
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.183  25.8 0.041  0.12 65.6  1694.
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.183  25.8 0.041  0.12  6.63  171.
#> # … with 1,994 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>