Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  dat = NULL,
  prior_model = NULL,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE,
  nocb = FALSE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the posterior distribution of population parameters
poso_estim_sir(dat=df_patient01,prior_model=mod_run001,
n_sample=1e3,n_resample=1e2)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.44555758 -0.638029371 -0.252370951
#> 2    0.31427838 -1.022476424 -0.273371778
#> 3    0.75616151 -0.227202561  0.078651563
#> 4    0.62869564 -0.246195806  0.247040886
#> 5    0.27467552 -0.101391204  0.915892955
#> 6    0.84756475 -0.140006769  1.124143889
#> 7    0.40516914 -0.919011681 -0.586216223
#> 8    0.69725407 -0.343789583  0.531142218
#> 9    0.23951516 -0.928924825  0.190472069
#> 10   0.67164235 -0.354344320  0.261156265
#> 11   0.11208804 -0.135407416  0.533279120
#> 12   0.50406455 -0.260902600 -0.039643789
#> 13   0.69725407 -0.343789583  0.531142218
#> 14   0.44555758 -0.638029371 -0.252370951
#> 15   0.31427838 -1.022476424 -0.273371778
#> 16   0.37371531 -0.600756537  0.348050818
#> 17   0.31257192 -0.602710420  0.079132686
#> 18   0.79755929 -0.197956946  0.151303623
#> 19   0.19293922 -0.292887095  0.230681445
#> 20   0.59779200 -0.341274984  0.449633213
#> 21   0.12735792 -0.689889066 -0.168604640
#> 22   0.56920175 -0.323016503  0.340778348
#> 23   0.67417637  0.060130062  0.699232410
#> 24   0.67417637  0.060130062  0.699232410
#> 25   0.60257045  0.115956530  0.617204809
#> 26   0.23269471 -0.580584150  0.219230234
#> 27   0.44715596 -0.709281583 -0.211076164
#> 28   0.44715596 -0.709281583 -0.211076164
#> 29   0.51816834 -0.349180120 -0.023301441
#> 30   0.59779200 -0.341274984  0.449633213
#> 31   0.26556629 -0.176054130  0.260497231
#> 32   0.67164235 -0.354344320  0.261156265
#> 33   0.75447761 -0.220901886  0.594642366
#> 34   0.67164235 -0.354344320  0.261156265
#> 35   0.40516914 -0.919011681 -0.586216223
#> 36   0.44715596 -0.709281583 -0.211076164
#> 37   0.84756475 -0.140006769  1.124143889
#> 38   0.60542585 -0.551115774 -0.431379506
#> 39   0.49860435 -0.386010316  1.061122923
#> 40   0.67417637  0.060130062  0.699232410
#> 41   0.44555758 -0.638029371 -0.252370951
#> 42   0.74837512 -0.269621248  0.584721950
#> 43   0.74837512 -0.269621248  0.584721950
#> 44   0.60542585 -0.551115774 -0.431379506
#> 45   0.75447761 -0.220901886  0.594642366
#> 46   0.60232645 -0.682342815 -0.287542321
#> 47   0.31427838 -1.022476424 -0.273371778
#> 48   0.36669493 -0.116700210  0.335933527
#> 49   0.44555758 -0.638029371 -0.252370951
#> 50   0.17161006 -0.904248026 -0.311020905
#> 51   0.56920175 -0.323016503  0.340778348
#> 52   0.69725407 -0.343789583  0.531142218
#> 53  -0.27017224 -0.159589077  0.540742311
#> 54   0.67164235 -0.354344320  0.261156265
#> 55   0.44715596 -0.709281583 -0.211076164
#> 56   0.75666356 -0.241081641  0.270829842
#> 57   0.67164235 -0.354344320  0.261156265
#> 58   0.74837512 -0.269621248  0.584721950
#> 59   0.44715596 -0.709281583 -0.211076164
#> 60   0.44715596 -0.709281583 -0.211076164
#> 61   0.67164235 -0.354344320  0.261156265
#> 62   0.31427838 -1.022476424 -0.273371778
#> 63   0.44555758 -0.638029371 -0.252370951
#> 64   0.63731078 -0.692742023 -0.699678759
#> 65   0.44715596 -0.709281583 -0.211076164
#> 66   0.60232645 -0.682342815 -0.287542321
#> 67   0.22821000  0.009984240  0.825529759
#> 68   0.60542585 -0.551115774 -0.431379506
#> 69   0.20267491 -0.228460061 -0.034576686
#> 70   0.59779200 -0.341274984  0.449633213
#> 71   0.45722696 -0.299266246  0.705954245
#> 72   0.75447761 -0.220901886  0.594642366
#> 73   0.75666356 -0.241081641  0.270829842
#> 74   0.24151639 -0.344051805  0.065825507
#> 75   0.44555758 -0.638029371 -0.252370951
#> 76   0.40516914 -0.919011681 -0.586216223
#> 77   0.75666356 -0.241081641  0.270829842
#> 78   0.51816834 -0.349180120 -0.023301441
#> 79   0.15367941  0.036775151  0.850349657
#> 80   0.44555758 -0.638029371 -0.252370951
#> 81   0.45806784  0.008819992  0.362005757
#> 82   0.08917863 -0.344485168 -0.001867226
#> 83   0.45722696 -0.299266246  0.705954245
#> 84   0.75616151 -0.227202561  0.078651563
#> 85   0.60542585 -0.551115774 -0.431379506
#> 86   0.62869564 -0.246195806  0.247040886
#> 87   0.75666356 -0.241081641  0.270829842
#> 88   0.75616151 -0.227202561  0.078651563
#> 89   0.66378816 -0.035394913  0.286962417
#> 90   0.37649411  0.079756416  0.796414152
#> 91   0.61289647 -0.285536058  0.815748654
#> 92   0.75666356 -0.241081641  0.270829842
#> 93   0.62869564 -0.246195806  0.247040886
#> 94   0.69725407 -0.343789583  0.531142218
#> 95   0.60232645 -0.682342815 -0.287542321
#> 96   0.17161006 -0.904248026 -0.311020905
#> 97   0.49860435 -0.386010316  1.061122923
#> 98   0.66378816 -0.035394913  0.286962417
#> 99   0.26853749 -0.599145877  0.400155583
#> 100  0.44715596 -0.709281583 -0.211076164
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1  0.446 -0.638 -0.252 
#>  2      2        4       70        1  0.314 -1.02  -0.273 
#>  3      3        4       70        1  0.756 -0.227  0.0787
#>  4      4        4       70        1  0.629 -0.246  0.247 
#>  5      5        4       70        1  0.275 -0.101  0.916 
#>  6      6        4       70        1  0.848 -0.140  1.12  
#>  7      7        4       70        1  0.405 -0.919 -0.586 
#>  8      8        4       70        1  0.697 -0.344  0.531 
#>  9      9        4       70        1  0.240 -0.929  0.190 
#> 10     10        4       70        1  0.672 -0.354  0.261 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  6.25  37.0 0.777 0.169 22.1  1124.        816.
#> 2      1    14     4    70     1  6.25  37.0 0.777 0.169  6.78    0.0461    251.
#> 3      2     1     4    70     1  5.48  25.2 0.761 0.218 31.3  1137.        788.
#> 4      2    14     4    70     1  5.48  25.2 0.761 0.218  5.59    0.0576    141.
#> 5      3     1     4    70     1  8.52  55.8 1.08  0.153 18.5   899.       1030.
#> 6      3    14     4    70     1  8.52  55.8 1.08  0.153  5.11    0.000702  285.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>