Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 1e+05,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object
A posologyr list, created by the posologyr function.

n_sample
Number of samples from the S-step

n_resample
Number of samples from the R-step

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)

#> 

#> $eta
#>             ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1    -4.687052e-02  0.0445779148       0       0
#> 2    -1.081355e-01 -0.0072255066       0       0
#> 3     9.128406e-02 -0.0373683715       0       0
#> 4    -1.049687e-02 -0.0090446478       0       0
#> 5    -5.095032e-02 -0.0047193341       0       0
#> 6    -8.863511e-02  0.1267839992       0       0
#> 7    -2.647532e-02 -0.0304030662       0       0
#> 8    -2.030334e-01 -0.0178086300       0       0
#> 9    -1.804569e-01  0.1657057687       0       0
#> 10   -6.586703e-01 -0.0636403175       0       0
#> 11   -9.362816e-02 -0.0908893349       0       0
#> 12   -9.535222e-01  0.0554142645       0       0
#> 13   -1.211815e-01  0.0776872174       0       0
#> 14   -1.315368e-01  0.0018913706       0       0
#> 15   -1.098638e-01 -0.0117013059       0       0
#> 16   -1.127235e-01  0.1826198105       0       0
#> 17   -6.438494e-02  0.1284123492       0       0
#> 18   -5.720033e-02 -0.1716342984       0       0
#> 19   -4.658354e-02 -0.1821903217       0       0
#> 20   -1.595831e-01  0.2272813915       0       0
#> 21   -2.010465e-01  0.1798060940       0       0
#> 22   -1.482908e-02  0.0436603738       0       0
#> 23   -1.187654e-01  0.1036479316       0       0
#> 24   -1.264960e-01  0.1150474511       0       0
#> 25   -6.514603e-02 -0.0314242130       0       0
#> 26   -1.595300e-01  0.1329531414       0       0
#> 27   -1.906500e-01 -0.1795742235       0       0
#> 28   -1.227407e-01  0.2280852458       0       0
#> 29   -9.857315e-02  0.1158393432       0       0
#> 30   -1.420784e-01  0.2557766014       0       0
#> 31   -1.595831e-01  0.2272813915       0       0
#> 32   -5.673012e-02 -0.0816740288       0       0
#> 33   -1.016312e-02 -0.0521318480       0       0
#> 34   -4.877819e-02  0.0921650012       0       0
#> 35   -1.826666e-01  0.2797295360       0       0
#> 36   -4.414021e-02  0.0592087422       0       0
#> 37   -1.751639e-01  0.2071269755       0       0
#> 38   -3.071259e-01  0.1434073368       0       0
#> 39   -1.052901e-01  0.1225021091       0       0
#> 40   -7.865540e-02  0.0298569014       0       0
#> 41   -1.623202e-01  0.1215591831       0       0
#> 42   -6.406905e-02  0.0184049171       0       0
#> 43   -1.210732e-01  0.0058439129       0       0
#> 44   -1.123308e-01  0.1189820804       0       0
#> 45   -7.078004e-03  0.0308222962       0       0
#> 46   -1.146847e-01  0.0626596992       0       0
#> 47   -5.817045e-02  0.1976650665       0       0
#> 48   -4.119900e-02 -0.0643614803       0       0
#> 49   -2.617096e-01  0.1516499962       0       0
#> 50   -1.401703e-01 -0.0550635188       0       0
#> 51   -1.953972e-01  0.2368954494       0       0
#> 52   -3.070370e-02 -0.0432675200       0       0
#> 53   -2.680535e-01  0.1695033669       0       0
#> 54    5.700036e-02 -0.0859397357       0       0
#> 55   -9.527178e-03  0.0089911689       0       0
#> 56   -1.436053e-01 -0.0554267910       0       0
#> 57   -9.883387e-02 -0.0334311083       0       0
#> 58   -4.324182e-02  0.2437970076       0       0
#> 59   -2.329375e-01  0.1949050979       0       0
#> 60   -2.612762e-01  0.1061798266       0       0
#> 61   -3.229324e-04  0.1508661718       0       0
#> 62   -3.206577e-01  0.2480272398       0       0
#> 63   -5.786396e-03  0.0241865389       0       0
#> 64   -3.473892e-01  0.2449577759       0       0
#> 65   -1.392440e-01  0.1922859703       0       0
#> 66   -1.667048e-01  0.3402289678       0       0
#> 67   -2.141899e-01  0.1464551324       0       0
#> 68    8.717620e-03 -0.0093311388       0       0
#> 69    5.470977e-02  0.0265545629       0       0
#> 70   -1.505534e-01 -0.0193491682       0       0
#> 71   -1.006943e-01 -0.0580607821       0       0
#> 72   -1.596236e-01  0.1157547550       0       0
#> 73   -1.667056e-01  0.1541181179       0       0
#> 74   -1.908141e-01  0.0547165677       0       0
#> 75   -1.135055e-01  0.1334465478       0       0
#> 76   -3.987090e-02  0.0227142388       0       0
#> 77   -9.649502e-02  0.0712373397       0       0
#> 78   -9.310809e-02  0.2114322190       0       0
#> 79   -1.123842e-01  0.0245510447       0       0
#> 80   -1.361448e-01  0.1675534803       0       0
#> 81   -3.687292e-01  0.2704966860       0       0
#> 82    1.300872e-02 -0.0818628756       0       0
#> 83   -5.467305e-02  0.0085890734       0       0
#> 84   -2.502200e-01  0.1291042436       0       0
#> 85   -1.482986e-02  0.1405057675       0       0
#> 86   -2.387978e-02  0.1074415264       0       0
#> 87   -3.245644e-02 -0.0094036084       0       0
#> 88   -7.549225e-02  0.1395424753       0       0
#> 89   -2.068517e-01  0.0856906192       0       0
#> 90   -2.045118e-01 -0.0587571996       0       0
#> 91   -1.970650e-01  0.3518804365       0       0
#> 92   -4.208874e-02  0.0758708265       0       0
#> 93   -1.254078e-01  0.2140500873       0       0
#> 94   -6.896198e-02  0.1432787295       0       0
#> 95   -2.510816e-01  0.2529076819       0       0
#> 96   -4.757481e-02  0.1284218685       0       0
#> 97   -3.215360e-01  0.2311871997       0       0
#> 98   -3.839253e-02  0.0123555925       0       0
#> 99   -1.088299e-01  0.1236684432       0       0
#> 100  -5.117641e-02 -0.0806635066       0       0
#> 101  -1.822389e-01  0.0652275615       0       0
#> 102  -6.811301e-02  0.1533693019       0       0
#> 103  -7.040223e-02 -0.0372801461       0       0
#> 104   4.434884e-02 -0.1020358644       0       0
#> 105  -4.166006e-02 -0.1873098831       0       0
#> 106  -1.061620e-01  0.0782305537       0       0
#> 107  -5.171921e-02 -0.0362029556       0       0
#> 108  -8.442082e-02  0.0552702897       0       0
#> 109  -7.677550e-02  0.0792618983       0       0
#> 110  -2.204779e-01  0.1019860204       0       0
#> 111   1.062295e-01 -0.1585424950       0       0
#> 112  -5.306161e-02 -0.0711742097       0       0
#> 113  -5.702232e-02 -0.0732465644       0       0
#> 114  -1.564113e-01  0.0532556334       0       0
#> 115  -4.213272e-02  0.0803212371       0       0
#> 116  -1.323463e-01  0.2377324553       0       0
#> 117  -8.436779e-02  0.1028876685       0       0
#> 118   3.714572e-02  0.1073090158       0       0
#> 119   5.527519e-02  0.0918161835       0       0
#> 120  -1.312914e-01  0.0857085936       0       0
#> 121  -1.650662e-01  0.2202137850       0       0
#> 122  -1.756404e-01  0.0543280465       0       0
#> 123  -2.421605e-01  0.0726626747       0       0
#> 124  -1.281279e-01  0.1373731706       0       0
#> 125  -2.125240e-01 -0.1126389118       0       0
#> 126  -6.451023e-02  0.0795680826       0       0
#> 127  -3.925658e-02 -0.0164328782       0       0
#> 128  -9.641279e-02 -0.0043720724       0       0
#> 129  -6.909374e-02 -0.0558340179       0       0
#> 130  -2.754658e-01  0.0955841301       0       0
#> 131  -1.269129e-01 -0.0366909212       0       0
#> 132  -2.635147e-02  0.0149881098       0       0
#> 133  -8.374020e-02  0.0682057859       0       0
#> 134  -1.952417e-01  0.0611661385       0       0
#> 135  -4.542161e-02  0.1133157614       0       0
#> 136  -1.682242e-01  0.0975619960       0       0
#> 137  -1.812760e-01  0.0848824401       0       0
#> 138  -1.083677e-01  0.0282858446       0       0
#> 139  -1.401844e-01 -0.1074056767       0       0
#> 140  -8.514915e-02 -0.0753898333       0       0
#> 141  -4.042230e-02  0.0086657961       0       0
#> 142  -9.572207e-02 -0.0618934628       0       0
#> 143  -6.565315e-02  0.0926128430       0       0
#> 144  -9.283323e-02 -0.1605301911       0       0
#> 145  -2.007226e-01  0.0331341141       0       0
#> 146  -1.140303e-01  0.0936460583       0       0
#> 147  -2.723674e-01  0.0654434573       0       0
#> 148  -1.639493e-01  0.1284391258       0       0
#> 149  -6.472673e-02 -0.0563374690       0       0
#> 150  -1.080046e-01  0.0496193554       0       0
#> 151  -4.976454e-02  0.0118940049       0       0
#> 152  -2.580821e-02 -0.0898365490       0       0
#> 153  -1.187010e-01  0.0784687582       0       0
#> 154  -1.967914e-01  0.2443585825       0       0
#> 155  -9.389834e-02  0.0849047142       0       0
#> 156  -1.117660e-01  0.1094153119       0       0
#> 157  -7.202080e-02 -0.0742017500       0       0
#> 158  -9.151365e-02 -0.0097211915       0       0
#> 159  -8.867951e-03  0.0227478585       0       0
#> 160  -2.328303e-02  0.1446094270       0       0
#> 161   1.420542e-02 -0.0794379249       0       0
#> 162  -5.091988e-02  0.0646365990       0       0
#> 163  -3.878372e-02  0.0198079179       0       0
#> 164  -5.164275e-02 -0.0701910508       0       0
#> 165  -1.100364e-01  0.0842107270       0       0
#> 166  -1.602350e-01  0.1579125690       0       0
#> 167  -1.001616e-02 -0.1906124473       0       0
#> 168  -1.032514e-01 -0.0506309989       0       0
#> 169   7.441706e-02 -0.0178899021       0       0
#> 170  -6.528662e-03 -0.0354875965       0       0
#> 171  -1.360423e-01  0.1474898755       0       0
#> 172  -8.464295e-02  0.1031866548       0       0
#> 173  -1.630663e-01  0.1339002325       0       0
#> 174  -1.802713e-01  0.0208693555       0       0
#> 175  -2.147831e-01  0.3482514405       0       0
#> 176  -9.838006e-02 -0.0303015235       0       0
#> 177  -4.396908e-04 -0.0178858900       0       0
#> 178  -3.215360e-01  0.2311871997       0       0
#> 179  -2.308649e-01  0.0858730615       0       0
#> 180  -9.104011e-02 -0.1136303983       0       0
#> 181  -6.207600e-02  0.0908934232       0       0
#> 182  -2.704210e-01  0.0534949071       0       0
#> 183  -1.382393e-01  0.2464604671       0       0
#> 184   1.085022e-02  0.0095371781       0       0
#> 185  -3.628556e-02  0.0607477815       0       0
#> 186  -2.723674e-01  0.0654434573       0       0
#> 187  -1.713437e-01  0.1547264019       0       0
#> 188  -2.058478e-01  0.0596442254       0       0
#> 189   2.104514e-02  0.0509953581       0       0
#> 190  -7.503216e-02  0.1363746642       0       0
#> 191  -1.648384e-01  0.0130447067       0       0
#> 192  -1.071029e-01 -0.0156230466       0       0
#> 193  -2.029097e-01  0.0620946640       0       0
#> 194  -3.742700e-01  0.2313247563       0       0
#> 195  -3.342204e-01  0.1464539398       0       0
#> 196  -2.628224e-01  0.1487552696       0       0
#> 197  -1.017525e-01  0.1454209395       0       0
#> 198  -4.460244e-02 -0.1266845645       0       0
#> 199  -4.102474e-02  0.2127753039       0       0
#> 200  -7.634468e-02  0.2209731181       0       0
#> 201  -1.552583e-01  0.1700917391       0       0
#> 202  -3.554174e-01  0.3058067977       0       0
#> 203  -1.850298e-01  0.1627446875       0       0
#> 204  -1.384417e-01  0.0930916311       0       0
#> 205  -2.242257e-01  0.0994402028       0       0
#> 206  -1.153877e-01  0.1810566500       0       0
#> 207  -5.331341e-02 -0.0376787069       0       0
#> 208  -1.078315e-01  0.0917326625       0       0
#> 209  -6.589192e-02  0.0659440338       0       0
#> 210  -3.794585e-02 -0.0982304133       0       0
#> 211  -2.449546e-01  0.0394623335       0       0
#> 212  -1.023685e-01  0.2511199177       0       0
#> 213  -4.310735e-02  0.0899708376       0       0
#> 214  -2.907321e-01  0.1172188738       0       0
#> 215  -2.951377e-02  0.0174582088       0       0
#> 216  -1.413947e-01  0.1593563341       0       0
#> 217  -1.436053e-01 -0.0554267910       0       0
#> 218   4.266536e-02  0.1228388582       0       0
#> 219  -1.073048e-01  0.1391745663       0       0
#> 220  -2.932003e-01  0.1656158875       0       0
#> 221  -3.723041e-02  0.1913137037       0       0
#> 222  -1.408721e-01  0.0146736741       0       0
#> 223  -2.689584e-01 -0.0480931834       0       0
#> 224  -1.440346e-01  0.3936407550       0       0
#> 225  -1.269670e-01 -0.0268752655       0       0
#> 226  -2.768306e-02 -0.1615691997       0       0
#> 227  -4.096015e-02 -0.0797239397       0       0
#> 228  -6.216254e-02  0.0961179888       0       0
#> 229   8.743561e-02 -0.1766824639       0       0
#> 230  -1.516057e-01  0.0631686584       0       0
#> 231  -1.401859e-01  0.1168152239       0       0
#> 232  -2.655550e-01  0.1395168729       0       0
#> 233  -7.647826e-02 -0.1271789513       0       0
#> 234  -1.826666e-01  0.2797295360       0       0
#> 235  -7.287838e-02 -0.1197365604       0       0
#> 236  -1.442204e-01  0.0187370295       0       0
#> 237  -1.181300e-01  0.1446007443       0       0
#> 238  -2.046311e-01 -0.0101853599       0       0
#> 239  -1.598728e-01  0.0757123284       0       0
#> 240  -7.384850e-02 -0.0087774489       0       0
#> 241  -2.199164e-01  0.1801719650       0       0
#> 242   1.935883e-02  0.0197217269       0       0
#> 243   3.484888e-02 -0.0062504466       0       0
#> 244  -7.706568e-02 -0.0243535821       0       0
#> 245  -1.495778e-01  0.2368909392       0       0
#> 246  -3.342204e-01  0.1464539398       0       0
#> 247  -1.517056e-01 -0.0223413330       0       0
#> 248  -1.921705e-01 -0.0218741671       0       0
#> 249  -8.436779e-02  0.1028876685       0       0
#> 250  -1.888769e-01  0.2405552486       0       0
#> 251  -1.156433e-01  0.1492963158       0       0
#> 252  -1.994217e-01  0.0940884957       0       0
#> 253   2.758269e-02  0.0545220263       0       0
#> 254  -1.407399e-01  0.0982956571       0       0
#> 255  -1.698331e-01  0.1155643008       0       0
#> 256  -1.614992e-01 -0.0108424087       0       0
#> 257  -3.366874e-01  0.3347038829       0       0
#> 258  -2.546581e-02  0.0978035285       0       0
#> 259  -8.841300e-02 -0.0193849327       0       0
#> 260  -1.595300e-01  0.1329531414       0       0
#> 261   1.204698e-04  0.0520910851       0       0
#> 262   1.850212e-02 -0.1344234316       0       0
#> 263  -2.489508e-01  0.0496180460       0       0
#> 264  -1.890590e-02  0.0569702583       0       0
#> 265  -1.496163e-01  0.0897764789       0       0
#> 266  -6.419154e-02 -0.0352939519       0       0
#> 267  -1.757677e-01  0.0031325842       0       0
#> 268  -1.515145e-01 -0.0321986046       0       0
#> 269  -1.506980e-01  0.1183426422       0       0
#> 270  -3.366874e-01  0.3347038829       0       0
#> 271  -8.440557e-02 -0.0228738662       0       0
#> 272  -1.254078e-01  0.2140500873       0       0
#> 273  -3.231858e-02 -0.0182917343       0       0
#> 274  -2.212011e-01  0.2737923210       0       0
#> 275   9.705832e-03 -0.0629030973       0       0
#> 276  -1.804569e-01  0.1657057687       0       0
#> 277  -9.045545e-02 -0.1273615136       0       0
#> 278  -7.743466e-02  0.1173933788       0       0
#> 279  -8.439382e-02  0.0872718196       0       0
#> 280  -1.068769e-01  0.1152068585       0       0
#> 281  -9.599855e-02  0.1058302924       0       0
#> 282   3.795145e-02  0.0232842133       0       0
#> 283  -1.740191e-01  0.1100428110       0       0
#> 284  -1.477613e-01  0.0050596094       0       0
#> 285   9.670398e-02  0.0844617695       0       0
#> 286   8.316033e-02 -0.0454251681       0       0
#> 287  -9.038570e-02  0.0159353474       0       0
#> 288  -1.101898e-01  0.1551447332       0       0
#> 289  -1.964249e-01  0.2751041636       0       0
#> 290  -7.748135e-02  0.0272500151       0       0
#> 291  -7.433320e-02 -0.0372288836       0       0
#> 292  -3.030839e-02 -0.0033271683       0       0
#> 293  -1.053333e-01  0.1077901602       0       0
#> 294  -1.879724e-01  0.1165720902       0       0
#> 295  -5.529301e-02  0.1654233684       0       0
#> 296  -5.607473e-02 -0.0726812590       0       0
#> 297  -1.578985e-02  0.0050322943       0       0
#> 298  -8.440557e-02 -0.0228738662       0       0
#> 299  -3.307693e-02  0.0332293237       0       0
#> 300  -1.131966e-01 -0.0185134679       0       0
#> 301  -3.944886e-02 -0.0274960191       0       0
#> 302   6.286846e-03 -0.0681483009       0       0
#> 303  -1.392440e-01  0.1922859703       0       0
#> 304  -7.647135e-02 -0.0692808010       0       0
#> 305  -6.903159e-02  0.1250207258       0       0
#> 306  -1.566673e-01  0.0015166634       0       0
#> 307  -3.321334e-02  0.0787636381       0       0
#> 308   1.786817e-02 -0.0639400265       0       0
#> 309  -6.167315e-02 -0.0903806251       0       0
#> 310  -2.028794e-01  0.0179550312       0       0
#> 311   1.931069e-02 -0.0556265353       0       0
#> 312  -1.240150e-01  0.0142087603       0       0
#> 313  -2.299043e-01  0.1707200713       0       0
#> 314   2.068423e-03 -0.0378261810       0       0
#> 315  -1.634090e-01  0.1314370689       0       0
#> 316  -1.635180e-01 -0.0956150266       0       0
#> 317  -1.749924e-01  0.2483017602       0       0
#> 318  -1.091775e-01  0.1739297908       0       0
#> 319  -1.452169e-01  0.0765634965       0       0
#> 320  -4.243696e-02  0.1847584340       0       0
#> 321  -6.590302e-02 -0.0010066728       0       0
#> 322  -3.018380e-02 -0.1603511052       0       0
#> 323   2.667032e-02 -0.0039108511       0       0
#> 324  -1.617532e-01  0.1085508414       0       0
#> 325  -1.509100e-01  0.1388199853       0       0
#> 326  -3.530534e-01  0.1689322566       0       0
#> 327  -1.633096e-01  0.0445776238       0       0
#> 328  -1.891204e-01  0.1223466804       0       0
#> 329  -3.594566e-02  0.0045614333       0       0
#> 330  -3.554174e-01  0.3058067977       0       0
#> 331  -5.534377e-02  0.2302993945       0       0
#> 332  -1.408721e-01  0.0146736741       0       0
#> 333  -2.020498e-01 -0.0893242329       0       0
#> 334  -1.347912e-01  0.1102450751       0       0
#> 335  -1.357231e-01  0.0415604182       0       0
#> 336  -2.141899e-01  0.1464551324       0       0
#> 337  -5.936925e-02  0.2259655797       0       0
#> 338  -1.452169e-01  0.0765634965       0       0
#> 339  -1.436091e-01  0.0812290096       0       0
#> 340  -2.908173e-01  0.2772469031       0       0
#> 341  -7.706568e-02 -0.0243535821       0       0
#> 342  -8.627820e-02 -0.0852295443       0       0
#> 343  -1.117660e-01  0.1094153119       0       0
#> 344  -1.254078e-01  0.2140500873       0       0
#> 345  -8.652537e-02  0.0706194717       0       0
#> 346  -5.427212e-02  0.1695928313       0       0
#> 347  -5.046300e-02 -0.0500663343       0       0
#> 348   3.706627e-02  0.1216707801       0       0
#> 349  -1.147226e-01  0.1766794816       0       0
#> 350  -2.890570e-02 -0.0515536295       0       0
#> 351  -1.512960e-03 -0.0754075635       0       0
#> 352   1.210976e-03 -0.0720932253       0       0
#> 353  -1.037534e-01  0.1476077456       0       0
#> 354  -2.204938e-01  0.1305238275       0       0
#> 355  -1.628087e-01  0.2777224869       0       0
#> 356  -5.699877e-02  0.0522793107       0       0
#> 357   4.272798e-02  0.0065949748       0       0
#> 358  -1.598564e-01  0.1659310479       0       0
#> 359  -9.059473e-02  0.0922218769       0       0
#> 360  -5.212775e-02 -0.1208610155       0       0
#> 361  -1.248694e-01  0.1339104458       0       0
#> 362  -1.073048e-01  0.1391745663       0       0
#> 363  -1.080046e-01  0.0496193554       0       0
#> 364  -1.020131e-01  0.0848410723       0       0
#> 365  -1.634090e-01  0.1314370689       0       0
#> 366  -3.079211e-01  0.0765716178       0       0
#> 367  -6.021115e-03  0.0241275235       0       0
#> 368  -7.566223e-02 -0.0249114126       0       0
#> 369  -1.127235e-01  0.1826198105       0       0
#> 370  -1.061367e-01  0.1681794064       0       0
#> 371  -1.195385e-01  0.0219394733       0       0
#> 372  -2.359325e-01  0.2604296630       0       0
#> 373  -2.032119e-01  0.1304329462       0       0
#> 374  -1.401798e-01  0.0811664848       0       0
#> 375  -1.886234e-01  0.2640906388       0       0
#> 376  -7.602891e-02 -0.0061684259       0       0
#> 377  -2.442594e-01  0.3050206568       0       0
#> 378  -1.300784e-01  0.2048950140       0       0
#> 379  -9.692633e-02 -0.0773722806       0       0
#> 380  -1.242796e-01  0.0599873701       0       0
#> 381  -2.861885e-01 -0.0364653991       0       0
#> 382  -1.970650e-01  0.3518804365       0       0
#> 383  -6.030168e-02  0.0219531610       0       0
#> 384   8.045142e-02 -0.1482273243       0       0
#> 385  -9.449849e-02 -0.0422554061       0       0
#> 386  -1.560335e-01  0.0097660394       0       0
#> 387  -9.452314e-02  0.1218862615       0       0
#> 388  -2.546581e-02  0.0978035285       0       0
#> 389  -1.409830e-01  0.0413656687       0       0
#> 390   5.026067e-02 -0.1662569754       0       0
#> 391  -1.584294e-01 -0.0467040355       0       0
#> 392  -2.680424e-01  0.0553334618       0       0
#> 393  -9.263136e-02 -0.1400940472       0       0
#> 394  -2.778612e-02  0.0228373031       0       0
#> 395  -1.175615e-01  0.1285658186       0       0
#> 396  -1.537002e-01  0.0609290395       0       0
#> 397  -7.553482e-02  0.1502137060       0       0
#> 398  -9.863061e-02 -0.1455940661       0       0
#> 399  -1.034249e-01  0.0904155256       0       0
#> 400  -1.010887e-01  0.1845586113       0       0
#> 401  -1.234054e-01  0.0388655562       0       0
#> 402  -1.758646e-01  0.0128019367       0       0
#> 403  -2.297483e-01  0.2136312914       0       0
#> 404   2.061468e-02 -0.0100437378       0       0
#> 405  -3.563981e-01  0.0671090750       0       0
#> 406  -3.272497e-02 -0.0584561160       0       0
#> 407  -5.715965e-02  0.0366175860       0       0
#> 408  -2.086152e-01  0.0070237313       0       0
#> 409   3.296241e-03 -0.0459349373       0       0
#> 410  -9.612987e-02  0.2794960089       0       0
#> 411  -7.428844e-02  0.1152670226       0       0
#> 412  -3.439568e-02 -0.0026429411       0       0
#> 413   1.195874e-01 -0.3212986781       0       0
#> 414  -1.298250e-01  0.0922025134       0       0
#> 415  -1.970650e-01  0.3518804365       0       0
#> 416  -2.350436e-01  0.0407091751       0       0
#> 417  -4.617461e-01  0.1832365817       0       0
#> 418  -7.715612e-02  0.1373897266       0       0
#> 419  -8.007922e-02  0.0256568129       0       0
#> 420  -1.386376e-02 -0.0437644920       0       0
#> 421  -9.182672e-02  0.0064223293       0       0
#> 422  -1.826666e-01  0.2797295360       0       0
#> 423  -2.579718e-01  0.0941330132       0       0
#> 424  -1.123308e-01  0.1189820804       0       0
#> 425  -1.017241e-01  0.1648481449       0       0
#> 426   5.358598e-02 -0.0992369324       0       0
#> 427  -8.436779e-02  0.1028876685       0       0
#> 428  -1.280264e-01  0.0276187452       0       0
#> 429  -2.133463e-01  0.1238214204       0       0
#> 430  -1.031221e-01  0.1144050363       0       0
#> 431  -1.131966e-01 -0.0185134679       0       0
#> 432  -1.149698e-01  0.0521081157       0       0
#> 433   6.596508e-03 -0.0178687838       0       0
#> 434  -1.432780e-01  0.3222595826       0       0
#> 435  -3.046253e-02  0.0786332859       0       0
#> 436  -5.551517e-02  0.0392643906       0       0
#> 437  -1.953972e-01  0.2368954494       0       0
#> 438  -2.775910e-02  0.0062687247       0       0
#> 439  -1.352157e-01  0.0126524625       0       0
#> 440  -1.245502e-01  0.0545411616       0       0
#> 441  -1.598564e-01  0.1659310479       0       0
#> 442  -1.672202e-01  0.0428974482       0       0
#> 443   1.051061e-02 -0.1794927474       0       0
#> 444  -3.838432e-01  0.0782528475       0       0
#> 445  -7.407299e-02  0.0985046724       0       0
#> 446  -2.597395e-01 -0.0651035820       0       0
#> 447  -5.798601e-02  0.0496509201       0       0
#> 448  -1.413529e-01 -0.0404734840       0       0
#> 449   2.154706e-02 -0.1231244026       0       0
#> 450  -1.373362e-01  0.1274145082       0       0
#> 451  -1.446531e-01  0.0940162853       0       0
#> 452  -8.542452e-02  0.0374160186       0       0
#> 453  -5.393832e-02 -0.0074153311       0       0
#> 454  -1.545138e-01  0.0401218658       0       0
#> 455  -2.782145e-01  0.1438056018       0       0
#> 456  -4.362875e-02  0.1412319884       0       0
#> 457  -1.875078e-01  0.2110650532       0       0
#> 458  -1.822389e-01  0.0652275615       0       0
#> 459  -1.442164e-01  0.0975453403       0       0
#> 460  -2.076426e-02 -0.0011231946       0       0
#> 461  -2.343359e-02  0.0361996811       0       0
#> 462  -3.585023e-02 -0.0279443864       0       0
#> 463  -9.431705e-02 -0.0557544220       0       0
#> 464  -4.354250e-01  0.1803820806       0       0
#> 465  -1.020717e-01  0.0615664152       0       0
#> 466  -1.368086e-01  0.1470608721       0       0
#> 467  -5.291389e-02  0.1840376311       0       0
#> 468  -9.372037e-02  0.0607437821       0       0
#> 469  -9.746579e-02  0.1226563257       0       0
#> 470  -1.240150e-01  0.0142087603       0       0
#> 471   5.314181e-02 -0.1194661827       0       0
#> 472  -1.145848e-01  0.1531802474       0       0
#> 473  -3.231315e-01  0.1461101710       0       0
#> 474  -1.144927e-01  0.2727724283       0       0
#> 475  -1.525182e-01  0.2125039594       0       0
#> 476  -3.625965e-02 -0.2125229870       0       0
#> 477  -2.059767e-01  0.1620772952       0       0
#> 478   6.895223e-03 -0.1461920714       0       0
#> 479  -3.321334e-02  0.0787636381       0       0
#> 480  -6.484761e-02  0.1168599730       0       0
#> 481  -1.090744e-01  0.2593979139       0       0
#> 482  -2.546581e-02  0.0978035285       0       0
#> 483   6.629065e-02  0.0272429251       0       0
#> 484  -1.796849e-01  0.0591076046       0       0
#> 485  -9.436964e-02  0.0950510154       0       0
#> 486  -4.388123e-02 -0.1385203706       0       0
#> 487  -6.504020e-02  0.0035013478       0       0
#> 488  -2.442594e-01  0.3050206568       0       0
#> 489  -6.080532e-02  0.0421009231       0       0
#> 490  -1.903075e-01  0.1363301945       0       0
#> 491  -1.140279e-01 -0.0076689297       0       0
#> 492  -2.166968e-01  0.0867363450       0       0
#> 493  -1.860235e-01  0.0571266323       0       0
#> 494  -6.589192e-02  0.0659440338       0       0
#> 495  -3.453147e-01  0.0442355381       0       0
#> 496  -1.243508e-01  0.1691018243       0       0
#> 497  -3.802156e-01  0.0313273871       0       0
#> 498  -1.569731e-01 -0.0075117348       0       0
#> 499  -1.499586e-01  0.0655082726       0       0
#> 500   6.803225e-05 -0.0116025461       0       0
#> 501  -1.389985e-01  0.0664458692       0       0
#> 502  -1.327366e-01 -0.0268026321       0       0
#> 503  -2.366046e-01  0.1142900685       0       0
#> 504   1.640377e-02 -0.0334579889       0       0
#> 505  -9.612987e-02  0.2794960089       0       0
#> 506  -7.119266e-02 -0.0087554127       0       0
#> 507  -5.090233e-02  0.0769839744       0       0
#> 508  -1.978163e-01 -0.0091851926       0       0
#> 509  -1.482908e-02  0.0436603738       0       0
#> 510  -2.439693e-01  0.1485414573       0       0
#> 511  -1.504900e-01  0.2047438881       0       0
#> 512  -7.447034e-02  0.2083916101       0       0
#> 513  -1.981368e-01  0.0771900318       0       0
#> 514  -1.905835e-01  0.1751608553       0       0
#> 515  -6.080532e-02  0.0421009231       0       0
#> 516  -6.451023e-02  0.0795680826       0       0
#> 517  -1.382393e-01  0.2464604671       0       0
#> 518  -1.306919e-01  0.1253269469       0       0
#> 519  -3.982408e-01  0.1127296766       0       0
#> 520  -8.364196e-02  0.0224329781       0       0
#> 521   2.257184e-02  0.0379772475       0       0
#> 522  -1.144927e-01  0.2727724283       0       0
#> 523  -1.323463e-01  0.2377324553       0       0
#> 524  -1.320836e-02  0.0119393507       0       0
#> 525  -1.129571e-01  0.0407985337       0       0
#> 526  -1.539777e-01  0.1210764134       0       0
#> 527  -7.912689e-02  0.1601497080       0       0
#> 528  -1.425666e-01  0.1955179183       0       0
#> 529  -1.513270e-01  0.1237263048       0       0
#> 530  -1.605991e-01 -0.0420548284       0       0
#> 531  -6.896198e-02  0.1432787295       0       0
#> 532  -1.006943e-01 -0.0580607821       0       0
#> 533  -1.127357e-01  0.1268709175       0       0
#> 534  -1.753703e-01  0.1364963034       0       0
#> 535  -1.252738e-01  0.0123157478       0       0
#> 536  -1.849259e-01 -0.0049709627       0       0
#> 537  -9.066539e-02  0.1796477666       0       0
#> 538  -1.628087e-01  0.2777224869       0       0
#> 539  -6.713966e-02  0.2745354457       0       0
#> 540  -2.004699e-02  0.0840796846       0       0
#> 541  -1.091518e-01  0.1639251276       0       0
#> 542  -5.121037e-02 -0.0400731120       0       0
#> 543  -1.269129e-01 -0.0366909212       0       0
#> 544  -1.766524e-01  0.2183280174       0       0
#> 545  -1.129269e-03  0.0142617911       0       0
#> 546  -2.216269e-02  0.1101513579       0       0
#> 547  -8.213414e-02  0.3195803142       0       0
#> 548  -1.239021e-01 -0.0985123286       0       0
#> 549  -4.020486e-02 -0.0399099519       0       0
#> 550  -6.709337e-02  0.0831887772       0       0
#> 551  -2.963027e-02 -0.0342456740       0       0
#> 552  -2.680424e-01  0.0553334618       0       0
#> 553   2.085706e-02 -0.1130755666       0       0
#> 554  -1.148026e-01  0.1145199365       0       0
#> 555  -1.389985e-01  0.0664458692       0       0
#> 556  -6.179872e-02  0.0374107873       0       0
#> 557  -4.218343e-02 -0.0371012935       0       0
#> 558  -1.382393e-01  0.2464604671       0       0
#> 559  -1.631185e-01  0.1205466755       0       0
#> 560  -4.279755e-02 -0.0728195764       0       0
#> 561  -8.867951e-03  0.0227478585       0       0
#> 562  -9.452314e-02  0.1218862615       0       0
#> 563  -7.458400e-02 -0.0330317929       0       0
#> 564  -1.204893e-01  0.1797128023       0       0
#> 565  -1.263753e-01 -0.0416050227       0       0
#> 566  -2.216269e-02  0.1101513579       0       0
#> 567  -8.320878e-02  0.0765010277       0       0
#> 568   1.077456e-01 -0.3288249811       0       0
#> 569  -1.409910e-01  0.0136651019       0       0
#> 570  -5.787046e-02  0.0028013813       0       0
#> 571  -3.725509e-02 -0.1130232472       0       0
#> 572  -6.191114e-02 -0.0876457244       0       0
#> 573  -2.914193e-01  0.0716681206       0       0
#> 574   6.307451e-03 -0.1028017411       0       0
#> 575  -9.005127e-02  0.1042915281       0       0
#> 576  -1.352749e-01  0.0697798261       0       0
#> 577  -8.040202e-02  0.1240501720       0       0
#> 578  -2.128370e-01  0.0701147005       0       0
#> 579  -2.261377e-01  0.0129792457       0       0
#> 580  -2.132951e-02 -0.1451099333       0       0
#> 581  -2.858180e-01  0.2410143017       0       0
#> 582  -5.602767e-02  0.0979232254       0       0
#> 583   1.786817e-02 -0.0639400265       0       0
#> 584  -2.129506e-03 -0.0370102915       0       0
#> 585   2.568500e-02 -0.0917606817       0       0
#> 586  -1.515145e-01 -0.0321986046       0       0
#> 587  -1.088299e-01  0.1236684432       0       0
#> 588  -2.455846e-01  0.1136250458       0       0
#> 589   5.438988e-02  0.0278927330       0       0
#> 590  -6.844014e-02 -0.0156464627       0       0
#> 591  -1.148908e-01  0.0378228569       0       0
#> 592  -3.073684e-01  0.1005921439       0       0
#> 593  -1.148908e-01  0.0378228569       0       0
#> 594  -1.869348e-01  0.1313081836       0       0
#> 595  -1.967914e-01  0.2443585825       0       0
#> 596  -2.375267e-02 -0.1037818282       0       0
#> 597  -1.705016e-01 -0.0014498981       0       0
#> 598  -1.440346e-01  0.3936407550       0       0
#> 599  -9.389774e-02 -0.0426837023       0       0
#> 600  -3.493539e-02  0.0138886815       0       0
#> 601  -1.518026e-01  0.0832192162       0       0
#> 602  -7.186086e-02  0.0064103649       0       0
#> 603  -1.090744e-01  0.2593979139       0       0
#> 604   2.276789e-02 -0.1716421515       0       0
#> 605  -5.943770e-02  0.1248286221       0       0
#> 606  -1.844123e-01 -0.0080906594       0       0
#> 607  -8.251052e-02 -0.0487403143       0       0
#> 608  -1.227407e-01  0.2280852458       0       0
#> 609  -1.401798e-01  0.0811664848       0       0
#> 610  -2.700516e-01  0.2873600052       0       0
#> 611  -2.909316e-01  0.2505404116       0       0
#> 612  -1.634090e-01  0.1314370689       0       0
#> 613  -8.820534e-02  0.1518209905       0       0
#> 614  -1.578137e-01  0.2017647160       0       0
#> 615  -1.504061e-01  0.1823521193       0       0
#> 616  -2.361959e-01  0.0435028715       0       0
#> 617  -1.617324e-01  0.1015302873       0       0
#> 618  -5.885826e-02 -0.0672936504       0       0
#> 619  -4.625611e-02 -0.0314222907       0       0
#> 620  -1.607735e-01  0.2009026288       0       0
#> 621  -5.427212e-02  0.1695928313       0       0
#> 622  -1.788638e-01  0.1190367666       0       0
#> 623  -2.204938e-01  0.1305238275       0       0
#> 624  -5.620970e-02 -0.0044562009       0       0
#> 625  -1.533079e-01  0.0314457038       0       0
#> 626  -8.912206e-02  0.0184606706       0       0
#> 627  -3.079211e-01  0.0765716178       0       0
#> 628  -1.112325e-01  0.0484845900       0       0
#> 629   1.755314e-02 -0.0432774736       0       0
#> 630  -1.604734e-02 -0.0659996703       0       0
#> 631  -3.084732e-01  0.2123950385       0       0
#> 632  -2.460126e-01  0.1816714921       0       0
#> 633  -2.419605e-02 -0.1059262030       0       0
#> 634  -5.607120e-02  0.0017579824       0       0
#> 635  -1.662069e-01  0.1509489061       0       0
#> 636  -1.516057e-01  0.0631686584       0       0
#> 637  -2.024138e-01  0.2474196429       0       0
#> 638  -1.239343e-02 -0.0207190497       0       0
#> 639  -1.886944e-01  0.1647363816       0       0
#> 640  -1.010887e-01  0.1845586113       0       0
#> 641  -1.231328e-01  0.1399448137       0       0
#> 642  -2.442594e-01  0.3050206568       0       0
#> 643  -4.171667e-02  0.0563925646       0       0
#> 644  -1.522624e-01 -0.0301622503       0       0
#> 645  -2.442594e-01  0.3050206568       0       0
#> 646  -6.094742e-02  0.0569654551       0       0
#> 647  -8.747730e-02  0.0462384496       0       0
#> 648  -1.698606e-01  0.0078934578       0       0
#> 649  -1.385833e-01 -0.0126724803       0       0
#> 650  -1.104695e-01  0.0181498394       0       0
#> 651  -4.572659e-02  0.1259671698       0       0
#> 652  -1.069818e-01  0.0822849814       0       0
#> 653  -9.220912e-02 -0.0794883624       0       0
#> 654  -9.942761e-02 -0.0531042856       0       0
#> 655  -2.861885e-01 -0.0364653991       0       0
#> 656  -1.134577e-01  0.0636179967       0       0
#> 657  -7.266680e-02 -0.0673369460       0       0
#> 658  -2.361665e-01  0.0848296099       0       0
#> 659  -7.142599e-02  0.2879868891       0       0
#> 660  -8.464342e-02 -0.2168536899       0       0
#> 661  -5.432844e-02 -0.0584815943       0       0
#> 662  -2.234276e-02  0.0079881418       0       0
#> 663  -9.516486e-02 -0.0325254431       0       0
#> 664  -1.987367e-01  0.0308658906       0       0
#> 665  -1.283992e-01  0.0965960155       0       0
#> 666  -5.423463e-02  0.0414697292       0       0
#> 667  -1.171926e-01  0.0020868688       0       0
#> 668  -1.123308e-01  0.1189820804       0       0
#> 669  -1.231328e-01  0.1399448137       0       0
#> 670  -2.052785e-01  0.0065588542       0       0
#> 671  -1.436053e-01 -0.0554267910       0       0
#> 672  -4.651890e-02  0.1434616102       0       0
#> 673  -4.989064e-02  0.0822437053       0       0
#> 674  -1.485272e-01  0.0617720401       0       0
#> 675  -3.859417e-02  0.0379269107       0       0
#> 676  -7.488640e-02  0.0902436630       0       0
#> 677  -4.102474e-02  0.2127753039       0       0
#> 678  -2.332661e-01  0.1958108544       0       0
#> 679  -1.254078e-01  0.2140500873       0       0
#> 680  -2.279673e-01  0.0493526467       0       0
#> 681  -1.581104e-01 -0.0372949923       0       0
#> 682  -1.886234e-01  0.2640906388       0       0
#> 683  -2.643741e-01 -0.0151050644       0       0
#> 684  -5.976752e-02  0.0111302898       0       0
#> 685  -7.142599e-02  0.2879868891       0       0
#> 686  -8.566167e-02 -0.2126173971       0       0
#> 687   9.201012e-03  0.0942565948       0       0
#> 688  -1.116300e-01  0.0670282587       0       0
#> 689  -3.551459e-03 -0.1148866296       0       0
#> 690  -2.485964e-01  0.1514430287       0       0
#> 691   9.479203e-03 -0.1811505518       0       0
#> 692  -1.036754e-01  0.1506983466       0       0
#> 693  -1.631185e-01  0.1205466755       0       0
#> 694  -4.925422e-02  0.0087673024       0       0
#> 695  -6.375014e-02 -0.0033465591       0       0
#> 696  -1.641128e-01  0.1554407438       0       0
#> 697  -1.970650e-01  0.3518804365       0       0
#> 698  -2.206327e-01  0.1059945513       0       0
#> 699  -4.851858e-02 -0.0386930463       0       0
#> 700  -3.131165e-02  0.0633056002       0       0
#> 701  -3.554174e-01  0.3058067977       0       0
#> 702  -1.174263e-01  0.2796359669       0       0
#> 703   3.129762e-02 -0.0086766510       0       0
#> 704  -2.854269e-01 -0.0542935232       0       0
#> 705  -1.568719e-01 -0.0779601770       0       0
#> 706  -6.481609e-03 -0.1681148907       0       0
#> 707  -9.539515e-02 -0.1769533699       0       0
#> 708  -2.478134e-01  0.1614562795       0       0
#> 709  -1.150806e-01  0.1616762908       0       0
#> 710  -3.225547e-02  0.0462200034       0       0
#> 711  -1.017525e-01  0.1454209395       0       0
#> 712  -4.074610e-02  0.1180170391       0       0
#> 713  -1.653741e-01 -0.0808586604       0       0
#> 714  -4.003448e-02  0.1051972735       0       0
#> 715  -2.206327e-01  0.1059945513       0       0
#> 716  -2.619732e-01  0.1473161465       0       0
#> 717  -2.624470e-02  0.0733330773       0       0
#> 718  -8.871593e-02  0.0276623870       0       0
#> 719   1.299020e-02 -0.0585050072       0       0
#> 720  -3.923718e-02 -0.0135422309       0       0
#> 721  -4.096640e-02 -0.1525341872       0       0
#> 722  -1.705016e-01 -0.0014498981       0       0
#> 723  -4.337889e-02 -0.0024214347       0       0
#> 724  -2.378975e-01  0.1613644843       0       0
#> 725  -9.372037e-02  0.0607437821       0       0
#> 726   4.500410e-03  0.0143382640       0       0
#> 727  -1.628087e-01  0.2777224869       0       0
#> 728  -7.686875e-02  0.0416913879       0       0
#> 729   4.452715e-02  0.0524819307       0       0
#> 730  -1.623202e-01  0.1215591831       0       0
#> 731   2.476974e-02  0.0391470746       0       0
#> 732  -7.433320e-02 -0.0372288836       0       0
#> 733  -5.798601e-02  0.0496509201       0       0
#> 734  -3.337524e-01  0.2090093522       0       0
#> 735  -1.185217e-01  0.1135500760       0       0
#> 736  -7.098363e-02  0.0398228711       0       0
#> 737  -1.219507e-02 -0.0933104434       0       0
#> 738  -2.806122e-01  0.0782921961       0       0
#> 739   5.427073e-02 -0.1217017577       0       0
#> 740  -8.439382e-02  0.0872718196       0       0
#> 741  -3.131468e-03 -0.0724969197       0       0
#> 742  -3.100542e-01  0.2245048283       0       0
#> 743  -1.440346e-01  0.3936407550       0       0
#> 744  -2.111927e-01 -0.0972566445       0       0
#> 745  -8.108665e-02 -0.0087750284       0       0
#> 746  -5.607473e-02 -0.0726812590       0       0
#> 747  -7.739412e-03 -0.1293093070       0       0
#> 748  -8.386250e-02  0.1279050349       0       0
#> 749   5.497466e-02 -0.0067898041       0       0
#> 750  -1.913013e-01  0.1991403555       0       0
#> 751  -3.431242e-01 -0.0299708152       0       0
#> 752   6.839143e-02  0.0725614639       0       0
#> 753  -1.180909e-01 -0.0108865868       0       0
#> 754  -1.921705e-01 -0.0218741671       0       0
#> 755  -6.207600e-02  0.0908934232       0       0
#> 756  -3.795733e-02  0.1617148385       0       0
#> 757  -9.251014e-02  0.2034756731       0       0
#> 758  -1.175831e-01 -0.1180251544       0       0
#> 759  -2.132951e-02 -0.1451099333       0       0
#> 760   2.763353e-02 -0.1696924287       0       0
#> 761  -2.329375e-01  0.1949050979       0       0
#> 762  -2.932003e-01  0.1656158875       0       0
#> 763  -6.371272e-02  0.1186051895       0       0
#> 764  -1.060783e-01  0.0956304035       0       0
#> 765  -5.331341e-02 -0.0376787069       0       0
#> 766  -1.175615e-01  0.1285658186       0       0
#> 767  -2.359325e-01  0.2604296630       0       0
#> 768  -1.280584e-01  0.0380539063       0       0
#> 769  -1.154387e-01  0.0720470504       0       0
#> 770  -9.529884e-02  0.2288599500       0       0
#> 771  -1.770077e-01  0.1699076265       0       0
#> 772  -6.592857e-02  0.0005429103       0       0
#> 773  -4.268109e-01 -0.0287897933       0       0
#> 774  -2.024138e-01  0.2474196429       0       0
#> 775  -3.606013e-01 -0.1076679244       0       0
#> 776  -1.382393e-01  0.2464604671       0       0
#> 777  -1.414352e-01  0.0414063540       0       0
#> 778  -1.269129e-01 -0.0366909212       0       0
#> 779  -7.382719e-02  0.1905955287       0       0
#> 780   5.300840e-02 -0.1135392374       0       0
#> 781  -1.047232e-01  0.1366953948       0       0
#> 782  -6.914706e-02  0.1008101091       0       0
#> 783  -7.645442e-02  0.0483610377       0       0
#> 784  -7.083090e-02 -0.0326850020       0       0
#> 785  -7.912689e-02  0.1601497080       0       0
#> 786  -8.549896e-02 -0.0765851594       0       0
#> 787  -4.216778e-02 -0.0825321842       0       0
#> 788  -1.794342e-01  0.1481340674       0       0
#> 789  -5.597149e-02  0.0030503888       0       0
#> 790  -1.156678e-01  0.0603001327       0       0
#> 791  -3.910774e-01  0.1325090994       0       0
#> 792  -1.232790e-01 -0.0255328077       0       0
#> 793  -8.318805e-02  0.0717342318       0       0
#> 794  -1.662069e-01  0.1509489061       0       0
#> 795  -8.937780e-03  0.0111544196       0       0
#> 796  -3.944886e-02 -0.0274960191       0       0
#> 797  -1.150806e-01  0.1616762908       0       0
#> 798  -1.197581e-01  0.0790986028       0       0
#> 799  -2.184167e-01  0.1233244995       0       0
#> 800  -3.775405e-02  0.0663146680       0       0
#> 801  -5.529301e-02  0.1654233684       0       0
#> 802  -4.166006e-02 -0.1873098831       0       0
#> 803  -2.163229e-01  0.1242208451       0       0
#> 804  -5.317805e-03 -0.0666214131       0       0
#> 805  -1.552583e-01  0.1700917391       0       0
#> 806  -7.647135e-02 -0.0692808010       0       0
#> 807  -2.726424e-01  0.2358194685       0       0
#> 808  -3.910774e-01  0.1325090994       0       0
#> 809  -1.519210e-01  0.1088651591       0       0
#> 810  -8.374020e-02  0.0682057859       0       0
#> 811  -3.754524e-02  0.1838290137       0       0
#> 812  -4.572659e-02  0.1259671698       0       0
#> 813  -1.551455e-01  0.0448840498       0       0
#> 814  -1.171580e-03 -0.1102040239       0       0
#> 815  -1.770077e-01  0.1699076265       0       0
#> 816  -1.667048e-01  0.3402289678       0       0
#> 817  -4.651890e-02  0.1434616102       0       0
#> 818  -1.086585e-01  0.0862362135       0       0
#> 819  -2.551968e-01  0.1151337965       0       0
#> 820   9.578564e-02 -0.0597924642       0       0
#> 821  -1.065036e-01 -0.0147152352       0       0
#> 822  -5.597149e-02  0.0030503888       0       0
#> 823  -1.061982e-01  0.1040801881       0       0
#> 824  -4.011619e-02 -0.0136489166       0       0
#> 825  -5.921032e-02  0.0484835985       0       0
#> 826  -9.857315e-02  0.1158393432       0       0
#> 827  -4.276541e-02  0.0156205239       0       0
#> 828  -2.312682e-01  0.0518819307       0       0
#> 829  -2.024138e-01  0.2474196429       0       0
#> 830  -3.357499e-01  0.1183298940       0       0
#> 831  -7.645442e-02  0.0483610377       0       0
#> 832  -7.142599e-02  0.2879868891       0       0
#> 833  -1.224482e-01  0.0879001323       0       0
#> 834  -1.069818e-01  0.0822849814       0       0
#> 835  -6.438494e-02  0.1284123492       0       0
#> 836  -4.687052e-02  0.0445779148       0       0
#> 837  -1.447418e-01  0.0473418581       0       0
#> 838  -5.909908e-02  0.0147389111       0       0
#> 839   1.033301e-01 -0.2517018729       0       0
#> 840  -1.368086e-01  0.1470608721       0       0
#> 841  -1.022777e-01 -0.0093856928       0       0
#> 842  -1.152537e-01  0.3331937809       0       0
#> 843  -3.342204e-01  0.1464539398       0       0
#> 844   1.101451e-02  0.0068728192       0       0
#> 845   5.825639e-02 -0.0925074825       0       0
#> 846  -1.329049e-01  0.0439801938       0       0
#> 847  -1.047232e-01  0.1366953948       0       0
#> 848  -1.488546e-01  0.0072908582       0       0
#> 849  -3.231315e-01  0.1461101710       0       0
#> 850  -3.035569e-01  0.2165370589       0       0
#> 851  -1.975957e-01  0.1244083373       0       0
#> 852  -8.055203e-02  0.0232987632       0       0
#> 853  -1.325365e-01  0.0809620129       0       0
#> 854  -1.704774e-01  0.1642687708       0       0
#> 855  -1.135055e-01  0.1334465478       0       0
#> 856  -2.890787e-01  0.3015542831       0       0
#> 857   3.625648e-02 -0.0622940206       0       0
#> 858  -3.329341e-01  0.0679346930       0       0
#> 859  -7.800785e-02  0.0264550651       0       0
#> 860  -1.382393e-01  0.2464604671       0       0
#> 861  -7.650917e-02 -0.0356818467       0       0
#> 862   3.980951e-02 -0.2354806834       0       0
#> 863  -2.202456e-01  0.0824667331       0       0
#> 864  -2.058564e-01  0.2434417235       0       0
#> 865  -2.293845e-01  0.0913087745       0       0
#> 866  -6.409560e-02  0.1185178934       0       0
#> 867  -2.871866e-02  0.1073337522       0       0
#> 868  -1.028252e-01 -0.1713507603       0       0
#> 869  -2.819969e-01 -0.0927051611       0       0
#> 870  -1.623732e-01 -0.1189287482       0       0
#> 871  -8.062735e-02  0.1118859688       0       0
#> 872  -6.577126e-02  0.1478542663       0       0
#> 873   2.763353e-02 -0.1696924287       0       0
#> 874  -1.233012e-01  0.0761919579       0       0
#> 875  -1.762553e-01  0.2503983140       0       0
#> 876  -5.956710e-02  0.0840332003       0       0
#> 877  -6.844014e-02 -0.0156464627       0       0
#> 878  -9.179011e-02  0.1131073648       0       0
#> 879  -2.042639e-01  0.2791463995       0       0
#> 880  -3.658104e-02  0.0091894670       0       0
#> 881  -5.010102e-02  0.1231953031       0       0
#> 882  -7.878005e-02 -0.0428932064       0       0
#> 883  -4.778433e-02  0.0255728675       0       0
#> 884  -6.732591e-02  0.0035483644       0       0
#> 885  -1.303399e-01  0.0702528439       0       0
#> 886  -8.983058e-02  0.1732395356       0       0
#> 887  -5.817045e-02  0.1976650665       0       0
#> 888  -2.036151e-01  0.2121194883       0       0
#> 889  -1.528186e-01 -0.0231478877       0       0
#> 890  -2.754048e-02  0.0702811166       0       0
#> 891  -2.236607e-01 -0.0722718224       0       0
#> 892  -3.481983e-01  0.1185277812       0       0
#> 893  -3.407885e-01  0.2895756827       0       0
#> 894  -1.749924e-01  0.2483017602       0       0
#> 895  -2.147831e-01  0.3482514405       0       0
#> 896  -5.607120e-02  0.0017579824       0       0
#> 897  -3.465152e-01  0.2144757571       0       0
#> 898  -1.804569e-01  0.1657057687       0       0
#> 899  -2.021312e-01 -0.0122276426       0       0
#> 900   4.627423e-03 -0.0223205513       0       0
#> 901  -1.672651e-01 -0.0311453403       0       0
#> 902  -1.014223e-01 -0.0536692915       0       0
#> 903   3.052044e-02 -0.1058430283       0       0
#> 904  -4.042264e-01  0.1978061254       0       0
#> 905  -7.634468e-02  0.2209731181       0       0
#> 906  -4.529570e-02  0.0603311492       0       0
#> 907  -3.834769e-01  0.1330937864       0       0
#> 908  -1.667048e-01  0.3402289678       0       0
#> 909  -1.281279e-01  0.1373731706       0       0
#> 910  -9.531953e-02  0.1813589475       0       0
#> 911  -9.389774e-02 -0.0426837023       0       0
#> 912  -3.337726e-01  0.1692126951       0       0
#> 913  -3.961811e-02  0.0496955170       0       0
#> 914  -2.169003e-01  0.0970636473       0       0
#> 915  -1.452825e-01  0.1459123866       0       0
#> 916  -1.432780e-01  0.3222595826       0       0
#> 917  -1.513270e-01  0.1237263048       0       0
#> 918  -1.875021e-01  0.2356118067       0       0
#> 919  -1.775991e-01 -0.0002313434       0       0
#> 920  -3.296224e-01  0.2143109942       0       0
#> 921  -2.328633e-02  0.0281959775       0       0
#> 922  -4.699820e-02  0.1018797750       0       0
#> 923  -7.488640e-02  0.0902436630       0       0
#> 924  -2.099622e-01  0.0642526128       0       0
#> 925  -2.253701e-01  0.0715343839       0       0
#> 926   1.487511e-02 -0.0900537391       0       0
#> 927  -1.788386e-01 -0.0062533052       0       0
#> 928  -2.498699e-01  0.1559379434       0       0
#> 929  -1.492023e-01  0.0857821856       0       0
#> 930  -2.318226e-01  0.0758013014       0       0
#> 931  -8.178080e-02  0.0741584719       0       0
#> 932  -2.519717e-01 -0.1675170763       0       0
#> 933  -1.713379e-02  0.1585964923       0       0
#> 934  -4.426043e-01  0.2138083558       0       0
#> 935  -6.409560e-02  0.1185178934       0       0
#> 936  -1.885906e-02 -0.0088008131       0       0
#> 937  -8.465247e-02  0.0092706092       0       0
#> 938  -2.409522e-01  0.0255808678       0       0
#> 939  -1.595831e-01  0.2272813915       0       0
#> 940  -1.382393e-01  0.2464604671       0       0
#> 941  -1.818005e-01  0.0153866617       0       0
#> 942  -3.797623e-02 -0.0804480450       0       0
#> 943  -2.147831e-01  0.3482514405       0       0
#> 944  -1.507159e-01  0.1148551729       0       0
#> 945  -4.423007e-02  0.0783740084       0       0
#> 946  -2.131115e-01  0.2109108317       0       0
#> 947  -5.095032e-02 -0.0047193341       0       0
#> 948  -2.105105e-02  0.2623096298       0       0
#> 949  -7.634468e-02  0.2209731181       0       0
#> 950  -9.729661e-02  0.1208994156       0       0
#> 951   3.422949e-02 -0.0161103332       0       0
#> 952  -1.528022e-01  0.2561682281       0       0
#> 953  -1.216467e-01  0.0267295972       0       0
#> 954  -7.549225e-02  0.1395424753       0       0
#> 955  -1.836865e-01  0.0131814554       0       0
#> 956  -1.049687e-02 -0.0090446478       0       0
#> 957  -5.951031e-02  0.0899345291       0       0
#> 958  -9.849737e-02  0.1125699352       0       0
#> 959  -6.629594e-02  0.1511350177       0       0
#> 960  -8.465247e-02  0.0092706092       0       0
#> 961  -1.252738e-01  0.0123157478       0       0
#> 962  -2.386491e-01  0.0848047136       0       0
#> 963  -2.528445e-01  0.0158975492       0       0
#> 964  -9.887563e-02  0.2111151703       0       0
#> 965  -4.298371e-01  0.1530173479       0       0
#> 966  -1.047232e-01  0.1366953948       0       0
#> 967  -2.097940e-01  0.0520128376       0       0
#> 968  -9.355739e-02  0.1197106299       0       0
#> 969  -1.385833e-01 -0.0126724803       0       0
#> 970  -1.091775e-01  0.1739297908       0       0
#> 971  -2.288885e-01  0.1630602267       0       0
#> 972  -1.652119e-01 -0.0156835590       0       0
#> 973   2.595515e-03 -0.0215218197       0       0
#> 974  -1.923476e-01 -0.0057404422       0       0
#> 975  -4.310735e-02  0.0899708376       0       0
#> 976  -2.042639e-01  0.2791463995       0       0
#> 977  -1.089392e-01  0.0500441673       0       0
#> 978  -1.152537e-01  0.3331937809       0       0
#> 979   4.864439e-02  0.0453348080       0       0
#> 980  -1.595094e-01  0.1453735164       0       0
#> 981  -3.197149e-01  0.3206349422       0       0
#> 982  -2.242257e-01  0.0994402028       0       0
#> 983   3.795145e-02  0.0232842133       0       0
#> 984  -1.472287e-01  0.2033709317       0       0
#> 985  -5.912717e-02  0.1005183865       0       0
#> 986  -4.459484e-02  0.1914499925       0       0
#> 987  -9.233843e-02 -0.1109575396       0       0
#> 988  -5.798032e-02 -0.0312500071       0       0
#> 989  -2.699254e-02 -0.1391959679       0       0
#> 990  -1.360101e-01  0.1293263719       0       0
#> 991  -1.536197e-01  0.0870613812       0       0
#> 992  -1.422535e-01  0.0730513907       0       0
#> 993  -2.495152e-01 -0.1351099824       0       0
#> 994  -8.408737e-03 -0.0120115331       0       0
#> 995  -3.112673e-01  0.0748254682       0       0
#> 996  -8.407718e-02  0.1669447873       0       0
#> 997  -2.498699e-01  0.1559379434       0       0
#> 998  -2.670810e-01  0.0281793148       0       0
#> 999  -4.763679e-02  0.0813573185       0       0
#> 1000 -2.029097e-01  0.0620946640       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.0469  0.0446 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.108  -0.00723
#>  3      3     0.21      80    65    19.8     0.041      0.12  0.0913 -0.0374 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0105 -0.00904
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.0510 -0.00472
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0886  0.127  
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0265 -0.0304 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.203  -0.0178 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.180   0.166  
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.659  -0.0636 
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.238  20.4 0.041  0.12 79.9  1627. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.238  20.4 0.041  0.12  4.71   95.8
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.224  19.3 0.041  0.12 85.1  1644. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.224  19.3 0.041  0.12  5.74  111. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.273  18.8 0.041  0.12 84.5  1585. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.273  18.8 0.041  0.12  3.60   67.6
#> # … with 1,994 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>