Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object

A posologyr list, created by the posologyr function.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01,n_sample=1e4,n_resample=1e3)
#> 
#> $eta
#>            ETA_Cl        ETA_Vc        ETA_Ka
#> 1     0.499627317 -0.6109912928 -0.4080762043
#> 2     0.678189730 -0.2316035839 -0.0414137216
#> 3     0.588656038 -0.5169057049  0.0208087330
#> 4     0.377652685 -0.2595962943  0.2569206114
#> 5     0.724992728 -0.2106123036  0.8821522758
#> 6     0.499060281 -0.8003335249 -0.7371727346
#> 7     0.724992728 -0.2106123036  0.8821522758
#> 8     0.562168663 -0.8310265025  0.1912981649
#> 9     0.214753934 -0.4524970894 -0.1911090866
#> 10    0.604534813 -0.5913970823 -0.4667145219
#> 11    0.286088052 -1.0191574466  0.0288623655
#> 12    0.160431886 -1.3112030263 -0.7382606122
#> 13    0.272810832 -0.1517590113  0.4605230395
#> 14    0.486217076 -0.3543872491  0.2370284288
#> 15    0.488745471 -0.3357001244  0.1559001421
#> 16    0.059257714 -0.2307624859  0.9574233896
#> 17    0.710554541 -0.3434791388 -0.2766488795
#> 18    0.456398098 -0.6442839728  0.0471418818
#> 19    0.548272924 -0.5718551112 -0.3482535258
#> 20    0.550899696 -0.6707786708  0.0938493587
#> 21    0.815651607 -0.2775730537  0.1302660986
#> 22    0.538077142 -0.2963925665 -0.0924875345
#> 23    0.579756819 -0.6510707747 -0.0183299634
#> 24    0.234081352 -0.8073720787  0.1318624281
#> 25    0.460996796 -0.4880909954 -0.1036765708
#> 26    0.622449836 -0.5019979967  0.0351213020
#> 27    0.622449836 -0.5019979967  0.0351213020
#> 28    0.486801754 -0.7019084695  0.2632276767
#> 29    0.585390161 -0.3987097235  0.1598145380
#> 30    0.335996948  0.1007173186  0.8339859774
#> 31    0.512849580 -0.4042207709  0.3387078706
#> 32    0.375393845 -0.6059725714 -0.1129216582
#> 33    0.125652744 -0.2103339808  0.4932162860
#> 34    0.867434758 -0.1074519986  0.2304054635
#> 35    0.566896273 -0.5133644287 -0.0146844362
#> 36    0.491049389 -0.1183437861 -0.0500159184
#> 37    0.415488014 -0.6537340249  0.0255461761
#> 38    0.766745659 -0.4300152905  0.0666734868
#> 39    0.604528349 -0.4321340141  0.0428331941
#> 40    0.230159648 -0.8926485120  0.7931458708
#> 41    0.687980308 -0.4539346978  0.1707708211
#> 42    0.238094067 -0.5775378325 -0.0714604997
#> 43    0.536411464 -0.5113742527  0.1342418112
#> 44    0.447428677 -0.5491538294  0.5517318049
#> 45    0.595481868 -0.3862767492 -0.1955553955
#> 46    0.394194662 -0.6111377422 -0.0372018787
#> 47    0.340841154 -1.0988306712 -0.2963597973
#> 48    0.625986463 -0.5755168690 -0.1160674525
#> 49    0.287363159 -0.8499850081 -0.6796414828
#> 50    0.643276313 -0.0241843890  1.0574182227
#> 51    0.453952887 -0.8932974880 -0.0905298854
#> 52    0.935223910  0.1162423892  0.8157133323
#> 53    0.248673669 -0.3818391674 -0.0696769070
#> 54    0.579756819 -0.6510707747 -0.0183299634
#> 55    0.195499862 -0.8856516707 -0.0144414466
#> 56    0.264698626 -0.1310576764  0.4406215781
#> 57    0.201371857 -1.0477175201 -0.0260467305
#> 58    0.395639984 -0.5714232677  0.0004347332
#> 59    0.446786336 -0.5419434114 -0.3332303903
#> 60    0.499383528 -0.7331630692 -0.1334910426
#> 61    0.344475385 -0.5274905444  0.6101056317
#> 62    0.400145465 -0.4812204369  0.0081800740
#> 63    0.563537853 -0.3500917472  0.3018707112
#> 64    0.186100812 -0.5878458741 -0.4128969823
#> 65    0.668977955 -0.5061001163 -0.1598800581
#> 66    0.571687829 -0.3307534735  0.5053009238
#> 67    0.570551133 -0.2615798711 -0.0293143781
#> 68    0.124735792 -1.1102668548  0.1721263025
#> 69    0.559397604 -0.5631371949 -0.3651745919
#> 70    0.448009156  0.0657124452  0.9848978960
#> 71    0.182059679 -1.1209950781  0.1162298662
#> 72    0.727980101 -0.0462542945  0.5400690727
#> 73    0.270258408 -0.8266619532  0.1170530168
#> 74    0.589840997 -0.3203187294  0.6457712526
#> 75    0.157934352 -0.0973191898  0.5475919521
#> 76    0.333722615 -0.2658663496  0.0380917820
#> 77    0.354367913 -0.9760211101 -0.3965770768
#> 78    0.550265725 -0.4716657725 -0.2550980120
#> 79    0.394194662 -0.6111377422 -0.0372018787
#> 80    0.160431886 -1.3112030263 -0.7382606122
#> 81    0.658923581 -0.3769216849  0.1052376964
#> 82    0.109571725 -0.3664374184  0.0020755816
#> 83    0.462971174 -0.3149622872  0.0396652875
#> 84    0.529306643 -0.3743206723  0.4181565011
#> 85    0.543881227 -0.5529193326  0.0241608148
#> 86    0.513042742 -0.4642284421  0.4491843488
#> 87    0.664289150 -0.0347229845  0.3700102071
#> 88    0.755456785 -0.1639874446  0.5752104325
#> 89    0.629130275 -0.5402049854 -0.1894482181
#> 90    0.444385945 -0.9624915387 -0.2362988578
#> 91    0.048209593 -0.2636143574  0.3177141186
#> 92    0.488630643 -0.4538462337  0.1622991162
#> 93    0.741884405 -0.0801056016  0.3660443124
#> 94    0.585390161 -0.3987097235  0.1598145380
#> 95    0.544022070 -0.2938113123 -0.2711860654
#> 96    0.484638801 -0.8978437068 -0.6991350380
#> 97    0.482232945 -0.7426386698  0.0194133642
#> 98    0.566896273 -0.5133644287 -0.0146844362
#> 99    0.593417734  0.1206044901  1.0828226978
#> 100   0.697305898  0.0561819695  0.3960745119
#> 101   0.486787335 -0.6501975692 -0.2880300132
#> 102   0.421660034 -0.9526954031  0.0239815834
#> 103   0.588656038 -0.5169057049  0.0208087330
#> 104   0.587032743 -0.2745762829  0.1419136925
#> 105   0.366115194 -0.4385279682 -0.1131402787
#> 106   0.364448340 -1.1085341788 -0.4627449308
#> 107   0.791800723  0.1470040432  0.7080084950
#> 108  -0.379095157  0.2697899281  0.8556440320
#> 109   0.566896273 -0.5133644287 -0.0146844362
#> 110   0.589840997 -0.3203187294  0.6457712526
#> 111   0.602291189 -0.5692892350  0.2564554486
#> 112   0.571163793 -0.4266272409  0.4883139892
#> 113   0.807769993 -0.1350515975  0.3447712206
#> 114   0.447578047 -0.2745898417  0.5073900474
#> 115   0.758055055 -0.2424490791  0.5025018499
#> 116   0.384993702 -0.1097881714  1.2837382000
#> 117   0.597043246 -0.7830465159 -0.3769068732
#> 118   0.569415470 -0.7913950408 -0.3718064238
#> 119   0.512849580 -0.4042207709  0.3387078706
#> 120   0.166943205 -0.2513536255  0.6398181891
#> 121   0.394194662 -0.6111377422 -0.0372018787
#> 122   0.444148483 -0.5581483980  0.6430898841
#> 123   0.542052770 -0.4601021798 -0.0201198140
#> 124   0.563537853 -0.3500917472  0.3018707112
#> 125   0.569591158 -0.3817920733  0.0880462438
#> 126   0.727607925 -0.0372549164  0.5264118007
#> 127   0.884370843  0.0781828790  0.4375292071
#> 128   0.432710971 -0.5360011517 -0.0511097222
#> 129   0.809902560 -0.1806018011  0.2174659622
#> 130   0.542132456 -0.1788207388  0.3733685915
#> 131   0.757914237 -0.3724883329  0.2198228426
#> 132   0.498275157 -0.6704979230 -0.1934383187
#> 133   0.604528349 -0.4321340141  0.0428331941
#> 134   0.424396799 -0.6633735320  0.2722304233
#> 135   0.592093456 -0.4990264620 -0.0801327951
#> 136   0.569117012 -0.6374391360  0.2493164643
#> 137   0.402388277 -0.3007315172  0.4886160460
#> 138   0.377652685 -0.2595962943  0.2569206114
#> 139   0.727980101 -0.0462542945  0.5400690727
#> 140   0.460113477 -0.4536816053  0.1460595627
#> 141   0.149801112 -0.9103415109 -0.1679437501
#> 142   0.522242489 -0.6190266571 -0.3420570531
#> 143   0.397179960 -0.2758479441  0.4047194225
#> 144   0.706010034 -0.5395383831 -0.1371347063
#> 145   0.400557652 -0.4750610381  0.5580670301
#> 146   0.302353108 -0.7582546863 -0.0682302901
#> 147   0.403658255 -0.4257849131  0.2575612802
#> 148   0.462628862 -0.2530602524  0.2858816037
#> 149   0.541808613 -0.6000684083  0.1389382737
#> 150   0.344734195 -0.2012448661  0.3383871609
#> 151   0.516220380 -0.1651065120 -0.0104352123
#> 152   0.260856641 -0.1980997770  0.0274646997
#> 153   0.563537853 -0.3500917472  0.3018707112
#> 154   0.687980308 -0.4539346978  0.1707708211
#> 155   0.780801973 -0.2795984890  0.2919030765
#> 156   0.223389288 -0.1607757221  0.4562929453
#> 157   0.826630356 -0.2692678497  0.0111192768
#> 158   0.691329167  0.2083434963  1.1421864272
#> 159   0.406866136 -0.1703454682  0.6703948313
#> 160   0.584934781 -0.4407511257  0.0410013340
#> 161   0.386683568 -0.8437197894 -0.3912133048
#> 162   0.521189165 -0.4514797709  0.2004061009
#> 163   0.502388442 -0.4606953329  0.5471098925
#> 164   0.167285553 -0.3208535987  0.2053257346
#> 165   0.724161432 -0.0890103234  0.6746425977
#> 166   0.400300689 -0.2131130595  0.0933892319
#> 167   0.386227517 -0.6610980128  0.0090702209
#> 168   0.480030977 -0.6300468308  0.3911939269
#> 169   0.513042742 -0.4642284421  0.4491843488
#> 170   0.552353145 -0.2481787892  0.7393398953
#> 171   0.586416673 -0.0185237850  1.2926839150
#> 172   0.304680845 -1.1186782177 -0.1106829582
#> 173   0.378884674 -0.7440585602  0.0623555745
#> 174   0.099957627 -0.5997926567  0.0142305497
#> 175   0.484638801 -0.8978437068 -0.6991350380
#> 176   0.452191080 -0.3034005773  0.1233361837
#> 177   0.514768522  0.2341380232  0.6989989647
#> 178   0.680509710 -0.0302459128  0.7906722059
#> 179   0.466485089 -0.9508115351 -0.3142549560
#> 180   0.569415470 -0.7913950408 -0.3718064238
#> 181  -0.412802132 -0.4526634780  0.4045130198
#> 182   0.403658255 -0.4257849131  0.2575612802
#> 183   0.719272836  0.2691239642  0.8927730821
#> 184   0.484638801 -0.8978437068 -0.6991350380
#> 185   0.324172273 -0.2866081740 -0.1716626542
#> 186   0.394194662 -0.6111377422 -0.0372018787
#> 187   0.396075647 -0.7561146262 -0.0429173404
#> 188   0.298301218 -1.1555540882 -0.5149929689
#> 189   0.608437160 -0.0934530264  0.4815594511
#> 190   0.766745659 -0.4300152905  0.0666734868
#> 191   0.298301218 -1.1555540882 -0.5149929689
#> 192   0.563537853 -0.3500917472  0.3018707112
#> 193   0.632587199 -0.5946529223  1.1901920089
#> 194   0.586416673 -0.0185237850  1.2926839150
#> 195   0.298606147 -0.4757160998 -0.1781541660
#> 196   0.542818530 -0.5699373894  0.1812170312
#> 197   0.521500963 -0.3044637385  0.5195351015
#> 198   0.488630643 -0.4538462337  0.1622991162
#> 199   0.187715557 -0.6354828281 -0.3646734639
#> 200   0.278501013 -0.7264388227 -0.1848974671
#> 201   0.495896947 -0.1053448720  0.3978629395
#> 202   0.633048396 -0.1115090821  0.5239766374
#> 203   0.585390161 -0.3987097235  0.1598145380
#> 204   0.160431886 -1.3112030263 -0.7382606122
#> 205   0.407102222  0.0116451600  0.6891978029
#> 206  -0.345812548 -0.3373564324  0.2593010114
#> 207   0.261384365 -0.5844263451 -0.1126519819
#> 208   0.354367913 -0.9760211101 -0.3965770768
#> 209   0.628008833 -0.1418131868  0.3707491156
#> 210   0.449542499 -0.7898360603 -0.1860791807
#> 211   0.395639984 -0.5714232677  0.0004347332
#> 212   0.367363510 -0.8261789715  0.0684590229
#> 213   0.569415470 -0.7913950408 -0.3718064238
#> 214   0.648512057 -0.2211584524  0.4508400772
#> 215   0.554736342 -0.4708936751 -0.0056957926
#> 216   0.629130275 -0.5402049854 -0.1894482181
#> 217   0.515880457 -0.2247558758  0.0279651753
#> 218   0.374708932 -0.5847235765  0.5654425527
#> 219   0.687980308 -0.4539346978  0.1707708211
#> 220   0.395639984 -0.5714232677  0.0004347332
#> 221   0.359856843 -0.3716079235  0.0066602109
#> 222   0.128527322 -0.6097156026 -0.0631195186
#> 223   0.893571285 -0.1201355305  0.5100947583
#> 224   0.402342699 -0.8921693998  0.0628382948
#> 225   0.663660003 -0.2926811567  0.6285436947
#> 226   0.577073853 -0.2044886826  0.1588314977
#> 227   0.648512057 -0.2211584524  0.4508400772
#> 228   0.589840997 -0.3203187294  0.6457712526
#> 229   0.593932198 -0.3825959353  0.3971249722
#> 230   0.302353108 -0.7582546863 -0.0682302901
#> 231   0.448021767 -0.3412599001  0.2732444069
#> 232   0.600938743 -0.5026073214 -0.0059720352
#> 233   0.498275157 -0.6704979230 -0.1934383187
#> 234   0.318268915 -0.8555953352 -0.2607156715
#> 235   0.624689390 -0.3017077682  0.5670477666
#> 236   0.264698626 -0.1310576764  0.4406215781
#> 237   0.284415444  0.0732038405  0.7174814685
#> 238   0.812196970 -0.4262816725 -0.0105745840
#> 239   0.558017444 -0.4773650214 -0.1588458573
#> 240   0.733260225 -0.4721719007 -0.3536593190
#> 241   0.550899696 -0.6707786708  0.0938493587
#> 242   0.091761411  0.1448300100  0.6307357305
#> 243   0.554736342 -0.4708936751 -0.0056957926
#> 244   0.314791744 -0.6146865541  0.1606769754
#> 245   0.789123569 -0.3899204053  0.5385236616
#> 246   0.403658255 -0.4257849131  0.2575612802
#> 247   0.521664869 -0.5763595949 -0.1443784148
#> 248   0.611003873 -0.2332037752  0.2322460442
#> 249   0.499383528 -0.7331630692 -0.1334910426
#> 250   0.559397604 -0.5631371949 -0.3651745919
#> 251   0.593932198 -0.3825959353  0.3971249722
#> 252   0.592093456 -0.4990264620 -0.0801327951
#> 253   0.713156594 -0.1408535941  0.4212162729
#> 254   0.021307780 -0.4384982246  0.2512529691
#> 255   0.621705416 -0.3492717518  0.2635675448
#> 256   0.867434758 -0.1074519986  0.2304054635
#> 257   0.873149106 -0.0739471957  0.1901985784
#> 258   0.405459087 -0.7923736137  0.0395969787
#> 259   0.032755822 -0.5295844035 -0.3019848015
#> 260   0.298895168 -0.6756703509  0.1902879704
#> 261   0.565530544 -0.1693509435  0.3734731092
#> 262   0.521664869 -0.5763595949 -0.1443784148
#> 263   0.597351656  0.1773381026  0.9965601132
#> 264   0.590178242 -0.3109612017  0.3799419439
#> 265   0.453952887 -0.8932974880 -0.0905298854
#> 266   0.453952887 -0.8932974880 -0.0905298854
#> 267   0.724161432 -0.0890103234  0.6746425977
#> 268   0.675933515 -0.1409540772  0.0686679425
#> 269   0.539944608  0.0279961640  0.7313057955
#> 270   0.724992728 -0.2106123036  0.8821522758
#> 271   0.858702002 -0.0686268506  0.5127078284
#> 272   0.318655345 -0.2385467522  0.4924622699
#> 273   0.529306643 -0.3743206723  0.4181565011
#> 274   0.679049459 -0.5486697337 -0.3444988209
#> 275   0.270093869 -0.0099762462  0.4688584969
#> 276   0.446786336 -0.5419434114 -0.3332303903
#> 277   0.668977955 -0.5061001163 -0.1598800581
#> 278   0.437020110 -0.6036846412 -0.0030927977
#> 279   0.658923581 -0.3769216849  0.1052376964
#> 280   0.375669901 -0.4846522402  0.4158024710
#> 281   0.724992728 -0.2106123036  0.8821522758
#> 282   0.714498473 -0.1106244905  0.2150525929
#> 283   0.261225597 -0.4079623908  0.3584549241
#> 284   0.780801973 -0.2795984890  0.2919030765
#> 285   0.310979768 -1.0209179481 -0.6432041419
#> 286   0.541567246 -0.2878155381  0.3007953008
#> 287   0.588104711 -0.5904667392 -0.4812826139
#> 288   0.514019466 -0.3425213052  0.0996971904
#> 289   0.769100033 -0.5135234190 -0.5113616121
#> 290   0.012918650 -1.2123102711  0.2067090292
#> 291   0.503826752 -0.2697259773  0.4547930209
#> 292   0.796058291 -0.2145221806  0.1694610011
#> 293   0.486846235 -0.5554311670 -0.1479233591
#> 294   0.494241949 -0.4607942409  0.1440474026
#> 295   0.711732110 -0.3056030792  0.6822534442
#> 296   0.808619065 -0.3114833322  0.2234667000
#> 297   0.660354952 -0.4521946039 -0.1274268605
#> 298   0.457199312 -0.6436084062 -0.2340390306
#> 299   0.240611636 -0.3471458433  0.0463813862
#> 300   0.706748905 -0.3295346899  0.4838138172
#> 301   0.498275157 -0.6704979230 -0.1934383187
#> 302   0.536411464 -0.5113742527  0.1342418112
#> 303   0.579756819 -0.6510707747 -0.0183299634
#> 304   0.588826659 -0.1149252525  0.2242921027
#> 305   0.650660482 -0.4141469432  0.0096168927
#> 306   0.460396833 -0.6637654509 -0.0088167482
#> 307   0.478189623 -0.0681446872  1.4089583963
#> 308   0.296949852 -0.7669784202  0.0232640647
#> 309   0.297199199 -0.8646868443 -0.3329031552
#> 310   0.591317446 -0.2051907881  0.3334664526
#> 311   0.452508987 -0.9484063828 -0.5711543690
#> 312   0.360960845 -0.4277505288  0.2111062745
#> 313   0.648512057 -0.2211584524  0.4508400772
#> 314   0.573889669 -0.2350441968  0.4913590786
#> 315   0.498275157 -0.6704979230 -0.1934383187
#> 316   0.395639984 -0.5714232677  0.0004347332
#> 317   0.281577654 -0.2032997238  0.2908259427
#> 318   0.466289759 -0.4513723316  0.5522578848
#> 319   0.447578047 -0.2745898417  0.5073900474
#> 320   0.815651607 -0.2775730537  0.1302660986
#> 321   0.544238786 -0.8326923533 -0.1517095152
#> 322   0.558017444 -0.4773650214 -0.1588458573
#> 323   0.542428563 -0.6667332604 -0.3173515532
#> 324   0.382399820 -0.6234184837  0.0649427323
#> 325   0.595604855 -0.3656412449  0.2514484929
#> 326   0.344475385 -0.5274905444  0.6101056317
#> 327   0.282742276 -0.1922157303  0.2749915186
#> 328   0.663660003 -0.2926811567  0.6285436947
#> 329   0.261628606 -0.9707824082 -0.3855029937
#> 330   0.508591617 -0.1486903215  0.1866855039
#> 331   0.421574472 -0.7065585820  0.2942734050
#> 332   0.592093456 -0.4990264620 -0.0801327951
#> 333   0.579756819 -0.6510707747 -0.0183299634
#> 334   0.656626101 -0.3293564155  0.0112728855
#> 335   0.541567246 -0.2878155381  0.3007953008
#> 336   0.364025410 -0.5996656699  0.2279666912
#> 337   0.477779555 -0.4236360001  0.4811780805
#> 338   0.557049558 -0.5511988086  1.2072843024
#> 339   0.584944599 -0.4141430759 -0.0035569425
#> 340   0.449542499 -0.7898360603 -0.1860791807
#> 341   0.179875805  0.2633368240  0.8834828523
#> 342   0.441955896 -0.5142669272 -0.0401700450
#> 343   0.706748905 -0.3295346899  0.4838138172
#> 344   0.504020895 -0.0004260568  0.9218597557
#> 345   0.429967272 -0.4251647481  0.5267537018
#> 346   0.395639984 -0.5714232677  0.0004347332
#> 347   0.704612778 -0.4489531122 -0.0390002008
#> 348   0.542818530 -0.5699373894  0.1812170312
#> 349   0.650660482 -0.4141469432  0.0096168927
#> 350   0.457199312 -0.6436084062 -0.2340390306
#> 351   0.614588912 -0.4627811204  0.8515802491
#> 352   0.269703996 -0.3891626245  0.2880176025
#> 353   0.809902560 -0.1806018011  0.2174659622
#> 354   0.441422217 -0.5328651557 -0.2198189328
#> 355   0.713156594 -0.1408535941  0.4212162729
#> 356   0.768007309 -0.3283422682 -0.0007660867
#> 357   0.462628862 -0.2530602524  0.2858816037
#> 358   0.353669336 -1.0269027305 -0.3949091167
#> 359   0.386683568 -0.8437197894 -0.3912133048
#> 360   0.193783245 -0.5368149649  0.2669710401
#> 361   0.836734787  0.0336372295  0.4828950624
#> 362   0.042922599 -0.4251651216  0.2657488512
#> 363   0.202011958 -0.2378760020  0.1965793164
#> 364   0.521189165 -0.4514797709  0.2004061009
#> 365   0.444148483 -0.5581483980  0.6430898841
#> 366   0.584934781 -0.4407511257  0.0410013340
#> 367   0.838791064 -0.2868096336  0.2990962483
#> 368   0.364448340 -1.1085341788 -0.4627449308
#> 369   0.653411238 -0.2984701811 -0.1693354909
#> 370   0.600938743 -0.5026073214 -0.0059720352
#> 371   0.382399820 -0.6234184837  0.0649427323
#> 372   0.486787335 -0.6501975692 -0.2880300132
#> 373   0.450137602 -0.5153667124 -0.1882298509
#> 374   0.182818647 -0.0973515022  0.4875487379
#> 375   0.133457689 -0.5622361794 -0.0660297392
#> 376   0.437020110 -0.6036846412 -0.0030927977
#> 377   0.298301218 -1.1555540882 -0.5149929689
#> 378   0.386683568 -0.8437197894 -0.3912133048
#> 379   0.565367767  0.0901305729  0.7169050908
#> 380   0.591317446 -0.2051907881  0.3334664526
#> 381   0.552120576 -0.3701135903  0.3245256668
#> 382   0.808729406 -0.1045582280  0.1692706355
#> 383   0.302353108 -0.7582546863 -0.0682302901
#> 384   0.179952822 -0.5172573826 -0.0807081445
#> 385   0.463833254 -0.6228419978 -0.0508452681
#> 386   0.344578398 -0.7965332972  0.0902728406
#> 387   0.544238786 -0.8326923533 -0.1517095152
#> 388   0.441422217 -0.5328651557 -0.2198189328
#> 389   0.329097297  0.0370855930  0.1570554230
#> 390   0.392172068 -0.6164790418 -0.0281440488
#> 391   0.691767235 -0.5352847855  0.1986682333
#> 392   0.663660003 -0.2926811567  0.6285436947
#> 393   0.550265725 -0.4716657725 -0.2550980120
#> 394   0.366554475 -0.9721809197 -0.4617619633
#> 395   0.498275157 -0.6704979230 -0.1934383187
#> 396   0.521189165 -0.4514797709  0.2004061009
#> 397   0.556572056 -0.3197255684  0.9483024757
#> 398   0.543881227 -0.5529193326  0.0241608148
#> 399   0.773443259 -0.2858532590  0.4119181432
#> 400   0.631427937 -0.2142573516  0.1765493076
#> 401   0.451375676 -0.4688628715  0.5266780721
#> 402   0.447578047 -0.2745898417  0.5073900474
#> 403   0.410026903 -0.3038550275  0.5294091837
#> 404   0.621705416 -0.3492717518  0.2635675448
#> 405   0.405459087 -0.7923736137  0.0395969787
#> 406   0.652515936 -0.1546629968  0.2489787199
#> 407   0.638662992 -0.3033193426  0.3528740976
#> 408   0.240750786 -1.2098253068 -0.5553506133
#> 409   0.635355036 -0.3471245633  0.7306908166
#> 410   0.542428563 -0.6667332604 -0.3173515532
#> 411   0.544238786 -0.8326923533 -0.1517095152
#> 412   0.245478019 -0.5697965158  0.2936362696
#> 413   0.668977955 -0.5061001163 -0.1598800581
#> 414   0.298301218 -1.1555540882 -0.5149929689
#> 415   0.727607925 -0.0372549164  0.5264118007
#> 416   0.618308717 -0.3918503994  0.5864440374
#> 417   0.348721967 -0.5893384893  0.1169609962
#> 418   0.704810536 -0.3294620488  0.4223278561
#> 419   0.424396799 -0.6633735320  0.2722304233
#> 420   0.773443259 -0.2858532590  0.4119181432
#> 421   0.331055123 -0.5861374618 -0.2775677326
#> 422   0.836250350 -0.1445569324  0.1000264984
#> 423   0.108527700 -0.9490191349 -0.2134837368
#> 424   0.453840352 -0.5634452174 -0.1806463250
#> 425   0.589840997 -0.3203187294  0.6457712526
#> 426   0.446334502 -0.5156434434 -0.0662054251
#> 427   0.701955085 -0.4560139529  0.0115884048
#> 428   0.609246554 -0.3893298055  0.0654004387
#> 429   0.369571213 -1.0343569824 -0.3503442551
#> 430   0.706748905 -0.3295346899  0.4838138172
#> 431   0.486553191 -0.2965787723  0.4945195023
#> 432   0.410282924 -0.6923881671 -0.4568291723
#> 433   0.424396799 -0.6633735320  0.2722304233
#> 434   0.540553067 -0.3591232265  0.3670039093
#> 435   0.590178242 -0.3109612017  0.3799419439
#> 436   0.363069606 -0.4513288438  0.6163470511
#> 437   0.434756462 -0.2949811453  0.5193829252
#> 438   0.643090264 -0.1224253158  0.7088964375
#> 439   0.397179960 -0.2758479441  0.4047194225
#> 440   0.873149106 -0.0739471957  0.1901985784
#> 441   0.796058291 -0.2145221806  0.1694610011
#> 442   0.396075647 -0.7561146262 -0.0429173404
#> 443   0.136031997 -0.5959755586  0.2349124885
#> 444   0.382018456 -0.2403798856  0.6766867459
#> 445   0.415488014 -0.6537340249  0.0255461761
#> 446   0.618308717 -0.3918503994  0.5864440374
#> 447   0.588741992 -0.7404371301 -0.3476323550
#> 448   0.313986569 -0.4729355204  0.0689678148
#> 449   0.679049459 -0.5486697337 -0.3444988209
#> 450   0.714496813 -0.2328782589  0.1189024970
#> 451   0.673220966 -0.1234234906  0.5435343198
#> 452   0.009530784 -0.3313339626  0.5633108527
#> 453   0.792933969  0.2799987243  1.2888029452
#> 454   0.656626101 -0.3293564155  0.0112728855
#> 455   0.218072485 -1.0527192740  0.2849819283
#> 456   0.844048634  0.1270731282  0.8969097928
#> 457   0.553367193 -0.6268295178  0.3825623503
#> 458   0.408364928 -0.9246746424  0.2278718868
#> 459   0.395639984 -0.5714232677  0.0004347332
#> 460   0.566896273 -0.5133644287 -0.0146844362
#> 461   0.727607925 -0.0372549164  0.5264118007
#> 462   0.714496813 -0.2328782589  0.1189024970
#> 463   0.720795395  0.0630292625  0.5947122737
#> 464   0.499110859 -0.4836553085  0.0179271125
#> 465   0.363932453 -0.3601226167 -0.0327885035
#> 466   0.499383528 -0.7331630692 -0.1334910426
#> 467   0.542818530 -0.5699373894  0.1812170312
#> 468   0.780801973 -0.2795984890  0.2919030765
#> 469   0.660354952 -0.4521946039 -0.1274268605
#> 470   0.548272924 -0.5718551112 -0.3482535258
#> 471   0.758055055 -0.2424490791  0.5025018499
#> 472   0.521189165 -0.4514797709  0.2004061009
#> 473   0.934462819  0.1715532271  0.4943677148
#> 474   0.588741992 -0.7404371301 -0.3476323550
#> 475   0.687980308 -0.4539346978  0.1707708211
#> 476   0.643090264 -0.1224253158  0.7088964375
#> 477   0.018915285 -0.2076773256  0.3875913029
#> 478   0.691767235 -0.5352847855  0.1986682333
#> 479   0.643276313 -0.0241843890  1.0574182227
#> 480   0.701955085 -0.4560139529  0.0115884048
#> 481   0.364448340 -1.1085341788 -0.4627449308
#> 482   0.314438036 -0.2999454154 -0.0459163665
#> 483   0.366115194 -0.4385279682 -0.1131402787
#> 484   0.553367193 -0.6268295178  0.3825623503
#> 485   0.614588912 -0.4627811204  0.8515802491
#> 486   0.882912562  0.0586557573  0.7712869237
#> 487   0.435179500 -0.3416570476  0.0533392852
#> 488   0.459494128 -0.7128751448 -0.2296207800
#> 489   0.542052770 -0.4601021798 -0.0201198140
#> 490   0.486553191 -0.2965787723  0.4945195023
#> 491   0.360714337 -0.3388716450  0.3245551908
#> 492   0.486426425 -0.1376061114  0.1538895114
#> 493   0.632612665 -0.6028117828 -0.0274301192
#> 494   0.099417710 -0.1339114830  0.3136005574
#> 495   0.300368938 -0.1618114518  0.3664226723
#> 496   0.569591158 -0.3817920733  0.0880462438
#> 497   0.361151591 -0.5993766917  0.1199718797
#> 498   0.581180751 -0.4126989700  0.3571307678
#> 499   0.873149106 -0.0739471957  0.1901985784
#> 500   0.392172068 -0.6164790418 -0.0281440488
#> 501   0.359916240 -0.5653373117  0.0918243603
#> 502   0.540553067 -0.3591232265  0.3670039093
#> 503   0.408097323 -0.8201777963  0.3075983514
#> 504   0.513042742 -0.4642284421  0.4491843488
#> 505   0.569092639 -0.5869084849  0.0628484541
#> 506   0.585390161 -0.3987097235  0.1598145380
#> 507   0.401534066 -0.1475975926  0.8573138449
#> 508   0.240750786 -1.2098253068 -0.5553506133
#> 509   0.559618208  0.0969165429  0.6461050627
#> 510   0.400559721 -0.9644068997 -0.7568555413
#> 511   0.486801754 -0.7019084695  0.2632276767
#> 512   0.438338528 -0.2417307052  0.5946369339
#> 513   0.450694918 -0.0883904142  0.8320005792
#> 514   0.486846235 -0.5554311670 -0.1479233591
#> 515   0.622449836 -0.5019979967  0.0351213020
#> 516   0.235255910 -0.0210490942  0.2767133637
#> 517   0.627959340 -0.1539228173  0.5864260334
#> 518   0.460396833 -0.6637654509 -0.0088167482
#> 519   0.250153897 -0.1219461299  0.0415585688
#> 520   0.929674808  0.2917008018  0.8324205624
#> 521   0.648512057 -0.2211584524  0.4508400772
#> 522   0.604745434 -0.8005475060 -0.2683964712
#> 523   0.297199199 -0.8646868443 -0.3329031552
#> 524  -0.240618730 -0.4128390292  0.6684769380
#> 525   0.281303679 -0.2594450506  0.3225061534
#> 526   0.522242489 -0.6190266571 -0.3420570531
#> 527   0.569591158 -0.3817920733  0.0880462438
#> 528   0.353669336 -1.0269027305 -0.3949091167
#> 529   0.554060464 -0.4976613537  0.2498930324
#> 530   0.542052770 -0.4601021798 -0.0201198140
#> 531   0.638662992 -0.3033193426  0.3528740976
#> 532   0.600938743 -0.5026073214 -0.0059720352
#> 533   0.706748905 -0.3295346899  0.4838138172
#> 534   0.253398915  0.0943849032  0.5764948943
#> 535   0.310509990 -0.1811916914  0.2524410902
#> 536   0.552120576 -0.3701135903  0.3245256668
#> 537   0.444385945 -0.9624915387 -0.2362988578
#> 538   0.586774382 -0.6251041982  0.0957501925
#> 539   0.632628302 -0.5861815637 -0.3758122409
#> 540   0.448021767 -0.3412599001  0.2732444069
#> 541   0.403658255 -0.4257849131  0.2575612802
#> 542   0.648512057 -0.2211584524  0.4508400772
#> 543   0.536411464 -0.5113742527  0.1342418112
#> 544   0.449542499 -0.7898360603 -0.1860791807
#> 545   0.724161432 -0.0890103234  0.6746425977
#> 546   0.597043246 -0.7830465159 -0.3769068732
#> 547   0.206943062 -0.3044303884  0.7843782380
#> 548   0.466289759 -0.4513723316  0.5522578848
#> 549   0.563537853 -0.3500917472  0.3018707112
#> 550   0.255541802 -0.3937043197 -0.0264192578
#> 551   0.091761411  0.1448300100  0.6307357305
#> 552   0.434756462 -0.2949811453  0.5193829252
#> 553   0.663660003 -0.2926811567  0.6285436947
#> 554   0.529890608 -0.0396235483  0.9990838986
#> 555   0.704612778 -0.4489531122 -0.0390002008
#> 556   0.638662992 -0.3033193426  0.3528740976
#> 557   0.561106944 -0.2651440725 -0.1287217606
#> 558   0.338651534 -0.5082220021  0.3076550409
#> 559   0.500438947 -0.2832233100  0.2268114943
#> 560   0.529890608 -0.0396235483  0.9990838986
#> 561   0.593932198 -0.3825959353  0.3971249722
#> 562   0.013508336 -0.3375600545 -0.0609053952
#> 563   0.706748905 -0.3295346899  0.4838138172
#> 564   0.298301218 -1.1555540882 -0.5149929689
#> 565   0.664874009 -0.1513899541  0.0675273210
#> 566   0.755456785 -0.1639874446  0.5752104325
#> 567   0.591317446 -0.2051907881  0.3334664526
#> 568   0.858702002 -0.0686268506  0.5127078284
#> 569   0.354367913 -0.9760211101 -0.3965770768
#> 570   0.741558840 -0.2541580256  0.2724238592
#> 571   0.405459087 -0.7923736137  0.0395969787
#> 572   0.541808613 -0.6000684083  0.1389382737
#> 573   0.266616307 -0.3288993350  0.8950116004
#> 574   0.434756462 -0.2949811453  0.5193829252
#> 575   0.574832889 -0.6533272987  0.2257753258
#> 576   0.529306643 -0.3743206723  0.4181565011
#> 577   0.704810536 -0.3294620488  0.4223278561
#> 578   0.834281011 -0.3297252034  0.3119704345
#> 579  -0.018783454 -0.2198046679  0.1903099059
#> 580   0.436688811 -0.2186366074  0.0537701396
#> 581   0.586416673 -0.0185237850  1.2926839150
#> 582   0.630642307 -0.5098421970  0.6534326113
#> 583   0.694848160 -0.0560723167  0.1604085117
#> 584   0.457199312 -0.6436084062 -0.2340390306
#> 585   0.517137951 -0.5002043838  0.5479908117
#> 586   0.597043246 -0.7830465159 -0.3769068732
#> 587   0.858702002 -0.0686268506  0.5127078284
#> 588   0.051956628 -0.2562357204  0.3983137810
#> 589   0.542052770 -0.4601021798 -0.0201198140
#> 590   0.450137602 -0.5153667124 -0.1882298509
#> 591   0.563289790 -0.0318025348  0.4153047836
#> 592   0.372501214 -0.4039336371  0.0155478578
#> 593   0.540553067 -0.3591232265  0.3670039093
#> 594   0.635355036 -0.3471245633  0.7306908166
#> 595   0.867434758 -0.1074519986  0.2304054635
#> 596   0.340841154 -1.0988306712 -0.2963597973
#> 597  -0.020066077  0.2068705635  0.6483892005
#> 598   0.390008380 -0.4004064154  0.1896346814
#> 599   0.341643317 -0.3330784616  0.5507623850
#> 600   0.652307087 -0.1377703809  0.4859756325
#> 601   0.663660003 -0.2926811567  0.6285436947
#> 602   0.258556178 -0.0976880411  1.0146620278
#> 603   0.483574727 -0.2894631592  0.4972823950
#> 604   0.452508987 -0.9484063828 -0.5711543690
#> 605   0.652515936 -0.1546629968  0.2489787199
#> 606   0.292057639 -0.3436674677 -0.0397240688
#> 607   0.649233058 -0.1161226482  0.0172150003
#> 608   0.595604855 -0.3656412449  0.2514484929
#> 609   0.451780299 -0.4627529046  0.0947918787
#> 610   0.353669336 -1.0269027305 -0.3949091167
#> 611   0.412357844 -0.8383649442 -0.4285191230
#> 612   0.807769993 -0.1350515975  0.3447712206
#> 613   0.764431198 -0.5755929307 -0.2848130735
#> 614   0.364448340 -1.1085341788 -0.4627449308
#> 615   0.935223910  0.1162423892  0.8157133323
#> 616   0.678189730 -0.2316035839 -0.0414137216
#> 617   0.085265139 -0.1168571611  0.6212157275
#> 618   0.571163793 -0.4266272409  0.4883139892
#> 619   0.438338528 -0.2417307052  0.5946369339
#> 620   0.609246554 -0.3893298055  0.0654004387
#> 621   0.456398098 -0.6442839728  0.0471418818
#> 622   0.499383528 -0.7331630692 -0.1334910426
#> 623   0.691767235 -0.5352847855  0.1986682333
#> 624   0.451780299 -0.4627529046  0.0947918787
#> 625   0.210939443 -0.3902984368  0.1635562787
#> 626   0.156737680 -0.7142724174  0.0386033424
#> 627   0.313628471 -0.8657323384  0.9952823631
#> 628   0.423671625 -0.1842146445  0.2901039616
#> 629   0.727607925 -0.0372549164  0.5264118007
#> 630   0.773443259 -0.2858532590  0.4119181432
#> 631   0.556572056 -0.3197255684  0.9483024757
#> 632   0.513787632 -0.0712591239  0.2674656834
#> 633   0.529890608 -0.0396235483  0.9990838986
#> 634   0.601223705 -0.2833174462  0.6968434617
#> 635   0.537731396 -0.3006104905  0.6336308732
#> 636   0.486267878 -0.2827264815  0.1901964713
#> 637   0.704810536 -0.3294620488  0.4223278561
#> 638   0.397131888 -0.1211933846  0.6072669564
#> 639   0.358049155 -0.3018006509  0.1197539342
#> 640   0.514687901 -0.5470952894 -0.3393160236
#> 641   0.447134252 -0.4377760119  0.4430582142
#> 642   0.572624930 -0.6301516180 -0.0623746439
#> 643   0.724992728 -0.2106123036  0.8821522758
#> 644   0.386227517 -0.6610980128  0.0090702209
#> 645   0.484059109 -0.3199501902  0.4145387516
#> 646   0.714496813 -0.2328782589  0.1189024970
#> 647   0.427273405 -0.5630162178 -0.3396310510
#> 648   0.093012727  0.0971846718  0.5385751174
#> 649   0.521189165 -0.4514797709  0.2004061009
#> 650   0.353669336 -1.0269027305 -0.3949091167
#> 651   0.360269660 -0.5950481484  0.5716799636
#> 652   0.826630356 -0.2692678497  0.0111192768
#> 653   0.754105295 -0.5030714788 -0.2258674569
#> 654   0.401948194 -0.6187516011  0.0233057965
#> 655   0.366554475 -0.9721809197 -0.4617619633
#> 656   0.529306643 -0.3743206723  0.4181565011
#> 657   0.486553191 -0.2965787723  0.4945195023
#> 658   0.657419274 -0.3488013624  0.5891294459
#> 659   0.366554475 -0.9721809197 -0.4617619633
#> 660   0.536411464 -0.5113742527  0.1342418112
#> 661   0.443076873 -0.3785672921  0.3554430483
#> 662   0.236514661 -0.7092475217 -0.2537887177
#> 663   0.498275157 -0.6704979230 -0.1934383187
#> 664   0.600938743 -0.5026073214 -0.0059720352
#> 665   0.288322284 -0.4404311094 -0.0067611906
#> 666   0.312319976 -0.4495775828  0.5106801800
#> 667   0.554060464 -0.4976613537  0.2498930324
#> 668   0.172675146 -0.9510944879 -0.6065510963
#> 669   0.265090446 -0.1454238284  0.3059188295
#> 670   0.622449836 -0.5019979967  0.0351213020
#> 671   0.858702002 -0.0686268506  0.5127078284
#> 672   0.563537853 -0.3500917472  0.3018707112
#> 673   0.660354952 -0.4521946039 -0.1274268605
#> 674   0.555314143 -0.3501426373  0.1942458579
#> 675   0.254021228 -0.5573125475  0.4123297599
#> 676   0.569092639 -0.5869084849  0.0628484541
#> 677   0.844048634  0.1270731282  0.8969097928
#> 678   0.933871249  0.2157023493  0.7861310464
#> 679   0.611003873 -0.2332037752  0.2322460442
#> 680   0.392458850 -0.8258682770 -0.4355652497
#> 681   0.447134252 -0.4377760119  0.4430582142
#> 682   0.550899696 -0.6707786708  0.0938493587
#> 683   0.482232945 -0.7426386698  0.0194133642
#> 684   0.516315594 -0.3113020522  0.1606688764
#> 685   0.333722615 -0.2658663496  0.0380917820
#> 686   0.455562173 -0.2367982811  0.4565294905
#> 687   0.757914237 -0.3724883329  0.2198228426
#> 688   0.528856250 -0.5733747025 -0.3465269238
#> 689   0.259183463 -0.3376072332  0.2623426953
#> 690   0.580136735 -0.3234166739  0.1084793721
#> 691   0.791800723  0.1470040432  0.7080084950
#> 692   0.557285587 -0.3287940440  0.2305388943
#> 693   0.933871249  0.2157023493  0.7861310464
#> 694   0.354367913 -0.9760211101 -0.3965770768
#> 695   0.440517875 -0.0330139559  0.4145045507
#> 696   0.297880828 -0.2400131613  0.2132650627
#> 697   0.577073853 -0.2044886826  0.1588314977
#> 698   0.443586091 -0.4248446569  0.0604056929
#> 699   0.668977955 -0.5061001163 -0.1598800581
#> 700   0.569117012 -0.6374391360  0.2493164643
#> 701   0.335003113 -0.8196334371 -0.1115912792
#> 702   0.581180751 -0.4126989700  0.3571307678
#> 703   0.542428563 -0.6667332604 -0.3173515532
#> 704   0.562892621 -0.8761745702 -0.6343268158
#> 705   0.482232945 -0.7426386698  0.0194133642
#> 706   0.429967272 -0.4251647481  0.5267537018
#> 707   0.363069606 -0.4513288438  0.6163470511
#> 708   0.514768522  0.2341380232  0.6989989647
#> 709   0.195581271 -0.2338393460  0.1526031619
#> 710   0.450694918 -0.0883904142  0.8320005792
#> 711   0.486801754 -0.7019084695  0.2632276767
#> 712   0.548041165 -0.4480862608 -0.5182729564
#> 713   0.446334502 -0.5156434434 -0.0662054251
#> 714   0.588741992 -0.7404371301 -0.3476323550
#> 715   0.724161432 -0.0890103234  0.6746425977
#> 716   0.563537853 -0.3500917472  0.3018707112
#> 717   0.425319618 -0.5607598104  0.0417204059
#> 718   0.569415470 -0.7913950408 -0.3718064238
#> 719   0.592093456 -0.4990264620 -0.0801327951
#> 720   0.371305938 -0.4929479991  0.1454229401
#> 721   0.363932453 -0.3601226167 -0.0327885035
#> 722   0.768007309 -0.3283422682 -0.0007660867
#> 723   0.638662992 -0.3033193426  0.3528740976
#> 724   0.552120576 -0.3701135903  0.3245256668
#> 725   0.399347782 -0.4367659281  0.0315375053
#> 726   0.437821884 -0.6131341922  0.0524423957
#> 727   0.443586091 -0.4248446569  0.0604056929
#> 728   0.353669336 -1.0269027305 -0.3949091167
#> 729   0.448952514 -0.4029141780  0.1794318409
#> 730   0.632612665 -0.6028117828 -0.0274301192
#> 731   0.717497661 -0.4044302892 -0.1407585287
#> 732   0.130720075 -0.6374456098 -0.2450052397
#> 733   0.845866759 -0.1586343789  0.6107993764
#> 734   0.716600278 -0.3790916861  0.4514476321
#> 735   0.650289973 -0.0401727452  0.6569365882
#> 736  -0.221797172  0.3012736458  1.0065592300
#> 737   0.741884405 -0.0801056016  0.3660443124
#> 738   0.642113142  0.1537617188  1.0381122872
#> 739   0.379271363 -0.4048623941 -0.1545996683
#> 740   0.850498335  0.0779450335  0.1915156398
#> 741   0.515096401  0.0495789039  0.4672317721
#> 742   0.768007309 -0.3283422682 -0.0007660867
#> 743   0.571687829 -0.3307534735  0.5053009238
#> 744   0.181623640 -0.3603470042  0.2306064635
#> 745   0.459494128 -0.7128751448 -0.2296207800
#> 746   0.716600278 -0.3790916861  0.4514476321
#> 747   0.595604855 -0.3656412449  0.2514484929
#> 748   0.536411464 -0.5113742527  0.1342418112
#> 749   0.355185228 -0.4406251316  0.2745621284
#> 750   0.438338528 -0.2417307052  0.5946369339
#> 751   0.219615216 -0.9182048920 -0.1531389779
#> 752   0.690292934 -0.5636993852  0.3092870884
#> 753   0.494241949 -0.4607942409  0.1440474026
#> 754   0.447810926 -0.2774332874  1.1496500079
#> 755   0.558017444 -0.4773650214 -0.1588458573
#> 756   0.609246554 -0.3893298055  0.0654004387
#> 757   0.419746758 -0.2546931348  0.2135114404
#> 758   0.563537853 -0.3500917472  0.3018707112
#> 759   0.450494498 -0.1665391667  0.4480682948
#> 760   0.668977955 -0.5061001163 -0.1598800581
#> 761   0.136790291  0.1269625528  0.6552873403
#> 762   0.652307087 -0.1377703809  0.4859756325
#> 763   0.479394220 -0.4666036749  0.1511681576
#> 764   0.858702002 -0.0686268506  0.5127078284
#> 765   0.513787632 -0.0712591239  0.2674656834
#> 766   0.367363510 -0.8261789715  0.0684590229
#> 767  -0.209890573 -0.2027929050  0.4185847248
#> 768   0.554736342 -0.4708936751 -0.0056957926
#> 769   0.187715557 -0.6354828281 -0.3646734639
#> 770   0.658923581 -0.3769216849  0.1052376964
#> 771   0.655188927  0.2265560192  0.4696825713
#> 772   0.577739322 -0.3248873953 -0.0047497988
#> 773   0.364334989 -1.0451425997 -0.1199067403
#> 774   0.592093456 -0.4990264620 -0.0801327951
#> 775  -0.128998585 -0.4572952970 -0.0769356484
#> 776   0.584934781 -0.4407511257  0.0410013340
#> 777   0.763505632  0.0676625613  0.6335338777
#> 778  -0.096519135 -0.3134233670  0.2769869269
#> 779   0.515880457 -0.2247558758  0.0279651753
#> 780   0.547429445  0.0139601916  0.3435441037
#> 781   0.727980101 -0.0462542945  0.5400690727
#> 782   0.625986463 -0.5755168690 -0.1160674525
#> 783   0.366115194 -0.4385279682 -0.1131402787
#> 784   0.078925294 -0.6508200928 -0.0410289236
#> 785   0.354367913 -0.9760211101 -0.3965770768
#> 786   0.090918438 -0.2973578699 -0.0618514902
#> 787   0.265090446 -0.1454238284  0.3059188295
#> 788   0.211536405 -0.2496912942  0.0454889841
#> 789   0.456398098 -0.6442839728  0.0471418818
#> 790   0.466485089 -0.9508115351 -0.3142549560
#> 791   0.711732110 -0.3056030792  0.6822534442
#> 792   0.201371857 -1.0477175201 -0.0260467305
#> 793   0.503748003 -0.2427530453  0.1341168224
#> 794   0.423671625 -0.1842146445  0.2901039616
#> 795   0.668977955 -0.5061001163 -0.1598800581
#> 796   0.768007309 -0.3283422682 -0.0007660867
#> 797   0.648512057 -0.2211584524  0.4508400772
#> 798   0.579756819 -0.6510707747 -0.0183299634
#> 799   0.365188149 -0.1582682728  0.1696951435
#> 800   0.438280058 -0.4125527981 -0.0038865081
#> 801   0.622449836 -0.5019979967  0.0351213020
#> 802   0.324542016 -0.2952238947  0.2761565458
#> 803   0.569092639 -0.5869084849  0.0628484541
#> 804   0.597043246 -0.7830465159 -0.3769068732
#> 805   0.477779555 -0.4236360001  0.4811780805
#> 806   0.557285587 -0.3287940440  0.2305388943
#> 807   0.569415470 -0.7913950408 -0.3718064238
#> 808   0.337485800 -0.3559755956  0.2240841632
#> 809   0.246642297 -0.2325949598  0.0585347467
#> 810   0.718349040 -0.3781103410 -0.3327550778
#> 811   0.631615614 -0.1712318320  0.2731492912
#> 812   0.724161432 -0.0890103234  0.6746425977
#> 813   0.758055055 -0.2424490791  0.5025018499
#> 814   0.796058291 -0.2145221806  0.1694610011
#> 815   0.650660482 -0.4141469432  0.0096168927
#> 816   0.695374776 -0.5278972223 -0.3879989097
#> 817   0.270258408 -0.8266619532  0.1170530168
#> 818   0.160431886 -1.3112030263 -0.7382606122
#> 819   0.528856250 -0.5733747025 -0.3465269238
#> 820   0.304680845 -1.1186782177 -0.1106829582
#> 821   0.479921632  0.0848845864  0.3108023595
#> 822   0.269108104  0.2640875813  0.7807098954
#> 823   0.836734787  0.0336372295  0.4828950624
#> 824   0.494241949 -0.4607942409  0.1440474026
#> 825   0.687980308 -0.4539346978  0.1707708211
#> 826   0.437821884 -0.6131341922  0.0524423957
#> 827   0.450494498 -0.1665391667  0.4480682948
#> 828   0.514687901 -0.5470952894 -0.3393160236
#> 829   0.607864040  0.1658584534  0.5683152161
#> 830   0.452508987 -0.9484063828 -0.5711543690
#> 831   0.629130275 -0.5402049854 -0.1894482181
#> 832   0.638662992 -0.3033193426  0.3528740976
#> 833   0.584944599 -0.4141430759 -0.0035569425
#> 834   0.543881227 -0.5529193326  0.0241608148
#> 835   0.145352556 -0.9293615395 -0.1570871679
#> 836   0.724161432 -0.0890103234  0.6746425977
#> 837   0.565367767  0.0901305729  0.7169050908
#> 838   0.547429445  0.0139601916  0.3435441037
#> 839   0.169222134 -0.2269181335  0.3555828598
#> 840   0.301823474 -0.8851777748  0.0297867884
#> 841   0.595604855 -0.3656412449  0.2514484929
#> 842   0.125069867 -0.5662798862 -0.3380245012
#> 843   0.628008833 -0.1418131868  0.3707491156
#> 844   0.550265725 -0.4716657725 -0.2550980120
#> 845   0.499383528 -0.7331630692 -0.1334910426
#> 846   0.536411464 -0.5113742527  0.1342418112
#> 847   0.129985749 -0.5022277011  0.1771852997
#> 848   0.380329283 -0.0934469640  1.2956808876
#> 849   0.395639984 -0.5714232677  0.0004347332
#> 850   0.704612778 -0.4489531122 -0.0390002008
#> 851   0.682424058 -0.4163177234  0.5446231618
#> 852   0.585390161 -0.3987097235  0.1598145380
#> 853   0.457199312 -0.6436084062 -0.2340390306
#> 854   0.516315594 -0.3113020522  0.1606688764
#> 855   0.511641784 -0.7677663540  0.0853658754
#> 856   0.410282924 -0.6923881671 -0.4568291723
#> 857   0.437020110 -0.6036846412 -0.0030927977
#> 858   0.757914237 -0.3724883329  0.2198228426
#> 859   0.520103329 -0.1785803611  0.3734570491
#> 860   0.550244646 -0.1956952838  0.1953012935
#> 861   0.293482709 -0.5437693803  0.1100294752
#> 862   0.542132456 -0.1788207388  0.3733685915
#> 863   0.359856843 -0.3716079235  0.0066602109
#> 864   0.845866759 -0.1586343789  0.6107993764
#> 865   0.399549796 -0.8666931197 -0.7884115090
#> 866   0.593932198 -0.3825959353  0.3971249722
#> 867   0.457058483 -0.7345171270  0.7537609799
#> 868   0.410383009 -0.2762660620  0.0059017656
#> 869   0.415488014 -0.6537340249  0.0255461761
#> 870   0.184698758 -0.4878075653  0.4617543412
#> 871   0.475347953 -0.7431585646 -0.2843218318
#> 872   0.621171872 -0.7045623815 -0.2368098092
#> 873   0.474201912 -0.2947185718  0.1579737626
#> 874   0.663660003 -0.2926811567  0.6285436947
#> 875   0.504020895 -0.0004260568  0.9218597557
#> 876   0.768007309 -0.3283422682 -0.0007660867
#> 877   0.466485089 -0.9508115351 -0.3142549560
#> 878   0.446334502 -0.5156434434 -0.0662054251
#> 879   0.590505456 -0.3871348954 -0.3718855280
#> 880   0.453952887 -0.8932974880 -0.0905298854
#> 881   0.526571085 -0.0523977223  0.7434825107
#> 882   0.562892621 -0.8761745702 -0.6343268158
#> 883   0.313005636 -0.9207090591 -0.5304953386
#> 884   0.449542499 -0.7898360603 -0.1860791807
#> 885   0.516053890 -0.8756037462 -0.9089821180
#> 886   0.486217076 -0.3543872491  0.2370284288
#> 887   0.597043246 -0.7830465159 -0.3769068732
#> 888   0.568549653 -0.7681406867  0.2856906639
#> 889   0.625986463 -0.5755168690 -0.1160674525
#> 890   0.419746758 -0.2546931348  0.2135114404
#> 891   0.453840352 -0.5634452174 -0.1806463250
#> 892   0.451238365 -0.4983855518  0.6889484887
#> 893   0.395639984 -0.5714232677  0.0004347332
#> 894   0.297199199 -0.8646868443 -0.3329031552
#> 895   0.571687829 -0.3307534735  0.5053009238
#> 896   0.313582678 -0.7320930879 -0.2032745923
#> 897   0.454896680 -0.5000603336  0.1240872561
#> 898   0.754965357 -0.3855683143 -0.1171646751
#> 899   0.386227517 -0.6610980128  0.0090702209
#> 900   0.597043246 -0.7830465159 -0.3769068732
#> 901   0.156968619 -0.5023714479  0.2491663292
#> 902   0.486787335 -0.6501975692 -0.2880300132
#> 903   0.444148483 -0.5581483980  0.6430898841
#> 904   0.316735531 -0.0608607610  0.7458103442
#> 905   0.624689390 -0.3017077682  0.5670477666
#> 906   0.710554541 -0.3434791388 -0.2766488795
#> 907   0.541567246 -0.2878155381  0.3007953008
#> 908   0.706748905 -0.3295346899  0.4838138172
#> 909   0.632628302 -0.5861815637 -0.3758122409
#> 910   0.608437160 -0.0934530264  0.4815594511
#> 911   0.520103329 -0.1785803611  0.3734570491
#> 912   0.604528349 -0.4321340141  0.0428331941
#> 913   0.515736645 -0.2278935864 -0.0489118941
#> 914   0.364448340 -1.1085341788 -0.4627449308
#> 915   0.565530544 -0.1693509435  0.3734731092
#> 916   0.453840352 -0.5634452174 -0.1806463250
#> 917   0.488630643 -0.4538462337  0.1622991162
#> 918   0.609246554 -0.3893298055  0.0654004387
#> 919   0.510452205  0.2809590179  1.0082867331
#> 920   0.494241949 -0.4607942409  0.1440474026
#> 921   0.609246554 -0.3893298055  0.0654004387
#> 922   0.590178242 -0.3109612017  0.3799419439
#> 923   0.590178242 -0.3109612017  0.3799419439
#> 924   0.395639984 -0.5714232677  0.0004347332
#> 925   0.484638801 -0.8978437068 -0.6991350380
#> 926   0.574832889 -0.6533272987  0.2257753258
#> 927   0.584934781 -0.4407511257  0.0410013340
#> 928   0.858702002 -0.0686268506  0.5127078284
#> 929   0.378447848 -0.2448526714  0.3504344008
#> 930   0.486553191 -0.2965787723  0.4945195023
#> 931   0.460396833 -0.6637654509 -0.0088167482
#> 932   0.412357844 -0.8383649442 -0.4285191230
#> 933   0.421339200 -0.2365788490  0.5306513054
#> 934   0.572624930 -0.6301516180 -0.0623746439
#> 935   0.488630643 -0.4538462337  0.1622991162
#> 936   0.438280058 -0.4125527981 -0.0038865081
#> 937   0.201371857 -1.0477175201 -0.0260467305
#> 938   0.463833254 -0.6228419978 -0.0508452681
#> 939   0.581180751 -0.4126989700  0.3571307678
#> 940   0.406866136 -0.1703454682  0.6703948313
#> 941   0.717497661 -0.4044302892 -0.1407585287
#> 942   0.361151591 -0.5993766917  0.1199718797
#> 943   0.768007309 -0.3283422682 -0.0007660867
#> 944   0.543881227 -0.5529193326  0.0241608148
#> 945   0.453840352 -0.5634452174 -0.1806463250
#> 946   0.618308717 -0.3918503994  0.5864440374
#> 947   0.455198162  0.2115012444  0.6536872572
#> 948   0.711732110 -0.3056030792  0.6822534442
#> 949   0.536411464 -0.5113742527  0.1342418112
#> 950   0.663660003 -0.2926811567  0.6285436947
#> 951   0.550899696 -0.6707786708  0.0938493587
#> 952   0.559397604 -0.5631371949 -0.3651745919
#> 953   0.124735792 -1.1102668548  0.1721263025
#> 954   0.815651607 -0.2775730537  0.1302660986
#> 955   0.588656038 -0.5169057049  0.0208087330
#> 956   0.741583133 -0.0879518094  0.4888897325
#> 957   0.447428677 -0.5491538294  0.5517318049
#> 958   0.767514388 -0.1685262001  0.2847224020
#> 959   0.539944608  0.0279961640  0.7313057955
#> 960   0.706748905 -0.3295346899  0.4838138172
#> 961   0.498275157 -0.6704979230 -0.1934383187
#> 962   0.386683568 -0.8437197894 -0.3912133048
#> 963   0.513083574 -0.6586953197  0.6969742406
#> 964   0.367363510 -0.8261789715  0.0684590229
#> 965   0.440584555  0.1697228675  0.8206657256
#> 966   0.553367193 -0.6268295178  0.3825623503
#> 967   0.595604855 -0.3656412449  0.2514484929
#> 968   0.557049558 -0.5511988086  1.2072843024
#> 969   0.360269660 -0.5950481484  0.5716799636
#> 970   0.452508987 -0.9484063828 -0.5711543690
#> 971   0.809902560 -0.1806018011  0.2174659622
#> 972   0.585390161 -0.3987097235  0.1598145380
#> 973   0.768007309 -0.3283422682 -0.0007660867
#> 974   0.614588912 -0.4627811204  0.8515802491
#> 975   0.724992728 -0.2106123036  0.8821522758
#> 976   0.556819121 -0.3678088066 -0.1323693651
#> 977   0.195499862 -0.8856516707 -0.0144414466
#> 978   0.727607925 -0.0372549164  0.5264118007
#> 979   0.858702002 -0.0686268506  0.5127078284
#> 980   0.406866136 -0.1703454682  0.6703948313
#> 981   0.542052770 -0.4601021798 -0.0201198140
#> 982   0.499383528 -0.7331630692 -0.1334910426
#> 983   0.424396799 -0.6633735320  0.2722304233
#> 984  -0.274609626 -0.0586515211  0.3868277436
#> 985   0.548272924 -0.5718551112 -0.3482535258
#> 986   0.359893846 -0.3768630080  0.0665056168
#> 987   0.585390161 -0.3987097235  0.1598145380
#> 988   0.425319618 -0.5607598104  0.0417204059
#> 989   0.210939443 -0.3902984368  0.1635562787
#> 990   0.566896273 -0.5133644287 -0.0146844362
#> 991   0.724978404 -0.3221077804  0.5381293219
#> 992   0.520975707 -0.0720266418  0.3027082042
#> 993   0.498275157 -0.6704979230 -0.1934383187
#> 994   0.562892621 -0.8761745702 -0.6343268158
#> 995  -0.284608024 -0.6591085539  0.1310184390
#> 996   0.650660482 -0.4141469432  0.0096168927
#> 997   0.136052793 -0.1395590903  0.5712903500
#> 998   0.652307087 -0.1377703809  0.4859756325
#> 999  -0.037538745 -0.3671225308  0.2846181265
#> 1000  0.614588912 -0.4627811204  0.8515802491
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1  0.500 -0.611 -0.408 
#>  2      2        4       70        1  0.678 -0.232 -0.0414
#>  3      3        4       70        1  0.589 -0.517  0.0208
#>  4      4        4       70        1  0.378 -0.260  0.257 
#>  5      5        4       70        1  0.725 -0.211  0.882 
#>  6      6        4       70        1  0.499 -0.800 -0.737 
#>  7      7        4       70        1  0.725 -0.211  0.882 
#>  8      8        4       70        1  0.562 -0.831  0.191 
#>  9      9        4       70        1  0.215 -0.452 -0.191 
#> 10     10        4       70        1  0.605 -0.591 -0.467 
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  6.59  38.0 0.665 0.173 19.1  1220.       726.
#> 2      1    14     4    70     1  6.59  38.0 0.665 0.173  6.55    0.215    249.
#> 3      2     1     4    70     1  7.88  55.5 0.959 0.142 17.2   983.       957.
#> 4      2    14     4    70     1  7.88  55.5 0.959 0.142  6.01    0.00377  333.
#> 5      3     1     4    70     1  7.21  41.7 1.02  0.173 23.6   940.       984.
#> 6      3    14     4    70     1  7.21  41.7 1.02  0.173  5.37    0.00162  224.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>