Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object
A posologyr list, created by the posologyr function.

n_sample
Number of samples from the S-step

n_resample
Number of samples from the R-step

return_model
A boolean. Returns a RxODE model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a RxODE model using the estimated distributions of ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01_tobra,n_sample=1e4,n_resample=1e3)

#> 

#> $eta
#>             ETA_ke         ETA_V
#> 1    -6.914706e-02  0.1008101091
#> 2    -4.266982e-02 -0.1123753186
#> 3    -1.432780e-01  0.3222595826
#> 4    -1.360423e-01  0.1474898755
#> 5    -1.224482e-01  0.0879001323
#> 6    -6.032019e-02 -0.2511921541
#> 7    -8.176453e-02 -0.1101743529
#> 8    -2.616130e-01  0.1589764363
#> 9    -1.191073e-01 -0.0228334634
#> 10    2.476974e-02  0.0391470746
#> 11   -2.359814e-01  0.1693137822
#> 12   -8.503098e-02 -0.0772295755
#> 13   -4.297231e-02 -0.0087920416
#> 14   -6.197825e-02  0.0995865729
#> 15   -6.438494e-02  0.1284123492
#> 16    4.500410e-03  0.0143382640
#> 17   -1.508722e-02 -0.0888743370
#> 18   -1.298250e-01  0.0922025134
#> 19   -4.851858e-02 -0.0386930463
#> 20   -1.210732e-01  0.0058439129
#> 21   -6.531553e-03 -0.0440634116
#> 22   -1.148026e-01  0.1145199365
#> 23   -1.773750e-03 -0.0553502162
#> 24   -1.188063e-01  0.0441367867
#> 25   -1.995179e-01  0.0057776196
#> 26   -4.131236e-01  0.2464345344
#> 27   -5.534377e-02  0.2302993945
#> 28   -1.240439e-01 -0.2270309682
#> 29   -1.383588e-01  0.0296167058
#> 30   -1.347250e-01  0.1036587349
#> 31   -1.045314e-01  0.2782394940
#> 32   -6.182940e-02 -0.0923479240
#> 33   -2.827625e-01  0.1891691997
#> 34   -1.964249e-01  0.2751041636
#> 35   -5.306161e-02 -0.0711742097
#> 36   -1.319983e-01 -0.0397738670
#> 37   -7.046256e-02  0.1853082062
#> 38   -2.455048e-01 -0.0381286446
#> 39   -4.297231e-02 -0.0087920416
#> 40    2.114059e-02 -0.0495831837
#> 41   -3.746507e-02  0.0293454388
#> 42   -1.970650e-01  0.3518804365
#> 43   -1.070014e-01  0.1520873141
#> 44   -1.057278e-01  0.1328636754
#> 45    3.131157e-02 -0.1544050612
#> 46   -4.475344e-02  0.0823844866
#> 47   -2.978325e-01  0.2419871112
#> 48    2.682404e-02 -0.0346372788
#> 49   -1.897335e-01  0.1521396968
#> 50   -1.228567e-01  0.0644305811
#> 51   -3.807732e-02 -0.2095009848
#> 52   -2.927018e-01  0.1660673703
#> 53   -1.751639e-01  0.2071269755
#> 54   -9.013242e-02 -0.0934316228
#> 55   -3.476987e-01  0.1563340996
#> 56   -1.794342e-01  0.1481340674
#> 57   -1.077635e-01  0.0937819917
#> 58   -8.492732e-02  0.1038068299
#> 59   -1.404935e-01  0.0090743540
#> 60   -1.224482e-01  0.0879001323
#> 61   -3.311594e-02 -0.1033170687
#> 62   -9.012989e-02  0.0567497080
#> 63   -9.310809e-02  0.2114322190
#> 64   -6.590302e-02 -0.0010066728
#> 65   -1.061367e-01  0.1681794064
#> 66   -1.740191e-01  0.1100428110
#> 67   -7.566223e-02 -0.0249114126
#> 68   -1.886234e-01  0.2640906388
#> 69   -3.451212e-02  0.1465832979
#> 70    1.233627e-02 -0.0285563211
#> 71   -3.274215e-02 -0.0578213435
#> 72   -3.551459e-03 -0.1148866296
#> 73   -4.261995e-02  0.0850490703
#> 74   -1.774686e-01 -0.0831465361
#> 75   -1.083677e-01  0.0282858446
#> 76    1.379257e-02  0.1019749280
#> 77   -6.182940e-02 -0.0923479240
#> 78   -1.137492e-01 -0.0407656088
#> 79   -1.179982e-01  0.1447594308
#> 80   -6.506285e-03 -0.0258647696
#> 81   -2.389760e-01  0.2228395592
#> 82   -9.433895e-02 -0.0570313198
#> 83    7.963138e-03 -0.0213318045
#> 84   -1.566365e-01  0.1717247686
#> 85   -3.359506e-02 -0.1415705729
#> 86   -1.271416e-01  0.0532732895
#> 87   -1.485507e-01  0.1419865674
#> 88   -4.645973e-02  0.1065121105
#> 89   -8.213414e-02  0.3195803142
#> 90   -2.442594e-01  0.3050206568
#> 91   -1.017525e-01  0.1454209395
#> 92    4.191096e-02 -0.1569152391
#> 93   -5.699877e-02  0.0522793107
#> 94   -2.123516e-02  0.0408645240
#> 95   -1.404744e-01  0.1658533782
#> 96   -8.381355e-02  0.0703715479
#> 97   -8.005407e-02 -0.0707429475
#> 98   -1.140226e-01  0.0544507247
#> 99   -1.104695e-01  0.0181498394
#> 100  -3.907466e-03 -0.0112139576
#> 101  -8.007922e-02  0.0256568129
#> 102  -2.166968e-01  0.0867363450
#> 103  -2.735252e-01  0.1091165314
#> 104   3.131157e-02 -0.1544050612
#> 105  -2.136138e-01  0.1596786602
#> 106  -1.528022e-01  0.2561682281
#> 107  -2.202456e-01  0.0824667331
#> 108  -1.326556e-02  0.0283795484
#> 109   5.314181e-02 -0.1194661827
#> 110   7.826745e-03 -0.1978241763
#> 111  -2.826804e-02  0.0040657177
#> 112  -4.693506e-02  0.0141438561
#> 113  -1.382304e-01  0.0302162238
#> 114  -1.970650e-01  0.3518804365
#> 115  -1.738946e-01  0.1591605041
#> 116  -1.385229e-03 -0.0571216527
#> 117  -1.494918e-01  0.0079721228
#> 118  -1.463174e-01  0.1666876351
#> 119  -3.115744e-02 -0.1410246258
#> 120  -2.018197e-01  0.0653640704
#> 121  -2.429184e-01 -0.0895989517
#> 122  -2.334386e-02 -0.1814704285
#> 123  -9.864152e-02  0.0949431183
#> 124  -3.175305e-01  0.2022291138
#> 125  -4.798104e-02  0.0672041916
#> 126   2.114059e-02 -0.0495831837
#> 127  -1.888769e-01  0.2405552486
#> 128  -1.357130e-01  0.0247267118
#> 129  -5.282244e-02  0.0593138357
#> 130  -1.387082e-01  0.0957946017
#> 131  -2.424842e-01  0.1590650303
#> 132  -2.393629e-01 -0.0116101877
#> 133  -6.619854e-02  0.0746673933
#> 134  -7.098363e-02  0.0398228711
#> 135  -1.201758e-01 -0.0884502452
#> 136  -6.396418e-02  0.0275493757
#> 137   1.184312e-02 -0.0204630363
#> 138  -1.361448e-01  0.1675534803
#> 139  -2.431189e-01  0.0989735627
#> 140  -2.564292e-01 -0.0029213166
#> 141  -2.213274e-02  0.0484164467
#> 142  -1.767561e-01  0.1990134842
#> 143  -5.990986e-02 -0.0636999626
#> 144  -1.994217e-01  0.0940884957
#> 145  -1.388027e-01  0.0591187465
#> 146  -1.584401e-01  0.0547166455
#> 147  -5.476720e-02 -0.0132251733
#> 148  -1.734990e-01 -0.0383124466
#> 149  -1.886052e-01 -0.0701758293
#> 150  -1.183511e-01  0.2168794963
#> 151  -2.339521e-01  0.0903288028
#> 152  -5.216928e-02  0.1709101987
#> 153  -9.826412e-02  0.0062001826
#> 154   3.836053e-03 -0.0214476748
#> 155   1.275058e-02 -0.0694538293
#> 156  -3.035569e-01  0.2165370589
#> 157  -1.239925e-01  0.2047964197
#> 158  -4.436207e-02  0.0988305495
#> 159  -5.291389e-02  0.1840376311
#> 160  -2.210261e-02  0.1657744742
#> 161  -2.640436e-01  0.1127985874
#> 162  -1.129571e-01  0.0407985337
#> 163  -1.780123e-01  0.0680661678
#> 164   2.114059e-02 -0.0495831837
#> 165  -5.895108e-02  0.0076284524
#> 166  -1.972993e-01  0.1563542345
#> 167  -2.510816e-01  0.2529076819
#> 168   1.270203e-01  0.0549499529
#> 169  -8.487899e-02  0.0345115372
#> 170  -1.129571e-01  0.0407985337
#> 171  -2.250290e-01  0.1968914222
#> 172  -1.098837e-01  0.2978441366
#> 173  -5.276618e-02  0.1112586519
#> 174  -2.694854e-01  0.1114039621
#> 175  -1.307832e-01 -0.0865032714
#> 176  -1.305504e-01  0.0849836461
#> 177  -1.698331e-01  0.1155643008
#> 178  -1.034249e-01  0.0904155256
#> 179  -1.981368e-01  0.0771900318
#> 180  -2.814455e-01  0.1988005471
#> 181  -6.409578e-02 -0.0375529042
#> 182  -1.850298e-01  0.1627446875
#> 183  -3.060816e-01  0.1149069125
#> 184  -1.069350e-01  0.2376865461
#> 185  -4.208874e-02  0.0758708265
#> 186   2.835309e-02 -0.0446664335
#> 187  -2.542719e-01  0.0991814638
#> 188  -1.667048e-01  0.3402289678
#> 189  -1.701794e-01  0.0491862239
#> 190  -1.098837e-01  0.2978441366
#> 191  -9.722759e-02 -0.0302881568
#> 192  -9.786136e-02  0.0015206040
#> 193  -2.855926e-02  0.0001641643
#> 194  -8.281825e-02  0.2006519221
#> 195  -9.751427e-02 -0.1225927471
#> 196  -1.566673e-01  0.0015166634
#> 197   1.487511e-02 -0.0900537391
#> 198  -8.178080e-02  0.0741584719
#> 199  -1.854421e-01 -0.0425988115
#> 200  -1.477613e-01  0.0050596094
#> 201  -4.234779e-02 -0.0960244152
#> 202  -3.286364e-01  0.1629053013
#> 203   1.970163e-02  0.0271659116
#> 204  -8.705980e-02  0.0934837790
#> 205  -8.841981e-02 -0.0458136704
#> 206  -1.788432e-01 -0.0133917296
#> 207  -9.296485e-02 -0.0476981231
#> 208  -2.960817e-01  0.1504327476
#> 209  -2.300235e-01 -0.0044116944
#> 210  -6.528662e-03 -0.0354875965
#> 211  -4.687052e-02  0.0445779148
#> 212  -1.455554e-01  0.1592409742
#> 213  -4.869636e-02  0.0309732367
#> 214   1.127425e-02  0.0607897033
#> 215  -6.701184e-02  0.1305297380
#> 216  -1.808963e-01  0.1635676961
#> 217  -4.954837e-02 -0.0075295054
#> 218  -2.199164e-01  0.1801719650
#> 219   2.064460e-02 -0.1217966868
#> 220  -1.427590e-01  0.0380921911
#> 221  -3.080766e-02  0.0851269521
#> 222   1.622545e-02 -0.1293286621
#> 223  -1.944441e-01  0.1783356638
#> 224  -5.065028e-02 -0.1103296071
#> 225   3.573085e-02  0.0749986674
#> 226  -2.552379e-01  0.2143695928
#> 227  -6.629594e-02  0.1511350177
#> 228  -8.103812e-02 -0.0500226110
#> 229  -2.439693e-01  0.1485414573
#> 230  -1.931965e-01  0.1072479694
#> 231  -1.210488e-01 -0.0899442688
#> 232  -1.128141e-01 -0.0902963999
#> 233   3.469757e-02 -0.1248187440
#> 234  -8.548013e-03  0.0595096638
#> 235  -2.261377e-01  0.0129792457
#> 236   6.615896e-03  0.1034966183
#> 237  -1.039987e-01 -0.0146189887
#> 238  -5.436742e-03  0.0395537988
#> 239  -6.004437e-02  0.0460243511
#> 240  -1.667048e-01  0.3402289678
#> 241  -1.975867e-01  0.0956909562
#> 242  -9.692633e-02 -0.0773722806
#> 243  -1.492023e-01  0.0857821856
#> 244  -1.328291e-01  0.1462537790
#> 245  -2.194066e-02  0.0792154319
#> 246  -1.928161e-01  0.2054331252
#> 247  -6.409560e-02  0.1185178934
#> 248  -5.091988e-02  0.0646365990
#> 249  -1.264082e-01 -0.1417088096
#> 250  -2.218594e-01 -0.0761124622
#> 251  -1.534993e-03  0.1350115382
#> 252  -1.575385e-01  0.0669496527
#> 253   9.659217e-03 -0.0287065066
#> 254  -8.005407e-02 -0.0707429475
#> 255  -1.271277e-01  0.0052600634
#> 256  -1.401743e-01  0.2028967747
#> 257   1.354035e-02 -0.0559077087
#> 258  -1.093695e-01  0.1115229307
#> 259  -1.879574e-01  0.1468180063
#> 260  -8.484085e-02  0.2741721113
#> 261  -2.060202e-01 -0.0122104833
#> 262  -6.732591e-02  0.0035483644
#> 263   3.604246e-02 -0.0449547320
#> 264  -2.437069e-01  0.0977705626
#> 265  -6.094742e-02  0.0569654551
#> 266  -2.134150e-01  0.0132194817
#> 267  -1.211815e-01  0.0776872174
#> 268  -1.802746e-01  0.0894944407
#> 269  -7.600923e-02 -0.0865207145
#> 270  -2.580472e-02  0.1858761236
#> 271  -8.138254e-02  0.1858168967
#> 272  -2.540381e-02  0.0159171110
#> 273  -6.111407e-02  0.1370873995
#> 274  -1.007392e-01  0.1571359331
#> 275   1.973905e-02 -0.0625703381
#> 276   1.927137e-02  0.0963567996
#> 277  -2.141899e-01  0.1464551324
#> 278  -8.488865e-02  0.0465654091
#> 279  -2.238979e-01  0.2063674831
#> 280  -6.409560e-02  0.1185178934
#> 281  -1.290641e-01 -0.0313066630
#> 282  -9.310809e-02  0.2114322190
#> 283  -1.201758e-01 -0.0884502452
#> 284  -1.017525e-01  0.1454209395
#> 285  -3.622472e-02  0.1477343306
#> 286  -3.528751e-02 -0.0672076277
#> 287  -1.404744e-01  0.1658533782
#> 288  -1.387288e-01  0.0893809099
#> 289  -9.569618e-03 -0.0470439087
#> 290  -6.059882e-02  0.0821147537
#> 291   5.675797e-02 -0.1002702792
#> 292  -1.406149e-01  0.0292946291
#> 293  -5.932769e-02  0.1797402126
#> 294  -8.429484e-02 -0.0424701103
#> 295  -1.474174e-01  0.1022189199
#> 296   3.237030e-02 -0.0561529362
#> 297  -1.596252e-01  0.0400794417
#> 298  -4.551269e-02  0.1239195302
#> 299  -2.603994e-01 -0.0524830311
#> 300  -1.406149e-01  0.0292946291
#> 301  -6.514603e-02 -0.0314242130
#> 302  -1.098085e-01  0.0418452203
#> 303  -2.147831e-01  0.3482514405
#> 304  -1.672202e-01  0.0428974482
#> 305  -9.436964e-02  0.0950510154
#> 306  -8.007922e-02  0.0256568129
#> 307  -9.711878e-02  0.2020503500
#> 308  -1.886234e-01  0.2640906388
#> 309  -6.306772e-02 -0.0228482271
#> 310   4.238569e-03 -0.0570670522
#> 311  -2.336664e-01  0.1855310686
#> 312  -2.656477e-01  0.1083497006
#> 313  -1.167720e-01  0.0522493117
#> 314  -2.909300e-01  0.1531351579
#> 315  -5.121037e-02 -0.0400731120
#> 316  -9.826412e-02  0.0062001826
#> 317   5.825639e-02 -0.0925074825
#> 318  -8.841981e-02 -0.0458136704
#> 319  -5.177009e-02  0.0358595071
#> 320  -2.750779e-01  0.0849865350
#> 321  -7.486242e-02  0.0140859869
#> 322  -1.083677e-01  0.0282858446
#> 323  -1.255441e-01  0.0623648042
#> 324  -2.898954e-01  0.0384133056
#> 325  -1.357130e-01  0.0247267118
#> 326  -1.472287e-01  0.2033709317
#> 327  -1.156678e-01  0.0603001327
#> 328  -1.363053e-01  0.0126515490
#> 329  -1.052901e-01  0.1225021091
#> 330  -1.206911e-01  0.1204783165
#> 331  -2.873135e-02 -0.0479617414
#> 332  -1.970650e-01  0.3518804365
#> 333  -5.930148e-02  0.0182262495
#> 334  -1.670305e-01  0.2315327970
#> 335  -7.646683e-02 -0.1355369253
#> 336   3.479654e-02  0.0387167838
#> 337   3.056019e-02 -0.0149115287
#> 338  -1.970650e-01  0.3518804365
#> 339  -8.436779e-02  0.1028876685
#> 340  -1.767561e-01  0.1990134842
#> 341  -6.197825e-02  0.0995865729
#> 342  -1.699088e-01  0.0681065483
#> 343  -7.069339e-02  0.3071629198
#> 344  -6.484761e-02  0.1168599730
#> 345  -1.992133e-01 -0.1040805929
#> 346  -1.082257e-01  0.2096635630
#> 347  -8.503098e-02 -0.0772295755
#> 348  -7.829706e-02 -0.0650848828
#> 349   3.617634e-02 -0.0874116945
#> 350  -4.989064e-02  0.0822437053
#> 351  -3.667801e-02 -0.0223690455
#> 352  -1.698331e-01  0.1155643008
#> 353  -7.833188e-03  0.0335055642
#> 354  -1.623202e-01  0.1215591831
#> 355  -1.129777e-01  0.0834062027
#> 356  -2.339521e-01  0.0903288028
#> 357  -1.150538e-01  0.0987142124
#> 358  -3.197149e-01  0.3206349422
#> 359  -3.175305e-01  0.2022291138
#> 360  -8.802423e-02  0.1737671789
#> 361  -1.504900e-01  0.2047438881
#> 362  -4.661531e-02 -0.1428423534
#> 363  -3.735332e-01 -0.0023275547
#> 364  -9.154542e-02  0.0801666278
#> 365  -3.129449e-01  0.0367952480
#> 366  -8.922346e-02 -0.0648196238
#> 367  -1.556618e-01  0.0999796807
#> 368  -9.599855e-02  0.1058302924
#> 369   3.665698e-02 -0.1083167325
#> 370  -5.052706e-02  0.1005441591
#> 371  -1.605885e-01 -0.0809168530
#> 372  -3.465152e-01  0.2144757571
#> 373  -1.739624e-01  0.1951777322
#> 374  -5.199728e-02  0.1560530530
#> 375  -7.620801e-03 -0.1144768913
#> 376  -6.333727e-02 -0.0548762912
#> 377  -5.177009e-02  0.0358595071
#> 378  -2.686939e-02 -0.0081215213
#> 379  -1.808963e-01  0.1635676961
#> 380  -3.368152e-01  0.1526311164
#> 381  -5.197564e-02  0.1467058945
#> 382  -1.140226e-01  0.0544507247
#> 383   4.509879e-02  0.0155215600
#> 384  -1.782434e-01  0.0846180304
#> 385  -1.617324e-01  0.1015302873
#> 386  -1.432780e-01  0.3222595826
#> 387  -8.431559e-02  0.0764191026
#> 388  -3.323549e-02  0.0872633081
#> 389  -1.018763e-01  0.0887186742
#> 390  -1.396531e-01  0.1683516432
#> 391  -5.164275e-02 -0.0701910508
#> 392  -6.484761e-02  0.1168599730
#> 393  -2.703914e-01  0.2082234091
#> 394   2.811050e-03 -0.0202552867
#> 395  -2.619732e-01  0.1473161465
#> 396  -4.213272e-02  0.0803212371
#> 397   2.123002e-02  0.0784200685
#> 398  -5.285709e-02  0.1365273849
#> 399   1.204698e-04  0.0520910851
#> 400  -1.970650e-01  0.3518804365
#> 401  -4.599072e-02  0.0098394471
#> 402  -4.877819e-02  0.0921650012
#> 403  -1.597914e-01 -0.0668156834
#> 404  -3.129792e-02 -0.1116073315
#> 405  -7.665699e-02 -0.1190212636
#> 406  -2.024138e-01  0.2474196429
#> 407  -1.982387e-01  0.1382508341
#> 408  -1.392440e-01  0.1922859703
#> 409  -1.144927e-01  0.2727724283
#> 410  -2.803086e-01  0.2175543226
#> 411  -1.529269e-01  0.1610750411
#> 412  -1.336902e-01 -0.0492721593
#> 413  -4.432318e-02 -0.0050892849
#> 414  -1.679648e-01  0.0789852889
#> 415  -3.354541e-01  0.2368213267
#> 416  -1.899217e-01  0.0968159684
#> 417   1.204698e-04  0.0520910851
#> 418  -1.282346e-01 -0.0191925386
#> 419  -1.312187e-01  0.1312735291
#> 420  -1.910860e-01  0.1058129284
#> 421  -8.983058e-02  0.1732395356
#> 422  -8.138254e-02  0.1858168967
#> 423  -7.613612e-02 -0.0341575666
#> 424  -5.715705e-02  0.0779321123
#> 425  -1.373362e-01  0.1274145082
#> 426  -4.700425e-02  0.0333126868
#> 427  -3.473892e-01  0.2449577759
#> 428  -3.905414e-02 -0.0051056341
#> 429  -1.311307e-01  0.1315268462
#> 430  -1.970650e-01  0.3518804365
#> 431  -1.151466e-01  0.1496136311
#> 432  -2.715078e-01  0.0143366477
#> 433  -1.798767e-01  0.1162829295
#> 434  -6.325316e-02  0.1916298105
#> 435  -8.040202e-02  0.1240501720
#> 436  -2.125347e-01  0.1279590987
#> 437   2.636994e-02 -0.0397919081
#> 438  -2.510816e-01  0.2529076819
#> 439  -2.975060e-01  0.0911864049
#> 440  -6.129262e-02  0.0045299169
#> 441  -1.059723e-01  0.1835860648
#> 442  -3.396895e-01  0.2070571920
#> 443  -2.431570e-01  0.1598175282
#> 444  -7.772173e-02  0.1617889385
#> 445   6.140099e-02 -0.1249123444
#> 446  -1.399746e-01  0.0690356199
#> 447  -2.136178e-01  0.1583053184
#> 448  -6.176233e-02  0.0461550421
#> 449  -5.658261e-02  0.0781034456
#> 450  -1.034245e-01  0.0809553106
#> 451  -1.479361e-01  0.2594854299
#> 452  -1.970650e-01  0.3518804365
#> 453  -1.686460e-01  0.0251165342
#> 454  -1.686460e-01  0.0251165342
#> 455  -2.955235e-01  0.1027594936
#> 456  -9.243245e-03  0.1624908593
#> 457  -1.850298e-01  0.1627446875
#> 458  -5.576281e-02  0.0702685276
#> 459  -2.350794e-01  0.1144450639
#> 460  -1.052901e-01  0.1225021091
#> 461  -6.878011e-02  0.0584692430
#> 462  -4.132530e-02  0.0577426266
#> 463  -9.380036e-03 -0.2226562294
#> 464  -1.416891e-01  0.1373967511
#> 465  -3.622472e-02  0.1477343306
#> 466  -9.849737e-02  0.1125699352
#> 467  -7.727396e-03 -0.0142700317
#> 468  -1.069263e-01 -0.0406346657
#> 469   3.353730e-02 -0.1473920317
#> 470  -1.150635e-01  0.0154838731
#> 471  -1.148228e-01  0.2154075366
#> 472  -1.264960e-01  0.1150474511
#> 473  -7.779454e-02 -0.0587485773
#> 474  -8.364196e-02  0.0224329781
#> 475  -1.704040e-01  0.1296529886
#> 476  -1.420784e-01  0.2557766014
#> 477  -1.792424e-01  0.0853798391
#> 478  -1.646634e-01  0.1836931992
#> 479  -5.798601e-02  0.0496509201
#> 480   1.275777e-02  0.0109026741
#> 481  -8.058950e-02  0.0575213054
#> 482  -2.734508e-01  0.2847352896
#> 483  -1.086585e-01  0.0862362135
#> 484  -2.519528e-03 -0.3234595104
#> 485  -1.427602e-01  0.1007132724
#> 486  -1.156520e-01 -0.0906932874
#> 487  -1.691153e-01  0.1711143338
#> 488   2.799306e-02  0.1151761185
#> 489  -9.353466e-02  0.1492190181
#> 490  -1.408721e-01  0.0146736741
#> 491  -8.705980e-02  0.0934837790
#> 492  -1.125174e-01  0.1908124853
#> 493  -1.088172e-01 -0.0886520754
#> 494  -4.216187e-03  0.0202040451
#> 495  -1.528022e-01  0.2561682281
#> 496  -1.584401e-01  0.0547166455
#> 497  -5.216928e-02  0.1709101987
#> 498  -7.454217e-02  0.0151765950
#> 499  -2.062962e-01  0.1316650457
#> 500  -2.162189e-02  0.0886285813
#> 501   1.868701e-02 -0.1596988100
#> 502  -6.304292e-02 -0.0619191268
#> 503  -2.431570e-01  0.1598175282
#> 504  -4.020486e-02 -0.0399099519
#> 505  -2.936657e-01  0.1352873036
#> 506  -1.261258e-01 -0.0149602694
#> 507  -5.530523e-02  0.1013463709
#> 508   1.640377e-02 -0.0334579889
#> 509  -2.909382e-02 -0.0710309353
#> 510  -1.566365e-01  0.1717247686
#> 511  -2.036562e-02 -0.0015971405
#> 512  -7.549225e-02  0.1395424753
#> 513  -1.253634e-01  0.1981056486
#> 514  -1.043837e-01  0.0864774844
#> 515  -2.024138e-01  0.2474196429
#> 516  -7.030049e-02  0.0165603736
#> 517   1.125808e-02  0.0043517381
#> 518  -3.429389e-02 -0.0307902880
#> 519  -6.100937e-02  0.2216439060
#> 520  -1.147226e-01  0.1766794816
#> 521  -6.987526e-02 -0.1291567413
#> 522  -5.768979e-02  0.1165760928
#> 523  -5.578719e-02  0.1003650341
#> 524  -2.740934e-02  0.0313522681
#> 525  -7.717840e-02  0.0501709090
#> 526  -1.650934e-01  0.0173138258
#> 527  -3.122090e-01  0.1829298505
#> 528  -1.472287e-01  0.2033709317
#> 529  -1.156166e-01  0.0982212324
#> 530   1.048031e-02  0.1035370669
#> 531  -1.275066e-01  0.1744108179
#> 532  -8.488865e-02  0.0465654091
#> 533  -1.052901e-01  0.1225021091
#> 534  -2.277070e-02 -0.0035408886
#> 535  -1.428444e-01  0.1846904296
#> 536  -9.527007e-02 -0.0087315556
#> 537  -1.704774e-01  0.1642687708
#> 538  -5.966514e-02  0.1217233828
#> 539  -3.455174e-02  0.1040190556
#> 540  -1.016233e-01 -0.0498279499
#> 541  -2.003608e-02  0.0278491548
#> 542  -1.248384e-01  0.1610285973
#> 543  -8.503098e-02 -0.0772295755
#> 544  -2.157883e-01  0.0615634374
#> 545  -1.244770e-01 -0.1174082537
#> 546  -2.442594e-01  0.3050206568
#> 547  -2.498699e-01  0.1559379434
#> 548  -1.547767e-03  0.0770414696
#> 549  -7.339852e-02  0.1670985162
#> 550  -1.595831e-01  0.2272813915
#> 551  -3.834067e-03  0.0818836401
#> 552  -1.070014e-01  0.1520873141
#> 553  -8.347765e-02 -0.0303805298
#> 554  -1.110133e-01 -0.0441769205
#> 555  -2.392902e-01  0.1019146562
#> 556  -8.374020e-02  0.0682057859
#> 557  -1.886234e-01  0.2640906388
#> 558  -2.962164e-01  0.0811604240
#> 559   2.584767e-02 -0.0008456347
#> 560  -3.348932e-01  0.2459780974
#> 561  -1.021668e-01  0.0441768890
#> 562  -4.726846e-02 -0.0172746534
#> 563  -9.310809e-02  0.2114322190
#> 564  -3.779168e-02  0.0877055735
#> 565  -7.696538e-02  0.0566713585
#> 566  -1.091107e-01 -0.0252740942
#> 567  -1.382492e-01  0.0908868683
#> 568  -6.318196e-02 -0.0152682586
#> 569   3.341752e-03 -0.0407373851
#> 570  -2.478134e-01  0.1614562795
#> 571  -1.328264e-01  0.0425531164
#> 572  -1.172650e-01 -0.0466145947
#> 573  -4.436207e-02  0.0988305495
#> 574  -1.231328e-01  0.1399448137
#> 575  -6.938057e-02 -0.0320135961
#> 576   1.036846e-01 -0.1105606812
#> 577  -1.082753e-02  0.0533214730
#> 578  -1.491010e-02  0.0381316546
#> 579  -1.301854e-01  0.1906735290
#> 580  -4.939367e-02  0.0196123448
#> 581  -1.942340e-01 -0.0248424772
#> 582  -1.689915e-01  0.0836128676
#> 583  -1.231328e-01  0.1399448137
#> 584   3.958741e-02 -0.0299053869
#> 585  -1.288996e-01 -0.0112136638
#> 586  -1.556618e-01  0.0999796807
#> 587   3.479654e-02  0.0387167838
#> 588  -1.160465e-01  0.0644236098
#> 589  -1.312279e-01  0.0628580281
#> 590  -1.467052e-01 -0.1750560520
#> 591  -8.488865e-02  0.0465654091
#> 592  -8.542452e-02  0.0374160186
#> 593  -1.059723e-01  0.1835860648
#> 594   8.048287e-02 -0.1196009966
#> 595  -1.227407e-01  0.2280852458
#> 596  -3.162788e-01  0.2676468306
#> 597  -8.381355e-02  0.0703715479
#> 598  -5.537535e-02  0.0066086099
#> 599  -9.431705e-02 -0.0557544220
#> 600  -3.304336e-02  0.0961055888
#> 601  -1.348624e-01  0.1933825575
#> 602  -1.724027e-01  0.0272373329
#> 603  -1.360960e-01 -0.1441435396
#> 604  -2.846464e-02 -0.0147775226
#> 605  -1.116815e-01  0.2203772268
#> 606  -2.359325e-01  0.2604296630
#> 607  -6.371272e-02  0.1186051895
#> 608  -1.982947e-01  0.1732959613
#> 609   8.875522e-03 -0.1832599307
#> 610  -4.700425e-02  0.0333126868
#> 611  -3.127166e-01  0.1230741717
#> 612  -1.220536e-01  0.2397407826
#> 613  -1.886234e-01  0.2640906388
#> 614  -2.909807e-01  0.0840193472
#> 615  -1.603423e-01  0.0974366392
#> 616  -1.404744e-01  0.1658533782
#> 617  -2.321764e-02  0.0136160176
#> 618  -3.045419e-01  0.1170019576
#> 619  -2.392902e-01  0.1019146562
#> 620  -8.488865e-02  0.0465654091
#> 621  -3.093173e-01  0.0154451401
#> 622  -2.885987e-01 -0.0327043081
#> 623  -5.534377e-02  0.2302993945
#> 624  -7.816832e-02  0.0569647482
#> 625  -5.895108e-02  0.0076284524
#> 626  -1.281279e-01  0.1373731706
#> 627  -9.679751e-02  0.0862493142
#> 628  -9.059473e-02  0.0922218769
#> 629  -1.086631e-01 -0.1999954861
#> 630  -8.547967e-02  0.0341414583
#> 631  -5.467305e-02  0.0085890734
#> 632  -1.977965e-01  0.0749498703
#> 633  -3.905414e-02 -0.0051056341
#> 634  -2.609967e-02  0.0764969499
#> 635   3.663543e-02 -0.0910590264
#> 636  -1.089392e-01  0.0500441673
#> 637  -4.778433e-02  0.0255728675
#> 638  -2.058564e-01  0.2434417235
#> 639  -5.177009e-02  0.0358595071
#> 640  -1.515076e-01 -0.1174006164
#> 641  -4.961367e-01  0.2889638125
#> 642  -1.413947e-01  0.1593563341
#> 643  -8.103812e-02 -0.0500226110
#> 644  -3.197149e-01  0.3206349422
#> 645  -1.117551e-01  0.0135171938
#> 646  -1.045314e-01  0.2782394940
#> 647  -1.164960e-01 -0.2037737379
#> 648  -1.322848e-01  0.0899024477
#> 649  -3.797623e-02 -0.0804480450
#> 650  -1.083364e-01 -0.0323506098
#> 651  -1.047218e-01  0.0608197270
#> 652  -1.679648e-01  0.0789852889
#> 653  -1.639493e-01  0.1284391258
#> 654  -5.699877e-02  0.0522793107
#> 655  -5.463423e-02  0.1189441471
#> 656  -1.682242e-01  0.0975619960
#> 657  -3.060816e-01  0.1149069125
#> 658  -6.780327e-02 -0.0263551171
#> 659  -1.736096e-01  0.1373083565
#> 660  -1.766282e-01 -0.0170185138
#> 661  -1.601342e-01 -0.0815929443
#> 662  -1.176231e-01 -0.0963626662
#> 663  -1.667056e-01  0.1541181179
#> 664  -2.482326e-01  0.0367992971
#> 665  -2.708449e-01  0.1966319181
#> 666  -7.343458e-02 -0.0595069857
#> 667  -3.108960e-02  0.1453943061
#> 668   1.415138e-01 -0.1976446323
#> 669  -1.337157e-01  0.0667771414
#> 670  -6.093040e-02  0.1837052322
#> 671  -1.713123e-01  0.0286909502
#> 672  -1.513724e-01  0.0750203031
#> 673  -1.439459e-01  0.1332307451
#> 674  -3.724002e-02 -0.0230493434
#> 675  -1.158851e-01  0.0487926371
#> 676  -1.362473e-01 -0.0628180172
#> 677  -1.420784e-01  0.2557766014
#> 678  -1.545033e-01 -0.0131471775
#> 679   5.929763e-02 -0.0792042404
#> 680  -1.466207e-01  0.1270935112
#> 681  -1.334120e-01  0.0129438022
#> 682  -1.964249e-01  0.2751041636
#> 683  -1.239925e-01  0.2047964197
#> 684  -1.623202e-01  0.1215591831
#> 685   4.004270e-03 -0.0309992933
#> 686  -2.359325e-01  0.2604296630
#> 687  -1.535133e-01  0.2487647747
#> 688  -1.918524e-01 -0.0201587518
#> 689  -5.071492e-02  0.2305148497
#> 690  -2.621671e-01  0.1802839918
#> 691  -2.056986e-01  0.1459398596
#> 692  -1.181300e-01  0.1446007443
#> 693   4.636881e-02 -0.0366639629
#> 694  -7.921778e-03 -0.0115047198
#> 695  -8.699876e-02  0.0137121713
#> 696  -1.513904e-01 -0.0277226384
#> 697  -1.574571e-01  0.0605272605
#> 698  -1.239925e-01  0.2047964197
#> 699  -2.067576e-01 -0.0339756700
#> 700  -6.919724e-02  0.0333599775
#> 701  -1.602064e-01  0.1471459131
#> 702  -8.462386e-02  0.1726924066
#> 703  -4.868278e-02 -0.0840964917
#> 704  -6.380008e-02 -0.0517215718
#> 705  -5.983925e-02  0.0906387476
#> 706  -1.439459e-01  0.1332307451
#> 707  -5.697206e-01  0.1223155759
#> 708   5.136732e-02  0.0323301339
#> 709   3.297676e-02  0.0978898305
#> 710  -7.481825e-02 -0.0382339069
#> 711  -1.944593e-01  0.1689626716
#> 712   1.197306e-02  0.1293871673
#> 713  -1.090953e-01  0.1406191614
#> 714  -7.816832e-02  0.0569647482
#> 715  -3.938489e-02 -0.0149956837
#> 716  -2.042639e-01  0.2791463995
#> 717  -7.948669e-02 -0.0980692873
#> 718   9.392325e-02 -0.1075184288
#> 719  -1.464220e-01  0.0750995879
#> 720  -3.391836e-02  0.1803159996
#> 721  -1.098837e-01  0.2978441366
#> 722  -1.061367e-01  0.1681794064
#> 723  -9.179011e-02  0.1131073648
#> 724  -1.934606e-01  0.0772065368
#> 725  -1.082753e-02  0.0533214730
#> 726  -2.173575e-01  0.0205662130
#> 727  -1.237574e-01  0.0215747351
#> 728   3.029550e-02 -0.0732124687
#> 729  -3.483516e-02 -0.0078252199
#> 730   1.370181e-02  0.0210967391
#> 731  -1.150635e-01  0.0154838731
#> 732   3.037055e-02 -0.1601871826
#> 733  -1.312914e-01  0.0857085936
#> 734  -1.886234e-01  0.2640906388
#> 735  -3.430876e-02  0.1567431327
#> 736  -1.231328e-01  0.1399448137
#> 737  -8.214038e-02  0.0114794950
#> 738  -1.978457e-01  0.2387137440
#> 739  -2.156685e-01  0.0829812140
#> 740  -8.591406e-02 -0.1274588814
#> 741  -1.717947e-01 -0.0007033376
#> 742  -1.812760e-01  0.0848824401
#> 743  -6.411236e-02  0.2658333246
#> 744  -4.903089e-02 -0.1541994765
#> 745  -2.205969e-01  0.0916220080
#> 746  -1.023685e-01  0.2511199177
#> 747  -1.188063e-01  0.0441367867
#> 748  -6.389360e-02  0.0803611398
#> 749   5.287859e-02 -0.0433240103
#> 750  -1.301854e-01  0.1906735290
#> 751  -1.463403e-01  0.1909538130
#> 752  -4.180271e-02 -0.0225266858
#> 753  -2.042639e-01  0.2791463995
#> 754  -1.089392e-01  0.0500441673
#> 755  -7.715612e-02  0.1373897266
#> 756  -2.888213e-01  0.1137244352
#> 757  -2.056986e-01  0.1459398596
#> 758  -1.691153e-01  0.1711143338
#> 759  -5.529301e-02  0.1654233684
#> 760  -6.318196e-02 -0.0152682586
#> 761  -5.763394e-02  0.0374732838
#> 762  -1.153877e-01  0.1810566500
#> 763  -3.140741e-01  0.0996844882
#> 764  -5.673012e-02 -0.0816740288
#> 765  -6.318196e-02 -0.0152682586
#> 766  -1.156678e-01  0.0603001327
#> 767  -1.446881e-01  0.0567660221
#> 768  -1.347912e-01  0.1102450751
#> 769  -1.215583e-01  0.0792832539
#> 770  -1.955052e-01  0.1171211150
#> 771  -6.919724e-02  0.0333599775
#> 772  -1.480136e-01 -0.0993115156
#> 773  -1.288996e-01 -0.0112136638
#> 774  -1.246712e-01 -0.0712578503
#> 775  -9.443571e-02  0.0376876512
#> 776  -1.148228e-01  0.2154075366
#> 777  -1.387082e-01  0.0957946017
#> 778  -1.300784e-01  0.2048950140
#> 779  -3.197149e-01  0.3206349422
#> 780   3.861110e-02  0.0546693305
#> 781  -1.370078e-01  0.1088464823
#> 782  -1.400337e-01  0.0937111435
#> 783  -2.998618e-02  0.0441347995
#> 784  -4.443731e-02 -0.0266626326
#> 785  -1.566365e-01  0.1717247686
#> 786   7.240004e-02 -0.1494912387
#> 787  -2.745979e-01  0.1960283332
#> 788  -6.167315e-02 -0.0903806251
#> 789  -1.397605e-01  0.0190430687
#> 790  -1.670305e-01  0.2315327970
#> 791  -1.502464e-01 -0.1329683673
#> 792  -4.184502e-01  0.1339869843
#> 793  -1.311307e-01  0.1315268462
#> 794  -1.578137e-01  0.2017647160
#> 795  -7.534573e-02 -0.0490653551
#> 796  -5.921032e-02  0.0484835985
#> 797  -1.732596e-02  0.0277597212
#> 798  -3.574668e-01  0.2206048114
#> 799  -4.261995e-02  0.0850490703
#> 800   1.325862e-01 -0.1034564166
#> 801  -1.298250e-01  0.0922025134
#> 802  -1.246572e-01  0.1001381507
#> 803  -8.464295e-02  0.1031866548
#> 804  -1.596050e-01  0.0874651178
#> 805  -7.454217e-02  0.0151765950
#> 806  -3.296224e-01  0.2143109942
#> 807  -4.476449e-02 -0.0969822997
#> 808  -1.879574e-01  0.1468180063
#> 809  -2.004699e-02  0.0840796846
#> 810  -9.289302e-02  0.0006842892
#> 811  -2.745979e-01  0.1960283332
#> 812  -9.322464e-02  0.0080541647
#> 813  -3.344053e-01  0.2123061703
#> 814  -1.211815e-01  0.0776872174
#> 815  -1.060783e-01  0.0956304035
#> 816  -1.091775e-01  0.1739297908
#> 817  -2.580331e-01  0.0253367029
#> 818  -1.525012e-01  0.0249116121
#> 819  -1.949491e-01  0.1032539644
#> 820  -2.018197e-01  0.0653640704
#> 821  -7.202080e-02 -0.0742017500
#> 822  -1.449724e-01  0.0717986661
#> 823  -7.705737e-03  0.0489915785
#> 824  -2.712993e-01  0.0628630850
#> 825  -7.030049e-02  0.0165603736
#> 826  -1.383588e-01  0.0296167058
#> 827  -1.010978e-01  0.1257342928
#> 828  -4.635027e-02  0.0022963255
#> 829   8.955445e-02 -0.2176491645
#> 830  -3.736516e-02  0.1068921947
#> 831  -6.129996e-02  0.1122191254
#> 832  -1.851125e-02 -0.0146349159
#> 833   3.665698e-02 -0.1083167325
#> 834  -7.917740e-02  0.2109844849
#> 835  -1.436912e-01  0.1017612788
#> 836  -9.436964e-02  0.0950510154
#> 837  -2.806499e-01  0.1302786680
#> 838  -6.187118e-02 -0.0614382251
#> 839  -2.780598e-01  0.1098624001
#> 840  -1.206911e-01  0.1204783165
#> 841  -1.471146e-01  0.1546333502
#> 842  -1.481106e-01 -0.0302509969
#> 843  -1.049247e-01  0.1329829777
#> 844  -2.324638e-01  0.1200578519
#> 845  -1.650794e-01  0.1199800120
#> 846  -1.010978e-01  0.1257342928
#> 847   2.257184e-02  0.0379772475
#> 848  -7.715612e-02  0.1373897266
#> 849  -2.552379e-01  0.2143695928
#> 850  -7.549225e-02  0.1395424753
#> 851  -6.914706e-02  0.1008101091
#> 852  -9.442605e-02  0.0923392456
#> 853  -5.983925e-02  0.0906387476
#> 854  -3.048045e-01  0.1283018169
#> 855  -2.287280e-01  0.0202422900
#> 856  -3.639452e-02  0.1556785425
#> 857  -7.182809e-02  0.0198672797
#> 858   1.085022e-02  0.0095371781
#> 859  -1.810268e-01  0.1984046822
#> 860  -1.047232e-01  0.1366953948
#> 861  -3.473892e-01  0.2449577759
#> 862  -1.931965e-01  0.1072479694
#> 863  -3.465152e-01  0.2144757571
#> 864   1.300872e-02 -0.0818628756
#> 865  -1.987022e-02 -0.1383649445
#> 866  -1.396142e-02  0.0856922428
#> 867  -1.298250e-01  0.0922025134
#> 868  -5.487209e-02 -0.1792200504
#> 869  -2.192064e-01  0.1716968128
#> 870  -3.063919e-01  0.0974603667
#> 871  -2.604057e-01 -0.0832509481
#> 872  -1.436059e-02  0.2056132945
#> 873  -4.822771e-02 -0.0958427890
#> 874  -1.057278e-01  0.1328636754
#> 875  -1.391350e-01 -0.0730571616
#> 876  -5.231472e-02 -0.0284592637
#> 877  -1.425666e-01  0.1955179183
#> 878   5.586131e-02  0.0395346623
#> 879  -5.768979e-02  0.1165760928
#> 880  -2.641495e-03 -0.2459236277
#> 881  -5.798601e-02  0.0496509201
#> 882   3.354512e-03  0.1236127208
#> 883  -2.442594e-01  0.3050206568
#> 884  -1.751639e-01  0.2071269755
#> 885  -9.887563e-02  0.2111151703
#> 886  -1.691153e-01  0.1711143338
#> 887  -1.860918e-01 -0.0285852098
#> 888  -7.433320e-02 -0.0372288836
#> 889  -7.746220e-02  0.0823801923
#> 890  -1.017525e-01  0.1454209395
#> 891  -2.822788e-01  0.2111194436
#> 892  -1.528022e-01  0.2561682281
#> 893  -6.098579e-02 -0.0013094395
#> 894  -2.767969e-01  0.1393690618
#> 895   6.803225e-05 -0.0116025461
#> 896  -1.454506e-01  0.0697480143
#> 897  -1.081355e-01 -0.0072255066
#> 898  -9.887563e-02  0.2111151703
#> 899  -2.907321e-01  0.1172188738
#> 900  -1.737738e-01  0.2645140806
#> 901  -1.017525e-01  0.1454209395
#> 902  -8.758304e-02  0.1109977357
#> 903  -1.263479e-01  0.1230733615
#> 904  -1.060783e-01  0.0956304035
#> 905  -9.151885e-02  0.0421646981
#> 906  -9.515327e-02  0.0683327559
#> 907  -6.514603e-02 -0.0314242130
#> 908  -1.061620e-01  0.0782305537
#> 909  -3.036676e-01  0.0280718775
#> 910  -9.711878e-02  0.2020503500
#> 911  -1.022777e-01 -0.0093856928
#> 912  -2.858180e-01  0.2410143017
#> 913  -5.597149e-02  0.0030503888
#> 914  -1.174263e-01  0.2796359669
#> 915  -1.724189e-01  0.1920108796
#> 916  -1.650794e-01  0.1199800120
#> 917  -1.254078e-01  0.2140500873
#> 918  -1.528022e-01  0.2561682281
#> 919  -9.250340e-02  0.0800681836
#> 920   7.358406e-02 -0.1764145559
#> 921  -1.116815e-01  0.2203772268
#> 922  -3.197149e-01  0.3206349422
#> 923  -3.556226e-02  0.0271776404
#> 924   1.973905e-02 -0.0625703381
#> 925  -2.449372e-01  0.1907037120
#> 926  -6.238317e-02  0.1131914568
#> 927  -1.447265e-01 -0.0171146426
#> 928  -1.220219e-01  0.0087341432
#> 929  -7.717840e-02  0.0501709090
#> 930  -6.506285e-03 -0.0258647696
#> 931  -3.650030e-01  0.1881017328
#> 932  -7.743466e-02  0.1173933788
#> 933  -2.266255e-01  0.1213557963
#> 934   2.656983e-02 -0.0402475103
#> 935  -1.885906e-02 -0.0088008131
#> 936  -8.944074e-02 -0.0382748675
#> 937  -1.858970e-01 -0.0407016214
#> 938  -1.650794e-01  0.1199800120
#> 939  -2.033783e-01  0.0471134668
#> 940  -1.668751e-01 -0.0264786786
#> 941  -5.763394e-02  0.0374732838
#> 942  -3.584557e-02 -0.0155466095
#> 943  -1.264960e-01  0.1150474511
#> 944  -2.058487e-01  0.1762715199
#> 945  -1.060783e-01  0.0956304035
#> 946  -2.769053e-02 -0.1026450074
#> 947  -1.300784e-01  0.2048950140
#> 948  -3.474757e-01  0.1644565528
#> 949  -1.792459e-01  0.0180505467
#> 950  -1.819263e-01  0.0461813480
#> 951  -1.146847e-01  0.0626596992
#> 952  -8.364196e-02  0.0224329781
#> 953  -1.220536e-01  0.2397407826
#> 954  -4.256670e-02 -0.2356202393
#> 955   4.849149e-02 -0.0628849541
#> 956  -2.641192e-01  0.0477950189
#> 957   6.146530e-02 -0.0903429459
#> 958  -6.406905e-02  0.0184049171
#> 959  -3.162788e-01  0.2676468306
#> 960   4.283728e-02 -0.0630179228
#> 961  -6.890982e-02  0.0241949658
#> 962  -1.532313e-01  0.0941862859
#> 963  -7.908502e-02 -0.0116091224
#> 964  -7.485328e-02  0.0680836745
#> 965  -2.914193e-01  0.0716681206
#> 966  -4.180271e-02 -0.0225266858
#> 967   4.321585e-02  0.0397269446
#> 968  -2.302104e-01  0.0898419247
#> 969  -4.324182e-02  0.2437970076
#> 970  -1.181300e-01  0.1446007443
#> 971  -9.746579e-02  0.1226563257
#> 972  -7.460353e-02 -0.0287043987
#> 973  -9.414181e-03 -0.0743628684
#> 974  -8.408381e-02 -0.0075330302
#> 975  -1.184499e-01 -0.0888957788
#> 976  -5.715705e-02  0.0779321123
#> 977  -7.485328e-02  0.0680836745
#> 978  -1.535133e-01  0.2487647747
#> 979  -9.745104e-02  0.0167729542
#> 980  -1.812760e-01  0.0848824401
#> 981  -4.567267e-02  0.1774931673
#> 982  -6.577126e-02  0.1478542663
#> 983  -5.930148e-02  0.0182262495
#> 984  -2.597485e-01  0.0513478120
#> 985  -1.100364e-01  0.0842107270
#> 986  -1.129571e-01  0.0407985337
#> 987  -1.065152e-01  0.0422267986
#> 988  -2.884782e-01  0.2073536578
#> 989  -1.056314e-01  0.0594411221
#> 990  -8.809149e-02  0.0897549227
#> 991  -1.181300e-01  0.1446007443
#> 992  -1.140681e-02 -0.1010407413
#> 993  -1.595831e-01  0.2272813915
#> 994  -2.888811e-01  0.2226492912
#> 995  -4.871620e-02 -0.0745515316
#> 996  -1.810268e-01  0.1984046822
#> 997  -1.812760e-01  0.0848824401
#> 998  -1.663957e-02  0.0083165555
#> 999  -7.403852e-02  0.0924258828
#> 1000 -6.404233e-03  0.1340769454
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 1,000 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.0691  0.101 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.0427 -0.112 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.143   0.322 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.136   0.147 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.122   0.0879
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0603 -0.251 
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0818 -0.110 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.262   0.159 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.119  -0.0228
#> 10     10     0.21      80    65    19.8     0.041      0.12  0.0248  0.0391
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2,000 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.232  21.5 0.041  0.12 75.8  1633. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.232  21.5 0.041  0.12  4.69  101. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.239  17.4 0.041  0.12 93.4  1625. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.239  17.4 0.041  0.12  5.45   94.8
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.216  26.9 0.041  0.12 61.5  1653. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.216  26.9 0.041  0.12  4.47  120. 
#> # … with 1,994 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>