Estimate the posterior distribution of individual parameters by SIR

Estimates the posterior distribution of individual parameters by Sequential Importance Resampling (SIR)

poso_estim_sir(
  object,
  n_sample = 10000,
  n_resample = 1000,
  return_model = TRUE
)

Arguments

object

A posologyr list, created by the posologyr function.

n_sample

Number of samples from the S-step

n_resample

Number of samples from the R-step

return_model

A boolean. Returns a rxode2 model using the estimated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of ETAs from the posterior distribution, estimated by Sequential Importance Resampling. If return_model is set to TRUE, a list of the dataframe of the posterior distribution of ETA, and a rxode2 model using the estimated distributions of ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the posterior distribution of population parameters
poso_estim_sir(patient01,n_sample=1e4,n_resample=1e3)
#> $eta
#>            ETA_Cl        ETA_Vc        ETA_Ka
#> 1     0.593350858 -4.559244e-01  0.3110338165
#> 2     0.783322355 -2.880338e-01  0.1651715412
#> 3     0.526453254 -7.381857e-01 -0.2459812392
#> 4     0.577793291 -3.978982e-01  0.7188403698
#> 5     0.449026838 -6.722857e-01 -0.3449094567
#> 6     0.532109310 -2.790201e-01  0.5045332616
#> 7     0.552043853 -5.541772e-01 -0.2038343609
#> 8     0.428626732 -1.285676e-01  0.5570232057
#> 9     0.757234376 -2.914318e-01  0.2739550422
#> 10    0.585119599 -3.715712e-01  0.3239969829
#> 11    0.648195786 -2.438098e-01  0.3384888413
#> 12    0.517242279 -2.406567e-01  1.0868602875
#> 13    0.489085113 -4.143278e-01 -0.0499112054
#> 14    0.645765952 -2.416312e-01  0.4913788639
#> 15    0.366560145 -2.741978e-01  0.4919368177
#> 16    0.607315954  1.916104e-01  0.8100374894
#> 17    0.648195786 -2.438098e-01  0.3384888413
#> 18   -0.005973738 -8.546492e-05  0.4118845832
#> 19    0.764106831 -2.203832e-01  0.2281586119
#> 20    0.163888781 -1.225429e+00 -0.7563132923
#> 21    0.368259543 -8.662150e-01  0.1160803796
#> 22    0.556442923 -3.035760e-01  0.2941478223
#> 23    0.436760735 -3.464795e-01  0.1050979713
#> 24    0.672421989 -6.054193e-01 -0.3552440114
#> 25    0.608943479 -3.943185e-01  0.3471474790
#> 26    0.674586332 -2.186576e-02  0.2258184790
#> 27    0.642436618 -3.096014e-01  1.0206356897
#> 28    0.507926367 -8.444463e-01  0.1555297731
#> 29    0.261637323 -1.261889e-01  0.6004965066
#> 30    0.272137385 -5.907316e-01  0.1895019834
#> 31    0.584769275 -2.923607e-02  0.5448847377
#> 32    0.368911983 -1.495586e-01  0.4893458503
#> 33    0.538854278 -3.902258e-01  0.6073351121
#> 34    0.128856571 -7.294398e-01  0.2321037465
#> 35    0.644114141 -6.862536e-01 -0.1444875728
#> 36    0.632546756 -2.454814e-01  0.2798429447
#> 37    0.238271432 -5.177131e-01  0.5035802412
#> 38    0.353679123 -3.012061e-01  0.2606980112
#> 39    0.639880801 -1.163314e-01  0.3755483357
#> 40    0.588258239 -3.805277e-01  0.3616792478
#> 41    0.594633305 -6.461874e-01 -0.2091563043
#> 42    0.381687605 -8.385829e-01 -0.6070146985
#> 43    0.512934913 -7.625702e-01  0.1566112353
#> 44    0.448883294 -8.337454e-01 -0.0649878065
#> 45    0.168130602 -4.429217e-01 -0.0317110506
#> 46    0.292572872 -6.333308e-01 -0.2342678405
#> 47    0.503072934  9.082159e-02  1.0386906279
#> 48    0.510623308 -3.073274e-01  0.1175567215
#> 49    0.379180273 -8.924712e-01 -0.3803001775
#> 50    0.721152315 -2.562438e-01  0.2438471748
#> 51    0.074271299 -7.862067e-01  0.1257464935
#> 52    0.558360435 -2.566795e-01  0.7095299285
#> 53    0.662207648 -5.834187e-01  0.1127272000
#> 54    0.212222759 -9.347593e-01 -0.7992976259
#> 55    0.272379360 -1.148700e+00 -0.4945985661
#> 56    0.628123482  2.439737e-02  0.8703916019
#> 57    0.351680153 -6.131780e-01  0.3480300339
#> 58    0.599749042 -2.392847e-01  0.0523817355
#> 59    0.635097538 -4.216498e-01 -0.2131359689
#> 60    0.532109310 -2.790201e-01  0.5045332616
#> 61    0.635097538 -4.216498e-01 -0.2131359689
#> 62    0.441599633 -2.589813e-01  0.1790411566
#> 63    0.677762516  8.534585e-02  0.6601944356
#> 64    0.177364107 -1.169776e+00 -0.1677328787
#> 65    0.077984516 -8.307965e-01  0.6471059825
#> 66    0.230163322 -1.970709e-01  0.2996280883
#> 67    0.442517546 -2.687531e-01  0.2161358272
#> 68    0.724902634 -5.403216e-02  0.9153594442
#> 69    0.408518619 -4.298715e-01  0.1424477714
#> 70    0.744696338 -2.307630e-01  0.0253665491
#> 71    0.317442739 -6.613272e-01 -0.2077336701
#> 72    0.615939454 -5.379870e-01 -0.2196287781
#> 73    0.149200840 -3.158523e-01  0.4973511627
#> 74    0.748374324 -3.522033e-01  0.3217783071
#> 75    0.648195786 -2.438098e-01  0.3384888413
#> 76    0.782804389 -1.230561e-01  0.5593653197
#> 77    0.429362512 -6.796696e-01  0.0568761747
#> 78    0.370541513 -3.290399e-01  0.0504144916
#> 79    0.584137050 -8.477569e-01 -0.6664834383
#> 80    0.403897432 -8.167510e-01  0.3771163220
#> 81    0.436861353 -5.854290e-01  0.1681859622
#> 82    0.520770009 -5.826589e-01  0.1660644431
#> 83    0.250533015 -8.770750e-01  0.2315841852
#> 84    0.436523346  3.458933e-02  1.5608720381
#> 85    0.623226871 -2.544954e-01  0.4074802169
#> 86    0.467847330 -4.169080e-01  0.2663687227
#> 87    0.621371011 -3.454422e-01 -0.1131285910
#> 88    0.315894383 -5.350913e-01  0.2590958976
#> 89    0.804245794 -1.282290e-01  0.7688563904
#> 90    0.679412792 -2.939501e-01  0.7264710332
#> 91    0.648195786 -2.438098e-01  0.3384888413
#> 92    0.590288655 -4.554547e-01  0.1860579026
#> 93    0.672421989 -6.054193e-01 -0.3552440114
#> 94    0.414408331 -1.430922e-01  0.4479315410
#> 95    0.524345425 -4.109664e-01  0.6193419756
#> 96    0.805210269 -5.120784e-02  0.7495099620
#> 97    0.244623792 -3.452596e-01 -0.0155414742
#> 98    0.644233056 -3.618588e-01  0.5398658697
#> 99    0.567708994 -1.605274e-01  0.0617065127
#> 100   0.752002325  6.117967e-03  0.6060843364
#> 101   0.340521012 -4.852684e-01  0.1287410295
#> 102   0.805210269 -5.120784e-02  0.7495099620
#> 103   0.363115386 -3.977127e-01  0.0342988392
#> 104   0.659454363 -2.233273e-01  0.2914213179
#> 105   0.783322355 -2.880338e-01  0.1651715412
#> 106   0.608943479 -3.943185e-01  0.3471474790
#> 107   0.583793464 -2.536374e-01  0.3991654172
#> 108   0.544277569 -5.992853e-01 -0.2664856370
#> 109   0.770089341 -1.510804e-01  0.0506789791
#> 110   0.407697870 -7.995766e-01  0.0091022467
#> 111   0.805210269 -5.120784e-02  0.7495099620
#> 112   0.316151033 -6.212005e-01  0.0772117373
#> 113   0.608943479 -3.943185e-01  0.3471474790
#> 114   0.593350858 -4.559244e-01  0.3110338165
#> 115   0.429362512 -6.796696e-01  0.0568761747
#> 116   0.416231162 -3.676731e-01 -0.1215392063
#> 117   0.750748763 -3.773923e-01  0.2374938517
#> 118   0.701547012 -5.628105e-01  0.2011065529
#> 119   0.177090943 -8.469938e-02  0.6850882904
#> 120   0.590288655 -4.554547e-01  0.1860579026
#> 121   0.672421989 -6.054193e-01 -0.3552440114
#> 122   0.435950015 -3.904839e-01 -0.0142967372
#> 123   0.634230572 -1.425292e-01  0.5569870258
#> 124   0.371924224 -7.047714e-01 -0.1305547081
#> 125   0.390666232 -5.331126e-01 -0.0336801421
#> 126   0.544277569 -5.992853e-01 -0.2664856370
#> 127   0.755948668 -1.120205e-01  0.4576099606
#> 128   0.615939454 -5.379870e-01 -0.2196287781
#> 129   0.713228998 -2.525679e-01  0.2003398689
#> 130   0.583793464 -2.536374e-01  0.3991654172
#> 131   0.564719272 -6.500368e-01  0.2321579090
#> 132   0.607645003  1.141888e-01  0.8889284168
#> 133   0.433114948 -1.428214e-01  0.3945448970
#> 134   0.755948668 -1.120205e-01  0.4576099606
#> 135   0.982425722  9.660679e-02  0.4943304661
#> 136   0.403897432 -8.167510e-01  0.3771163220
#> 137   0.678252212  1.104094e-01  0.5951892858
#> 138   0.615939454 -5.379870e-01 -0.2196287781
#> 139   0.108272733 -5.358955e-01  0.3918890003
#> 140   0.845481055 -1.049904e-01  0.6661489092
#> 141   0.166828801 -2.830137e-01  0.4587394653
#> 142   0.574984052 -1.019722e-01  0.3842329842
#> 143   0.488053881 -2.258137e-01  0.8647453701
#> 144   0.448883294 -8.337454e-01 -0.0649878065
#> 145   0.548455152 -4.266172e-01  0.3917507046
#> 146   0.507270232 -4.604752e-01  0.4179064941
#> 147   0.249853202 -8.815013e-01  0.0476049561
#> 148   0.381798127 -6.491273e-01  0.0019473093
#> 149   0.577697092 -6.610146e-01 -0.0761771440
#> 150   0.656332163 -4.103918e-01 -0.0209797670
#> 151   0.486400430 -5.310261e-01 -0.1917685435
#> 152   0.486400430 -5.310261e-01 -0.1917685435
#> 153   0.399040234 -5.655249e-01  0.3919373970
#> 154   0.748374324 -3.522033e-01  0.3217783071
#> 155   0.185570858 -5.458347e-01 -0.3139503823
#> 156   0.234273492 -7.283026e-01 -0.3324160759
#> 157   0.510229147 -3.943776e-01  0.0578260620
#> 158   0.505999681 -7.966699e-01 -0.4136578479
#> 159   0.900910422 -2.713757e-01  0.2891482540
#> 160   0.459802796 -2.133903e-01  0.5015794874
#> 161   0.632677784 -2.790618e-01  0.3937965735
#> 162   0.835596377 -1.748454e-01  0.5968263163
#> 163   0.074271299 -7.862067e-01  0.1257464935
#> 164   0.520923446 -6.641042e-01 -0.2209728459
#> 165   0.401619765 -2.217401e-01  0.0287423309
#> 166   0.710019943 -3.205738e-01  0.0280784958
#> 167   0.657127811 -2.995528e-01  0.1240659947
#> 168   0.445274775 -2.120666e-01  0.9955616608
#> 169   0.429362512 -6.796696e-01  0.0568761747
#> 170   0.602223101 -3.244704e-01  0.5681096091
#> 171  -0.261699710 -9.038298e-01  0.2072689020
#> 172   0.691163701 -1.516967e-01  1.6704002614
#> 173   0.368395143 -4.506515e-01  0.5491257308
#> 174   0.535355672 -8.141930e-01 -0.1366474433
#> 175   0.354891031 -3.950728e-01  0.1536995811
#> 176   0.743337247 -5.168349e-01 -0.6784750836
#> 177   0.398345758 -9.344078e-01  0.0297452395
#> 178   0.366814610 -8.115863e-01 -0.2384443412
#> 179   0.364433365 -9.505407e-01 -0.2192605382
#> 180   0.604864836 -2.236177e-01  0.3762682936
#> 181   0.756082446 -3.628444e-01  0.1734604610
#> 182   0.549627254 -4.228628e-01 -0.2610124956
#> 183   0.468620058 -7.350085e-01 -0.0770598555
#> 184   0.593350858 -4.559244e-01  0.3110338165
#> 185   0.366393963 -3.508074e-01  0.5847888313
#> 186   0.836028487 -1.823020e-01  0.3787170554
#> 187   0.407697870 -7.995766e-01  0.0091022467
#> 188   0.307288705 -6.414479e-01 -0.1629205745
#> 189   0.764106831 -2.203832e-01  0.2281586119
#> 190   0.784629285 -1.249499e-01  0.4880350023
#> 191   0.764106831 -2.203832e-01  0.2281586119
#> 192   0.678193887 -9.366769e-02  0.9191288732
#> 193   0.422230220 -3.954542e-01  0.8448324525
#> 194   0.588089944 -5.116597e-01 -0.0911808177
#> 195   0.544277569 -5.992853e-01 -0.2664856370
#> 196   0.561668702 -6.902955e-01 -0.2680225618
#> 197   0.720282742 -3.114853e-01 -0.1187426232
#> 198   0.347666399 -3.296001e-01  0.5246696331
#> 199   0.744696338 -2.307630e-01  0.0253665491
#> 200   0.637614997 -6.553868e-02  0.4862096598
#> 201   0.592986240 -1.560459e-01  0.1373980186
#> 202   0.303724530 -3.981489e-01  0.0851979488
#> 203   0.710828107 -3.787493e-01 -0.0995455894
#> 204   0.465569640 -3.288849e-01  0.8931608019
#> 205   0.459575558 -3.561530e-01  0.3533101601
#> 206   0.480123503 -5.719025e-01 -0.0186538211
#> 207   0.595533513 -3.841880e-01  0.0624969249
#> 208   0.374229546 -2.997070e-01  0.3923300442
#> 209   0.733653849 -3.634514e-01 -0.1165163764
#> 210   0.408518619 -4.298715e-01  0.1424477714
#> 211   0.468620058 -7.350085e-01 -0.0770598555
#> 212   0.696324376 -1.063341e-01  0.2169611667
#> 213   0.784629285 -1.249499e-01  0.4880350023
#> 214   0.107840739 -1.242024e-01  0.3990420262
#> 215   0.814047616 -2.968562e-01  0.5985407473
#> 216   0.632677784 -2.790618e-01  0.3937965735
#> 217   0.637117694 -5.010288e-01 -0.0163940628
#> 218   0.477449672 -5.544117e-01 -0.2163620373
#> 219   0.348781558 -8.064140e-01 -0.1935796641
#> 220   0.547621899 -4.747561e-01  0.1164947153
#> 221   0.511248394 -1.039213e-01  0.3325529123
#> 222   0.547621899 -4.747561e-01  0.1164947153
#> 223   0.585119599 -3.715712e-01  0.3239969829
#> 224   0.520923446 -6.641042e-01 -0.2209728459
#> 225   0.856272269  3.579527e-02  0.6225082578
#> 226   0.429362512 -6.796696e-01  0.0568761747
#> 227   0.690384308 -1.112801e-01  1.2935257885
#> 228   0.472527278 -1.631860e-01  0.5091226158
#> 229   0.738089346 -5.083571e-01 -0.1744905043
#> 230   0.492525661 -7.777571e-01 -0.4494138113
#> 231   0.399855279 -7.168713e-01  0.1798173945
#> 232   0.473176768 -9.061856e-01 -0.4365924548
#> 233   0.748374324 -3.522033e-01  0.3217783071
#> 234   0.428626732 -1.285676e-01  0.5570232057
#> 235   0.755948668 -1.120205e-01  0.4576099606
#> 236   0.371924224 -7.047714e-01 -0.1305547081
#> 237   0.214102475 -5.148216e-01  0.9423414553
#> 238   0.724902634 -5.403216e-02  0.9153594442
#> 239   0.448363199 -5.899503e-01  0.0237764062
#> 240   0.752002325  6.117967e-03  0.6060843364
#> 241   0.724902634 -5.403216e-02  0.9153594442
#> 242   0.635097538 -4.216498e-01 -0.2131359689
#> 243   0.740591346 -5.512911e-02  0.7390198780
#> 244   0.637049886  1.275622e-01  0.8465630398
#> 245   0.364433365 -9.505407e-01 -0.2192605382
#> 246   0.814047616 -2.968562e-01  0.5985407473
#> 247   0.659368427 -1.448032e-01  0.6094278631
#> 248   0.746569888 -3.212300e-01  0.3150573295
#> 249   0.714289649 -2.919856e-01  0.2220296838
#> 250   0.379225046 -3.811862e-01  0.4252382340
#> 251   0.680056335 -3.772264e-01  0.3020396915
#> 252   0.465747019 -8.088830e-01 -0.6042918030
#> 253   0.592986240 -1.560459e-01  0.1373980186
#> 254   0.259862981 -1.581124e-01  0.2428883670
#> 255   0.547420206 -4.200971e-01  0.1206842168
#> 256   0.623226871 -2.544954e-01  0.4074802169
#> 257   0.520923446 -6.641042e-01 -0.2209728459
#> 258   0.728408571  2.048086e-01  1.2927125679
#> 259   0.351680153 -6.131780e-01  0.3480300339
#> 260   0.589354938 -8.565276e-03  0.6282006667
#> 261   0.541423669 -3.018876e-01  0.2847904272
#> 262   0.823666383  2.155409e-02  0.2456147949
#> 263   0.845481055 -1.049904e-01  0.6661489092
#> 264   0.614934205 -8.639657e-03  0.2886548486
#> 265   0.643912963 -3.289597e-01  0.2791175717
#> 266   0.448217110 -6.334806e-01 -0.1520780309
#> 267   0.505999681 -7.966699e-01 -0.4136578479
#> 268   0.486400430 -5.310261e-01 -0.1917685435
#> 269   0.486400430 -5.310261e-01 -0.1917685435
#> 270  -0.263719801 -5.111568e-01 -0.1270823223
#> 271   0.509567963 -4.681300e-01  0.2287899397
#> 272   0.805647796 -1.867938e-01  0.4291724022
#> 273   0.593350858 -4.559244e-01  0.3110338165
#> 274   0.877457484 -2.256131e-01  0.2683269313
#> 275   0.364918315 -4.994674e-01  0.2585559051
#> 276   0.677606874 -4.651721e-01 -0.0953820547
#> 277   0.920614757  4.815350e-03  0.2634697511
#> 278   0.434359762 -6.425417e-01  0.3466628601
#> 279   0.499327196 -4.806236e-01 -0.1409412324
#> 280   0.714289649 -2.919856e-01  0.2220296838
#> 281   0.578028570 -4.130846e-01  0.2883536217
#> 282   0.805647796 -1.867938e-01  0.4291724022
#> 283   0.462850443 -3.790177e-01 -0.0098725053
#> 284   0.575161380 -2.676681e-01 -0.1309425452
#> 285   0.733653849 -3.634514e-01 -0.1165163764
#> 286   0.623226871 -2.544954e-01  0.4074802169
#> 287   0.688709897 -2.625539e-01 -0.0205189106
#> 288   0.577697092 -6.610146e-01 -0.0761771440
#> 289   0.600942354 -4.990908e-01  0.1511872520
#> 290   0.567708994 -1.605274e-01  0.0617065127
#> 291   0.497888904  1.051552e-01  0.7193014277
#> 292   0.538633064 -4.687045e-01  0.6827850446
#> 293   0.227856637 -2.181062e-01  0.6757783898
#> 294   0.382010904 -1.896653e-01  0.2569877436
#> 295   0.657127811 -2.995528e-01  0.1240659947
#> 296   0.293897379 -7.264297e-01  0.1809269100
#> 297   0.588258239 -3.805277e-01  0.3616792478
#> 298   0.804245794 -1.282290e-01  0.7688563904
#> 299   0.516156041 -5.040512e-01 -0.0268062905
#> 300   0.060257708 -9.928131e-01 -0.3290638344
#> 301   0.227891269 -4.096901e-01  0.1392472992
#> 302   0.551883801 -5.247510e-01  0.0886974812
#> 303   0.708245660 -2.832233e-01 -0.0076840747
#> 304   0.510229147 -3.943776e-01  0.0578260620
#> 305   0.523910806 -6.652104e-01 -0.3188112208
#> 306   0.784629285 -1.249499e-01  0.4880350023
#> 307   0.450569216 -7.363590e-01  0.5146942629
#> 308   0.493308079  4.077762e-02  1.0260510975
#> 309   0.593112391 -4.451551e-01  0.1517529765
#> 310   0.271970806 -9.338122e-01  0.2603590014
#> 311   0.836028487 -1.823020e-01  0.3787170554
#> 312   0.552043853 -5.541772e-01 -0.2038343609
#> 313   0.632546756 -2.454814e-01  0.2798429447
#> 314   0.643912963 -3.289597e-01  0.2791175717
#> 315   0.389591347 -6.670816e-01  0.1099298561
#> 316   0.306723011 -9.111968e-01  0.0034371432
#> 317   0.320864804 -8.127956e-01 -0.2713695769
#> 318   0.344403634 -4.293977e-01  1.0058381066
#> 319   0.549627254 -4.228628e-01 -0.2610124956
#> 320   0.364433365 -9.505407e-01 -0.2192605382
#> 321   0.645467411 -1.291593e-01  0.0899274837
#> 322   0.746569888 -3.212300e-01  0.3150573295
#> 323   0.523910806 -6.652104e-01 -0.3188112208
#> 324   0.017474255 -3.125038e-01  0.2518616614
#> 325   0.396831246 -5.040956e-01 -0.0778142895
#> 326   0.492525661 -7.777571e-01 -0.4494138113
#> 327   0.440259324  4.655827e-02  0.4949944933
#> 328   0.615939454 -5.379870e-01 -0.2196287781
#> 329   0.585119599 -3.715712e-01  0.3239969829
#> 330   0.581306531 -3.175423e-01  0.6381925200
#> 331   0.637614997 -6.553868e-02  0.4862096598
#> 332   0.576278609 -3.135657e-01  0.0788084571
#> 333   0.603819001 -1.859605e-01  0.2396973162
#> 334   0.049905884 -5.260243e-01  0.0279022992
#> 335   0.674586332 -2.186576e-02  0.2258184790
#> 336   0.671903386 -4.941620e-01  0.3541257905
#> 337   0.672637770 -6.699005e-01 -0.3745436581
#> 338   0.538854278 -3.902258e-01  0.6073351121
#> 339   0.588258239 -3.805277e-01  0.3616792478
#> 340   0.632546756 -2.454814e-01  0.2798429447
#> 341   0.229689512 -8.021361e-01  0.2981475410
#> 342   0.337529364 -2.977638e-01  0.5289414161
#> 343   0.507449139 -2.138341e-01  0.4677717192
#> 344   0.710019943 -3.205738e-01  0.0280784958
#> 345   0.817277383 -2.949467e-01 -0.1250724519
#> 346   0.043748297 -6.026617e-01  0.2094918705
#> 347   0.600942354 -4.990908e-01  0.1511872520
#> 348   0.394083478 -7.810359e-01 -0.3382743599
#> 349   0.749820914 -1.598008e-01  0.9102406234
#> 350   0.590981464  8.206783e-02  0.9547959575
#> 351   0.250151980 -6.565963e-01 -0.1933450730
#> 352   0.644233056 -3.618588e-01  0.5398658697
#> 353   0.416355054  6.672438e-02  0.9312712867
#> 354   0.623589670 -5.066256e-01  0.1322400569
#> 355   0.900396676 -2.058962e-01 -0.0107293586
#> 356   0.805210269 -5.120784e-02  0.7495099620
#> 357   0.549901524 -4.649401e-01  0.3258372420
#> 358   0.599749042 -2.392847e-01  0.0523817355
#> 359   0.782804389 -1.230561e-01  0.5593653197
#> 360   0.379180273 -8.924712e-01 -0.3803001775
#> 361   0.422650930 -6.210379e-01 -0.4008864576
#> 362   0.599749042 -2.392847e-01  0.0523817355
#> 363   0.551883801 -5.247510e-01  0.0886974812
#> 364   0.084796152 -4.006775e-01  0.2841973825
#> 365   0.537105064 -3.281645e-01  0.4903216634
#> 366   0.508578512 -5.870037e-01  0.4917318576
#> 367   0.291822216 -2.367719e-01  0.5430270140
#> 368   0.456224905 -7.778298e-01 -0.4748066121
#> 369   0.600942354 -4.990908e-01  0.1511872520
#> 370   0.193494288 -1.014295e+00 -0.3992477904
#> 371   0.551883801 -5.247510e-01  0.0886974812
#> 372   0.548455152 -4.266172e-01  0.3917507046
#> 373   0.569205799 -1.056344e+00 -0.7697510714
#> 374   0.211011152 -2.261018e-02  0.5275552200
#> 375   0.504253510 -6.189435e-01  0.4356198426
#> 376   0.376801586 -4.514124e-01  0.6652674412
#> 377   0.720724014 -4.602511e-01 -0.2404747045
#> 378   0.745675206 -5.429329e-01 -0.3945019445
#> 379   0.676273062 -6.339105e-01 -0.4625058042
#> 380  -0.094045397 -3.295268e-01  0.7909619410
#> 381   0.365826496 -1.889788e-03  0.5875204913
#> 382   0.524345425 -4.109664e-01  0.6193419756
#> 383   0.768254754 -8.578182e-02  0.8556270930
#> 384   0.376801586 -4.514124e-01  0.6652674412
#> 385   0.714289649 -2.919856e-01  0.2220296838
#> 386   0.676007840 -3.426761e-01  0.7109712026
#> 387   0.512934913 -7.625702e-01  0.1566112353
#> 388   0.593350858 -4.559244e-01  0.3110338165
#> 389   0.624231043 -3.176884e-01 -0.0375486128
#> 390   0.520770009 -5.826589e-01  0.1660644431
#> 391   0.528613652 -4.351960e-01  0.7630974177
#> 392   0.348781558 -8.064140e-01 -0.1935796641
#> 393   0.477449672 -5.544117e-01 -0.2163620373
#> 394   0.507463464 -3.673248e-01  0.1091212085
#> 395   0.583793464 -2.536374e-01  0.3991654172
#> 396   0.559480500 -1.781126e-01  0.4374700312
#> 397   0.516156041 -5.040512e-01 -0.0268062905
#> 398   0.551883801 -5.247510e-01  0.0886974812
#> 399   0.770190983 -2.344266e-01  0.2321999164
#> 400   0.559480500 -1.781126e-01  0.4374700312
#> 401   0.426465495  1.774345e-02  0.8230779417
#> 402   0.438494314 -1.045713e-02  0.7201309800
#> 403   0.593350858 -4.559244e-01  0.3110338165
#> 404   0.157882039 -5.384849e-01  0.2876993763
#> 405   0.341546523 -6.154995e-01  0.2432556881
#> 406   0.752002325  6.117967e-03  0.6060843364
#> 407   0.623589670 -5.066256e-01  0.1322400569
#> 408   0.678193887 -9.366769e-02  0.9191288732
#> 409   0.476688932 -3.793997e-02  0.5285227061
#> 410   0.649373210 -1.690043e-01  0.2558348152
#> 411   0.511733018 -1.909648e-01  0.1941408789
#> 412   0.752002325  6.117967e-03  0.6060843364
#> 413   0.192534331 -5.542634e-01  0.0433641082
#> 414   0.445701382 -4.473227e-01  0.3573777441
#> 415   0.257273747 -8.650205e-01 -0.1043906966
#> 416   0.642436618 -3.096014e-01  1.0206356897
#> 417   0.410268863 -3.742628e-01  0.2915421846
#> 418   0.746569888 -3.212300e-01  0.3150573295
#> 419   0.429362512 -6.796696e-01  0.0568761747
#> 420   0.347666399 -3.296001e-01  0.5246696331
#> 421   0.396831246 -5.040956e-01 -0.0778142895
#> 422   0.665355184 -1.851128e-01  0.0597237171
#> 423   0.345987208 -5.989845e-01 -0.0488895409
#> 424   0.724902634 -5.403216e-02  0.9153594442
#> 425   0.609965398 -3.495247e-01  0.3750695023
#> 426   0.214570511 -1.542935e-01  0.4050745623
#> 427   0.764106831 -2.203832e-01  0.2281586119
#> 428   0.430838735 -5.282053e-01  0.2912414182
#> 429   0.763464014  1.302456e-01  0.8186591232
#> 430   0.449118917 -8.808611e-01  0.2914141003
#> 431   0.602223101 -3.244704e-01  0.5681096091
#> 432   0.558360435 -2.566795e-01  0.7095299285
#> 433   0.659164276 -6.382710e-01 -0.3558369088
#> 434   0.480123503 -5.719025e-01 -0.0186538211
#> 435   0.471890127 -6.529326e-01 -0.3452922221
#> 436   0.566449742 -4.158066e-01  0.2072529138
#> 437   0.714289649 -2.919856e-01  0.2220296838
#> 438   0.678585471 -1.755864e-01  0.6162667653
#> 439   0.608943479 -3.943185e-01  0.3471474790
#> 440   0.464913144 -4.140672e-02  0.3649211730
#> 441   0.215103328 -9.057674e-01 -0.5976333475
#> 442   0.672421989 -6.054193e-01 -0.3552440114
#> 443   0.384520854 -4.460456e-01 -0.1034287978
#> 444   0.662953241 -5.669931e-01  0.3030210027
#> 445   0.621371011 -3.454422e-01 -0.1131285910
#> 446   0.643912963 -3.289597e-01  0.2791175717
#> 447   0.594633305 -6.461874e-01 -0.2091563043
#> 448   0.537105064 -3.281645e-01  0.4903216634
#> 449   0.317442739 -6.613272e-01 -0.2077336701
#> 450   0.477449672 -5.544117e-01 -0.2163620373
#> 451   0.414837287 -6.344746e-01 -0.2698998082
#> 452   0.234673230 -4.272357e-01 -0.1693143628
#> 453   0.755948668 -1.120205e-01  0.4576099606
#> 454   0.472527278 -1.631860e-01  0.5091226158
#> 455   0.371924224 -7.047714e-01 -0.1305547081
#> 456   0.757234376 -2.914318e-01  0.2739550422
#> 457   0.837428030  1.888835e-01  0.8106157681
#> 458   0.707244992 -3.087660e-01 -0.1579569724
#> 459   0.141021508 -4.046460e-01  0.0387565779
#> 460   0.432881420 -7.878095e-01 -0.2708468484
#> 461   0.314029840 -3.643150e-01  0.1532121371
#> 462   0.386514824 -4.419072e-01  0.2657324569
#> 463   0.634230572 -1.425292e-01  0.5569870258
#> 464   0.551883801 -5.247510e-01  0.0886974812
#> 465   0.511649271 -3.257325e-01  0.4880000542
#> 466   0.216284814 -4.112872e-01  0.4907159498
#> 467   0.784629285 -1.249499e-01  0.4880350023
#> 468   0.537105064 -3.281645e-01  0.4903216634
#> 469  -0.551379473  4.389429e-02  0.5198644436
#> 470  -0.202371246 -5.173068e-01 -0.0881726364
#> 471   0.021026191 -6.502304e-01 -0.0554084458
#> 472   0.920614757  4.815350e-03  0.2634697511
#> 473   0.772898372 -3.003373e-01  0.3118079962
#> 474   0.436760735 -3.464795e-01  0.1050979713
#> 475   0.714289649 -2.919856e-01  0.2220296838
#> 476   0.604291862 -5.617658e-01  0.4607442049
#> 477   0.376833187 -5.896055e-01  0.1860172421
#> 478   0.501071993 -4.519316e-01  0.4668503402
#> 479   0.749820914 -1.598008e-01  0.9102406234
#> 480   0.724902634 -5.403216e-02  0.9153594442
#> 481   0.478285844 -7.541659e-01  0.2360577348
#> 482   0.645765952 -2.416312e-01  0.4913788639
#> 483   0.463648938 -6.553450e-01  0.1439635457
#> 484  -0.101930041 -3.877616e-02  0.8049835350
#> 485   0.644562275 -1.412364e-01  0.4165147513
#> 486   0.523910806 -6.652104e-01 -0.3188112208
#> 487   0.542738800 -8.829751e-01 -0.3371784247
#> 488   0.384109592 -3.123773e-01  0.6640479796
#> 489   0.177364107 -1.169776e+00 -0.1677328787
#> 490   0.721152315 -2.562438e-01  0.2438471748
#> 491   0.523910806 -6.652104e-01 -0.3188112208
#> 492   0.679828484 -3.889512e-01  0.3147578797
#> 493   0.471890127 -6.529326e-01 -0.3452922221
#> 494   0.648195786 -2.438098e-01  0.3384888413
#> 495   0.495232022 -2.258763e-01  0.3229901619
#> 496   0.856272269  3.579527e-02  0.6225082578
#> 497   0.831108301 -1.053882e-01  0.6556810033
#> 498   0.561668702 -6.902955e-01 -0.2680225618
#> 499   0.332321460 -2.826474e-01  0.3912279464
#> 500   0.837428030  1.888835e-01  0.8106157681
#> 501  -0.143882776  5.622670e-02  0.4475858958
#> 502   0.760085191  1.539284e-01  0.6716064289
#> 503   0.253424921 -1.084298e-01  1.3031878021
#> 504   0.713228998 -2.525679e-01  0.2003398689
#> 505   0.492976212 -6.573783e-01  0.2104361620
#> 506   0.645765952 -2.416312e-01  0.4913788639
#> 507   0.675033211 -4.111685e-01  0.0161866498
#> 508   0.106233652 -1.114541e+00 -0.0599480124
#> 509   0.459575558 -3.561530e-01  0.3533101601
#> 510   0.635097538 -4.216498e-01 -0.2131359689
#> 511   0.239791497 -3.916593e-01  0.2756093129
#> 512   0.172022987 -9.798900e-02  0.5840368715
#> 513   0.777212700 -2.935970e-01  0.1834490101
#> 514   0.563714276 -2.655455e-01  0.3149603741
#> 515   0.621840910 -3.221962e-01  0.7670927626
#> 516   0.527151866 -6.386879e-01 -0.3390576286
#> 517   0.524345425 -4.109664e-01  0.6193419756
#> 518   0.458956754 -1.055591e-01  0.5305659982
#> 519   0.748374324 -3.522033e-01  0.3217783071
#> 520   0.427519797 -4.353952e-01 -0.3054135962
#> 521   0.648195786 -2.438098e-01  0.3384888413
#> 522   0.633267786 -3.853727e-01 -0.0655063832
#> 523   0.477449672 -5.544117e-01 -0.2163620373
#> 524   0.305117803 -3.709511e-01  0.5640666159
#> 525   0.520923446 -6.641042e-01 -0.2209728459
#> 526   0.764106831 -2.203832e-01  0.2281586119
#> 527   0.342637624 -7.648332e-01  0.4662735383
#> 528   0.658441193 -4.251212e-01  0.9435505457
#> 529   0.672637770 -6.699005e-01 -0.3745436581
#> 530   0.455693716 -1.619269e-01  0.5881281316
#> 531   0.829554730 -4.263860e-01 -0.4644904517
#> 532   0.371674680 -3.981974e-01 -0.2244205445
#> 533   0.586101796 -5.614519e-01 -0.3179659884
#> 534   0.677606874 -4.651721e-01 -0.0953820547
#> 535   0.424132453 -4.898846e-01  0.1084687720
#> 536   0.728408571  2.048086e-01  1.2927125679
#> 537   0.467278253 -3.740001e-01  0.1687208574
#> 538   0.407717503 -1.039176e+00 -0.1575951323
#> 539   0.434359762 -6.425417e-01  0.3466628601
#> 540   0.551883801 -5.247510e-01  0.0886974812
#> 541   0.468620058 -7.350085e-01 -0.0770598555
#> 542   0.323757409 -5.819956e-02  0.5974908432
#> 543   0.839064863 -2.123456e-01  0.6110296285
#> 544   0.679412792 -2.939501e-01  0.7264710332
#> 545   0.749820914 -1.598008e-01  0.9102406234
#> 546   0.336680662 -7.067461e-01  0.1109217403
#> 547   0.272379360 -1.148700e+00 -0.4945985661
#> 548   0.588089944 -5.116597e-01 -0.0911808177
#> 549   0.323728068  3.572496e-02  0.6397039148
#> 550   0.548455152 -4.266172e-01  0.3917507046
#> 551   0.381687605 -8.385829e-01 -0.6070146985
#> 552   0.639880801 -1.163314e-01  0.3755483357
#> 553   0.632546756 -2.454814e-01  0.2798429447
#> 554   0.642310684 -3.424110e-01  0.2999074092
#> 555   0.604864836 -2.236177e-01  0.3762682936
#> 556   0.609965398 -3.495247e-01  0.3750695023
#> 557   0.659368427 -1.448032e-01  0.6094278631
#> 558   0.708245660 -2.832233e-01 -0.0076840747
#> 559   0.721152315 -2.562438e-01  0.2438471748
#> 560   0.788851525 -9.706636e-02  0.3239277882
#> 561   0.404529321 -4.779090e-01  0.2348114155
#> 562   0.600942354 -4.990908e-01  0.1511872520
#> 563   0.665355184 -1.851128e-01  0.0597237171
#> 564   0.720282742 -3.114853e-01 -0.1187426232
#> 565   0.755948668 -1.120205e-01  0.4576099606
#> 566   0.115809960 -9.019800e-03  0.4143993311
#> 567   0.478285844 -7.541659e-01  0.2360577348
#> 568   0.682501737 -2.860677e-01 -0.1441459646
#> 569   0.551476998 -4.953304e-01  0.7210137555
#> 570   0.516341759 -4.228832e-01  0.0784885761
#> 571   0.417067627 -4.860407e-01  0.3430306313
#> 572   0.428626732 -1.285676e-01  0.5570232057
#> 573   0.527151866 -6.386879e-01 -0.3390576286
#> 574   0.659368427 -1.448032e-01  0.6094278631
#> 575   0.734879183 -1.169869e-01  0.6214184805
#> 576   0.710828107 -3.787493e-01 -0.0995455894
#> 577   0.461092827  9.651932e-02  0.5743824628
#> 578   0.724902634 -5.403216e-02  0.9153594442
#> 579   0.681346380 -3.008207e-01  0.7245055588
#> 580  -0.031065557 -4.259113e-02  0.8716116923
#> 581   0.672637770 -6.699005e-01 -0.3745436581
#> 582   0.617607509 -2.329249e-01  0.2997968153
#> 583   0.594633305 -6.461874e-01 -0.2091563043
#> 584   0.381687605 -8.385829e-01 -0.6070146985
#> 585   0.455693716 -1.619269e-01  0.5881281316
#> 586  -0.073034139 -5.568083e-01  0.4402724536
#> 587   0.757234376 -2.914318e-01  0.2739550422
#> 588   0.856272269  3.579527e-02  0.6225082578
#> 589   0.634230572 -1.425292e-01  0.5569870258
#> 590   0.479740013 -4.504948e-01  0.2265237422
#> 591   0.633267786 -3.853727e-01 -0.0655063832
#> 592   0.441599633 -2.589813e-01  0.1790411566
#> 593   0.509567963 -4.681300e-01  0.2287899397
#> 594   0.577793291 -3.978982e-01  0.7188403698
#> 595   0.680056335 -3.772264e-01  0.3020396915
#> 596   0.553652253 -1.651489e-02  0.4456701293
#> 597   0.260375372 -9.953618e-01 -0.1271117296
#> 598   0.506271382 -3.580561e-02  0.1975482714
#> 599   0.576278609 -3.135657e-01  0.0788084571
#> 600   0.748374324 -3.522033e-01  0.3217783071
#> 601   0.659164276 -6.382710e-01 -0.3558369088
#> 602   0.756082446 -3.628444e-01  0.1734604610
#> 603   0.563714276 -2.655455e-01  0.3149603741
#> 604   0.677606874 -4.651721e-01 -0.0953820547
#> 605   0.385126758 -7.949904e-01  0.4126714035
#> 606   0.432881420 -7.878095e-01 -0.2708468484
#> 607   0.748374324 -3.522033e-01  0.3217783071
#> 608   0.678193887 -9.366769e-02  0.9191288732
#> 609   0.310354595 -2.639734e-01  0.3746363767
#> 610   0.124665566 -8.677485e-01 -0.3224822584
#> 611   0.009229749 -6.147766e-01 -0.2559717096
#> 612   0.372197199 -2.341822e-01  0.0434787855
#> 613   0.413870416  4.103282e-02  0.9220544878
#> 614   0.383455172 -1.430446e-01  0.6003957312
#> 615   0.445701382 -4.473227e-01  0.3573777441
#> 616   0.657127811 -2.995528e-01  0.1240659947
#> 617   0.510229147 -3.943776e-01  0.0578260620
#> 618   0.508578512 -5.870037e-01  0.4917318576
#> 619   0.583793464 -2.536374e-01  0.3991654172
#> 620   0.445274775 -2.120666e-01  0.9955616608
#> 621   0.371003900 -7.232611e-02  1.1210352715
#> 622   0.433628991 -1.125923e-01  0.4965494267
#> 623   0.440259324  4.655827e-02  0.4949944933
#> 624   0.428902490 -8.144624e-01 -0.2155868864
#> 625   0.221066546 -7.090722e-01  1.0665768366
#> 626   0.520923446 -6.641042e-01 -0.2209728459
#> 627   0.551883801 -5.247510e-01  0.0886974812
#> 628   0.272137385 -5.907316e-01  0.1895019834
#> 629   0.455028482 -4.201190e-01 -0.1886750843
#> 630   0.504253510 -6.189435e-01  0.4356198426
#> 631   0.633267786 -3.853727e-01 -0.0655063832
#> 632   0.385597158 -3.921926e-01  0.4250122817
#> 633   0.480123503 -5.719025e-01 -0.0186538211
#> 634   0.523381647 -2.605360e-01  0.2014923479
#> 635   0.621840910 -3.221962e-01  0.7670927626
#> 636   0.547541571 -6.363692e-01 -0.0130952056
#> 637   0.679412792 -2.939501e-01  0.7264710332
#> 638   0.421979277 -5.554172e-01  0.2834130906
#> 639   0.544277569 -5.992853e-01 -0.2664856370
#> 640   0.520923446 -6.641042e-01 -0.2209728459
#> 641   0.381533548 -1.899402e-01  0.0992468049
#> 642   0.507449139 -2.138341e-01  0.4677717192
#> 643   0.471183472 -6.246591e-01  0.4088695575
#> 644   0.455599488 -5.125149e-01 -0.2899885075
#> 645   0.750748763 -3.773923e-01  0.2374938517
#> 646   0.594030124 -1.105908e-01  0.1212560870
#> 647   0.608943479 -3.943185e-01  0.3471474790
#> 648   0.645765952 -2.416312e-01  0.4913788639
#> 649   0.689273872  1.949064e-01  0.4647907382
#> 650   0.334936300 -8.526092e-01 -0.0680244324
#> 651   0.364433365 -9.505407e-01 -0.2192605382
#> 652   0.604864836 -2.236177e-01  0.3762682936
#> 653   0.217832688 -4.945246e-01 -0.0388582829
#> 654   0.413795867 -9.450093e-01 -0.1039264155
#> 655   0.637614997 -6.553868e-02  0.4862096598
#> 656   0.639880801 -1.163314e-01  0.3755483357
#> 657   0.433031553 -3.272726e-01  0.5952782851
#> 658   0.606731731 -7.017408e-01 -0.2914873738
#> 659   0.642310684 -3.424110e-01  0.2999074092
#> 660   0.366814610 -8.115863e-01 -0.2384443412
#> 661   0.662322521 -2.618829e-01 -0.0103167908
#> 662   0.494282278 -1.303075e-01  0.0247952818
#> 663   0.836028487 -1.823020e-01  0.3787170554
#> 664   0.379180273 -8.924712e-01 -0.3803001775
#> 665   0.503160776 -3.561467e-01  0.2979357296
#> 666   0.701203734  7.909825e-02  0.5007969511
#> 667   0.782804389 -1.230561e-01  0.5593653197
#> 668   0.492237951 -2.317074e-02  1.3928863702
#> 669   0.634230572 -1.425292e-01  0.5569870258
#> 670   0.432881420 -7.878095e-01 -0.2708468484
#> 671   0.385126758 -7.949904e-01  0.4126714035
#> 672   0.287008326 -3.701084e-01  0.0758312416
#> 673   0.486685843 -2.376337e-01  0.0225631472
#> 674   0.177364107 -1.169776e+00 -0.1677328787
#> 675   0.840919963 -7.126163e-02 -0.0260861859
#> 676   0.173468992 -5.969495e-01 -0.0009886615
#> 677   0.445274775 -2.120666e-01  0.9955616608
#> 678   0.679412792 -2.939501e-01  0.7264710332
#> 679   0.549901524 -4.649401e-01  0.3258372420
#> 680   0.371924224 -7.047714e-01 -0.1305547081
#> 681   0.306723011 -9.111968e-01  0.0034371432
#> 682   0.404529321 -4.779090e-01  0.2348114155
#> 683   0.588258239 -3.805277e-01  0.3616792478
#> 684   0.606731731 -7.017408e-01 -0.2914873738
#> 685   0.590288655 -4.554547e-01  0.1860579026
#> 686   0.632883990 -2.991580e-01  0.1027099874
#> 687   0.454305805 -3.433615e-01  0.0974471488
#> 688   0.560411858 -6.942586e-01 -0.4466663608
#> 689   0.390909887 -4.178321e-01  0.2103889630
#> 690   0.167837007  4.718885e-02  0.5157669552
#> 691   0.644233056 -3.618588e-01  0.5398658697
#> 692   0.637049886  1.275622e-01  0.8465630398
#> 693   0.507270232 -4.604752e-01  0.4179064941
#> 694   0.332321460 -2.826474e-01  0.3912279464
#> 695   0.434359762 -6.425417e-01  0.3466628601
#> 696   0.471890127 -6.529326e-01 -0.3452922221
#> 697   0.422163865 -4.598432e-01  0.6300131339
#> 698   0.358071267 -3.271486e-01  0.1959454688
#> 699   0.547621899 -4.747561e-01  0.1164947153
#> 700   0.320864804 -8.127956e-01 -0.2713695769
#> 701   0.390909887 -4.178321e-01  0.2103889630
#> 702   0.758232729  1.684363e-01  1.1207797658
#> 703   0.858199326 -7.728986e-02  0.4986066416
#> 704   0.604864836 -2.236177e-01  0.3762682936
#> 705   0.212219626 -5.230406e-02  0.7532063200
#> 706   0.718662531 -3.300560e-01 -0.2106767528
#> 707   0.708068636 -2.450956e-01  0.2480497702
#> 708   0.186131671  8.417272e-02  0.4891446391
#> 709  -0.437420944 -3.062726e-01  0.5862140255
#> 710   0.750748763 -3.773923e-01  0.2374938517
#> 711   0.365966888 -8.114035e-01 -0.0934948098
#> 712   0.463340197 -7.320644e-01  0.0006770978
#> 713   0.764106831 -2.203832e-01  0.2281586119
#> 714   0.772898372 -3.003373e-01  0.3118079962
#> 715   0.551476998 -4.953304e-01  0.7210137555
#> 716   0.621840910 -3.221962e-01  0.7670927626
#> 717   0.421979277 -5.554172e-01  0.2834130906
#> 718   0.637117694 -5.010288e-01 -0.0163940628
#> 719   0.748374324 -3.522033e-01  0.3217783071
#> 720   0.458503923 -1.179023e-01  0.6413383580
#> 721   0.761453318 -2.456769e-01  0.1138727760
#> 722   0.159199158 -4.568460e-01 -0.2088568119
#> 723   0.595533513 -3.841880e-01  0.0624969249
#> 724   0.541423669 -3.018876e-01  0.2847904272
#> 725   0.483654407 -3.395139e-01  0.0795130669
#> 726   0.644233056 -3.618588e-01  0.5398658697
#> 727   0.755948668 -1.120205e-01  0.4576099606
#> 728   0.105350866 -4.589278e-02  0.1495869079
#> 729   0.615080088 -4.055646e-01  0.1289377623
#> 730   0.486685843 -2.376337e-01  0.0225631472
#> 731   0.523381647 -2.605360e-01  0.2014923479
#> 732   0.463340197 -7.320644e-01  0.0006770978
#> 733   0.379225046 -3.811862e-01  0.4252382340
#> 734   0.160685271 -1.063898e-01  0.5193690633
#> 735   0.548351065 -1.435985e-01  0.8825678446
#> 736   0.681702077 -2.596513e-01  0.2802040451
#> 737   0.558360435 -2.566795e-01  0.7095299285
#> 738   0.637049886  1.275622e-01  0.8465630398
#> 739   0.308277167 -4.700871e-01  0.0459198897
#> 740   0.366869427 -4.502838e-01  0.4123277931
#> 741   0.788851525 -9.706636e-02  0.3239277882
#> 742   0.507270232 -4.604752e-01  0.4179064941
#> 743   0.772574207 -5.158414e-01 -0.1384238015
#> 744   0.509567963 -4.681300e-01  0.2287899397
#> 745   0.648195786 -2.438098e-01  0.3384888413
#> 746   0.355262668 -1.212416e-01  0.2614309332
#> 747   0.742363625 -3.012939e-01 -0.0581729323
#> 748   0.675033211 -4.111685e-01  0.0161866498
#> 749   0.384520854 -4.460456e-01 -0.1034287978
#> 750   0.615080088 -4.055646e-01  0.1289377623
#> 751   0.507449139 -2.138341e-01  0.4677717192
#> 752   0.632546756 -2.454814e-01  0.2798429447
#> 753   0.569809646 -1.718119e-01  0.1569629253
#> 754   0.368259543 -8.662150e-01  0.1160803796
#> 755   0.635097538 -4.216498e-01 -0.2131359689
#> 756   0.594633305 -6.461874e-01 -0.2091563043
#> 757   0.674165869 -5.548351e-01 -0.3456277860
#> 758   0.277893140 -8.578706e-01 -0.0493468724
#> 759   0.368395143 -4.506515e-01  0.5491257308
#> 760   0.603819001 -1.859605e-01  0.2396973162
#> 761   0.690384308 -1.112801e-01  1.2935257885
#> 762   0.677745529 -2.315386e-01  0.9717013804
#> 763   0.644233056 -3.618588e-01  0.5398658697
#> 764   0.593112391 -4.451551e-01  0.1517529765
#> 765   0.348414147 -5.075430e-01 -0.0483172476
#> 766   0.544277569 -5.992853e-01 -0.2664856370
#> 767   0.675033211 -4.111685e-01  0.0161866498
#> 768   0.662207648 -5.834187e-01  0.1127272000
#> 769   0.306723011 -9.111968e-01  0.0034371432
#> 770   0.719984341 -5.152758e-01 -0.0681015701
#> 771   0.805210269 -5.120784e-02  0.7495099620
#> 772   0.524345425 -4.109664e-01  0.6193419756
#> 773   0.721152315 -2.562438e-01  0.2438471748
#> 774   0.602223101 -3.244704e-01  0.5681096091
#> 775   0.632546756 -2.454814e-01  0.2798429447
#> 776   0.468620058 -7.350085e-01 -0.0770598555
#> 777   0.241197378 -4.586828e-01 -0.3390580887
#> 778   0.469834069  1.937757e-01  0.4078856845
#> 779   0.868612348 -1.983837e-01  0.2047187786
#> 780   0.279588047 -8.941154e-01  0.0167813530
#> 781   0.678585471 -1.755864e-01  0.6162667653
#> 782   0.676007840 -3.426761e-01  0.7109712026
#> 783   0.075703109 -1.162157e+00  0.1835868636
#> 784   0.681702077 -2.596513e-01  0.2802040451
#> 785   0.734879183 -1.169869e-01  0.6214184805
#> 786   0.552043853 -5.541772e-01 -0.2038343609
#> 787   0.720724014 -4.602511e-01 -0.2404747045
#> 788   0.238271432 -5.177131e-01  0.5035802412
#> 789   0.770190983 -2.344266e-01  0.2321999164
#> 790   0.308063094 -2.239205e-01  0.0233801909
#> 791   0.677606874 -4.651721e-01 -0.0953820547
#> 792   0.734879183 -1.169869e-01  0.6214184805
#> 793   0.537105064 -3.281645e-01  0.4903216634
#> 794   0.585119599 -3.715712e-01  0.3239969829
#> 795   0.609965398 -3.495247e-01  0.3750695023
#> 796   0.644233056 -3.618588e-01  0.5398658697
#> 797   0.340521012 -4.852684e-01  0.1287410295
#> 798   0.636667081 -6.571933e-01  0.1538288318
#> 799   0.845481055 -1.049904e-01  0.6661489092
#> 800   0.609965398 -3.495247e-01  0.3750695023
#> 801   0.516341759 -4.228832e-01  0.0784885761
#> 802   0.407717503 -1.039176e+00 -0.1575951323
#> 803   0.303266300 -4.761944e-01  0.2998876419
#> 804   0.574984052 -1.019722e-01  0.3842329842
#> 805   0.416677826 -4.601709e-01  0.3375113008
#> 806   0.334936300 -8.526092e-01 -0.0680244324
#> 807   0.620255357 -7.908585e-02  0.3136820696
#> 808   0.724033676 -5.699124e-01 -0.0460595155
#> 809   0.595533513 -3.841880e-01  0.0624969249
#> 810   0.681702077 -2.596513e-01  0.2802040451
#> 811   0.455693716 -1.619269e-01  0.5881281316
#> 812   0.149200840 -3.158523e-01  0.4973511627
#> 813   0.600942354 -4.990908e-01  0.1511872520
#> 814   0.508578512 -5.870037e-01  0.4917318576
#> 815   0.587282569 -3.164016e-01 -0.0172514472
#> 816   0.659368427 -1.448032e-01  0.6094278631
#> 817   0.547541571 -6.363692e-01 -0.0130952056
#> 818   0.574984052 -1.019722e-01  0.3842329842
#> 819   0.676007840 -3.426761e-01  0.7109712026
#> 820   0.676007840 -3.426761e-01  0.7109712026
#> 821   0.065786601  1.233652e-01  1.2648079000
#> 822   0.313176455 -2.505333e-01  0.7818269062
#> 823   0.646343685 -3.578694e-01  0.6202815306
#> 824   0.471890127 -6.529326e-01 -0.3452922221
#> 825   0.472731876  2.121097e-01  1.2719785113
#> 826   0.805647796 -1.867938e-01  0.4291724022
#> 827   0.570709889 -2.618270e-02  0.5317711436
#> 828   0.628123482  2.439737e-02  0.8703916019
#> 829   0.382021243 -4.987490e-01  0.3845672966
#> 830   0.512934913 -7.625702e-01  0.1566112353
#> 831   0.440727101 -5.379753e-01 -0.4055950491
#> 832   0.721152315 -2.562438e-01  0.2438471748
#> 833   0.161722754 -1.993058e-01  0.4075143511
#> 834   0.432881420 -7.878095e-01 -0.2708468484
#> 835   0.721294410 -5.291841e-01  0.0711421337
#> 836   0.724902634 -5.403216e-02  0.9153594442
#> 837   0.529567049  2.633879e-01  0.7031581149
#> 838   0.433031553 -3.272726e-01  0.5952782851
#> 839   0.202614659 -1.099996e-01  0.3894199915
#> 840   0.092555062 -6.884904e-01 -0.3622030615
#> 841   0.667503577 -1.775822e-02  1.0245307576
#> 842   0.403897432 -8.167510e-01  0.3771163220
#> 843   0.547541571 -6.363692e-01 -0.0130952056
#> 844   0.465569640 -3.288849e-01  0.8931608019
#> 845   0.238271432 -5.177131e-01  0.5035802412
#> 846   0.416677826 -4.601709e-01  0.3375113008
#> 847   0.464425379 -5.006514e-01  0.5876401999
#> 848   0.628745926 -5.809379e-01  0.5309801903
#> 849   0.623589670 -5.066256e-01  0.1322400569
#> 850   0.432881420 -7.878095e-01 -0.2708468484
#> 851   0.293897379 -7.264297e-01  0.1809269100
#> 852   0.636667081 -6.571933e-01  0.1538288318
#> 853   0.674165869 -5.548351e-01 -0.3456277860
#> 854   0.757234376 -2.914318e-01  0.2739550422
#> 855   0.560411858 -6.942586e-01 -0.4466663608
#> 856   0.160049263  6.409222e-02  0.3057469082
#> 857   0.255796218 -5.819872e-01 -0.1967224897
#> 858   0.868612348 -1.983837e-01  0.2047187786
#> 859   0.176621791 -9.171630e-01 -0.3685446581
#> 860   0.335961077 -4.927080e-01  0.0221162236
#> 861   0.514473814 -7.800970e-01  0.3476938321
#> 862   0.508327682  4.723193e-02  0.3052876218
#> 863   0.710828107 -3.787493e-01 -0.0995455894
#> 864   0.572406407  1.195062e-02  0.2668315054
#> 865   0.471890127 -6.529326e-01 -0.3452922221
#> 866   0.720724014 -4.602511e-01 -0.2404747045
#> 867   0.141021508 -4.046460e-01  0.0387565779
#> 868   0.360982691 -2.557283e-01 -0.1233753052
#> 869   0.111162095  2.223812e-01  0.7857841537
#> 870   0.370775911 -3.670309e-01  0.0457083183
#> 871   0.538633064 -4.687045e-01  0.6827850446
#> 872   0.584137050 -8.477569e-01 -0.6664834383
#> 873   0.066931692 -5.623937e-03  0.3986686120
#> 874   0.465747019 -8.088830e-01 -0.6042918030
#> 875   0.492976212 -6.573783e-01  0.2104361620
#> 876   0.106074617 -6.024026e-01  0.3874172246
#> 877   0.442517546 -2.687531e-01  0.2161358272
#> 878   0.679828484 -3.889512e-01  0.3147578797
#> 879   0.678193887 -9.366769e-02  0.9191288732
#> 880   0.289920197 -6.625904e-01  0.1883354207
#> 881   0.169792928 -8.644922e-01 -0.8549706061
#> 882   0.659164276 -6.382710e-01 -0.3558369088
#> 883   0.588089944 -5.116597e-01 -0.0911808177
#> 884   0.624231043 -3.176884e-01 -0.0375486128
#> 885   0.381798127 -6.491273e-01  0.0019473093
#> 886   0.720724014 -4.602511e-01 -0.2404747045
#> 887   0.632546756 -2.454814e-01  0.2798429447
#> 888   0.740591346 -5.512911e-02  0.7390198780
#> 889   0.595530281 -1.593324e-02  0.0715208297
#> 890   0.368259543 -8.662150e-01  0.1160803796
#> 891   0.464425379 -5.006514e-01  0.5876401999
#> 892   0.657494838 -3.109330e-01 -0.1058116377
#> 893   0.785648476  1.593701e-01  0.8484666275
#> 894   0.604864836 -2.236177e-01  0.3762682936
#> 895   0.633267786 -3.853727e-01 -0.0655063832
#> 896   0.299348081 -6.185502e-01  0.0095232845
#> 897   0.782804389 -1.230561e-01  0.5593653197
#> 898   0.720724014 -4.602511e-01 -0.2404747045
#> 899   0.770190983 -2.344266e-01  0.2321999164
#> 900   0.615939454 -5.379870e-01 -0.2196287781
#> 901   0.385126758 -7.949904e-01  0.4126714035
#> 902   0.280829809 -7.674151e-01 -0.4938880164
#> 903   0.551883801 -5.247510e-01  0.0886974812
#> 904   0.608943479 -3.943185e-01  0.3471474790
#> 905   0.672421989 -6.054193e-01 -0.3552440114
#> 906   0.644233056 -3.618588e-01  0.5398658697
#> 907   0.588089944 -5.116597e-01 -0.0911808177
#> 908   0.503160776 -3.561467e-01  0.2979357296
#> 909   0.549901524 -4.649401e-01  0.3258372420
#> 910   0.578028570 -4.130846e-01  0.2883536217
#> 911   0.357032848  9.053337e-02  0.9030820365
#> 912   0.667503577 -1.775822e-02  1.0245307576
#> 913   0.495378749 -5.157677e-01  0.6833739336
#> 914   0.763464014  1.302456e-01  0.8186591232
#> 915   0.492237951 -2.317074e-02  1.3928863702
#> 916   0.594633305 -6.461874e-01 -0.2091563043
#> 917   0.679828484 -3.889512e-01  0.3147578797
#> 918   0.428902490 -8.144624e-01 -0.2155868864
#> 919   0.623226871 -2.544954e-01  0.4074802169
#> 920   0.516341759 -4.228832e-01  0.0784885761
#> 921   0.835596377 -1.748454e-01  0.5968263163
#> 922   0.719984341 -5.152758e-01 -0.0681015701
#> 923   0.454305805 -3.433615e-01  0.0974471488
#> 924   0.856272269  3.579527e-02  0.6225082578
#> 925   0.384109592 -3.123773e-01  0.6640479796
#> 926   0.186153624 -1.078641e+00 -0.4254307197
#> 927  -0.039744012 -7.830703e-03  0.4847378732
#> 928   0.416231162 -3.676731e-01 -0.1215392063
#> 929   0.632677784 -2.790618e-01  0.3937965735
#> 930   0.637117694 -5.010288e-01 -0.0163940628
#> 931   0.632677784 -2.790618e-01  0.3937965735
#> 932   0.576278609 -3.135657e-01  0.0788084571
#> 933   0.720452198 -5.952054e-01  0.2422817817
#> 934   0.768254754 -8.578182e-02  0.8556270930
#> 935   0.438494314 -1.045713e-02  0.7201309800
#> 936   0.705716304 -6.381503e-01 -0.1743351952
#> 937   0.364433365 -9.505407e-01 -0.2192605382
#> 938   0.554316350 -1.109275e-01  0.2278148571
#> 939   0.637117694 -5.010288e-01 -0.0163940628
#> 940   0.637614997 -6.553868e-02  0.4862096598
#> 941   0.628123482  2.439737e-02  0.8703916019
#> 942   0.070848414 -4.135877e-01  0.7816182275
#> 943   0.756082446 -3.628444e-01  0.1734604610
#> 944   0.678585471 -1.755864e-01  0.6162667653
#> 945   0.724902634 -5.403216e-02  0.9153594442
#> 946   0.178609809 -2.293844e-01  0.3766444252
#> 947   0.541519599 -3.878320e-01  0.5431720223
#> 948   0.268080651 -6.871386e-01  0.2224515788
#> 949   0.429362512 -6.796696e-01  0.0568761747
#> 950   0.740591346 -5.512911e-02  0.7390198780
#> 951   0.845481055 -1.049904e-01  0.6661489092
#> 952   0.777212700 -2.935970e-01  0.1834490101
#> 953   0.299348081 -6.185502e-01  0.0095232845
#> 954   0.734879183 -1.169869e-01  0.6214184805
#> 955   0.407697870 -7.995766e-01  0.0091022467
#> 956   0.508578512 -5.870037e-01  0.4917318576
#> 957   0.492525661 -7.777571e-01 -0.4494138113
#> 958   0.659368427 -1.448032e-01  0.6094278631
#> 959   0.510623308 -3.073274e-01  0.1175567215
#> 960   0.644562275 -1.412364e-01  0.4165147513
#> 961   0.805647796 -1.867938e-01  0.4291724022
#> 962  -0.042901311 -3.067684e-03  0.4662950433
#> 963   0.448363199 -5.899503e-01  0.0237764062
#> 964   0.220317672 -5.043838e-01  0.0030202474
#> 965   0.186153624 -1.078641e+00 -0.4254307197
#> 966   0.492976212 -6.573783e-01  0.2104361620
#> 967   0.611776812 -1.960184e-01  0.1115391659
#> 968   0.320864804 -8.127956e-01 -0.2713695769
#> 969   0.644114141 -6.862536e-01 -0.1444875728
#> 970   0.468620058 -7.350085e-01 -0.0770598555
#> 971   0.734879183 -1.169869e-01  0.6214184805
#> 972   0.537105064 -3.281645e-01  0.4903216634
#> 973   0.348781558 -8.064140e-01 -0.1935796641
#> 974   0.307447221 -4.792562e-01  0.1201092383
#> 975   0.425620368 -2.462319e-01  0.4426839720
#> 976   0.772898372 -3.003373e-01  0.3118079962
#> 977   0.623226871 -2.544954e-01  0.4074802169
#> 978   0.424432722 -7.977020e-01 -0.7111186928
#> 979   0.029528545 -8.377585e-01  0.5421732697
#> 980   0.644233056 -3.618588e-01  0.5398658697
#> 981   0.644233056 -3.618588e-01  0.5398658697
#> 982   0.679145483 -4.287712e-02  0.4200332572
#> 983   0.713228998 -2.525679e-01  0.2003398689
#> 984   0.348414147 -5.075430e-01 -0.0483172476
#> 985   0.719984341 -5.152758e-01 -0.0681015701
#> 986   0.022782949 -2.256559e-01  0.0602405162
#> 987   0.756082446 -3.628444e-01  0.1734604610
#> 988   0.548455152 -4.266172e-01  0.3917507046
#> 989   0.526453254 -7.381857e-01 -0.2459812392
#> 990   0.690384308 -1.112801e-01  1.2935257885
#> 991   0.858199326 -7.728986e-02  0.4986066416
#> 992   0.468620058 -7.350085e-01 -0.0770598555
#> 993   0.594633305 -6.461874e-01 -0.2091563043
#> 994   0.472731876  2.121097e-01  1.2719785113
#> 995   0.399855279 -7.168713e-01  0.1798173945
#> 996   0.607645003  1.141888e-01  0.8889284168
#> 997   0.029538791 -1.479103e+00  0.1773679432
#> 998   0.215103328 -9.057674e-01 -0.5976333475
#> 999   0.653961225 -2.438471e-01 -0.1863769286
#> 1000  0.243200548 -3.353638e-01  0.0701527346
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 1,000 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>  <dbl>  <dbl>
#>  1      1        4       70        1  0.593 -0.456  0.311
#>  2      2        4       70        1  0.783 -0.288  0.165
#>  3      3        4       70        1  0.526 -0.738 -0.246
#>  4      4        4       70        1  0.578 -0.398  0.719
#>  5      5        4       70        1  0.449 -0.672 -0.345
#>  6      6        4       70        1  0.532 -0.279  0.505
#>  7      7        4       70        1  0.552 -0.554 -0.204
#>  8      8        4       70        1  0.429 -0.129  0.557
#>  9      9        4       70        1  0.757 -0.291  0.274
#> 10     10        4       70        1  0.585 -0.372  0.324
#> # … with 990 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 2,000 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20    Cc       depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>       <dbl> <dbl>
#> 1      1     1     4    70     1  7.24  44.4 1.36  0.163 26.6      7.33e+2 1178.
#> 2      1    14     4    70     1  7.24  44.4 1.36  0.163  5.43     1.44e-5  241.
#> 3      2     1     4    70     1  8.75  52.5 1.18  0.167 20.6      8.38e+2 1080.
#> 4      2    14     4    70     1  8.75  52.5 1.18  0.167  4.48     1.83e-4  235.
#> 5      3     1     4    70     1  6.77  33.5 0.782 0.202 24.2      1.12e+3  809.
#> 6      3    14     4    70     1  6.77  33.5 0.782 0.202  4.99     4.31e-2  167.
#> # … with 1,994 more rows, and 1 more variable: AUC <dbl>
#>