Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>            ETA_Cl        ETA_Vc       ETA_Ka
#> 1    0.2986017454  0.0465202736  0.327874377
#> 2    0.2987749816 -0.3212886187 -0.437645678
#> 3    0.1503503071 -0.3376155887 -0.282794157
#> 4    0.0316298782 -0.0049328168 -0.050625059
#> 5   -0.4908979864  0.4188000560  0.615199092
#> 6    0.7034450065 -0.0328024614 -0.434168654
#> 7   -0.1951971266 -0.5901072198 -0.357886558
#> 8    0.5800184131 -0.3545261707 -0.431446825
#> 9    0.2243636646  0.8699375061 -0.347790473
#> 10  -0.0285443221 -0.6040010270  0.446117418
#> 11  -0.3303024701 -0.1688718107  0.261221825
#> 12   0.2939906679 -0.2242210917 -0.559739062
#> 13  -0.2071401556 -0.3428251753 -0.525140733
#> 14  -0.3854152369  0.5872515922  0.602547225
#> 15  -0.2756454993  0.4333765093 -1.121915971
#> 16   0.4921646028  0.8560768139 -0.225408683
#> 17   0.0645096578 -0.7647559127 -0.306251133
#> 18   0.6557369358  0.0547782712 -0.827495102
#> 19  -0.2766689183 -0.4839344277  0.241602552
#> 20  -0.1177224569 -0.4149067944  0.001155736
#> 21   0.1907377900  0.2683790232 -0.584089573
#> 22   0.0867197590 -0.4737321237  0.750652491
#> 23  -0.0901664851 -0.1161741593  0.564380442
#> 24  -0.1169231447  0.9641269906  0.584693717
#> 25   0.0873019470  0.1370781049  0.024111806
#> 26  -0.1973509698  0.5730034166  1.297711536
#> 27  -0.0009908155 -0.0748391193 -0.759963370
#> 28   0.5257506465 -0.4230996185  0.511810732
#> 29   0.1593810847 -0.2376556910  0.843093011
#> 30   0.2631123963  0.1010828809  0.261538856
#> 31   0.7561422500  0.0540743007 -0.346521019
#> 32  -0.2340963939 -0.1967394014  0.199441781
#> 33  -0.0832971714 -0.0627815252 -0.832341496
#> 34   0.6008096505 -0.1162523429  0.188957545
#> 35   0.0257116178 -0.8110934915  0.212632312
#> 36  -0.8665505143  0.9061831256 -0.220348867
#> 37   0.3966838189  0.9782472298 -0.307675053
#> 38   0.1030034131  0.3394703094 -0.164119189
#> 39   0.0548731576  0.0004774005 -0.402637777
#> 40  -0.2333694520 -0.9153840381  0.319485870
#> 41  -0.4402004701  0.3807344652 -0.222505713
#> 42   0.0907886859  0.1212598672 -0.336889666
#> 43  -0.1848234455 -0.0884780864 -0.063940128
#> 44  -0.0146229712 -0.6872268498  0.410290074
#> 45   0.2065508546  0.5164569587  0.252930650
#> 46   0.6656033643 -0.1122304869  0.710459114
#> 47  -0.2225025946 -0.4311293779  0.253903870
#> 48  -0.6824255480  0.6107449032  0.180864507
#> 49  -0.1475716074  0.1224192915  0.483145175
#> 50   0.1844812129  0.7932790767  0.246690772
#> 51   0.0511713286  0.3126618764  0.118115645
#> 52   0.5300039324  0.3320304768  1.289081370
#> 53  -0.6878677560  0.6210122843  0.053687697
#> 54   0.3162788067 -0.3007492055 -0.371708273
#> 55   0.2517651396 -0.0904108140 -0.071468380
#> 56   0.4767778576 -0.4696527510  0.218998234
#> 57  -0.8836641306 -0.9537433494 -0.618826566
#> 58   0.5984282964 -0.0127164876 -0.238774262
#> 59   0.0002174395  0.3839963574 -0.770526136
#> 60  -1.0396747443  0.4029726573 -0.530463613
#> 61   0.5328024588  0.4346907397  0.550425839
#> 62   0.2629877005  0.6929571772 -0.221519280
#> 63   0.1706838020 -0.6491106741  0.088779881
#> 64   0.1913639684 -0.6242726679  0.345027257
#> 65   0.1671478776 -0.2285932561 -0.259049931
#> 66   0.5786578186  0.0594820441 -0.124387628
#> 67  -0.8579898529 -0.0153709397  0.067328263
#> 68  -0.2150682787 -0.7846071851 -0.457031219
#> 69   0.0344685453 -0.0279970496 -0.947917845
#> 70   0.0408205448 -0.0175302126 -1.087422199
#> 71  -0.3136954993  0.9076050025 -0.049498425
#> 72  -0.2322415158  0.6995824657 -0.720973245
#> 73   0.2910222004  0.5899612592 -0.136804212
#> 74   0.7505189424 -0.2488973991  0.160000398
#> 75  -0.1296107278 -0.5816412204  0.547580837
#> 76  -0.1542124113 -0.0952146489  0.085797803
#> 77  -0.0051585114  0.4123420424  0.328145857
#> 78   0.5337829223  0.3784329960  0.262919975
#> 79   0.7748296310  0.3343697948 -0.259914459
#> 80  -0.6127526183  0.6735150630 -0.321261078
#> 81  -0.3902157623  0.2978465963  0.902351942
#> 82  -0.0102510210  0.5289049914 -0.278835524
#> 83  -0.4418616357  0.5269728440 -0.829806615
#> 84  -0.0753875214 -0.1125283374 -0.068061255
#> 85  -0.3304563931  0.7906856309  0.295189948
#> 86  -0.3078579949  0.6144848397  0.041121722
#> 87   0.2038495418 -0.7261215345 -0.436580987
#> 88   0.0380096670  0.8587950991 -0.059562065
#> 89   0.1907218239  0.0357229807  0.135710308
#> 90  -0.8299869323 -0.4794316938 -0.271235073
#> 91  -0.1997241249 -0.1077125504 -0.117581178
#> 92  -0.1752132666  0.2622991788  0.680262137
#> 93   0.2787622087  0.1230711639  0.581761670
#> 94  -0.0150283547  0.1067066813  0.395978727
#> 95  -0.2041837888 -0.6283204205 -0.232251292
#> 96  -0.0455233542 -0.3901278146 -0.765912513
#> 97   0.0334451723  0.1184517837 -0.363598325
#> 98   0.8553625843  0.1711106536  0.976990148
#> 99  -0.5685511474 -0.2188477155  0.582290366
#> 100  0.3169523560  0.6854001053  0.056760473
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl   ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>
#>  1      1        4       70        1  0.299   0.0465   0.328 
#>  2      2        4       70        1  0.299  -0.321   -0.438 
#>  3      3        4       70        1  0.150  -0.338   -0.283 
#>  4      4        4       70        1  0.0316 -0.00493 -0.0506
#>  5      5        4       70        1 -0.491   0.419    0.615 
#>  6      6        4       70        1  0.703  -0.0328  -0.434 
#>  7      7        4       70        1 -0.195  -0.590   -0.358 
#>  8      8        4       70        1  0.580  -0.355   -0.431 
#>  9      9        4       70        1  0.224   0.870   -0.348 
#> 10     10        4       70        1 -0.0285 -0.604    0.446 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  5.39  73.3 1.39  0.0735  16.9    7.20e+2 1238.
#> 2      1    14     4    70     1  5.39  73.3 1.39  0.0735  10.5    1.05e-5  768.
#> 3      2     1     4    70     1  5.39  50.8 0.646 0.106   14.4    1.24e+3  729.
#> 4      2    14     4    70     1  5.39  50.8 0.646 0.106   10.9    2.81e-1  555.
#> 5      3     1     4    70     1  4.65  49.9 0.754 0.0931  16.5    1.14e+3  824.
#> 6      3    14     4    70     1  4.65  49.9 0.754 0.0931  12.7    6.35e-2  634.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>