Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

Usage

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, $model an rxode2 model using the estimated ETAs, $event the data.table used to solve the returned rxode2 model.

Examples

# model
mod_run001 <- function() {
  ini({
    THETA_Cl <- 4.0
    THETA_Vc <- 70.0
    THETA_Ka <- 1.0
    ETA_Cl ~ 0.2
    ETA_Vc ~ 0.2
    ETA_Ka ~ 0.2
    prop.sd <- sqrt(0.05)
  })
  model({
    TVCl <- THETA_Cl
    TVVc <- THETA_Vc
    TVKa <- THETA_Ka

    Cl <- TVCl*exp(ETA_Cl)
    Vc <- TVVc*exp(ETA_Vc)
    Ka <- TVKa*exp(ETA_Ka)

    K20 <- Cl/Vc
    Cc <- centr/Vc

    d/dt(depot) = -Ka*depot
    d/dt(centr) = Ka*depot - K20*centr
    Cc ~ prop(prop.sd)
  })
}
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#>  
#>  
#>  
#>  
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1   -0.626118495  0.114181258 -1.089977423
#> 2   -0.002491555  0.277966827  0.513585283
#> 3   -0.814741627 -0.110607238 -0.109209384
#> 4   -0.126429720 -0.247621892  0.281289321
#> 5    0.923507208 -0.729400635  0.229164299
#> 6   -0.833164068 -0.233451094 -0.023524289
#> 7    0.242835347 -0.408786690  0.209365022
#> 8    0.162316736 -0.583409609  0.329943601
#> 9    0.844565080 -0.043578776 -0.418523660
#> 10  -0.007133196 -0.369751261 -0.676365686
#> 11   0.418307135  0.078928106  0.108979453
#> 12   0.726073134  0.050104954 -0.059925286
#> 13  -0.854217084 -0.124878691 -0.140177303
#> 14   0.477314594  0.031320537 -0.285824640
#> 15  -0.022344982 -0.112466815  0.198919317
#> 16   1.232260201  0.020809466  0.258359350
#> 17   0.052858355 -0.854947394  0.385536795
#> 18  -0.108778777 -0.092164995  0.008576480
#> 19   0.013219971  0.245890352 -1.017015082
#> 20   1.199676043 -0.161543056  0.095415592
#> 21   0.480462085 -0.297436507  0.498174667
#> 22  -0.109968218 -0.526622321 -0.436413663
#> 23   0.476308114  0.058884898  0.218521447
#> 24  -0.760017399 -0.657733272  0.127075897
#> 25   0.598067870  0.105853796  0.589558724
#> 26   0.234299580  0.271345975 -0.049164727
#> 27   0.077002004 -0.040395591  0.860592505
#> 28   0.580658894  0.334869635  0.248751082
#> 29  -0.245188102  0.496646302 -1.168271430
#> 30  -0.069628373  0.194041413 -0.170813730
#> 31   0.189702450  0.475433666  0.468998541
#> 32  -0.017040133  0.217412406  0.748135847
#> 33  -0.158475130  0.423219641  0.588902649
#> 34  -0.132661452 -0.173167175 -0.351256162
#> 35  -0.472587094 -0.355776943 -0.785430249
#> 36  -0.308817936 -0.249787573 -0.240003136
#> 37   0.101574342  0.437578343 -0.093415162
#> 38  -0.625835383  0.115621390 -0.197578722
#> 39   0.254285588  0.951156441  0.190002471
#> 40  -0.753233600  0.111535869  0.479787852
#> 41   0.912033660  0.201001840  0.622438164
#> 42   0.190766359  0.048113024  0.009970508
#> 43   0.269943050 -0.117460907 -0.236246880
#> 44   0.085931834 -0.512596075  0.378425286
#> 45   0.036546129 -0.583666071 -0.422577520
#> 46   0.203187738 -0.382458185 -0.128303442
#> 47   0.400239006  0.030099460 -0.072751070
#> 48  -0.369984355  0.839198826  0.342762477
#> 49   0.438249957  0.591118430 -0.500749906
#> 50   0.230136035 -0.674889964  0.685462828
#> 51   0.191920539  0.054606307 -0.508934617
#> 52  -0.249551952  0.470709543  0.303069339
#> 53   0.017217521 -0.159378512  0.350098525
#> 54   0.359743811 -0.849733032  0.418495455
#> 55  -0.138212034  0.117646994 -0.800777022
#> 56  -0.352520072 -0.506702694  0.162630373
#> 57  -0.127852962  0.231508675 -0.046022156
#> 58  -0.435617165  0.568261928  0.429711796
#> 59   0.343782648  0.463282325 -0.211928742
#> 60  -0.570347095 -0.136677720  0.989133385
#> 61  -0.465848262 -0.512741053 -0.749229147
#> 62   0.682420512  0.247839297  0.891346008
#> 63  -0.068924890  1.146838273  0.474940456
#> 64   0.511028685  0.502596010 -0.177544465
#> 65  -0.368173579 -0.258885074  0.788790589
#> 66   0.059475896  0.168375618  0.509245544
#> 67   0.555109723  0.273735392 -0.192024613
#> 68   0.608416801 -0.031688406 -0.121710828
#> 69  -1.094188573  0.029286516 -0.491268116
#> 70  -0.283165888 -0.922894327  1.184638417
#> 71  -0.515815439 -0.152337889  0.351672035
#> 72  -0.568190736  0.242453072  0.033588380
#> 73   0.249775243  0.185775389 -0.649488201
#> 74   0.420920174 -0.151576730 -0.033797831
#> 75   0.017979951  0.055589127 -0.446512611
#> 76   0.551588806  0.152242459 -0.211398977
#> 77   0.316964005 -0.683771124  0.106179842
#> 78  -0.587108379  0.334081340 -0.698779487
#> 79   0.031776028 -0.286008644 -0.377983025
#> 80   0.301978609  0.515805336 -0.754227850
#> 81  -0.403751120  0.589263704  0.492019811
#> 82   0.538341344 -0.640083750  0.618456538
#> 83   0.001397963 -0.034832047  0.197412704
#> 84   0.057656072 -0.371283104 -0.225213596
#> 85  -0.533812565 -0.336180890  0.651072068
#> 86  -0.370562765  0.129591078 -0.214686445
#> 87  -0.270487910  0.652981124  0.066938642
#> 88  -0.641000683 -0.004607784 -0.094914840
#> 89  -0.405327650 -0.940111168  0.846736541
#> 90  -0.432959036 -0.045885473  0.107313183
#> 91   0.027234813 -0.973841605 -0.052708658
#> 92   0.050219756  0.003526815  0.839752588
#> 93   0.965425280  0.317393983  0.343005395
#> 94  -0.137836338  0.452580985 -0.411012369
#> 95   0.251951230  0.144218670  0.163981760
#> 96   0.505277641 -0.421050740  0.097419956
#> 97   0.632991621 -0.171610636 -0.077853793
#> 98  -0.099167163 -0.451474970  0.214986875
#> 99   0.717512770 -0.677539567 -0.633265146
#> 100  0.392106741  0.279120500  0.944639121
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 8
#>    sim.id THETA_Cl THETA_Vc THETA_Ka prop.sd   ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1   0.224 -0.626    0.114  -1.09  
#>  2      2        4       70        1   0.224 -0.00249  0.278   0.514 
#>  3      3        4       70        1   0.224 -0.815   -0.111  -0.109 
#>  4      4        4       70        1   0.224 -0.126   -0.248   0.281 
#>  5      5        4       70        1   0.224  0.924   -0.729   0.229 
#>  6      6        4       70        1   0.224 -0.833   -0.233  -0.0235
#>  7      7        4       70        1   0.224  0.243   -0.409   0.209 
#>  8      8        4       70        1   0.224  0.162   -0.583   0.330 
#>  9      9        4       70        1   0.224  0.845   -0.0436 -0.419 
#> 10     10        4       70        1   0.224 -0.00713 -0.370  -0.676 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 15,100 × 14
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20   rxCc     Cc depot
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>  <dbl> <dbl>
#> 1      1   0       4    70     1  2.14  78.5 0.336 0.0273 0      0         0 
#> 2      1   0.1     4    70     1  2.14  78.5 0.336 0.0273 0.0847 0.0847  393.
#> 3      1   0.2     4    70     1  2.14  78.5 0.336 0.0273 0.335  0.335   774.
#> 4      1   0.3     4    70     1  2.14  78.5 0.336 0.0273 0.744  0.744  1141.
#> 5      1   0.4     4    70     1  2.14  78.5 0.336 0.0273 1.31   1.31   1497.
#> 6      1   0.5     4    70     1  2.14  78.5 0.336 0.0273 2.02   2.02   1841.
#> # ℹ 15,094 more rows
#> # ℹ 2 more variables: centr <dbl>, AUC <dbl>
#> 
#> $event
#> ── EventTable with 152 records ──
#> 1 dosing records (see $get.dosing(); add with add.dosing or et)
#> 151 observation times (see $get.sampling(); add with add.sampling or et)
#> ── First part of : ──
#> # A tibble: 152 × 5
#>       id  time   amt evid          dur       
#>    <int> <dbl> <dbl> <evid>        <rate/dur>
#>  1     1   0      NA 0:Observation NA        
#>  2     1   0    2000 1:Dose (Add)   0.5      
#>  3     1   0.1    NA 0:Observation NA        
#>  4     1   0.2    NA 0:Observation NA        
#>  5     1   0.3    NA 0:Observation NA        
#>  6     1   0.4    NA 0:Observation NA        
#>  7     1   0.5    NA 0:Observation NA        
#>  8     1   0.6    NA 0:Observation NA        
#>  9     1   0.7    NA 0:Observation NA        
#> 10     1   0.8    NA 0:Observation NA        
#> # ℹ 142 more rows
#>