Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object

A posologyr list, created by the posologyr function.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01,n_simul=100)
#> $eta
#>           ETA_Cl        ETA_Vc      ETA_Ka
#> 1    0.063591840 -0.2865121928  0.32851026
#> 2    0.164124046 -1.5151566289  0.15134175
#> 3   -0.054746768 -0.6140230779  0.53048082
#> 4   -0.832665186 -0.8814548395 -0.12756659
#> 5   -0.597871890 -0.0060326881 -0.50115075
#> 6    0.214349540  0.2057602360 -0.08641598
#> 7    0.428518392 -0.1595456993  0.58776671
#> 8   -0.024554835  0.2843689288  0.66022205
#> 9   -0.028751421 -0.2563085144 -0.54415564
#> 10   0.269156377  0.2297256552  0.19979487
#> 11  -0.289095038  0.0401761571  0.63776406
#> 12   0.999042533  0.5574288237 -0.76658532
#> 13  -0.376546012  0.3182271307  0.40033280
#> 14  -0.456351526  0.3104542895 -0.64700156
#> 15   0.276160996 -0.5222566882 -0.17409281
#> 16   0.059207830 -0.1317558007 -0.03730049
#> 17  -0.058717867 -0.5083605802 -0.05716461
#> 18   0.414982064  0.0892174350 -0.45182490
#> 19   0.759642600 -0.0039730164  0.40118049
#> 20   0.028589229 -0.3796791495 -0.54087251
#> 21   0.717712493 -0.0191094801  0.50466710
#> 22  -0.268451303  0.3129309648 -0.03300686
#> 23  -0.198413391  0.1764675183  1.15246327
#> 24  -0.228047886  0.6323820482 -0.23162126
#> 25  -0.287959742 -0.9841542706  0.29654048
#> 26  -0.048157083  0.3454201256 -0.79466199
#> 27  -0.553613070 -0.2814895079 -0.12198088
#> 28   0.412071954 -0.4189865796 -0.19369133
#> 29  -0.146793705 -0.4256697119  0.06243449
#> 30  -0.321482874  0.4485877747  0.60812205
#> 31   0.435446091  0.7116739273 -0.06690919
#> 32  -0.094713934  0.0704024274  0.68642811
#> 33   0.509175653 -0.0644279028  0.27708218
#> 34   0.184543726  0.1211242721  0.16759358
#> 35   0.313665794 -0.0590024004 -0.24108562
#> 36   0.599896312  0.0192712256  0.11607434
#> 37  -0.240057353 -0.4434044560 -0.18147127
#> 38   0.381728399  0.4121336263  0.17822758
#> 39  -0.978209945 -0.2499885553 -0.20416011
#> 40  -0.225049610  0.6050524452  0.15769040
#> 41   0.356258463 -0.0412838705  0.75719289
#> 42   0.108403526  0.4733192165  0.01284945
#> 43   0.643366812 -0.5443162327 -0.38593506
#> 44   0.245506234  0.5265789404 -0.20400480
#> 45   0.492704581  1.3602966458 -0.35474463
#> 46  -0.641989685  0.1465238907  0.35976958
#> 47  -0.292215262  0.0123570002  0.07860249
#> 48   0.223343122  0.1495284964 -0.04525669
#> 49   0.326197270 -0.4045583936  0.28906501
#> 50  -0.033853149 -1.0599352269 -0.68313699
#> 51  -0.398306063 -0.2046069007 -0.11459251
#> 52   0.016579728  0.3172052179 -0.91658546
#> 53  -0.638267765 -0.0613897917 -1.49885549
#> 54  -0.393177668  0.4231895627 -0.39000543
#> 55   0.315066286 -0.0135989028 -0.23802612
#> 56  -0.206886476 -0.6344506662 -0.54421188
#> 57  -0.380695782  0.1352875748  0.41842187
#> 58   0.454662865  0.2871291150 -0.12736877
#> 59   0.698395176  0.8745551209  0.08813133
#> 60   0.002066597  0.5781615388 -0.48416731
#> 61   0.602052445  0.1627114402 -0.45554390
#> 62   0.667053096  0.6384063432 -0.24614729
#> 63   0.146398232  0.7718500752  0.52811988
#> 64  -0.221185661  0.4668284221  0.12477663
#> 65   0.155264797 -0.5140739200 -0.49413689
#> 66  -0.627852760  0.3251556311  0.07314759
#> 67  -0.572461977  0.0736930863 -0.44518113
#> 68   0.384719963 -0.1092963263 -0.10564023
#> 69   0.355586459 -0.1164171575 -0.03802341
#> 70  -0.223645662  0.6320866217  0.06860153
#> 71  -0.037293344 -0.0506443167 -0.51500718
#> 72   0.440669049 -0.0236515034  0.09222835
#> 73   0.422761175 -0.3255240786 -0.00457742
#> 74   0.302136030  0.3455368886  0.45977094
#> 75  -0.693687762 -0.2364536150  0.64338532
#> 76   0.213993782 -0.2228821972 -0.20377654
#> 77   0.478991538  0.0127877320  0.09126109
#> 78  -0.021061640  0.9062995351  0.26600065
#> 79   0.366936656 -0.9642218010 -0.21181354
#> 80   0.317332636 -0.2274703882 -0.24813865
#> 81   0.033924757 -0.4441533627  0.09830820
#> 82  -0.097086010 -0.7189316295 -0.04816339
#> 83   0.466315679 -0.6312621273  0.60830528
#> 84   0.360517656 -0.2426326671  0.25458867
#> 85   0.099957234 -0.0210815092 -0.10434897
#> 86   0.257386116  0.4094191413  0.11460680
#> 87   0.521113776  0.1107341039  0.50121907
#> 88  -0.231318277  0.3261272056  0.17652141
#> 89  -0.139257986 -0.0938767121  0.15778646
#> 90  -0.089096212  0.5142339480 -0.56119311
#> 91  -0.483938316  0.0002131897 -1.31751660
#> 92   0.294016380  0.7367092403  0.23769939
#> 93  -0.350252995  0.2616738974 -0.28230521
#> 94  -0.117801793  0.6781376672 -0.16881622
#> 95  -0.302984248 -0.5978910957 -0.44101143
#> 96   0.586815230  0.5927443389  0.05029313
#> 97   0.041766475 -0.4433018422  0.47065851
#> 98  -0.352479399 -0.6790116839 -0.94533162
#> 99   0.150443954  0.9629798578  0.71592808
#> 100 -0.201053605 -0.0467728665  0.10896298
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl   ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>
#>  1      1        4       70        1  0.0636 -0.287    0.329 
#>  2      2        4       70        1  0.164  -1.52     0.151 
#>  3      3        4       70        1 -0.0547 -0.614    0.530 
#>  4      4        4       70        1 -0.833  -0.881   -0.128 
#>  5      5        4       70        1 -0.598  -0.00603 -0.501 
#>  6      6        4       70        1  0.214   0.206   -0.0864
#>  7      7        4       70        1  0.429  -0.160    0.588 
#>  8      8        4       70        1 -0.0246  0.284    0.660 
#>  9      9        4       70        1 -0.0288 -0.256   -0.544 
#> 10     10        4       70        1  0.269   0.230    0.200 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc     depot  centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>     <dbl>  <dbl>
#> 1      1     1     4    70     1  4.26  52.6  1.39 0.0811 23.5    7.20e+2 1234. 
#> 2      1    14     4    70     1  4.26  52.6  1.39 0.0811 13.3    1.04e-5  696. 
#> 3      2     1     4    70     1  4.71  15.4  1.16 0.306  65.6    8.48e+2 1009. 
#> 4      2    14     4    70     1  4.71  15.4  1.16 0.306   2.62   2.29e-4   40.2
#> 5      3     1     4    70     1  3.79  37.9  1.70 0.100  35.9    5.76e+2 1360. 
#> 6      3    14     4    70     1  3.79  37.9  1.70 0.100  14.2    1.46e-7  538. 
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>