Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1   -0.088713429  0.657778814  0.482539929
#> 2    0.024779356  0.698031088 -0.561559905
#> 3   -0.304029379 -0.499987189  0.016071504
#> 4    0.451357827 -0.846511350 -0.704708476
#> 5    0.543587732  0.251604918  0.316886613
#> 6   -0.330443435 -0.107931104 -0.393796622
#> 7   -0.136087817 -0.483145924  0.139740759
#> 8   -0.689545759 -0.862492706  0.103687178
#> 9    0.420246256 -0.549074178  0.047185767
#> 10   0.425258603  0.862404356 -0.167390190
#> 11  -0.227509156  0.490325163  0.176819358
#> 12  -0.239046457 -0.113336591  0.357634726
#> 13   0.611836325 -0.007100804 -0.087267629
#> 14  -0.359398492 -0.813314450  0.233305397
#> 15   0.795502466  0.199594328  0.562177110
#> 16   0.054789244 -0.389553037  0.482093493
#> 17  -0.962858199  0.467670321 -0.455429897
#> 18   0.657466098  0.099285889  0.308068118
#> 19   0.184699317  0.280267768  0.041000396
#> 20  -0.584463968 -0.029128631 -0.493139413
#> 21  -1.055908743 -0.196718060 -0.403538741
#> 22  -0.489590553  0.127412217 -0.011038822
#> 23   0.159909243 -0.096954495 -0.004201785
#> 24  -0.474577951 -0.140669221 -0.538539754
#> 25   0.183522904  0.022811374  0.313685457
#> 26  -0.367552647 -0.033672415 -0.061232256
#> 27  -0.189274828  0.533978446 -0.469404763
#> 28  -0.657634274  0.128731772  0.847483033
#> 29   1.051023817 -0.435081054  0.202482521
#> 30   0.856535556 -0.680246237 -0.007453003
#> 31  -0.194833273 -0.675718041 -0.728366580
#> 32  -0.910815595 -0.234757413 -0.195455592
#> 33   0.028085043  0.497306684  0.241112751
#> 34  -0.267231270  0.027329298 -1.054405756
#> 35  -0.060056485  0.615777949 -0.586502007
#> 36   0.225776754  0.555060904  0.311025225
#> 37   0.222120429 -0.193915454  0.202972937
#> 38   0.404186490 -0.151245337 -0.342301520
#> 39   0.364625781  0.567361622  0.151514594
#> 40  -1.025132129  0.448902967  0.026359040
#> 41   0.289285640 -0.246130026 -0.473581338
#> 42  -0.154038533  0.927733919  0.181588188
#> 43  -0.022025325  0.872619302  0.134119925
#> 44  -0.489804904  0.253779648  0.277949613
#> 45   0.462010225 -0.311165299 -0.127382476
#> 46   0.260468366  0.368789886 -0.311692292
#> 47   0.257120952 -0.283153065  0.686607586
#> 48   0.108468325 -1.001809097  0.476099240
#> 49  -0.302927339  0.057778694  0.148540853
#> 50  -0.188303023  0.590328845 -1.392815286
#> 51  -0.867989622 -0.031555704 -0.130444325
#> 52  -0.108838058 -0.463916571  0.215364862
#> 53  -0.269351149  0.525022161 -0.808047559
#> 54  -0.479084981  0.604059651 -0.288989776
#> 55  -0.690879045  0.666040888 -0.327805402
#> 56   0.019824115 -0.581495193 -0.034490510
#> 57   0.100676719  0.351308076  0.057216712
#> 58  -0.353533119  0.013805637  0.363635507
#> 59  -0.531403314 -0.172698824  0.225684048
#> 60  -0.127281637 -0.921468066  0.365019656
#> 61   0.282144592 -0.593071630  0.345984144
#> 62   0.926555192 -0.629831396  0.184001929
#> 63  -0.158215134 -0.787065510  1.338552988
#> 64   0.495007072 -1.073195859 -0.706941213
#> 65  -0.536996930 -0.184222258 -0.625729577
#> 66   0.497247849 -0.991201755  0.128326509
#> 67  -0.363989123 -0.571189872  0.174799835
#> 68   0.279521913  0.427851248 -0.225328712
#> 69   0.089714046  0.281764087  0.117748885
#> 70  -0.311261757  0.382812185  0.137124681
#> 71   0.581333200  0.270634546  0.253852640
#> 72  -0.544889684 -0.355512812  0.186011634
#> 73  -0.869792795  0.163932335 -0.223194866
#> 74   0.140964820  0.551690769 -0.038489024
#> 75   1.115526759  0.337677529 -0.494461345
#> 76   0.275930487  0.429192004 -0.055415346
#> 77  -0.530956256 -0.271757881 -0.019338360
#> 78   0.021425997  1.208914818  0.161434293
#> 79   0.278166139  0.150441467  0.743808277
#> 80   0.151298485 -0.888936251 -0.259950783
#> 81   0.103944767 -0.107277416  0.464530432
#> 82   0.566073487 -0.057631857 -0.094237606
#> 83  -0.118308179  0.700750884 -0.664477413
#> 84  -0.420754231  0.063564301 -0.048467328
#> 85  -0.009610177  0.186680771  0.024464290
#> 86   0.044863656  0.237561169  0.289888628
#> 87  -0.416610518 -0.143699649  0.341219857
#> 88  -1.181518399  0.899520309  0.265760067
#> 89  -0.063059367  0.066039348 -0.605469917
#> 90   0.208931133 -0.123476949 -0.665562659
#> 91   0.288140477  0.250935100 -0.080051566
#> 92   0.101704856 -0.055483309  1.352698049
#> 93   0.517078953  1.104117086 -0.752468252
#> 94  -0.291951481 -0.080034155 -0.151091387
#> 95  -0.206103903  0.426983923 -0.153528703
#> 96   0.611757708  0.194217066 -0.194746053
#> 97  -0.653219723 -0.290014889 -0.036819270
#> 98  -0.666456522  0.181437121  0.753227158
#> 99   0.054177406  0.738175261  0.702728105
#> 100  0.576569459  0.386006889  0.438725816
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>  <dbl>   <dbl>
#>  1      1        4       70        1 -0.0887  0.658  0.483 
#>  2      2        4       70        1  0.0248  0.698 -0.562 
#>  3      3        4       70        1 -0.304  -0.500  0.0161
#>  4      4        4       70        1  0.451  -0.847 -0.705 
#>  5      5        4       70        1  0.544   0.252  0.317 
#>  6      6        4       70        1 -0.330  -0.108 -0.394 
#>  7      7        4       70        1 -0.136  -0.483  0.140 
#>  8      8        4       70        1 -0.690  -0.862  0.104 
#>  9      9        4       70        1  0.420  -0.549  0.0472
#> 10     10        4       70        1  0.425   0.862 -0.167 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  3.66 135.  1.62  0.0271 10.2  6.10e+2 1373.
#> 2      1    14     4    70     1  3.66 135.  1.62  0.0271 10.4  4.35e-7 1402.
#> 3      2     1     4    70     1  4.10 141.  0.570 0.0291  4.86 1.31e+3  683.
#> 4      2    14     4    70     1  4.10 141.  0.570 0.0291 10.0  7.89e-1 1411.
#> 5      3     1     4    70     1  2.95  42.5 1.02  0.0695 24.1  9.43e+2 1025.
#> 6      3    14     4    70     1  2.95  42.5 1.02  0.0695 19.4  1.73e-3  826.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>