Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object

A posologyr list, created by the posologyr function.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01,n_simul=100)
#> $eta
#>           ETA_Cl        ETA_Vc      ETA_Ka
#> 1   -0.044049876 -0.0120377910 -0.67027441
#> 2    0.228016122  0.1299564567  0.34524178
#> 3    0.102536504  0.5058590258 -0.72456296
#> 4    0.401063409 -0.1478569654 -0.24533496
#> 5   -0.375433913  0.1854364430  0.48514232
#> 6    0.234783681  0.1170565292  0.60424573
#> 7    0.372724027 -0.5448452386  0.68084639
#> 8    0.257131265  0.1767940490 -0.45423636
#> 9    0.128327055 -0.5324107204  0.81557320
#> 10  -0.056913535 -0.0698750625  0.12315829
#> 11   0.225201700  0.0541327146 -0.04198025
#> 12   0.170864063 -0.3670821648 -0.01721197
#> 13  -0.423071538 -0.6819435131  0.31928197
#> 14  -0.438234120 -0.0132368657  0.16005287
#> 15   0.922851010 -0.4253075984 -0.23450608
#> 16  -0.424191418  0.1506061114  0.75787693
#> 17  -0.366820399 -0.5525847135  0.34973046
#> 18  -0.042726006 -0.4199159606 -0.24966535
#> 19   0.036229102 -0.0466033706 -0.74870508
#> 20   0.011623185 -0.6062065390 -0.01102187
#> 21  -0.629581380 -0.8231650720 -0.23342516
#> 22   0.277918892 -0.0294934173  0.23743597
#> 23  -0.421071642 -0.0102589189  0.20203115
#> 24   0.792381401  0.4832724753  0.24004799
#> 25   0.198939160 -0.6279284349  0.06702564
#> 26   0.598526460  0.2149585063  0.49762888
#> 27   0.986743004 -0.1622235649 -0.58130351
#> 28   0.390405373 -0.1801576592  0.28588111
#> 29   0.945842465  0.2626887237  0.86995526
#> 30   0.188663167 -0.0807140761  0.09716079
#> 31  -0.088969783 -0.0987169941  0.36513652
#> 32   0.153035610 -0.8652340308  0.00344665
#> 33  -0.046721325  0.1436209388 -0.46852093
#> 34   0.019381484 -0.2269917817  0.08542614
#> 35   0.245389966 -0.7717380624  0.51636261
#> 36  -0.200068396 -0.0800793009 -0.35835238
#> 37  -0.196678963 -0.1695677624  0.04753315
#> 38  -0.230722849  0.5023724703  0.50883933
#> 39  -0.177874919 -0.4366989942 -0.16817051
#> 40  -0.428397443 -0.0844715122  0.04401541
#> 41   0.635130014 -0.1657014662  0.30958870
#> 42   0.030360191  0.1401526841  0.27986799
#> 43   0.898371820 -0.1456220655  0.22517109
#> 44  -0.296908332  0.0551430655  0.33620295
#> 45   0.281826752 -0.6290201045  0.50347373
#> 46   0.402354097 -0.7839088627  0.45952466
#> 47   0.020846183 -0.0161649359 -0.62646759
#> 48   0.663887876 -0.1808053691  0.29801961
#> 49  -0.691912975  0.2449973748 -0.60046476
#> 50   0.297704366 -0.4921825785  0.01435859
#> 51   0.166263155  0.0969946114 -0.23230679
#> 52  -0.245983954 -0.2776330763 -0.73100830
#> 53  -0.718131516 -0.0871710488  0.76674236
#> 54  -0.000993685 -0.4190102182 -0.12660044
#> 55  -0.073444784  0.5128064728  0.20990251
#> 56   0.234629226  0.2050359311 -0.41390813
#> 57   0.388626059  0.2618384686  0.38087620
#> 58  -0.581130235 -0.0006429916  0.55263727
#> 59  -0.458361528 -0.1587482709  0.25907577
#> 60   0.236538944 -0.0848981951 -0.28405298
#> 61  -0.386701151  0.2567882581 -0.44162266
#> 62   0.175366318  0.1335698544  0.14936255
#> 63   0.033052480 -0.3606192422 -0.35467447
#> 64   1.304840666  0.0817008211  0.32078064
#> 65   0.239918862 -0.4816338722  0.87636420
#> 66   0.362422650 -0.9232239825  0.54838174
#> 67  -0.277644836 -0.4080551509 -0.05595675
#> 68   0.135944394  0.8840124642  0.37540071
#> 69  -0.106096587  1.1770633848 -0.58027532
#> 70   0.010670007 -0.3751817959 -0.36897061
#> 71   0.013493993 -0.2704261993 -0.36098706
#> 72  -0.400580488  0.3082472003 -0.18859022
#> 73   0.103279556 -0.0140901018  0.08438148
#> 74   1.399658915  0.3270396956  0.07236936
#> 75   0.708801368  0.3839428016  0.74711377
#> 76   0.414587225 -0.3531146770  0.31706221
#> 77  -0.239801063  0.0161834972  0.66255731
#> 78   0.104787315 -0.2035054088  0.54112668
#> 79   0.036988153 -0.5488424126  0.10670877
#> 80   0.069491223  0.6717979079  0.44525905
#> 81   0.377935935 -0.1818528448 -0.01921628
#> 82   0.403731181  0.4790696560  0.18929109
#> 83  -0.453990549 -0.3539746654 -0.15536628
#> 84  -0.411228396 -0.5727552303 -0.12951299
#> 85  -0.047945264 -0.1471551784  0.36044417
#> 86   0.222000679  0.1174855901  0.25511885
#> 87   0.078522459 -0.2515646120 -0.34135292
#> 88  -0.581944586  0.6466931919 -0.16495548
#> 89  -0.263145253 -0.3926270900 -0.03191925
#> 90  -0.406586328  0.0071772693  0.62375169
#> 91  -0.647605135 -0.0484198982 -0.57263443
#> 92   0.033520462 -0.3783284351  0.09077106
#> 93   0.478486216  0.0120541333 -0.05227890
#> 94  -0.112268914  0.4451587220  0.98641093
#> 95   0.213375909  0.8886303488  0.19742391
#> 96  -0.099078059 -0.3548218378 -0.93162550
#> 97  -0.389812715 -0.2570484548  0.13225109
#> 98  -0.397577350  0.1479877830 -0.11833040
#> 99   0.440411319  0.1913958438  0.13459354
#> 100 -0.517234868 -0.3658871641 -0.13418693
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1 -0.0440 -0.0120 -0.670
#>  2      2        4       70        1  0.228   0.130   0.345
#>  3      3        4       70        1  0.103   0.506  -0.725
#>  4      4        4       70        1  0.401  -0.148  -0.245
#>  5      5        4       70        1 -0.375   0.185   0.485
#>  6      6        4       70        1  0.235   0.117   0.604
#>  7      7        4       70        1  0.373  -0.545   0.681
#>  8      8        4       70        1  0.257   0.177  -0.454
#>  9      9        4       70        1  0.128  -0.532   0.816
#> 10     10        4       70        1 -0.0569 -0.0699  0.123
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  3.83  69.2 0.512 0.0553  8.95    1.37e+3  619.
#> 2      1    14     4    70     1  3.83  69.2 0.512 0.0553 15.1     1.77e+0 1046.
#> 3      2     1     4    70     1  5.02  79.7 1.41  0.0630 15.8     7.08e+2 1256.
#> 4      2    14     4    70     1  5.02  79.7 1.41  0.0630 11.0     7.52e-6  880.
#> 5      3     1     4    70     1  4.43 116.  0.485 0.0382  5.14    1.39e+3  597.
#> 6      3    14     4    70     1  4.43 116.  0.485 0.0382 11.0     2.56e+0 1282.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>