Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>          ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.29240046 -0.302467432  0.635009421
#> 2    0.46710774  0.097760304 -0.851714341
#> 3   -0.85224345 -0.653869432 -0.082863776
#> 4   -0.38665110  0.316860621  0.309884857
#> 5    0.15028676  0.072231445 -0.464912937
#> 6    0.48613294  0.248919979  0.045008529
#> 7   -0.50890063 -0.816630530 -0.672791163
#> 8   -0.48753623  0.709146602 -0.221721618
#> 9   -0.06378355 -0.510103564  0.649600841
#> 10  -0.22320311  0.603759897  0.147165699
#> 11  -1.11929508  0.704451692  0.357904087
#> 12   0.46206295  0.312185857 -0.369949437
#> 13   0.14919488 -0.110902728  0.194557156
#> 14  -0.43847122  0.480165117  0.555094326
#> 15   0.59973803  0.863324957 -0.390062532
#> 16   0.80953804  0.392441856 -0.082862307
#> 17  -0.67428582  0.729051284  0.195096354
#> 18  -0.43017432  0.166290536  0.078287926
#> 19  -0.37856897 -0.051466708  0.183714997
#> 20   0.33336075 -0.434787331  0.015203142
#> 21  -0.58414471  0.869987920 -0.074821637
#> 22  -0.71294426 -0.501238049 -0.410333504
#> 23  -0.81063921 -0.212075992  0.520778452
#> 24   0.19523950 -0.001228841 -0.208077362
#> 25   0.42143804  0.377545960 -0.157549253
#> 26  -0.54323620 -0.608272190  0.772877889
#> 27  -0.30849309  0.538921546 -0.078775701
#> 28   0.52352441  0.422995008 -0.412149782
#> 29  -0.33734245  0.585401974 -0.131989693
#> 30  -0.42021363  0.392428032 -0.710826981
#> 31   0.53240586  0.135718930  0.004564378
#> 32  -0.04784371  0.290650845  0.427265854
#> 33  -0.25154641 -0.279028619 -0.041668698
#> 34   0.05307384  0.097353127  0.463490856
#> 35  -0.23738906 -0.161263927  0.737411402
#> 36  -0.28385146  0.013332445 -0.528907406
#> 37  -0.59397611  0.045478364 -0.915383013
#> 38   0.15884363  0.448263431 -0.931869174
#> 39   0.06567524  0.314073299 -1.399574329
#> 40  -0.12503379  0.156477123 -0.627578611
#> 41   0.78570763 -0.110447541 -0.078058750
#> 42   0.52542982  0.220824512 -0.616535322
#> 43  -0.15666617 -0.505110441 -0.925110717
#> 44   0.27449581 -0.572307464  0.348047298
#> 45   0.18234723  0.038494519 -0.687027624
#> 46   0.12631165  0.050762710  0.794954054
#> 47   0.21028303  0.201513889 -0.082393226
#> 48  -0.23062825  0.038619829 -0.119008959
#> 49  -0.15235417  0.388660017  0.375257290
#> 50  -0.23833276 -0.032178159  0.877117337
#> 51  -0.24051740 -0.692741874  0.689739286
#> 52  -0.63941784  0.447342919  0.234618905
#> 53  -0.61277976 -0.277391472 -0.055269950
#> 54   0.30455242 -0.149730569  0.294105636
#> 55   0.27004027 -0.680717296 -0.103224889
#> 56  -0.76272226 -0.906629044 -0.082317655
#> 57  -0.20530678  0.400044344 -0.478261429
#> 58  -0.17541108  0.837604472 -0.073023620
#> 59  -0.65901957 -0.138111096  0.492537550
#> 60   0.93998199 -0.130473906 -0.006762794
#> 61   0.06665932  0.569419142  0.278363188
#> 62   0.80025250 -0.266428480  0.452425145
#> 63  -0.21622239  0.109080254 -0.051127860
#> 64  -0.23932245  0.178694844  0.513250886
#> 65   0.47494610  0.378435144 -0.459917328
#> 66   0.24811125 -0.460369720 -0.470282814
#> 67   0.10923430 -0.422973361 -0.241072561
#> 68   0.09744668  0.456481108 -0.525822335
#> 69   0.43149733 -0.228821218 -0.027179019
#> 70  -0.29291198 -0.528821775  0.901290606
#> 71   0.81103694 -0.604605626  0.272956762
#> 72   0.04916817  0.780552757 -0.421212528
#> 73  -0.09387032  0.072339450  0.257411587
#> 74   0.43554917 -0.460511569 -0.397656537
#> 75   0.31136682 -0.116693783  0.341943487
#> 76   0.37550949  0.227986998 -0.259606841
#> 77  -0.34791398  0.010273354  0.253953620
#> 78   0.32508748  0.116769334 -0.240167507
#> 79   0.55893873  0.470031733  0.121411336
#> 80   0.01917492  0.473872583 -0.349872145
#> 81   0.23396935 -0.277154005  0.248778824
#> 82   0.29716751  0.214738655 -0.107935861
#> 83   0.43157435 -0.485257727 -0.055918879
#> 84  -0.25603968 -0.206225295  0.163754509
#> 85  -0.35067292  0.015892138  0.369233292
#> 86  -0.09831728  0.361310652 -0.689600916
#> 87   0.50428607  0.511394871 -0.394424597
#> 88   0.13719222 -0.342536870  0.119378437
#> 89  -0.08984965  0.410146015  0.048891661
#> 90  -0.62975505  0.294901712  0.752029527
#> 91  -0.14292041  0.389444648 -0.345075807
#> 92   0.34159530 -0.552544715 -0.514006069
#> 93  -0.41989493  0.013954432 -0.073246462
#> 94  -0.84727896 -1.256192295 -0.484494808
#> 95   0.76851048 -0.548817839  0.743366184
#> 96   0.43028985 -0.545459179 -1.075965281
#> 97  -0.21788798 -0.102046814  0.712432262
#> 98   0.46817517  0.884463020 -0.222604216
#> 99   1.03607719 -0.489571881  0.425923052
#> 100  1.05365108  0.085974265 -0.016173460
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1  0.292  -0.302   0.635 
#>  2      2        4       70        1  0.467   0.0978 -0.852 
#>  3      3        4       70        1 -0.852  -0.654  -0.0829
#>  4      4        4       70        1 -0.387   0.317   0.310 
#>  5      5        4       70        1  0.150   0.0722 -0.465 
#>  6      6        4       70        1  0.486   0.249   0.0450
#>  7      7        4       70        1 -0.509  -0.817  -0.673 
#>  8      8        4       70        1 -0.488   0.709  -0.222 
#>  9      9        4       70        1 -0.0638 -0.510   0.650 
#> 10     10        4       70        1 -0.223   0.604   0.147 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  5.36  51.7 1.89  0.104  27.5  5.04e+2 1425.
#> 2      1    14     4    70     1  5.36  51.7 1.89  0.104   9.85 1.10e-8  509.
#> 3      2     1     4    70     1  6.38  77.2 0.427 0.0827  6.83 1.46e+3  527.
#> 4      2    14     4    70     1  6.38  77.2 0.427 0.0827 10.2  5.67e+0  789.
#> 5      3     1     4    70     1  1.71  36.4 0.920 0.0469 26.6  1.01e+3  969.
#> 6      3    14     4    70     1  1.71  36.4 0.920 0.0469 30.4  6.43e-3 1106.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>