Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc        ETA_Ka
#> 1   -0.339543400  0.206595140 -0.1195110925
#> 2   -0.856949085 -0.355622486  1.0792330571
#> 3    0.693714605  0.342436881  0.1749119655
#> 4   -0.308307161  0.085294477 -0.0222103122
#> 5    0.121336655 -0.922232769  0.0324035470
#> 6   -0.073696362  0.106714581  0.3607969538
#> 7   -0.229660289  0.005639216 -0.4616547105
#> 8   -0.160786311  0.128015727 -0.0008484833
#> 9   -0.338583602  0.112357677  0.6451749555
#> 10   0.330180718  0.370140895  0.6354151972
#> 11   0.412778712 -0.204214095 -0.7151505206
#> 12  -0.987103182  0.381833266 -0.1384300372
#> 13   0.023170669 -0.752943778  0.5358180713
#> 14  -0.323458914 -0.067557654 -0.1764024783
#> 15  -0.248406524 -0.651660197 -0.2796383969
#> 16  -0.799544115  0.095865765  0.2578133504
#> 17   0.163203957 -0.271092546 -0.4675516921
#> 18  -0.577975000 -0.398151474 -0.5347044082
#> 19   0.245157093 -0.388038086  0.0955765743
#> 20   0.245355602 -0.132975610  0.0328155916
#> 21  -0.103588489 -0.059922368 -0.7050122746
#> 22  -0.399229132 -0.061145474  0.0364302837
#> 23  -0.673534032  0.476337378 -0.5584505179
#> 24  -0.051829857 -0.343469077 -0.0244868428
#> 25   0.205320877 -0.476993844 -0.3059509436
#> 26  -0.125836573 -0.580332517 -0.0814602852
#> 27  -0.380397913  0.681845560  0.0728530375
#> 28  -0.021292621 -0.006831388 -0.0797725684
#> 29   0.007792051  0.410427886 -0.2016134652
#> 30   0.112592309  0.399868934 -0.1339409723
#> 31  -0.339242372  0.420641903  0.4880308272
#> 32  -0.146095317 -0.282284818 -0.2740679582
#> 33  -0.506520040 -0.189025828  0.3242589323
#> 34   0.183102036  0.211654967 -0.1151652486
#> 35   0.315760180  0.054283462 -0.3311633117
#> 36   0.091968159 -0.251315453 -0.0222043506
#> 37   0.172820285 -0.179604244 -0.5947912936
#> 38  -0.525977441 -0.398304795  0.2180632053
#> 39  -0.159376088  0.239872542 -0.3333519328
#> 40   0.912474269  0.228805760  0.7388150903
#> 41   0.191321762 -0.090464301 -1.1664971576
#> 42   0.034715628 -0.148902781 -0.0639329382
#> 43  -0.126013446 -0.340682875 -0.0575711453
#> 44  -0.196126261 -0.344072937  0.1310400744
#> 45   0.233121254 -0.002910270 -0.3664469455
#> 46  -0.496672818  0.320192111  0.5766715984
#> 47  -0.459640649 -0.508915097  0.2681975329
#> 48  -0.510932647 -0.621685994  0.4054888865
#> 49  -0.349070722 -1.031533196  0.0789243571
#> 50   0.213334369  0.964805696  0.7330169424
#> 51  -0.157793321 -0.465683872  0.0048620761
#> 52   0.325747328  0.271068565  0.1151425878
#> 53  -0.234183107  0.016122266 -0.1006040498
#> 54   0.064399073 -0.462669828  0.3444748322
#> 55   0.885089652  0.474965069  1.0515827469
#> 56   0.273021618 -0.057002053 -0.0127435573
#> 57   0.451063968  0.144973575  0.5514230087
#> 58   0.421275194 -0.655824948  0.5753149621
#> 59   0.631729761  0.269056665 -0.2816058203
#> 60  -0.367469264 -0.512408014 -0.0147753883
#> 61  -0.029264927 -0.317812306 -0.2331866218
#> 62  -0.227797869 -0.093899913 -0.3087121906
#> 63  -0.546803321 -0.326718015  0.3199012149
#> 64  -0.692708948  0.008667657 -0.1530141932
#> 65  -0.433959729 -0.591775861  0.0021232180
#> 66   0.475369365 -0.302474876 -0.1020144105
#> 67   0.309896513 -0.179497243  0.3793254300
#> 68   0.048723637  0.661496922 -0.1927757856
#> 69  -0.025913010  0.076230903 -0.1609988150
#> 70  -0.031230855  0.281177738 -0.1698390038
#> 71  -0.314608705 -0.080409069 -0.5882668137
#> 72  -0.467888224  0.283152199  0.1920559399
#> 73   0.666672106 -0.232438766 -0.3060493343
#> 74   0.046631436  0.861881264  0.0568417162
#> 75  -0.226032713 -0.413384120  0.7422616791
#> 76   0.294331480  0.091280204 -0.7446711553
#> 77  -0.008474957 -0.188975517 -0.0685421327
#> 78   0.294524172 -0.452147163 -0.7563548566
#> 79   0.125266937 -0.665399772 -0.2474150585
#> 80   0.985300751  0.198107564  0.0578887677
#> 81   0.210154500 -0.297130052  0.1500712407
#> 82   0.470293769 -0.871263356 -0.7513577878
#> 83   0.507431123  0.239430680  0.0083553893
#> 84  -0.408497640  0.768850735  0.3283896781
#> 85   0.506184579  0.081793213 -0.6416711866
#> 86  -0.203196601 -0.439614145  0.0858012318
#> 87  -0.163756880  0.298360829  0.2670387356
#> 88   0.327030764  0.331046879 -0.7657731099
#> 89   0.306884057 -0.296169354  0.3614439121
#> 90   0.319501805 -0.590170831 -0.2716944968
#> 91   0.926673303  0.868689141  0.1147392502
#> 92  -0.681698079  0.961894273  0.4284031431
#> 93   0.077431998  0.118116425 -0.0940788546
#> 94   0.021799167 -0.301507602  0.7301811297
#> 95   0.283569155 -0.262043430 -0.2844613748
#> 96   0.673876874  0.159957977  0.3647605739
#> 97  -0.196524988 -0.067194264 -0.7119452988
#> 98  -0.294866187  0.245086441  0.3544343429
#> 99  -0.190358129 -0.058494765  0.1340906111
#> 100  0.391347315  0.188504209 -0.2128329615
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl   ETA_Vc    ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>     <dbl>
#>  1      1        4       70        1 -0.340   0.207   -0.120   
#>  2      2        4       70        1 -0.857  -0.356    1.08    
#>  3      3        4       70        1  0.694   0.342    0.175   
#>  4      4        4       70        1 -0.308   0.0853  -0.0222  
#>  5      5        4       70        1  0.121  -0.922    0.0324  
#>  6      6        4       70        1 -0.0737  0.107    0.361   
#>  7      7        4       70        1 -0.230   0.00564 -0.462   
#>  8      8        4       70        1 -0.161   0.128   -0.000848
#>  9      9        4       70        1 -0.339   0.112    0.645   
#> 10     10        4       70        1  0.330   0.370    0.635   
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc     depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>     <dbl> <dbl>
#> 1      1     1     4    70     1  2.85  86.1 0.887 0.0331 11.0   1.04e+ 3  950.
#> 2      1    14     4    70     1  2.85  86.1 0.887 0.0331 15.3   1.01e- 2 1318.
#> 3      2     1     4    70     1  1.70  49.1 2.94  0.0346 35.2   2.40e+ 2 1729.
#> 4      2    14     4    70     1  1.70  49.1 2.94  0.0346 25.6  -5.09e-10 1257.
#> 5      3     1     4    70     1  8.00  98.6 1.19  0.0812 11.4   8.31e+ 2 1128.
#> 6      3    14     4    70     1  8.00  98.6 1.19  0.0812  7.13  1.56e- 4  703.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>