Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>          ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.62665990  0.044957630       0       0
#> 2    0.48148979  0.130804815       0       0
#> 3    0.04372297  0.123178979       0       0
#> 4    0.19458752 -0.153448921       0       0
#> 5   -0.12495651  0.033718658       0       0
#> 6    0.04983392 -0.066296423       0       0
#> 7    0.25284885 -0.025283957       0       0
#> 8    0.29513713 -0.056587228       0       0
#> 9   -0.10166639  0.104107272       0       0
#> 10  -0.49013102 -0.098700270       0       0
#> 11   0.73244940  0.012935179       0       0
#> 12  -0.19267536  0.123574931       0       0
#> 13  -0.36896348  0.034228916       0       0
#> 14  -0.38125436  0.040865822       0       0
#> 15   0.19119891  0.089339812       0       0
#> 16   0.18152317 -0.022197401       0       0
#> 17  -0.07207033 -0.101323164       0       0
#> 18   0.26605910 -0.096134234       0       0
#> 19  -0.15974729 -0.059973417       0       0
#> 20  -0.43361677 -0.118015799       0       0
#> 21   0.02197230  0.101399471       0       0
#> 22   0.01508975 -0.024045491       0       0
#> 23   0.12286271  0.061048259       0       0
#> 24   0.36951389  0.003432946       0       0
#> 25   0.13608251 -0.220774046       0       0
#> 26  -0.25041015 -0.131099442       0       0
#> 27   0.21547139 -0.018664374       0       0
#> 28   0.11382125 -0.082203083       0       0
#> 29  -0.07308638  0.078745097       0       0
#> 30  -0.19151835  0.048215035       0       0
#> 31   0.08881165  0.048378551       0       0
#> 32  -0.59680461 -0.127776904       0       0
#> 33  -0.04823973 -0.318634246       0       0
#> 34  -0.57304347  0.254940853       0       0
#> 35  -0.29741874  0.112821178       0       0
#> 36   0.04348932 -0.166521619       0       0
#> 37  -0.33482933  0.034778223       0       0
#> 38   0.02739093  0.141119857       0       0
#> 39  -0.19602999  0.024713023       0       0
#> 40  -0.30975449  0.091472055       0       0
#> 41   0.43939321 -0.067858520       0       0
#> 42  -0.44226965 -0.006151901       0       0
#> 43  -0.10857326  0.045499707       0       0
#> 44   0.42907958 -0.056347038       0       0
#> 45  -0.48734690 -0.093767524       0       0
#> 46   0.31193272  0.105370722       0       0
#> 47   0.09463286  0.114808529       0       0
#> 48   0.00410981  0.226357995       0       0
#> 49   0.06761154  0.103657246       0       0
#> 50  -0.10221789  0.022561999       0       0
#> 51   0.02028231  0.099782373       0       0
#> 52   0.14024887  0.259582413       0       0
#> 53  -0.03875573  0.085521175       0       0
#> 54   0.41533283 -0.125510860       0       0
#> 55   0.20593044  0.005735578       0       0
#> 56   0.23009706  0.118244251       0       0
#> 57   0.24988877 -0.099280951       0       0
#> 58   0.29702027 -0.043971424       0       0
#> 59   0.04450779 -0.037128939       0       0
#> 60   0.09842954 -0.100032342       0       0
#> 61   0.27722384 -0.034280120       0       0
#> 62   0.08599876 -0.029398443       0       0
#> 63  -0.32735707 -0.110007955       0       0
#> 64   0.01016133  0.187123329       0       0
#> 65   0.13440573  0.176493958       0       0
#> 66   0.09667399  0.087154083       0       0
#> 67  -0.02772744 -0.003926715       0       0
#> 68  -0.36735805  0.013176725       0       0
#> 69  -0.05168285 -0.064321443       0       0
#> 70   0.04224009  0.074211706       0       0
#> 71  -0.09278490  0.075073559       0       0
#> 72  -0.04307056 -0.044366382       0       0
#> 73   0.11020359 -0.155611398       0       0
#> 74   0.23201338 -0.076162428       0       0
#> 75   0.15466998 -0.082635802       0       0
#> 76   0.44660435  0.005898338       0       0
#> 77   0.35787390 -0.170663460       0       0
#> 78   0.12120905 -0.230755765       0       0
#> 79   0.17398940 -0.268613798       0       0
#> 80  -0.42692257  0.089475883       0       0
#> 81   0.14055962  0.084141269       0       0
#> 82  -0.15437642  0.060914829       0       0
#> 83   0.45565388 -0.182354488       0       0
#> 84  -0.23812691  0.057738778       0       0
#> 85  -0.43804877 -0.067960922       0       0
#> 86  -0.06375270 -0.169911771       0       0
#> 87  -0.30073600 -0.128611070       0       0
#> 88  -0.11409419 -0.118226946       0       0
#> 89  -0.04977425 -0.007745256       0       0
#> 90   0.36492247 -0.094961591       0       0
#> 91   0.35464487  0.025660613       0       0
#> 92  -0.25031101 -0.059128315       0       0
#> 93   0.22043003 -0.077484693       0       0
#> 94   0.24765101 -0.038187404       0       0
#> 95  -0.08454069 -0.336958234       0       0
#> 96   0.58340298  0.142732117       0       0
#> 97   0.32056502 -0.085910251       0       0
#> 98   0.27463132  0.039104017       0       0
#> 99   0.47707567  0.131542930       0       0
#> 100  0.10782103 -0.015328251       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.627   0.0450
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.481   0.131 
#>  3      3     0.21      80    65    19.8     0.041      0.12  0.0437  0.123 
#>  4      4     0.21      80    65    19.8     0.041      0.12  0.195  -0.153 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.125   0.0337
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.0498 -0.0663
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.253  -0.0253
#>  8      8     0.21      80    65    19.8     0.041      0.12  0.295  -0.0566
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.102   0.104 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.490  -0.0987
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21     Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.133  20.4 0.041  0.12 21.6   440.  
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.133  20.4 0.041  0.12  3.71   75.6 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.403  22.2 0.041  0.12 16.2   360.  
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.403  22.2 0.041  0.12  0.257   5.70
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.260  22.0 0.041  0.12 18.2   400.  
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.260  22.0 0.041  0.12  0.868  19.1 
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>