Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1   -0.152913779  0.433529602 -0.148697471
#> 2    0.044203135 -0.341842501 -0.262233780
#> 3    0.994491748  0.450159803  0.252778741
#> 4    0.948480396  0.280762695  0.399131834
#> 5   -0.623852459  0.174650884  0.996849138
#> 6    0.574237519 -0.064632781 -0.508568985
#> 7   -0.358199637 -0.406360795 -0.079344852
#> 8    0.140389081  0.645069852 -0.090563816
#> 9    0.786740996  0.008152017 -0.109420784
#> 10   0.161044614 -0.733824113 -0.489286486
#> 11   0.014490267 -0.317392314  0.735793155
#> 12  -0.633605223  0.221299887  0.431187155
#> 13  -0.793703265  0.325869799  0.171640746
#> 14   0.012429528 -0.105853511 -0.452336977
#> 15   0.572035710 -0.152178758  1.049908478
#> 16   0.505660236  0.110757063  0.195346531
#> 17  -0.808705199 -0.007276988  0.058306251
#> 18  -0.045643542  0.945483808  0.218479945
#> 19  -0.595932585 -0.008505751  0.141773230
#> 20  -0.139991630 -0.033513011 -0.018855255
#> 21  -0.355158227 -0.638217618 -0.210814072
#> 22   0.329736239 -0.111058616  0.377644562
#> 23   0.114608260  0.711071607 -0.081631618
#> 24  -0.080451033 -0.220808481  0.508584971
#> 25  -0.338899361 -0.132769260  0.174683711
#> 26   0.829322186 -0.071889804 -0.360971003
#> 27  -0.374670685 -0.123396213  0.834875906
#> 28  -0.204583298 -0.485813112 -0.044773091
#> 29  -0.511922108 -0.123484170  0.310156982
#> 30  -0.112590917  0.608757556  0.342174426
#> 31   0.366173230  0.198519952  0.213942204
#> 32  -0.447943890 -0.082723023  0.902780555
#> 33  -0.798027691 -0.494730461 -0.134286122
#> 34   0.460909523  0.242347946 -0.661442477
#> 35  -0.264213012  0.179962219  0.154553139
#> 36   0.476506654  0.332565824  0.343420435
#> 37  -0.771814824 -0.274902327  0.654177750
#> 38  -0.933230574 -0.098348923  0.372161925
#> 39   0.199698898  0.104372726  0.078689098
#> 40   0.009206382 -0.752668532  0.004983360
#> 41  -0.397845897 -0.216064238  0.004636063
#> 42  -0.097954690  0.009804983 -0.845067391
#> 43   0.801073668  0.326981704 -0.118625452
#> 44  -0.074151747 -0.036248649  0.620744829
#> 45  -0.821405372 -0.184765889  0.109080470
#> 46   0.125742097  0.025143973 -0.210711395
#> 47  -0.201594750  0.108116347  0.372871863
#> 48  -0.051079637 -0.291795166 -0.098750266
#> 49   0.209482138  0.588070055  0.552939855
#> 50   0.328996364  0.504367104  0.022790431
#> 51   0.445168250  0.245900992 -0.037625666
#> 52  -0.258889754 -0.235328952 -0.404925843
#> 53   0.229214346 -0.415065123 -0.514798625
#> 54   0.468529832 -1.032799252  0.238513946
#> 55   0.659272296 -0.414951373  0.536109517
#> 56   0.163476707  0.307363191 -0.565407710
#> 57  -0.757061393  0.119685170 -0.089259651
#> 58   0.357480262  0.236219387  0.824563991
#> 59  -0.303731977  0.315503081  0.154570595
#> 60   0.815974655  0.150413289 -0.189834962
#> 61   0.096728548 -0.182897124 -0.061828349
#> 62   0.181746724  0.057132021 -0.298459782
#> 63   0.548405829  0.132135826 -0.482457941
#> 64   0.366756109  0.936819154 -1.074977116
#> 65  -1.572222795 -0.770730092 -0.277210748
#> 66   0.474077401  0.349507440 -0.106848859
#> 67   0.290001029  0.266219530 -0.129630507
#> 68   0.158460096  0.243032972 -0.147826584
#> 69  -0.022919788  0.014039573 -0.138058535
#> 70   0.211104069 -0.110744025 -0.153458465
#> 71   0.313787614 -0.033969095  0.544121164
#> 72  -0.348534175 -0.165086573 -0.680412268
#> 73  -0.514561639 -0.289093301  0.160747529
#> 74   0.387899104 -0.230303917 -0.789991284
#> 75   1.102437762  0.489658659 -0.990151812
#> 76  -0.131798040 -0.977799207  0.713913239
#> 77   0.433136362 -0.268465473  0.144822268
#> 78   0.138757404  0.676402391 -0.285350179
#> 79  -0.560895378 -0.845275475  0.803928955
#> 80   0.301899013  0.243470516  0.152039725
#> 81   0.140441205  0.526959996  0.080722973
#> 82   0.575884620  0.519631502 -0.339393041
#> 83  -1.531594502  0.509641664  0.595724386
#> 84  -0.305399890  0.328664564  0.288376513
#> 85  -0.553923158 -0.758051980  0.283886414
#> 86   0.410554296  0.239468634 -0.303922237
#> 87  -0.473409755 -0.660859444 -0.787589621
#> 88  -0.177610751 -0.892335982 -0.183971265
#> 89   0.196992709 -0.320339153  0.083745999
#> 90  -0.186673754  0.287791915  0.267556112
#> 91  -0.617264961 -0.710066139  0.695965366
#> 92  -0.218146362  0.432588831 -0.049485096
#> 93   0.016921001 -0.798093625 -0.124103905
#> 94   0.064352333  0.207985538  0.005112638
#> 95  -0.442570155  0.157781883  0.662008893
#> 96   0.097975799  0.429238854 -0.316350229
#> 97   0.332746961 -0.079633310  0.351946052
#> 98  -0.281941902  0.176455239 -0.075722997
#> 99  -0.227528133  0.049616721 -0.733783383
#> 100 -0.204357435  0.070144301 -0.791663501
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl   ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>
#>  1      1        4       70        1 -0.153   0.434   -0.149 
#>  2      2        4       70        1  0.0442 -0.342   -0.262 
#>  3      3        4       70        1  0.994   0.450    0.253 
#>  4      4        4       70        1  0.948   0.281    0.399 
#>  5      5        4       70        1 -0.624   0.175    0.997 
#>  6      6        4       70        1  0.574  -0.0646  -0.509 
#>  7      7        4       70        1 -0.358  -0.406   -0.0793
#>  8      8        4       70        1  0.140   0.645   -0.0906
#>  9      9        4       70        1  0.787   0.00815 -0.109 
#> 10     10        4       70        1  0.161  -0.734   -0.489 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  3.43 108.  0.862 0.0318  8.62    1.06e+3  931.
#> 2      1    14     4    70     1  3.43 108.  0.862 0.0318 12.4     1.44e-2 1341.
#> 3      2     1     4    70     1  4.18  49.7 0.769 0.0841 16.9     1.13e+3  840.
#> 4      2    14     4    70     1  4.18  49.7 0.769 0.0841 14.2     5.12e-2  707.
#> 5      3     1     4    70     1 10.8  110.  1.29  0.0985 10.7     7.75e+2 1173.
#> 6      3    14     4    70     1 10.8  110.  1.29  0.0985  5.09    4.16e-5  559.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>