Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.031593674  1.018304729  0.003604231
#> 2    0.260757099  0.117779260 -0.851799112
#> 3    0.344674180 -0.491621352  0.437415015
#> 4    0.348223095  0.356790784 -0.298383012
#> 5   -0.236874674 -0.337603093 -0.088335657
#> 6    0.243430710  0.402269230 -0.274055925
#> 7    0.534555909  0.001144003  0.960512882
#> 8    0.637173005 -0.018567829 -0.325907511
#> 9    0.358979914  0.372753590  1.269822102
#> 10  -0.731336653  0.563204426  0.731549062
#> 11   0.462872416  0.101504117 -0.147881089
#> 12   0.295277396 -0.052776261  0.560266061
#> 13   0.202767814 -0.100089450  0.538064838
#> 14   0.347315121 -0.182024434 -0.234104983
#> 15  -0.263023961 -0.309992118  0.162212089
#> 16   0.972123686 -0.238146591  0.379062770
#> 17   0.211195933 -0.522783940  0.097953134
#> 18   0.156418699  0.082286540  0.109045013
#> 19   0.191911473  0.093287084 -0.293187794
#> 20   0.215541217  0.521469937  0.083352762
#> 21   0.506604300 -0.119622623  0.500137983
#> 22  -0.335886395  0.795916060 -0.154774267
#> 23   0.037742920  0.154088421 -0.153089242
#> 24  -0.122816865  0.041707818  0.601793409
#> 25   0.114182376 -0.500330672  0.333921287
#> 26  -0.063076941 -0.193821715  0.316069102
#> 27   0.249372255  0.556340636 -0.092155489
#> 28   0.061082029 -0.158801595  0.363072367
#> 29   1.189363982  1.034944062 -0.933345515
#> 30  -0.748192084 -0.340966436  0.076016098
#> 31   0.938354932  0.139271070  0.230451545
#> 32  -0.261116183 -0.086272510  0.326144207
#> 33   0.240608302  0.086493214 -0.418999062
#> 34  -0.492036614  0.439728992 -0.291617556
#> 35  -0.492256178 -0.534195813 -0.019141206
#> 36  -0.125567578 -0.271717819  0.319731319
#> 37   0.370799419 -0.185063630  0.383247094
#> 38  -0.631406589  0.667959959 -0.014967768
#> 39  -0.007386424 -1.111443262 -0.311305349
#> 40  -0.933988617 -0.021018472 -0.202591598
#> 41  -0.153742490  0.477754510  0.616298022
#> 42   0.349561558  0.527463565  0.947681704
#> 43  -0.294202338 -0.371968558  0.057069920
#> 44  -0.291237758 -0.406501110  0.323410964
#> 45   0.244312445 -0.032000329  0.068304169
#> 46  -0.204582661 -0.150617960  0.611564731
#> 47   0.446783444 -0.250470285  0.612967818
#> 48  -0.493876765  0.134085877  0.425689564
#> 49   0.307250637  0.386890520 -0.835670639
#> 50  -0.038612464 -0.167109416  0.330010505
#> 51   0.206893031  0.286864625 -0.117722587
#> 52  -0.120633683 -0.481805029 -0.151949757
#> 53  -0.605255139 -0.642191543 -0.062933940
#> 54   0.443374048 -0.354468496 -0.345418690
#> 55  -0.222584328  0.309158243  0.368864058
#> 56   0.053563763  0.029955900  0.275462547
#> 57  -0.860291239  0.791646491 -0.428013700
#> 58   0.804204112 -0.213431175  0.141982680
#> 59  -0.348465116  0.770541158  0.417175625
#> 60   0.160561219 -0.062321918 -0.324324339
#> 61   0.033212790  0.085759519 -0.091371406
#> 62   0.279061796 -0.344522974 -0.224868634
#> 63  -0.387760986 -0.479844574 -0.358231425
#> 64   0.141075851 -0.277479811 -0.618507095
#> 65  -0.225616074  0.150498058  0.621597525
#> 66   0.078265709 -0.369957141  0.190294420
#> 67   0.184069166 -0.604396453 -0.310720073
#> 68  -0.146460568  0.379104330  0.727632119
#> 69   0.150101090 -0.635243621  1.067375823
#> 70   0.809088379 -0.354727283  0.155052586
#> 71   0.098249848  0.210783581  0.263303231
#> 72  -0.871700223  0.026975524  0.192206387
#> 73  -0.181512326  0.208830380 -0.572009083
#> 74  -0.387837499  0.500387116  0.104560906
#> 75  -0.031893468  0.457683743 -0.094770111
#> 76  -0.089241857  0.251748122 -0.495987907
#> 77  -0.384656713  0.056268078  0.640463012
#> 78  -0.342615699 -0.599068715 -1.061944652
#> 79   1.033383466  0.881969051 -0.108706754
#> 80   0.275930163 -0.051366888 -0.303249586
#> 81   0.591363119 -0.002490939 -0.011452023
#> 82   0.341678100  0.335226262  0.010201604
#> 83  -0.582918136  0.238643748  0.414894181
#> 84  -0.426374450 -0.267549344  1.019267377
#> 85   0.304654085 -0.389446357 -0.179981046
#> 86   0.043227715  0.577114869  0.046193835
#> 87  -0.183143250 -0.288270561 -0.590119330
#> 88   0.594492304 -0.241546231  0.301227432
#> 89   0.304343483  0.519535167  0.022783017
#> 90  -0.382857324  0.915563319 -0.900375440
#> 91   0.267363719  0.297818410 -0.166012086
#> 92   0.009438450 -0.738548528  0.270737622
#> 93  -0.110501085  0.184930905  0.527625710
#> 94  -0.204643252  0.198652972 -0.748440199
#> 95  -0.020450348 -0.099849076 -0.335846250
#> 96   0.500383882 -1.218122837 -0.306041797
#> 97   0.066334832  0.485836662  0.056915158
#> 98  -0.547176006 -0.242079704  0.288200884
#> 99  -0.515042215 -0.329207699 -0.406086070
#> 100 -0.132483486  0.315543092  0.219861002
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl   ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>    <dbl>
#>  1      1        4       70        1  0.0316  1.02     0.00360
#>  2      2        4       70        1  0.261   0.118   -0.852  
#>  3      3        4       70        1  0.345  -0.492    0.437  
#>  4      4        4       70        1  0.348   0.357   -0.298  
#>  5      5        4       70        1 -0.237  -0.338   -0.0883 
#>  6      6        4       70        1  0.243   0.402   -0.274  
#>  7      7        4       70        1  0.535   0.00114  0.961  
#>  8      8        4       70        1  0.637  -0.0186  -0.326  
#>  9      9        4       70        1  0.359   0.373    1.27   
#> 10     10        4       70        1 -0.731   0.563    0.732  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  4.13 194.  1.00  0.0213  5.36    9.52e+2 1038.
#> 2      1    14     4    70     1  4.13 194.  1.00  0.0213  7.87    2.05e-3 1525.
#> 3      2     1     4    70     1  5.19  78.7 0.427 0.0659  6.74    1.46e+3  531.
#> 4      2    14     4    70     1  5.19  78.7 0.427 0.0659 12.0     5.68e+0  949.
#> 5      3     1     4    70     1  5.65  42.8 1.55  0.132  29.9     6.42e+2 1279.
#> 6      3    14     4    70     1  5.65  42.8 1.55  0.132   8.33    1.16e-6  357.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>