Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>           ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1   -0.060968322  0.1346077810       0       0
#> 2   -0.163051035 -0.0225454206       0       0
#> 3    0.210807515  0.0844137257       0       0
#> 4    0.240446018 -0.0274419663       0       0
#> 5    0.113663181  0.0398067323       0       0
#> 6    0.080001541 -0.0055107247       0       0
#> 7    0.506985753  0.0218253751       0       0
#> 8   -0.050356836 -0.0282845864       0       0
#> 9   -0.006191521 -0.0209614889       0       0
#> 10  -0.149813397  0.0987709620       0       0
#> 11   0.463227472 -0.0640848970       0       0
#> 12   0.503375636 -0.0512628082       0       0
#> 13   0.413732449 -0.1337700722       0       0
#> 14  -0.039954250 -0.0783342157       0       0
#> 15  -0.050564252  0.0786916339       0       0
#> 16  -0.238644188 -0.0693268750       0       0
#> 17  -0.070907844 -0.0940179716       0       0
#> 18  -0.253983671 -0.0686839609       0       0
#> 19   0.098979228 -0.0351389504       0       0
#> 20  -0.042137417 -0.0297754847       0       0
#> 21  -0.025514306 -0.0364561384       0       0
#> 22  -0.165130965 -0.0732782454       0       0
#> 23   0.101780660 -0.0518667610       0       0
#> 24  -0.398358943  0.1378010141       0       0
#> 25  -0.079564860  0.0976783883       0       0
#> 26  -0.198875467  0.1042672118       0       0
#> 27   0.148831225 -0.1354654034       0       0
#> 28  -0.004943539 -0.0000401736       0       0
#> 29   0.011613604  0.0155800724       0       0
#> 30  -0.232834393 -0.0539323027       0       0
#> 31   0.163164205  0.0165747119       0       0
#> 32  -0.546665499 -0.0755801058       0       0
#> 33   0.379085162 -0.0011198787       0       0
#> 34   0.313428542 -0.0600350785       0       0
#> 35  -0.395757735  0.1225276612       0       0
#> 36  -0.683302892 -0.0913832557       0       0
#> 37  -0.044369846  0.0789764040       0       0
#> 38   0.329638904  0.0526005202       0       0
#> 39   0.311005753  0.1233421094       0       0
#> 40   0.051812214 -0.1341193906       0       0
#> 41  -0.022283976  0.1255768983       0       0
#> 42  -0.047551574  0.0390971542       0       0
#> 43  -0.124439182 -0.1222942731       0       0
#> 44   0.124428029 -0.1063171579       0       0
#> 45  -0.203222531 -0.1789872704       0       0
#> 46  -0.628238190  0.0244139918       0       0
#> 47  -0.484488948  0.1835910165       0       0
#> 48   0.117325807  0.0642654074       0       0
#> 49  -0.356469368 -0.0172197040       0       0
#> 50  -0.358465080  0.0390900313       0       0
#> 51   0.278894094 -0.0102148327       0       0
#> 52  -0.009200683 -0.1358150941       0       0
#> 53   0.157188263  0.2178758897       0       0
#> 54   0.559111934  0.1334199501       0       0
#> 55   0.097381120  0.0409099637       0       0
#> 56  -0.151782517 -0.1396939688       0       0
#> 57   0.053322918 -0.0573648088       0       0
#> 58   0.586172592 -0.0017183729       0       0
#> 59   0.338678927  0.1096711507       0       0
#> 60   0.275598732 -0.0768845268       0       0
#> 61  -0.354224657  0.0131669812       0       0
#> 62  -0.293131393 -0.0544838471       0       0
#> 63  -0.507796782  0.0876017697       0       0
#> 64  -0.125630167 -0.1536689695       0       0
#> 65  -0.195938151  0.1327789120       0       0
#> 66  -0.116434593 -0.0321410508       0       0
#> 67   0.357148289 -0.1045534410       0       0
#> 68   0.048022105  0.0951830326       0       0
#> 69  -0.100194075 -0.0011833871       0       0
#> 70  -0.089183700  0.1029217156       0       0
#> 71  -0.298811607  0.0949611993       0       0
#> 72   0.175006876 -0.0532782254       0       0
#> 73  -0.095355118  0.1055230207       0       0
#> 74   0.265041991  0.0961866938       0       0
#> 75   0.034206387 -0.2030016041       0       0
#> 76   0.057097346 -0.0110621459       0       0
#> 77  -0.189564779 -0.0079788968       0       0
#> 78  -0.174555396  0.0025919883       0       0
#> 79  -0.260144649  0.0707832962       0       0
#> 80  -0.384180761  0.1026477590       0       0
#> 81   0.497079114 -0.0923792902       0       0
#> 82   0.292091944  0.0732377373       0       0
#> 83  -0.188685859 -0.0771866921       0       0
#> 84   0.236794885  0.2867962420       0       0
#> 85   0.021019782 -0.0460204574       0       0
#> 86  -0.331255635  0.2591400119       0       0
#> 87   0.073550452  0.0801622456       0       0
#> 88  -0.283540820  0.1489227633       0       0
#> 89   0.180292350  0.0223135410       0       0
#> 90   0.079054215 -0.0337309049       0       0
#> 91  -0.305707744  0.0137814295       0       0
#> 92  -0.162021090 -0.0259202522       0       0
#> 93  -0.129978342 -0.0633258973       0       0
#> 94   0.071569299 -0.0802236729       0       0
#> 95  -0.116359447 -0.0202684573       0       0
#> 96   0.118695856  0.0762269214       0       0
#> 97  -0.360085086  0.0231907932       0       0
#> 98   0.195852062  0.2529643760       0       0
#> 99  -0.209186048  0.0477090206       0       0
#> 100  0.323612668 -0.0034479630       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21   ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>    <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.0610   0.135  
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.163   -0.0225 
#>  3      3     0.21      80    65    19.8     0.041      0.12  0.211    0.0844 
#>  4      4     0.21      80    65    19.8     0.041      0.12  0.240   -0.0274 
#>  5      5     0.21      80    65    19.8     0.041      0.12  0.114    0.0398 
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.0800  -0.00551
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.507    0.0218 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0504  -0.0283 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.00619 -0.0210 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.150    0.0988 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.234  22.3 0.041  0.12 73.2  1631. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.234  22.3 0.041  0.12  4.45   99.1
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.212  19.0 0.041  0.12 87.1  1658. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.212  19.0 0.041  0.12  6.61  126. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.307  21.2 0.041  0.12 72.9  1544. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.307  21.2 0.041  0.12  2.31   49.0
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>