Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc        ETA_Ka
#> 1   -0.085890461  0.709785799 -0.1276153055
#> 2   -0.183236788 -0.299495976  0.5230789589
#> 3    0.701088378 -0.133972832 -0.2151410925
#> 4   -0.628631277  0.022698026 -0.2976174817
#> 5   -0.131416785 -0.698641584  0.4198811886
#> 6    0.777011055  0.179659925 -0.6975238305
#> 7   -0.009408740 -0.172434660 -0.4833160018
#> 8   -0.029721945  0.464479351  0.3289083215
#> 9    0.213856236  0.375641207 -0.4153553040
#> 10  -0.175469846 -0.525354516 -0.4228754054
#> 11  -0.599364690  0.015821500  0.5301410360
#> 12  -0.232786496 -0.079599636  0.1969052012
#> 13   0.548440702 -0.011475730  0.2952382473
#> 14  -0.347130610  0.250217573  0.7889850066
#> 15  -0.104825366 -0.864102961  0.3038567038
#> 16   0.613082770 -0.445028613 -0.4915806031
#> 17  -0.623684401 -0.069811824 -0.9648319417
#> 18   0.014148364 -0.290300550  0.5456567086
#> 19  -0.828746344  0.441178457 -0.3177000597
#> 20  -0.382049533 -0.370028907 -0.2862403684
#> 21  -0.426476961  0.231237078 -0.3248719336
#> 22   0.377703675 -0.145759047  0.2673803625
#> 23   0.381343761 -0.129657283  1.4358859169
#> 24   0.398036554  0.295024423  0.4579943231
#> 25  -0.026718563  0.335103611 -0.0482765813
#> 26   0.089703494  0.060322042 -0.1763333821
#> 27   0.401624355 -0.552438470 -0.1981180744
#> 28  -0.016830470  0.228576104  0.2626686035
#> 29   0.649153771 -0.589698923  0.0002002498
#> 30   0.367671438  0.028834193 -0.1266141224
#> 31  -0.392202766 -0.128617167 -0.0712523408
#> 32   0.162472445 -0.018779139 -0.1996299161
#> 33   0.257775737  0.168320145  0.2948284334
#> 34   0.809433970 -0.662809310  0.4368551216
#> 35  -0.306553390  0.087161284 -0.1791058972
#> 36  -0.536034876 -0.132160797 -0.5157129350
#> 37  -0.314693026 -0.043070167 -0.1289146971
#> 38   0.164000420  0.641979701  0.7153432779
#> 39  -0.075506362 -0.174732705  0.0875154943
#> 40  -0.216055452 -0.031443785  0.4366327609
#> 41   0.378640568 -0.160276668 -0.4293408180
#> 42  -0.117732256  0.008560496 -0.1989228577
#> 43  -0.067241688  0.369743361 -0.1083861930
#> 44  -0.214307896  0.958557860  0.0587962754
#> 45  -0.708943624 -0.186308879  0.1941011790
#> 46   0.074606416  0.787663242 -0.1380275242
#> 47  -0.310787924  0.071922929 -0.3795914660
#> 48   0.028186376 -0.347881907 -0.7712114153
#> 49  -0.051018870  0.514006837  0.2426269853
#> 50   0.124268360  0.727837179 -0.2887380894
#> 51  -0.401236181 -0.160046564 -0.4321824604
#> 52   0.273601260  0.322473226 -0.5985528887
#> 53  -0.719116530 -0.150001321 -0.9575097410
#> 54  -0.069866021  0.073396025  0.3601171328
#> 55  -0.145839708 -0.102879564  0.1217667266
#> 56  -0.144791094  1.023342238 -0.0918657938
#> 57  -0.538046151  0.517850073  0.1292814356
#> 58  -0.048000584  0.278687886 -0.0683035512
#> 59  -0.264435064 -0.156772900 -0.2923588308
#> 60  -0.061654958  0.161694363 -0.0992417820
#> 61   0.523041333 -0.727838654 -0.2513508710
#> 62  -0.048062178  0.027278319  1.3119897386
#> 63   0.372141150  0.145466056  0.3536797356
#> 64   0.267930274  0.308620149  0.3402285958
#> 65  -0.368592997 -0.863577361  0.8906548833
#> 66   0.316591789 -0.415107019 -0.1986952988
#> 67   0.672976749 -0.337160924 -0.6026391284
#> 68  -0.714453219 -0.701654930 -0.2873635040
#> 69  -0.261293187 -0.252806796 -0.0481177323
#> 70   0.677473140 -0.233524637  0.4154433002
#> 71  -0.469183010 -0.479845960  0.2092685944
#> 72   0.954077095  0.680934789  0.4550119325
#> 73  -0.044554175  0.204658573  0.0742854681
#> 74  -0.562557686 -0.229945196 -0.0752843218
#> 75   0.031807077  1.096754774 -0.1068302456
#> 76  -0.502197231  0.036064387  0.0263583019
#> 77  -0.102129070  0.302921682 -0.5276893336
#> 78   0.499146576  0.817978026 -0.7560802841
#> 79   0.002179547  0.673568748 -0.6955083783
#> 80  -0.094116386 -0.338031370 -0.0508623174
#> 81   0.398168419 -0.310817205  1.2935456696
#> 82  -0.505971585 -0.356607739 -0.8599110747
#> 83  -0.316408393  0.256039503 -0.6915717745
#> 84  -0.113253216 -1.000943753 -0.1171174982
#> 85  -0.489026076 -0.296446538 -0.2607647311
#> 86   0.186279951  0.117003245  0.3829098518
#> 87   1.070826361  0.443479742 -0.5624480565
#> 88   0.391513398  0.105992887 -0.1938154859
#> 89  -0.048582522 -0.462992038 -0.2115678248
#> 90   0.192970670  0.446075725  0.7450532418
#> 91   0.220043873 -0.101875267  0.3479393863
#> 92  -0.276681137  0.075769923  0.2008613761
#> 93  -0.184349376 -1.128261281 -0.5484235004
#> 94   0.136938156 -0.210761722 -0.2457185949
#> 95   0.312268359 -0.071299703 -0.8094183727
#> 96  -0.547456448 -0.440272304 -0.2633317450
#> 97  -0.327831822  0.423627538  0.0440176753
#> 98  -0.670775359 -0.475975230 -0.4027093757
#> 99  -0.296284821 -0.607648384 -0.1178243380
#> 100 -0.543636846  0.087369982 -0.1116125029
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka   ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1 -0.0859   0.710  -0.128
#>  2      2        4       70        1 -0.183   -0.299   0.523
#>  3      3        4       70        1  0.701   -0.134  -0.215
#>  4      4        4       70        1 -0.629    0.0227 -0.298
#>  5      5        4       70        1 -0.131   -0.699   0.420
#>  6      6        4       70        1  0.777    0.180  -0.698
#>  7      7        4       70        1 -0.00941 -0.172  -0.483
#>  8      8        4       70        1 -0.0297   0.464   0.329
#>  9      9        4       70        1  0.214    0.376  -0.415
#> 10     10        4       70        1 -0.175   -0.525  -0.423
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc   depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>   <dbl> <dbl>
#> 1      1     1     4    70     1  3.67 142.  0.880 0.0258  6.66 1.04e+3  948.
#> 2      1    14     4    70     1  3.67 142.  0.880 0.0258 10.2  1.12e-2 1445.
#> 3      2     1     4    70     1  3.33  51.9 1.69  0.0642 26.5  5.81e+2 1377.
#> 4      2    14     4    70     1  3.33  51.9 1.69  0.0642 16.6  1.74e-7  860.
#> 5      3     1     4    70     1  8.06  61.2 0.806 0.132  13.9  1.10e+3  852.
#> 6      3    14     4    70     1  8.06  61.2 0.806 0.132   6.38 3.08e-2  391.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>