Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object = NULL, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>            ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1    0.2760469916 -1.103391e-01       0       0
#> 2    0.2123217716 -1.025323e-01       0       0
#> 3   -0.4332796076  2.160092e-01       0       0
#> 4    0.1841515084  8.314281e-03       0       0
#> 5    0.0991815130 -1.873895e-02       0       0
#> 6   -0.0008222685  4.532465e-02       0       0
#> 7   -0.0799782748 -5.276183e-02       0       0
#> 8    0.1600808977 -4.101946e-02       0       0
#> 9   -0.2478820303  1.170131e-02       0       0
#> 10  -0.0242566503 -2.251104e-01       0       0
#> 11   0.0724615037 -6.067992e-02       0       0
#> 12  -0.4451492797 -1.141321e-01       0       0
#> 13   0.2765827107 -8.425464e-03       0       0
#> 14   0.1241226438  4.534709e-02       0       0
#> 15   0.3459065505  5.264848e-05       0       0
#> 16   0.0478741550  1.532322e-01       0       0
#> 17  -0.2445880966  4.620112e-02       0       0
#> 18  -0.1964167001  1.469577e-01       0       0
#> 19  -0.5061162545 -7.853829e-02       0       0
#> 20  -0.2222394500 -4.740376e-02       0       0
#> 21   0.2591230571  2.252739e-02       0       0
#> 22   0.1671111967 -7.801496e-02       0       0
#> 23  -0.0661428502  6.991889e-02       0       0
#> 24  -0.3541246988 -7.494164e-02       0       0
#> 25   0.3451216835 -1.401530e-01       0       0
#> 26   0.4601745567 -2.030081e-01       0       0
#> 27   0.4535769249  8.511164e-02       0       0
#> 28   0.1999382000  7.846619e-02       0       0
#> 29   0.1148818212  2.240800e-01       0       0
#> 30  -0.0796957569  3.261784e-02       0       0
#> 31  -0.6023454314 -3.602944e-02       0       0
#> 32  -0.2013580527 -2.291209e-02       0       0
#> 33  -0.4668571821 -8.547308e-02       0       0
#> 34   0.0334157577 -3.110813e-02       0       0
#> 35   0.5288704823  4.236432e-02       0       0
#> 36   0.0374843153 -2.046570e-01       0       0
#> 37   0.0759482221 -1.974550e-01       0       0
#> 38  -0.3248376117 -8.604247e-02       0       0
#> 39  -0.1817786336 -2.386919e-01       0       0
#> 40   0.5561600722 -1.044828e-01       0       0
#> 41   0.2251872319  1.990750e-01       0       0
#> 42   0.1766485819 -2.351933e-01       0       0
#> 43   0.1551819191  1.584004e-01       0       0
#> 44   0.0679339174  2.592005e-02       0       0
#> 45  -0.3460114897  9.234846e-02       0       0
#> 46   0.0276523389  1.962052e-01       0       0
#> 47   0.3412336705 -1.988754e-01       0       0
#> 48   0.0287815485  4.286052e-03       0       0
#> 49   0.4547153905  1.293878e-01       0       0
#> 50   0.1721695358 -3.293519e-02       0       0
#> 51   0.1910167703 -9.062503e-02       0       0
#> 52  -0.0990469571  1.109665e-01       0       0
#> 53  -0.3859827199  1.725208e-01       0       0
#> 54  -0.0068438338  1.737898e-01       0       0
#> 55   0.5281678567  9.084942e-02       0       0
#> 56   0.0166975959 -1.414630e-01       0       0
#> 57  -0.1541484169  2.441565e-02       0       0
#> 58  -0.6521012068 -6.926882e-02       0       0
#> 59  -0.0705294531  7.406453e-02       0       0
#> 60   0.1057104636 -6.689009e-02       0       0
#> 61   0.4027423598 -9.510482e-02       0       0
#> 62   0.1070295172  1.066252e-01       0       0
#> 63  -0.0765787123 -7.659124e-02       0       0
#> 64   0.1617748474 -1.218232e-01       0       0
#> 65  -0.0418204385 -1.397452e-01       0       0
#> 66   0.2996533014 -9.004619e-02       0       0
#> 67  -0.4024937860  6.661144e-02       0       0
#> 68   0.0888982667  7.158667e-02       0       0
#> 69   0.5929019381 -3.930382e-02       0       0
#> 70   0.0257729068  1.777915e-02       0       0
#> 71  -0.0308680344  2.004741e-01       0       0
#> 72   0.1559960519 -6.333718e-02       0       0
#> 73   0.1294663482 -2.610334e-01       0       0
#> 74  -0.4878756618  3.960546e-03       0       0
#> 75  -0.3154128893  1.638702e-02       0       0
#> 76   0.1114229608  5.154291e-02       0       0
#> 77   0.3608020942  7.777132e-02       0       0
#> 78   0.1080253517 -7.391018e-03       0       0
#> 79   0.0809165705  4.700881e-02       0       0
#> 80   0.1751978490 -1.919959e-02       0       0
#> 81   0.0351491941 -1.201661e-01       0       0
#> 82  -0.3499789011  8.382569e-02       0       0
#> 83  -0.2379049785  1.125718e-01       0       0
#> 84   0.0720199569  1.543139e-01       0       0
#> 85   0.0275614101  3.014959e-03       0       0
#> 86   0.0105049035 -4.102538e-02       0       0
#> 87   0.1827048554  4.784915e-02       0       0
#> 88   0.3455565798 -1.194516e-01       0       0
#> 89  -0.4227714339 -7.589431e-02       0       0
#> 90  -0.0167457533 -5.841418e-02       0       0
#> 91  -0.1892594260 -1.235315e-01       0       0
#> 92   0.0168560178 -2.008946e-01       0       0
#> 93  -0.4049136291  2.124442e-01       0       0
#> 94  -0.0394525705 -1.298238e-01       0       0
#> 95  -0.3312562306  3.438841e-02       0       0
#> 96   0.2844170614  3.434849e-02       0       0
#> 97  -0.2881400000 -1.498940e-01       0       0
#> 98  -0.0734692385  1.789153e-01       0       0
#> 99  -0.3880441833  3.898916e-02       0       0
#> 100  0.0947099729  1.620971e-01       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21    ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>     <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12  0.276    -0.110  
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.212    -0.103  
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.433     0.216  
#>  4      4     0.21      80    65    19.8     0.041      0.12  0.184     0.00831
#>  5      5     0.21      80    65    19.8     0.041      0.12  0.0992   -0.0187 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.000822  0.0453 
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0800   -0.0528 
#>  8      8     0.21      80    65    19.8     0.041      0.12  0.160    -0.0410 
#>  9      9     0.21      80    65    19.8     0.041      0.12 -0.248     0.0117 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.0243   -0.225  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21     Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.328  17.4 0.041  0.12 21.8   380. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.328  17.4 0.041  0.12  0.586  10.2
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.308  17.6 0.041  0.12 22.0   386. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.308  17.6 0.041  0.12  0.694  12.2
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.161  24.2 0.041  0.12 17.8   430. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.161  24.2 0.041  0.12  2.26   54.5
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>