Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>            ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.1871304234 -0.077520784       0       0
#> 2    0.1997402775 -0.056734575       0       0
#> 3   -0.5091322100 -0.112221040       0       0
#> 4   -0.0167784031  0.045238812       0       0
#> 5   -0.1872637669  0.016627134       0       0
#> 6   -0.0757815470 -0.125791808       0       0
#> 7    0.4338216712 -0.027143968       0       0
#> 8   -0.2652606311 -0.141240884       0       0
#> 9    0.0817488972 -0.002591941       0       0
#> 10   0.2679231151  0.135027054       0       0
#> 11   0.4146767136 -0.029977936       0       0
#> 12  -0.0145154266  0.008463797       0       0
#> 13   0.1279002551  0.178998043       0       0
#> 14  -0.0860691391  0.091894946       0       0
#> 15   0.1683063583 -0.047992244       0       0
#> 16  -0.4741740674 -0.132266271       0       0
#> 17  -0.4574688482 -0.104565768       0       0
#> 18  -0.3384982234  0.268083211       0       0
#> 19   0.0227203028  0.106029241       0       0
#> 20  -0.3038359479 -0.250907876       0       0
#> 21   0.5552821354 -0.210698532       0       0
#> 22   0.1084783519 -0.216380647       0       0
#> 23   0.2940786396 -0.066088337       0       0
#> 24  -0.1402123543  0.252354459       0       0
#> 25  -0.3157584166  0.147582501       0       0
#> 26   0.1917585195  0.142487153       0       0
#> 27  -0.0422006990 -0.115049342       0       0
#> 28  -0.2791197974 -0.005809084       0       0
#> 29  -0.3424307941 -0.056904121       0       0
#> 30   0.1844916046  0.201814294       0       0
#> 31   0.1238734036  0.236182786       0       0
#> 32  -0.3214603271  0.070283466       0       0
#> 33  -0.5214182580 -0.202949379       0       0
#> 34   0.2798043789 -0.056051081       0       0
#> 35  -0.1675399404 -0.045624741       0       0
#> 36   0.0699729641  0.132473073       0       0
#> 37   0.4721293284  0.120266860       0       0
#> 38  -0.0568992169 -0.109529981       0       0
#> 39  -0.0259081973 -0.008468644       0       0
#> 40  -0.0690867501 -0.052593590       0       0
#> 41  -0.2495373816 -0.056214083       0       0
#> 42   0.3279748176  0.124281225       0       0
#> 43  -0.3938503377 -0.075441971       0       0
#> 44  -0.3629206356  0.183980405       0       0
#> 45   0.0481161011  0.037295492       0       0
#> 46  -0.3845139182 -0.160900202       0       0
#> 47  -0.4491838523  0.022166612       0       0
#> 48  -0.2523644115 -0.056792830       0       0
#> 49   0.3234918783 -0.035663145       0       0
#> 50  -0.2501970927 -0.067157455       0       0
#> 51   0.3535022362 -0.131864739       0       0
#> 52   0.0281052596 -0.139351904       0       0
#> 53  -0.3103624065 -0.047307207       0       0
#> 54   0.2041592833 -0.211516995       0       0
#> 55  -0.6301054715 -0.069921431       0       0
#> 56  -0.1498568260  0.014931330       0       0
#> 57  -0.5356852904 -0.083868642       0       0
#> 58   0.2160442746 -0.172227652       0       0
#> 59  -0.0213050805 -0.091433170       0       0
#> 60  -0.5047183015 -0.150362172       0       0
#> 61   0.4359539499 -0.066643885       0       0
#> 62  -0.0335831909 -0.067365319       0       0
#> 63  -0.1388044479 -0.108694331       0       0
#> 64  -0.6733912149 -0.112332531       0       0
#> 65   0.1657199975 -0.019431559       0       0
#> 66   0.8761423218  0.045291255       0       0
#> 67  -0.6887312194 -0.081408176       0       0
#> 68  -0.1349085887 -0.223568244       0       0
#> 69   0.1559137208  0.131669596       0       0
#> 70   0.3550476328  0.019541918       0       0
#> 71   0.3556854530  0.053655119       0       0
#> 72   0.0158177439 -0.071081508       0       0
#> 73  -0.2197224161  0.133446770       0       0
#> 74   0.1890340051  0.281618138       0       0
#> 75   0.1713105875  0.096711334       0       0
#> 76  -0.0003647564  0.068309798       0       0
#> 77   0.0962960348  0.029822278       0       0
#> 78   0.5984477925  0.116611286       0       0
#> 79  -0.0169358422 -0.052532272       0       0
#> 80   0.0365876127  0.058150430       0       0
#> 81  -0.3998391980 -0.035770081       0       0
#> 82  -0.0193618474 -0.063850234       0       0
#> 83   0.2476184086 -0.051575020       0       0
#> 84   0.3550994113  0.186818828       0       0
#> 85  -0.0487526725 -0.114417174       0       0
#> 86  -0.3860995268 -0.064467250       0       0
#> 87   0.0582398986  0.077255382       0       0
#> 88   0.1352811318  0.025540682       0       0
#> 89   0.4865161927 -0.116488475       0       0
#> 90  -0.2817944094 -0.049485198       0       0
#> 91  -0.4067290518  0.271176180       0       0
#> 92   0.0015260991 -0.045470060       0       0
#> 93   0.6154377930  0.198097599       0       0
#> 94  -0.5931124932  0.122131833       0       0
#> 95  -0.1992834471 -0.035534368       0       0
#> 96   0.2092659544  0.091355871       0       0
#> 97   0.2237864771  0.007060852       0       0
#> 98  -0.5160849502 -0.028270388       0       0
#> 99   0.0990012843 -0.026439441       0       0
#> 100 -0.1705526939  0.021802450       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.187  -0.0775 
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.200  -0.0567 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.509  -0.112  
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.0168  0.0452 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.187   0.0166 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0758 -0.126  
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.434  -0.0271 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.265  -0.141  
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.0817 -0.00259
#> 10     10     0.21      80    65    19.8     0.041      0.12  0.268   0.135  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.207  18.0 0.041  0.12 92.4  1665. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.207  18.0 0.041  0.12  7.36  133. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.304  18.4 0.041  0.12 84.2  1548. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.304  18.4 0.041  0.12  2.75   50.5
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.150  17.4 0.041  0.12 99.8  1737. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.150  17.4 0.041  0.12 14.3   250. 
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>