Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object

A posologyr list, created by the posologyr function.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = RxODE::RxODE({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc      ETA_Ka
#> 1   -0.214464327 -0.446957933  0.05915979
#> 2   -0.527568806  0.486935508 -0.59028016
#> 3    0.777945981 -0.054203901 -0.16055617
#> 4    0.706976334  0.490346155 -0.47642608
#> 5   -0.642517316  0.495048071  0.64363849
#> 6    0.211684437 -0.739354666  0.05385514
#> 7   -0.157035146  0.246894292  0.54499040
#> 8   -0.187328199  0.364740624  0.25860246
#> 9   -0.248964103  0.112158866 -0.56367848
#> 10   0.296174009 -0.435895467  0.06460930
#> 11  -0.002066311 -0.237225047 -0.06236761
#> 12   0.002511815 -0.199824088  0.08301982
#> 13  -0.798472734  0.069404502  0.69705486
#> 14   1.211417329 -0.053470134  0.42715094
#> 15   0.256491029  0.148091237 -0.35808681
#> 16   0.128330731 -0.317348000  0.97092004
#> 17  -0.344871572  0.352777145 -0.31983416
#> 18   0.416592168  0.171519147  0.30955507
#> 19   0.006891023 -0.405500274  0.43862768
#> 20   0.455347682  0.469651041 -0.45472724
#> 21   0.595373377 -0.144539236 -0.58372131
#> 22  -0.520018886 -0.111235028 -0.10132007
#> 23   0.235856680  0.572045118  0.05496113
#> 24   0.861352478  0.053155918  0.32880295
#> 25   0.437498939  0.179551581 -0.92123584
#> 26   0.512941258  0.405507874 -0.28568020
#> 27  -0.044533227  0.114103516  0.52627009
#> 28   0.532525406  0.323733662  0.74529081
#> 29   0.114290354  0.188872342 -0.25056337
#> 30  -0.319940964 -0.264936889  0.86927409
#> 31  -0.402341767  0.188221685 -0.95264808
#> 32   0.281220530  0.399478160 -0.21017404
#> 33   0.281173684  0.359122929  0.26958066
#> 34   0.673614584  0.107027065  0.98865376
#> 35  -0.635344949  0.351653672  0.04856801
#> 36   0.251915703 -0.160578391 -0.14692754
#> 37   0.091400831 -0.059184082 -0.18681167
#> 38  -0.018940160  0.445284896  0.22371080
#> 39   0.279373123  0.094899020 -0.46223133
#> 40   0.142630909  0.034660246  0.28337350
#> 41   0.071621078  0.162484973 -0.29839435
#> 42   0.794135373 -0.285884315 -0.05840766
#> 43  -0.455993711 -0.365282936 -0.48565651
#> 44  -0.273920136  0.277073903 -0.87713943
#> 45  -0.177005453  0.690006730 -0.52987552
#> 46  -0.152651345 -0.589398785 -0.75414679
#> 47   0.087895242 -0.657865851 -0.64719474
#> 48   0.292256040  0.035516882  0.28579495
#> 49  -1.206481310  0.179273495  0.43539924
#> 50  -1.220912040 -0.303119623  0.47224181
#> 51  -0.026571372  0.002559226 -0.43062038
#> 52  -0.016331409 -0.270799131 -0.22167304
#> 53   0.918400767  0.102659923  0.17940602
#> 54   0.008943154 -0.400466593  0.92253215
#> 55  -0.625731344  1.206408596  0.59380674
#> 56  -1.248417935  0.275752815 -0.31149308
#> 57  -0.055305328 -0.355091444  0.10819081
#> 58  -0.110110488 -0.337442920  0.18132943
#> 59   0.363640989  0.102413281 -0.54009870
#> 60   0.060546791  0.335767300 -0.37632093
#> 61   0.259970123  0.770211054  0.23245349
#> 62  -0.493728667  0.252696884  0.86285894
#> 63   0.155709433  0.013887796 -0.05179391
#> 64   0.099348492 -0.639653461  0.38296701
#> 65  -0.192399064 -0.912121415 -0.55319971
#> 66  -0.066613206 -0.237735622  0.20170324
#> 67  -0.393414299 -0.257313178 -0.71183708
#> 68  -0.466218963 -0.450223788  0.23664058
#> 69  -0.543431736 -0.202337098  0.24698612
#> 70   0.339544706  0.933962411  0.49406086
#> 71  -0.715742666  0.959722583  0.05508681
#> 72  -0.262480042 -0.403359300 -0.24594086
#> 73  -0.413319191 -0.149178108 -0.27910241
#> 74  -0.408778597  0.439077796 -0.38597259
#> 75   0.226797513  0.529302722  0.36630444
#> 76  -1.024683971 -1.356197393  0.17818241
#> 77   0.334130206 -0.275311662  0.35850889
#> 78   0.533043451 -0.201007974  0.08599592
#> 79  -0.791045163 -0.773630346  0.85626183
#> 80  -0.462530278  0.257617793 -0.19784882
#> 81   0.382586309 -0.780519959 -0.19473397
#> 82   0.099364380  0.055902422  0.56649278
#> 83   0.307588598 -0.866914967 -0.29313195
#> 84   0.428414412 -0.075233281 -0.76709047
#> 85  -0.685796320 -0.424006059  0.20107468
#> 86  -0.403125032  0.768482540  0.32901216
#> 87   0.190830505  0.057343416  0.03886849
#> 88  -0.878202577 -0.752508821 -0.07909878
#> 89  -0.806930104  0.281190369  0.58901512
#> 90   0.707942311  0.173055706 -0.06574685
#> 91  -0.397234086 -0.360491303  0.12632473
#> 92  -0.155868328 -0.068332829  0.17622806
#> 93  -0.017048067  0.773408464  1.03075975
#> 94  -0.014709255  0.216679203 -0.20068181
#> 95  -0.380040141 -0.479539699 -0.13922224
#> 96   0.442121745  0.026631041  0.04617817
#> 97  -0.073669217  0.032122848  0.06135895
#> 98   0.014403008  0.154381137  0.46576159
#> 99   0.145343526 -0.026179706  0.02104973
#> 100  0.047266427  0.116547646 -1.30374696
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>  <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1 -0.214 -0.447   0.0592
#>  2      2        4       70        1 -0.528  0.487  -0.590 
#>  3      3        4       70        1  0.778 -0.0542 -0.161 
#>  4      4        4       70        1  0.707  0.490  -0.476 
#>  5      5        4       70        1 -0.643  0.495   0.644 
#>  6      6        4       70        1  0.212 -0.739   0.0539
#>  7      7        4       70        1 -0.157  0.247   0.545 
#>  8      8        4       70        1 -0.187  0.365   0.259 
#>  9      9        4       70        1 -0.249  0.112  -0.564 
#> 10     10        4       70        1  0.296 -0.436   0.0646
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  3.23  44.8 1.06  0.0721 23.5     9.13e+2 1053.
#> 2      1    14     4    70     1  3.23  44.8 1.06  0.0721 17.8     9.35e-4  796.
#> 3      2     1     4    70     1  2.36 114.  0.554 0.0207  5.88    1.32e+3  670.
#> 4      2    14     4    70     1  2.36 114.  0.554 0.0207 13.7     9.84e-1 1562.
#> 5      3     1     4    70     1  8.71  66.3 0.852 0.131  13.4     1.06e+3  885.
#> 6      3    14     4    70     1  8.71  66.3 0.852 0.131   5.86    1.65e-2  389.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>