Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>            ETA_Cl       ETA_Vc      ETA_Ka
#> 1    0.1513439615  0.170439840 -0.38448579
#> 2    0.0034329324  0.280711069 -0.07838674
#> 3   -0.5469140653 -0.890079215 -0.49871856
#> 4    1.0186625970  0.606740964  0.70401426
#> 5    0.5257590752  0.066483819 -0.00480137
#> 6   -0.1346363601  0.353831248  0.03329770
#> 7   -0.5298512489 -0.295713649  0.22895437
#> 8   -0.7721781272 -0.505385397  0.40902638
#> 9    0.0844111487  0.348997806 -0.21190538
#> 10  -0.3085450304  0.025460777  0.30478139
#> 11  -0.7255466667 -0.219825281 -0.05991493
#> 12  -0.6198033860  0.117252881 -0.12301990
#> 13  -0.5932537067  0.137748339 -0.64592033
#> 14  -0.3306061659 -0.255628471 -0.12589111
#> 15   0.0979708486  0.255791801 -0.40007389
#> 16  -0.0948478683 -0.391625878 -0.30309497
#> 17   0.3758526616 -0.133536690  0.37796210
#> 18   0.2848546159  0.304851717  0.01032083
#> 19   0.4715045696  0.402340791 -0.15685745
#> 20   0.3179667441 -0.659975095 -0.10219923
#> 21  -0.6378213088  0.600954944 -0.58369483
#> 22  -0.3658464741  0.131167642  0.14837627
#> 23   0.5473631676 -0.034995584 -0.13548251
#> 24   0.0756283900  0.011787136  0.31485710
#> 25   0.1521384141  0.054321898 -0.45425796
#> 26  -0.6449478513  0.061094290 -0.05127289
#> 27  -0.2246867072  0.830888484 -0.13869445
#> 28   0.3603566134  0.497580952 -0.43264605
#> 29   0.5862968631 -0.471880859  0.32079636
#> 30   0.2545999951 -0.297347444 -0.61221367
#> 31  -0.3009252880 -0.562660719  0.61577386
#> 32  -0.0163709531  1.202480440 -0.46134199
#> 33  -0.3181388611  0.460613051 -1.09220922
#> 34   0.6942573352 -0.463638124 -0.35215564
#> 35  -0.1782411542 -0.222454082 -0.59555449
#> 36  -0.2357388064  0.669763814  0.25514341
#> 37  -0.0665535064 -0.464623347  0.52893873
#> 38  -0.1595293898 -0.466211136 -0.24275175
#> 39  -0.0277645736  0.754160650 -0.86815622
#> 40  -1.1128920792  1.438367146  0.08737397
#> 41  -0.3561483456 -0.414435864  0.35597643
#> 42   0.1628832098 -0.334200389 -0.07243068
#> 43  -0.3713133882 -0.361174106 -0.27494708
#> 44  -0.1061178528 -0.156984973  0.52417915
#> 45   0.8879237683 -0.206468735 -0.38979430
#> 46  -1.1030969697 -0.233233790  0.11939967
#> 47   0.2783466170 -0.244576682 -0.43676724
#> 48  -0.3317944815 -0.333334074 -0.05375465
#> 49   0.2103708785 -0.254048458  0.14533834
#> 50   0.4482637207  0.122624524  0.44492847
#> 51   0.8403896010 -0.237777291  0.16812194
#> 52  -0.1876124076 -0.234930446  0.86520616
#> 53  -0.4545005479  0.786540422 -0.30012847
#> 54  -0.5018849205  0.344982341  0.25193138
#> 55   0.4355687665 -0.685141285 -0.19449638
#> 56   0.1670684198 -0.369026490 -0.14870635
#> 57  -0.2558710106  0.346893373 -0.59012299
#> 58   0.2413166800 -0.259340605  0.30056777
#> 59  -0.1123359412 -0.731065841 -0.01472370
#> 60  -0.1959804325  0.424113069 -0.08716340
#> 61  -0.6829049505  0.008324778  0.46983739
#> 62  -0.2171146812 -0.042386788 -0.31356107
#> 63  -0.4095484431  0.484827262 -0.22650500
#> 64  -0.1499146849 -0.879866775  0.34852313
#> 65   0.1303542548  0.180439883  0.51328072
#> 66  -0.2517715984  0.057196604  0.26706492
#> 67  -0.8222323340 -0.389335874 -0.31986290
#> 68   0.0489797721  0.537171180  0.10760235
#> 69   0.4333742842 -0.243352492  0.23179238
#> 70   0.3445198064 -0.336259274 -0.66061765
#> 71  -0.3533213046  0.223926950 -0.23713156
#> 72   0.5671870452  0.073782694  0.42139844
#> 73   0.3986309626  0.184951714 -0.20306148
#> 74  -0.5825659218  0.474277475  0.67098485
#> 75  -0.5085016070  0.516796689 -0.53069889
#> 76  -0.2643280422  0.869509020 -0.19832605
#> 77  -0.0006992199 -0.111442224  0.05874151
#> 78   0.2399604838  0.937887238 -0.18092087
#> 79  -0.2868887008 -0.024210093  0.09324171
#> 80   0.1642403616 -0.024294277  0.06768569
#> 81  -0.1860123269  0.257753337 -0.41123138
#> 82  -0.1425528653 -0.640786893  0.06867645
#> 83  -0.6743700195  0.429774386 -0.66960010
#> 84   0.0481364065  0.114077760  1.38494234
#> 85   0.0806036260  0.797039288  0.18561092
#> 86  -0.0032475964  0.237296766 -0.23810747
#> 87  -0.1031380469 -0.016144264  1.51042112
#> 88  -0.0799067990 -0.644057757 -0.12342059
#> 89  -0.5310291340  0.218097563  0.40683127
#> 90   0.0381772962 -0.447844907 -0.30126337
#> 91   0.2989108472  0.735068999 -0.64854345
#> 92  -0.5396568276 -0.159777600  0.37991045
#> 93   0.3248342479  0.207853664  0.43335143
#> 94   0.5790226205 -0.110566862  0.17112095
#> 95  -0.3996550148 -0.071962939 -0.07133921
#> 96  -0.7724732211  0.309387133  0.07033402
#> 97  -0.1676819227 -0.955911015  0.08583296
#> 98  -0.4197468703 -0.511095222  0.43522874
#> 99  -0.0493842701 -0.112081220 -0.41382835
#> 100  0.4111924279 -0.403496604 -1.01117444
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka   ETA_Cl  ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>
#>  1      1        4       70        1  0.151    0.170  -0.384  
#>  2      2        4       70        1  0.00343  0.281  -0.0784 
#>  3      3        4       70        1 -0.547   -0.890  -0.499  
#>  4      4        4       70        1  1.02     0.607   0.704  
#>  5      5        4       70        1  0.526    0.0665 -0.00480
#>  6      6        4       70        1 -0.135    0.354   0.0333 
#>  7      7        4       70        1 -0.530   -0.296   0.229  
#>  8      8        4       70        1 -0.772   -0.505   0.409  
#>  9      9        4       70        1  0.0844   0.349  -0.212  
#> 10     10        4       70        1 -0.309    0.0255  0.305  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  4.65  83.0 0.681 0.0561  9.34 1206.       775.
#> 2      1    14     4    70     1  4.65  83.0 0.681 0.0561 12.1     0.173   1008.
#> 3      2     1     4    70     1  4.01  92.7 0.925 0.0433 10.5  1009.       973.
#> 4      2    14     4    70     1  4.01  92.7 0.925 0.0433 12.5     0.00608 1157.
#> 5      3     1     4    70     1  2.31  28.7 0.607 0.0805 24.5  1273.       703.
#> 6      3    14     4    70     1  2.31  28.7 0.607 0.0805 26.5     0.474    761.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>