Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)
#> $eta
#>          ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.09617919 -0.154255395       0       0
#> 2    0.24497492  0.009726151       0       0
#> 3    0.52712444 -0.066609441       0       0
#> 4    0.21562952 -0.206301562       0       0
#> 5   -0.39552497 -0.024635792       0       0
#> 6   -0.42310849 -0.002815486       0       0
#> 7   -0.03656118  0.062359561       0       0
#> 8   -0.21249359  0.184279262       0       0
#> 9    0.40205271  0.070704863       0       0
#> 10  -0.24435472 -0.099235309       0       0
#> 11   0.17102878 -0.033008082       0       0
#> 12   0.31533417 -0.050612273       0       0
#> 13   0.06789326 -0.044504453       0       0
#> 14   0.24822396 -0.079374048       0       0
#> 15   0.44164053  0.123414905       0       0
#> 16  -0.48305121 -0.168737624       0       0
#> 17   0.05739619 -0.015809290       0       0
#> 18  -0.08303550  0.084834367       0       0
#> 19   0.13767494 -0.003867043       0       0
#> 20  -0.35639022 -0.140531269       0       0
#> 21   0.14369579 -0.038486017       0       0
#> 22  -0.14534897  0.066161516       0       0
#> 23  -0.02533554 -0.094573322       0       0
#> 24   0.02352899 -0.155283152       0       0
#> 25   0.22085285  0.162404682       0       0
#> 26  -0.10885007  0.044094932       0       0
#> 27  -0.23883505  0.007383595       0       0
#> 28  -0.54404495 -0.105307495       0       0
#> 29  -0.18623723 -0.064586483       0       0
#> 30  -0.18834745  0.054396025       0       0
#> 31   0.36780327  0.096912212       0       0
#> 32   0.46022560 -0.063812852       0       0
#> 33   0.70595346 -0.002986226       0       0
#> 34   0.22084871 -0.049391243       0       0
#> 35  -0.02528544 -0.125536899       0       0
#> 36   0.06788597 -0.244133196       0       0
#> 37   0.08564596 -0.140270424       0       0
#> 38   0.17144722  0.124162297       0       0
#> 39  -0.23363477 -0.105413850       0       0
#> 40  -0.52535448  0.044217544       0       0
#> 41  -0.01291266  0.017200804       0       0
#> 42  -0.13316730  0.058697791       0       0
#> 43  -0.04074574  0.128093846       0       0
#> 44   0.11049385  0.224889065       0       0
#> 45   0.26284186 -0.146102488       0       0
#> 46  -0.25770770  0.130648326       0       0
#> 47  -0.36105031 -0.082503652       0       0
#> 48   0.08083572  0.062013071       0       0
#> 49   0.29010436  0.114609992       0       0
#> 50  -0.06979008 -0.001859437       0       0
#> 51   0.06118693  0.021653716       0       0
#> 52   0.18763637 -0.064857587       0       0
#> 53  -0.07483517  0.069856619       0       0
#> 54   0.18241762  0.189783479       0       0
#> 55   0.09519915  0.299828340       0       0
#> 56   0.28872798  0.261522298       0       0
#> 57   0.27464622 -0.080774440       0       0
#> 58  -0.09562491 -0.220134643       0       0
#> 59  -0.19960266 -0.141212430       0       0
#> 60  -0.68390432  0.001032125       0       0
#> 61   0.31711756 -0.085713340       0       0
#> 62  -0.46562049  0.130642541       0       0
#> 63   0.44664838  0.033682344       0       0
#> 64  -0.06562780  0.023931800       0       0
#> 65  -0.21436597  0.119276853       0       0
#> 66   0.14922250 -0.133167329       0       0
#> 67  -0.20654228  0.006230759       0       0
#> 68  -0.08304898  0.083639593       0       0
#> 69   0.16679265 -0.023146276       0       0
#> 70   0.34484560 -0.183409714       0       0
#> 71  -0.17577933  0.136673216       0       0
#> 72  -0.03493618  0.109794076       0       0
#> 73  -0.22606798 -0.095538927       0       0
#> 74  -0.20635920  0.022323025       0       0
#> 75   0.02935217  0.031682066       0       0
#> 76  -0.09067996  0.077367307       0       0
#> 77  -0.19595156 -0.025051975       0       0
#> 78   0.10349567  0.089717393       0       0
#> 79  -0.10073415  0.040989191       0       0
#> 80   0.10290830  0.178571252       0       0
#> 81  -0.31002866 -0.108517408       0       0
#> 82   0.28205315  0.039848324       0       0
#> 83  -0.14089047 -0.016880751       0       0
#> 84  -0.42172042  0.130155475       0       0
#> 85   0.28997455 -0.076440556       0       0
#> 86  -0.32944695 -0.139096944       0       0
#> 87  -0.48622515 -0.043552983       0       0
#> 88   0.06742811 -0.101653717       0       0
#> 89   0.07309992 -0.134774983       0       0
#> 90  -0.37833110  0.091771281       0       0
#> 91   0.35788144  0.145689762       0       0
#> 92  -0.37357130  0.198854160       0       0
#> 93  -0.11395939 -0.038782187       0       0
#> 94  -0.03214416 -0.147921949       0       0
#> 95   0.12814901  0.129171616       0       0
#> 96   0.12752549  0.085974845       0       0
#> 97   0.32851990 -0.103536566       0       0
#> 98   0.55624113  0.171319305       0       0
#> 99   0.35479020 -0.034546515       0       0
#> 100 -0.54627654 -0.077717476       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.0962 -0.154  
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.245   0.00973
#>  3      3     0.21      80    65    19.8     0.041      0.12  0.527  -0.0666 
#>  4      4     0.21      80    65    19.8     0.041      0.12  0.216  -0.206  
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.396  -0.0246 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.423  -0.00282
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0366  0.0624 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.212   0.184  
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.402   0.0707 
#> 10     10     0.21      80    65    19.8     0.041      0.12 -0.244  -0.0992 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.226  16.7 0.041  0.12 98.3  1640. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.226  16.7 0.041  0.12  6.46  108. 
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.318  19.7 0.041  0.12 77.9  1532. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.318  19.7 0.041  0.12  2.27   44.6
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.422  18.2 0.041  0.12 77.9  1419. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.422  18.2 0.041  0.12  1.10   20.0
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>