Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1    0.505346020 -0.625821776 -0.280056363
#> 2    0.304213510  0.879235364  0.004524492
#> 3   -0.046585294 -0.052921446  0.297308206
#> 4    0.202384465  0.254921382 -0.123010488
#> 5   -0.387309164  0.087430217  0.356277430
#> 6   -0.085832608 -0.043711307  0.473225287
#> 7   -0.464199024 -0.141938130  0.482580097
#> 8   -0.123639414 -0.119277159  0.285005639
#> 9   -0.034662057 -0.360844599 -0.951792739
#> 10   0.399117058 -0.126433890  0.439490300
#> 11   1.069819898  0.247043026 -0.205148131
#> 12   0.299724427  0.681185081 -0.012252132
#> 13  -0.566507673 -0.420471976 -0.563932440
#> 14   0.491551664  0.020433425 -0.162170484
#> 15  -0.515994663 -0.663627713  0.289615555
#> 16   0.033624093 -0.993579453  0.254559254
#> 17   0.412363899  0.663118256 -0.446466241
#> 18   0.048990719 -0.197452351  0.507524933
#> 19   1.081134770  0.493525286 -0.516283884
#> 20   0.010421516 -0.685475194  1.069345106
#> 21   0.105250152 -1.231555102  0.461189930
#> 22   0.103936799  0.058623840 -0.626592072
#> 23  -0.053284712  0.157176725 -0.577871960
#> 24  -0.450999727 -0.757526891 -0.409351508
#> 25  -0.206893537  0.164817549  0.171547549
#> 26   0.034320526 -0.150337878 -0.148092910
#> 27   0.049575434 -0.327939968 -0.216474431
#> 28   0.703943247 -0.625784902 -0.475954434
#> 29  -0.789703019 -0.794604246 -0.911112775
#> 30   0.341480565  0.044990016 -0.016198011
#> 31   0.342721757 -0.042099574 -0.776061041
#> 32   0.162295019  0.578867315 -0.192217268
#> 33  -0.214846837 -0.395339542  0.031760065
#> 34   0.119898414 -0.105266121 -0.832759673
#> 35   0.157633021  1.073185728  0.332835065
#> 36   0.168210918  0.186368093  0.028343538
#> 37  -0.192002781 -0.523290107  0.419132131
#> 38  -0.160309802  1.107434763 -0.309904382
#> 39  -0.340198906 -0.403623614 -0.139456529
#> 40   0.133206758  0.122712711  0.641363169
#> 41  -0.691257709 -0.113880955 -0.177435542
#> 42  -0.533356587 -0.322766620  0.148596703
#> 43   0.459566125 -0.160991885 -0.261052127
#> 44  -0.759770614  0.511448863 -0.131993505
#> 45   0.139765123  0.218550082 -0.757991600
#> 46  -0.368822666  0.498258560 -0.152771433
#> 47   0.326764052 -0.258660511  0.130577476
#> 48  -0.053676684 -0.480272372  0.385943561
#> 49   0.133201761  0.016113931  0.629225821
#> 50  -0.367333714 -0.184948152  0.250652424
#> 51  -0.228460122  0.293941756  0.198523560
#> 52   0.463056731 -0.217130423 -0.573938177
#> 53   0.047130070 -0.181563592 -0.008247740
#> 54   1.127291325  0.331882622 -0.309910198
#> 55  -0.393590435  0.580371206 -0.232764430
#> 56   0.011691850 -0.798811753 -0.260568763
#> 57  -0.111418780 -0.069040988  1.199094126
#> 58   0.506194961  0.125021911 -0.126708131
#> 59   0.364384171  0.057618294 -0.226699490
#> 60   0.072510283  0.636894725 -0.373646006
#> 61   0.053793104 -0.103578490 -0.157380627
#> 62   0.405093694  0.735077692 -0.244642133
#> 63  -0.297302510 -0.204867712 -0.019222066
#> 64  -0.143560858 -0.582010394  0.414040687
#> 65   0.554000079 -0.363556420  0.929450313
#> 66   0.067166521 -0.132932213  0.152798539
#> 67  -0.451758215  0.271307963 -0.161478441
#> 68  -0.365360038  0.115153547 -0.130123490
#> 69   0.370021791 -0.007847479 -0.183912661
#> 70   0.296605996 -0.493131259 -0.332869238
#> 71   0.313163279  0.411103215 -0.453488915
#> 72  -0.496343986  1.237353648  0.151083943
#> 73  -0.037287225 -0.372499160  0.158401093
#> 74  -0.303743054  0.482472403 -0.230705883
#> 75   0.682587219 -0.356142690  0.258922448
#> 76   0.057796162  0.344154102  0.330119071
#> 77  -0.282367845 -0.170609450 -0.338938817
#> 78  -0.736869116 -0.408973164  0.351462904
#> 79  -0.282772511  0.143776478  0.429607253
#> 80  -0.639336686  0.212441603  0.199576195
#> 81  -0.191787527 -0.431438102 -0.540205169
#> 82  -0.241336060  0.262287607  0.015976826
#> 83  -0.238178794 -0.223354877  0.403521297
#> 84   0.175853247 -0.875458177 -0.908754170
#> 85   0.245305837 -0.204194455  0.413112812
#> 86  -0.193284451  0.411880706 -0.259957755
#> 87  -0.595418114 -0.564696102 -0.179374048
#> 88   0.195290834  0.246724734  0.716837470
#> 89   0.860382063 -0.196475858 -0.420384544
#> 90  -0.063112222 -0.302103266 -0.496159115
#> 91  -0.209820798  0.651155990  0.676921094
#> 92  -0.288690120 -0.584078461  0.038489864
#> 93   0.707764697 -0.003419764  0.458427520
#> 94  -0.572053968 -0.109984168 -0.291179572
#> 95  -0.485118317 -0.141948675 -0.121857352
#> 96  -0.056636821  0.720781275 -0.507401637
#> 97  -0.347055578  0.576723510  0.157644808
#> 98   0.468899120  0.069625870  0.394624240
#> 99   0.319402444 -0.257603914  0.193858351
#> 100 -0.008964651 -0.207799812  0.372659905
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1      1        4       70        1  0.505  -0.626  -0.280  
#>  2      2        4       70        1  0.304   0.879   0.00452
#>  3      3        4       70        1 -0.0466 -0.0529  0.297  
#>  4      4        4       70        1  0.202   0.255  -0.123  
#>  5      5        4       70        1 -0.387   0.0874  0.356  
#>  6      6        4       70        1 -0.0858 -0.0437  0.473  
#>  7      7        4       70        1 -0.464  -0.142   0.483  
#>  8      8        4       70        1 -0.124  -0.119   0.285  
#>  9      9        4       70        1 -0.0347 -0.361  -0.952  
#> 10     10        4       70        1  0.399  -0.126   0.439  
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  6.63  37.4 0.756 0.177  21.3     1.14e+3  797.
#> 2      1    14     4    70     1  6.63  37.4 0.756 0.177   6.11    6.18e-2  229.
#> 3      2     1     4    70     1  5.42 169.  1.00  0.0322  6.13    9.51e+2 1034.
#> 4      2    14     4    70     1  5.42 169.  1.00  0.0322  7.87    2.03e-3 1328.
#> 5      3     1     4    70     1  3.82  66.4 1.35  0.0575 18.5     7.43e+2 1225.
#> 6      3    14     4    70     1  3.82  66.4 1.35  0.0575 14.3     1.86e-5  948.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>