Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

Usage

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, $model an rxode2 model using the estimated ETAs, $event the data.table used to solve the returned rxode2 model.

Examples

# model
mod_run001 <- function() {
  ini({
    THETA_Cl <- 4.0
    THETA_Vc <- 70.0
    THETA_Ka <- 1.0
    ETA_Cl ~ 0.2
    ETA_Vc ~ 0.2
    ETA_Ka ~ 0.2
    prop.sd <- sqrt(0.05)
  })
  model({
    TVCl <- THETA_Cl
    TVVc <- THETA_Vc
    TVKa <- THETA_Ka

    Cl <- TVCl*exp(ETA_Cl)
    Vc <- TVVc*exp(ETA_Vc)
    Ka <- TVKa*exp(ETA_Ka)

    K20 <- Cl/Vc
    Cc <- centr/Vc

    d/dt(depot) = -Ka*depot
    d/dt(centr) = Ka*depot - K20*centr
    Cc ~ prop(prop.sd)
  })
}
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#>  
#>  
#>  
#>  
#> $eta
#>          ETA_Cl        ETA_Vc       ETA_Ka
#> 1    0.12788489 -0.8474363696  0.972933500
#> 2    0.15044931 -0.6165957725 -0.288257826
#> 3   -0.26290512  0.5912277997 -1.384240405
#> 4    0.06358117 -1.0524378667  0.292384536
#> 5   -0.04738660  0.3020810818  0.017593578
#> 6    0.08829972  1.0646795643  0.177153842
#> 7   -0.56754035 -0.3004684498  0.086744472
#> 8   -0.17353057  0.0297399528 -0.239102222
#> 9   -0.37204606 -0.0681723758  0.667158179
#> 10   0.13349896  0.0284245681  0.181822898
#> 11  -0.12622407  0.6860145096 -0.234266909
#> 12  -0.24704093 -0.0306282674 -0.549257201
#> 13  -0.16917794  0.4382400402 -0.012830985
#> 14   0.19623634  0.7074781593 -0.136171803
#> 15   0.28794308 -0.0661337956 -0.073951692
#> 16  -0.10032863 -0.0136452244 -0.871130218
#> 17  -0.37228285 -0.0663627695 -0.336515849
#> 18   0.50185856 -0.1880645348  0.398147597
#> 19   0.16715147  0.7078051051 -0.294626263
#> 20   0.24462343 -0.0151938765 -0.366482185
#> 21  -0.78361129 -0.4687277771  0.172885315
#> 22  -0.14383402  0.6361751832 -0.223703774
#> 23  -0.14275268 -0.4155070297  0.001112369
#> 24   0.07919262 -0.0923541265 -1.028548245
#> 25   0.32541141 -0.3991711915  0.158204676
#> 26  -0.41127178  0.2498722022  0.534857269
#> 27  -0.24468315  0.1749402395  0.541677063
#> 28  -0.20073218  0.1098069426  0.994336524
#> 29  -0.28347955 -0.4408523470  0.175145338
#> 30   0.06856882  0.0635693077  0.328021625
#> 31   0.72545211 -0.3352657120 -0.722663525
#> 32   0.25653428  0.0234340351 -0.406303331
#> 33   0.37959016  0.3802766057 -0.076283200
#> 34  -0.31170568  0.7406108899 -0.065524473
#> 35   0.05754786  0.7310744062 -0.549295961
#> 36   0.08559765  0.1137295326 -0.326162522
#> 37   0.12743126  0.4451818055 -0.237158833
#> 38  -0.06270716 -0.8739258173  0.548334921
#> 39  -0.55583095 -0.3039715802  0.445685750
#> 40  -0.12034899 -0.2824848002  0.556723874
#> 41   0.41352429 -0.3565260587 -0.343045638
#> 42  -0.25671247  0.4915157307  0.366315463
#> 43  -0.01111643 -0.4709172790 -0.087432011
#> 44  -0.38215972 -0.7138946687 -0.677261388
#> 45  -0.76348667  0.1181884539 -0.546778736
#> 46  -0.21289274  0.2845195196  0.837286700
#> 47   0.12095204  0.1063290947  1.054290892
#> 48  -0.65677056  0.6579158907  0.208758285
#> 49  -0.01757885 -0.0588808083  0.371926925
#> 50  -0.24673067 -0.1805240163 -0.097830097
#> 51   0.39191046  0.4159538681 -0.037967541
#> 52  -0.87001416 -0.0101960308 -0.244342202
#> 53   0.69353025  0.3160638259  0.007746790
#> 54   0.76718629 -0.5305894330  0.687736930
#> 55  -0.15600010 -0.5372649408  0.233664676
#> 56  -0.22696478  0.3977303316 -0.325538825
#> 57   0.34939402  0.3356843830 -0.460650352
#> 58  -0.33446585 -0.4592726994 -0.762361942
#> 59   0.56771730 -0.0018453990  0.164049535
#> 60   0.02221040 -0.2074472251 -0.245654268
#> 61  -0.43756583  0.4004970612 -0.323059623
#> 62   0.22864983 -0.1643911881 -0.385340511
#> 63   0.04960210 -0.9753202418 -0.156391704
#> 64   0.41269427  0.9290324438  0.488513823
#> 65   0.48279891  0.5271998013 -0.145369898
#> 66  -0.01509115 -0.3794736224  0.633903308
#> 67   0.07928208  0.2879443073 -0.147487550
#> 68   0.20664080  0.0326334612  0.362415054
#> 69   0.06363990  0.8187491113  0.398934107
#> 70  -0.06412535  0.4812014182 -0.191628561
#> 71   0.44473418  0.3355002381 -0.190077991
#> 72  -0.35786214 -0.6259152421  0.309481419
#> 73   0.39099910 -0.8330321749  0.258486502
#> 74   0.60591300  0.7522020684  0.003734992
#> 75   0.14105591 -0.0772794560  0.199033893
#> 76   0.11255801  0.0686682135 -0.014146813
#> 77   0.29403252 -0.7716550381 -0.377551574
#> 78  -0.10415015 -0.4098232879 -0.277269197
#> 79   0.04687274  0.3381934530  0.466881915
#> 80  -0.01657469 -0.0004508222 -0.274059996
#> 81   0.34703845 -0.1353410256 -0.518273028
#> 82  -0.77759568  0.0783792867 -0.009475904
#> 83   0.09837731 -0.2966746333 -0.151063236
#> 84   0.86918648  0.5425733777 -0.484762337
#> 85  -0.20056939 -0.0337251850  0.880821557
#> 86   0.22423405  0.2266253611  0.098053962
#> 87  -0.06289958 -0.0805403600 -0.328896523
#> 88   0.91974567  1.2345015201  1.176161346
#> 89   0.87918576  0.2155617721  0.643564140
#> 90  -0.82927912 -0.3592766936 -0.700110707
#> 91  -0.34914053  0.0990080614 -0.229746377
#> 92  -0.22935254 -0.0841394013 -0.248344505
#> 93  -0.22056017  0.1123495210  0.058659532
#> 94  -0.24618351  0.1502172258 -0.043350296
#> 95   0.61466996  0.3262538319 -0.020061632
#> 96  -0.26323832 -0.1269916789  0.930418768
#> 97   0.73455148 -0.0915474373  0.805957437
#> 98  -0.03921193  0.1359026367  0.274076390
#> 99   0.14639569 -0.1606572049 -0.790413751
#> 100  0.08872456  0.5419509531 -0.251199157
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 8
#>    sim.id THETA_Cl THETA_Vc THETA_Ka prop.sd  ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1   0.224  0.128  -0.847   0.973 
#>  2      2        4       70        1   0.224  0.150  -0.617  -0.288 
#>  3      3        4       70        1   0.224 -0.263   0.591  -1.38  
#>  4      4        4       70        1   0.224  0.0636 -1.05    0.292 
#>  5      5        4       70        1   0.224 -0.0474  0.302   0.0176
#>  6      6        4       70        1   0.224  0.0883  1.06    0.177 
#>  7      7        4       70        1   0.224 -0.568  -0.300   0.0867
#>  8      8        4       70        1   0.224 -0.174   0.0297 -0.239 
#>  9      9        4       70        1   0.224 -0.372  -0.0682  0.667 
#> 10     10        4       70        1   0.224  0.133   0.0284  0.182 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 15,100 × 14
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka   K20  rxCc    Cc depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#> 1      1   0       4    70     1  4.55  30.0  2.65 0.152  0     0       0    0  
#> 2      1   0.1     4    70     1  4.55  30.0  2.65 0.152  1.61  1.61  351.  48.3
#> 3      1   0.2     4    70     1  4.55  30.0  2.65 0.152  5.90  5.90  621. 177. 
#> 4      1   0.3     4    70     1  4.55  30.0  2.65 0.152 12.2  12.2   828. 366. 
#> 5      1   0.4     4    70     1  4.55  30.0  2.65 0.152 20.0  20.0   987. 600. 
#> 6      1   0.5     4    70     1  4.55  30.0  2.65 0.152 28.9  28.9  1109. 867. 
#> # ℹ 15,094 more rows
#> # ℹ 1 more variable: AUC <dbl>
#> 
#> $event
#> ── EventTable with 152 records ──
#> 1 dosing records (see $get.dosing(); add with add.dosing or et)
#> 151 observation times (see $get.sampling(); add with add.sampling or et)
#> ── First part of : ──
#> # A tibble: 152 × 5
#>       id  time   amt evid          dur       
#>    <int> <dbl> <dbl> <evid>        <rate/dur>
#>  1     1   0      NA 0:Observation NA        
#>  2     1   0    2000 1:Dose (Add)   0.5      
#>  3     1   0.1    NA 0:Observation NA        
#>  4     1   0.2    NA 0:Observation NA        
#>  5     1   0.3    NA 0:Observation NA        
#>  6     1   0.4    NA 0:Observation NA        
#>  7     1   0.5    NA 0:Observation NA        
#>  8     1   0.6    NA 0:Observation NA        
#>  9     1   0.7    NA 0:Observation NA        
#> 10     1   0.8    NA 0:Observation NA        
#> # ℹ 142 more rows
#>