Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl        ETA_Vc       ETA_Ka
#> 1    0.114174341 -0.1880565753  0.003755305
#> 2   -0.468654185 -0.1650081274 -0.048307652
#> 3    0.124171287  0.5443422175  0.380923828
#> 4   -0.263687445  0.1336741097  0.017868552
#> 5   -0.436656973 -0.3962395430 -0.388147494
#> 6   -0.704036618  0.7004220814  0.630893206
#> 7    0.128233421  1.0901681621  0.267496331
#> 8   -0.029013934 -0.0845881604  0.348386486
#> 9    0.051852605  0.2553888796  0.603738181
#> 10   0.183824127  0.0400282993 -0.430968831
#> 11   0.611973878 -0.0731418156  1.015913681
#> 12  -0.320713800  0.1080102488  0.231533119
#> 13  -0.529441148  0.2155538646 -0.108053500
#> 14   1.056556987  0.3779287652 -0.401767070
#> 15   0.480011622  0.1694421801  0.098174551
#> 16  -0.647125240 -0.1374110676 -0.163859259
#> 17   0.813676528  0.2537627960 -0.208646684
#> 18   0.669472793 -0.2267276603  0.008945843
#> 19   0.209680202 -0.4299640046 -0.080147642
#> 20  -0.314747083 -0.0279226338 -1.045664634
#> 21  -0.431063714  0.3893657728 -0.180609479
#> 22   0.156913091 -0.2627413622  0.364311226
#> 23  -0.270675686 -0.1374730820 -0.822507068
#> 24   0.014524844  0.3877179017 -0.400738099
#> 25  -0.870286221 -0.8804081757 -0.780769044
#> 26  -0.467053566  0.0425861260 -0.067819973
#> 27  -0.054806081  0.2908767270  0.003769247
#> 28   0.310824793 -0.5259410414 -0.261671853
#> 29   0.061774266 -0.0706192120 -0.553180559
#> 30   0.297606250 -0.7293841325 -0.029778127
#> 31   0.402214136  0.4115925288 -0.267667890
#> 32   0.369239849  0.6447114104 -0.071122775
#> 33   0.044447645  0.3571235635 -0.266565742
#> 34  -0.491904798 -0.2185419968  0.121175899
#> 35   0.694723271  0.0156580142 -0.487334386
#> 36   0.455883033 -0.0145170789 -0.011406821
#> 37  -0.375762880  0.6271749830 -0.534720452
#> 38   0.089847762  0.7454299635  0.031973434
#> 39   0.439845014 -0.4438660036 -0.215772467
#> 40  -0.380209164  0.3794630942 -0.088342008
#> 41  -0.618272455  0.4838294460  0.030288518
#> 42  -0.483235024 -0.7141516091  0.316244180
#> 43   0.632294567 -0.5081940631  0.464738398
#> 44   0.102942292  0.3519299834  0.246976683
#> 45   0.333765702 -0.3583955167 -0.417263888
#> 46  -0.736661215  0.0654944299 -0.246362338
#> 47  -0.163003099 -0.0002802555 -0.008863757
#> 48  -0.037576843 -0.4219828006  0.183745843
#> 49  -0.027273248 -0.0059495950 -0.302268562
#> 50   0.373446844 -0.0776013013 -0.668030870
#> 51   0.547461575 -0.3521258700 -0.784284586
#> 52  -0.402852660  1.0000671970  0.386808728
#> 53   0.514821847  0.5808225599 -0.487222013
#> 54   0.855030262  0.0864644972 -0.339104541
#> 55   0.285669576 -0.1795445468  0.719565301
#> 56  -0.229226986 -0.1443522209  0.203269469
#> 57   0.408718690  0.9445996103  0.900203085
#> 58   0.276943291 -0.0737319409 -0.094994499
#> 59  -0.507498905 -0.3816422883 -0.653168827
#> 60  -0.299032114  0.4360641639  1.045305063
#> 61   0.247551897  0.5285372220 -0.727332934
#> 62   0.057796095 -0.5455014755 -0.227292018
#> 63   0.452560721  0.4857433272 -0.299809721
#> 64   0.567419912 -0.2948475141  0.600337571
#> 65  -0.107965218  0.4263053767  0.088356337
#> 66  -0.020198654 -0.0346632624 -0.340938282
#> 67  -0.919387877  0.2211137793 -0.248731776
#> 68  -0.604041246  0.3716359811  0.080302843
#> 69   0.489701454  0.1779297865  0.175794366
#> 70   0.284585788  0.5276159049 -0.188693580
#> 71  -0.425954665 -0.3873767552  0.606140073
#> 72  -0.081274930 -0.0978498550 -0.143519196
#> 73  -0.273818695  0.1900034166 -0.190085368
#> 74  -0.067904063 -0.1099702641 -0.215438723
#> 75  -0.070501314  1.1619047068 -0.341202127
#> 76  -0.396163111  0.0890981139 -0.676855168
#> 77   0.172787799  0.0832798835 -0.202423342
#> 78  -0.296496151 -0.5214924652  0.054355612
#> 79  -0.464988047 -0.0732635873  0.020838050
#> 80   0.870745145  0.1651030208  0.511273214
#> 81  -0.070329077  0.0124320857  0.357219762
#> 82  -0.105855326  0.7503192867  0.110409878
#> 83   0.089445464 -0.3547191344  0.444360123
#> 84   0.018466955  0.0083654383 -0.524785865
#> 85  -0.071759397 -0.2757814073  0.407699606
#> 86   0.031034661 -0.1041058267 -0.007757751
#> 87   0.214749365  0.3127217830  0.140054955
#> 88  -0.533646945  0.8360999879  0.342969451
#> 89   0.079664590  0.0580322561 -1.078208838
#> 90   0.094034627 -0.3563465694  0.272836043
#> 91  -0.623153112  0.2557405303  0.278745030
#> 92  -0.149643731  0.5048252858 -0.326242798
#> 93  -0.432039516  0.6529012831 -0.692182328
#> 94  -0.042548171 -0.4634844620 -0.774414781
#> 95  -0.485119984  0.1381103739 -0.591609314
#> 96  -0.657404826 -0.4044548465  0.563563999
#> 97   0.623721417  0.1330361697 -0.314451900
#> 98   0.006299256  0.5233086090 -0.145571051
#> 99   0.488164358  0.1841559486  0.744546065
#> 100  0.222661084 -0.0331813740  0.409242155
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1      1        4       70        1  0.114  -0.188   0.00376
#>  2      2        4       70        1 -0.469  -0.165  -0.0483 
#>  3      3        4       70        1  0.124   0.544   0.381  
#>  4      4        4       70        1 -0.264   0.134   0.0179 
#>  5      5        4       70        1 -0.437  -0.396  -0.388  
#>  6      6        4       70        1 -0.704   0.700   0.631  
#>  7      7        4       70        1  0.128   1.09    0.267  
#>  8      8        4       70        1 -0.0290 -0.0846  0.348  
#>  9      9        4       70        1  0.0519  0.255   0.604  
#> 10     10        4       70        1  0.184   0.0400 -0.431  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  4.48  58.0 1.00  0.0773  17.5    9.52e+2 1014.
#> 2      1    14     4    70     1  4.48  58.0 1.00  0.0773  12.9    2.05e-3  749.
#> 3      2     1     4    70     1  2.50  59.4 0.953 0.0422  16.7    9.88e+2  994.
#> 4      2    14     4    70     1  2.50  59.4 0.953 0.0422  19.7    4.12e-3 1172.
#> 5      3     1     4    70     1  4.53 121.  1.46  0.0375  10.7    6.82e+2 1295.
#> 6      3    14     4    70     1  4.53 121.  1.46  0.0375  10.2    3.72e-6 1225.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>