Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

Usage

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, $model an rxode2 model using the estimated ETAs, $event the data.table used to solve the returned rxode2 model.

Examples

# model
mod_run001 <- function() {
  ini({
    THETA_Cl <- 4.0
    THETA_Vc <- 70.0
    THETA_Ka <- 1.0
    ETA_Cl ~ 0.2
    ETA_Vc ~ 0.2
    ETA_Ka ~ 0.2
    prop.sd <- sqrt(0.05)
  })
  model({
    TVCl <- THETA_Cl
    TVVc <- THETA_Vc
    TVKa <- THETA_Ka

    Cl <- TVCl*exp(ETA_Cl)
    Vc <- TVVc*exp(ETA_Vc)
    Ka <- TVKa*exp(ETA_Ka)

    K20 <- Cl/Vc
    Cc <- centr/Vc

    d/dt(depot) = -Ka*depot
    d/dt(centr) = Ka*depot - K20*centr
    Cc ~ prop(prop.sd)
  })
}
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#>  
#>  
#>  
#>  
#> $eta
#>            ETA_Cl       ETA_Vc      ETA_Ka
#> 1    1.509733e-02 -0.016186364 -0.52679899
#> 2   -5.065591e-01  0.159582056 -0.12083643
#> 3   -3.402884e-01 -0.169764753  0.62537101
#> 4   -1.556625e-02 -0.083026187 -0.40598895
#> 5   -4.225858e-01  0.388054331  0.76373040
#> 6    5.766502e-02  0.472789504  0.01560593
#> 7    1.272383e-01  0.479519960 -0.01932000
#> 8   -4.678293e-02  0.143768214 -0.16376177
#> 9    1.442827e-01 -0.210819602 -0.11415598
#> 10   1.373633e-02 -0.783946398 -0.14714587
#> 11  -6.077227e-01  0.310746294 -0.42885063
#> 12  -1.534199e-01  0.055884964 -0.34751056
#> 13   2.428810e-01 -0.338450231  0.14203658
#> 14  -4.800486e-01  0.408987673 -0.26153526
#> 15   4.765925e-02  0.335450664 -0.52468882
#> 16   1.105351e-01 -0.401941402 -0.05765038
#> 17   3.776152e-01 -0.140522589 -0.35108615
#> 18   2.810179e-01 -0.471003701 -0.13318097
#> 19   2.336511e-01  0.216193619  0.03164949
#> 20  -7.452898e-01  0.119738012  0.23306728
#> 21   7.650526e-02 -0.434502128 -0.04988265
#> 22  -2.007812e-01 -0.526480474 -0.17362858
#> 23  -5.092078e-01 -0.061090908 -0.17439658
#> 24   2.337833e-01 -0.010218500  0.34012453
#> 25  -3.547032e-01 -0.581185456  0.24508276
#> 26   2.959496e-01  0.195121715  0.26155660
#> 27  -7.395556e-02 -0.028523080 -0.27195278
#> 28   2.567295e-01 -0.165186564 -0.17145398
#> 29  -1.880491e-01  0.146760354 -0.78936597
#> 30  -4.386245e-01 -0.527549991  0.16738600
#> 31  -5.700471e-01  0.301636092 -0.16243644
#> 32   1.938019e-01  0.269485385 -0.18439931
#> 33   2.610015e-01 -0.040020795 -0.02163562
#> 34  -3.958267e-01 -0.441250063  0.96602276
#> 35   7.838248e-02 -0.483769430  0.19542081
#> 36  -5.022530e-01 -0.627425065  0.25488459
#> 37  -3.196284e-01  0.651172607 -0.13790850
#> 38   3.364246e-01  0.779906791  0.64544032
#> 39   2.436184e-01  0.245160056 -0.82777827
#> 40  -5.346926e-01  0.009676858  0.50180182
#> 41   8.009718e-01  0.503864539  0.22347708
#> 42   5.034496e-01  0.435619372 -0.06418450
#> 43   4.020605e-01  0.294447273 -0.09108545
#> 44  -3.438734e-02 -0.472621016  0.15122555
#> 45  -1.779681e-01 -0.310812946 -0.35791899
#> 46   8.133532e-02  0.902326045  0.27394667
#> 47   2.340458e-01  0.824833303  0.30876212
#> 48   1.007434e-01 -0.251093484 -0.34481015
#> 49  -2.598501e-01  0.631180458  0.71997301
#> 50   2.023043e-01 -0.194051882  0.18650263
#> 51   1.505199e-01  0.099820484 -0.16501937
#> 52   9.241291e-02 -0.415205479  0.13583375
#> 53  -8.863234e-01  0.815655132  0.52001258
#> 54   5.723886e-01  0.706054216  0.19429468
#> 55   1.485582e-01 -0.328449059 -0.57102991
#> 56   5.132540e-01 -0.061339112 -0.65184530
#> 57   6.244343e-01 -0.389777674 -0.32398450
#> 58   3.184719e-01  0.342900862  0.36099688
#> 59   4.555447e-02 -0.178798934 -0.19149676
#> 60  -3.826410e-01 -0.392417878  0.60643613
#> 61   2.214500e-01 -0.477350884  0.46744042
#> 62   6.189829e-01 -1.045731659  0.26309320
#> 63  -2.384927e-01  0.218179937 -0.02156128
#> 64   7.140586e-01  1.466480496  0.37082036
#> 65   4.729326e-01 -0.222973041 -0.63497526
#> 66   1.882671e-02 -0.120701693  0.38346047
#> 67   6.358933e-01 -0.272008317  0.05674449
#> 68   4.416018e-01 -0.106041060  0.21729501
#> 69  -8.509248e-01  0.514105036  0.41726888
#> 70  -1.672575e-01 -0.311480882 -0.20727406
#> 71  -7.834966e-02  0.752990336 -0.41392643
#> 72  -2.864462e-01  0.098062354  0.11429091
#> 73   8.666932e-01 -0.713095263 -0.35054982
#> 74  -7.509757e-01 -0.378566836  0.61839686
#> 75  -3.261728e-01 -0.317464254  0.42147870
#> 76   6.126478e-01 -0.740255177  0.53250207
#> 77  -2.018874e-01 -0.691384766 -0.18377623
#> 78   4.067081e-01  0.065882537  0.27262689
#> 79   1.834668e-01  0.599527805 -0.20296871
#> 80   1.917972e-01  0.236414328 -0.35693471
#> 81   2.725721e-01  1.414693131  1.00542110
#> 82  -8.184843e-05 -0.222826401  0.15659665
#> 83   3.276706e-01 -0.663560795  0.30690260
#> 84  -4.664156e-01 -0.299638370  0.17032271
#> 85  -2.121784e-01  1.057803427  0.05927526
#> 86  -6.178694e-01 -0.245640977  0.11130505
#> 87   3.157618e-01  0.450638415 -0.15717022
#> 88  -1.754226e-01  0.778997712 -0.34891351
#> 89  -5.971167e-01  0.117111697 -0.44168223
#> 90  -1.136389e+00 -1.205529113 -0.41517518
#> 91  -6.521712e-01  0.413118127 -0.44770183
#> 92   1.638941e-01  0.191758884 -0.31244355
#> 93   1.953781e-01  0.810612503  0.12871112
#> 94  -1.139908e-01  0.899697081 -0.83085976
#> 95   5.242689e-02  0.846107522 -0.02433113
#> 96   2.049053e-01 -0.224359547  0.53554546
#> 97  -5.278126e-02  0.046590127  0.18526015
#> 98   4.875914e-01  0.503152935 -0.02522384
#> 99  -1.559060e-02 -0.016935492  0.71474995
#> 100  7.081456e-02  0.315723149  0.48204204
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 8
#>    sim.id THETA_Cl THETA_Vc THETA_Ka prop.sd  ETA_Cl  ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1      1        4       70        1   0.224  0.0151 -0.0162 -0.527 
#>  2      2        4       70        1   0.224 -0.507   0.160  -0.121 
#>  3      3        4       70        1   0.224 -0.340  -0.170   0.625 
#>  4      4        4       70        1   0.224 -0.0156 -0.0830 -0.406 
#>  5      5        4       70        1   0.224 -0.423   0.388   0.764 
#>  6      6        4       70        1   0.224  0.0577  0.473   0.0156
#>  7      7        4       70        1   0.224  0.127   0.480  -0.0193
#>  8      8        4       70        1   0.224 -0.0468  0.144  -0.164 
#>  9      9        4       70        1   0.224  0.144  -0.211  -0.114 
#> 10     10        4       70        1   0.224  0.0137 -0.784  -0.147 
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 15,100 × 14
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20  rxCc    Cc depot
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <dbl>
#> 1      1   0       4    70     1  4.06  68.9 0.590 0.0590 0     0        0 
#> 2      1   0.1     4    70     1  4.06  68.9 0.590 0.0590 0.168 0.168  388.
#> 3      1   0.2     4    70     1  4.06  68.9 0.590 0.0590 0.657 0.657  755.
#> 4      1   0.3     4    70     1  4.06  68.9 0.590 0.0590 1.45  1.45  1100.
#> 5      1   0.4     4    70     1  4.06  68.9 0.590 0.0590 2.52  2.52  1425.
#> 6      1   0.5     4    70     1  4.06  68.9 0.590 0.0590 3.86  3.86  1732.
#> # ℹ 15,094 more rows
#> # ℹ 2 more variables: centr <dbl>, AUC <dbl>
#> 
#> $event
#> ── EventTable with 152 records ──
#> 1 dosing records (see $get.dosing(); add with add.dosing or et)
#> 151 observation times (see $get.sampling(); add with add.sampling or et)
#> ── First part of : ──
#> # A tibble: 152 × 5
#>       id  time   amt evid          dur       
#>    <int> <dbl> <dbl> <evid>        <rate/dur>
#>  1     1   0      NA 0:Observation NA        
#>  2     1   0    2000 1:Dose (Add)   0.5      
#>  3     1   0.1    NA 0:Observation NA        
#>  4     1   0.2    NA 0:Observation NA        
#>  5     1   0.3    NA 0:Observation NA        
#>  6     1   0.4    NA 0:Observation NA        
#>  7     1   0.5    NA 0:Observation NA        
#>  8     1   0.6    NA 0:Observation NA        
#>  9     1   0.7    NA 0:Observation NA        
#> 10     1   0.8    NA 0:Observation NA        
#> # ℹ 142 more rows
#>