Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object

A posologyr list, created by the posologyr function.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc       ETA_Ka
#> 1   -0.939138388 -0.487005523 -0.092696794
#> 2    0.057924426 -0.581906989  0.134103718
#> 3   -0.086908391  0.009428939  0.133161745
#> 4   -0.209150641 -0.190565618  0.355978159
#> 5   -0.327235857  0.401419297  0.324673915
#> 6   -0.425740891  0.317052671 -0.492720185
#> 7    0.027385231  0.054614354  0.293404354
#> 8    0.416770647  1.268300611 -0.251913342
#> 9    0.144473326 -0.203380657  0.399299232
#> 10   0.129717255 -0.045565608 -0.784547971
#> 11  -0.838973561 -0.359839919 -0.094414962
#> 12  -0.758920441 -0.441648760 -0.030553165
#> 13   0.180460240  0.299652389 -0.539329156
#> 14  -0.503582841  0.344511492  0.054960112
#> 15   0.377336417  0.180091621 -0.076530123
#> 16  -0.651012478  0.399012853 -0.345385033
#> 17  -0.820536167  0.227145341  0.070306315
#> 18   0.536585989 -0.486694992 -0.297072597
#> 19  -0.098503903  0.467221066 -0.221438413
#> 20  -0.139102435  0.045348227  0.592667710
#> 21  -0.218099703  0.214010415  0.423708979
#> 22   0.096527629  0.245114794  0.063082189
#> 23  -0.737180224 -0.129858577  0.070108865
#> 24  -0.249584550 -0.065204600  1.050451291
#> 25   0.668926551  0.256769520 -0.222328539
#> 26  -0.730133861  0.213049246  0.268669717
#> 27   0.033804271  0.076648065  0.130147350
#> 28  -0.490621759 -0.915632210  0.240872654
#> 29  -0.004471913 -0.116717771 -0.117058759
#> 30  -0.114023695  0.264686923  0.055389025
#> 31  -0.064417907  1.007980775  0.677244708
#> 32   0.577643317 -0.936485324  0.484528196
#> 33   0.967314428 -0.139699562  0.086494945
#> 34   0.039591504 -0.197562026 -0.136320219
#> 35   0.330794155  0.381485614 -0.828162578
#> 36   0.104082495  0.149613606  0.185564464
#> 37  -0.158164028  0.495447550 -0.263932270
#> 38   0.736954298  0.348707955 -0.605389076
#> 39  -1.019089297 -0.389559910 -0.074179200
#> 40  -0.180617191 -0.460867263 -0.610538898
#> 41   0.345371214 -0.131758707 -0.121067999
#> 42  -0.213283619  0.188623528  0.453796697
#> 43  -0.536577204  0.350609508  0.550358290
#> 44   0.508666584 -0.240138647  0.349258746
#> 45   0.238522873 -0.654232022  0.374744852
#> 46   0.213199111 -0.341265131  0.024635747
#> 47  -0.098246045  0.245869667  0.009687690
#> 48  -0.256973388 -0.096452454 -0.067980883
#> 49   0.372393882  0.345824648  0.268184514
#> 50  -0.431840947 -0.832383142 -0.927041099
#> 51  -0.046449696 -0.287471633 -0.423930925
#> 52  -0.095297486 -0.152668074  1.015998287
#> 53   0.697656137  0.533587801 -0.078860977
#> 54   0.177089208 -0.235385338 -0.029097125
#> 55  -0.054528385 -0.793058116  0.003181416
#> 56   0.208074821 -0.521503749  0.127021435
#> 57  -0.168165298  0.145437699  0.296212405
#> 58  -0.449390024 -0.895165167 -0.822561943
#> 59  -0.334297056 -0.603890858  0.180697714
#> 60   0.027266187  0.385760222  0.284167193
#> 61   0.336326498  0.424923905 -0.093800950
#> 62   0.110403502 -0.680927442 -0.314654863
#> 63   0.032518690  0.127142397 -0.713831487
#> 64   0.104934657 -0.641524307  0.890986213
#> 65  -0.037795568  0.338047457  0.132888704
#> 66  -0.197004969 -0.127031367 -0.066580803
#> 67  -0.675819830 -0.137295638  0.631674885
#> 68  -0.302085436  0.231740590 -0.200170946
#> 69  -1.324895114  0.378532700  0.475585501
#> 70  -0.070518569 -0.077131609 -0.229851788
#> 71  -0.176973060 -0.101215528 -0.087296804
#> 72  -0.168427977  1.246764044 -0.099025405
#> 73  -0.296784531  0.170147413 -0.229067441
#> 74  -0.148108384 -0.856234253  0.039391299
#> 75   0.149046746  0.098318706 -0.171605337
#> 76   0.804747650 -0.072010376  0.023413116
#> 77  -0.002591147 -0.036671344 -0.119752353
#> 78   0.007632986  0.028754925  0.847422528
#> 79  -0.177975026 -0.283498987  0.432516928
#> 80   0.112450793 -0.786321610 -0.473491417
#> 81   0.275522377  0.365282805 -0.461168376
#> 82  -0.544310111  0.504601831  0.042564609
#> 83  -0.022916538  0.002121141 -0.862723929
#> 84   0.005824090  0.355189774 -0.451268463
#> 85  -0.533811019  0.518785251  0.426219269
#> 86  -0.352828953 -0.024756506  0.458590499
#> 87  -0.166895610  0.232530849  0.109848769
#> 88  -0.802766751  0.065895373 -0.457785578
#> 89  -0.100328296  0.766674387  0.593990276
#> 90  -0.384164662 -0.126531995 -0.050380021
#> 91  -0.701192831  0.431512884  0.306409240
#> 92  -0.668536925  0.062987329 -0.046053135
#> 93  -0.393375366 -0.413328731 -0.124051079
#> 94   0.433473730  0.501865349 -0.133120523
#> 95  -0.148591324  0.116935256  0.265092708
#> 96   1.036633595 -0.029273022 -0.841014910
#> 97   0.587434571 -0.391758292 -0.285745763
#> 98  -0.095623267 -0.350972460  0.017476243
#> 99   0.137042605  0.049943564 -0.634710057
#> 100  0.911280289  0.723593127  0.332386851
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl   ETA_Vc  ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>    <dbl>   <dbl>
#>  1      1        4       70        1 -0.939  -0.487   -0.0927
#>  2      2        4       70        1  0.0579 -0.582    0.134 
#>  3      3        4       70        1 -0.0869  0.00943  0.133 
#>  4      4        4       70        1 -0.209  -0.191    0.356 
#>  5      5        4       70        1 -0.327   0.401    0.325 
#>  6      6        4       70        1 -0.426   0.317   -0.493 
#>  7      7        4       70        1  0.0274  0.0546   0.293 
#>  8      8        4       70        1  0.417   1.27    -0.252 
#>  9      9        4       70        1  0.144  -0.203    0.399 
#> 10     10        4       70        1  0.130  -0.0456  -0.785 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  1.56  43.0 0.911 0.0364  22.5    1.02e+3  966.
#> 2      1    14     4    70     1  1.56  43.0 0.911 0.0364  29.4    7.28e-3 1264.
#> 3      2     1     4    70     1  4.24  39.1 1.14  0.108   27.8    8.60e+2 1087.
#> 4      2    14     4    70     1  4.24  39.1 1.14  0.108   12.7    3.01e-4  498.
#> 5      3     1     4    70     1  3.67  70.7 1.14  0.0519  15.8    8.61e+2 1114.
#> 6      3    14     4    70     1  3.67  70.7 1.14  0.0519  14.5    3.05e-4 1026.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>