Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>           ETA_ke         ETA_V ETA_k12 ETA_k21
#> 1    0.037108553 -0.0752024286       0       0
#> 2   -0.031020492 -0.0228677533       0       0
#> 3    0.193168244  0.0272920131       0       0
#> 4   -0.621168055 -0.2106179622       0       0
#> 5    0.188576875  0.0727452552       0       0
#> 6   -0.015057271  0.0888136805       0       0
#> 7   -0.312400833 -0.1765543523       0       0
#> 8   -0.375556035  0.1287064821       0       0
#> 9    0.342176163 -0.4184368792       0       0
#> 10   0.156702793 -0.0098637580       0       0
#> 11   0.006082632  0.1448782135       0       0
#> 12   0.329566950 -0.0247166587       0       0
#> 13  -0.033300154 -0.0476983024       0       0
#> 14  -0.092893296  0.0368017924       0       0
#> 15  -0.061224002  0.1230643038       0       0
#> 16  -0.135941875  0.0930097043       0       0
#> 17  -0.115817843 -0.0743683847       0       0
#> 18   0.079930760  0.1343283964       0       0
#> 19   0.020677515 -0.1322933359       0       0
#> 20  -0.054590317  0.1618046422       0       0
#> 21   0.361232823 -0.1592931021       0       0
#> 22   0.107803714  0.1806224253       0       0
#> 23  -0.266762577 -0.0647440400       0       0
#> 24   0.433984515 -0.0763822630       0       0
#> 25   0.271252454 -0.0625607944       0       0
#> 26  -0.294030923  0.0574774832       0       0
#> 27  -0.404293960  0.0462627025       0       0
#> 28   0.083129177 -0.1199870955       0       0
#> 29   0.294815873  0.0713537074       0       0
#> 30   0.024605854 -0.1814497068       0       0
#> 31   0.003865701 -0.0011502576       0       0
#> 32   0.273090279 -0.0415009508       0       0
#> 33   0.266749927 -0.1381555878       0       0
#> 34  -0.630762293 -0.1016904699       0       0
#> 35  -0.044769151  0.1180194348       0       0
#> 36   0.042809638 -0.1685544650       0       0
#> 37  -0.077550980  0.2484949809       0       0
#> 38  -0.112769647 -0.0488894955       0       0
#> 39  -0.111275424  0.2448771221       0       0
#> 40   0.240081801  0.1649983659       0       0
#> 41  -0.402654032 -0.1310714424       0       0
#> 42   0.087685468  0.1602580987       0       0
#> 43  -0.203930509 -0.0676406758       0       0
#> 44  -0.085970202  0.0107710824       0       0
#> 45  -0.249984961 -0.1264916987       0       0
#> 46   0.178723824  0.0008891906       0       0
#> 47   0.019941973 -0.0382223760       0       0
#> 48  -0.581815303  0.1124787896       0       0
#> 49   0.282450351  0.0688597201       0       0
#> 50   0.086858801  0.0991451253       0       0
#> 51   0.102752156 -0.0999221684       0       0
#> 52  -0.247292103  0.0082200215       0       0
#> 53   0.245889090  0.0177279161       0       0
#> 54  -0.218198358 -0.0384591112       0       0
#> 55  -0.171108345  0.0394054247       0       0
#> 56  -0.373170160  0.0155306337       0       0
#> 57  -0.204453775 -0.0263991462       0       0
#> 58  -0.113767611 -0.0850780140       0       0
#> 59  -0.133623612 -0.0722273435       0       0
#> 60   0.110150238 -0.1761292365       0       0
#> 61  -0.001330133 -0.0186421533       0       0
#> 62  -0.286397278  0.0451476058       0       0
#> 63   0.071396771 -0.1590968927       0       0
#> 64  -0.043674747 -0.0253190161       0       0
#> 65  -0.334244962 -0.0018036533       0       0
#> 66   0.170459224  0.0400226783       0       0
#> 67   0.145328689 -0.0437438494       0       0
#> 68  -0.151780616  0.1112655738       0       0
#> 69   0.304604484  0.0092458761       0       0
#> 70   0.344989052 -0.1101828278       0       0
#> 71   0.268740320  0.0885148060       0       0
#> 72   0.616230078  0.0548066119       0       0
#> 73   0.222584103  0.1462123249       0       0
#> 74   0.142137646 -0.0477516064       0       0
#> 75   0.331306187  0.0053976153       0       0
#> 76  -0.470442277  0.1080198562       0       0
#> 77   0.128655337  0.0201263604       0       0
#> 78   0.215241108  0.0976086456       0       0
#> 79   0.006232624 -0.0099165380       0       0
#> 80   0.031353518 -0.1052028280       0       0
#> 81   0.045566473 -0.0730977323       0       0
#> 82   0.491344614  0.0230740944       0       0
#> 83  -0.347984970 -0.2146544378       0       0
#> 84   0.321549427 -0.0817408221       0       0
#> 85   0.240474609 -0.0187042744       0       0
#> 86   0.224094700 -0.0023904148       0       0
#> 87   0.073104561 -0.0291808770       0       0
#> 88   0.167741839  0.0099577066       0       0
#> 89   0.171243130 -0.0777699825       0       0
#> 90  -0.055080539 -0.0980097900       0       0
#> 91   0.138180080 -0.0940682341       0       0
#> 92   0.018358280  0.1090982310       0       0
#> 93  -0.312830250 -0.1098658247       0       0
#> 94  -0.582822118  0.0241017816       0       0
#> 95  -0.714219762  0.0838882234       0       0
#> 96  -0.102914992 -0.1355351643       0       0
#> 97  -0.122514965 -0.0255954120       0       0
#> 98  -0.231009201  0.0149775781       0       0
#> 99   0.332597265 -0.2065939347       0       0
#> 100  0.151333816 -0.1503729561       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12  0.0371 -0.0752 
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.0310 -0.0229 
#>  3      3     0.21      80    65    19.8     0.041      0.12  0.193   0.0273 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.621  -0.211  
#>  5      5     0.21      80    65    19.8     0.041      0.12  0.189   0.0727 
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0151  0.0888 
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.312  -0.177  
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.376   0.129  
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.342  -0.418  
#> 10     10     0.21      80    65    19.8     0.041      0.12  0.157  -0.00986
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.258  18.1 0.041  0.12 88.7  1602. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.258  18.1 0.041  0.12  4.30   77.7
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.241  19.0 0.041  0.12 85.3  1622. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.241  19.0 0.041  0.12  4.84   92.2
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.302  20.0 0.041  0.12 77.5  1550. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.302  20.0 0.041  0.12  2.57   51.4
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>