Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>           ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1    0.081475592  0.021645524       0       0
#> 2    0.093446970  0.087244728       0       0
#> 3   -0.564848923 -0.177985694       0       0
#> 4   -0.182944145  0.139526674       0       0
#> 5   -0.162571739 -0.015036469       0       0
#> 6    0.389513301 -0.012534577       0       0
#> 7   -0.073839447 -0.075057159       0       0
#> 8   -0.183427067  0.018431194       0       0
#> 9    0.245847461  0.390111376       0       0
#> 10   0.050262778  0.064159784       0       0
#> 11  -0.446656316 -0.002972407       0       0
#> 12   0.386759763 -0.319906574       0       0
#> 13   0.486140701  0.144783846       0       0
#> 14   0.427035470 -0.072063684       0       0
#> 15   0.190252328 -0.145989681       0       0
#> 16   0.347352367  0.083954725       0       0
#> 17  -0.046537492  0.110512344       0       0
#> 18  -1.006616334  0.051417482       0       0
#> 19   0.138662773 -0.109175025       0       0
#> 20  -0.076852360 -0.156400158       0       0
#> 21  -0.388361954  0.045328543       0       0
#> 22   0.300173690 -0.129002105       0       0
#> 23  -0.493072674 -0.198246478       0       0
#> 24  -0.285232212 -0.017916939       0       0
#> 25  -0.379900792  0.013472833       0       0
#> 26   0.298018846  0.001240399       0       0
#> 27   0.234821883 -0.015330376       0       0
#> 28   0.058347792  0.144597312       0       0
#> 29  -0.186889615  0.052319274       0       0
#> 30   0.331406019 -0.113071460       0       0
#> 31  -0.893942276 -0.044313588       0       0
#> 32   0.205031089 -0.143102118       0       0
#> 33   0.273492107  0.023140758       0       0
#> 34  -0.203744785  0.084550381       0       0
#> 35  -0.110581763  0.035266826       0       0
#> 36  -0.477840757  0.061521647       0       0
#> 37  -0.064557494 -0.143493638       0       0
#> 38  -0.130692265 -0.159858091       0       0
#> 39  -0.027832735  0.085734580       0       0
#> 40   0.115858279 -0.100083089       0       0
#> 41   0.058730254 -0.016816115       0       0
#> 42   0.274537516 -0.096168661       0       0
#> 43  -0.326950520 -0.052099825       0       0
#> 44  -0.053790287 -0.130112261       0       0
#> 45  -0.286576166 -0.067391556       0       0
#> 46   0.136974570 -0.043966167       0       0
#> 47   0.089783035 -0.090601999       0       0
#> 48   0.042240233 -0.077113884       0       0
#> 49  -0.280531604 -0.156705376       0       0
#> 50   0.144255586 -0.016924176       0       0
#> 51   0.137249231  0.113242966       0       0
#> 52  -0.328577539  0.068675448       0       0
#> 53   0.389385275 -0.115839053       0       0
#> 54  -0.059896066  0.036117624       0       0
#> 55   0.205411004  0.127835756       0       0
#> 56  -0.080836502  0.147060729       0       0
#> 57  -0.575715033 -0.132386568       0       0
#> 58   0.529496755 -0.156823465       0       0
#> 59   0.030086652 -0.141324392       0       0
#> 60  -0.532951545 -0.277340686       0       0
#> 61  -0.112691192  0.170002620       0       0
#> 62  -0.526962218 -0.030655145       0       0
#> 63   0.217233409  0.003702696       0       0
#> 64  -0.010285030 -0.129200070       0       0
#> 65  -0.321874384  0.039138292       0       0
#> 66  -0.076781069 -0.083457423       0       0
#> 67  -0.107870774  0.153375347       0       0
#> 68  -0.009890999 -0.020362585       0       0
#> 69  -0.257970761 -0.103641268       0       0
#> 70   0.246574864  0.187308677       0       0
#> 71   0.036641116  0.115953975       0       0
#> 72   0.009916217  0.031205821       0       0
#> 73   0.304693337 -0.004738326       0       0
#> 74  -0.029726491  0.035259869       0       0
#> 75  -0.104056400  0.035386058       0       0
#> 76  -0.133957569 -0.027997322       0       0
#> 77   0.008728244 -0.192266750       0       0
#> 78  -0.093498435 -0.149047016       0       0
#> 79   0.197452313 -0.105177748       0       0
#> 80  -0.097485016  0.013706065       0       0
#> 81  -0.220812813  0.059567773       0       0
#> 82  -0.215055762  0.034835177       0       0
#> 83  -0.305029257  0.100306259       0       0
#> 84  -0.166182968  0.011688668       0       0
#> 85   0.213149797 -0.128682540       0       0
#> 86   0.070235481 -0.098578126       0       0
#> 87  -0.036631818  0.092612006       0       0
#> 88  -0.089289915 -0.086105623       0       0
#> 89   0.178562471 -0.115515998       0       0
#> 90  -0.084624956  0.057304098       0       0
#> 91   0.137372290  0.007762195       0       0
#> 92  -0.473567043  0.029366342       0       0
#> 93   0.148094035  0.018763294       0       0
#> 94  -0.276088410 -0.012233969       0       0
#> 95  -0.127579056 -0.129121953       0       0
#> 96  -0.110325899 -0.124885741       0       0
#> 97  -0.038817972 -0.042771628       0       0
#> 98   0.148549026 -0.002506138       0       0
#> 99   0.216119632 -0.086992712       0       0
#> 100 -0.369292940  0.060107765       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12  0.0815  0.0216
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.0934  0.0872
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.565  -0.178 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.183   0.140 
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.163  -0.0150
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.390  -0.0125
#>  7      7     0.21      80    65    19.8     0.041      0.12 -0.0738 -0.0751
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.183   0.0184
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.246   0.390 
#> 10     10     0.21      80    65    19.8     0.041      0.12  0.0503  0.0642
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21     Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.270  19.9 0.041  0.12  79.8  1588. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.270  19.9 0.041  0.12   3.48   69.3
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.273  21.2 0.041  0.12  74.6  1584. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.273  21.2 0.041  0.12   3.16   67.2
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.142  16.3 0.041  0.12 107.   1747. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.142  16.3 0.041  0.12  16.8   274. 
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>