Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object

A posologyr list, created by the posologyr function.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
patient01 <- posologyr(prior_model=mod_run001,
                       dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01,n_simul=100)
#> $eta
#>           ETA_Cl      ETA_Vc       ETA_Ka
#> 1   -0.139628306  0.22569793  0.755714210
#> 2   -0.067442227  0.11022444  0.002164101
#> 3    0.403272430  0.09334127  0.084261088
#> 4   -0.426545423 -0.45272286  1.327223582
#> 5   -0.011305250  0.60976602 -0.576230357
#> 6    0.445331271  0.35012148  0.510934912
#> 7   -0.067237618  0.11109807 -0.406266081
#> 8   -0.149599857 -0.22349912 -0.292946729
#> 9    0.035044053 -0.99091557  0.749637786
#> 10  -0.341895104 -0.22673152 -0.490978294
#> 11  -0.454091404  0.50075535  0.274290619
#> 12   0.003146252 -0.01252064  0.237919659
#> 13  -0.042019802 -0.26843620  0.217882638
#> 14  -0.100247872  0.73993722 -0.674159134
#> 15  -0.566798844  0.30782457 -0.120070368
#> 16   0.305599544 -0.14545879 -0.015227501
#> 17  -0.220813695  0.64982712  0.542593607
#> 18  -0.050763078 -0.09268304  0.008684489
#> 19   0.096203518  0.33982551 -0.622073912
#> 20   0.578411942  0.77585070  0.395931293
#> 21   0.159122174  0.04662136 -0.275877957
#> 22   0.794498330  0.05446868 -0.493062051
#> 23  -0.801227441 -0.37577782 -0.777918683
#> 24   0.059385269 -0.14096825 -0.321383404
#> 25  -0.716154175 -0.60146931 -0.719697625
#> 26   0.333995329  0.34727135 -0.154709374
#> 27  -0.281049646 -0.40863568  0.008421959
#> 28  -0.657273366  0.45233267 -0.461023332
#> 29   0.388560283  0.05844178  0.735413911
#> 30  -0.425582419 -0.92298144  0.525476571
#> 31  -0.256374661  0.42134007  0.920448056
#> 32  -0.021457808  0.30117893  0.496402632
#> 33  -0.489355534 -0.94958738 -0.482475279
#> 34   0.126889985  0.23436303 -0.979836385
#> 35  -0.385509579 -0.31687906 -0.174632551
#> 36   0.152160639  0.72444431  0.667287866
#> 37   0.362399919  0.16544619 -0.199547754
#> 38   0.198854329  0.29509804  0.276265379
#> 39   0.101014829 -0.03421079 -0.167304684
#> 40   0.608340532 -0.22133854 -1.015036217
#> 41   0.695064573  0.08564166 -0.127647020
#> 42   0.981182985 -0.43231702 -0.545552173
#> 43   0.662086562  0.63594423  0.489260907
#> 44   0.568879948  0.90732276 -0.127872569
#> 45  -0.575996899  0.81072486 -0.004462399
#> 46  -0.764059616  0.19170317  0.397969792
#> 47   0.200786788 -0.52129894 -0.406957246
#> 48  -0.442819833 -0.06046027 -0.273928333
#> 49   0.286843917  0.68621676  0.214893918
#> 50   1.032320575 -0.25756071 -0.101885705
#> 51  -0.053015527  0.21687329 -0.169403614
#> 52   0.043401659 -0.52192030  0.325860500
#> 53  -0.120404207  0.06615057 -0.222835034
#> 54   0.529357211 -0.37001364  0.280439822
#> 55   0.079314514  1.04021700  0.458750877
#> 56  -0.393661976 -0.85007761 -0.095633527
#> 57  -0.879890640 -0.50772326 -0.568781709
#> 58   0.351896477 -0.09602567 -0.192624526
#> 59  -0.099068690  0.26794574  0.069830410
#> 60   0.194963694  0.66296985  0.258953005
#> 61   0.369479630  0.48239449  0.127319816
#> 62  -0.425308071 -0.45057951 -0.167897739
#> 63   0.347592833 -0.41813712  0.837005602
#> 64  -0.508887700  0.06047301  0.511859497
#> 65  -0.329366695 -0.34594782  0.184132684
#> 66   0.390617259 -0.85140186  0.603811990
#> 67  -0.013337379  0.45014839  0.207343957
#> 68   0.275430146 -0.40933639  0.277345603
#> 69   0.070891842  0.41081801 -0.152086215
#> 70  -0.219776691  0.37834127 -0.552503274
#> 71   0.230304141  0.59206471 -0.283355393
#> 72   0.023908128  0.03932251  0.600560760
#> 73  -0.150062388  0.71217934  0.798732370
#> 74  -0.113351952 -0.07603207  0.736089774
#> 75   0.120145271  0.47983164 -0.343353634
#> 76   0.160891837 -0.03103689  0.354759995
#> 77  -0.226464181  0.25843332  0.727401025
#> 78   0.713056185 -0.13368917 -0.207846401
#> 79   0.297255613  0.15147882  0.059709668
#> 80  -0.234506308  0.47778593 -0.214551948
#> 81   0.616773875 -0.09505608 -0.082956185
#> 82  -0.197060793  0.19219536 -0.799608044
#> 83   0.883272444  0.34481093 -0.325267570
#> 84   0.619331496  0.44794613 -0.099379737
#> 85   0.876360627 -0.84526007  0.502492817
#> 86   0.565735149  0.50444589  0.003431245
#> 87   0.538800126  0.07683623 -0.078600894
#> 88  -0.673679751 -0.36622180  0.017560202
#> 89   0.558842946  0.68406681  0.353313146
#> 90   0.832170477  0.39375252 -0.088841234
#> 91  -0.294563070  0.22253592  0.105196159
#> 92   0.416311529  0.33644360 -0.065476570
#> 93  -0.056361246  0.31148789 -0.304272419
#> 94   0.018959311  1.20716293 -0.532648351
#> 95  -0.240389326  0.27934137  0.170553359
#> 96   0.380295730  0.14282885 -0.825467579
#> 97   0.078200463  0.37306175 -0.376999360
#> 98  -0.786350828  0.45588140  0.238058377
#> 99   0.416507206  0.16354174  0.241541350
#> 100 -0.050068952 -0.09392201 -0.812984078
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc   ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>    <dbl>
#>  1      1        4       70        1 -0.140   0.226   0.756  
#>  2      2        4       70        1 -0.0674  0.110   0.00216
#>  3      3        4       70        1  0.403   0.0933  0.0843 
#>  4      4        4       70        1 -0.427  -0.453   1.33   
#>  5      5        4       70        1 -0.0113  0.610  -0.576  
#>  6      6        4       70        1  0.445   0.350   0.511  
#>  7      7        4       70        1 -0.0672  0.111  -0.406  
#>  8      8        4       70        1 -0.150  -0.223  -0.293  
#>  9      9        4       70        1  0.0350 -0.991   0.750  
#> 10     10        4       70        1 -0.342  -0.227  -0.491  
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc     depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>     <dbl> <dbl>
#> 1      1     1     4    70     1  3.48  87.7  2.13 0.0397 17.6   4.24e+ 2 1546.
#> 2      1    14     4    70     1  3.48  87.7  2.13 0.0397 13.5  -6.03e-10 1181.
#> 3      2     1     4    70     1  3.74  78.2  1.00 0.0478 13.1   9.53e+ 2 1025.
#> 4      2    14     4    70     1  3.74  78.2  1.00 0.0478 13.9   2.09e- 3 1088.
#> 5      3     1     4    70     1  5.99  76.8  1.09 0.0779 13.9   8.95e+ 2 1068.
#> 6      3    14     4    70     1  5.99  76.8  1.09 0.0779  9.60  6.45e- 4  738.
#> # … with 194 more rows, and 1 more variable: AUC <dbl>
#>