Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                             dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)
#> $eta
#>           ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.030374260 -0.161829148       0       0
#> 2   -0.102757335  0.048017450       0       0
#> 3   -0.008150549 -0.024723836       0       0
#> 4   -0.117610421  0.206626778       0       0
#> 5    0.066439439  0.097090345       0       0
#> 6    0.213708594  0.159437052       0       0
#> 7    0.321362136  0.044435652       0       0
#> 8   -0.096179186 -0.154255395       0       0
#> 9    0.244974916  0.009726151       0       0
#> 10   0.527124438 -0.066609441       0       0
#> 11   0.215629524 -0.206301562       0       0
#> 12  -0.395524970 -0.024635792       0       0
#> 13  -0.423108491 -0.002815486       0       0
#> 14  -0.036561179  0.062359561       0       0
#> 15  -0.212493590  0.184279262       0       0
#> 16   0.402052708  0.070704863       0       0
#> 17  -0.244354721 -0.099235309       0       0
#> 18   0.171028783 -0.033008082       0       0
#> 19   0.315334174 -0.050612273       0       0
#> 20   0.067893262 -0.044504453       0       0
#> 21   0.248223957 -0.079374048       0       0
#> 22   0.441640529  0.123414905       0       0
#> 23  -0.483051206 -0.168737624       0       0
#> 24   0.057396195 -0.015809290       0       0
#> 25  -0.083035502  0.084834367       0       0
#> 26   0.137674938 -0.003867043       0       0
#> 27  -0.356390218 -0.140531269       0       0
#> 28   0.143695791 -0.038486017       0       0
#> 29  -0.145348973  0.066161516       0       0
#> 30  -0.025335540 -0.094573322       0       0
#> 31   0.023528988 -0.155283152       0       0
#> 32   0.220852855  0.162404682       0       0
#> 33  -0.108850074  0.044094932       0       0
#> 34  -0.238835047  0.007383595       0       0
#> 35  -0.544044954 -0.105307495       0       0
#> 36  -0.186237230 -0.064586483       0       0
#> 37  -0.188347449  0.054396025       0       0
#> 38   0.367803271  0.096912212       0       0
#> 39   0.460225596 -0.063812852       0       0
#> 40   0.705953460 -0.002986226       0       0
#> 41   0.220848714 -0.049391243       0       0
#> 42  -0.025285436 -0.125536899       0       0
#> 43   0.067885966 -0.244133196       0       0
#> 44   0.085645963 -0.140270424       0       0
#> 45   0.171447216  0.124162297       0       0
#> 46  -0.233634773 -0.105413850       0       0
#> 47  -0.525354483  0.044217544       0       0
#> 48  -0.012912660  0.017200804       0       0
#> 49  -0.133167297  0.058697791       0       0
#> 50  -0.040745745  0.128093846       0       0
#> 51   0.110493852  0.224889065       0       0
#> 52   0.262841860 -0.146102488       0       0
#> 53  -0.257707697  0.130648326       0       0
#> 54  -0.361050311 -0.082503652       0       0
#> 55   0.080835724  0.062013071       0       0
#> 56   0.290104357  0.114609992       0       0
#> 57  -0.069790081 -0.001859437       0       0
#> 58   0.061186935  0.021653716       0       0
#> 59   0.187636367 -0.064857587       0       0
#> 60  -0.074835169  0.069856619       0       0
#> 61   0.182417617  0.189783479       0       0
#> 62   0.095199149  0.299828340       0       0
#> 63   0.288727978  0.261522298       0       0
#> 64   0.274646218 -0.080774440       0       0
#> 65  -0.095624906 -0.220134643       0       0
#> 66  -0.199602656 -0.141212430       0       0
#> 67  -0.683904325  0.001032125       0       0
#> 68   0.317117559 -0.085713340       0       0
#> 69  -0.465620486  0.130642541       0       0
#> 70   0.446648384  0.033682344       0       0
#> 71  -0.065627796  0.023931800       0       0
#> 72  -0.214365971  0.119276853       0       0
#> 73   0.149222499 -0.133167329       0       0
#> 74  -0.206542281  0.006230759       0       0
#> 75  -0.083048976  0.083639593       0       0
#> 76   0.166792654 -0.023146276       0       0
#> 77   0.344845595 -0.183409714       0       0
#> 78  -0.175779330  0.136673216       0       0
#> 79  -0.034936179  0.109794076       0       0
#> 80  -0.226067975 -0.095538927       0       0
#> 81  -0.206359205  0.022323025       0       0
#> 82   0.029352167  0.031682066       0       0
#> 83  -0.090679955  0.077367307       0       0
#> 84  -0.195951556 -0.025051975       0       0
#> 85   0.103495665  0.089717393       0       0
#> 86  -0.100734151  0.040989191       0       0
#> 87   0.102908303  0.178571252       0       0
#> 88  -0.310028656 -0.108517408       0       0
#> 89   0.282053145  0.039848324       0       0
#> 90  -0.140890473 -0.016880751       0       0
#> 91  -0.421720424  0.130155475       0       0
#> 92   0.289974553 -0.076440556       0       0
#> 93  -0.329446952 -0.139096944       0       0
#> 94  -0.486225149 -0.043552983       0       0
#> 95   0.067428112 -0.101653717       0       0
#> 96   0.073099923 -0.134774983       0       0
#> 97  -0.378331095  0.091771281       0       0
#> 98   0.357881437  0.145689762       0       0
#> 99  -0.373571302  0.198854160       0       0
#> 100 -0.113959392 -0.038782187       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 × 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21   ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>    <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.0304  -0.162  
#>  2      2     0.21      80    65    19.8     0.041      0.12 -0.103    0.0480 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.00815 -0.0247 
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.118    0.207  
#>  5      5     0.21      80    65    19.8     0.041      0.12  0.0664   0.0971 
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.214    0.159  
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.321    0.0444 
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0962  -0.154  
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.245    0.00973
#> 10     10     0.21      80    65    19.8     0.041      0.12  0.527   -0.0666 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 × 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.242  16.6 0.041  0.12 98.0  1622. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.242  16.6 0.041  0.12  5.56   92.0
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.225  20.4 0.041  0.12 80.4  1642. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.225  20.4 0.041  0.12  5.36  109. 
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.247  19.0 0.041  0.12 85.1  1615. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.247  19.0 0.041  0.12  4.59   87.1
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>