Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object = NULL, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul = 0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>           ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.462198173  0.059092596       0       0
#> 2    0.215377300  0.060574719       0       0
#> 3   -0.335236160 -0.261383394       0       0
#> 4   -0.175892330 -0.236890019       0       0
#> 5   -0.169895410 -0.221594098       0       0
#> 6   -0.075787368 -0.075950589       0       0
#> 7    0.107621339 -0.007389852       0       0
#> 8   -0.021621892  0.088628581       0       0
#> 9    0.059297626 -0.079204240       0       0
#> 10   0.427788885 -0.029987274       0       0
#> 11   0.100376919  0.061767566       0       0
#> 12  -0.115090410 -0.057151201       0       0
#> 13   0.299652371  0.073155932       0       0
#> 14   0.201515442 -0.002780858       0       0
#> 15   0.197691593 -0.256933896       0       0
#> 16   0.186151761 -0.054349600       0       0
#> 17   0.003125341  0.185211669       0       0
#> 18  -0.128783383  0.103017566       0       0
#> 19  -0.151598807 -0.130505388       0       0
#> 20   0.073190336 -0.064659602       0       0
#> 21  -0.165867197 -0.010737987       0       0
#> 22   0.715642430  0.176669063       0       0
#> 23  -0.111390825  0.121587208       0       0
#> 24   0.038356470 -0.221529861       0       0
#> 25  -0.236179577  0.056985530       0       0
#> 26  -0.179242442  0.085379839       0       0
#> 27   0.156051514 -0.057857091       0       0
#> 28  -0.546498718 -0.003148088       0       0
#> 29  -0.103424519  0.080955311       0       0
#> 30   0.824672018  0.120595044       0       0
#> 31   0.296721835  0.070723919       0       0
#> 32   0.043586409  0.081618941       0       0
#> 33   0.324569226 -0.005756239       0       0
#> 34  -0.176107558 -0.136148556       0       0
#> 35  -0.121073165  0.005843913       0       0
#> 36  -0.237881174 -0.033525274       0       0
#> 37  -0.611966932 -0.095560512       0       0
#> 38   0.318836559 -0.116860082       0       0
#> 39   0.009024212 -0.140649124       0       0
#> 40   0.252783535  0.006695324       0       0
#> 41   0.655607309 -0.096640117       0       0
#> 42   0.107015655  0.056704681       0       0
#> 43   0.547938501 -0.013390488       0       0
#> 44  -0.214189901  0.146455132       0       0
#> 45   0.433553133  0.202913324       0       0
#> 46  -0.228727976  0.020242290       0       0
#> 47  -0.215784697 -0.207091284       0       0
#> 48  -0.004027379 -0.044516939       0       0
#> 49   0.188802074 -0.047700778       0       0
#> 50  -0.048525145  0.063014179       0       0
#> 51   0.337018751  0.039865399       0       0
#> 52  -0.007289793 -0.165451572       0       0
#> 53  -0.287378520 -0.041079395       0       0
#> 54   0.371996163 -0.014491292       0       0
#> 55   0.146891382  0.220216341       0       0
#> 56  -0.231822609  0.075801301       0       0
#> 57   0.313071484  0.184229084       0       0
#> 58   0.269082516  0.200638203       0       0
#> 59   0.143219576 -0.021402112       0       0
#> 60  -0.690641022  0.108727901       0       0
#> 61  -0.046539674 -0.195214304       0       0
#> 62  -0.106811097  0.148546442       0       0
#> 63   0.247387440  0.117468154       0       0
#> 64  -0.224982701 -0.264423353       0       0
#> 65  -0.170501606 -0.001449898       0       0
#> 66  -0.374201993  0.142803112       0       0
#> 67  -0.054915464 -0.210488394       0       0
#> 68  -0.114027948 -0.007668930       0       0
#> 69  -0.114684712  0.062659699       0       0
#> 70   0.332101048 -0.207670472       0       0
#> 71  -0.172418859  0.192010880       0       0
#> 72  -0.406612458  0.007224421       0       0
#> 73  -0.112814928 -0.144418359       0       0
#> 74  -0.051258319 -0.006877915       0       0
#> 75  -0.017325960  0.027759721       0       0
#> 76  -0.418450219  0.133986984       0       0
#> 77  -0.098281679  0.038237598       0       0
#> 78   0.094798949  0.046097504       0       0
#> 79  -0.156411282  0.053255633       0       0
#> 80   0.651289508 -0.084592298       0       0
#> 81   0.191722714 -0.160891140       0       0
#> 82   0.098499920  0.002118682       0       0
#> 83   0.046435930  0.177040426       0       0
#> 84  -0.512154674  0.001195509       0       0
#> 85  -0.037455316  0.209526304       0       0
#> 86  -0.133499765 -0.110723580       0       0
#> 87   0.425325013  0.029063468       0       0
#> 88   0.010841191 -0.325992340       0       0
#> 89   0.297474480  0.025962045       0       0
#> 90  -0.027950379  0.029667190       0       0
#> 91   0.163824543  0.115279264       0       0
#> 92   0.294413126 -0.092125379       0       0
#> 93   0.268982631 -0.085930559       0       0
#> 94  -0.140872059  0.014673674       0       0
#> 95   0.175334140 -0.059448331       0       0
#> 96   0.096486574 -0.123762038       0       0
#> 97   0.461190065  0.031024993       0       0
#> 98   0.071564554  0.045326937       0       0
#> 99  -0.076344682  0.220973118       0       0
#> 100  0.308052873  0.089055754       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21  ETA_ke    ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>   <dbl>    <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.462   0.0591 
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.215   0.0606 
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.335  -0.261  
#>  4      4     0.21      80    65    19.8     0.041      0.12 -0.176  -0.237  
#>  5      5     0.21      80    65    19.8     0.041      0.12 -0.170  -0.222  
#>  6      6     0.21      80    65    19.8     0.041      0.12 -0.0758 -0.0760 
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.108  -0.00739
#>  8      8     0.21      80    65    19.8     0.041      0.12 -0.0216  0.0886 
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.0593 -0.0792 
#> 10     10     0.21      80    65    19.8     0.041      0.12  0.428  -0.0300 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21     Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.157  20.7 0.041  0.12 20.9   432. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.157  20.7 0.041  0.12  2.78   57.5
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.309  20.7 0.041  0.12 18.6   386. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.309  20.7 0.041  0.12  0.585  12.1
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.178  15.0 0.041  0.12 28.4   425. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.178  15.0 0.041  0.12  3.02   45.2
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>