Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(object = NULL, n_simul = 1000, return_model = TRUE)

Arguments

object
A posologyr list, created by the posologyr function.

n_simul
An integer, the number of simulations to be run. For n_simul = 0, all ETAs are set to 0.

return_model
A boolean. Returns a RxODE model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a dataframe of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a RxODE model using the simulated ETAs.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,0.5,1.0,14.0),
                        DV=c(NA,NA,25.0,5.5),
                        AMT=c(1000,-1000,0,0),
                        EVID=c(10102,10102,0,0),
                        DUR=c(0.5,0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
patient01_tobra <- posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)
# estimate the prior distribution of population parameters
poso_simu_pop(patient01_tobra,n_simul=100)

#> $eta
#>           ETA_ke        ETA_V ETA_k12 ETA_k21
#> 1   -0.680005766  0.014804719       0       0
#> 2    0.110951371 -0.091159910       0       0
#> 3   -0.260920435 -0.069183479       0       0
#> 4    0.050433877  0.060232312       0       0
#> 5    0.151852020 -0.210935995       0       0
#> 6    0.319512471 -0.039919497       0       0
#> 7    0.005109306  0.318304521       0       0
#> 8    0.233805739  0.114527836       0       0
#> 9    0.109250787  0.016823356       0       0
#> 10   0.491609086  0.125276291       0       0
#> 11  -0.234970638 -0.067973486       0       0
#> 12   0.329916506 -0.046731470       0       0
#> 13  -0.188737653 -0.091816688       0       0
#> 14  -0.112124457 -0.236205227       0       0
#> 15  -0.512792707  0.123063613       0       0
#> 16  -0.227416449  0.003483139       0       0
#> 17   0.026310904  0.097568658       0       0
#> 18   0.311185701  0.253049413       0       0
#> 19  -0.041843058  0.041305593       0       0
#> 20  -0.273947407  0.211491718       0       0
#> 21   0.155806368 -0.082672399       0       0
#> 22  -0.293998830  0.167341585       0       0
#> 23  -0.105697099 -0.088283763       0       0
#> 24  -0.233989306 -0.083287953       0       0
#> 25   0.073378206 -0.001554476       0       0
#> 26  -0.112814752  0.072873278       0       0
#> 27   0.340408165 -0.018729595       0       0
#> 28   0.215606611  0.130081523       0       0
#> 29  -0.304070347 -0.002813693       0       0
#> 30  -0.247722014 -0.110921530       0       0
#> 31   0.053727683  0.143582003       0       0
#> 32   0.519854462  0.056696711       0       0
#> 33   0.170933909 -0.089478220       0       0
#> 34  -0.116408097  0.120838426       0       0
#> 35   0.085848716  0.103859692       0       0
#> 36  -0.249808237  0.055279478       0       0
#> 37  -0.232917434 -0.266571625       0       0
#> 38   0.409124183 -0.017963249       0       0
#> 39   0.139815108 -0.041226638       0       0
#> 40   0.207288260  0.095485889       0       0
#> 41   0.790255821 -0.086838172       0       0
#> 42   0.321392563 -0.065809717       0       0
#> 43   0.569273971 -0.144433403       0       0
#> 44  -0.179629313 -0.021196112       0       0
#> 45  -0.006148376 -0.044012178       0       0
#> 46  -0.548139448 -0.044265674       0       0
#> 47   0.398454055  0.128183402       0       0
#> 48   0.072963864 -0.066549733       0       0
#> 49   0.266232093 -0.156943101       0       0
#> 50   0.268816004 -0.043543980       0       0
#> 51   0.403802726 -0.019784538       0       0
#> 52   0.046765848 -0.006687424       0       0
#> 53   0.159946186 -0.161512205       0       0
#> 54   0.460721350 -0.037934478       0       0
#> 55   0.157573404  0.036590224       0       0
#> 56  -0.034695658 -0.060371174       0       0
#> 57   0.018653265  0.251382841       0       0
#> 58  -0.091374729  0.074000579       0       0
#> 59   0.092760210  0.038018714       0       0
#> 60  -0.463198671  0.017923206       0       0
#> 61   0.137478655 -0.197679980       0       0
#> 62  -0.162264491 -0.014456895       0       0
#> 63  -0.211640651 -0.051527853       0       0
#> 64   0.127141721  0.164165719       0       0
#> 65  -0.447543199  0.004184452       0       0
#> 66   0.170865660  0.114818340       0       0
#> 67  -0.204639735 -0.010788207       0       0
#> 68  -0.310182195  0.063232523       0       0
#> 69   0.172181426  0.113636727       0       0
#> 70   0.582846086  0.103821138       0       0
#> 71   0.605252949 -0.054373390       0       0
#> 72   0.093711844  0.067674964       0       0
#> 73   0.285542079  0.037241609       0       0
#> 74   0.226942734  0.178008808       0       0
#> 75  -0.346027340  0.027622279       0       0
#> 76  -0.323571061 -0.029467299       0       0
#> 77   0.206275392  0.118901357       0       0
#> 78  -0.028842460  0.021561188       0       0
#> 79  -0.252042859 -0.196028377       0       0
#> 80  -0.161380184  0.120785005       0       0
#> 81  -0.246721513 -0.231197949       0       0
#> 82   0.169550460 -0.082369416       0       0
#> 83   0.588756009 -0.033063445       0       0
#> 84   0.296403683  0.091610093       0       0
#> 85  -0.079403546 -0.177431008       0       0
#> 86  -0.174672792 -0.188339789       0       0
#> 87  -0.171328120  0.109720037       0       0
#> 88  -0.131409904 -0.137166922       0       0
#> 89  -0.201433560 -0.042099166       0       0
#> 90   0.684213216  0.066052112       0       0
#> 91   0.085863669 -0.035325450       0       0
#> 92  -0.066580469 -0.054688746       0       0
#> 93  -0.293810505 -0.009721644       0       0
#> 94   0.537366366  0.128086186       0       0
#> 95  -0.270201726  0.146467713       0       0
#> 96  -0.029647480  0.226349683       0       0
#> 97   0.121970690  0.152599934       0       0
#> 98  -0.011350287  0.074588276       0       0
#> 99  -0.241434065 -0.031367709       0       0
#> 100  0.217330560 -0.002941455       0       0
#> 
#> $model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#> # A tibble: 100 x 9
#>    sim.id THETA_ke CLCREAT    WT THETA_V THETA_k12 THETA_k21   ETA_ke   ETA_V
#>     <int>    <dbl>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>    <dbl>   <dbl>
#>  1      1     0.21      80    65    19.8     0.041      0.12 -0.680    0.0148
#>  2      2     0.21      80    65    19.8     0.041      0.12  0.111   -0.0912
#>  3      3     0.21      80    65    19.8     0.041      0.12 -0.261   -0.0692
#>  4      4     0.21      80    65    19.8     0.041      0.12  0.0504   0.0602
#>  5      5     0.21      80    65    19.8     0.041      0.12  0.152   -0.211 
#>  6      6     0.21      80    65    19.8     0.041      0.12  0.320   -0.0399
#>  7      7     0.21      80    65    19.8     0.041      0.12  0.00511  0.318 
#>  8      8     0.21      80    65    19.8     0.041      0.12  0.234    0.115 
#>  9      9     0.21      80    65    19.8     0.041      0.12  0.109    0.0168
#> 10     10     0.21      80    65    19.8     0.041      0.12  0.492    0.125 
#> # … with 90 more rows
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 200 x 14
#>   sim.id  time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21     Cc centr
#>    <int> <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>
#> 1      1     1 -1.39  2.97  -3.19  -2.12 0.126  19.8 0.041  0.12 22.4   442. 
#> 2      1    14 -1.39  2.97  -3.19  -2.12 0.126  19.8 0.041  0.12  4.15   82.0
#> 3      2     1 -1.39  2.97  -3.19  -2.12 0.278  17.8 0.041  0.12 22.2   395. 
#> 4      2    14 -1.39  2.97  -3.19  -2.12 0.278  17.8 0.041  0.12  0.903  16.0
#> 5      3     1 -1.39  2.97  -3.19  -2.12 0.192  18.2 0.041  0.12 23.2   421. 
#> 6      3    14 -1.39  2.97  -3.19  -2.12 0.192  18.2 0.041  0.12  2.14   38.9
#> # … with 194 more rows, and 2 more variables: periph <dbl>, AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#>