Estimate the prior distribution of population parameters

Estimates the prior distribution of population parameters by Monte Carlo simulations

poso_simu_pop(
  dat = NULL,
  prior_model = NULL,
  n_simul = 1000,
  return_model = TRUE
)

Arguments

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

n_simul

An integer, the number of simulations to be run. For n_simul =0, all ETAs are set to 0.

return_model

A boolean. Returns a rxode2 model using the simulated ETAs if set to TRUE.

Value

If return_model is set to FALSE, a list of one element: a dataframe $eta of the individual values of ETA. If return_model is set to TRUE, a list of the dataframe of the individual values of ETA, and a rxode2 model using the simulated ETAs.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# estimate the prior distribution of population parameters
poso_simu_pop(dat=df_patient01,prior_model=mod_run001,n_simul=100)
#> $eta
#>           ETA_Cl       ETA_Vc      ETA_Ka
#> 1   -0.405685873  0.255727046 -0.19994987
#> 2    0.168863232  0.362470457  0.36978196
#> 3   -0.397274612 -0.025681302  0.38382423
#> 4    0.300818507  0.838894622  0.51024836
#> 5   -0.593018012  0.322290611  0.12563880
#> 6   -0.094713710  0.193955707 -0.23453661
#> 7   -0.307652180 -0.413525151 -0.25415091
#> 8    0.225731088 -0.381095872  0.47093112
#> 9   -0.116372383  0.160926553  0.10695373
#> 10   0.089726221  0.053581916  0.53553320
#> 11   0.261037582  0.105805281 -0.10772076
#> 12  -0.451417238  0.091871003 -0.22624305
#> 13  -0.529238757 -0.068759016  0.14064092
#> 14  -0.473099562 -0.336727386  0.20104400
#> 15   0.401830945  0.071903819 -0.40063230
#> 16  -0.353078928 -0.358343735  0.52607602
#> 17   0.285518170  0.469113124 -1.00410888
#> 18  -0.004799354  0.098936377 -0.14999348
#> 19  -0.139963889  0.341711624 -0.70195938
#> 20   0.756648984  0.166537053 -0.76328425
#> 21  -0.646615298  0.320843667  0.86904169
#> 22   0.305218188 -0.069892099 -0.19948560
#> 23   0.040831534 -0.028159882  0.75580429
#> 24   0.301294673 -0.502043299 -0.40771127
#> 25   0.278066119  0.001336468  0.11578399
#> 26  -0.131118458 -0.083852708  0.03019897
#> 27   0.669627884 -0.229700953  0.05497179
#> 28  -0.120985201 -0.277572997  0.13633472
#> 29  -0.047814106 -0.210900395 -0.03367084
#> 30   0.038222927 -0.079867392  0.31473728
#> 31  -0.838271343 -0.086193967 -0.32549485
#> 32  -0.229816655 -0.551278828  0.42453581
#> 33   0.262472280 -0.314042381 -0.23578973
#> 34  -0.480036824 -0.019650865  0.32182896
#> 35  -0.211150040  0.013837432 -0.25612592
#> 36   0.248828516  0.126327979 -0.38885632
#> 37  -0.373563389  0.089796875 -0.43369736
#> 38   0.322363643  0.511216606 -0.02644541
#> 39  -0.643179708 -0.064446830  0.04386968
#> 40  -0.044162963  0.862266584 -0.74980888
#> 41   0.289650436  0.092206489  0.52547264
#> 42   0.966601613 -0.365461121  0.03297253
#> 43   0.264271990  0.302167657 -0.64038830
#> 44  -0.043747700  0.040431782 -0.01957620
#> 45  -0.127791961 -0.261948063 -0.39546700
#> 46   0.029649086 -0.342314641  0.24409704
#> 47   0.275299027 -0.091149872 -0.29137703
#> 48  -0.238015074  0.156354784  0.43957466
#> 49   0.203048017 -0.717635153 -0.77542924
#> 50   0.285532177 -0.869346814 -0.29224191
#> 51   0.202781671  0.187541692  0.11178506
#> 52   0.216243902  0.222473391 -0.67473809
#> 53   0.506165217 -0.034542599 -0.03055766
#> 54  -0.219306520  0.110051707  0.09309839
#> 55  -0.046686266  0.137879259 -0.28200030
#> 56  -0.006276491  0.182935981  0.37301158
#> 57   0.605869628 -0.184665609  0.38401660
#> 58  -0.069895369  0.234078584 -0.19646084
#> 59  -0.694746813  0.198764143 -0.13836866
#> 60  -0.237380814 -0.978029888  0.32981412
#> 61   0.017306712 -0.500146243  0.56021582
#> 62   0.027218891 -0.270374998  0.01604977
#> 63   0.630428858 -0.683705763 -0.78858996
#> 64  -0.225972160 -1.281142858 -0.49715821
#> 65  -0.603456747  0.947598042 -0.38490606
#> 66   0.060522260 -0.372518557  0.56220804
#> 67   0.194349803 -0.414470444  0.51841969
#> 68   0.356872010  0.413791584 -0.04697531
#> 69  -0.263396625 -0.044948193  0.62095434
#> 70   0.155155536  0.567951941  0.13047124
#> 71  -0.085047125  0.151981605  0.09265632
#> 72   0.042892431  0.122012727 -0.05298408
#> 73  -0.514237959  0.129619659 -1.22214162
#> 74   0.558987418  0.480893210 -0.14276880
#> 75   0.209894366  0.426497688 -0.37816601
#> 76  -0.067459369 -0.574746547 -0.49992508
#> 77  -0.194199400  0.143188463 -0.60799019
#> 78  -0.323650512  0.105670465  0.17196343
#> 79   0.767931892  0.676420488  0.02158285
#> 80  -0.223061147 -0.162688216 -0.88181021
#> 81   0.310689610 -0.595051248  0.06945858
#> 82   0.773209170  0.059206792  0.41502296
#> 83  -0.170130618 -0.090561497 -0.01888843
#> 84   0.301699965  0.166102897 -0.16159094
#> 85   0.146137057 -0.355585943  0.28340259
#> 86  -0.481017347 -0.013319583 -0.14784661
#> 87   0.165211246  0.502416410 -0.22257059
#> 88   0.242502764 -0.108216504 -0.17939949
#> 89  -0.605324417 -0.547328146 -0.42436153
#> 90  -0.180478592 -0.353612222  0.27601742
#> 91  -0.453667929 -0.775820891  0.35124463
#> 92  -0.335635693  0.358789784 -0.12666541
#> 93   0.684363813  0.075004843  0.37758052
#> 94  -0.210610255 -0.016802484  0.57048132
#> 95   0.076725310  0.636690142  0.53890615
#> 96  -0.023863121 -0.132823865 -0.60013555
#> 97  -0.506439688  0.126798566  0.82510155
#> 98   0.003155287 -0.258136441 -0.48321102
#> 99   0.060681301  0.171496959  0.30772064
#> 100  0.571632214  0.085845212 -0.30852275
#> 
#> $model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> # A tibble: 100 × 7
#>    sim.id THETA_Cl THETA_Vc THETA_Ka  ETA_Cl  ETA_Vc ETA_Ka
#>     <int>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>  <dbl>
#>  1      1        4       70        1 -0.406   0.256  -0.200
#>  2      2        4       70        1  0.169   0.362   0.370
#>  3      3        4       70        1 -0.397  -0.0257  0.384
#>  4      4        4       70        1  0.301   0.839   0.510
#>  5      5        4       70        1 -0.593   0.322   0.126
#>  6      6        4       70        1 -0.0947  0.194  -0.235
#>  7      7        4       70        1 -0.308  -0.414  -0.254
#>  8      8        4       70        1  0.226  -0.381   0.471
#>  9      9        4       70        1 -0.116   0.161   0.107
#> 10     10        4       70        1  0.0897  0.0536  0.536
#> # ℹ 90 more rows
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> 
#> Simulation without uncertainty in parameters, omega, or sigma matricies
#> 
#> ── First part of data (object): ──
#> # A tibble: 200 × 13
#>   sim.id  time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc      depot centr
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>      <dbl> <dbl>
#> 1      1     1     4    70     1  2.67  90.4 0.819 0.0295  9.94    1.09e+3  899.
#> 2      1    14     4    70     1  2.67  90.4 0.819 0.0295 15.3     2.60e-2 1383.
#> 3      2     1     4    70     1  4.74 101.  1.45  0.0471 12.7     6.90e+2 1282.
#> 4      2    14     4    70     1  4.74 101.  1.45  0.0471 10.8     4.65e-6 1082.
#> 5      3     1     4    70     1  2.69  68.2 1.47  0.0394 19.0     6.80e+2 1296.
#> 6      3    14     4    70     1  2.69  68.2 1.47  0.0394 17.5     3.51e-6 1195.
#> # ℹ 194 more rows
#> # ℹ 1 more variable: AUC <dbl>
#>