Creates a posologyr list from a prior model and an individual event record

Creates a list for a posologyr prior model, an individual event record, and an rxSolve solve object, created from the prior ppk model and the individual event record.

posologyr(prior_model = NULL, dat = NULL, nocb = FALSE)

Arguments

prior_model

A posologyr prior population pharmacokinetics model, a list of six objects.

dat

Dataframe. An individual subject dataset following the structure of NONMEM/rxode2 event records.

nocb

A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. If FALSE, the last observation carried forward (locf) style will be used. Defaults to FALSE.

posologyr will check the validity of the compiled rxode2 model. If prior_model$ppk_model$isValid() returns FALSE, posologyr will call rxode2 to recompile the model before solving it using prior_model and dat.

Value

A list of eight objects: the posologyr prior population pharmacokinetics model given as prior_model parameter (6 objects), the individual event record (tdm_data) given as dat parameter, and the solved model (solved_ppk_model).

ppk_model

A rxode2 model implementing the structural population pharmacokinetics model with the individual model (i.e. the model of inter-individual variability) and the covariates

error_model

A function of the residual error model

theta

A named vector of the population estimates of the fixed effects parameters (called THETAs, following NONMEM terminology)

omega

A named square variance-covariance matrix of the population parameters inter-individual variability

covariates

A character vector of the covariates of the model

sigma

The estimates of the parameters of the residual error model

tdm_data

A dataframe. The individual subject dataset given as dat parameter

solved_ppk_model

An rxSolve solve object, created with prior_ppk_model and using dat as the event record.

Examples

# model
mod_run001 <- list(
ppk_model = rxode2::rxode({
  centr(0) = 0;
  depot(0) = 0;

  TVCl = THETA_Cl;
  TVVc = THETA_Vc;
  TVKa = THETA_Ka;

  Cl = TVCl*exp(ETA_Cl);
  Vc = TVVc*exp(ETA_Vc);
  Ka = TVKa*exp(ETA_Ka);

  K20 = Cl/Vc;
  Cc = centr/Vc;

  d/dt(depot) = -Ka*depot;
  d/dt(centr) = Ka*depot - K20*centr;
  d/dt(AUC) = Cc;
}),
error_model = function(f,sigma) {
  dv <- cbind(f,1)
  g <- diag(dv%*%sigma%*%t(dv))
  return(sqrt(g))
},
theta = c(THETA_Cl=4.0, THETA_Vc=70.0, THETA_Ka=1.0),
omega = lotri::lotri({ETA_Cl + ETA_Vc + ETA_Ka ~
    c(0.2,
      0, 0.2,
      0, 0, 0.2)}),
sigma = lotri::lotri({prop + add ~ c(0.05,0.0,0.00)}))
#>  
# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA))
# loading the model and Patient01's event record
posologyr(prior_model=mod_run001,
          dat=df_patient01)
#> $ppk_model
#> rxode2 NA model named rx_5d00d8f882dfd041ea778d7db6a4391f model (✔ ready). 
#> $state: depot, centr, AUC
#> $params: THETA_Cl, THETA_Vc, THETA_Ka, ETA_Cl, ETA_Vc, ETA_Ka
#> $lhs: TVCl, TVVc, TVKa, Cl, Vc, Ka, K20, Cc
#> 
#> $error_model
#> function(f,sigma) {
#>   dv <- cbind(f,1)
#>   g <- diag(dv%*%sigma%*%t(dv))
#>   return(sqrt(g))
#> }
#> <environment: 0x5584c6cc7fa8>
#> 
#> $theta
#> THETA_Cl THETA_Vc THETA_Ka 
#>        4       70        1 
#> 
#> $omega
#>        ETA_Cl ETA_Vc ETA_Ka
#> ETA_Cl    0.2    0.0    0.0
#> ETA_Vc    0.0    0.2    0.0
#> ETA_Ka    0.0    0.0    0.2
#> 
#> $sigma
#>      prop add
#> prop 0.05   0
#> add  0.00   0
#> 
#> $tdm_data
#>   ID TIME   DV  AMT EVID DUR
#> 1  1    0   NA 2000    1 0.5
#> 2  1    1 25.0    0    0  NA
#> 3  1   14  5.5    0    0  NA
#> 
#> $solved_ppk_model
#> ── Solved rxode2 object ──
#> ── Parameters ($params): ──
#> THETA_Cl THETA_Vc THETA_Ka   ETA_Cl   ETA_Vc   ETA_Ka 
#>        4       70        1        0        0        0 
#> ── Initial Conditions ($inits): ──
#> depot centr   AUC 
#>     0     0     0 
#> ── First part of data (object): ──
#> # A tibble: 2 × 12
#>    time  TVCl  TVVc  TVKa    Cl    Vc    Ka    K20    Cc     depot centr    AUC
#>   <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <dbl>     <dbl> <dbl>  <dbl>
#> 1     1     4    70     1     4    70     1 0.0571  14.6 955.      1020.   6.39
#> 2    14     4    70     1     4    70     1 0.0571  13.8   0.00216  967. 258.  
#> 
#> $interpolation
#> [1] "locf"
#>