Creates a posologyr list from a prior model and an individual event record

Creates a list for a posologyr prior model, an individual event record, and an rxSolve solve object, created from the prior ppk model and the individual event record.

posologyr(prior_model = NULL, dat = NULL, nocb = TRUE)

Arguments

prior_model
A posologyr prior population pharmacokinetics model, a list of six objects.

dat
Dataframe. An individual subject dataset following the structure of NONMEM/RxODE event records.

nocb
A boolean. for time-varying covariates: the next observation carried backward (nocb) interpolation style, similar to NONMEM. Defaults to TRUE. If FALSE, the last observation carried forward (locf) style will be used.

posologyr will check the validity of the compiled RxODE model. If prior_model$ppk_model$isValid() returns FALSE, posologyr will call RxODE to recompile the model before solving it using prior_model and dat.

Value

A list of eight objects: the posologyr prior population pharmacokinetics model given as prior_model parameter (6 objects), the individual event record (tdm_data) given as dat parameter, and the solved model (solved_ppk_model).

ppk_model

A RxODE model implementing the structural population pharmacokinetics model with the individual model (i.e. the model of inter-individual variability) and the covariates

error_model

A function of the residual error model

theta

A named vector of the population estimates of the fixed effects parameters (called THETAs, following NONMEM terminology)

omega

A named square variance-covariance matrix of the population parameters inter-individual variability

covariates

A character vector of the covariates of the model

sigma

The estimates of the parameters of the residual error model

tdm_data

A dataframe. The individual subject dataset given as dat parameter

solved_ppk_model

An rxSolve solve object, created with prior_ppk_model and using dat as the event record.

Examples

# df_patient01: event table for Patient01, following a 30 minutes intravenous
# infusion of tobramycin
df_patient01 <- data.frame(ID=1,
                        TIME=c(0.0,1.0,14.0),
                        DV=c(NA,25.0,5.5),
                        AMT=c(2000,0,0),
                        EVID=c(1,0,0),
                        DUR=c(0.5,NA,NA),
                        CLCREAT=80,WT=65)
# loading a tobramycin model and Patient01's event record
posologyr(prior_model=mod_tobramycin_2cpt_fictional,
                                dat=df_patient01)

#> $ppk_model
#> RxODE NA model named rx_0a92bbb9235fd902ddfefecdc7b25627 model (✓ ready). 
#> $state: centr, periph, AUC
#> $params: THETA_ke, CLCREAT, WT, THETA_V, THETA_k12, THETA_k21, ETA_ke, ETA_V
#> $lhs: tTVke, tTVV, tTVk12, tTVk21, ke, V, k12, k21, Cc
#> 
#> $error_model
#> function (f, sigma) 
#> {
#>     g <- sigma[1] + sigma[2] * f
#>     return(g)
#> }
#> <bytecode: 0x558eae236750>
#> 
#> $theta
#>  THETA_ke   THETA_V THETA_k12 THETA_k21 
#>     0.210    19.800     0.041     0.120 
#> 
#> $omega
#>          ETA_ke   ETA_V ETA_k12 ETA_k21
#> ETA_ke  0.08075 0.00000       0       0
#> ETA_V   0.00000 0.01203       0       0
#> ETA_k12 0.00000 0.00000       0       0
#> ETA_k21 0.00000 0.00000       0       0
#> 
#> $covariates
#> [1] "CLCREAT" "WT"     
#> 
#> $sigma
#>     additive_a proportional_b 
#>          0.000          0.198 
#> 
#> $tdm_data
#>   ID TIME   DV  AMT EVID DUR CLCREAT WT
#> 1  1    0   NA 2000    1 0.5      80 65
#> 2  1    1 25.0    0    0  NA      80 65
#> 3  1   14  5.5    0    0  NA      80 65
#> 
#> $solved_ppk_model
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂ Solved RxODE object ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> ── Parameters ($params): ───────────────────────────────────────────────────────
#>  THETA_ke   THETA_V THETA_k12 THETA_k21    ETA_ke     ETA_V 
#>     0.210    19.800     0.041     0.120     0.000     0.000 
#> ── Initial Conditions ($inits): ────────────────────────────────────────────────
#>  centr periph    AUC 
#>      0      0      0 
#> ── First part of data (object): ────────────────────────────────────────────────
#> # A tibble: 2 × 13
#>    time tTVke  tTVV tTVk12 tTVk21    ke     V   k12   k21    Cc  centr periph
#>   <dbl> <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl>  <dbl>
#> 1     1 -1.39  2.97  -3.19  -2.12 0.249  19.5 0.041  0.12 82.9  1613.    52.5
#> 2    14 -1.39  2.97  -3.19  -2.12 0.249  19.5 0.041  0.12  4.38   85.3   96.2
#> # … with 1 more variable: AUC <dbl>
#> ▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂
#> 
#> $interpolation
#> [1] "nocb"
#>